Crate script

The script crate contains all matters DOM.

Modules

conversions 🔒

css 🔒

dom 🔒

The implementation of the DOM.

drag 🔒

engine 🔒

event_loop 🔒

fetch 🔒

layout_dom

A safe wrapper for DOM nodes that prevents layout from mutating the DOM, from letting DOM nodes escape, and from generally doing anything that it isn’t supposed to. This is accomplished via a simple whitelist of allowed operations, along with some lifetime magic to prevent nodes from escaping.

messaging 🔒

mime 🔒

modules 🔒

navigation 🔒

The listener that encapsulates all state for an in-progress document request. Any redirects that are encountered are followed. Whenever a non-redirect response is received, it is forwarded to the appropriate script thread.

realms 🔒

routed_promise 🔒

runtime 🔒

tasks 🔒

test

unminify 🔒

url 🔒

window_named_properties 🔒

xpath 🔒

Bindings to the xpath crate

Macros

make_atomic_setter

make_bool_getter

make_bool_setter

make_clamped_uint_setter

make_dimension_setter

make_dimension_uint_getter

make_dimension_uint_setter

make_form_action_getter

make_getter

make_int_getter

make_int_setter

make_labels_getter

make_legacy_color_setter

make_limited_int_setter

make_limited_uint_setter

make_nonzero_dimension_setter

make_setter

make_uint_getter

make_uint_setter

make_url_getter

make_url_setter

maplike

Usage:

native_fn

Defines a macro native_fn! to create a JavaScript function from a Rust function pointer.

native_raw_obj_fn

Defines a macro native_raw_obj_fn! to create a raw JavaScript function object.

setlike

Usage:

Structs

JSEngineSetup

ScriptThread

ServiceWorkerManager

A structure managing all registrations and workers for a given origin.

Functions

init