Struct script::animations::Animations

pub(crate) struct Animations {
    pub sets: DocumentAnimationSet,
    has_running_animations: Cell<bool>,
    rooted_nodes: DomRefCell<FxHashMap<NoTrace<OpaqueNode>, Dom<Node>>>,
    pending_events: DomRefCell<Vec<TransitionOrAnimationEvent>>,
    timeline_value_at_last_dirty: Cell<f64>,
}

The set of animations for a document.

Fields
sets: DocumentAnimationSet

The map of nodes to their animation states.

has_running_animations: Cell<bool>

Whether or not we have animations that are running.

rooted_nodes: DomRefCell<FxHashMap<NoTrace<OpaqueNode>, Dom<Node>>>

A list of nodes with in-progress CSS transitions or pending events.

pending_events: DomRefCell<Vec<TransitionOrAnimationEvent>>

A list of pending animation-related events.

timeline_value_at_last_dirty: Cell<f64>

The timeline value at the last time all animations were marked dirty. This is used to prevent marking animations dirty when the timeline has not changed.

Implementations

impl Animations

pub(crate) fn new() -> Self

pub(crate) fn clear(&self)

pub(crate) fn mark_animating_nodes_as_dirty( &self, current_timeline_value: f64 ) -> bool

pub(crate) fn update_for_new_timeline_value(&self, window: &Window, now: f64)

pub(crate) fn cancel_animations_for_node(&self, node: &Node)

Cancel animations for the given node, if any exist.

pub(crate) fn do_post_reflow_update(&self, window: &Window, now: f64)

Processes any new animations that were discovered after reflow. Collect messages that trigger events for any animations that changed state.

fn update_running_animations_presence(&self, window: &Window, new_value: bool)

fn handle_animation_presence_or_pending_events_change(&self, window: &Window)

pub(crate) fn running_animation_count(&self) -> usize

fn start_pending_animations( &self, key: &AnimationSetKey, set: &mut ElementAnimationSet, now: f64, pipeline_id: PipelineId )

Walk through the list of pending animations and start all of the ones that have left the delay phase.

fn finish_running_animations( &self, key: &AnimationSetKey, set: &mut ElementAnimationSet, now: f64, pipeline_id: PipelineId )

Walk through the list of running animations and remove all of the ones that have ended.

fn handle_canceled_animations( &self, key: &AnimationSetKey, set: &mut ElementAnimationSet, now: f64, pipeline_id: PipelineId )

Send events for canceled animations. Currently this only handles canceled transitions, but eventually this should handle canceled CSS animations as well.

fn handle_new_animations( &self, key: &AnimationSetKey, set: &mut ElementAnimationSet, now: f64, pipeline_id: PipelineId )

fn root_newly_animating_dom_nodes( &self, sets: &FxHashMap<AnimationSetKey, ElementAnimationSet> )

Ensure that all nodes with new animations are rooted. This should be called immediately after a restyle, to ensure that these addresses are still valid.

fn unroot_unused_nodes( &self, sets: &FxHashMap<AnimationSetKey, ElementAnimationSet> )

fn add_transition_event( &self, key: &AnimationSetKey, transition: &Transition, event_type: TransitionOrAnimationEventType, now: f64, pipeline_id: PipelineId )

fn add_animation_event( &self, key: &AnimationSetKey, animation: &Animation, event_type: TransitionOrAnimationEventType, now: f64, pipeline_id: PipelineId )

pub(crate) fn send_pending_events(&self, window: &Window)