Module script::cors

A partial implementation of CORS For now this library is XHR-specific. For stuff involving <img>, <iframe>, <form>, etc please check what the request mode should be and compare with the fetch spec This library will eventually become the core of the Fetch crate with CORSRequest being expanded into FetchRequest (etc)

Structs

CORSCache
A CORS cache object. Anchor it somewhere to the user agent.

CORSCacheEntry

CORSRequest

CORSResponse

Enums

HeaderOrMethod
Union type for CORS cache entries Each entry might pertain to a header or method

RequestMode
http://fetch.spec.whatwg.org/#concept-request-mode This only covers some of the request modes. The same-origin and no CORS modes are unnecessary for XHR.

Functions

allow_cross_origin_request
Perform a CORS check on a header list and CORS request http://fetch.spec.whatwg.org/#cors-check