Struct script::cors::CORSRequest

pub struct CORSRequest {
    pub origin: Url,
    pub destination: Url,
    pub mode: RequestMode,
    pub method: Method,
    pub headers: Headers,
    pub preflight_flag: bool,
}

Fields

origin

destination

mode

method

headers

preflight_flag

CORS preflight flag (http://fetch.spec.whatwg.org/#concept-http-fetch) Indicates that a CORS preflight request and/or cache check is to be performed

Methods

impl CORSRequest

fn maybe_new(referer: Url, destination: Url, mode: RequestMode, method: Method, headers: Headers) -> Result<Option<CORSRequest>, ()>

Creates a CORS request if necessary. Will return an error when fetching is forbidden

fn http_fetch(&self) -> CORSResponse

http://fetch.spec.whatwg.org/#concept-http-fetch This method assumes that the CORS flag is set This does not perform the full HTTP fetch, rather it handles part of the CORS filtering if self.mode is ForcedPreflight, then the CORS-with-forced-preflight fetch flag is set as well