script::css::stylesheet_loader

Trait StylesheetOwner

pub(crate) trait StylesheetOwner {
    // Required methods
    fn parser_inserted(&self) -> bool;
    fn potentially_render_blocking(&self) -> bool;
    fn referrer_policy(&self, cx: &mut JSContext) -> ReferrerPolicy;
    fn increment_pending_loads_count(&self);
    fn load_finished(&self, successful: bool) -> Option<bool>;
    fn set_origin_clean(&self, cx: &mut JSContext, origin_clean: bool);
}

Required Methods

fn parser_inserted(&self) -> bool

Returns whether this element was inserted by the parser (i.e., it should trigger a document-load-blocking load).

fn potentially_render_blocking(&self) -> bool

https://html.spec.whatwg.org/multipage/#potentially-render-blocking

fn referrer_policy(&self, cx: &mut JSContext) -> ReferrerPolicy

Which referrer policy should loads triggered by this owner follow

fn increment_pending_loads_count(&self)

Notes that a new load is pending to finish.

fn load_finished(&self, successful: bool) -> Option<bool>

Returns None if there are still pending loads, or whether any load has failed since the loads started.

fn set_origin_clean(&self, cx: &mut JSContext, origin_clean: bool)

Sets origin_clean flag.