script

Module document_collection

Structs

DocumentCollection 🔒

The collection of all Documents managed by the crate::script_thread::ScriptThread. This is stored as a mapping of PipelineId to Document, but for updating the rendering, Documents should be processed in order via Self::documents_in_order.

DocumentTree 🔒

A tree representation of Documents managed by the ScriptThread, which is used to generate an ordered set of Documents for the update the rendering step of the HTML5 specification.

DocumentTreeNode 🔒

DocumentsIter 🔒