Struct script::document_loader::DocumentLoader

pub struct DocumentLoader {
    resource_threads: ResourceThreads,
    blocking_loads: Vec<LoadType>,
    events_inhibited: bool,
    cancellers: Vec<FetchCanceller>,
}

Fields
resource_threads: ResourceThreadsblocking_loads: Vec<LoadType>events_inhibited: boolcancellers: Vec<FetchCanceller>
Implementations

impl DocumentLoader

pub fn new(existing: &DocumentLoader) -> DocumentLoader

pub fn new_with_threads( resource_threads: ResourceThreads, initial_load: Option<ServoUrl>, ) -> DocumentLoader

pub fn cancel_all_loads(&mut self) -> bool

fn add_blocking_load(&mut self, load: LoadType)

Add a load to the list of blocking loads.

pub fn fetch_async_with_callback( &mut self, load: LoadType, request: RequestBuilder, callback: BoxedFetchCallback, )

Initiate a new fetch given a response callback.

pub fn fetch_async_background( &mut self, request: RequestBuilder, callback: BoxedFetchCallback, cancel_override: Option<IpcReceiver<()>>, )

Initiate a new fetch that does not block the document load event.

pub fn finish_load(&mut self, load: &LoadType)

Mark an in-progress network request complete.

pub fn is_blocked(&self) -> bool

pub fn is_only_blocked_by_iframes(&self) -> bool

pub fn inhibit_events(&mut self)

pub fn events_inhibited(&self) -> bool

pub fn resource_threads(&self) -> &ResourceThreads