Struct script::document_loader::LoadBlocker

pub struct LoadBlocker {
    doc: JS<Document>,
    load: Option<LoadType>,
}

Canary value ensuring that manually added blocking loads (ie. ones that weren't created via DocumentLoader::prepare_async_load) are always removed by the time that the owner is destroyed.

Fields

doc

The document whose load event is blocked by this object existing.

load

The load that is blocking the document's load event.

Methods

impl LoadBlocker

fn new(doc: &Document, load: LoadType) -> LoadBlocker

Mark the document's load event as blocked on this new load.

fn terminate(blocker: &mut Option<LoadBlocker>)

Remove this load from the associated document's list of blocking loads.

fn url(&self) -> Option<&Url>

Return the url associated with this load.