Module script::dom::abstractworkerglobalscope

Structs

SendableWorkerScriptChan
A ScriptChan that can be cloned freely and will silently send a TrustedWorkerAddress with common event loop messages. While this SendableWorkerScriptChan is alive, the associated Worker object will remain alive.

WorkerThreadWorkerChan
A ScriptChan that can be cloned freely and will silently send a TrustedWorkerAddress with worker event loop messages. While this SendableWorkerScriptChan is alive, the associated Worker object will remain alive.

Traits

WorkerEventLoopMethods

Functions

run_worker_event_loop