script::dom::audio::baseaudiocontext

Struct BaseAudioContext

#[repr(C)]
pub(crate) struct BaseAudioContext {
    eventtarget: EventTarget,
    audio_context_impl: Arc<Mutex<AudioContext>>,
    destination: MutNullableDom<AudioDestinationNode>,
    listener: MutNullableDom<AudioListener>,
    in_flight_resume_promises_queue: DomRefCell<VecDeque<(Box<[Rc<Promise>]>, ErrorResult)>>,
    pending_resume_promises: DomRefCell<Vec<Rc<Promise>>>,
    decode_resolvers: DomRefCell<HashMap<String, DecodeResolver>>,
    sample_rate: f32,
    state: Cell<AudioContextState>,
    channel_count: u32,
}

Fields
eventtarget: EventTargetaudio_context_impl: Arc<Mutex<AudioContext>>destination: MutNullableDom<AudioDestinationNode>

https://webaudio.github.io/web-audio-api/#dom-baseaudiocontext-destination

listener: MutNullableDom<AudioListener>in_flight_resume_promises_queue: DomRefCell<VecDeque<(Box<[Rc<Promise>]>, ErrorResult)>>

Resume promises which are soon to be fulfilled by a queued task.

pending_resume_promises: DomRefCell<Vec<Rc<Promise>>>

https://webaudio.github.io/web-audio-api/#pendingresumepromises

decode_resolvers: DomRefCell<HashMap<String, DecodeResolver>>sample_rate: f32

https://webaudio.github.io/web-audio-api/#dom-baseaudiocontext-samplerate

state: Cell<AudioContextState>

https://webaudio.github.io/web-audio-api/#dom-baseaudiocontext-state Although servo-media already keeps track of the control thread state, we keep a state flag here as well. This is so that we can synchronously throw when trying to do things on the context when the context has just been “closed()”.

channel_count: u32
Implementations

impl BaseAudioContext

pub(crate) fn new_inherited( options: BaseAudioContextOptions, pipeline_id: PipelineId, ) -> Fallible<BaseAudioContext>

pub(crate) fn is_offline(&self) -> bool

Tells whether this is an OfflineAudioContext or not.

pub(crate) fn audio_context_impl(&self) -> Arc<Mutex<AudioContext>>

pub(crate) fn destination_node(&self) -> NodeId

pub(crate) fn listener(&self) -> NodeId

pub(crate) fn is_allowed_to_start(&self) -> bool

fn is_detached_array_buffer(array_buffer: &ArrayBuffer) -> bool

fn push_pending_resume_promise(&self, promise: &Rc<Promise>)

fn take_pending_resume_promises(&self, result: ErrorResult)

Takes the pending resume promises.

The result with which these promises will be fulfilled is passed here and this method returns nothing because we actually just move the current list of pending resume promises to the in_flight_resume_promises_queue field.

Each call to this method must be followed by a call to fulfill_in_flight_resume_promises, to actually fulfill the promises which were taken and moved to the in-flight queue.

fn fulfill_in_flight_resume_promises<F>(&self, cx: &mut JSContext, f: F)
where F: FnOnce(),

Fulfills the next in-flight resume promises queue after running a closure.

See the comment on take_pending_resume_promises for why this method does not take a list of promises to fulfill. Callers cannot just pop the front list off of in_flight_resume_promises_queue and later fulfill the promises because that would mean putting #[cfg_attr(crown, expect(crown::unrooted_must_root))] on even more functions, potentially hiding actual safety bugs.

pub(crate) fn control_thread_state(&self) -> ProcessingState

Control thread processing state

pub(crate) fn set_state_attribute(&self, state: AudioContextState)

Set audio context state

pub(crate) fn resume(&self)

pub(crate) fn channel_count(&self) -> u32

impl BaseAudioContext

fn __assert_parent_type(&self)