Module script::dom::bindings::callback

Base classes to work with IDL callbacks.

Structs

CallSetup
A class that performs whatever setup we need to safely make a call while this class is on the stack. After new returns, the call is safe to make.

CallbackFunction
A common base class for representing IDL callback function types.

CallbackInterface
A common base class for representing IDL callback interface types.

Enums

ExceptionHandling
The exception handling used for a call.

Traits

CallbackContainer
A trait to be implemented by concrete IDL callback function and callback interface types.

Functions

wrap_call_this_object
Wraps the reflector for p into the compartment of cx.