Struct script::dom::bindings::callback::CallSetup

pub struct CallSetup {
    // some fields omitted
}

A class that performs whatever setup we need to safely make a call while this class is on the stack. After new returns, the call is safe to make.

Methods

impl CallSetup

fn new<T: CallbackContainer>(callback: T, handling: ExceptionHandling) -> CallSetup

Performs the setup needed to make a call.

fn get_context(&self) -> *mut JSContext

Returns the JSContext used for the call.