Struct script::dom::bindings::callback::CallbackInterface

pub struct CallbackInterface {
    object: CallbackObject,
}

A common base class for representing IDL callback interface types.

Fields
object: CallbackObject
Methods

impl CallbackInterface

fn callback(&self) -> *mut JSObject

Returns the underlying JSObject.

impl CallbackInterface

fn new() -> CallbackInterface

Create a new CallbackInterface object for the given JSObject.

fn init(&mut self, callback: *mut JSObject)

Initialize the callback function with a value. Should be called once this object is done moving.

fn get_callable_property(&self, cx: *mut JSContext, name: &str) -> Fallible<JSVal>

Returns the property with the given name, if it is a callable object, or an error otherwise.