Module script::dom::bindings::cell

A shareable mutable container for the DOM.

Structs

DomRefCell

A mutable field in the DOM.

Ref

Wraps a borrowed reference to a value in a RefCell box. A wrapper type for an immutably borrowed value from a RefCell<T>.

RefCell

A mutable memory location with dynamically checked borrow rules

RefMut

A wrapper type for a mutably borrowed value from a RefCell<T>.

Functions

ref_filter_map

Make a new Ref for a optional component of the borrowed data, e.g. an enum variant.