script::dom::bindings::cell

Struct RefMut
ยท

pub(crate) struct RefMut<'b, T>
where
    T: 'b + ?Sized,
{
    value: NonNull<T>,
    borrow: BorrowRefMut<'b>,
    marker: PhantomData<&'b mut T>,
}

A wrapper type for a mutably borrowed value from a RefCell<T>.

See the module-level documentation for more.

Fields
value: NonNull<T>borrow: BorrowRefMut<'b>marker: PhantomData<&'b mut T>
Implementations

impl<'b, T> RefMut<'b, T>
where T: ?Sized,

pub fn map<U, F>(orig: RefMut<'b, T>, f: F) -> RefMut<'b, U>
where F: FnOnce(&mut T) -> &mut U, U: ?Sized,

Makes a new RefMut for a component of the borrowed data, e.g., an enum variant.

The RefCell is already mutably borrowed, so this cannot fail.

This is an associated function that needs to be used as RefMut::map(...). A method would interfere with methods of the same name on the contents of a RefCell used through Deref.

Examples

use std::cell::{RefCell, RefMut};

let c = RefCell::new((5, 'b'));
{
    let b1: RefMut<'_, (u32, char)> = c.borrow_mut();
    let mut b2: RefMut<'_, u32> = RefMut::map(b1, |t| &mut t.0);
    assert_eq!(*b2, 5);
    *b2 = 42;
}
assert_eq!(*c.borrow(), (42, 'b'));

pub fn filter_map<U, F>( orig: RefMut<'b, T>, f: F, ) -> Result<RefMut<'b, U>, RefMut<'b, T>>
where F: FnOnce(&mut T) -> Option<&mut U>, U: ?Sized,

Makes a new RefMut for an optional component of the borrowed data. The original guard is returned as an Err(..) if the closure returns None.

The RefCell is already mutably borrowed, so this cannot fail.

This is an associated function that needs to be used as RefMut::filter_map(...). A method would interfere with methods of the same name on the contents of a RefCell used through Deref.

Examples

use std::cell::{RefCell, RefMut};

let c = RefCell::new(vec![1, 2, 3]);

{
    let b1: RefMut<'_, Vec<u32>> = c.borrow_mut();
    let mut b2: Result<RefMut<'_, u32>, _> = RefMut::filter_map(b1, |v| v.get_mut(1));

    if let Ok(mut b2) = b2 {
        *b2 += 2;
    }
}

assert_eq!(*c.borrow(), vec![1, 4, 3]);

pub fn try_map<U, E>( orig: RefMut<'b, T>, f: impl FnOnce(&mut T) -> Result<&mut U, E>, ) -> Result<RefMut<'b, U>, (RefMut<'b, T>, E)>
where U: ?Sized,

๐ฌThis is a nightly-only experimental API. (refcell_try_map)

Tries to makes a new RefMut for a component of the borrowed data. On failure, the original guard is returned alongside with the error returned by the closure.

The RefCell is already mutably borrowed, so this cannot fail.

This is an associated function that needs to be used as RefMut::try_map(...). A method would interfere with methods of the same name on the contents of a RefCell used through Deref.

Examples

#![feature(refcell_try_map)]
use std::cell::{RefCell, RefMut};
use std::str::{from_utf8_mut, Utf8Error};

let c = RefCell::new(vec![0x68, 0x65, 0x6C, 0x6C, 0x6F]);
{
    let b1: RefMut<'_, Vec<u8>> = c.borrow_mut();
    let b2: Result<RefMut<'_, str>, _> = RefMut::try_map(b1, |v| from_utf8_mut(v));
    let mut b2 = b2.unwrap();
    assert_eq!(&*b2, "hello");
    b2.make_ascii_uppercase();
}
assert_eq!(*c.borrow(), "HELLO".as_bytes());

let c = RefCell::new(vec![0xFF]);
let b1: RefMut<'_, Vec<u8>> = c.borrow_mut();
let b2: Result<_, (RefMut<'_, Vec<u8>>, Utf8Error)> = RefMut::try_map(b1, |v| from_utf8_mut(v));
let (b3, e) = b2.unwrap_err();
assert_eq!(*b3, vec![0xFF]);
assert_eq!(e.valid_up_to(), 0);

pub fn map_split<U, V, F>( orig: RefMut<'b, T>, f: F, ) -> (RefMut<'b, U>, RefMut<'b, V>)
where F: FnOnce(&mut T) -> (&mut U, &mut V), U: ?Sized, V: ?Sized,

Splits a RefMut into multiple RefMuts for different components of the borrowed data.

The underlying RefCell will remain mutably borrowed until both returned RefMuts go out of scope.

The RefCell is already mutably borrowed, so this cannot fail.

This is an associated function that needs to be used as RefMut::map_split(...). A method would interfere with methods of the same name on the contents of a RefCell used through Deref.

Examples

use std::cell::{RefCell, RefMut};

let cell = RefCell::new([1, 2, 3, 4]);
let borrow = cell.borrow_mut();
let (mut begin, mut end) = RefMut::map_split(borrow, |slice| slice.split_at_mut(2));
assert_eq!(*begin, [1, 2]);
assert_eq!(*end, [3, 4]);
begin.copy_from_slice(&[4, 3]);
end.copy_from_slice(&[2, 1]);

pub const fn leak(orig: RefMut<'b, T>) -> &'b mut T

๐ฌThis is a nightly-only experimental API. (cell_leak)

Converts into a mutable reference to the underlying data.

The underlying RefCell can not be borrowed from again and will always appear already mutably borrowed, making the returned reference the only to the interior.

This is an associated function that needs to be used as RefMut::leak(...). A method would interfere with methods of the same name on the contents of a RefCell used through Deref.

Examples

#![feature(cell_leak)]
use std::cell::{RefCell, RefMut};
let cell = RefCell::new(0);

let value = RefMut::leak(cell.borrow_mut());
assert_eq!(*value, 0);
*value = 1;

assert!(cell.try_borrow_mut().is_err());