script::dom::bindings::cell

Trait AtomicSafeBorrowMut

pub(crate) trait AtomicSafeBorrowMut {
    type Target;

    // Required method
    fn safe_borrow_mut<'a: 'b, 'b>(
        &'a self,
        no_gc: &'b NoGC,
    ) -> AtomicRefMutNoGC<'b, Self::Target>;
}

Required Associated Types

type Target

Required Methods

fn safe_borrow_mut<'a: 'b, 'b>( &'a self, no_gc: &'b NoGC, ) -> AtomicRefMutNoGC<'b, Self::Target>

A version of [DomRefCell::safe_borrow_mut] for AtomicRefCell. The resulting borrowed mutable reference statically guarantees that no garbage collection can occur while the borrow is live.