script::dom::bindings::constructor

Function get_constructor_object_from_local_name

fn get_constructor_object_from_local_name(
    name: LocalName,
    cx: JSContext,
    global: HandleObject<'_>,
    rval: MutableHandleObject<'_>,
) -> bool

Returns the constructor object for the element associated with the given local name. This list should only include elements marked with the HTMLConstructor extended attribute.