script::dom::bindings

Module conversions

Conversions of Rust values to and from JSVal.

IDL type	Argument type	Return type

any	JSVal

boolean	bool

byte	i8

octet	u8

short	i16

unsigned short	u16

long	i32

unsigned long	u32

long long	i64

unsigned long long	u64

unrestricted float	f32

float	Finite<f32>

unrestricted double	f64

double	Finite<f64>

DOMString	DOMString

USVString	USVString

ByteString	ByteString

object	*mut JSObject

interface types	&T	DomRoot<T>

dictionary types	&T	unsupported

enumeration types	T

callback function types	Rc<T>

nullable types	Option<T>

sequences	Vec<T>

union types	T

Enums

ConversionBehavior 🔒

Behavior for converting out-of-range integers.

ConversionResult 🔒

An enum to better support enums through FromJSValConvertible::from_jsval.

PrototypeCheck 🔒

StringificationBehavior 🔒

Behavior for stringification of JSVals.

Constants

DOM_OBJECT_SLOT 🔒

The index of the slot wherein a pointer to the reflected DOM object is stored for non-proxy bindings.

Traits

DerivedFrom 🔒

A trait to mark an IDL interface as deriving from another one.

FromJSValConvertible 🔒

A trait to convert JSVals to Rust types.

IDLInterface 🔒

A trait to check whether a given JSObject implements an IDL interface.

ToJSValConvertible 🔒

A trait to convert Rust types to JSVals.

Functions

get_dom_class 🔒 ⚠

Get the DOMClass from obj, or Err(()) if obj is not a DOM object.

get_property 🔒

Get a property from a JS object, and convert it to a Rust value.

get_property_jsval 🔒

Get a property from a JS object.

is_array_like 🔒

Returns whether value is an array-like object (Array, FileList, HTMLCollection, HTMLFormControlsCollection, HTMLOptionsCollection, NodeList, DOMTokenList).

is_dom_proxy 🔒 ⚠

Returns whether obj is a DOM object implemented as a proxy.

jsid_to_string 🔒

Convert id to a DOMString. Returns None if id is not a string or integer.

native_from_handlevalue 🔒

Get a *const T for a DOM object accessible from a HandleValue. Caller is responsible for throwing a JS exception if needed in case of error.

native_from_object 🔒 ⚠

Get a *const T for a DOM object accessible from a JSObject.

native_from_object_static 🔒 ⚠

Get a *const T for a DOM object accessible from a JSObject, where the DOM object is guaranteed not to be a wrapper.

private_from_object 🔒 ⚠

Get the private pointer of a DOM object from a given reflector.

private_from_proto_check 🔒 ⚠

Get a *const libc::c_void for the given DOM object, unwrapping any wrapper around it first, and checking if the object is of the correct type.

root_from_handleobject 🔒

Get a DomRoot<T> for a DOM object accessible from a HandleObject.

root_from_handlevalue 🔒

Get a DomRoot<T> for a DOM object accessible from a HandleValue. Caller is responsible for throwing a JS exception if needed in case of error.

root_from_object 🔒 ⚠

Get a DomRoot<T> for the given DOM object, unwrapping any wrapper around it first, and checking if the object is of the correct type.

root_from_object_static 🔒

Get a DomRoot<T> for the given DOM object, unwrapping any wrapper around it first, and checking if the object is of the correct type.