Function script::dom::bindings::conversions::native_from_reflector_jsmanaged

pub fn native_from_reflector_jsmanaged<T>(obj: *mut JSObject) -> Result<Unrooted<T>, ()> where T: Reflectable + IDLInterface

Get an Unrooted<T> for the given DOM object, unwrapping any wrapper around it first, and checking if the object is of the correct type.

Returns Err(()) if obj is an opaque security wrapper or if the object is not a reflector for a DOM object of the given type (as defined by the proto_id and proto_depth).