Module script::dom::bindings::error

Utilities to throw exceptions from Rust bindings.

Structs

ErrorInfo
A struct encapsulating information about a runtime script error.

Enums

Error
DOM exceptions that can be thrown by a native DOM method.

Functions

report_pending_exception⚠
Report a pending exception, thereby clearing it.

throw_dom_exception
Set a pending exception for the given result on cx.

throw_invalid_this⚠
Throw an exception to signal that a JSObject can not be converted to a given DOM type.

Type Definitions

ErrorResult
The return type for IDL operations that can throw DOM exceptions and return ().

Fallible
The return type for IDL operations that can throw DOM exceptions.