Module script::dom::bindings::global

Abstractions for global scopes.

This module contains smart pointers to global scopes, to simplify writing code that works in workers as well as window scopes.

Enums

GlobalField
A traced reference to a global object, for use in fields of traced Rust structures.

GlobalRef
A freely-copyable reference to a rooted global object.

GlobalRoot
A stack-based rooted reference to a global object.

GlobalUnrooted
An unrooted reference to a global object.

Functions

global_object_for_js_object
Returns the global object of the realm that the given JS object was created in.