Enum script::dom::bindings::global::GlobalRef

pub enum GlobalRef<'a> {
    Window(JSRef<'a, Window>),
    Worker(JSRef<'a, WorkerGlobalScope>),
}

A freely-copyable reference to a rooted global object.

Variants

Window

A reference to a Window object.

Worker

A reference to a WorkerGlobalScope object.

Methods

impl<'a> GlobalRef<'a>

fn get_cx(&self) -> *mut JSContext

Get the JSContext for the JSRuntime associated with the thread this global object is on.

fn as_window<'b>(&'b self) -> JSRef<'b, Window>

Extract a Window, causing task failure if the global object is not a Window.

fn resource_task(&self) -> ResourceTask

Get the ResourceTask for this global scope.

fn get_next_worker_id(&self) -> WorkerId

Get next worker id.

fn get_url(&self) -> Url

Get the URL for this global scope.

fn script_chan(&self) -> Box<ScriptChan + Send>

ScriptChan used to send messages to the event loop of this global's thread.