Module script::dom::bindings::import::base

Re-exports

pub use crate::dom::bindings::callback::wrap_call_this_object;

pub use crate::dom::bindings::callback::CallSetup;

pub use crate::dom::bindings::callback::CallbackContainer;

pub use crate::dom::bindings::callback::CallbackFunction;

pub use crate::dom::bindings::callback::CallbackInterface;

pub use crate::dom::bindings::callback::CallbackObject;

pub use crate::dom::bindings::callback::ExceptionHandling;

pub use crate::dom::bindings::codegen::Bindings::AudioNodeBinding::ChannelCountMode;

pub use crate::dom::bindings::codegen::Bindings::AudioNodeBinding::ChannelCountModeValues;

pub use crate::dom::bindings::codegen::Bindings::AudioNodeBinding::ChannelInterpretation;

pub use crate::dom::bindings::codegen::Bindings::AudioNodeBinding::ChannelInterpretationValues;

pub use crate::dom::bindings::codegen::UnionTypes;

pub use crate::dom::bindings::conversions::root_from_handlevalue;

pub use crate::dom::bindings::conversions::StringificationBehavior;

pub use crate::dom::bindings::error::Error::JSFailed;

pub use crate::dom::bindings::error::throw_dom_exception;

pub use crate::dom::bindings::error::Fallible;

pub use crate::dom::bindings::num::Finite;

pub use crate::dom::bindings::reflector::DomObject;

pub use crate::dom::bindings::root::DomRoot;

pub use crate::dom::bindings::str::ByteString;

pub use crate::dom::bindings::str::DOMString;

pub use crate::dom::bindings::str::USVString;

pub use crate::dom::bindings::trace::RootedTraceableBox;

pub use crate::dom::bindings::utils::get_dictionary_property;

pub use crate::dom::bindings::utils::set_dictionary_property;

pub use crate::dom::globalscope::GlobalScope;

pub use crate::script_runtime::JSContext as SafeJSContext;

Modules

ptr

Manually manage memory through raw pointers.

Structs

HandleValueArray

A generic handle to an array of rooted values.

JSContext

JSObject

Rc

A single-threaded reference-counting pointer. ‘Rc’ stands for ‘Reference Counted’.

Enums

ConversionBehavior

Behavior for converting out-of-range integers.

ConversionResult

An enum to better support enums through FromJSValConvertible::from_jsval.

Traits

FromJSValConvertible

A trait to convert JSVals to Rust types.

ToJSValConvertible

A trait to convert Rust types to JSVals.

Functions

CurrentGlobalOrNull⚠

Get the current realm’s global. Returns nullptr if no realm has been entered.

IsCallable⚠

Return true if the given object is callable. In ES6 terms, an object is callable if it has a [Call] internal method.

JS_CallFunctionValue⚠

JS_NewObject⚠

JS_WrapValue⚠

NullValue

ObjectOrNullValue

ObjectValue

UndefinedValue

maybe_resume_unwind

If there is a pending panic, resume unwinding.

throw_type_error⚠

Throw a TypeError with the given message.

Type Aliases

HandleObject

HandleValue

Heap

JSVal

MutableHandleObject

MutableHandleValue

RawHandleValue

Reference to a T that has been rooted elsewhere. This is most useful as a parameter type, which guarantees that the T lvalue is properly rooted. See “Move GC Stack Rooting” above.