Module script::dom::bindings::import::module

Re-exports

pub use crate::dom::bindings::codegen::Bindings::AnalyserNodeBinding::AnalyserOptions;

pub use crate::dom::bindings::codegen::Bindings::AudioNodeBinding::AudioNode_Binding;

pub use crate::dom::bindings::codegen::Bindings::AudioNodeBinding::ChannelCountMode;

pub use crate::dom::bindings::codegen::Bindings::AudioNodeBinding::ChannelCountModeValues;

pub use crate::dom::bindings::codegen::Bindings::AudioNodeBinding::ChannelInterpretation;

pub use crate::dom::bindings::codegen::Bindings::AudioNodeBinding::ChannelInterpretationValues;

pub use crate::dom::bindings::codegen::Bindings::EventTargetBinding::EventTarget_Binding;

pub use crate::dom::bindings::codegen::InterfaceObjectMap;

pub use crate::dom::bindings::codegen::PrototypeList;

pub use crate::dom::bindings::codegen::RegisterBindings;

pub use crate::dom::bindings::constant::ConstantSpec;

pub use crate::dom::bindings::constant::ConstantVal;

pub use crate::dom::bindings::constructor::call_default_constructor;

pub use crate::dom::bindings::constructor::call_html_constructor;

pub use crate::dom::bindings::constructor::pop_current_element_queue;

pub use crate::dom::bindings::constructor::push_new_element_queue;

pub use crate::dom::bindings::conversions::is_array_like;

pub use crate::dom::bindings::conversions::jsid_to_string;

pub use crate::dom::bindings::conversions::native_from_handlevalue;

pub use crate::dom::bindings::conversions::native_from_object_static;

pub use crate::dom::bindings::conversions::IDLInterface;

pub use crate::dom::bindings::conversions::StringificationBehavior;

pub use crate::dom::bindings::conversions::DOM_OBJECT_SLOT;

pub use crate::dom::bindings::error::throw_constructor_without_new;

pub use crate::dom::bindings::error::Error;

pub use crate::dom::bindings::error::ErrorResult;

pub use crate::dom::bindings::finalize::finalize_common;

pub use crate::dom::bindings::finalize::finalize_global;

pub use crate::dom::bindings::finalize::finalize_weak_referenceable;

pub use crate::dom::bindings::guard::Condition;

pub use crate::dom::bindings::guard::Guard;

pub use crate::dom::bindings::inheritance::Castable;

pub use crate::dom::bindings::interface::create_callback_interface_object;

pub use crate::dom::bindings::interface::create_global_object;

pub use crate::dom::bindings::interface::create_interface_prototype_object;

pub use crate::dom::bindings::interface::create_named_constructors;

pub use crate::dom::bindings::interface::create_noncallback_interface_object;

pub use crate::dom::bindings::interface::define_dom_interface;

pub use crate::dom::bindings::interface::define_guarded_methods;

pub use crate::dom::bindings::interface::define_guarded_properties;

pub use crate::dom::bindings::interface::get_desired_proto;

pub use crate::dom::bindings::interface::get_per_interface_object_handle;

pub use crate::dom::bindings::interface::is_exposed_in;

pub use crate::dom::bindings::interface::ConstructorClassHook;

pub use crate::dom::bindings::interface::InterfaceConstructorBehavior;

pub use crate::dom::bindings::interface::NonCallbackInterfaceObjectClass;

pub use crate::dom::bindings::interface::ProtoOrIfaceIndex;

pub use crate::dom::bindings::iterable::Iterable;

pub use crate::dom::bindings::iterable::IteratorType;

pub use crate::dom::bindings::like::Maplike;

pub use crate::dom::bindings::like::Setlike;

pub use crate::dom::bindings::namespace::create_namespace_object;

pub use crate::dom::bindings::namespace::NamespaceObjectClass;

pub use crate::dom::bindings::proxyhandler;

pub use crate::dom::bindings::proxyhandler::ensure_expando_object;

pub use crate::dom::bindings::proxyhandler::get_expando_object;

pub use crate::dom::bindings::proxyhandler::set_property_descriptor;

pub use crate::dom::bindings::record::Record;

pub use crate::dom::bindings::reflector::DomObjectIteratorWrap;

pub use crate::dom::bindings::reflector::DomObjectWrap;

pub use crate::dom::bindings::reflector::Reflector;

pub use crate::dom::bindings::root::Dom;

pub use crate::dom::bindings::root::DomSlice;

pub use crate::dom::bindings::root::MaybeUnreflectedDom;

pub use crate::dom::bindings::root::Root;

pub use crate::dom::bindings::utils::enumerate_global;

pub use crate::dom::bindings::utils::exception_to_promise;

pub use crate::dom::bindings::utils::generic_getter;

pub use crate::dom::bindings::utils::generic_lenient_getter;

pub use crate::dom::bindings::utils::generic_lenient_setter;

pub use crate::dom::bindings::utils::generic_method;

pub use crate::dom::bindings::utils::generic_setter;

pub use crate::dom::bindings::utils::generic_static_promise_method;

pub use crate::dom::bindings::utils::get_array_index_from_id;

pub use crate::dom::bindings::utils::get_property_on_prototype;

pub use crate::dom::bindings::utils::has_property_on_prototype;

pub use crate::dom::bindings::utils::resolve_global;

pub use crate::dom::bindings::utils::trace_global;

pub use crate::dom::bindings::utils::AsVoidPtr;

pub use crate::dom::bindings::utils::DOMClass;

pub use crate::dom::bindings::utils::DOMJSClass;

pub use crate::dom::bindings::utils::ProtoOrIfaceArray;

pub use crate::dom::bindings::utils::DOM_PROTO_UNFORGEABLE_HOLDER_SLOT;

pub use crate::dom::bindings::utils::JSCLASS_DOM_GLOBAL;

pub use crate::dom::bindings::weakref::WeakReferenceable;

pub use crate::dom::bindings::weakref::DOM_WEAK_SLOT;

pub use crate::dom::types::AnalyserNode;

pub use crate::dom::types::AudioNode;

pub use crate::dom::types::BaseAudioContext;

pub use crate::dom::types::EventTarget;

pub use crate::mem::malloc_size_of_including_raw_self;

pub use crate::realms::AlreadyInRealm;

pub use crate::realms::InRealm;

pub use crate::script_runtime::CanGc;

pub use super::base::*;

Modules

cmp

Utilities for comparing and ordering values.

jsapi

typedarray

High-level, safe bindings for JS typed array APIs. Allows creating new typed arrays or wrapping existing JS reflectors, and prevents reinterpreting existing buffers as different types except in well-defined cases.

Macros

pref

A convenience macro for accessing a preference value using its static path. Passing an invalid path is a compile-time error.

typedarray

Structs

CString

A type representing an owned, C-compatible, nul-terminated string with no nul bytes in the middle.

CallArgs

CustomAutoRooterGuard

An RAII guard used to root underlying data in CustomAutoRooter until the guard is dropped (falls out of scope). The underlying data can be accessed through this guard via its Deref and DerefMut implementations. This structure is created by root method on CustomAutoRooter or by the auto_root! macro.

GCContext

Handle

JSAutoRealm

JSClass

JSClassOps

JSFunctionSpec

JSJitGetterCallArgs

A class, expected to be passed by value, which represents the CallArgs for a JSJitGetterOp.

JSJitInfo

This struct contains metadata passed from the DOM to the JS Engine for JIT optimizations on DOM property accessors.

JSJitMethodCallArgs

JSJitSetterCallArgs

A class, expected to be passed by value, which represents the CallArgs for a JSJitSetterOp.

JSNativeWrapper

Wrapper to relace JSNative for JSPropertySpecs and JSFunctionSpecs. This will allow us to pass one JSJitInfo per function with the property/function spec, without additional field overhead.

JSPropertySpec

Description of a property. JS_DefineProperties and JS_InitClass take arrays of these and define many properties at once. JS_PSG, JS_PSGS and JS_PS_END are helper macros for defining such arrays.

JSPropertySpec_AccessorsOrValue_Accessors

JSPropertySpec_ValueWrapper

JSTracer

JSTypedMethodJitInfo

MutableHandle

NonNull

*mut T but non-zero and covariant.

ObjectOpResult

Per ES6, the [[DefineOwnProperty]] internal method has three different possible outcomes:

PropertyDescriptor

A structure that represents a property on an object, or the absence of a property. Use {,Mutable}Handle to interact with instances of this structure rather than interacting directly with member fields.

ProxyTraps

RawHandle

Reference to a T that has been rooted elsewhere. This is most useful as a parameter type, which guarantees that the T lvalue is properly rooted. See “Move GC Stack Rooting” above.

RawMutableHandle

Similar to a handle, but the underlying storage can be changed. This is useful for outparams.

RawMutableHandleIdVector

RootedGuard

Rust API for keeping a Rooted value in the context’s root stack. Example usage: rooted!(in(cx) let x = UndefinedValue());. RootedGuard::new also works, but the macro is preferred.

__BindgenBitfieldUnit

Enums

JSJitInfo_AliasSet

An enum that describes what this getter/setter/method aliases. This determines what things can be hoisted past this call, and if this call is movable what it can be hoisted past.

JSJitInfo_ArgType

JSJitInfo_OpType

JSPropertySpec_Kind

JSPropertySpec_ValueWrapper_Type

JSValueType

SymbolCode

Constants

JSCLASS_FOREGROUND_FINALIZE

JSCLASS_GLOBAL_SLOT_COUNT

JSCLASS_IS_DOMJSCLASS

JSCLASS_IS_GLOBAL

JSCLASS_RESERVED_SLOTS_MASK

JSCLASS_RESERVED_SLOTS_SHIFT

JSITER_HIDDEN

JSITER_OWNONLY

JSITER_SYMBOLS

JSPROP_ENUMERATE

The property is visible in for/in loops.

JSPROP_PERMANENT

The property is non-configurable: it can’t be deleted, and if it’s an accessor descriptor, its getter and setter can’t be changed.

JSPROP_READONLY

The property is non-writable. This flag is only valid for data properties.

Statics

UndefinedHandleValue

Traits

GCMethods

A trait for types which can place appropriate GC barriers.

JSTraceable

A trait to allow tracing (only) DOM objects. Types that can be traced.

ToJSValConvertible

A trait to convert Rust types to JSVals.

Functions

AppendToIdVector⚠

Call⚠

CreateProxyHandler⚠

GetPropertyKeys⚠

GetProxyReservedSlot⚠

GetRealmErrorPrototype⚠

GetRealmFunctionPrototype⚠

GetRealmIteratorPrototype⚠

GetRealmObjectPrototype⚠

GetWellKnownSymbol⚠

Get one of the well-known symbols defined by ES6. A single set of well-known symbols is shared by all compartments in a JSRuntime.

JS_AtomizeAndPinString⚠

JS_CALLEE⚠

JS_CopyOwnPropertiesAndPrivateFields⚠

JS_DefineProperty⚠

JS_DefinePropertyById2⚠

JS_ForwardGetPropertyTo⚠

Get the value of the property obj[id], or undefined if no such property exists. This is the C++ equivalent of vp = Reflect.get(obj, id, receiver).

JS_GetProperty⚠

JS_GetPropertyDescriptorById⚠

JS_GetReservedSlot⚠

JS_HasPropertyById⚠

Compute the expression id in obj.

JS_InitializePropertiesFromCompatibleNativeObject⚠

JS_NewObjectWithGivenProto⚠

JS_NewObjectWithoutMetadata⚠

JS_NewPlainObject⚠

Creates a new plain object, like new Object(), with Object.prototype as [[Prototype]].

JS_SetImmutablePrototype⚠

JS_SetProperty⚠

JS_SetPrototype⚠

JS_SetReservedSlot⚠

JS_WrapObject⚠

NewProxyObject⚠

PrivateValue

RUST_INTERNED_STRING_TO_JSID⚠

RUST_SYMBOL_TO_JSID⚠

SetProxyReservedSlot⚠

get_context_realm⚠

get_object_class⚠

get_object_realm⚠

int_to_jsid⚠

wrap_panic

Generic wrapper for JS engine callbacks panic-catching

Type Aliases

ArrayBuffer

ArrayBufferView

Float32Array

Float64Array

RawHandleId

Reference to a T that has been rooted elsewhere. This is most useful as a parameter type, which guarantees that the T lvalue is properly rooted. See “Move GC Stack Rooting” above.

RawHandleObject

Reference to a T that has been rooted elsewhere. This is most useful as a parameter type, which guarantees that the T lvalue is properly rooted. See “Move GC Stack Rooting” above.

RawMutableHandleObject

Similar to a handle, but the underlying storage can be changed. This is useful for outparams.

RawMutableHandleValue

Similar to a handle, but the underlying storage can be changed. This is useful for outparams.

Uint8Array

Uint8ClampedArray

jsid

Unions

JSJitInfo__bindgen_ty_1

JSJitInfo__bindgen_ty_2

JSJitInfo__bindgen_ty_3

JSPropertySpec_Accessor

JSPropertySpec_AccessorsOrValue

JSPropertySpec_Name

JSPropertySpec_ValueWrapper__bindgen_ty_1