Struct script::dom::bindings::import::module::jsapi::JS::TransitiveCompileOptions

#[repr(C, packed(1))]
pub struct TransitiveCompileOptions {

    pub filename_: *const i8,
    pub introducerFilename_: *const i8,
    pub sourceMapURL_: *const u16,
    pub mutedErrors_: bool,
    pub forceStrictMode_: bool,
    pub shouldResistFingerprinting_: bool,
    pub sourcePragmas_: bool,
    pub skipFilenameValidation_: bool,
    pub hideScriptFromDebugger_: bool,
    pub deferDebugMetadata_: bool,
    pub eagerDelazificationStrategy_: DelazificationOption,
    pub selfHostingMode: bool,
    pub asmJSOption: AsmJSOption,
    pub throwOnAsmJSValidationFailureOption: bool,
    pub forceAsync: bool,
    pub discardSource: bool,
    pub sourceIsLazy: bool,
    pub allowHTMLComments: bool,
    pub nonSyntacticScope: bool,
    pub topLevelAwait: bool,
    pub importAssertions: bool,
    pub borrowBuffer: bool,
    pub usePinnedBytecode: bool,
    pub allocateInstantiationStorage: bool,
    pub deoptimizeModuleGlobalVars: bool,
    pub introductionType: *const i8,
    pub introductionLineno: u32,
    pub introductionOffset: u32,
    pub hasIntroductionInfo: bool,

}

The common base class for the CompileOptions hierarchy.

Use this in code that needs to propagate compile options from one compilation unit to another.

Fields
filename_: *const i8introducerFilename_: *const i8sourceMapURL_: *const u16mutedErrors_: bool

The Web Platform allows scripts to be loaded from arbitrary cross-origin sources. This allows an attack by which a malicious website loads a sensitive file (say, a bank statement) cross-origin (using the user’s cookies), and sniffs the generated syntax errors (via a window.onerror handler) for juicy morsels of its contents.

To counter this attack, HTML5 specifies that script errors should be sanitized (“muted”) when the script is not same-origin with the global for which it is loaded. Callers should set this flag for cross-origin scripts, and it will be propagated appropriately to child scripts and passed back in JSErrorReports.

forceStrictMode_: boolshouldResistFingerprinting_: boolsourcePragmas_: boolskipFilenameValidation_: boolhideScriptFromDebugger_: booldeferDebugMetadata_: booleagerDelazificationStrategy_: DelazificationOptionselfHostingMode: boolasmJSOption: AsmJSOptionthrowOnAsmJSValidationFailureOption: boolforceAsync: booldiscardSource: boolsourceIsLazy: boolallowHTMLComments: boolnonSyntacticScope: booltopLevelAwait: boolimportAssertions: boolborrowBuffer: boolusePinnedBytecode: boolallocateInstantiationStorage: booldeoptimizeModuleGlobalVars: boolintroductionType: *const i8

|introductionType| is a statically allocated C string. See JSScript.h for more information.

introductionLineno: u32introductionOffset: u32hasIntroductionInfo: bool