Module script::dom::bindings::import::module::jsapi::mozilla::detail

Structs

AbsReturnType

AlignasHelper

AlignedChecker

AllowDeprecatedAbs

AllowDeprecatedAbsFixed

AsVariantTemporary

AsVariantTemporary stores a value of type T to allow construction of a Variant value via type inference. Because T is copied and there’s no guarantee that the copy can be elided, AsVariantTemporary is best used with primitive or very small types.

AssertionConditionType

EmptyWrapper

EntrySlot

FloatingPointTrait

FuzzyEqualsEpsilon

HasFreeLSB

HasPointerType

HasPointerTypeHelper

HashTable

HashTableEntry

HashTable_AddPtr

HashTable_DoubleHash

HashTable_Enum

HashTable_FakeSlot

HashTable_Iterator

HashTable_ModIterator

HashTable_Ptr

HashTable_Range

IsMaybeImpl

IsPackableVariant

IsPackableVariant_EVbool

IsPackableVariant_VEbool

IsResult

LazyAssertedCastT

LinkedListElementTraits

LinkedList supports refcounted elements using this adapter class. Clients using LinkedList<RefPtr> will get a data structure that holds a strong reference to T as long as T is in the list.

MakeUnsignedChar

MaybePoisoner

PointerType

ResultImplementationNullIsOkBase

SelectResultImpl

SelectVariantType

SelectVariantType takes a type T and a list of variant types Variants and yields a type Type, selected from Variants, that can store a value of type T or a reference to type T. If no such type was found, Type is not defined. SelectVariantType also has a count member that contains the total number of selectable types (which will be used to check that a requested type is not ambiguously present twice.)

SelectVariantTypeHelper

SelectVariantTypeHelper is used in the implementation of SelectVariantType.

TempPtrToSetterT

ToStorageTypeArgument

UniqueSelector

UnusedZero

UnusedZeroEnum

VariantTag

VectorTesting

WrapToSignedHelper

WrappingAddHelper

WrappingMultiplyHelper

WrappingSubtractHelper

Enums

HashTable_FailureBehavior

HashTable_LookupReason

HashTable_RebuildStatus

PackingStrategy

StorageType

Type Aliases

AlignedStorageOrEmpty

AssertionConditionType_ValueT

AtomicBaseIncDec_Base

AtomicBase_Intrinsics

AtomicBase_ValueType

AtomicIntrinsics_Base

EntrySlot_Entry

EntrySlot_NonConstT

HashTableEntry_NonConstT

HashTable_Entry

HashTable_Key

HashTable_Lookup

HashTable_NonConstT

HashTable_Slot

IntrinsicAddSub_Base

IntrinsicBase_OrderedOp

IntrinsicBase_ValueType

IntrinsicIncDec_Base

IntrinsicMemoryOps_Base

IsMaybe

IsPackableVariant_Impl

LinkedListElementTraits_ClientType

LinkedListElementTraits_ConstClientType

LinkedListElementTraits_ConstRawType

LinkedListElementTraits_RawType

MakeUnsignedChar_Type

PointerTypeImpl_Type

PointerType_Type

ResultImplementationNullIsOkBase_ErrorStorageType

SelectResultImpl_Type

UniqueSelector_SingleObject

UniquePtr is a smart pointer that wholly owns a resource. Ownership may be transferred out of a UniquePtr through explicit action, but otherwise the resource is destroyed when the UniquePtr is destroyed.

UnsignedIntType

UnusedZeroEnum_StorageType

VariantTag_Type

WrapToSignedHelper_SignedType

WrappingAddHelper_UnsignedT

WrappingMultiplyHelper_UnsignedT

WrappingSubtractHelper_UnsignedT

supports_os