Type Alias script::dom::bindings::import::module::jsapi::mozilla::HashNumber

pub type HashNumber = u32;

Implementations

impl u32

pub const MIN: u32 = 0u32

The smallest value that can be represented by this integer type.

Examples

Basic usage:

assert_eq!(u32::MIN, 0);

pub const MAX: u32 = 4_294_967_295u32

The largest value that can be represented by this integer type (232 − 1).

Examples

Basic usage:

assert_eq!(u32::MAX, 4294967295);

pub const BITS: u32 = 32u32

The size of this integer type in bits.

Examples

assert_eq!(u32::BITS, 32);

pub fn from_str_radix(src: &str, radix: u32) -> Result<u32, ParseIntError>

Converts a string slice in a given base to an integer.

The string is expected to be an optional + sign followed by digits. Leading and trailing whitespace represent an error. Digits are a subset of these characters, depending on radix:

0-9

a-z

A-Z

Panics

This function panics if radix is not in the range from 2 to 36.

Examples

Basic usage:

assert_eq!(u32::from_str_radix("A", 16), Ok(10));

pub const fn count_ones(self) -> u32

Returns the number of ones in the binary representation of self.

Examples

Basic usage:

let n = 0b01001100u32;

assert_eq!(n.count_ones(), 3);

pub const fn count_zeros(self) -> u32

Returns the number of zeros in the binary representation of self.

Examples

Basic usage:

assert_eq!(u32::MAX.count_zeros(), 0);

pub const fn leading_zeros(self) -> u32

Returns the number of leading zeros in the binary representation of self.

Depending on what you’re doing with the value, you might also be interested in the ilog2 function which returns a consistent number, even if the type widens.

Examples

Basic usage:

let n = u32::MAX >> 2;

assert_eq!(n.leading_zeros(), 2);

pub const fn trailing_zeros(self) -> u32

Returns the number of trailing zeros in the binary representation of self.

Examples

Basic usage:

let n = 0b0101000u32;

assert_eq!(n.trailing_zeros(), 3);

pub const fn leading_ones(self) -> u32

Returns the number of leading ones in the binary representation of self.

Examples

Basic usage:

let n = !(u32::MAX >> 2);

assert_eq!(n.leading_ones(), 2);

pub const fn trailing_ones(self) -> u32

Returns the number of trailing ones in the binary representation of self.

Examples

Basic usage:

let n = 0b1010111u32;

assert_eq!(n.trailing_ones(), 3);

pub const fn rotate_left(self, n: u32) -> u32

Shifts the bits to the left by a specified amount, n, wrapping the truncated bits to the end of the resulting integer.

Please note this isn’t the same operation as the << shifting operator!

Examples

Basic usage:

let n = 0x10000b3u32;
let m = 0xb301;

assert_eq!(n.rotate_left(8), m);

pub const fn rotate_right(self, n: u32) -> u32

Shifts the bits to the right by a specified amount, n, wrapping the truncated bits to the beginning of the resulting integer.

Please note this isn’t the same operation as the >> shifting operator!

Examples

Basic usage:

let n = 0xb301u32;
let m = 0x10000b3;

assert_eq!(n.rotate_right(8), m);

pub const fn swap_bytes(self) -> u32

Reverses the byte order of the integer.

Examples

Basic usage:

let n = 0x12345678u32;
let m = n.swap_bytes();

assert_eq!(m, 0x78563412);

pub const fn reverse_bits(self) -> u32

Reverses the order of bits in the integer. The least significant bit becomes the most significant bit, second least-significant bit becomes second most-significant bit, etc.

Examples

Basic usage:

let n = 0x12345678u32;
let m = n.reverse_bits();

assert_eq!(m, 0x1e6a2c48);
assert_eq!(0, 0u32.reverse_bits());

pub const fn from_be(x: u32) -> u32

Converts an integer from big endian to the target’s endianness.

On big endian this is a no-op. On little endian the bytes are swapped.

Examples

Basic usage:

let n = 0x1Au32;

if cfg!(target_endian = "big") {
    assert_eq!(u32::from_be(n), n)
} else {
    assert_eq!(u32::from_be(n), n.swap_bytes())
}

pub const fn from_le(x: u32) -> u32

Converts an integer from little endian to the target’s endianness.

On little endian this is a no-op. On big endian the bytes are swapped.

Examples

Basic usage:

let n = 0x1Au32;

if cfg!(target_endian = "little") {
    assert_eq!(u32::from_le(n), n)
} else {
    assert_eq!(u32::from_le(n), n.swap_bytes())
}

pub const fn to_be(self) -> u32

Converts self to big endian from the target’s endianness.

On big endian this is a no-op. On little endian the bytes are swapped.

Examples

Basic usage:

let n = 0x1Au32;

if cfg!(target_endian = "big") {
    assert_eq!(n.to_be(), n)
} else {
    assert_eq!(n.to_be(), n.swap_bytes())
}

pub const fn to_le(self) -> u32

Converts self to little endian from the target’s endianness.

On little endian this is a no-op. On big endian the bytes are swapped.

Examples

Basic usage:

let n = 0x1Au32;

if cfg!(target_endian = "little") {
    assert_eq!(n.to_le(), n)
} else {
    assert_eq!(n.to_le(), n.swap_bytes())
}

pub const fn checked_add(self, rhs: u32) -> Option<u32>

Checked integer addition. Computes self + rhs, returning None if overflow occurred.

Examples

Basic usage:

assert_eq!((u32::MAX - 2).checked_add(1), Some(u32::MAX - 1));
assert_eq!((u32::MAX - 2).checked_add(3), None);

pub unsafe fn unchecked_add(self, rhs: u32) -> u32

🔬This is a nightly-only experimental API. (unchecked_math)

Unchecked integer addition. Computes self + rhs, assuming overflow cannot occur.

Safety

This results in undefined behavior when self + rhs > u32::MAX or self + rhs < u32::MIN, i.e. when checked_add would return None.

pub const fn checked_add_signed(self, rhs: i32) -> Option<u32>

Checked addition with a signed integer. Computes self + rhs, returning None if overflow occurred.

Examples

Basic usage:

assert_eq!(1u32.checked_add_signed(2), Some(3));
assert_eq!(1u32.checked_add_signed(-2), None);
assert_eq!((u32::MAX - 2).checked_add_signed(3), None);

pub const fn checked_sub(self, rhs: u32) -> Option<u32>

Checked integer subtraction. Computes self - rhs, returning None if overflow occurred.

Examples

Basic usage:

assert_eq!(1u32.checked_sub(1), Some(0));
assert_eq!(0u32.checked_sub(1), None);

pub unsafe fn unchecked_sub(self, rhs: u32) -> u32

🔬This is a nightly-only experimental API. (unchecked_math)

Unchecked integer subtraction. Computes self - rhs, assuming overflow cannot occur.

Safety

This results in undefined behavior when self - rhs > u32::MAX or self - rhs < u32::MIN, i.e. when checked_sub would return None.

pub const fn checked_mul(self, rhs: u32) -> Option<u32>

Checked integer multiplication. Computes self * rhs, returning None if overflow occurred.

Examples

Basic usage:

assert_eq!(5u32.checked_mul(1), Some(5));
assert_eq!(u32::MAX.checked_mul(2), None);

pub unsafe fn unchecked_mul(self, rhs: u32) -> u32

🔬This is a nightly-only experimental API. (unchecked_math)

Unchecked integer multiplication. Computes self * rhs, assuming overflow cannot occur.

Safety

This results in undefined behavior when self * rhs > u32::MAX or self * rhs < u32::MIN, i.e. when checked_mul would return None.

pub const fn checked_div(self, rhs: u32) -> Option<u32>

Checked integer division. Computes self / rhs, returning None if rhs == 0.

Examples

Basic usage:

assert_eq!(128u32.checked_div(2), Some(64));
assert_eq!(1u32.checked_div(0), None);

pub const fn checked_div_euclid(self, rhs: u32) -> Option<u32>

Checked Euclidean division. Computes self.div_euclid(rhs), returning None if rhs == 0.

Examples

Basic usage:

assert_eq!(128u32.checked_div_euclid(2), Some(64));
assert_eq!(1u32.checked_div_euclid(0), None);

pub const fn checked_rem(self, rhs: u32) -> Option<u32>

Checked integer remainder. Computes self % rhs, returning None if rhs == 0.

Examples

Basic usage:

assert_eq!(5u32.checked_rem(2), Some(1));
assert_eq!(5u32.checked_rem(0), None);

pub const fn checked_rem_euclid(self, rhs: u32) -> Option<u32>

Checked Euclidean modulo. Computes self.rem_euclid(rhs), returning None if rhs == 0.

Examples

Basic usage:

assert_eq!(5u32.checked_rem_euclid(2), Some(1));
assert_eq!(5u32.checked_rem_euclid(0), None);

pub const fn ilog(self, base: u32) -> u32

Returns the logarithm of the number with respect to an arbitrary base, rounded down.

This method might not be optimized owing to implementation details; ilog2 can produce results more efficiently for base 2, and ilog10 can produce results more efficiently for base 10.

Panics

This function will panic if self is zero, or if base is less than 2.

Examples

assert_eq!(5u32.ilog(5), 1);

pub const fn ilog2(self) -> u32

Returns the base 2 logarithm of the number, rounded down.

Panics

This function will panic if self is zero.

Examples

assert_eq!(2u32.ilog2(), 1);

pub const fn ilog10(self) -> u32

Returns the base 10 logarithm of the number, rounded down.

Panics

This function will panic if self is zero.

Example

assert_eq!(10u32.ilog10(), 1);

pub const fn checked_ilog(self, base: u32) -> Option<u32>

Returns the logarithm of the number with respect to an arbitrary base, rounded down.

Returns None if the number is zero, or if the base is not at least 2.

This method might not be optimized owing to implementation details; checked_ilog2 can produce results more efficiently for base 2, and checked_ilog10 can produce results more efficiently for base 10.

Examples

assert_eq!(5u32.checked_ilog(5), Some(1));

pub const fn checked_ilog2(self) -> Option<u32>

Returns the base 2 logarithm of the number, rounded down.

Returns None if the number is zero.

Examples

assert_eq!(2u32.checked_ilog2(), Some(1));

pub const fn checked_ilog10(self) -> Option<u32>

Returns the base 10 logarithm of the number, rounded down.

Returns None if the number is zero.

Examples

assert_eq!(10u32.checked_ilog10(), Some(1));

pub const fn checked_neg(self) -> Option<u32>

Checked negation. Computes -self, returning None unless self == 0.

Note that negating any positive integer will overflow.

Examples

Basic usage:

assert_eq!(0u32.checked_neg(), Some(0));
assert_eq!(1u32.checked_neg(), None);

pub const fn checked_shl(self, rhs: u32) -> Option<u32>

Checked shift left. Computes self << rhs, returning None if rhs is larger than or equal to the number of bits in self.

Examples

Basic usage:

assert_eq!(0x1u32.checked_shl(4), Some(0x10));
assert_eq!(0x10u32.checked_shl(129), None);

pub unsafe fn unchecked_shl(self, rhs: u32) -> u32

🔬This is a nightly-only experimental API. (unchecked_math)

Unchecked shift left. Computes self << rhs, assuming that rhs is less than the number of bits in self.

Safety

This results in undefined behavior if rhs is larger than or equal to the number of bits in self, i.e. when checked_shl would return None.

pub const fn checked_shr(self, rhs: u32) -> Option<u32>

Checked shift right. Computes self >> rhs, returning None if rhs is larger than or equal to the number of bits in self.

Examples

Basic usage:

assert_eq!(0x10u32.checked_shr(4), Some(0x1));
assert_eq!(0x10u32.checked_shr(129), None);

pub unsafe fn unchecked_shr(self, rhs: u32) -> u32

🔬This is a nightly-only experimental API. (unchecked_math)

Unchecked shift right. Computes self >> rhs, assuming that rhs is less than the number of bits in self.

Safety

This results in undefined behavior if rhs is larger than or equal to the number of bits in self, i.e. when checked_shr would return None.

pub const fn checked_pow(self, exp: u32) -> Option<u32>

Checked exponentiation. Computes self.pow(exp), returning None if overflow occurred.

Examples

Basic usage:

assert_eq!(2u32.checked_pow(5), Some(32));
assert_eq!(u32::MAX.checked_pow(2), None);

pub const fn saturating_add(self, rhs: u32) -> u32

Saturating integer addition. Computes self + rhs, saturating at the numeric bounds instead of overflowing.

Examples

Basic usage:

assert_eq!(100u32.saturating_add(1), 101);
assert_eq!(u32::MAX.saturating_add(127), u32::MAX);

pub const fn saturating_add_signed(self, rhs: i32) -> u32

Saturating addition with a signed integer. Computes self + rhs, saturating at the numeric bounds instead of overflowing.

Examples

Basic usage:

assert_eq!(1u32.saturating_add_signed(2), 3);
assert_eq!(1u32.saturating_add_signed(-2), 0);
assert_eq!((u32::MAX - 2).saturating_add_signed(4), u32::MAX);

pub const fn saturating_sub(self, rhs: u32) -> u32

Saturating integer subtraction. Computes self - rhs, saturating at the numeric bounds instead of overflowing.

Examples

Basic usage:

assert_eq!(100u32.saturating_sub(27), 73);
assert_eq!(13u32.saturating_sub(127), 0);

pub const fn saturating_mul(self, rhs: u32) -> u32

Saturating integer multiplication. Computes self * rhs, saturating at the numeric bounds instead of overflowing.

Examples

Basic usage:

assert_eq!(2u32.saturating_mul(10), 20);
assert_eq!((u32::MAX).saturating_mul(10), u32::MAX);

pub const fn saturating_div(self, rhs: u32) -> u32

Saturating integer division. Computes self / rhs, saturating at the numeric bounds instead of overflowing.

Examples

Basic usage:

assert_eq!(5u32.saturating_div(2), 2);

let _ = 1u32.saturating_div(0);

pub const fn saturating_pow(self, exp: u32) -> u32

Saturating integer exponentiation. Computes self.pow(exp), saturating at the numeric bounds instead of overflowing.

Examples

Basic usage:

assert_eq!(4u32.saturating_pow(3), 64);
assert_eq!(u32::MAX.saturating_pow(2), u32::MAX);

pub const fn wrapping_add(self, rhs: u32) -> u32

Wrapping (modular) addition. Computes self + rhs, wrapping around at the boundary of the type.

Examples

Basic usage:

assert_eq!(200u32.wrapping_add(55), 255);
assert_eq!(200u32.wrapping_add(u32::MAX), 199);

pub const fn wrapping_add_signed(self, rhs: i32) -> u32

Wrapping (modular) addition with a signed integer. Computes self + rhs, wrapping around at the boundary of the type.

Examples

Basic usage:

assert_eq!(1u32.wrapping_add_signed(2), 3);
assert_eq!(1u32.wrapping_add_signed(-2), u32::MAX);
assert_eq!((u32::MAX - 2).wrapping_add_signed(4), 1);

pub const fn wrapping_sub(self, rhs: u32) -> u32

Wrapping (modular) subtraction. Computes self - rhs, wrapping around at the boundary of the type.

Examples

Basic usage:

assert_eq!(100u32.wrapping_sub(100), 0);
assert_eq!(100u32.wrapping_sub(u32::MAX), 101);

pub const fn wrapping_mul(self, rhs: u32) -> u32

Wrapping (modular) multiplication. Computes self * rhs, wrapping around at the boundary of the type.

Examples

Basic usage:

Please note that this example is shared between integer types. Which explains why u8 is used here.

assert_eq!(10u8.wrapping_mul(12), 120);
assert_eq!(25u8.wrapping_mul(12), 44);

pub const fn wrapping_div(self, rhs: u32) -> u32

Wrapping (modular) division. Computes self / rhs. Wrapped division on unsigned types is just normal division. There’s no way wrapping could ever happen. This function exists, so that all operations are accounted for in the wrapping operations.

Panics

This function will panic if rhs is 0.

Examples

Basic usage:

assert_eq!(100u32.wrapping_div(10), 10);

pub const fn wrapping_div_euclid(self, rhs: u32) -> u32

Wrapping Euclidean division. Computes self.div_euclid(rhs). Wrapped division on unsigned types is just normal division. There’s no way wrapping could ever happen. This function exists, so that all operations are accounted for in the wrapping operations. Since, for the positive integers, all common definitions of division are equal, this is exactly equal to self.wrapping_div(rhs).

Panics

This function will panic if rhs is 0.

Examples

Basic usage:

assert_eq!(100u32.wrapping_div_euclid(10), 10);

pub const fn wrapping_rem(self, rhs: u32) -> u32

Wrapping (modular) remainder. Computes self % rhs. Wrapped remainder calculation on unsigned types is just the regular remainder calculation. There’s no way wrapping could ever happen. This function exists, so that all operations are accounted for in the wrapping operations.

Panics

This function will panic if rhs is 0.

Examples

Basic usage:

assert_eq!(100u32.wrapping_rem(10), 0);

pub const fn wrapping_rem_euclid(self, rhs: u32) -> u32

Wrapping Euclidean modulo. Computes self.rem_euclid(rhs). Wrapped modulo calculation on unsigned types is just the regular remainder calculation. There’s no way wrapping could ever happen. This function exists, so that all operations are accounted for in the wrapping operations. Since, for the positive integers, all common definitions of division are equal, this is exactly equal to self.wrapping_rem(rhs).

Panics

This function will panic if rhs is 0.

Examples

Basic usage:

assert_eq!(100u32.wrapping_rem_euclid(10), 0);

pub const fn wrapping_neg(self) -> u32

Wrapping (modular) negation. Computes -self, wrapping around at the boundary of the type.

Since unsigned types do not have negative equivalents all applications of this function will wrap (except for -0). For values smaller than the corresponding signed type’s maximum the result is the same as casting the corresponding signed value. Any larger values are equivalent to MAX + 1 - (val - MAX - 1) where MAX is the corresponding signed type’s maximum.

Examples

Basic usage:

assert_eq!(0_u32.wrapping_neg(), 0);
assert_eq!(u32::MAX.wrapping_neg(), 1);
assert_eq!(13_u32.wrapping_neg(), (!13) + 1);
assert_eq!(42_u32.wrapping_neg(), !(42 - 1));

pub const fn wrapping_shl(self, rhs: u32) -> u32

Panic-free bitwise shift-left; yields self << mask(rhs), where mask removes any high-order bits of rhs that would cause the shift to exceed the bitwidth of the type.

Note that this is not the same as a rotate-left; the RHS of a wrapping shift-left is restricted to the range of the type, rather than the bits shifted out of the LHS being returned to the other end. The primitive integer types all implement a rotate_left function, which may be what you want instead.

Examples

Basic usage:

assert_eq!(1u32.wrapping_shl(7), 128);
assert_eq!(1u32.wrapping_shl(128), 1);

pub const fn wrapping_shr(self, rhs: u32) -> u32

Panic-free bitwise shift-right; yields self >> mask(rhs), where mask removes any high-order bits of rhs that would cause the shift to exceed the bitwidth of the type.

Note that this is not the same as a rotate-right; the RHS of a wrapping shift-right is restricted to the range of the type, rather than the bits shifted out of the LHS being returned to the other end. The primitive integer types all implement a rotate_right function, which may be what you want instead.

Examples

Basic usage:

assert_eq!(128u32.wrapping_shr(7), 1);
assert_eq!(128u32.wrapping_shr(128), 128);

pub const fn wrapping_pow(self, exp: u32) -> u32

Wrapping (modular) exponentiation. Computes self.pow(exp), wrapping around at the boundary of the type.

Examples

Basic usage:

assert_eq!(3u32.wrapping_pow(5), 243);
assert_eq!(3u8.wrapping_pow(6), 217);

pub const fn overflowing_add(self, rhs: u32) -> (u32, bool)

Calculates self + rhs

Returns a tuple of the addition along with a boolean indicating whether an arithmetic overflow would occur. If an overflow would have occurred then the wrapped value is returned.

Examples

Basic usage

assert_eq!(5u32.overflowing_add(2), (7, false));
assert_eq!(u32::MAX.overflowing_add(1), (0, true));

pub fn carrying_add(self, rhs: u32, carry: bool) -> (u32, bool)

🔬This is a nightly-only experimental API. (bigint_helper_methods)

Calculates self + rhs + carry and returns a tuple containing the sum and the output carry.

Performs “ternary addition” of two integer operands and a carry-in bit, and returns an output integer and a carry-out bit. This allows chaining together multiple additions to create a wider addition, and can be useful for bignum addition.

This can be thought of as a 32-bit “full adder”, in the electronics sense.

If the input carry is false, this method is equivalent to overflowing_add, and the output carry is equal to the overflow flag. Note that although carry and overflow flags are similar for unsigned integers, they are different for signed integers.

Examples

#![feature(bigint_helper_methods)]

//    3  MAX    (a = 3 × 2^32 + 2^32 - 1)
// +  5    7    (b = 5 × 2^32 + 7)
// ---------
//    9    6    (sum = 9 × 2^32 + 6)

let (a1, a0): (u32, u32) = (3, u32::MAX);
let (b1, b0): (u32, u32) = (5, 7);
let carry0 = false;

let (sum0, carry1) = a0.carrying_add(b0, carry0);
assert_eq!(carry1, true);
let (sum1, carry2) = a1.carrying_add(b1, carry1);
assert_eq!(carry2, false);

assert_eq!((sum1, sum0), (9, 6));

pub const fn overflowing_add_signed(self, rhs: i32) -> (u32, bool)

Calculates self + rhs with a signed rhs

Returns a tuple of the addition along with a boolean indicating whether an arithmetic overflow would occur. If an overflow would have occurred then the wrapped value is returned.

Examples

Basic usage:

assert_eq!(1u32.overflowing_add_signed(2), (3, false));
assert_eq!(1u32.overflowing_add_signed(-2), (u32::MAX, true));
assert_eq!((u32::MAX - 2).overflowing_add_signed(4), (1, true));

pub const fn overflowing_sub(self, rhs: u32) -> (u32, bool)

Calculates self - rhs

Returns a tuple of the subtraction along with a boolean indicating whether an arithmetic overflow would occur. If an overflow would have occurred then the wrapped value is returned.

Examples

Basic usage

assert_eq!(5u32.overflowing_sub(2), (3, false));
assert_eq!(0u32.overflowing_sub(1), (u32::MAX, true));

pub fn borrowing_sub(self, rhs: u32, borrow: bool) -> (u32, bool)

🔬This is a nightly-only experimental API. (bigint_helper_methods)

Calculates self − rhs − borrow and returns a tuple containing the difference and the output borrow.

Performs “ternary subtraction” by subtracting both an integer operand and a borrow-in bit from self, and returns an output integer and a borrow-out bit. This allows chaining together multiple subtractions to create a wider subtraction, and can be useful for bignum subtraction.

Examples

#![feature(bigint_helper_methods)]

//    9    6    (a = 9 × 2^32 + 6)
// -  5    7    (b = 5 × 2^32 + 7)
// ---------
//    3  MAX    (diff = 3 × 2^32 + 2^32 - 1)

let (a1, a0): (u32, u32) = (9, 6);
let (b1, b0): (u32, u32) = (5, 7);
let borrow0 = false;

let (diff0, borrow1) = a0.borrowing_sub(b0, borrow0);
assert_eq!(borrow1, true);
let (diff1, borrow2) = a1.borrowing_sub(b1, borrow1);
assert_eq!(borrow2, false);

assert_eq!((diff1, diff0), (3, u32::MAX));

pub const fn abs_diff(self, other: u32) -> u32

Computes the absolute difference between self and other.

Examples

Basic usage:

assert_eq!(100u32.abs_diff(80), 20u32);
assert_eq!(100u32.abs_diff(110), 10u32);

pub const fn overflowing_mul(self, rhs: u32) -> (u32, bool)

Calculates the multiplication of self and rhs.

Returns a tuple of the multiplication along with a boolean indicating whether an arithmetic overflow would occur. If an overflow would have occurred then the wrapped value is returned.

Examples

Basic usage:

Please note that this example is shared between integer types. Which explains why u32 is used here.

assert_eq!(5u32.overflowing_mul(2), (10, false));
assert_eq!(1_000_000_000u32.overflowing_mul(10), (1410065408, true));

pub const fn overflowing_div(self, rhs: u32) -> (u32, bool)

Calculates the divisor when self is divided by rhs.

Returns a tuple of the divisor along with a boolean indicating whether an arithmetic overflow would occur. Note that for unsigned integers overflow never occurs, so the second value is always false.

Panics

This function will panic if rhs is 0.

Examples

Basic usage

assert_eq!(5u32.overflowing_div(2), (2, false));

pub const fn overflowing_div_euclid(self, rhs: u32) -> (u32, bool)

Calculates the quotient of Euclidean division self.div_euclid(rhs).

Returns a tuple of the divisor along with a boolean indicating whether an arithmetic overflow would occur. Note that for unsigned integers overflow never occurs, so the second value is always false. Since, for the positive integers, all common definitions of division are equal, this is exactly equal to self.overflowing_div(rhs).

Panics

This function will panic if rhs is 0.

Examples

Basic usage

assert_eq!(5u32.overflowing_div_euclid(2), (2, false));

pub const fn overflowing_rem(self, rhs: u32) -> (u32, bool)

Calculates the remainder when self is divided by rhs.

Returns a tuple of the remainder after dividing along with a boolean indicating whether an arithmetic overflow would occur. Note that for unsigned integers overflow never occurs, so the second value is always false.

Panics

This function will panic if rhs is 0.

Examples

Basic usage

assert_eq!(5u32.overflowing_rem(2), (1, false));

pub const fn overflowing_rem_euclid(self, rhs: u32) -> (u32, bool)

Calculates the remainder self.rem_euclid(rhs) as if by Euclidean division.

Returns a tuple of the modulo after dividing along with a boolean indicating whether an arithmetic overflow would occur. Note that for unsigned integers overflow never occurs, so the second value is always false. Since, for the positive integers, all common definitions of division are equal, this operation is exactly equal to self.overflowing_rem(rhs).

Panics

This function will panic if rhs is 0.

Examples

Basic usage

assert_eq!(5u32.overflowing_rem_euclid(2), (1, false));

pub const fn overflowing_neg(self) -> (u32, bool)

Negates self in an overflowing fashion.

Returns !self + 1 using wrapping operations to return the value that represents the negation of this unsigned value. Note that for positive unsigned values overflow always occurs, but negating 0 does not overflow.

Examples

Basic usage

assert_eq!(0u32.overflowing_neg(), (0, false));
assert_eq!(2u32.overflowing_neg(), (-2i32 as u32, true));

pub const fn overflowing_shl(self, rhs: u32) -> (u32, bool)

Shifts self left by rhs bits.

Returns a tuple of the shifted version of self along with a boolean indicating whether the shift value was larger than or equal to the number of bits. If the shift value is too large, then value is masked (N-1) where N is the number of bits, and this value is then used to perform the shift.

Examples

Basic usage

assert_eq!(0x1u32.overflowing_shl(4), (0x10, false));
assert_eq!(0x1u32.overflowing_shl(132), (0x10, true));

pub const fn overflowing_shr(self, rhs: u32) -> (u32, bool)

Shifts self right by rhs bits.

Returns a tuple of the shifted version of self along with a boolean indicating whether the shift value was larger than or equal to the number of bits. If the shift value is too large, then value is masked (N-1) where N is the number of bits, and this value is then used to perform the shift.

Examples

Basic usage

assert_eq!(0x10u32.overflowing_shr(4), (0x1, false));
assert_eq!(0x10u32.overflowing_shr(132), (0x1, true));

pub const fn overflowing_pow(self, exp: u32) -> (u32, bool)

Raises self to the power of exp, using exponentiation by squaring.

Returns a tuple of the exponentiation along with a bool indicating whether an overflow happened.

Examples

Basic usage:

assert_eq!(3u32.overflowing_pow(5), (243, false));
assert_eq!(3u8.overflowing_pow(6), (217, true));

pub const fn pow(self, exp: u32) -> u32

Raises self to the power of exp, using exponentiation by squaring.

Examples

Basic usage:

assert_eq!(2u32.pow(5), 32);

pub fn isqrt(self) -> u32

🔬This is a nightly-only experimental API. (isqrt)

Returns the square root of the number, rounded down.

Examples

Basic usage:

#![feature(isqrt)]
assert_eq!(10u32.isqrt(), 3);

pub const fn div_euclid(self, rhs: u32) -> u32

Performs Euclidean division.

Since, for the positive integers, all common definitions of division are equal, this is exactly equal to self / rhs.

Panics

This function will panic if rhs is 0.

Examples

Basic usage:

assert_eq!(7u32.div_euclid(4), 1); // or any other integer type

pub const fn rem_euclid(self, rhs: u32) -> u32

Calculates the least remainder of self (mod rhs).

Since, for the positive integers, all common definitions of division are equal, this is exactly equal to self % rhs.

Panics

This function will panic if rhs is 0.

Examples

Basic usage:

assert_eq!(7u32.rem_euclid(4), 3); // or any other integer type

pub const fn div_floor(self, rhs: u32) -> u32

🔬This is a nightly-only experimental API. (int_roundings)

Calculates the quotient of self and rhs, rounding the result towards negative infinity.

This is the same as performing self / rhs for all unsigned integers.

Panics

This function will panic if rhs is zero.

Examples

Basic usage:

#![feature(int_roundings)]
assert_eq!(7_u32.div_floor(4), 1);

pub const fn div_ceil(self, rhs: u32) -> u32

Calculates the quotient of self and rhs, rounding the result towards positive infinity.

Panics

This function will panic if rhs is zero.

Overflow behavior

On overflow, this function will panic if overflow checks are enabled (default in debug mode) and wrap if overflow checks are disabled (default in release mode).

Examples

Basic usage:

assert_eq!(7_u32.div_ceil(4), 2);

pub const fn next_multiple_of(self, rhs: u32) -> u32

Calculates the smallest value greater than or equal to self that is a multiple of rhs.

Panics

This function will panic if rhs is zero.

Overflow behavior

On overflow, this function will panic if overflow checks are enabled (default in debug mode) and wrap if overflow checks are disabled (default in release mode).

Examples

Basic usage:

assert_eq!(16_u32.next_multiple_of(8), 16);
assert_eq!(23_u32.next_multiple_of(8), 24);

pub const fn checked_next_multiple_of(self, rhs: u32) -> Option<u32>

Calculates the smallest value greater than or equal to self that is a multiple of rhs. Returns None if rhs is zero or the operation would result in overflow.

Examples

Basic usage:

assert_eq!(16_u32.checked_next_multiple_of(8), Some(16));
assert_eq!(23_u32.checked_next_multiple_of(8), Some(24));
assert_eq!(1_u32.checked_next_multiple_of(0), None);
assert_eq!(u32::MAX.checked_next_multiple_of(2), None);

pub const fn is_power_of_two(self) -> bool

Returns true if and only if self == 2^k for some k.

Examples

Basic usage:

assert!(16u32.is_power_of_two());
assert!(!10u32.is_power_of_two());

pub const fn next_power_of_two(self) -> u32

Returns the smallest power of two greater than or equal to self.

When return value overflows (i.e., self > (1 << (N-1)) for type uN), it panics in debug mode and the return value is wrapped to 0 in release mode (the only situation in which method can return 0).

Examples

Basic usage:

assert_eq!(2u32.next_power_of_two(), 2);
assert_eq!(3u32.next_power_of_two(), 4);

pub const fn checked_next_power_of_two(self) -> Option<u32>

Returns the smallest power of two greater than or equal to n. If the next power of two is greater than the type’s maximum value, None is returned, otherwise the power of two is wrapped in Some.

Examples

Basic usage:

assert_eq!(2u32.checked_next_power_of_two(), Some(2));
assert_eq!(3u32.checked_next_power_of_two(), Some(4));
assert_eq!(u32::MAX.checked_next_power_of_two(), None);

pub fn wrapping_next_power_of_two(self) -> u32

🔬This is a nightly-only experimental API. (wrapping_next_power_of_two)

Returns the smallest power of two greater than or equal to n. If the next power of two is greater than the type’s maximum value, the return value is wrapped to 0.

Examples

Basic usage:

#![feature(wrapping_next_power_of_two)]

assert_eq!(2u32.wrapping_next_power_of_two(), 2);
assert_eq!(3u32.wrapping_next_power_of_two(), 4);
assert_eq!(u32::MAX.wrapping_next_power_of_two(), 0);

pub const fn to_be_bytes(self) -> [u8; 4]

Return the memory representation of this integer as a byte array in big-endian (network) byte order.

Examples

let bytes = 0x12345678u32.to_be_bytes();
assert_eq!(bytes, [0x12, 0x34, 0x56, 0x78]);

pub const fn to_le_bytes(self) -> [u8; 4]

Return the memory representation of this integer as a byte array in little-endian byte order.

Examples

let bytes = 0x12345678u32.to_le_bytes();
assert_eq!(bytes, [0x78, 0x56, 0x34, 0x12]);

pub const fn to_ne_bytes(self) -> [u8; 4]

Return the memory representation of this integer as a byte array in native byte order.

As the target platform’s native endianness is used, portable code should use to_be_bytes or to_le_bytes, as appropriate, instead.

Examples

let bytes = 0x12345678u32.to_ne_bytes();
assert_eq!(
    bytes,
    if cfg!(target_endian = "big") {
        [0x12, 0x34, 0x56, 0x78]
    } else {
        [0x78, 0x56, 0x34, 0x12]
    }
);

pub const fn from_be_bytes(bytes: [u8; 4]) -> u32

Create a native endian integer value from its representation as a byte array in big endian.

Examples

let value = u32::from_be_bytes([0x12, 0x34, 0x56, 0x78]);
assert_eq!(value, 0x12345678);

When starting from a slice rather than an array, fallible conversion APIs can be used:

fn read_be_u32(input: &mut &[u8]) -> u32 {
    let (int_bytes, rest) = input.split_at(std::mem::size_of::<u32>());
    *input = rest;
    u32::from_be_bytes(int_bytes.try_into().unwrap())
}

pub const fn from_le_bytes(bytes: [u8; 4]) -> u32

Create a native endian integer value from its representation as a byte array in little endian.

Examples

let value = u32::from_le_bytes([0x78, 0x56, 0x34, 0x12]);
assert_eq!(value, 0x12345678);

When starting from a slice rather than an array, fallible conversion APIs can be used:

fn read_le_u32(input: &mut &[u8]) -> u32 {
    let (int_bytes, rest) = input.split_at(std::mem::size_of::<u32>());
    *input = rest;
    u32::from_le_bytes(int_bytes.try_into().unwrap())
}

pub const fn from_ne_bytes(bytes: [u8; 4]) -> u32

Create a native endian integer value from its memory representation as a byte array in native endianness.

As the target platform’s native endianness is used, portable code likely wants to use from_be_bytes or from_le_bytes, as appropriate instead.

Examples

let value = u32::from_ne_bytes(if cfg!(target_endian = "big") {
    [0x12, 0x34, 0x56, 0x78]
} else {
    [0x78, 0x56, 0x34, 0x12]
});
assert_eq!(value, 0x12345678);

When starting from a slice rather than an array, fallible conversion APIs can be used:

fn read_ne_u32(input: &mut &[u8]) -> u32 {
    let (int_bytes, rest) = input.split_at(std::mem::size_of::<u32>());
    *input = rest;
    u32::from_ne_bytes(int_bytes.try_into().unwrap())
}

pub const fn min_value() -> u32

👎Deprecating in a future Rust version: replaced by the MIN associated constant on this type

New code should prefer to use u32::MIN instead.

Returns the smallest value that can be represented by this integer type.

pub const fn max_value() -> u32

👎Deprecating in a future Rust version: replaced by the MAX associated constant on this type

New code should prefer to use u32::MAX instead.

Returns the largest value that can be represented by this integer type.

pub fn widening_mul(self, rhs: u32) -> (u32, u32)

🔬This is a nightly-only experimental API. (bigint_helper_methods)

Calculates the complete product self * rhs without the possibility to overflow.

This returns the low-order (wrapping) bits and the high-order (overflow) bits of the result as two separate values, in that order.

If you also need to add a carry to the wide result, then you want Self::carrying_mul instead.

Examples

Basic usage:

Please note that this example is shared between integer types. Which explains why u32 is used here.

#![feature(bigint_helper_methods)]
assert_eq!(5u32.widening_mul(2), (10, 0));
assert_eq!(1_000_000_000u32.widening_mul(10), (1410065408, 2));

pub fn carrying_mul(self, rhs: u32, carry: u32) -> (u32, u32)

🔬This is a nightly-only experimental API. (bigint_helper_methods)

Calculates the “full multiplication” self * rhs + carry without the possibility to overflow.

This returns the low-order (wrapping) bits and the high-order (overflow) bits of the result as two separate values, in that order.

Performs “long multiplication” which takes in an extra amount to add, and may return an additional amount of overflow. This allows for chaining together multiple multiplications to create “big integers” which represent larger values.

If you don’t need the carry, then you can use Self::widening_mul instead.

Examples

Basic usage:

Please note that this example is shared between integer types. Which explains why u32 is used here.

#![feature(bigint_helper_methods)]
assert_eq!(5u32.carrying_mul(2, 0), (10, 0));
assert_eq!(5u32.carrying_mul(2, 10), (20, 0));
assert_eq!(1_000_000_000u32.carrying_mul(10, 0), (1410065408, 2));
assert_eq!(1_000_000_000u32.carrying_mul(10, 10), (1410065418, 2));
assert_eq!(u32::MAX.carrying_mul(u32::MAX, u32::MAX), (0, u32::MAX));

This is the core operation needed for scalar multiplication when implementing it for wider-than-native types.

#![feature(bigint_helper_methods)]
fn scalar_mul_eq(little_endian_digits: &mut Vec<u16>, multiplicand: u16) {
    let mut carry = 0;
    for d in little_endian_digits.iter_mut() {
        (*d, carry) = d.carrying_mul(multiplicand, carry);
    }
    if carry != 0 {
        little_endian_digits.push(carry);
    }
}

let mut v = vec![10, 20];
scalar_mul_eq(&mut v, 3);
assert_eq!(v, [30, 60]);

assert_eq!(0x87654321_u64 * 0xFEED, 0x86D3D159E38D);
let mut v = vec![0x4321, 0x8765];
scalar_mul_eq(&mut v, 0xFEED);
assert_eq!(v, [0xE38D, 0xD159, 0x86D3]);

If carry is zero, this is similar to overflowing_mul, except that it gives the value of the overflow instead of just whether one happened:

#![feature(bigint_helper_methods)]
let r = u8::carrying_mul(7, 13, 0);
assert_eq!((r.0, r.1 != 0), u8::overflowing_mul(7, 13));
let r = u8::carrying_mul(13, 42, 0);
assert_eq!((r.0, r.1 != 0), u8::overflowing_mul(13, 42));

The value of the first field in the returned tuple matches what you’d get by combining the wrapping_mul and wrapping_add methods:

#![feature(bigint_helper_methods)]
assert_eq!(
    789_u16.carrying_mul(456, 123).0,
    789_u16.wrapping_mul(456).wrapping_add(123),
);

pub fn midpoint(self, rhs: u32) -> u32

🔬This is a nightly-only experimental API. (num_midpoint)

Calculates the middle point of self and rhs.

midpoint(a, b) is (a + b) >> 1 as if it were performed in a sufficiently-large signed integral type. This implies that the result is always rounded towards negative infinity and that no overflow will ever occur.

Examples

#![feature(num_midpoint)]
assert_eq!(0u32.midpoint(4), 2);
assert_eq!(1u32.midpoint(4), 2);