Type Alias script::dom::bindings::import::module::RawHandleObject

pub type RawHandleObject = Handle<*mut JSObject>;

Reference to a T that has been rooted elsewhere. This is most useful as a parameter type, which guarantees that the T lvalue is properly rooted. See “Move GC Stack Rooting” above.

If you want to add additional methods to Handle for a specific specialization, define a HandleOperations specialization containing them.

Aliased Type

struct RawHandleObject {
    pub ptr: *const *mut JSObject,
    pub _phantom_0: PhantomData<UnsafeCell<*mut JSObject>>,
}

Fields
ptr: *const *mut JSObject_phantom_0: PhantomData<UnsafeCell<*mut JSObject>>
Implementations

impl<T> Handle<T>

pub fn get(&self) -> Twhere T: Copy,

pub unsafe fn from_marked_location(ptr: *const T) -> Handle<T>

impl Handle<*mut JSObject>

pub fn null() -> Handle<*mut JSObject>