Module script::dom::bindings::import::module::typedarray

High-level, safe bindings for JS typed array APIs. Allows creating new typed arrays or wrapping existing JS reflectors, and prevents reinterpreting existing buffers as different types except in well-defined cases.

Structs

ArrayBufferU8

A kind of typed array.

ArrayBufferViewU8

A kind of typed array.

ClampedU8

A kind of typed array.

Float32

A kind of typed array.

Float64

A kind of typed array.

Int8

A kind of typed array.

Int16

A kind of typed array.

Int32

A kind of typed array.

TypedArray

A typed array wrapper.

Uint8

A kind of typed array.

Uint16

A kind of typed array.

Uint32

A kind of typed array.

Enums

CreateWith

Traits

JSObjectStorage

Trait that specifies how pointers to wrapped objects are stored. It supports two variants, one with bare pointer (to be rooted on stack using CustomAutoRooter) and wrapped in a Box<Heap>, which can be stored in a heap-allocated structure, to be rooted with JSTraceable-implementing tracers (currently implemented in Servo).

TypedArrayElement

Internal trait used to associate an element type with an underlying representation and various functions required to manipulate typed arrays of that element type.

TypedArrayElementCreator

Internal trait for creating new typed arrays.

Type Aliases

ArrayBuffer

ArrayBufferView

Float32Array

Float64Array

HeapArrayBuffer

HeapArrayBufferView

HeapFloat32Array

HeapFloat64Array

HeapInt8Array

HeapInt16Array

HeapInt32Array

HeapUint8Array

HeapUint8ClampedArray

HeapUint16Array

HeapUint32Array

Int8Array

Int16Array

Int32Array

Uint8Array

Uint8ClampedArray

Uint16Array

Uint32Array