Type Alias script::dom::bindings::import::module::typedarray::HeapArrayBufferView

pub type HeapArrayBufferView = TypedArray<ArrayBufferViewU8, Box<Heap<*mut JSObject>, Global>>;

Aliased Type

struct HeapArrayBufferView {
    object: Box<Heap<*mut JSObject>, Global>,
    computed: Cell<Option<(*mut u8, usize)>>,
}

Fields
object: Box<Heap<*mut JSObject>, Global>computed: Cell<Option<(*mut u8, usize)>>
Implementations

impl<T, S> TypedArray<T, S>where T: TypedArrayElement, S: JSObjectStorage,

pub fn from(object: *mut JSObject) -> Result<TypedArray<T, S>, ()>

Create a typed array representation that wraps an existing JS reflector. This operation will fail if attempted on a JS object that does not match the expected typed array details.

pub fn len(&self) -> usize

Returns the number of elements in the underlying typed array.

pub unsafe fn underlying_object(&self) -> &S

Unsafety

Returned wrapped pointer to the underlying JSObject is meant to be read-only, modifying it can lead to Undefined Behaviour and violation of TypedArray API guarantees.

Practically, this exists only to implement JSTraceable trait in Servo for Box<Heap<*mut JSObject>> variant.

pub fn to_vec(&self) -> Vec<<T as TypedArrayElement>::Element, Global>where <T as TypedArrayElement>::Element: Clone,

Retrieves an owned data that’s represented by the typed array.

pub unsafe fn as_slice(&self) -> &[<T as TypedArrayElement>::Element]

Unsafety

The returned slice can be invalidated if the underlying typed array is neutered.

pub unsafe fn as_mut_slice( &mut self ) -> &mut [<T as TypedArrayElement>::Element]

Unsafety

The returned slice can be invalidated if the underlying typed array is neutered.

The underlying JSObject can be aliased, which can lead to Undefined Behavior due to mutable aliasing.

pub fn is_shared(&self) -> bool

Return a boolean flag which denotes whether the underlying buffer is a SharedArrayBuffer.

impl<T, S> TypedArray<T, S>where T: TypedArrayElementCreator + TypedArrayElement, S: JSObjectStorage,

pub unsafe fn create( cx: *mut JSContext, with: CreateWith<'_, <T as TypedArrayElement>::Element>, result: MutableHandle<'_, *mut JSObject> ) -> Result<(), ()>

Create a new JS typed array, optionally providing initial data that will be copied into the newly-allocated buffer. Returns the new JS reflector.

pub unsafe fn update(&mut self, data: &[<T as TypedArrayElement>::Element])

Update an existed JS typed array

impl<S> TypedArray<ArrayBufferViewU8, S>where S: JSObjectStorage,

pub fn get_array_type(&self) -> Type