Struct script::dom::bindings::interface::InterfaceConstructorBehavior

pub struct InterfaceConstructorBehavior(ClassOps);

The constructor behavior of a non-callback interface object.

Methods

impl InterfaceConstructorBehavior

const fn throw() -> Self

An interface constructor that unconditionally throws a type error.

const fn call(hook: unsafe extern fn(_: *mut JSContext, _: u32, _: *mut Value) -> bool)
-> Self

An interface constructor that calls a native Rust function.