Struct script::dom::bindings::interface::NonCallbackInterfaceObjectClass

pub struct NonCallbackInterfaceObjectClass {
    pub class: Class,
    pub proto_id: ID,
    pub proto_depth: u16,
    pub representation: &'static [u8],
}

The class of a non-callback interface object.

Fields
class: Class

The SpiderMonkey Class structure.

proto_id: ID

The prototype id of that interface, used in the hasInstance hook.

proto_depth: u16

The prototype depth of that interface, used in the hasInstance hook.

representation: &'static [u8]

The string representation of the object.

Methods

impl NonCallbackInterfaceObjectClass

const fn new(constructor_behavior: &'static InterfaceConstructorBehavior,
string_rep: &'static [u8],
proto_id: ID,
proto_depth: u16)
-> NonCallbackInterfaceObjectClass

Create a new NonCallbackInterfaceObjectClass structure.

fn as_jsclass(&self) -> &JSClass

cast own reference to JSClass reference