Module script::dom::bindings::iterable

Implementation of iterable<...> and iterable<..., ...> WebIDL declarations.

Structs

IterableIterator
An iterator over the iterable entries of a given DOM interface.

Enums

IteratorType
The values that an iterator will iterate over.

Traits

Iterable
A DOM object that can be iterated over using a pair value iterator.

Functions

dict_return

key_and_value_return