Struct script::dom::bindings::iterable::IterableIterator

pub struct IterableIterator<T: DomObject + JSTraceable + Iterable> {
    reflector: Reflector,
    iterable: JS<T>,
    type_: IteratorType,
    index: Cell<u32>,
}

An iterator over the iterable entries of a given DOM interface.

Fields
reflector: Reflector iterable: JS<T> type_: IteratorType index: Cell<u32>
Methods

impl<T: DomObject + JSTraceable + Iterable> IterableIterator<T>

fn new(iterable: &T,
type_: IteratorType,
wrap: unsafe fn(*mut JSContext, &GlobalScope, Box<IterableIterator<T>>) -> Root<Self>)
-> Root<Self>

Create a new iterator instance for the provided iterable DOM interface.

fn Next(&self, cx: *mut JSContext) -> Fallible<NonZero<*mut JSObject>>

Return the next value from the iterable object.