script::dom::bindings::iterable

Trait Iterable

pub trait Iterable {
    type Key: ToJSValConvertible;
    type Value: ToJSValConvertible;

    // Required methods
    fn get_iterable_length(&self, cx: &mut JSContext) -> u32;
    fn get_value_at_index(&self, cx: &mut JSContext, index: u32) -> Self::Value;
    fn get_key_at_index(&self, cx: &mut JSContext, index: u32) -> Self::Key;
}

A DOM object that can be iterated over using a pair value iterator.

Required Associated Types

type Key: ToJSValConvertible

The type of the key of the iterator pair.

type Value: ToJSValConvertible

The type of the value of the iterator pair.

Required Methods

fn get_iterable_length(&self, cx: &mut JSContext) -> u32

Return the number of entries that can be iterated over.

fn get_value_at_index(&self, cx: &mut JSContext, index: u32) -> Self::Value

Return the value at the provided index.

fn get_key_at_index(&self, cx: &mut JSContext, index: u32) -> Self::Key

Return the key at the provided index.