Module script::dom::bindings::js

Smart pointers for the JS-managed DOM objects.

The DOM is made up of DOM objects whose lifetime is entirely controlled by the whims of the SpiderMonkey garbage collector. The types in this module are designed to ensure that any interactions with said Rust types only occur on values that will remain alive the entire time.

Here is a brief overview of the important types:

JSRef<T>: a freely-copyable reference to a rooted DOM object.

Root<T>: a stack-based reference to a rooted DOM object.

JS<T>: a reference to a DOM object that can automatically be traced by the GC when encountered as a field of a Rust structure.

Temporary<T>: a reference to a DOM object that will remain rooted for the duration of its lifetime.

The rule of thumb is as follows:

All methods return Temporary<T>, to ensure the value remains alive until it is stored somewhere that is reachable by the GC.

All functions take JSRef<T> arguments, to ensure that they will remain uncollected for the duration of their usage.

All DOM structs contain JS<T> fields and derive the JSTraceable trait, to ensure that they are transitively marked as reachable by the GC if the enclosing value is reachable.

All methods for type T are implemented for JSRef<T>, to ensure that the self value will not be collected for the duration of the method call.

Both Temporary<T> and JS<T> do not allow access to their inner value without explicitly creating a stack-based root via the root method. This returns a Root<T>, which causes the JS-owned value to be uncollectable for the duration of the Root object's lifetime. A JSRef<T> can be obtained from a Root<T> by calling the r method. These JSRef<T> values are not allowed to outlive their originating Root<T>, to ensure that all interactions with the enclosed value only occur when said value is uncollectable, and will cause static lifetime errors if misused.

Other miscellaneous helper traits:

OptionalRootable and OptionalRootedRootable: make rooting Option values easy via a root method

ResultRootable: make rooting successful Result values easy

TemporaryPushable: allows mutating vectors of JS<T> with new elements of JSRef/Temporary

RootedReference: makes obtaining an Option<JSRef<T>> from an Option<Root<T>> easy

Structs

JS
A traced reference to a DOM object. Must only be used as a field in other DOM objects.

JSRef
A reference to a DOM object that is guaranteed to be alive. This is freely copyable.

LayoutJS
An unrooted reference to a DOM object for use in layout. Layout*Helpers traits must be implemented on this.

MutHeap
A holder that provides interior mutability for GC-managed values such as JSVal and JS<T>.

MutNullableJS
A mutable JS<T> value, with nullability represented by an enclosing Option wrapper. Must be used in place of traditional internal mutability to ensure that the proper GC barriers are enforced.

Root
A rooted reference to a DOM object.

RootCollection
An opaque, LIFO rooting mechanism. This tracks roots and ensures that they are destructed in a LIFO order.

RootCollectionPtr
A pointer to a RootCollection, for use in global variables.

Temporary
A type that represents a JS-owned value that is rooted for the lifetime of this value. Importantly, it requires explicit rooting in order to interact with the inner value. Can be assigned into JS-owned member fields (i.e. JS<T> types) safely via the JS<T>::assign method or OptionalSettable::assign (for Option<JS<T>> fields).

Unrooted
An unrooted, JS-owned value. Must not be held across a GC.

Traits

Assignable
Trait that allows extracting a JS<T> value from a variety of rooting-related containers, which in general is an unsafe operation since they can outlive the rooted lifetime of the original value.

Comparable
A trait for comparing smart pointers ignoring the lifetimes

HeapGCValue
A trait to be implemented for JS-managed types that can be stored in mutable member fields.

OptionalOptionalRootedRootable
Root a rootable Option<Option> type (used for Option<Option<JS<T>>>)

OptionalRootable
Root a rootable Option type (used for Option<Temporary<T>>)

OptionalRootedReference
Get an Option<Option<JSRef<T>>> out of an Option<Option<Root<T>>>

OptionalRootedRootable
Root a rootable Option type (used for Option<JS<T>>)

OptionalUnrootable
Return an unrooted type for storing in optional DOM fields

ResultRootable
Root a rootable Result type (any of Temporary<T> or JS<T>)

RootedReference
Get an Option<JSRef<T>> out of an Option<Root<T>>

TemporaryPushable
Provides a facility to push unrooted values onto lists of rooted values. This is safe under the assumption that said lists are reachable via the GC graph, and therefore the new values are transitively rooted for the lifetime of their new owner.