Struct script::dom::bindings::js::JS

pub struct JS<T> {
    // some fields omitted
}

A traced reference to a DOM object. Must only be used as a field in other DOM objects.

Methods

impl<T> JS<T>

unsafe fn to_layout(self) -> LayoutJS<T>

Returns LayoutJS<T> containing the same pointer.

impl<T: Reflectable> JS<T>

fn root(&self) -> Root<T>

Root this JS-owned value to prevent its collection as garbage.

impl<U: Reflectable> JS<U>

fn from_rooted<T: Assignable<U>>(root: T) -> JS<U>

Create a JS<T> from any JS-managed pointer.

impl<T: Reflectable> JS<T>

fn assign(&mut self, val: Temporary<T>)

Store an unrooted value in this field. This is safe under the assumption that JS values are only used as fields in DOM types that are reachable in the GC graph, so this unrooted value becomes transitively rooted for the lifetime of its new owner.

impl<From> JS<From>

unsafe fn transmute_copy<To>(&self) -> JS<To>

Return self as a JS of another type.