Struct script::dom::bindings::js::LayoutJS

pub struct LayoutJS<T> {
    // some fields omitted
}

An unrooted reference to a DOM object for use in layout. Layout*Helpers traits must be implemented on this.

Methods

impl<T: Reflectable> LayoutJS<T>

unsafe fn get_jsobject(&self) -> *mut JSObject

Get the reflector.

impl LayoutJS<Node>

unsafe fn from_trusted_node_address(inner: TrustedNodeAddress) -> LayoutJS<Node>

Create a new JS-owned value wrapped from an address known to be a Node pointer.

impl<T: Reflectable> LayoutJS<T>

unsafe fn unsafe_get(&self) -> *const T

Returns an unsafe pointer to the interior of this JS object. This is the only method that be safely accessed from layout. (The fact that this is unsafe is what necessitates the layout wrappers.)

impl<From> LayoutJS<From>

unsafe fn transmute_copy<To>(&self) -> LayoutJS<To>

Return self as a LayoutJS of another type.