Struct script::dom::bindings::js::MutHeap

pub struct MutHeap<T: HeapGCValue + Copy> {
    // some fields omitted
}

A holder that provides interior mutability for GC-managed values such as JSVal and JS<T>.

Must be used in place of traditional interior mutability to ensure proper GC barriers are enforced.

Methods

impl<T: HeapGCValue + Copy> MutHeap<T>

fn new(initial: T) -> MutHeap<T>

Create a new MutHeap.

fn set(&self, val: T)

Set this MutHeap to the given value, calling write barriers as appropriate.

fn get(&self) -> T

Set the value in this MutHeap, calling read barriers as appropriate.