Struct script::dom::bindings::js::MutHeapJSVal

pub struct MutHeapJSVal {
    val: UnsafeCell<Heap<JSVal>>,
}

A holder that provides interior mutability for GC-managed JSVals.

Must be used in place of traditional interior mutability to ensure proper GC barriers are enforced.

Fields
val: UnsafeCell<Heap<JSVal>>
Methods

impl MutHeapJSVal

fn new() -> MutHeapJSVal

Create a new MutHeapJSVal.

fn set(&self, val: JSVal)

Set this MutHeapJSVal to the given value, calling write barriers as appropriate.

fn get(&self) -> JSVal

Get the value in this MutHeapJSVal, calling read barriers as appropriate.