Struct script::dom::bindings::js::MutNullableHeap

pub struct MutNullableHeap<T: HeapGCValue> {
    ptr: UnsafeCell<Option<T>>,
}

A holder that provides interior mutability for GC-managed values such as JS<T>, with nullability represented by an enclosing Option wrapper. Essentially a Cell<Option<JS<T>>>, but safer.

This should only be used as a field in other DOM objects; see warning on JS<T>.

Fields

ptr

Methods

impl<T: Reflectable> MutNullableHeap<JS<T>>

fn new(initial: Option<&T>) -> MutNullableHeap<JS<T>>

Create a new MutNullableHeap.

fn or_init<F>(&self, cb: F) -> Root<T> where F: FnOnce() -> Root<T>

Retrieve a copy of the current inner value. If it is None, it is initialized with the result of cb first.

unsafe fn get_inner_as_layout(&self) -> Option<LayoutJS<T>>

Retrieve a copy of the inner optional JS<T> as LayoutJS<T>. For use by layout, which can't use safe types like Temporary.

fn get(&self) -> Option<Root<T>>

Get a rooted value out of this object

fn set(&self, val: Option<&T>)

Set this MutNullableHeap to the given value.