Struct script::dom::bindings::js::MutNullableJS

pub struct MutNullableJS<T: Reflectable> {
    // some fields omitted
}

A mutable JS<T> value, with nullability represented by an enclosing Option wrapper. Must be used in place of traditional internal mutability to ensure that the proper GC barriers are enforced.

Methods

impl<U: Reflectable> MutNullableJS<U>

fn new<T: Assignable<U>>(initial: Option<T>) -> MutNullableJS<U>

Create a new MutNullableJS

impl<T: Reflectable> MutNullableJS<T>

fn assign<U: Assignable<T>>(&self, val: Option<U>)

Store an unrooted value in this field. This is safe under the assumption that MutNullableJS<T> values are only used as fields in DOM types that are reachable in the GC graph, so this unrooted value becomes transitively rooted for the lifetime of its new owner.

fn clear(&self)

Set the inner value to null, without making API users jump through useless type-ascription hoops.

fn get(&self) -> Option<Temporary<T>>

Retrieve a copy of the current optional inner value.

unsafe fn get_inner_as_layout(&self) -> Option<LayoutJS<T>>

Retrieve a copy of the inner optional JS<T> as LayoutJS<T>. For use by layout, which can't use safe types like Temporary.

fn or_init<F>(&self, cb: F) -> Temporary<T> where F: FnOnce() -> Temporary<T>

Retrieve a copy of the current inner value. If it is None, it is initialized with the result of cb first.