Struct script::dom::bindings::js::Root

pub struct Root<T: DomObject> {
    ptr: NonZero<*const T>,
    root_list: *const RootCollection,
}

A rooted reference to a DOM object.

The JS value is pinned for the duration of this object's lifetime; roots are additive, so this object's destruction will not invalidate other roots for the same JS value. Roots cannot outlive the associated RootCollection object.

Fields
ptr: NonZero<*const T>

Reference to rooted value that must not outlive this container

root_list: *const RootCollection

List that ensures correct dynamic root ordering

Methods

impl<T: Castable> Root<T>

fn upcast<U>(root: Root<T>) -> Root<U> where U: Castable, T: DerivedFrom<U>

Cast a DOM object root upwards to one of the interfaces it derives from.

fn downcast<U>(root: Root<T>) -> Option<Root<U>> where U: DerivedFrom<T>

Cast a DOM object root downwards to one of the interfaces it might implement.

impl<T: DomObject> Root<T>

fn new(unrooted: NonZero<*const T>) -> Root<T>

Create a new stack-bounded root for the provided JS-owned value. It cannot outlive its associated RootCollection, and it gives out references which cannot outlive this new Root.

fn from_ref(unrooted: &T) -> Root<T>

Generate a new root from a reference