Struct script::dom::bindings::js::Temporary

pub struct Temporary<T> {
    // some fields omitted
}

A type that represents a JS-owned value that is rooted for the lifetime of this value. Importantly, it requires explicit rooting in order to interact with the inner value. Can be assigned into JS-owned member fields (i.e. JS<T> types) safely via the JS<T>::assign method or OptionalSettable::assign (for Option<JS<T>> fields).

Methods

impl<T: Reflectable> Temporary<T>

fn new(inner: JS<T>) -> Temporary<T>

Create a new Temporary value from a JS-owned value.

fn from_unrooted(unrooted: Unrooted<T>) -> Temporary<T>

Create a new Temporary value from an unrooted value.

fn from_rooted<'a>(root: JSRef<'a, T>) -> Temporary<T>

Create a new Temporary value from a rooted value.

fn root(self) -> Root<T>

Create a stack-bounded root for this value.

unsafe fn transmute<To>(self) -> Temporary<To>

Returns self as a Temporary of another type. For use by InheritTypes only.