Struct script::dom::bindings::js::Unrooted

pub struct Unrooted<T> {
    // some fields omitted
}

An unrooted, JS-owned value. Must not be held across a GC.

This is used in particular to wrap pointers extracted from a reflector.

Methods

impl<T: Reflectable> Unrooted<T>

unsafe fn from_raw(raw: *const T) -> Unrooted<T>

Create a new JS-owned value wrapped from a raw Rust pointer.

fn from_js(ptr: JS<T>) -> Unrooted<T>

Create a new unrooted value from a JS<T>.

fn from_temporary(ptr: Temporary<T>) -> Unrooted<T>

Create a new unrooted value from a Temporary<T>.

fn reflector<'a>(&'a self) -> &'a Reflector

Get the Reflector for this pointer.

unsafe fn unsafe_get(&self) -> *const T

Returns an unsafe pointer to the interior of this object.

fn root(self) -> Root<T>

Create a stack-bounded root for this value.