Trait script::dom::bindings::js::Comparable

pub trait Comparable<T> {
    fn equals(&self, other: T) -> bool;
}

A trait for comparing smart pointers ignoring the lifetimes

Required Methods

fn equals(&self, other: T) -> bool

Returns whether the other value points to the same object.