Trait script::dom::bindings::js::TemporaryPushable

pub trait TemporaryPushable<T> {
    fn push_unrooted(&mut self, val: &T);
    fn insert_unrooted(&mut self, index: usize, val: &T);
}

Provides a facility to push unrooted values onto lists of rooted values. This is safe under the assumption that said lists are reachable via the GC graph, and therefore the new values are transitively rooted for the lifetime of their new owner.

Required Methods

fn push_unrooted(&mut self, val: &T)

Push a new value onto this container.

fn insert_unrooted(&mut self, index: usize, val: &T)

Insert a new value into this container.