Function script::dom::bindings::proxyhandler::get_property_descriptor

pub unsafe fn get_property_descriptor(cx: *mut JSContext, proxy: *mut JSObject, id: jsid, set: bool, desc: *mut JSPropertyDescriptor) -> bool

Invoke the [[GetOwnProperty]] trap (getOwnPropertyDescriptor) on proxy, with argument id and return the result, if it is not undefined. Otherwise, walk along the prototype chain to find a property with that name.