Struct script::dom::bindings::refcounted::Trusted

pub struct Trusted<T> {
    // some fields omitted
}

A safe wrapper around a raw pointer to a DOM object that can be shared among tasks for use in asynchronous operations. The underlying DOM object is guaranteed to live at least as long as the last outstanding Trusted<T> instance.

Methods

impl<T: Reflectable> Trusted<T>

fn new(cx: *mut JSContext, ptr: JSRef<T>, script_chan: Box<ScriptChan + Send>) -> Trusted<T>

Create a new Trusted<T> instance from an existing DOM pointer. The DOM object will be prevented from being GCed for the duration of the resulting Trusted<T> object's lifetime.

fn to_temporary(&self) -> Temporary<T>

Obtain a usable DOM pointer from a pinned Trusted<T> value. Fails if used on a different thread than the original value from which this Trusted<T> was obtained.