Struct script::dom::bindings::refcounted::TrustedPromise

pub struct TrustedPromise {
    dom_object: *const Promise,
    owner_thread: *const c_void,
}

A safe wrapper around a DOM Promise object that can be shared among threads for use in asynchronous operations. The underlying DOM object is guaranteed to live at least as long as the last outstanding TrustedPromise instance. These values cannot be cloned, only created from existing Rc values.

Fields
dom_object: *const Promise owner_thread: *const c_void
Methods

impl TrustedPromise

fn new(promise: Rc<Promise>) -> TrustedPromise

Create a new TrustedPromise instance from an existing DOM object. The object will be prevented from being GCed for the duration of the resulting TrustedPromise object's lifetime.

fn root(self) -> Rc<Promise>

Obtain a usable DOM Promise from a pinned TrustedPromise value. Fails if used on a different thread than the original value from which this TrustedPromise was obtained.