Trait script::dom::bindings::reflector::DomObject

pub trait DomObject: JSTraceable + 'static {
    // Required method
    fn reflector(&self) -> &Reflector;

    // Provided method
    fn global(&self) -> Root<Dom<GlobalScope>>
       where Self: Sized { ... }
}

A trait to provide access to the Reflector for a DOM object.

Required Methods

fn reflector(&self) -> &Reflector

Returns the receiver’s reflector.

Provided Methods

fn global(&self) -> Root<Dom<GlobalScope>>
where Self: Sized,

Returns the global scope of the realm that the DomObject was created in.