Trait script::dom::bindings::reflector::Reflectable

pub trait Reflectable {
    fn reflector(&self) -> &Reflector;
    fn init_reflector(&mut self, obj: *mut JSObject);

    fn global(&self) -> GlobalRoot where Self: Sized { ... }
}

A trait to provide access to the Reflector for a DOM object.

Required Methods

fn reflector(&self) -> &Reflector

Returns the receiver's reflector.

fn init_reflector(&mut self, obj: *mut JSObject)

Initializes the Reflector

Provided Methods

fn global(&self) -> GlobalRoot where Self: Sized

Returns the global object of the realm that the Reflectable was created in.