script::dom::bindings::root

Struct LayoutDom

#[repr(transparent)]
pub struct LayoutDom<'dom, T> {
    value: &'dom T,
}

An unrooted reference to a DOM object for use in layout. Layout*Helpers traits must be implemented on this.

Fields
value: &'dom T
Implementations

impl LayoutDom<'_, Node>

pub(crate) unsafe fn from_trusted_node_address( inner: TrustedNodeAddress, ) -> Self

Create a new JS-owned value wrapped from an address known to be a Node pointer.

impl<'dom, T> LayoutDom<'dom, T>
where T: 'dom + DomObject,

pub fn unsafe_get(self) -> &'dom T

Returns a reference to the interior of this JS object. The fact that this is unsafe is what necessitates the layout wrappers.

pub(crate) unsafe fn to_layout_slice( slice: &'dom [Dom<T>], ) -> &'dom [LayoutDom<'dom, T>]

Transforms a slice of Dom<T> into a slice of LayoutDom<T>.

impl<'dom, T> LayoutDom<'dom, T>
where T: Castable,

pub(crate) fn upcast<U>(&self) -> LayoutDom<'dom, U>
where U: Castable, T: DerivedFrom<U>,

Cast a DOM object root upwards to one of the interfaces it derives from.

pub(crate) fn downcast<U>(&self) -> Option<LayoutDom<'dom, U>>
where U: DerivedFrom<T>,

Cast a DOM object downwards to one of the interfaces it might implement.

pub(crate) fn is<U>(&self) -> bool
where U: DerivedFrom<T>,

Returns whether this inner object is a U.

pub(crate) unsafe fn as_ref(self) -> &'dom T

Get a reference to the internal value.

SAFETY

This function effectively circumvents all the safety provided by LayoutDom as it allows performing arbitrary (potentially mutating) operations on the value. Use with caution!

impl<T> LayoutDom<'_, T>
where T: DomObject,

pub(crate) unsafe fn get_jsobject(&self) -> *mut JSObject

Get the reflector.

impl<'dom> LayoutDom<'dom, CharacterData>

pub(crate) fn data_for_layout(self) -> AtomicRef<'dom, str>

impl<'dom> LayoutDom<'dom, Document>

pub(crate) fn is_html_document_for_layout(&self) -> bool

pub(crate) fn quirks_mode(self) -> QuirksMode

pub(crate) fn shared_style_locks(self) -> &'dom SharedRwLocks

pub(crate) fn flush_shadow_root_stylesheets_if_necessary( self, stylist: &mut Stylist, guard: &SharedRwLockReadGuard<'_>, )

pub(crate) fn elements_with_id(self, id: &Atom) -> &[LayoutDom<'dom, Element>]

pub(crate) fn url_for_layout(self) -> ServoUrl

pub(crate) fn selection_for_layout(&self) -> Option<LayoutDom<'dom, Selection>>

impl<'dom> LayoutDom<'dom, DocumentFragment>

pub(crate) fn shadowroot_host_for_layout(self) -> LayoutDom<'dom, Element>

impl<'dom> LayoutDom<'dom, Attr>

pub(crate) fn value(self) -> &'dom AttrValue

pub(crate) fn local_name(self) -> &'dom LocalName

pub(crate) fn namespace(self) -> &'dom Namespace

impl<'dom> LayoutDom<'dom, Element>

pub(crate) fn is_root(&self) -> bool

pub(crate) fn is_body_element_of_html_element_root(&self) -> bool

Returns true if this element is the body child of an html element root element.

pub(crate) fn each_attr_name_for_layout<F>(self, callback: F)
where F: FnMut(&LocalName),

Iterate over attribute local names for layout, handling mixed Raw and Dom entries.

pub(crate) fn has_class_or_part_for_layout( self, name: &AtomIdent, attr_name: &LocalName, case_sensitivity: CaseSensitivity, ) -> bool

pub(crate) fn get_classes_for_layout(self) -> Option<&'dom [Atom]>

pub(crate) fn get_parts_for_layout(self) -> Option<&'dom [Atom]>

pub(crate) fn style_data(self) -> Option<&'dom StyleData>

pub(crate) unsafe fn initialize_style_data(self)

pub(crate) unsafe fn clear_style_data(self)

pub(crate) fn synthesize_presentational_hints_for_legacy_attributes<V>( self, hints: &mut V, )
where V: Push<ApplicableDeclarationBlock>,

pub(crate) fn get_span(self) -> Option<u32>

pub(crate) fn get_colspan(self) -> Option<u32>

pub(crate) fn get_rowspan(self) -> Option<u32>

pub(crate) fn is_html_element(&self) -> bool

pub(crate) fn id_attribute(self) -> *const Option<Atom>

pub(crate) fn style_attribute( self, ) -> *const Option<ServoArc<Locked<PropertyDeclarationBlock>>>

pub(crate) fn local_name(self) -> &'dom LocalName

pub(crate) fn namespace(self) -> &'dom Namespace

pub(crate) fn get_lang_attr_val_for_layout(self) -> Option<&'dom str>

pub(crate) fn get_lang_for_layout(self) -> AtomString

pub(crate) fn get_state_for_layout(self) -> ElementState

pub(crate) fn insert_selector_flags(self, flags: ElementSelectorFlags)

pub(crate) fn get_selector_flags(self) -> ElementSelectorFlags

pub(crate) fn get_shadow_root_for_layout( self, ) -> Option<LayoutDom<'dom, ShadowRoot>>

pub(crate) fn get_attr_for_layout( self, namespace: &Namespace, name: &LocalName, ) -> Option<&'dom AttrValue>

pub(crate) fn get_attr_val_for_layout( self, namespace: &Namespace, name: &LocalName, ) -> Option<&'dom str>

pub(crate) fn get_attr_vals_for_layout( self, name: &LocalName, ) -> impl Iterator<Item = &'dom AttrValue>

pub(crate) fn each_custom_state_for_layout( self, callback: impl FnMut(&AtomIdent), )

impl LayoutDom<'_, HTMLBodyElement>

pub(crate) fn get_background_color(self) -> Option<AbsoluteColor>

pub(crate) fn get_color(self) -> Option<AbsoluteColor>

impl LayoutDom<'_, HTMLCanvasElement>

pub(crate) fn data(self) -> HTMLCanvasData

impl LayoutDom<'_, HTMLIFrameElement>

pub(crate) fn pipeline_id(self) -> Option<PipelineId>

pub(crate) fn browsing_context_id(self) -> Option<BrowsingContextId>

pub(crate) fn get_width(self) -> LengthOrPercentageOrAuto

pub(crate) fn get_height(self) -> LengthOrPercentageOrAuto

impl<'dom> LayoutDom<'dom, HTMLImageElement>

fn current_request(self) -> &'dom ImageRequest

fn dimension_attribute_source(self) -> LayoutDom<'dom, Element>

pub(crate) fn image_url(self) -> Option<ServoUrl>

pub(crate) fn image_data(self) -> (Option<Image>, Option<ImageMetadata>)

pub(crate) fn image_density(self) -> Option<f64>

pub(crate) fn showing_broken_image_icon(self) -> bool

pub(crate) fn get_width(self) -> LengthOrPercentageOrAuto

pub(crate) fn get_height(self) -> LengthOrPercentageOrAuto

impl LayoutDom<'_, HTMLVideoElement>

pub(crate) fn data(self) -> HTMLMediaData

pub(crate) fn get_width(self) -> LengthOrPercentageOrAuto

pub(crate) fn get_height(self) -> LengthOrPercentageOrAuto

impl<'dom> LayoutDom<'dom, HTMLInputElement>

pub(crate) fn size_for_layout(self) -> u32

Textual input, specifically text entry and domain specific input has a default preferred size.

https://html.spec.whatwg.org/multipage/#the-input-element-as-a-text-entry-widget https://html.spec.whatwg.org/multipage/#the-input-element-as-domain-specific-widgets

pub(crate) fn selection_for_layout(self) -> Option<RangeAny<Utf32CodeUnits>>

impl LayoutDom<'_, HTMLTextAreaElement>

pub(crate) fn selection_for_layout(self) -> Option<RangeAny<Utf32CodeUnits>>

pub(crate) fn get_cols(self) -> u32

pub(crate) fn get_rows(self) -> u32

impl LayoutDom<'_, HTMLHRElement>

pub(crate) fn get_color(self) -> Option<AbsoluteColor>

pub(crate) fn get_width(self) -> LengthOrPercentageOrAuto

pub(crate) fn get_size_info(self) -> Option<SizePresentationalHint>

impl<'dom> LayoutDom<'dom, HTMLTableCellElement>

pub(crate) fn get_background_color(self) -> Option<AbsoluteColor>

pub(crate) fn get_colspan(self) -> Option<u32>

pub(crate) fn get_rowspan(self) -> Option<u32>

pub(crate) fn get_table(self) -> Option<LayoutDom<'dom, HTMLTableElement>>

pub(crate) fn get_width(self) -> LengthOrPercentageOrAuto

pub(crate) fn get_height(self) -> LengthOrPercentageOrAuto

impl<'dom> LayoutDom<'dom, HTMLTableColElement>

pub(crate) fn get_span(self) -> Option<u32>

pub(crate) fn get_width(self) -> LengthOrPercentageOrAuto

impl LayoutDom<'_, HTMLTableElement>

pub(crate) fn get_background_color(self) -> Option<AbsoluteColor>

pub(crate) fn get_border(self) -> Option<u32>

pub(crate) fn get_cellpadding(self) -> Option<u32>

pub(crate) fn get_cellspacing(self) -> Option<u32>

pub(crate) fn get_width(self) -> LengthOrPercentageOrAuto

pub(crate) fn get_height(self) -> LengthOrPercentageOrAuto

impl LayoutDom<'_, HTMLTableRowElement>

pub(crate) fn get_background_color(self) -> Option<AbsoluteColor>

pub(crate) fn get_height(self) -> LengthOrPercentageOrAuto

impl LayoutDom<'_, HTMLTableSectionElement>

pub(crate) fn get_background_color(self) -> Option<AbsoluteColor>

pub(crate) fn get_height(self) -> LengthOrPercentageOrAuto

impl LayoutDom<'_, HTMLFontElement>

pub(crate) fn get_color(self) -> Option<AbsoluteColor>

pub(crate) fn get_face(self) -> Option<Atom>

pub(crate) fn get_size(self) -> Option<u32>

impl<'dom> LayoutDom<'dom, Node>

pub(crate) fn parent_node_ref(self) -> Option<LayoutDom<'dom, Node>>

pub(crate) fn type_id_for_layout(self) -> NodeTypeId

pub(crate) fn is_element_for_layout(&self) -> bool

pub(crate) fn is_text_node_for_layout(&self) -> bool

pub(crate) fn composed_parent_node_ref(self) -> Option<LayoutDom<'dom, Node>>

pub(crate) fn traversal_parent(self) -> Option<LayoutDom<'dom, Element>>

pub(crate) fn first_child_ref(self) -> Option<LayoutDom<'dom, Node>>

pub(crate) fn last_child_ref(self) -> Option<LayoutDom<'dom, Node>>

pub(crate) fn prev_sibling_ref(self) -> Option<LayoutDom<'dom, Node>>

pub(crate) fn next_sibling_ref(self) -> Option<LayoutDom<'dom, Node>>

pub(crate) fn owner_doc_for_layout(self) -> LayoutDom<'dom, Document>

pub(crate) fn containing_shadow_root_for_layout( self, ) -> Option<LayoutDom<'dom, ShadowRoot>>

pub(crate) fn assigned_slot_for_layout( self, ) -> Option<LayoutDom<'dom, HTMLSlotElement>>

pub(crate) unsafe fn get_flag(self, flag: NodeFlags) -> bool

pub(crate) unsafe fn set_flag(self, flag: NodeFlags, value: bool)

pub(crate) fn layout_data(self) -> Option<&'dom GenericLayoutData>

pub(crate) unsafe fn initialize_layout_data( self, new_data: Box<GenericLayoutData>, )

Initialize the style data of this node.

Safety

This method is unsafe because it modifies the given node during layout. Callers should ensure that no other layout thread is attempting to read or modify the opaque layout data of this node.

pub(crate) unsafe fn clear_layout_data(self)

Clear the style and opaque layout data of this node.

Safety

This method is unsafe because it modifies the given node during layout. Callers should ensure that no other layout thread is attempting to read or modify the opaque layout data of this node.

pub(crate) fn is_single_line_text_inner_editor(&self) -> bool

Whether this element serve as a container of editable text for a text input that is implemented as an UA widget.

pub(crate) fn is_text_container_of_single_line_input(&self) -> bool

Whether this element serve as a container of any text inside a text input that is implemented as an UA widget.

pub(crate) fn text_content(self) -> AtomicRef<'dom, str>

pub(crate) fn selection_for_text_node(&self) -> Option<RangeAny<Utf32CodeUnits>>

pub(crate) fn text_node_paints_caret(&self) -> bool

pub(crate) fn image_url(self) -> Option<ServoUrl>

pub(crate) fn image_data(self) -> Option<(Option<Image>, Option<ImageMetadata>)>

pub(crate) fn image_density(self) -> Option<f64>

pub(crate) fn showing_broken_image_icon(self) -> bool

pub(crate) fn canvas_data(self) -> Option<HTMLCanvasData>

pub(crate) fn media_data(self) -> Option<HTMLMediaData>

pub(crate) fn svg_data(self) -> Option<SVGElementData<'dom>>

pub(crate) fn iframe_browsing_context_id(self) -> Option<BrowsingContextId>

pub(crate) fn iframe_pipeline_id(self) -> Option<PipelineId>

pub(crate) fn opaque(self) -> OpaqueNode

pub(crate) fn implemented_pseudo_element(&self) -> Option<PseudoElement>

pub(crate) fn is_in_ua_widget(&self) -> bool

pub(crate) fn is_root_of_user_agent_widget(&self) -> bool

pub(crate) fn children_count(&self) -> u32

impl<'dom> LayoutDom<'dom, Selection>

pub(crate) fn range_for_layout(&self) -> &Option<SelectionRange>

impl<'dom> LayoutDom<'dom, ShadowRoot>

pub(crate) fn get_host_for_layout(self) -> LayoutDom<'dom, Element>

pub(crate) fn get_style_data_for_layout(self) -> &'dom CascadeData

pub(crate) fn is_user_agent_widget(&self) -> bool

pub(crate) unsafe fn flush_stylesheets_for_layout( self, stylist: &mut Stylist, guard: &SharedRwLockReadGuard<'_>, )

impl<'dom> LayoutDom<'dom, SVGElement>

pub(crate) fn synthesize_presentational_hints( self, document: LayoutDom<'dom, Document>, push: &mut impl FnMut(PropertyDeclaration), )

fn parse_svg_attribute<F>( self, parser_context: &ParserContext<'_>, attr_name: &str, parse: F, push: &mut impl FnMut(PropertyDeclaration), )
where F: for<'i, 't> FnOnce(&ParserContext<'_>, &mut Parser<'i, 't>) -> Result<PropertyDeclaration, ParseError<'i>>,

impl<'dom> LayoutDom<'dom, SVGSVGElement>

pub(crate) fn data(self) -> SVGElementData<'dom>