script::dom::bindings::str

Function max_week_in_year

fn max_week_in_year(year: i32) -> u32

https://html.spec.whatwg.org/multipage/#week-number-of-the-last-day

A week-year with a number year has 53 weeks if it corresponds to either a year year in the proleptic Gregorian calendar that has a Thursday as its first day (January 1st), or a year year in the proleptic Gregorian calendar that has a Wednesday as its first day (January 1st) and where year is a number divisible by 400, or a number divisible by 4 but not by 100. All other week-years have 52 weeks.