Struct script::dom::bindings::str::ByteString

pub struct ByteString(_);

Encapsulates the IDL ByteString type.

Methods

impl ByteString

fn new(value: Vec<u8>) -> ByteString

Creates a new ByteString.

fn as_str<'a>(&'a self) -> Option<&'a str>

Returns self as a string, if it encodes valid UTF-8, and None otherwise.

fn as_slice<'a>(&'a self) -> &'a [u8]

Returns the underlying vector as a slice.

fn len(&self) -> usize

Returns the length.

fn eq_ignore_case(&self, other: &ByteString) -> bool

Compare self to other, matching A–Z and a–z as equal.

fn to_lower(&self) -> ByteString

Returns self with A–Z replaced by a–z.

fn is_token(&self) -> bool

Returns whether self is a token, as defined by RFC 2616.

fn is_field_value(&self) -> bool

Returns whether self is a field-value, as defined by RFC 2616.