script::dom::bindings::str

Trait ToInputValueString

pub(crate) trait ToInputValueString {
    // Required methods
    fn to_date_string(&self) -> String;
    fn to_month_string(&self) -> String;
    fn to_week_string(&self) -> String;
    fn to_time_string(&self) -> String;
    fn to_local_date_time_string(&self) -> String;
}

Required Methods

fn to_date_string(&self) -> String

fn to_month_string(&self) -> String

fn to_week_string(&self) -> String

fn to_time_string(&self) -> String

fn to_local_date_time_string(&self) -> String

A valid normalized local date and time string should be “{date}T{time}” where date and time are both valid, and the time string must be as short as possible https://html.spec.whatwg.org/multipage/#valid-normalised-local-date-and-time-string