script::dom::bindings

Module structuredclone

This module implements structured cloning, as defined by HTML.

Structs

StructuredDataReader 🔒

Reader and writer structs for results from, and inputs to, structured-data read/write operations. https://html.spec.whatwg.org/multipage/#safe-passing-of-structured-data

StructuredDataWriter 🔒

A data holder for transferred and serialized objects.

Enums

OperationError 🔒

StructuredCloneTags 🔒

StructuredData 🔒

Statics

STRUCTURED_CLONE_CALLBACKS 🔒

Functions

can_transfer_callback 🔒 ⚠

can_transfer_for_type 🔒 ⚠

free_transfer_callback 🔒 ⚠

read 🔒

Read structured serialized data, possibly containing transferred objects. Returns a vec of rooted transfer-received ports, or an error.

read_callback 🔒 ⚠

read_object 🔒 ⚠

read_transfer_callback 🔒 ⚠

reader_for_type 🔒

receive_object 🔒

receiver_for_type 🔒

report_error_callback 🔒 ⚠

sab_cloned_callback 🔒 ⚠

serialize_for_type 🔒

transfer_for_type 🔒

try_serialize 🔒 ⚠

try_transfer 🔒 ⚠

write 🔒

Writes a structured clone. Returns a DataClone error if that fails.

write_callback 🔒 ⚠

write_object 🔒 ⚠

write_transfer_callback 🔒 ⚠

https://html.spec.whatwg.org/multipage/#structuredserializewithtransfer

Type Aliases

SerializeOperation 🔒

TransferOperation 🔒