Struct script::dom::bindings::trace::RootedVec

pub struct RootedVec<T> {
    // some fields omitted
}

A vector of items that are rooted for the lifetime of this struct

Methods

impl<T: VecRootableType> RootedVec<T>

fn new() -> RootedVec<T>

Create a vector of items of type T that is rooted for the lifetime of this struct