script::dom::bindings

Module utils

Various utilities to glue JavaScript and the DOM implementation together.

Modules

call_policies 🔒

Structs

CallPolicyInfo 🔒

Controls various details of an IDL operation, such as whether a [LegacyLenientThis] attribute is specified and preconditions that affect the outcome of the “perform a security check” steps.

DOMClass 🔒

The struct that holds inheritance information for DOM object reflectors.

DOMJSClass 🔒

The JSClass used for DOM object reflectors.

GlobalStaticData 🔒

Static data associated with a global object.

Constants

DOM_CALLBACKS 🔒

Traits

CallPolicy 🔒

Functions

define_all_exposed_interfaces 🔒

Eagerly define all relevant WebIDL interface constructors on the provided global object.

define_dictionary_property 🔒

Define an own enumerable data property with name property on object. Returns Err(()) on JSAPI failure, or null object, and Ok(()) otherwise.

extract_exception_info 🔒 ⚠

get_array_index_from_id 🔒

Get an array index from the given jsid. Returns None if the given jsid is not an integer.

has_own_property 🔒

Checks whether object has an own property named property. Returns Err(()) on JSAPI failure (there is a pending exception), and Ok(false) for null objects or when the property is not own.

has_property_on_prototype 🔒

Computes whether proxy has a property id on its prototype and stores the result in found.

instance_class_has_proto_at_depth 🔒 ⚠

instance_class_is_error 🔒 ⚠

https://searchfox.org/mozilla-central/rev/c18faaae88b30182e487fa3341bc7d923e22f23a/xpcom/base/CycleCollectedJSRuntime.cpp#792

is_platform_object 🔒

is_platform_object_static 🔒

Returns wether obj is a platform object using static unwrap https://heycam.github.io/webidl/#dfn-platform-object

set_dictionary_property 🔒

Set the property with name property from object. Returns Err(()) on JSAPI failure, or null object, and Ok(()) otherwise

to_frozen_array 🔒

Returns a JSVal representing the frozen JavaScript array

Type Aliases

ProtoOrIfaceArray 🔒

An array of *mut JSObject of size PROTO_OR_IFACE_LENGTH.