Struct script::dom::bindings::utils::Prefable

pub struct Prefable<T: 'static> {
    pub pref: Option<&'static str>,
    pub specs: &'static [T],
    pub terminator: bool,
}

A container around JS member specifications that are conditionally enabled.

Fields

pref

If present, the name of the preference used to conditionally enable these specs.

specs

The underlying slice of specifications.

terminator

Whether the specifications contain special terminating entries that should be included or not.

Methods

impl<T> Prefable<T>

fn specs(&self) -> &'static [T]

Retrieve the slice represented by this container, unless the condition guarding it is false.