script::dom::bindings

Module weakref

Structs

DOMTracker 🔒

MutableWeakRef 🔒

A mutable weak reference to a JS-managed DOM object. On tracing, the contained weak reference is dropped if the pointee was already collected.

WeakRef 🔒

A weak reference to a JS-managed DOM object.

WeakRefEntry 🔒

An entry of a vector of weak references. Passed to the closure given to WeakRefVec::update.

WeakRefVec 🔒

A vector of weak references. On tracing, the vector retains only references which still point to live objects.

Traits

WeakReferenceable 🔒

Trait implemented by weak-referenceable interfaces.