Enum script::dom::blob::BlobImpl

pub enum BlobImpl {
    File(FileBlob),
    Memory(Vec<u8>),
    Sliced(JS<Blob>, RelativePos),
}

Different backends of Blob

Variants
File(FileBlob)

File-based blob, whose content lives in the net process

Memory(Vec<u8>)

Memory-based blob, whose content lives in the script process

Sliced(JS<Blob>, RelativePos)

Sliced blob, including parent blob reference and relative positions of current slicing range, IMPORTANT: The depth of tree is only two, i.e. the parent Blob must be either File-based or Memory-based

Methods

impl BlobImpl

fn new_from_bytes(bytes: Vec<u8>) -> BlobImpl

Construct memory-backed BlobImpl

fn new_from_file(file_id: Uuid, name: PathBuf, size: u64) -> BlobImpl

Construct file-backed BlobImpl from File ID