Struct script::dom::blob::Blob

pub struct Blob {
    reflector_: Reflector,
    blob_impl: DOMRefCell<BlobImpl>,
    type_string: String,
    is_closed: Cell<bool>,
}

Fields
reflector_: Reflector blob_impl: DOMRefCell<BlobImpl> type_string: String

content-type string

is_closed: Cell<bool>
Methods

impl Blob

fn type_id(&self) -> &'static BlobTypeId

impl Blob

fn new(global: &GlobalScope,
blob_impl: BlobImpl,
typeString: String)
-> Root<Blob>

fn new_inherited(blob_impl: BlobImpl, type_string: String) -> Blob

fn new_sliced(parent: &Blob,
rel_pos: RelativePos,
relative_content_type: DOMString)
-> Root<Blob>

fn Constructor(global: &GlobalScope,
blobParts: Option<Vec<BlobOrString>>,
blobPropertyBag: &BlobPropertyBag)
-> Fallible<Root<Blob>>

fn get_bytes(&self) -> Result<Vec<u8>, ()>

Get a slice to inner data, this might incur synchronous read and caching

fn get_blob_url_id(&self) -> Uuid

Get a FileID representing the Blob content, used by URL.createObjectURL

fn promote(&self, set_valid: bool) -> Uuid

Promote non-Slice blob: 1. Memory-based: The bytes in data slice will be transferred to file manager thread. 2. File-based: If set_valid, then activate the FileID so it can serve as URL Depending on set_valid, the returned FileID can be part of valid or invalid Blob URL.

fn create_sliced_url_id(&self,
parent_id: &Uuid,
rel_pos: &RelativePos,
parent_len: u64)
-> Uuid

Get a FileID representing sliced parent-blob content

fn clean_up_file_resource(&self)

Cleanups at the time of destruction/closing

fn send_to_file_manager(&self, msg: FileManagerThreadMsg)