Struct script::dom::browsingcontext::BrowsingContext

#[repr(C)]

pub struct BrowsingContext {
    reflector: Reflector,
    frame_id: FrameId,
    currently_active: Cell<Option<PipelineId>>,
    discarded: Cell<bool>,
    frame_element: Option<JS<Element>>,
    parent: Option<JS<BrowsingContext>>,
}

Fields
reflector: Reflector

The WindowProxy object. Unlike other reflectors, we mutate this field because we have to brain-transplant the reflector when the WindowProxy changes Window.

frame_id: FrameId

The frame id of the browsing context. In the case that this is a nested browsing context, this is the frame id of the container.

currently_active: Cell<Option<PipelineId>>

The pipeline id of the currently active document. May be None, when the currently active document is in another script thread. We do not try to keep the pipeline id for documents in other threads, as this would require the constellation notifying many script threads about the change, which could be expensive.

discarded: Cell<bool>

Has this browsing context been discarded?

frame_element: Option<JS<Element>>

The containing iframe element, if this is a same-origin iframe

parent: Option<JS<BrowsingContext>>

The parent browsing context, if this is a nested browsing context

Methods

impl BrowsingContext

fn new_inherited(frame_id: FrameId,
currently_active: Option<PipelineId>,
frame_element: Option<&Element>,
parent: Option<&BrowsingContext>)
-> BrowsingContext

fn new(window: &Window,
frame_id: FrameId,
frame_element: Option<&Element>,
parent: Option<&BrowsingContext>)
-> Root<BrowsingContext>

fn new_dissimilar_origin(global_to_clone_from: &GlobalScope,
frame_id: FrameId,
parent: Option<&BrowsingContext>)
-> Root<BrowsingContext>

fn discard(&self)

fn is_discarded(&self) -> bool

fn frame_id(&self) -> FrameId

fn frame_element(&self) -> Option<&Element>

fn parent(&self) -> Option<&BrowsingContext>

fn top(&self) -> &BrowsingContext

fn set_window_proxy(&self, window: &GlobalScope, traps: &ProxyTraps)

Change the Window that this browsing context's WindowProxy resolves to.

fn set_currently_active(&self, window: &Window)

fn unset_currently_active(&self)

fn currently_active(&self) -> Option<PipelineId>

fn window_proxy(&self) -> *mut JSObject