script::dom

Module console

Structs

Console 🔒

https://developer.mozilla.org/en-US/docs/Web/API/Console

Enums

IncludeStackTrace 🔒

Constants

MAX_LOG_CHILDREN 🔒

The maximum elements in an object logged by console methods.

MAX_LOG_DEPTH 🔒

The maximum object depth logged by console methods.

Functions

accessor_value_from_property_descriptor 🔒

apply_sprintf_substitutions 🔒

Apply sprintf-style substitutions to console format arguments per the WHATWG Console spec.

console_argument_from_handle_value 🔒

console_map_object_from_handle_value 🔒

console_object_from_handle_value 🔒

format_float_substitution 🔒

format_integer_substitution 🔒

get_js_stack 🔒

handle_value_to_string 🔒

maybe_stringify_dom_object 🔒

stringify_debugger_value 🔒

An implementation of https://console.spec.whatwg.org/#printer. This produces a string version of the argument that is printed to the console.

stringify_handle_value 🔒

stringify_handle_values 🔒