script::dom

Module countqueuingstrategy

Structs

CountQueuingStrategy 🔒

Constants

_IMPL_DOMOBJECT_FOR_CountQueuingStrategy 🔒

Functions

count_queuing_strategy_size 🔒 ⚠

https://streams.spec.whatwg.org/#count-queuing-strategy-size-function

extract_high_water_mark 🔒

Extract the high water mark from a QueuingStrategy. If the high water mark is not set, return the default value.

extract_size_algorithm 🔒

Extract the size algorithm from a QueuingStrategy. If the size algorithm is not set, return a fallback function which always returns 1.