script::dom::css

Module fontface

Structs

FontFace

https://drafts.csswg.org/css-font-loading/#fontface-interface

FontFaceErrorReporter 🔒

Constants

_IMPL_DOMOBJECT_FOR_FontFace 🔒

Functions

parse_font_face_descriptors 🔒

Given the various font face descriptors, construct the equivalent @font-face css rule as a string and parse it using style crate. Returns Err(Error::Syntax) if parsing fails.

serialize_parsed_descriptors 🔒