script::dom::css::fontface

Struct FontFace

#[repr(C)]
pub struct FontFace {
    reflector: Reflector,
    status: Cell<FontFaceLoadStatus>,
    family_name: DomRefCell<DOMString>,
    descriptors: DomRefCell<FontFaceDescriptors>,
    font_face_set: MutNullableDom<FontFaceSet>,
    template: RefCell<Option<(LowercaseFontFamilyName, FontTemplate)>>,
    urls: DomRefCell<Option<SourceList>>,
    font_status_promise: Rc<Promise>,
    css_font_face_rule: DomRefCell<Option<Arc<FontFaceRuleInfo>>>,
}

https://drafts.csswg.org/css-font-loading/#fontface-interface

Fields
reflector: Reflectorstatus: Cell<FontFaceLoadStatus>family_name: DomRefCell<DOMString>descriptors: DomRefCell<FontFaceDescriptors>font_face_set: MutNullableDom<FontFaceSet>

A reference to the FontFaceSet that this FontFace is a member of, if it has been added to one. None otherwise. The spec suggests that a FontFace can be a member of multiple FontFaceSets, but this doesn’t seem to be the case in practice, as the FontFaceSet constructor is not exposed on the global scope.

template: RefCell<Option<(LowercaseFontFamilyName, FontTemplate)>>

This holds the FontTemplate resulting from loading this FontFace, to be used when the FontFace is added to the global FontFaceSet and thus the [FontContext].

urls: DomRefCell<Option<SourceList>>

https://drafts.csswg.org/css-font-loading/#m-fontface-urls-slot

font_status_promise: Rc<Promise>

https://drafts.csswg.org/css-font-loading/#dom-fontface-fontstatuspromise-slot

css_font_face_rule: DomRefCell<Option<Arc<FontFaceRuleInfo>>>

The @font-face rule that this FontFace object is css-connected to, if any.

Implementations

impl FontFace

fn __assert_parent_type(&self)

impl FontFace

fn new_failed_font_face( cx: &mut JSContext, global: &GlobalScope, proto: Option<HandleObject<'_>>, ) -> DomRoot<Self>

Construct a FontFace to be used in the case of failure in parsing the font face descriptors.

fn new_inherited( family_name: DOMString, urls: Option<SourceList>, descriptors: &Descriptors, font_status_promise: Rc<Promise>, ) -> Self

https://drafts.csswg.org/css-font-loading/#font-face-constructor

pub(crate) fn new( cx: &mut JSContext, global: &GlobalScope, proto: Option<HandleObject<'_>>, font_family: DOMString, urls: Option<SourceList>, descriptors: &Descriptors, font_status_promise: Rc<Promise>, ) -> DomRoot<Self>

https://drafts.csswg.org/css-font-loading/#font-face-constructor

pub(crate) fn new_inherited_for_web_font( family_name: DOMString, descriptors: FontFaceDescriptors, src: Option<SourceList>, font_status_promise: Rc<Promise>, font_face_rule: ServoArc<FontFaceRuleInfo>, ) -> Self

Constructs a unrooted FontFace object for a font that is backed by a @font-face rule.

pub(crate) fn new_for_web_font( cx: &mut JSContext, global: &GlobalScope, font_face_rule: ServoArc<FontFaceRuleInfo>, ) -> Option<DomRoot<Self>>

Constructs a FontFace object for a font that is backed by a @font-face rule.

pub(crate) fn disconnect_from_css(&self)

Mark this font face as not css-connected.

pub(crate) fn is_css_connected(&self) -> bool

https://drafts.csswg.org/css-font-loading/#css-connected

pub(crate) fn css_font_face_rule( &self, ) -> Ref<'_, Option<ServoArc<FontFaceRuleInfo>>>

pub(crate) fn is_connected_to_font_face_rule( &self, target_rule: &ServoArc<FontFaceRuleInfo>, ) -> bool

Return true if the FontFace is css-connected and was created by the provided @font-face rule.

fn load_from_data( &self, cx: &mut JSContext, global: &GlobalScope, data: Vec<u8>, )

Step 3 of https://drafts.csswg.org/css-font-loading/#font-face-constructor

pub(super) fn set_associated_font_face_set(&self, font_face_set: &FontFaceSet)

pub(super) fn template(&self) -> Option<(LowercaseFontFamilyName, FontTemplate)>

fn validate_and_set_descriptors( &self, new_descriptors: FontFaceDescriptors, ) -> ErrorResult

Implements the body of the setter for the descriptor attributes of the FontFace interface.

https://drafts.csswg.org/css-font-loading/#fontface-interface: On setting, parse the string according to the grammar for the corresponding @font-face descriptor. If it does not match the grammar, throw a SyntaxError; otherwise, set the attribute to the serialization of the parsed value.

fn font_face_rule(&self, global: &GlobalScope) -> Fallible<FontFaceRule>