script::dom::css::fontfaceset

Struct FontFaceSet

#[repr(C)]
pub(crate) struct FontFaceSet {
    target: EventTarget,
    promise: RefCell<Rc<Promise>>,
    set_entries: DomRefCell<Vec<Dom<FontFace>>>,
}

https://drafts.csswg.org/css-font-loading/#FontFaceSet-interface

Fields
target: EventTargetpromise: RefCell<Rc<Promise>>

https://drafts.csswg.org/css-font-loading/#dom-fontfaceset-readypromise-slot

set_entries: DomRefCell<Vec<Dom<FontFace>>>
Implementations

impl FontFaceSet

fn __assert_parent_type(&self)

impl FontFaceSet

fn new_inherited(promise: Rc<Promise>) -> Self

pub(crate) fn new( cx: &mut JSContext, global: &GlobalScope, proto: Option<HandleObject<'_>>, ) -> DomRoot<Self>

pub(super) fn handle_font_face_status_changed( &self, cx: &mut JSContext, font_face: &FontFace, )

pub(crate) fn fulfill_ready_promise_if_needed(&self, cx: &mut JSContext) -> bool

Fulfill the font ready promise, returning true if it was not already fulfilled beforehand.

pub(crate) fn waiting_to_fullfill_promise(&self) -> bool

fn contains_face(&self, target: &FontFace) -> bool

fn delete_face(&self, target: &FontFace) -> bool

Removes a face from the set’s set entries.

pub(crate) fn switch_to_loading(&self, cx: &mut JSContext)

https://drafts.csswg.org/css-font-loading/#switch-the-fontfaceset-to-loading

fn flush_author_font_set(&self, cx: &mut JSContext)

Runs the CSS cascade to ensure that new @font-face rules have an entry in this set.

pub(crate) fn notify_font_face_rules_removed( &self, removed_font_face_rules: &[ServoArc<FontFaceRuleInfo>], )

Marks the entries corresponding to removed @font-face rules as not css-connected.

fn query_fonts( &self, target_family: &FamilyName, sample_text: &str, ) -> Vec<DomRoot<FontFace>>

Uses the font matching rules to select font faces within self that can be used to render the provided text.

This is used to implement Step 6 of https://drafts.csswg.org/css-font-loading/#find-the-matching-font-faces.

fn find_the_matching_font_faces( &self, document: &Document, font: &str, sample_text: &str, ) -> Result<Vec<DomRoot<FontFace>>, FontQuerySyntaxError>

https://drafts.csswg.org/css-font-loading/#find-the-matching-font-faces