script::dom::debugger::debuggerglobalscope

Struct DebuggerGlobalScope

#[repr(C)]
pub(crate) struct DebuggerGlobalScope {
    global_scope: GlobalScope,
    devtools_to_script_sender: GenericSender<DevtoolScriptControlMsg>,
    get_possible_breakpoints_result_sender: RefCell<Option<GenericSender<Vec<RecommendedBreakpointLocation>>>>,
    eval_result_sender: RefCell<Option<GenericSender<EvaluateJSReply>>>,
    get_list_frame_result_sender: RefCell<Option<GenericSender<Vec<String>>>>,
    get_environment_result_sender: RefCell<Option<GenericSender<String>>>,
    pipeline_id: PipelineId,
    origin: MutableOrigin,
}

Global scope for interacting with the devtools Debugger API.

https://firefox-source-docs.mozilla.org/js/Debugger/

Fields
global_scope: GlobalScopedevtools_to_script_sender: GenericSender<DevtoolScriptControlMsg>get_possible_breakpoints_result_sender: RefCell<Option<GenericSender<Vec<RecommendedBreakpointLocation>>>>eval_result_sender: RefCell<Option<GenericSender<EvaluateJSReply>>>get_list_frame_result_sender: RefCell<Option<GenericSender<Vec<String>>>>get_environment_result_sender: RefCell<Option<GenericSender<String>>>pipeline_id: PipelineIdorigin: MutableOrigin
Implementations

impl DebuggerGlobalScope

fn __assert_parent_type(&self)

impl DebuggerGlobalScope

pub(crate) fn new( debugger_pipeline_id: PipelineId, script_to_devtools_sender: Option<GenericCallback<ScriptToDevtoolsControlMsg>>, devtools_to_script_sender: GenericSender<DevtoolScriptControlMsg>, mem_profiler_chan: ProfilerChan, time_profiler_chan: ProfilerChan, script_to_constellation_sender: ScriptToConstellationSender, script_to_embedder_chan: ScriptToEmbedderChan, resource_threads: ResourceThreads, storage_threads: StorageThreads, gpu_id_hub: Arc<IdentityHub>, cx: &mut JSContext, ) -> DomRoot<Self>

Create a new heap-allocated DebuggerGlobalScope.

debugger_pipeline_id is the pipeline id to use when creating the debugger’s GlobalScope:

in normal script threads, it should be set to PipelineId::new(), because those threads can generate pipeline ids, and they may contain debuggees from more than one pipeline

in web worker threads, it should be set to the pipeline id of the page that created the thread, because those threads can’t generate pipeline ids, and they only contain one debuggee from one pipeline

pub(crate) fn origin(&self) -> MutableOrigin

pub(crate) fn as_global_scope(&self) -> &GlobalScope

pub(crate) fn execute(&self, cx: &mut JSContext)

pub(crate) fn fire_add_debuggee( &self, cx: &mut JSContext, debuggee_global: &GlobalScope, debuggee_pipeline_id: PipelineId, debuggee_worker_id: Option<WorkerId>, )

pub(crate) fn fire_eval( &self, cx: &mut JSContext, code: DOMString, debuggee_pipeline_id: PipelineId, debuggee_worker_id: Option<WorkerId>, frame_actor_id: Option<String>, eager: bool, result_sender: GenericSender<EvaluateJSReply>, )

pub(crate) fn fire_get_possible_breakpoints( &self, cx: &mut JSContext, spidermonkey_id: u32, result_sender: GenericSender<Vec<RecommendedBreakpointLocation>>, )

pub(crate) fn fire_set_breakpoint( &self, cx: &mut JSContext, spidermonkey_id: u32, script_id: u32, offset: u32, )

pub(crate) fn fire_interrupt(&self, cx: &mut JSContext)

pub(crate) fn fire_list_frames( &self, cx: &mut JSContext, pipeline_id: PipelineId, start: u32, count: u32, result_sender: GenericSender<Vec<String>>, )

pub(crate) fn fire_get_environment( &self, cx: &mut JSContext, request: GetEnvironmentRequest, result_sender: GenericSender<String>, )

pub(crate) fn fire_resume( &self, cx: &mut JSContext, resume_limit_type: Option<String>, frame_actor_id: Option<String>, )

pub(crate) fn fire_clear_breakpoint( &self, cx: &mut JSContext, spidermonkey_id: u32, script_id: u32, offset: u32, )

pub(crate) fn fire_blackbox( &self, cx: &mut JSContext, spidermonkey_id: u32, coverage: BlackboxCoverage, )

pub(crate) fn fire_unblackbox( &self, cx: &mut JSContext, spidermonkey_id: u32, coverage: BlackboxCoverage, )

pub(crate) fn pipeline_id(&self) -> PipelineId