Struct script::dom::dedicatedworkerglobalscope::SendableWorkerScriptChan

pub struct SendableWorkerScriptChan {
    // some fields omitted
}

A ScriptChan that can be cloned freely and will silently send a TrustedWorkerAddress with every message. While this SendableWorkerScriptChan is alive, the associated Worker object will remain alive.