Struct script::dom::dissimilaroriginlocation::DissimilarOriginLocation

#[repr(C)]

pub struct DissimilarOriginLocation {
    reflector: Reflector,
    window: JS<DissimilarOriginWindow>,
}

Represents a dissimilar-origin Location that exists in another script thread.

Since the Location is in a different script thread, we cannot access it directly, but some of its accessors (for example setting location.href) still need to function.

Fields
reflector: Reflector

The reflector. Once we have XOWs, this will have a cross-origin wrapper placed around it.

window: JS<DissimilarOriginWindow>

The window associated with this location.

Methods

impl DissimilarOriginLocation

fn new_inherited(window: &DissimilarOriginWindow) -> DissimilarOriginLocation

fn new(window: &DissimilarOriginWindow) -> Root<DissimilarOriginLocation>