Struct script::dom::dissimilaroriginwindow::DissimilarOriginWindow

#[repr(C)]
pub struct DissimilarOriginWindow {
    globalscope: GlobalScope,
    window_proxy: Dom<WindowProxy>,
    location: MutNullableDom<DissimilarOriginLocation>,
}

Represents a dissimilar-origin Window that exists in another script thread.

Since the Window is in a different script thread, we cannot access it directly, but some of its accessors (for example window.parent) still need to function.

In windowproxy.rs, we create a custom window proxy for these windows, that throws security exceptions for most accessors. This is not a replacement for XOWs, but provides belt-and-braces security.

Fields
globalscope: GlobalScope

The global for this window.

window_proxy: Dom<WindowProxy>

The window proxy for this window.

location: MutNullableDom<DissimilarOriginLocation>

The location of this window, initialized lazily.

Implementations

impl DissimilarOriginWindow

fn __assert_parent_type(&self)

impl DissimilarOriginWindow

pub fn new( global_to_clone_from: &GlobalScope, window_proxy: &WindowProxy ) -> Root<Dom<Self>>

pub fn window_proxy(&self) -> Root<Dom<WindowProxy>>

impl DissimilarOriginWindow

fn post_message_impl( &self, target_origin: &USVString, cx: JSContext, message: HandleValue<'_>, transfer: CustomAutoRooterGuard<'_, Vec<*mut JSObject>> ) -> Result<(), Error>

https://html.spec.whatwg.org/multipage/#window-post-message-steps

pub fn post_message( &self, target_origin: &USVString, data: StructuredSerializedData ) -> Result<(), Error>

https://html.spec.whatwg.org/multipage/#window-post-message-steps