script::dom::document

Module document

Structs

Document 🔒

https://dom.spec.whatwg.org/#document

DocumentSizes 🔒

Holds DOM object memory sizes for fine-grained memory reports.

NavigationTiming 🔒

https://html.spec.whatwg.org/multipage/#document-load-timing-info

PendingInOrderScriptVec 🔒

PendingScript 🔒

PendingScrollEvent 🔒

RefreshRedirectDue 🔒

RenderingUpdateReason 🔒

Reasons why a Document might need a rendering update that is otherwise untracked via other Document properties.

SameOriginDescendantNavigablesIterator 🔒

Iterator for same origin descendant navigables in a shadow-including tree order, returning the active documents of the navigables. https://html.spec.whatwg.org/multipage/#descendant-navigables

SameoriginAncestorNavigablesIterator 🔒

Iterator for same origin ancestor navigables, returning the active documents of the navigables. https://html.spec.whatwg.org/multipage/#ancestor-navigables

Enums

AnimationFrameCallback 🔒

DeclarativeRefresh 🔒

Information about a declarative refresh

FireMouseEventType 🔒

HasBrowsingContext 🔒

IsHTMLDocument 🔒

TheEndLoadingPhase 🔒

Constants

_IMPL_DOMOBJECT_FOR_Document 🔒

Functions

get_registrable_domain_suffix_of_or_is_equal_to 🔒

is_named_element_with_id_attribute 🔒

is_named_element_with_name_attribute 🔒

update_with_current_instant 🔒

url_has_network_scheme 🔒

https://url.spec.whatwg.org/#network-scheme