script::dom::document::document

Struct Document

#[repr(C)]
pub(crate) struct Document {

    node: Node,
    document_or_shadow_root: DocumentOrShadowRoot,
    window: Dom<Window>,
    implementation: MutNullableDom<DOMImplementation>,
    content_type: Mime,
    last_modified: Option<String>,
    encoding: Cell<&'static Encoding>,
    has_browsing_context: bool,
    is_html_document: bool,
    activity: Cell<DocumentActivity>,
    url: DomRefCell<ServoUrl>,
    about_base_url: DomRefCell<Option<ServoUrl>>,
    quirks_mode: Cell<QuirksMode>,
    event_handler: DocumentEventHandler,
    focus_handler: DocumentFocusHandler,
    embedder_controls: DocumentEmbedderControls,
    id_map: TreeOrderedIndexMap,
    name_map: TreeOrderedIndexMap,
    tag_map: DomRefCell<HashMapTracedValues<LocalName, Dom<HTMLCollection>, FxBuildHasher>>,
    tagns_map: DomRefCell<HashMapTracedValues<QualName, Dom<HTMLCollection>, FxBuildHasher>>,
    classes_map: DomRefCell<HashMapTracedValues<Vec<Atom>, Dom<HTMLCollection>>>,
    images: MutNullableDom<HTMLCollection>,
    embeds: MutNullableDom<HTMLCollection>,
    links: MutNullableDom<HTMLCollection>,
    forms: MutNullableDom<HTMLCollection>,
    scripts: MutNullableDom<HTMLCollection>,
    anchors: MutNullableDom<HTMLCollection>,
    applets: MutNullableDom<HTMLCollection>,
    iframes: RefCell<IFrameCollection>,
    shared_style_locks: SharedRwLocks,
    stylesheets: DomRefCell<DocumentStylesheetSet<ServoStylesheetInDocument>>,
    stylesheet_list: MutNullableDom<StyleSheetList>,
    ready_state: Cell<DocumentReadyState>,
    current_script: MutNullableDom<HTMLScriptElement>,
    current_the_end_loading_phase: Cell<TheEndLoadingPhase>,
    pending_parsing_blocking_script: DomRefCell<Option<PendingScript>>,
    script_blocking_stylesheet_set: DomRefCell<IndexSet<StylesheetContextId>>,
    render_blocking_element_count: Cell<u32>,
    deferred_scripts: PendingInOrderScriptVec,
    asap_in_order_scripts_list: PendingInOrderScriptVec,
    asap_scripts_set: DomRefCell<Vec<Dom<HTMLScriptElement>>>,
    animation_frame_ident: Cell<u32>,
    animation_frame_list: DomRefCell<VecDeque<(u32, Option<AnimationFrameCallback>)>>,
    running_animation_callbacks: Cell<bool>,
    loader: DomRefCell<DocumentLoader>,
    current_parser: MutNullableDom<ServoParser>,
    base_element: MutNullableDom<HTMLBaseElement>,
    target_base_element: MutNullableDom<HTMLBaseElement>,
    appropriate_template_contents_owner_document: MutNullableDom<Document>,
    pending_restyles: DomRefCell<FxHashMap<Dom<Element>, NoTrace<PendingRestyle>>>,
    needs_restyle: Cell<RestyleReason>,
    origin: DomRefCell<MutableOrigin>,
    referrer: Option<String>,
    target_element: MutNullableDom<Element>,
    policy_container: DomRefCell<PolicyContainer>,
    preloaded_resources: DomRefCell<PreloadedResources>,
    ignore_destructive_writes_counter: Cell<u32>,
    ignore_opens_during_unload_counter: Cell<u32>,
    spurious_animation_frames: Cell<u8>,
    fullscreen_element: MutNullableDom<Element>,
    form_id_listener_map: DomRefCell<HashMapTracedValues<Atom, HashSet<Dom<Element>>, FxBuildHasher>>,
    interactive_time: DomRefCell<ProgressiveWebMetrics>,
    tti_window: DomRefCell<InteractiveWindow>,
    canceller: FetchCanceller,
    throw_on_dynamic_markup_insertion_counter: Cell<u64>,
    page_showing: Cell<bool>,
    salvageable: Cell<bool>,
    active_parser_was_aborted: Cell<bool>,
    fired_unload: Cell<bool>,
    responsive_images: DomRefCell<Vec<Dom<HTMLImageElement>>>,
    navigation_timing: Rc<NavigationTiming>,
    resource_fetch_timing: RefCell<Option<ResourceFetchTiming>>,
    script_and_layout_blockers: Cell<u32>,
    delayed_tasks: DomRefCell<Vec<Box<dyn NonSendTaskBox>>>,
    completely_loaded: Cell<bool>,
    shadow_roots: DomRefCell<HashSet<Dom<ShadowRoot>>>,
    shadow_roots_styles_changed: Cell<bool>,
    media_controls: DomRefCell<HashMap<String, Dom<ShadowRoot>>>,
    dirty_canvases: DomRefCell<Vec<Dom<HTMLCanvasElement>>>,
    has_pending_animated_image_update: Cell<bool>,
    selection: MutNullableDom<Selection>,
    timeline: Dom<DocumentTimeline>,
    animations: Animations,
    image_animation_manager: DomRefCell<ImageAnimationManager>,
    dirty_root: MutNullableDom<Element>,
    declarative_refresh: DomRefCell<Option<DeclarativeRefresh>>,
    resize_observers: DomRefCell<Vec<Dom<ResizeObserver>>>,
    fonts: MutNullableDom<FontFaceSet>,
    visibility_state: Cell<DocumentVisibilityState>,
    status_code: Option<u16>,
    is_initial_about_blank: Cell<bool>,
    allow_declarative_shadow_roots: Cell<bool>,
    inherited_insecure_requests_policy: Cell<Option<InsecureRequestsPolicy>>,
    has_trustworthy_ancestor_origin: Cell<bool>,
    intersection_observer_task_queued: Cell<bool>,
    intersection_observers: DomRefCell<Vec<Dom<IntersectionObserver>>>,
    highlighted_dom_node: MutNullableDom<Node>,
    lcp_candidates: DomRefCell<HashMapTracedValues<LCPCandidateID, Dom<Element>>>,
    adopted_stylesheets: DomRefCell<Vec<Dom<CSSStyleSheet>>>,
    adopted_stylesheets_frozen_types: CachedFrozenArray,
    pending_scroll_events: DomRefCell<Vec<PendingScrollEvent>>,
    rendering_update_reasons: Cell<RenderingUpdateReason>,
    waiting_on_canvas_image_updates: Cell<bool>,
    root_removal_noted: Cell<bool>,
    current_rendering_epoch: Cell<Epoch>,
    custom_element_reaction_stack: Rc<CustomElementReactionStack>,
    active_sandboxing_flag_set: Cell<SandboxingFlagSet>,
    creation_sandboxing_flag_set: Cell<SandboxingFlagSet>,
    favicon: RefCell<Option<Image>>,
    websockets: DOMTracker<WebSocket>,
    details_name_groups: DomRefCell<Option<DetailsNameGroups>>,
    protocol_handler_automation_mode: RefCell<CustomHandlersAutomationMode>,
    layout_animations_test_enabled: bool,
    state_override: DomRefCell<FxHashMap<CommandName, bool>>,
    value_override: DomRefCell<FxHashMap<CommandName, DOMString>>,
    default_single_line_container_name: Cell<DefaultSingleLineContainerName>,
    css_styling_flag: Cell<bool>,
    accessibility_data: DomRefCell<AccessibilityData>,
    iframe_load_in_progress: Cell<bool>,
    mute_iframe_load: Cell<bool>,
    timers: OneshotTimers,
    pipeline_id: PipelineId,
    task_manager: Rc<TaskManager>,
    image_cache: Arc<dyn ImageCache>,
    history: MutNullableDom<History>,
    theme: Cell<Option<Theme>>,
    window_detached: Cell<bool>,
    ancestor_origins_list: MutNullableDom<DOMStringList>,
    internal_ancestor_origin_objects_list: RefCell<Option<Vec<ImmutableOrigin>>>,

}

https://dom.spec.whatwg.org/#document

Fields
node: Nodedocument_or_shadow_root: DocumentOrShadowRootwindow: Dom<Window>implementation: MutNullableDom<DOMImplementation>content_type: Mimelast_modified: Option<String>encoding: Cell<&'static Encoding>has_browsing_context: boolis_html_document: boolactivity: Cell<DocumentActivity>url: DomRefCell<ServoUrl>

https://html.spec.whatwg.org/multipage/#the-document%27s-address

about_base_url: DomRefCell<Option<ServoUrl>>

https://html.spec.whatwg.org/multipage/#concept-document-about-base-url

quirks_mode: Cell<QuirksMode>event_handler: DocumentEventHandler

A helper used to process and store data related to input event handling.

focus_handler: DocumentFocusHandler

A helper used to process and store data related to focus handling.

embedder_controls: DocumentEmbedderControls

A helper to handle showing and hiding user interface controls in the embedding layer.

id_map: TreeOrderedIndexMapname_map: TreeOrderedIndexMaptag_map: DomRefCell<HashMapTracedValues<LocalName, Dom<HTMLCollection>, FxBuildHasher>>tagns_map: DomRefCell<HashMapTracedValues<QualName, Dom<HTMLCollection>, FxBuildHasher>>classes_map: DomRefCell<HashMapTracedValues<Vec<Atom>, Dom<HTMLCollection>>>images: MutNullableDom<HTMLCollection>embeds: MutNullableDom<HTMLCollection>links: MutNullableDom<HTMLCollection>forms: MutNullableDom<HTMLCollection>scripts: MutNullableDom<HTMLCollection>anchors: MutNullableDom<HTMLCollection>applets: MutNullableDom<HTMLCollection>iframes: RefCell<IFrameCollection>

Information about the <iframes> in this Document.

shared_style_locks: SharedRwLocks

Shared locks used for style attributes, author-origin stylesheets, and user and user agent stylesheets in this document. Can be acquired once for accessing many objects. This is shared with the owning ScriptThread.

stylesheets: DomRefCell<DocumentStylesheetSet<ServoStylesheetInDocument>>

List of stylesheets associated with nodes in this document. |None| if the list needs to be refreshed.

stylesheet_list: MutNullableDom<StyleSheetList>ready_state: Cell<DocumentReadyState>current_script: MutNullableDom<HTMLScriptElement>

The script element that is currently executing.

current_the_end_loading_phase: Cell<TheEndLoadingPhase>pending_parsing_blocking_script: DomRefCell<Option<PendingScript>>

https://html.spec.whatwg.org/multipage/#pending-parsing-blocking-script

script_blocking_stylesheet_set: DomRefCell<IndexSet<StylesheetContextId>>

https://html.spec.whatwg.org/multipage/#script-blocking-style-sheet-set

A Document has a script-blocking style sheet set, which is an ordered set, initially empty.

render_blocking_element_count: Cell<u32>

Number of elements that block the rendering of the page. https://html.spec.whatwg.org/multipage/#implicitly-potentially-render-blocking

deferred_scripts: PendingInOrderScriptVec

https://html.spec.whatwg.org/multipage/#list-of-scripts-that-will-execute-when-the-document-has-finished-parsing

asap_in_order_scripts_list: PendingInOrderScriptVec

https://html.spec.whatwg.org/multipage/#list-of-scripts-that-will-execute-in-order-as-soon-as-possible

asap_scripts_set: DomRefCell<Vec<Dom<HTMLScriptElement>>>

https://html.spec.whatwg.org/multipage/#set-of-scripts-that-will-execute-as-soon-as-possible

animation_frame_ident: Cell<u32>

https://html.spec.whatwg.org/multipage/#animation-frame-callback-identifier Current identifier of animation frame callback

animation_frame_list: DomRefCell<VecDeque<(u32, Option<AnimationFrameCallback>)>>

https://html.spec.whatwg.org/multipage/#list-of-animation-frame-callbacks List of animation frame callbacks

running_animation_callbacks: Cell<bool>

Whether we’re in the process of running animation callbacks.

Tracking this is not necessary for correctness. Instead, it is an optimization to avoid sending needless ChangeRunningAnimationsState messages to Paint.

loader: DomRefCell<DocumentLoader>

Tracks all outstanding loads related to this document.

current_parser: MutNullableDom<ServoParser>

The current active HTML parser, to allow resuming after interruptions.

base_element: MutNullableDom<HTMLBaseElement>

The cached first base element with an href attribute.

target_base_element: MutNullableDom<HTMLBaseElement>

The cached first base element, used for its target (doesn’t need a href)

appropriate_template_contents_owner_document: MutNullableDom<Document>

This field is set to the document itself for inert documents. https://html.spec.whatwg.org/multipage/#appropriate-template-contents-owner-document

pending_restyles: DomRefCell<FxHashMap<Dom<Element>, NoTrace<PendingRestyle>>>

Information on elements needing restyle to ship over to layout when the time comes.

needs_restyle: Cell<RestyleReason>

A collection of reasons that the Document needs to be restyled at the next opportunity for a reflow. If this is empty, then the Document does not need to be restyled.

origin: DomRefCell<MutableOrigin>

The document’s origin.

referrer: Option<String>

https://html.spec.whatwg.org/multipage/#dom-document-referrer

target_element: MutNullableDom<Element>

https://html.spec.whatwg.org/multipage/#target-element

policy_container: DomRefCell<PolicyContainer>

https://html.spec.whatwg.org/multipage/#concept-document-policy-container

preloaded_resources: DomRefCell<PreloadedResources>

https://html.spec.whatwg.org/multipage/#map-of-preloaded-resources

ignore_destructive_writes_counter: Cell<u32>

https://html.spec.whatwg.org/multipage/#ignore-destructive-writes-counter

ignore_opens_during_unload_counter: Cell<u32>

https://html.spec.whatwg.org/multipage/#ignore-opens-during-unload-counter

spurious_animation_frames: Cell<u8>

The number of spurious requestAnimationFrame() requests we’ve received.

A rAF request is considered spurious if nothing was actually reflowed.

fullscreen_element: MutNullableDom<Element>

Entry node for fullscreen.

form_id_listener_map: DomRefCell<HashMapTracedValues<Atom, HashSet<Dom<Element>>, FxBuildHasher>>

Map from ID to set of form control elements that have that ID as their ‘form’ content attribute. Used to reset form controls whenever any element with the same ID as the form attribute is inserted or removed from the document. See https://html.spec.whatwg.org/multipage/#form-owner It is safe to use FxBuildHasher here as Atoms are in the string_cache

interactive_time: DomRefCell<ProgressiveWebMetrics>tti_window: DomRefCell<InteractiveWindow>canceller: FetchCanceller

RAII canceller for Fetch

throw_on_dynamic_markup_insertion_counter: Cell<u64>

https://html.spec.whatwg.org/multipage/#throw-on-dynamic-markup-insertion-counter

page_showing: Cell<bool>

https://html.spec.whatwg.org/multipage/#page-showing

salvageable: Cell<bool>

Whether the document is salvageable.

active_parser_was_aborted: Cell<bool>

Whether the document was aborted with an active parser

fired_unload: Cell<bool>

Whether the unload event has already been fired.

responsive_images: DomRefCell<Vec<Dom<HTMLImageElement>>>

List of responsive images

navigation_timing: Rc<NavigationTiming>

NavigationTiming information for this Document. https://html.spec.whatwg.org/multipage/#load-timing-info

resource_fetch_timing: RefCell<Option<ResourceFetchTiming>>

A ResourceFetchTiming that holds timing information for this Document.

script_and_layout_blockers: Cell<u32>

Number of outstanding requests to prevent JS or layout from running.

delayed_tasks: DomRefCell<Vec<Box<dyn NonSendTaskBox>>>

List of tasks to execute as soon as last script/layout blocker is removed.

completely_loaded: Cell<bool>

https://html.spec.whatwg.org/multipage/#completely-loaded

shadow_roots: DomRefCell<HashSet<Dom<ShadowRoot>>>

Set of shadow roots connected to the document tree.

shadow_roots_styles_changed: Cell<bool>

Whether any of the shadow roots need the stylesheets flushed.

media_controls: DomRefCell<HashMap<String, Dom<ShadowRoot>>>

List of registered media controls. We need to keep this list to allow the media controls to access the “privileged” document.servoGetMediaControls(id) API, where id needs to match any of the registered ShadowRoots hosting the media controls UI.

dirty_canvases: DomRefCell<Vec<Dom<HTMLCanvasElement>>>

A set of dirty HTML canvas elements that need their WebRender images updated the next time the rendering is updated.

has_pending_animated_image_update: Cell<bool>

Whether or not animated images need to have their contents updated.

selection: MutNullableDom<Selection>

https://w3c.github.io/slection-api/#dfn-selection

timeline: Dom<DocumentTimeline>

A timeline for animations which is used for synchronizing animations. https://drafts.csswg.org/web-animations/#timeline

animations: Animations

Animations for this Document

image_animation_manager: DomRefCell<ImageAnimationManager>

Image Animation Manager for this Document

dirty_root: MutNullableDom<Element>

The nearest inclusive ancestors to all the nodes that require a restyle.

declarative_refresh: DomRefCell<Option<DeclarativeRefresh>>

https://html.spec.whatwg.org/multipage/#will-declaratively-refresh

resize_observers: DomRefCell<Vec<Dom<ResizeObserver>>>

https://drafts.csswg.org/resize-observer/#dom-document-resizeobservers-slot

Note: we are storing, but never removing, resize observers. The lifetime of resize observers is specified at https://drafts.csswg.org/resize-observer/#lifetime. But implementing it comes with known problems:

https://bugzilla.mozilla.org/show_bug.cgi?id=1596992

https://github.com/w3c/csswg-drafts/issues/4518

fonts: MutNullableDom<FontFaceSet>

The set of all fonts loaded by this document. https://drafts.csswg.org/css-font-loading/#font-face-source

visibility_state: Cell<DocumentVisibilityState>

https://html.spec.whatwg.org/multipage/#visibility-state

status_code: Option<u16>

https://www.iana.org/assignments/http-status-codes/http-status-codes.xhtml

is_initial_about_blank: Cell<bool>

https://html.spec.whatwg.org/multipage/#is-initial-about:blank

allow_declarative_shadow_roots: Cell<bool>

https://dom.spec.whatwg.org/#document-allow-declarative-shadow-roots

inherited_insecure_requests_policy: Cell<Option<InsecureRequestsPolicy>>

https://w3c.github.io/webappsec-upgrade-insecure-requests/#insecure-requests-policy

has_trustworthy_ancestor_origin: Cell<bool>intersection_observer_task_queued: Cell<bool>

https://w3c.github.io/IntersectionObserver/#document-intersectionobservertaskqueued

intersection_observers: DomRefCell<Vec<Dom<IntersectionObserver>>>

Active intersection observers that should be processed by this document in the update intersection observation steps. https://w3c.github.io/IntersectionObserver/#run-the-update-intersection-observations-steps

Let observer list be a list of all IntersectionObservers whose root is in the DOM tree of document. For the top-level browsing context, this includes implicit root observers.

Details of which document that should process an observers is discussed further at https://github.com/w3c/IntersectionObserver/issues/525.

The lifetime of an intersection observer is specified at https://github.com/w3c/IntersectionObserver/issues/525.

highlighted_dom_node: MutNullableDom<Node>

The node that is currently highlighted by the devtools

lcp_candidates: DomRefCell<HashMapTracedValues<LCPCandidateID, Dom<Element>>>

Resolved LCP candidate elements, keyed by their LCPCandidateID.

adopted_stylesheets: DomRefCell<Vec<Dom<CSSStyleSheet>>>

The constructed stylesheet that is adopted by this Document. https://drafts.csswg.org/cssom/#dom-documentorshadowroot-adoptedstylesheets

adopted_stylesheets_frozen_types: CachedFrozenArray

Cached frozen array of Self::adopted_stylesheets

pending_scroll_events: DomRefCell<Vec<PendingScrollEvent>>

https://drafts.csswg.org/cssom-view/#document-pending-scroll-events

Each Document has an associated list of pending scroll events, which stores pairs of (EventTarget, DOMString), initially empty.

rendering_update_reasons: Cell<RenderingUpdateReason>

Other reasons that a rendering update might be required for this Document.

waiting_on_canvas_image_updates: Cell<bool>

Whether or not this Document is waiting on canvas image updates. If it is waiting it will not do any new layout until the canvas images are up-to-date in the renderer.

root_removal_noted: Cell<bool>

Whether we have already noted that the document element was removed.

current_rendering_epoch: Cell<Epoch>

The current rendering epoch, which is used to track updates in the renderer.

Every display list update also advances the Epoch, so that the renderer knows when a particular display list is ready in order to take a screenshot.

Canvas image updates happen asynchronously and are tagged with this Epoch. Until those asynchronous updates are complete, the Document will not perform any more rendering updates.

custom_element_reaction_stack: Rc<CustomElementReactionStack>

The global custom element reaction stack for this script thread.

active_sandboxing_flag_set: Cell<SandboxingFlagSet>

https://html.spec.whatwg.org/multipage/#active-sandboxing-flag-set,

creation_sandboxing_flag_set: Cell<SandboxingFlagSet>

The SandboxingFlagSet use to create the browsing context for this Document. These are cached here as they cannot always be retrieved readily if the owner of browsing context (either <iframe> or popup) might be in a different ScriptThread.

See https://html.spec.whatwg.org/multipage/#determining-the-creation-sandboxing-flags.

favicon: RefCell<Option<Image>>

The cached favicon for that document.

websockets: DOMTracker<WebSocket>

All websockets created that are associated with this document.

details_name_groups: DomRefCell<Option<DetailsNameGroups>>

https://html.spec.whatwg.org/multipage/#details-name-group

protocol_handler_automation_mode: RefCell<CustomHandlersAutomationMode>

https://html.spec.whatwg.org/multipage/#registerprotocolhandler()-automation-mode

layout_animations_test_enabled: bool

Reflect the value of that preferences to prevent paying the cost of a RwLock access.

state_override: DomRefCell<FxHashMap<CommandName, bool>>

https://w3c.github.io/editing/docs/execCommand/#state-override

value_override: DomRefCell<FxHashMap<CommandName, DOMString>>

https://w3c.github.io/editing/docs/execCommand/#value-override

default_single_line_container_name: Cell<DefaultSingleLineContainerName>

https://w3c.github.io/editing/docs/execCommand/#default-single-line-container-name

css_styling_flag: Cell<bool>

https://w3c.github.io/editing/docs/execCommand/#css-styling-flag

accessibility_data: DomRefCell<AccessibilityData>

Data necessary for maintaining the accessibility tree.

iframe_load_in_progress: Cell<bool>

https://html.spec.whatwg.org/multipage/#iframe-load-in-progress

mute_iframe_load: Cell<bool>

https://html.spec.whatwg.org/multipage/#mute-iframe-load

timers: OneshotTimers

The mechanism by which time-outs and intervals are scheduled. https://html.spec.whatwg.org/multipage/#timers

pipeline_id: PipelineIdtask_manager: Rc<TaskManager>

A TaskManager for this Window.

image_cache: Arc<dyn ImageCache>history: MutNullableDom<History>

https://html.spec.whatwg.org/multipage/#doc-history

theme: Cell<Option<Theme>>

Theme specific for this document, set by a meta element

window_detached: Cell<bool>

True if this document is no longer the active document of its associated window.

ancestor_origins_list: MutNullableDom<DOMStringList>

https://html.spec.whatwg.org/multipage/#concept-document-ancestor-origins-list

internal_ancestor_origin_objects_list: RefCell<Option<Vec<ImmutableOrigin>>>

https://html.spec.whatwg.org/multipage/#concept-document-internal-ancestor-origin-objects-list

Implementations

impl Document

fn __assert_parent_type(&self)

impl Document

pub(crate) fn history(&self, cx: &mut JSContext) -> DomRoot<History>

pub(crate) fn image_cache(&self) -> StdArc<dyn ImageCache>

pub(crate) fn task_manager(&self) -> Rc<TaskManager>

pub(crate) fn timers(&self) -> &OneshotTimers

pub(crate) fn pipeline_id(&self) -> PipelineId

fn fully_exit_fullscreen(&self, cx: &mut JSContext)

https://fullscreen.spec.whatwg.org/#fully-exit-fullscreen

fn unloading_cleanup_steps(&self, cx: &mut JSContext)

https://html.spec.whatwg.org/multipage/#unloading-document-cleanup-steps

pub(crate) fn track_websocket(&self, websocket: &WebSocket)

fn close_outstanding_websockets(&self) -> bool

fn document_element_changed(&self)

pub(crate) fn note_dirty_element(&self, no_gc: &NoGC, element: &Element)

This is a port of Gecko’s restyle root architecture. The idea is that we track a node which is the root of restyle damage. Below that root, certain nodes can be marked with a HAS_DIRTY_DESCENDANTS flag which means they should be traversed during styling and damage propagation. Above the dirty root nothing should be marked with the HAS_DIRTY_DESCENDANTS flag.

The overall algorithm is as follows:

When the first dirty element is noted, we just set it as the restyle root.

When additional dirty elements are noted, we propagate the given bit up the tree, until we either reach the dirty root or the document root.

If we reach the document root, we then propagate the HAS_DIRTY_DESCENDANTS flags up the tree until we cross the path of the new root. Once we find this common ancestor, we record it as the restyle root, and then clear the bits between the new restyle root and the document root.

fn set_dirty_root(&self, no_gc: &NoGC, new_dirty_root: Option<&Element>)

pub(crate) fn take_dirty_root(&self) -> Option<DomRoot<Element>>

pub(crate) fn loader(&self) -> Ref<'_, DocumentLoader>

pub(crate) fn loader_mut(&self) -> RefMut<'_, DocumentLoader>

pub(crate) fn has_browsing_context(&self) -> bool

pub(crate) fn browsing_context(&self) -> Option<DomRoot<WindowProxy>>

https://html.spec.whatwg.org/multipage/#concept-document-bc

pub(crate) fn webview_id(&self) -> WebViewId

pub(crate) fn window(&self) -> &Window

pub(crate) fn is_html_document(&self) -> bool

pub(crate) fn is_xhtml_document(&self) -> bool

pub(crate) fn is_fully_active(&self) -> bool

https://html.spec.whatwg.org/multipage/#fully-active

pub(crate) fn is_active(&self) -> bool

https://html.spec.whatwg.org/multipage/#nav-document

pub(crate) fn current_rendering_epoch(&self) -> Epoch

pub(crate) fn selection(&self) -> Option<DomRoot<Selection>>

Get the Selection instance for this Document if there is one or None.

pub(crate) fn set_activity( &self, cx: &mut JSContext, activity: DocumentActivity, )

pub(crate) fn origin(&self) -> Ref<'_, MutableOrigin>

pub(crate) fn mark_as_internal(&self)

Part of https://html.spec.whatwg.org/multipage/#navigate-ua-inline TODO: Remove this when we create documents after processing headers

pub(crate) fn set_protocol_handler_automation_mode( &self, mode: CustomHandlersAutomationMode, )

pub(crate) fn url(&self) -> ServoUrl

https://dom.spec.whatwg.org/#concept-document-url

pub(crate) fn set_url(&self, url: ServoUrl)

pub(crate) fn about_base_url(&self) -> Option<ServoUrl>

pub(crate) fn set_about_base_url(&self, about_base_url: Option<ServoUrl>)

pub(crate) fn fallback_base_url(&self) -> ServoUrl

https://html.spec.whatwg.org/multipage/#fallback-base-url

pub(crate) fn base_url(&self) -> ServoUrl

https://html.spec.whatwg.org/multipage/#document-base-url

pub(crate) fn add_restyle_reason(&self, reason: RestyleReason)

pub(crate) fn clear_restyle_reasons(&self)

pub(crate) fn stylesheets_changed_since_last_reflow(&self) -> bool

pub(crate) fn restyle_reason(&self, no_gc: &NoGC) -> RestyleReason

pub(crate) fn base_element(&self) -> Option<DomRoot<HTMLBaseElement>>

Returns the first base element in the DOM that has an href attribute.

pub(crate) fn target_base_element(&self) -> Option<DomRoot<HTMLBaseElement>>

Returns the first base element in the DOM (doesn’t need to have an href attribute).

pub(crate) fn refresh_base_element(&self, cx: &mut JSContext)

Refresh the cached first base element in the DOM.

pub(crate) fn quirks_mode(&self) -> QuirksMode

pub(crate) fn set_quirks_mode(&self, new_mode: QuirksMode)

pub(crate) fn encoding(&self) -> &'static Encoding

pub(crate) fn set_encoding(&self, encoding: &'static Encoding)

pub(crate) fn content_and_heritage_changed(&self, no_gc: &NoGC, node: &Node)

pub(crate) fn unregister_element_id(&self, cx: &mut JSContext, id: &Atom)

Remove any existing association between the provided id and any elements in this document.

pub(crate) fn register_element_id( &self, cx: &mut JSContext, element: &Element, id: &Atom, )

Associate an element present in this document with the provided id.

pub(crate) fn unregister_element_name(&self, name: &Atom)

Remove any existing association between the provided name and any elements in this document.

pub(crate) fn register_element_name(&self, element: &Element, name: &Atom)

Associate an element present in this document with the provided name.

pub(crate) fn register_form_id_listener<T: ?Sized + FormControl>( &self, id: DOMString, listener: &T, )

pub(crate) fn unregister_form_id_listener<T: ?Sized + FormControl>( &self, id: DOMString, listener: &T, )

fn find_a_potential_indicated_element( &self, cx: &mut JSContext, fragment: &str, ) -> Option<DomRoot<Element>>

https://html.spec.whatwg.org/multipage/#find-a-potential-indicated-element

fn select_indicated_part( &self, cx: &mut JSContext, fragment: &str, ) -> Option<DomRoot<Node>>

Attempt to find a named element in this page’s document. https://html.spec.whatwg.org/multipage/#the-indicated-part-of-the-document

pub(crate) fn scroll_to_the_fragment(&self, cx: &mut JSContext, fragment: &str)

https://html.spec.whatwg.org/multipage/#scroll-to-the-fragment-identifier

fn get_anchor_by_name( &self, cx: &mut JSContext, name: &str, ) -> Option<DomRoot<Element>>

pub(crate) fn notify_embedder_of_load_completion(&self)

pub(crate) fn set_document_readiness_to_loading_for_initialization(&self)

Set the readyState of this Document to loading for the purposes of the “initialize a document object” part of the specification.

See https://html.spec.whatwg.org/multipage/#initialise-the-document-object.

pub(crate) fn update_the_current_document_readiness( &self, cx: &mut JSContext, state: DocumentReadyState, )

https://html.spec.whatwg.org/multipage/#update-the-current-document-readiness

pub(crate) fn scripting_enabled(&self) -> bool

Return whether scripting is enabled or not https://html.spec.whatwg.org/multipage/#concept-n-script

pub(crate) fn title_changed(&self)

Handles any updates when the document’s title has changed.

fn title(&self) -> Option<DOMString>

Determine the title of the Document according to the specification at: https://html.spec.whatwg.org/multipage/#document.title. The difference here is that when the title isn’t specified None is returned.

pub(crate) fn send_title_to_embedder(&self)

Sends this document’s title to the constellation.

pub(crate) fn send_to_embedder(&self, msg: EmbedderMsg)

pub(crate) fn dirty_all_nodes(&self, no_gc: &NoGC)

pub(crate) fn run_the_scroll_steps(&self, cx: &mut JSContext)

https://drafts.csswg.org/cssom-view/#document-run-the-scroll-steps

pub(crate) fn handle_viewport_scroll_event(&self)

https://drafts.csswg.org/cssom-view/#scrolling-events

Whenever a viewport gets scrolled (whether in response to user interaction or by an API), the user agent must run these steps:

pub(crate) fn finish_handle_scroll_event(&self, event_target: &EventTarget)

https://drafts.csswg.org/cssom-view/#scrolling-events

These are the shared steps 3 and 4 from all scroll targets listed in the first section of the specification.

pub(crate) fn node_from_nodes_and_strings( &self, cx: &mut JSContext, nodes: Vec<NodeOrString<DomTypeHolder>>, ) -> Fallible<DomRoot<Node>>

pub(crate) fn get_body_attribute(&self, local_name: &LocalName) -> DOMString

pub(crate) fn set_body_attribute( &self, cx: &mut JSContext, local_name: &LocalName, value: DOMString, )

pub(crate) fn set_current_script(&self, script: Option<&HTMLScriptElement>)

pub(crate) fn has_a_stylesheet_that_is_blocking_scripts(&self) -> bool

https://html.spec.whatwg.org/multipage/#has-a-style-sheet-that-is-blocking-scripts

pub(crate) fn add_script_blocking_stylesheet(&self, id: StylesheetContextId)

pub(crate) fn remove_script_blocking_stylesheet(&self, id: StylesheetContextId)

pub(crate) fn render_blocking_element_count(&self) -> u32

pub(crate) fn increment_render_blocking_element_count(&self)

https://html.spec.whatwg.org/multipage/#block-rendering

pub(crate) fn decrement_render_blocking_element_count(&self)

https://html.spec.whatwg.org/multipage/#unblock-rendering

pub(crate) fn allows_adding_render_blocking_elements(&self) -> bool

https://html.spec.whatwg.org/multipage/#allows-adding-render-blocking-elements

pub(crate) fn is_render_blocked(&self) -> bool

https://html.spec.whatwg.org/multipage/#render-blocked

pub(crate) fn invalidate_stylesheets(&self, no_gc: &NoGC)

pub(crate) fn has_active_request_animation_frame_callbacks(&self) -> bool

Whether or not this Document has any active requestAnimationFrame callbacks registered.

pub(crate) fn request_animation_frame( &self, callback: AnimationFrameCallback, ) -> u32

https://html.spec.whatwg.org/multipage/#dom-window-requestanimationframe

pub(crate) fn cancel_animation_frame(&self, ident: u32)

https://html.spec.whatwg.org/multipage/#dom-window-cancelanimationframe

pub(crate) fn run_the_animation_frame_callbacks( &self, cx: &mut CurrentRealm<'_>, )

https://html.spec.whatwg.org/multipage/#run-the-animation-frame-callbacks

pub(crate) fn policy_container(&self) -> Ref<'_, PolicyContainer>

pub(crate) fn set_policy_container(&self, policy_container: PolicyContainer)

pub(crate) fn set_csp_list(&self, csp_list: Option<CspList>)

pub(crate) fn enforce_csp_policy(&self, policy: CspPolicy)

https://www.w3.org/TR/CSP/#enforced

pub(crate) fn get_csp_list(&self) -> Ref<'_, Option<CspList>>

pub(crate) fn preloaded_resources(&self) -> Ref<'_, PreloadedResources>

pub(crate) fn insert_preloaded_resource( &self, key: PreloadKey, preload_id: PreloadId, )

pub(crate) fn fetch_blocking<Listener: FetchResponseListener>( &self, load: LoadType, request: RequestBuilder, listener: Listener, )

pub(crate) fn fetch_background<Listener: FetchResponseListener>( &self, request_builder: RequestBuilder, listener: Listener, )

fn deferred_fetch_control_document(&self) -> DomRoot<Document>

https://fetch.spec.whatwg.org/#deferred-fetch-control-document

pub(crate) fn available_deferred_fetch_quota( &self, origin: ImmutableOrigin, ) -> isize

https://fetch.spec.whatwg.org/#available-deferred-fetch-quota

pub(crate) fn update_document_for_history_step_application( &self, old_url: &ServoUrl, new_url: &ServoUrl, )

https://html.spec.whatwg.org/multipage/#update-document-for-history-step-application

pub(crate) fn finish_load_for_dropped_blocker(&self, load: LoadType)

pub(crate) fn finish_load(&self, load: LoadType, cx: &mut JSContext)

Step 8 of https://html.spec.whatwg.org/multipage/#the-end https://html.spec.whatwg.org/multipage/#delay-the-load-event

pub(crate) fn check_if_unloading_is_cancelled( &self, cx: &mut JSContext, recursive_flag: bool, ) -> bool

https://html.spec.whatwg.org/multipage/#checking-if-unloading-is-canceled

pub(crate) fn unload(&self, cx: &mut JSContext, recursive_flag: bool)

fn completely_finish_loading(&self)

https://html.spec.whatwg.org/multipage/#completely-finish-loading

fn queue_document_completion(&self, cx: &mut JSContext)

Step 9 of https://html.spec.whatwg.org/multipage/#the-end

pub(crate) fn completely_loaded(&self) -> bool

pub(crate) fn start_the_end_loading_phase(&self)

pub(crate) fn set_pending_parsing_blocking_script( &self, script: &HTMLScriptElement, load: Option<Result<Script, ()>>, )

pub(crate) fn has_pending_parsing_blocking_script(&self) -> bool

pub(crate) fn pending_parsing_blocking_script_loaded( &self, element: &HTMLScriptElement, result: Result<Script, ()>, cx: &mut JSContext, )

https://html.spec.whatwg.org/multipage/#prepare-a-script step 22.d.

fn process_pending_parsing_blocking_script(&self, cx: &mut JSContext)

pub(crate) fn add_asap_script(&self, script: &HTMLScriptElement)

pub(crate) fn asap_script_loaded( &self, cx: &mut JSContext, element: &HTMLScriptElement, result: Result<Script, ()>, )

https://html.spec.whatwg.org/multipage/#the-end step 5. https://html.spec.whatwg.org/multipage/#prepare-a-script step 22.d.

pub(crate) fn push_asap_in_order_script(&self, script: &HTMLScriptElement)

pub(crate) fn asap_in_order_script_loaded( &self, cx: &mut JSContext, element: &HTMLScriptElement, result: Result<Script, ()>, )

https://html.spec.whatwg.org/multipage/#the-end step 5. https://html.spec.whatwg.org/multipage/#prepare-a-script step> 22.c.

pub(crate) fn add_deferred_script(&self, script: &HTMLScriptElement)

https://html.spec.whatwg.org/multipage/#list-of-scripts-that-will-execute-when-the-document-has-finished-parsing

pub(crate) fn deferred_script_loaded( &self, cx: &mut JSContext, element: &HTMLScriptElement, result: Result<Script, ()>, )

https://html.spec.whatwg.org/multipage/#the-end step 3. https://html.spec.whatwg.org/multipage/#prepare-a-script step 22.d.

fn process_deferred_scripts(&self, cx: &mut JSContext)

Step 5 of https://html.spec.whatwg.org/multipage/#the-end

fn dispatch_dom_content_loaded(&self)

Step 6 of https://html.spec.whatwg.org/multipage/#the-end

fn has_finished_all_asap_scripts(&self) -> bool

fn wait_until_asap_scripts_have_executed(&self)

Step 7 of https://html.spec.whatwg.org/multipage/#the-end

pub(crate) fn wait_until_load_blockers_have_resolved(&self, cx: &mut JSContext)

Step 8 of https://html.spec.whatwg.org/multipage/#the-end

pub(crate) fn destroy_document_and_its_descendants(&self, cx: &mut JSContext)

https://html.spec.whatwg.org/multipage/#destroy-a-document-and-its-descendants

pub(crate) fn destroy(&self, cx: &mut JSContext)

https://html.spec.whatwg.org/multipage/#destroy-a-document

fn active_parser(&self) -> Option<DomRoot<ServoParser>>

https://html.spec.whatwg.org/multipage/#active-parser

pub(crate) fn abort(&self, cx: &mut JSContext)

https://html.spec.whatwg.org/multipage/#abort-a-document

pub(crate) fn abort_a_document_and_its_descendants(&self, cx: &mut JSContext)

https://html.spec.whatwg.org/multipage/#abort-a-document-and-its-descendants

pub(crate) fn notify_constellation_load(&self)

pub(crate) fn set_current_parser(&self, script: Option<&ServoParser>)

pub(crate) fn get_current_parser(&self) -> Option<DomRoot<ServoParser>>

pub(crate) fn get_current_parser_line(&self) -> u32

pub(crate) fn set_ancestor_origins_list( &self, ancestor_origins_list: &DOMStringList, )

https://html.spec.whatwg.org/multipage/#concept-document-ancestor-origins-list

pub(crate) fn ancestor_origins_list(&self) -> Option<DomRoot<DOMStringList>>

https://html.spec.whatwg.org/multipage/#concept-document-ancestor-origins-list

pub(crate) fn set_internal_ancestor_origin_objects_list( &self, internal_ancestor_origin_objects_list: Vec<ImmutableOrigin>, )

https://html.spec.whatwg.org/multipage/#concept-document-internal-ancestor-origin-objects-list

pub(crate) fn internal_ancestor_origin_objects_list( &self, ) -> Ref<'_, Option<Vec<ImmutableOrigin>>>

https://html.spec.whatwg.org/multipage/#concept-document-internal-ancestor-origin-objects-list

pub(crate) fn iframes(&self) -> Ref<'_, IFrameCollection>

A reference to the IFrameCollection of this Document, holding information about <iframe>s found within it.

pub(crate) fn iframes_mut(&self) -> RefMut<'_, IFrameCollection>

A mutable reference to the IFrameCollection of this Document, holding information about <iframe>s found within it.

pub(crate) fn set_navigation_start(&self, navigation_start: CrossProcessInstant)

pub(crate) fn get_interactive_metrics(&self) -> Ref<'_, ProgressiveWebMetrics>

pub(crate) fn has_recorded_tti_metric(&self) -> bool

pub(crate) fn start_tti(&self)

pub(crate) fn record_tti_if_necessary(&self)

check tti for this document if it’s been 10s since this doc encountered a task over 50ms, then we consider the main thread available and try to set tti

pub(crate) fn is_cookie_averse(&self) -> bool

https://html.spec.whatwg.org/multipage/#cookie-averse-document-object

pub(crate) fn custom_element_registry( &self, ) -> Option<DomRoot<CustomElementRegistry>>

pub(crate) fn set_custom_element_registry( &self, registry: &CustomElementRegistry, )

pub(crate) fn effective_global_custom_element_registry( &self, ) -> Option<DomRoot<CustomElementRegistry>>

https://dom.spec.whatwg.org/#effective-global-custom-element-registry

pub(crate) fn teardown_custom_element_registry(&self)

Cleans up any active promises https://github.com/servo/servo/issues/15318

pub(crate) fn increment_throw_on_dynamic_markup_insertion_counter(&self)

pub(crate) fn decrement_throw_on_dynamic_markup_insertion_counter(&self)

pub(crate) fn react_to_environment_changes(&self, cx: &JSContext)

pub(crate) fn register_responsive_image(&self, img: &HTMLImageElement)

pub(crate) fn unregister_responsive_image(&self, img: &HTMLImageElement)

pub(crate) fn register_media_controls(&self, id: &str, controls: &ShadowRoot)

pub(crate) fn unregister_media_controls(&self, id: &str)

pub(crate) fn mark_canvas_as_dirty(&self, canvas: &Dom<HTMLCanvasElement>)

pub(crate) fn needs_rendering_update(&self, no_gc: &NoGC) -> bool

Whether or not this Document needs a rendering update, due to changed contents or pending events. This is used to decide whether or not to schedule a call to the “update the rendering” algorithm.

pub(crate) fn update_the_rendering( &self, cx: &mut JSContext, ) -> (ReflowPhasesRun, ReflowStatistics)

An implementation of step 22 from https://html.spec.whatwg.org/multipage/#update-the-rendering:

pub(crate) fn handle_no_longer_waiting_on_asynchronous_image_updates(&self)

pub(crate) fn waiting_on_canvas_image_updates(&self) -> bool

pub(crate) fn maybe_fulfill_font_ready_promise( &self, cx: &mut JSContext, ) -> bool

From https://drafts.csswg.org/css-font-loading/#fontfaceset-pending-on-the-environment:

A FontFaceSet is pending on the environment if any of the following are true:

the document is still loading

the document has pending stylesheet requests

the document has pending layout operations which might cause the user agent to request a font, or which depend on recently-loaded fonts

Returns true if the promise was fulfilled.

pub(crate) fn id_map(&self) -> &TreeOrderedIndexMap

pub(crate) fn add_resize_observer(&self, resize_observer: &ResizeObserver)

https://drafts.csswg.org/resize-observer/#dom-resizeobserver-resizeobserver

pub(crate) fn gather_active_resize_observations_at_depth( &self, no_gc: &NoGC, depth: &ResizeObservationDepth, ) -> bool

https://drafts.csswg.org/resize-observer/#gather-active-observations-h https://drafts.csswg.org/resize-observer/#has-active-resize-observations

pub(crate) fn broadcast_active_resize_observations( &self, cx: &mut JSContext, ) -> ResizeObservationDepth

https://drafts.csswg.org/resize-observer/#broadcast-active-resize-observations

pub(crate) fn has_skipped_resize_observations(&self) -> bool

https://drafts.csswg.org/resize-observer/#has-skipped-observations-h

pub(crate) fn deliver_resize_loop_error_notification(&self, cx: &mut JSContext)

https://drafts.csswg.org/resize-observer/#deliver-resize-loop-error-notification

pub(crate) fn status_code(&self) -> Option<u16>

pub(crate) fn encoding_parse_a_url( &self, url: &str, ) -> Result<ServoUrl, ParseError>

https://html.spec.whatwg.org/multipage/#encoding-parsing-a-url

pub(crate) fn allowed_to_use_feature(&self, _feature: PermissionName) -> bool

https://html.spec.whatwg.org/multipage/#allowed-to-use

pub(crate) fn add_intersection_observer( &self, intersection_observer: &IntersectionObserver, )

Add an IntersectionObserver to the Document, to be processed in the Document’s event loop. https://github.com/w3c/IntersectionObserver/issues/525

pub(crate) fn remove_intersection_observer( &self, intersection_observer: &IntersectionObserver, )

Remove an IntersectionObserver from Document, ommiting it from the event loop. An observer without any target, ideally should be removed to be conformant with https://w3c.github.io/IntersectionObserver/#lifetime.

pub(crate) fn update_intersection_observer_steps( &self, cx: &mut JSContext, time: CrossProcessInstant, )

https://w3c.github.io/IntersectionObserver/#update-intersection-observations-algo

fn update_single_intersection_observer_steps( &self, cx: &mut JSContext, intersection_observer: &IntersectionObserver, time: CrossProcessInstant, )

Step 2.1-2.2 of https://w3c.github.io/IntersectionObserver/#update-intersection-observations-algo

pub(crate) fn notify_intersection_observers(&self, cx: &mut JSContext)

https://w3c.github.io/IntersectionObserver/#notify-intersection-observers-algo

pub(crate) fn queue_an_intersection_observer_task(&self)

https://w3c.github.io/IntersectionObserver/#queue-intersection-observer-task

pub(crate) fn store_lcp_candidate(&self, id: LCPCandidateID, element: &Element)

pub(crate) fn handle_paint_metric( &self, cx: &mut JSContext, metric_type: ProgressiveWebMetricType, metric_value: CrossProcessInstant, first_reflow: bool, )

fn write( &self, cx: &mut JSContext, text: Vec<TrustedHTMLOrString<DomTypeHolder>>, line_feed: bool, containing_class: &str, field: &str, ) -> ErrorResult

https://html.spec.whatwg.org/multipage/#document-write-steps

pub(crate) fn details_name_groups<'a: 'b, 'b>( &'a self, no_gc: &'b NoGC, ) -> RefMut<'b, DetailsNameGroups>

pub(crate) fn accessibility_data_mut(&self) -> RefMut<'_, AccessibilityData>

pub(crate) fn accessibility_active(&self) -> bool

pub(crate) fn rooted_nodes_for_accessibility_integrity_check( &self, ) -> Option<FxHashSet<OpaqueNode>>

pub(crate) fn get_document_element_unrooted<'a>( &self, no_gc: &'a NoGC, ) -> Option<UnrootedDom<'a, Element>>

pub(crate) fn collect_reports( &self, reports: &mut Vec<Report>, ops: &mut MallocSizeOfOps, ) -> HashSet<*const JSObject>

impl Document

pub(crate) fn new_inherited( window: &Window, has_browsing_context: HasBrowsingContext, url: Option<ServoUrl>, about_base_url: Option<ServoUrl>, origin: MutableOrigin, is_html_document: IsHTMLDocument, content_type: Option<Mime>, last_modified: Option<String>, activity: DocumentActivity, doc_loader: DocumentLoader, referrer: Option<String>, status_code: Option<u16>, canceller: FetchCanceller, is_initial_about_blank: bool, allow_declarative_shadow_roots: bool, inherited_insecure_requests_policy: Option<InsecureRequestsPolicy>, has_trustworthy_ancestor_origin: bool, custom_element_reaction_stack: Rc<CustomElementReactionStack>, creation_sandboxing_flag_set: SandboxingFlagSet, timeline: &DocumentTimeline, pipeline_id: PipelineId, image_cache: StdArc<dyn ImageCache>, ) -> Document

pub(crate) fn detach_window(&self)

pub(crate) fn window_detached(&self) -> bool

pub(crate) fn insecure_requests_policy(&self) -> InsecureRequestsPolicy

Returns a policy value that should be used for fetches initiated by this document.

pub(crate) fn event_handler(&self) -> &DocumentEventHandler

Get the Document’s DocumentEventHandler.

pub(crate) fn focus_handler(&self) -> &DocumentFocusHandler

Get the Document’s DocumentFocusHandler.

pub(crate) fn embedder_controls(&self) -> &DocumentEmbedderControls

Get the Document’s DocumentEmbedderControls.

fn has_pending_scroll_events(&self) -> bool

Whether or not this Document has any pending scroll events to be processed during “update the rendering.”

pub(crate) fn add_rendering_update_reason(&self, reason: RenderingUpdateReason)

Add a RenderingUpdateReason to this Document which will trigger a rendering update at a later time.

pub(crate) fn clear_rendering_update_reasons(&self)

Clear all RenderingUpdateReasons from this Document.

pub(crate) fn add_script_and_layout_blocker(&self)

Prevent any JS or layout from running until the corresponding call to remove_script_and_layout_blocker. Used to isolate periods in which the DOM is in an unstable state and should not be exposed to arbitrary web content. Any attempts to invoke content JS or query layout during that time will trigger a panic. add_delayed_task will cause the provided task to be executed as soon as the last blocker is removed.

pub(crate) fn remove_script_and_layout_blocker(&self, cx: &mut JSContext)

Terminate the period in which JS or layout is disallowed from running. If no further blockers remain, any delayed tasks in the queue will be executed in queue order until the queue is empty.

pub(crate) fn add_delayed_task<T: 'static + NonSendTaskBox>(&self, task: T)

Enqueue a task to run as soon as any JS and layout blockers are removed.

pub(crate) fn is_safe_to_run_script_or_layout(&self) -> bool

Returns true if the DOM is in a state that will allow running content JS or performing a layout operation.

pub(crate) fn ensure_safe_to_run_script_or_layout(&self)

Assert that the DOM is in a state that will allow running content JS or performing a layout operation.

pub(crate) fn new( cx: &mut JSContext, window: &Window, has_browsing_context: HasBrowsingContext, url: Option<ServoUrl>, about_base_url: Option<ServoUrl>, origin: MutableOrigin, doctype: IsHTMLDocument, content_type: Option<Mime>, last_modified: Option<String>, activity: DocumentActivity, doc_loader: DocumentLoader, referrer: Option<String>, status_code: Option<u16>, canceller: FetchCanceller, is_initial_about_blank: bool, allow_declarative_shadow_roots: bool, inherited_insecure_requests_policy: Option<InsecureRequestsPolicy>, has_trustworthy_ancestor_origin: bool, custom_element_reaction_stack: Rc<CustomElementReactionStack>, creation_sandboxing_flag_set: SandboxingFlagSet, pipeline_id: PipelineId, image_cache: StdArc<dyn ImageCache>, ) -> DomRoot<Document>

fn new_with_proto( cx: &mut JSContext, window: &Window, proto: Option<HandleObject<'_>>, has_browsing_context: HasBrowsingContext, url: Option<ServoUrl>, about_base_url: Option<ServoUrl>, origin: MutableOrigin, doctype: IsHTMLDocument, content_type: Option<Mime>, last_modified: Option<String>, activity: DocumentActivity, doc_loader: DocumentLoader, referrer: Option<String>, status_code: Option<u16>, canceller: FetchCanceller, is_initial_about_blank: bool, allow_declarative_shadow_roots: bool, inherited_insecure_requests_policy: Option<InsecureRequestsPolicy>, has_trustworthy_ancestor_origin: bool, custom_element_reaction_stack: Rc<CustomElementReactionStack>, creation_sandboxing_flag_set: SandboxingFlagSet, pipeline_id: PipelineId, image_cache: StdArc<dyn ImageCache>, ) -> DomRoot<Document>

pub(crate) fn get_redirect_count(&self) -> u16

pub(crate) fn set_resource_fetch_timing(&self, timing: ResourceFetchTiming)

pub(crate) fn resource_fetch_timing( &self, ) -> Ref<'_, Option<ResourceFetchTiming>>

pub(crate) fn navigation_timing(&self) -> Rc<NavigationTiming>

pub(crate) fn performance_timing_attribute( &self, name: &str, ) -> Fallible<Option<CrossProcessInstant>>

pub(crate) fn elements_by_name_count(&self, name: &DOMString) -> u32

pub(crate) fn nth_element_by_name<'a>( &self, no_gc: &'a NoGC, index: u32, name: &DOMString, ) -> Option<UnrootedDom<'a, Node>>

fn is_element_in_get_by_name(node: &Node, name: &DOMString) -> bool

fn count_node_list<F: Fn(&Node) -> bool>(&self, callback: F) -> u32

fn nth_in_node_list<'a, F: Fn(&Node) -> bool>( &self, no_gc: &'a NoGC, index: u32, callback: F, ) -> Option<UnrootedDom<'a, Node>>

fn get_html_element(&self) -> Option<DomRoot<HTMLHtmlElement>>

pub(crate) fn shared_style_locks(&self) -> &SharedRwLocks

Return a reference to the per-ScriptThread shared locks used for stylesheets.

pub(crate) fn style_shared_author_lock(&self) -> &SharedRwLock

Return a reference to the per-ScriptThread shared lock used for author stylesheets.

pub(crate) fn flush_stylesheets_for_reflow(&self) -> bool

Flushes the stylesheet list, and returns whether any stylesheet changed.

pub(crate) fn salvageable(&self) -> bool

pub(crate) fn make_document_unsalvageable(&self)

https://html.spec.whatwg.org/multipage/#make-document-unsalvageable

pub(crate) fn appropriate_template_contents_owner_document( &self, cx: &mut JSContext, ) -> DomRoot<Document>

https://html.spec.whatwg.org/multipage/#appropriate-template-contents-owner-document

pub(crate) fn get_element_by_id( &self, no_gc: &NoGC, id: &Atom, ) -> Option<DomRoot<Element>>

pub(crate) fn ensure_pending_restyle( &self, el: &Element, ) -> RefMut<'_, PendingRestyle>

pub(crate) fn element_attr_will_change(&self, el: &Element, attr: AttrRef<'_>)

pub(crate) fn set_referrer_policy(&self, policy: ReferrerPolicy)

pub(crate) fn get_referrer_policy(&self) -> ReferrerPolicy

pub(crate) fn set_target_element(&self, node: Option<&Element>)

pub(crate) fn incr_ignore_destructive_writes_counter(&self)

pub(crate) fn decr_ignore_destructive_writes_counter(&self)

pub(crate) fn is_prompting_or_unloading(&self) -> bool

fn incr_ignore_opens_during_unload_counter(&self)

fn decr_ignore_opens_during_unload_counter(&self)

pub(crate) fn set_fullscreen_element(&self, element: Option<&Element>)

fn reset_form_owner_for_listeners(&self, cx: &mut JSContext, id: &Atom)

pub(crate) fn register_shadow_root(&self, shadow_root: &ShadowRoot)

pub(crate) fn unregister_shadow_root(&self, shadow_root: &ShadowRoot)

pub(crate) fn invalidate_shadow_roots_stylesheets(&self)

pub(crate) fn flush_shadow_root_stylesheets_if_necessary_for_layout( &self, stylist: &mut Stylist, guard: &SharedRwLockReadGuard<'_>, )

pub(crate) fn stylesheet_count(&self) -> usize

pub(crate) fn stylesheet_at( &self, cx: &mut JSContext, index: usize, ) -> Option<DomRoot<CSSStyleSheet>>

pub(crate) fn add_owned_stylesheet( &self, no_gc: &NoGC, owner_node: &Element, sheet: Arc<Stylesheet>, )

Add a stylesheet owned by owner_node to the list of document sheets, in the correct tree position. Additionally, ensure that the owned stylesheet is inserted before any constructed stylesheet.

https://drafts.csswg.org/cssom/#documentorshadowroot-final-css-style-sheets

pub(crate) fn append_constructed_stylesheet( &self, cssom_stylesheet: &CSSStyleSheet, )

Append a constructed stylesheet to the back of document stylesheet set. Because it would be the last element, we therefore would not mess with the ordering.

https://drafts.csswg.org/cssom/#documentorshadowroot-final-css-style-sheets

pub(crate) fn add_stylesheet_to_stylist( &self, stylesheet: Arc<Stylesheet>, before_stylesheet: Option<Arc<Stylesheet>>, )

pub(crate) fn remove_stylesheet( &self, owner: StylesheetSource, stylesheet: &Arc<Stylesheet>, )

Remove a stylesheet owned by owner from the list of document sheets.

pub(crate) fn get_elements_with_id( &self, cx: &mut JSContext, id: &Atom, ) -> Ref<'_, [Dom<Element>]>

pub(crate) fn get_elements_with_name( &self, cx: &mut JSContext, name: &Atom, ) -> Ref<'_, [Dom<Element>]>

pub(crate) fn drain_pending_restyles( &self, no_gc: &NoGC, ) -> Vec<(TrustedNodeAddress, PendingRestyle)>

pub(crate) fn advance_animation_timeline_for_testing(&self, delta: TimeDuration)

pub(crate) fn maybe_mark_animating_nodes_as_dirty(&self, no_gc: &NoGC)

pub(crate) fn current_animation_timeline_value(&self) -> f64

pub(crate) fn animations(&self) -> &Animations

pub(crate) fn update_animations_post_reflow(&self)

pub(crate) fn cancel_animations_for_node(&self, node: &Node)

pub(crate) fn remove_style_and_layout_data_from_subtree( &self, no_gc: &NoGC, subtree_root: &Node, )

Clear style and layout data on this Node and all descendants. This is used to clean up the data when a Node becomes detached from the flat tree. Note that this operates on shadow-including descendants.

pub(crate) fn clean_up_style_and_layout_data_for_node(&self, node: &Node)

pub(crate) fn update_animations_and_send_events( &self, cx: &mut CurrentRealm<'_>, )

An implementation of https://drafts.csswg.org/web-animations-1/#update-animations-and-send-events.

pub(crate) fn image_animation_manager(&self) -> Ref<'_, ImageAnimationManager>

pub(crate) fn set_has_pending_animated_image_update(&self)

pub(crate) fn shared_declarative_refresh_steps( &self, content: &[u8], from_meta_element: bool, )

https://html.spec.whatwg.org/multipage/#shared-declarative-refresh-steps

pub(crate) fn will_declaratively_refresh(&self) -> bool

pub(crate) fn set_declarative_refresh(&self, refresh: DeclarativeRefresh)

fn update_visibility_state( &self, cx: &mut JSContext, visibility_state: DocumentVisibilityState, )

https://html.spec.whatwg.org/multipage/#visibility-state

pub(crate) fn is_initial_about_blank(&self) -> bool

https://html.spec.whatwg.org/multipage/#is-initial-about:blank

pub(crate) fn allow_declarative_shadow_roots(&self) -> bool

https://dom.spec.whatwg.org/#document-allow-declarative-shadow-roots

pub(crate) fn has_trustworthy_ancestor_origin(&self) -> bool

pub(crate) fn has_trustworthy_ancestor_or_current_origin(&self) -> bool

pub(crate) fn highlight_dom_node(&self, node: Option<&Node>)

pub(crate) fn highlighted_dom_node(&self) -> Option<DomRoot<Node>>

pub(crate) fn custom_element_reaction_stack( &self, ) -> Rc<CustomElementReactionStack>

pub(crate) fn active_sandboxing_flag_set(&self) -> SandboxingFlagSet

pub(crate) fn has_active_sandboxing_flag(&self, flag: SandboxingFlagSet) -> bool

pub(crate) fn set_active_sandboxing_flag_set(&self, flags: SandboxingFlagSet)

pub(crate) fn creation_sandboxing_flag_set(&self) -> SandboxingFlagSet

pub(crate) fn creation_sandboxing_flag_set_considering_parent_iframe( &self, ) -> SandboxingFlagSet

pub(crate) fn viewport_scrolling_box( &self, flags: ScrollContainerQueryFlags, ) -> ScrollingBox

pub(crate) fn notify_embedder_favicon(&self)

pub(crate) fn set_favicon(&self, favicon: Image)

pub(crate) fn fullscreen_element(&self) -> Option<DomRoot<Element>>

pub(crate) fn state_override(&self, command_name: &CommandName) -> Option<bool>

https://w3c.github.io/editing/docs/execCommand/#state-override

pub(crate) fn set_state_override( &self, command_name: CommandName, state: Option<bool>, )

https://w3c.github.io/editing/docs/execCommand/#state-override

pub(crate) fn value_override( &self, command_name: &CommandName, ) -> Option<DOMString>

https://w3c.github.io/editing/docs/execCommand/#value-override

pub(crate) fn set_value_override( &self, command_name: CommandName, value: Option<DOMString>, )

https://w3c.github.io/editing/docs/execCommand/#value-override

pub(crate) fn clear_command_overrides(&self)

https://w3c.github.io/editing/docs/execCommand/#value-override and https://w3c.github.io/editing/docs/execCommand/#state-override

pub(crate) fn default_single_line_container_name( &self, ) -> DefaultSingleLineContainerName

https://w3c.github.io/editing/docs/execCommand/#default-single-line-container-name

pub(crate) fn set_default_single_line_container_name( &self, value: DefaultSingleLineContainerName, )

https://w3c.github.io/editing/docs/execCommand/#default-single-line-container-name

pub(crate) fn css_styling_flag(&self) -> bool

https://w3c.github.io/editing/docs/execCommand/#css-styling-flag

pub(crate) fn set_css_styling_flag(&self, value: bool)

https://w3c.github.io/editing/docs/execCommand/#css-styling-flag

pub(crate) fn mute_iframe_load_flag(&self) -> bool

pub(crate) fn set_iframe_load_in_progress(&self, value: bool)

pub(crate) fn theme(&self) -> Option<Theme>

pub(crate) fn set_theme(&self, new_theme: Option<Theme>)

impl Document

pub(crate) fn create_element( &self, cx: &mut JSContext, name: &str, ) -> DomRoot<Element>

impl Document

fn selection_if_command_is_enabled( &self, cx: &mut JSContext, command_name: CommandName, ) -> Option<DomRoot<Selection>>

https://w3c.github.io/editing/docs/execCommand/#enabled

fn command_if_command_is_supported( &self, command_id: &DOMString, ) -> Option<CommandName>

https://w3c.github.io/editing/docs/execCommand/#supported

impl Document

pub(crate) fn enter_fullscreen( &self, cx: &mut CurrentRealm<'_>, pending: &Element, ) -> Rc<Promise>

https://fullscreen.spec.whatwg.org/#dom-element-requestfullscreen

pub(crate) fn exit_fullscreen(&self, cx: &mut JSContext) -> Rc<Promise>

https://fullscreen.spec.whatwg.org/#exit-fullscreen

pub(crate) fn get_allow_fullscreen(&self) -> bool

impl Document

fn internal_ancestor_origin_objects_list_creation_steps( &self, referrer_policy: &ReferrerPolicy, ) -> Vec<ImmutableOrigin>

https://html.spec.whatwg.org/multipage/#internal-ancestor-origin-objects-list-creation-steps

fn ancestor_origins_list_creation_steps( &self, cx: &mut JSContext, ) -> DomRoot<DOMStringList>

https://html.spec.whatwg.org/multipage/#ancestor-origins-list-creation-steps