script::dom::document::document_event_handler

Function keyboard_event_fires_keypress_event

fn keyboard_event_fires_keypress_event(keyboard_event: &KeyboardEvent) -> bool

Whether a given EmbedderKeyboardEvent should trigger a keypress event.

From https://w3c.github.io/uievents/#keypress:

If supported by a user agent, this event MUST be dispatched when a key is pressed down, if and only if that key normally produces a character value. The keypress event type is device dependent and relies on the capabilities of the input devices and how they are mapped in the operating system.