script::dom::document::documentorshadowroot

Struct DocumentOrShadowRoot

pub(crate) struct DocumentOrShadowRoot {
    window: Dom<Window>,
    custom_element_registry: MutNullableDom<CustomElementRegistry>,
}

Fields
window: Dom<Window>custom_element_registry: MutNullableDom<CustomElementRegistry>
Implementations

impl DocumentOrShadowRoot

pub(crate) fn new(window: &Window) -> Self

pub(crate) fn custom_element_registry( &self, ) -> Option<DomRoot<CustomElementRegistry>>

https://dom.spec.whatwg.org/#dom-documentorshadowroot-customelementregistry

pub(crate) fn set_custom_element_registry( &self, registry: Option<&CustomElementRegistry>, )

pub(crate) fn retarget_hit_test_result( &self, this: &Node, node: &Node, ) -> Option<DomRoot<Element>>

Retarget the result of elementsFromPoint or elementFromPoint according to the resolution in https://github.com/w3c/csswg-drafts/issues/556.

pub(crate) fn element_from_point( &self, this: &Node, x: Finite<f64>, y: Finite<f64>, document_element: Option<DomRoot<Element>>, has_browsing_context: bool, ) -> Option<DomRoot<Element>>

https://drafts.csswg.org/cssom-view/#dom-document-elementfrompoint

pub(crate) fn elements_from_point( &self, this: &Node, x: Finite<f64>, y: Finite<f64>, document_element: Option<DomRoot<Element>>, has_browsing_context: bool, ) -> Vec<DomRoot<Element>>

https://drafts.csswg.org/cssom-view/#dom-document-elementsfrompoint

pub(crate) fn active_element(&self, this: &Node) -> Option<DomRoot<Element>>

https://html.spec.whatwg.org/multipage/#dom-documentorshadowroot-activeelement-dev

pub(crate) fn remove_stylesheet( owner: StylesheetSource, s: &Arc<Stylesheet>, stylesheets: StylesheetSetRef<'_, ServoStylesheetInDocument>, )

Remove a stylesheet owned by owner from the list of document sheets.

pub(crate) fn add_stylesheet( owner: StylesheetSource, stylesheets: StylesheetSetRef<'_, ServoStylesheetInDocument>, sheet: Arc<Stylesheet>, insertion_point: Option<ServoStylesheetInDocument>, style_shared_lock: &StyleSharedRwLock, )

Add a stylesheet owned by owner to the list of document sheets, in the correct tree position.

fn set_adopted_stylesheet( adopted_stylesheets: &mut Vec<Dom<CSSStyleSheet>>, incoming_stylesheets: &[Dom<CSSStyleSheet>], owner: &StyleSheetListOwner, ) -> ErrorResult

Inner part of adopted stylesheet. We are setting it by, assuming it is a FrozenArray instead of an ObservableArray. Thus, it would have a completely different workflow compared to the spec. The workflow here is actually following Gecko’s implementation of AdoptedStylesheet before the implementation of ObservableArray.

The main purpose from this function is to set the &mut adopted_stylesheet to match incoming_stylesheet and update the corresponding Styleset in a Document or a ShadowRoot. In case of duplicates, the setter will respect the last duplicates.

https://drafts.csswg.org/cssom/#dom-documentorshadowroot-adoptedstylesheets

pub(crate) fn set_adopted_stylesheet_from_jsval( cx: &mut JSContext, adopted_stylesheets: &DomRefCell<Vec<Dom<CSSStyleSheet>>>, incoming_value: HandleValue<'_>, owner: &StyleSheetListOwner, ) -> ErrorResult

Set adoptedStylesheet given a js value by converting and passing the converted values to the inner DocumentOrShadowRoot::set_adopted_stylesheet.

impl DocumentOrShadowRoot

pub(crate) fn get_fullscreen_element( node: &Node, fullscreen_element: Option<DomRoot<Element>>, ) -> Option<DomRoot<Element>>

https://fullscreen.spec.whatwg.org/#dom-document-fullscreenelement