script::dom::document::focus

Struct DocumentFocusHandler

pub(crate) struct DocumentFocusHandler {
    window: Dom<Window>,
    focused_area: DomRefCell<FocusableArea>,
    focus_sequence: Cell<FocusSequenceNumber>,
    has_focus: Cell<bool>,
    sequential_focus_navigation_starting_point: MutNullableDom<Node>,
}

The DocumentFocusHandler is a structure responsible for handling and storing data related to focus for the Document. It exists to decrease the size of the Document. structure.

Fields
window: Dom<Window>

The Window element for this DocumentFocusHandler.

focused_area: DomRefCell<FocusableArea>

The focused area of the Document.

https://html.spec.whatwg.org/multipage/#focused-area-of-the-document

focus_sequence: Cell<FocusSequenceNumber>

The last sequence number sent to the constellation.

has_focus: Cell<bool>

Indicates whether the container is included in the top-level browsing context’s focus chain (not considering system focus). Permanently true for a top-level document.

sequential_focus_navigation_starting_point: MutNullableDom<Node>

https://html.spec.whatwg.org/multipage/#sequential-focus-navigation-starting-point

Implementations

impl DocumentFocusHandler

pub(crate) fn new(window: &Window, has_focus: bool) -> Self

pub(crate) fn has_focus(&self) -> bool

pub(crate) fn set_has_focus(&self, has_focus: bool)

pub(crate) fn focused_area(&self) -> Ref<'_, FocusableArea>

Return the element that currently has focus. If None is returned the viewport itself has focus.

pub(crate) fn set_focused_area(&self, new_focusable_area: FocusableArea)

Set the element that currently has focus and update the focus state for both the previously set element (if any) and the new one, as well as the new one. This will not do anything if the new element is the same as the previous one. Note that this will not fire any focus events. If that is necessary the DocumentFocusHandler::focus should be used.

pub fn focus_sequence(&self) -> FocusSequenceNumber

Get the last sequence number sent to the constellation.

Received focus-related messages with sequence numbers less than the one returned by this method must be discarded.

fn increment_fetch_focus_sequence(&self) -> FocusSequenceNumber

Generate the next sequence number for focus-related messages.

pub(crate) fn current_focus_chain(&self) -> Vec<FocusableArea>

https://html.spec.whatwg.org/multipage/#current-focus-chain-of-a-top-level-traversable

pub(crate) fn focus(&self, cx: &mut JSContext, new_focus_target: &FocusableArea)

Reassign the focus context to the element that last requested focus during this transaction, or the document if no elements requested it.

pub(crate) fn focus_update_steps( &self, cx: &mut JSContext, new_focus_chain: RootedGuard<'_, Vec<FocusableArea>>, old_focus_chain: RootedGuard<'_, Vec<FocusableArea>>, new_focus_target: &FocusableArea, )

https://html.spec.whatwg.org/multipage/#focus-update-steps

pub(crate) fn fire_focus_event( &self, cx: &mut JSContext, focus_event_type: FocusEventType, event_target: &EventTarget, related_target: Option<&EventTarget>, )

https://html.spec.whatwg.org/multipage/#fire-a-focus-event

pub(crate) fn perform_focus_fixup_rule(&self, cx: &mut JSContext)

https://html.spec.whatwg.org/multipage/#focus-fixup-rule

For each doc of docs, if the focused area of doc is not a focusable area, then run the focusing steps for doc’s viewport, and set doc’s relevant global object’s navigation API’s focus changed during ongoing navigation to false.

TODO: Handle the “focus changed during ongoing navigation” flag.

pub(crate) fn set_sequential_focus_navigation_starting_point(&self, node: &Node)

fn sequential_focus_navigation_starting_point(&self) -> Option<DomRoot<Node>>

pub(crate) fn sequential_focus_navigation_via_keyboard_event( &self, cx: &mut JSContext, event: &KeyboardEvent, )

fn sequential_focus_navigation( &self, cx: &mut JSContext, direction: SequentialFocusDirection, )

https://html.spec.whatwg.org/multipage/#sequential-focus-navigation:currently-focused-area-of-a-top-level-traversable

fn sequential_focus_navigation_loop( &self, cx: &mut JSContext, starting_point: Option<DomRoot<Node>>, direction: SequentialFocusDirection, allow_focusing_viewport: bool, )

The inner loop (“Loop”) of: https://html.spec.whatwg.org/multipage/#sequential-focus-navigation:currently-focused-area-of-a-top-level-traversable

fn sequentially_focus_child_iframe_local_or_remote( &self, cx: &mut JSContext, iframe_element: &HTMLIFrameElement, direction: SequentialFocusDirection, )

fn sequentially_focus_parent_local_or_remote( &self, cx: &mut JSContext, direction: SequentialFocusDirection, )

pub(crate) fn sequential_focus_from_another_document( &self, cx: &mut JSContext, browsing_context_id: Option<BrowsingContextId>, direction: SequentialFocusDirection, )