script::dom::documentorshadowroot

Struct DocumentOrShadowRoot

pub(crate) struct DocumentOrShadowRoot {
    window: Dom<Window>,
}

Fields
window: Dom<Window>
Implementations

impl DocumentOrShadowRoot

pub(crate) fn new(window: &Window) -> Self

pub(crate) fn element_from_point( &self, x: Finite<f64>, y: Finite<f64>, document_element: Option<DomRoot<Element>>, has_browsing_context: bool, ) -> Option<DomRoot<Element>>

pub(crate) fn elements_from_point( &self, x: Finite<f64>, y: Finite<f64>, document_element: Option<DomRoot<Element>>, has_browsing_context: bool, ) -> Vec<DomRoot<Element>>

pub(crate) fn get_active_element( &self, focused_element: Option<DomRoot<Element>>, body: Option<DomRoot<HTMLElement>>, document_element: Option<DomRoot<Element>>, ) -> Option<DomRoot<Element>>

pub(crate) fn remove_stylesheet( owner: StylesheetSource, s: &Arc<Stylesheet>, stylesheets: StylesheetSetRef<'_, ServoStylesheetInDocument>, )

Remove a stylesheet owned by owner from the list of document sheets.

pub(crate) fn add_stylesheet( owner: StylesheetSource, stylesheets: StylesheetSetRef<'_, ServoStylesheetInDocument>, sheet: Arc<Stylesheet>, insertion_point: Option<ServoStylesheetInDocument>, style_shared_lock: &StyleSharedRwLock, )

Add a stylesheet owned by owner to the list of document sheets, in the correct tree position.

pub(crate) fn unregister_named_element( &self, id_map: &DomRefCell<HashMapTracedValues<Atom, Vec<Dom<Element>>, FxBuildHasher>>, to_unregister: &Element, id: &Atom, )

Remove any existing association between the provided id/name and any elements in this document.

pub(crate) fn register_named_element( &self, id_map: &DomRefCell<HashMapTracedValues<Atom, Vec<Dom<Element>>, FxBuildHasher>>, element: &Element, id: &Atom, root: DomRoot<Node>, )

Associate an element present in this document with the provided id/name.

fn set_adopted_stylesheet( adopted_stylesheets: &mut Vec<Dom<CSSStyleSheet>>, incoming_stylesheets: &[Dom<CSSStyleSheet>], owner: &StyleSheetListOwner, ) -> ErrorResult

Inner part of adopted stylesheet. We are setting it by, assuming it is a FrozenArray instead of an ObservableArray. Thus, it would have a completely different workflow compared to the spec. The workflow here is actually following Geckoโs implementation of AdoptedStylesheet before the implementation of ObservableArray.

The main purpose from this function is to set the &mut adopted_stylesheet to match incoming_stylesheet and update the corresponding Styleset in a Document or a ShadowRoot. In case of duplicates, the setter will respect the last duplicates.

https://drafts.csswg.org/cssom/#dom-documentorshadowroot-adoptedstylesheets

pub(crate) fn set_adopted_stylesheet_from_jsval( context: JSContext, adopted_stylesheets: &mut Vec<Dom<CSSStyleSheet>>, incoming_value: HandleValue<'_>, owner: &StyleSheetListOwner, can_gc: CanGc, ) -> ErrorResult

Set adoptedStylesheet given a js value by converting and passing the converted values to the inner DocumentOrShadowRoot::set_adopted_stylesheet.