script::dom::element::element

Struct Element

#[repr(C)]
pub struct Element {

    node: Node,
    local_name: LocalName,
    tag_name: TagName,
    namespace: Namespace,
    prefix: DomRefCell<Option<Prefix>>,
    attrs: AttributeStorage,
    id_attribute: DomRefCell<Option<Atom>>,
    is: DomRefCell<Option<LocalName>>,
    style_attribute: DomRefCell<Option<Arc<Locked<PropertyDeclarationBlock>>>>,
    attr_list: MutNullableDom<NamedNodeMap>,
    class_list: MutNullableDom<DOMTokenList>,
    state: Cell<ElementState>,
    selector_flags: AtomicUsize,
    rare_data: DomRefCell<Option<Box<ElementRareData>>>,
    style_data: DomRefCell<Option<Box<StyleData>>>,

}

https://dom.spec.whatwg.org/#element

Fields
node: Nodelocal_name: LocalNametag_name: TagNamenamespace: Namespaceprefix: DomRefCell<Option<Prefix>>attrs: AttributeStorageid_attribute: DomRefCell<Option<Atom>>is: DomRefCell<Option<LocalName>>

https://dom.spec.whatwg.org/#concept-element-is-value

style_attribute: DomRefCell<Option<Arc<Locked<PropertyDeclarationBlock>>>>attr_list: MutNullableDom<NamedNodeMap>class_list: MutNullableDom<DOMTokenList>state: Cell<ElementState>selector_flags: AtomicUsize

These flags are set by the style system to indicate the that certain operations may require restyling this element or its descendants.

rare_data: DomRefCell<Option<Box<ElementRareData>>>style_data: DomRefCell<Option<Box<StyleData>>>

Style data for this node. This is accessed and mutated by style passes and is used to lay out this node and populate layout data.

Implementations

impl Element

fn __assert_parent_type(&self)

impl Element

fn has_uncollapsed_selection(&self) -> bool

fn has_selectable_text(&self) -> bool

fn select_all(&self)

impl Element

fn inclusive_ancestor_element_in_non_ua_shadow_root(&self) -> DomRoot<Element>

Find the first inclusive ancestor of this Element that is not in a UA shadow root.

impl Element

pub(crate) fn get_attribute_string_value( &self, local_name: &LocalName, ) -> Option<String>

Callers should convert the LocalName to ASCII lowercase before calling. https://dom.spec.whatwg.org/#concept-element-attributes-get-by-name

pub(crate) fn get_attribute_string_value_with_namespace( &self, namespace: &Namespace, local_name: &LocalName, ) -> Option<String>

pub(crate) fn get_int_attribute( &self, local_name: &LocalName, default: i32, ) -> i32

pub(crate) fn set_atomic_attribute( &self, cx: &mut JSContext, local_name: &LocalName, value: DOMString, )

pub(crate) fn set_bool_attribute( &self, cx: &mut JSContext, local_name: &LocalName, value: bool, )

pub(crate) fn get_url_attribute(&self, local_name: &LocalName) -> USVString

pub(crate) fn set_url_attribute( &self, cx: &mut JSContext, local_name: &LocalName, value: USVString, )

pub(crate) fn get_trusted_type_url_attribute( &self, local_name: &LocalName, ) -> TrustedScriptURLOrUSVString<DomTypeHolder>

pub(crate) fn get_trusted_html_attribute( &self, local_name: &LocalName, ) -> TrustedHTMLOrString<DomTypeHolder>

pub(crate) fn get_string_attribute(&self, local_name: &LocalName) -> DOMString

pub(crate) fn set_string_attribute( &self, cx: &mut JSContext, local_name: &LocalName, value: DOMString, )

pub(crate) fn get_nullable_string_attribute( &self, local_name: &LocalName, ) -> Option<DOMString>

Used for string attribute reflections where absence of the attribute returns null, e.g. element.ariaLabel returning null when the aria-label attribute is absent.

pub(crate) fn set_nullable_string_attribute( &self, cx: &mut JSContext, local_name: &LocalName, value: Option<DOMString>, )

Used for string attribute reflections where setting null/undefined removes the attribute, e.g. element.ariaLabel = null removing the aria-label attribute.

pub(crate) fn set_nullable_tokenlist_attribute( &self, cx: &mut JSContext, local_name: &LocalName, value: Option<DOMString>, )

pub(crate) fn any_tokenlist_attribute( &self, local_name: &LocalName, f: impl Fn(&Atom) -> bool, ) -> bool

Returns true if any attribute in the tokenlist fulfill f. Equivalent to get_tokenlist_attribute(name).iter().any(f).

pub(crate) fn get_tokenlist_attribute( &self, local_name: &LocalName, ) -> Vec<Atom>

pub(crate) fn set_tokenlist_attribute( &self, cx: &mut JSContext, local_name: &LocalName, value: DOMString, )

pub(crate) fn get_uint_attribute( &self, local_name: &LocalName, default: u32, ) -> u32

fn compute_attribute_value_with_style_fast_path( &self, attr: AttrRef<'_>, ) -> AttrValue

Ensure that for styles, we clone the already-parsed property declaration block. This does two things:

It uses the same fast-path as CSSStyleDeclaration

It also avoids the CSP checks when cloning (it shouldn’t run any when cloning existing valid attributes)

pub(crate) fn copy_all_attributes_to_other_element( &self, cx: &mut JSContext, target_element: &Element, )

https://dom.spec.whatwg.org/#concept-node-clone

impl Element

pub(crate) fn create( cx: &mut JSContext, name: QualName, is: Option<LocalName>, document: &Document, creator: ElementCreator, mode: CustomElementCreationMode, proto: Option<HandleObject<'_>>, ) -> DomRoot<Element>

pub(crate) fn new_inherited( local_name: LocalName, namespace: Namespace, prefix: Option<Prefix>, document: &Document, ) -> Element

pub(crate) fn new_inherited_with_state( state: ElementState, local_name: LocalName, namespace: Namespace, prefix: Option<Prefix>, document: &Document, ) -> Element

pub(crate) fn set_had_duplicate_attributes(&self, no_gc: &NoGC)

pub(crate) fn new( cx: &mut JSContext, local_name: LocalName, namespace: Namespace, prefix: Option<Prefix>, document: &Document, proto: Option<HandleObject<'_>>, ) -> DomRoot<Element>

fn rare_data(&self) -> Ref<'_, Option<Box<ElementRareData>>>

fn rare_data_mut(&self) -> RefMut<'_, Option<Box<ElementRareData>>>

pub(crate) fn ensure_rare_data<'a: 'b, 'b>( &'a self, no_gc: &'b NoGC, ) -> RefMut<'b, Box<ElementRareData>>

pub(crate) fn clean_up_style_data(&self)

pub(crate) fn restyle(&self, no_gc: &NoGC, damage: NodeDamage)

pub(crate) fn has_dirty_descendants(&self) -> bool

pub(crate) fn note_dirty_descendants(&self, no_gc: &NoGC)

pub(crate) fn set_is(&self, is: LocalName)

pub(crate) fn get_is(&self) -> Option<LocalName>

https://dom.spec.whatwg.org/#concept-element-is-value

pub(crate) fn set_initial_custom_element_state_to_uncustomized(&self)

This is a performance optimization. Element::create can simply call element.set_custom_element_state(CustomElementState::Uncustomized) to initialize uncustomized, built-in elements with the right state, which currently just means that the DEFINED state should be true for styling. However set_custom_element_state has a high performance cost and it is unnecessary if the element is being created as an uncustomized built-in element.

See https://github.com/servo/servo/issues/37745 for more details.

pub(crate) fn set_custom_element_state( &self, state: CustomElementState, no_gc: &NoGC, )

https://dom.spec.whatwg.org/#concept-element-custom-element-state

pub(crate) fn get_custom_element_state(&self) -> CustomElementState

pub(crate) fn is_custom(&self) -> bool

https://dom.spec.whatwg.org/#concept-element-custom

pub(crate) fn set_custom_element_definition( &self, definition: Rc<CustomElementDefinition>, no_gc: &NoGC, )

pub(crate) fn get_custom_element_definition( &self, ) -> Option<Rc<CustomElementDefinition>>

pub(crate) fn clear_custom_element_definition(&self, no_gc: &NoGC)

pub(crate) fn push_callback_reaction( &self, function: Rc<Function<DomTypeHolder>>, args: Box<[Heap<JSVal>]>, no_gc: &NoGC, )

pub(crate) fn push_upgrade_reaction( &self, definition: Rc<CustomElementDefinition>, no_gc: &NoGC, )

pub(crate) fn clear_reaction_queue(&self)

pub(crate) fn invoke_reactions(&self, cx: &mut JSContext)

pub(crate) fn has_css_layout_box(&self) -> bool

pub(crate) fn is_potentially_scrollable_body(&self) -> bool

https://drafts.csswg.org/cssom-view/#potentially-scrollable

pub(crate) fn is_potentially_scrollable_body_for_scrolling_element( &self, ) -> bool

https://drafts.csswg.org/cssom-view/#potentially-scrollable

fn is_potentially_scrollable_body_shared_logic( &self, treat_overflow_clip_on_parent_as_hidden: bool, ) -> bool

https://drafts.csswg.org/cssom-view/#potentially-scrollable

pub(crate) fn establishes_scroll_container(&self) -> bool

Whether this element is styled such that it establishes a scroll container. https://www.w3.org/TR/css-overflow-3/#scroll-container

pub(crate) fn establishes_scroll_container_without_reflow(&self) -> bool

Like Self::establishes_scroll_container, but without forcing a reflow.

pub(crate) fn has_overflow(&self, no_gc: &NoGC) -> bool

fn has_scrolling_box(&self, no_gc: &NoGC) -> bool

Whether or not this element has a scrolling box according to https://drafts.csswg.org/cssom-view/#scrolling-box.

This is true if:

The element has a layout box.

The style specifies that overflow should be scrollable (auto, hidden or scroll).

The fragment actually has content that overflows the box.

pub(crate) fn shadow_root(&self) -> Option<DomRoot<ShadowRoot>>

pub(crate) fn shadow_root_unrooted<'a>( &self, no_gc: &'a NoGC, ) -> Option<UnrootedDom<'a, ShadowRoot>>

pub(crate) fn is_shadow_host(&self) -> bool

pub(crate) fn attach_shadow( &self, cx: &mut JSContext, is_ua_widget: IsUserAgentWidget, mode: ShadowRootMode, clonable: bool, serializable: bool, delegates_focus: bool, slot_assignment_mode: SlotAssignmentMode, ) -> Fallible<DomRoot<ShadowRoot>>

https://dom.spec.whatwg.org/#dom-element-attachshadow

pub(crate) fn attach_ua_shadow_root( &self, cx: &mut JSContext, use_ua_widget_styling: bool, ) -> DomRoot<ShadowRoot>

Attach a UA widget shadow root with its default parameters. Additionally mark ShadowRoot to use styling configuration for a UA widget.

The general trait of these elements is that it would hide the implementation. Thus, we would make it inaccessible (i.e., closed mode, not cloneable, and not serializable).

With UA shadow root element being assumed as one element, any focus should be delegated to its host.

pub(crate) fn is_translate_enabled(&self) -> bool

pub(crate) fn directionality(&self) -> String

pub(crate) fn is_root(&self) -> bool

pub(crate) fn registered_intersection_observers_mut<'a: 'b, 'b>( &'a self, no_gc: &'b NoGC, ) -> RefMut<'b, Vec<IntersectionObserverRegistration>>

Return all IntersectionObserverRegistration for this element. Lazily initialize the raredata if it does not exist.

pub(crate) fn registered_intersection_observers( &self, ) -> Option<Ref<'_, Vec<IntersectionObserverRegistration>>>

pub(crate) fn get_intersection_observer_registration( &self, observer: &IntersectionObserver, ) -> Option<Ref<'_, IntersectionObserverRegistration>>

pub(crate) fn add_initial_intersection_observer_registration( &self, observer: &IntersectionObserver, no_gc: &NoGC, )

Add a new IntersectionObserverRegistration with initial value to the element.

pub(crate) fn remove_intersection_observer( &self, observer: &IntersectionObserver, no_gc: &NoGC, )

Removes a certain IntersectionObserver.

pub(crate) fn scrolling_box( &self, flags: ScrollContainerQueryFlags, ) -> Option<ScrollingBox>

Get the ScrollingBox that contains this element, if one does. position: fixed elements do not have a containing ScrollingBox.

pub(crate) fn scroll_into_view_with_options( &self, cx: &mut JSContext, behavior: ScrollBehavior, block: ScrollAxisState, inline: ScrollAxisState, container: Option<&Element>, inner_target_rect: Option<Rect<Au, CSSPixel>>, )

https://drafts.csswg.org/cssom-view/#scroll-a-target-into-view

pub(crate) fn ensure_contenteditable_selection_range( &self, cx: &mut JSContext, document: &Document, ) -> DomRoot<Range>

pub(crate) fn handle_scroll_event(&self)

https://drafts.csswg.org/cssom-view/#scrolling-events

Whenever an element gets scrolled (whether in response to user interaction or by an API), the user agent must run these steps:

pub(crate) fn style(&self) -> Option<ServoArc<ComputedValues>>

pub(crate) fn is_styled(&self) -> bool

pub(crate) fn is_display_none(&self) -> bool

pub(crate) fn check_style_on_self_or_eager_pseudos( &self, check_styles_fn: impl Fn(&ComputedValues) -> bool, ) -> bool

impl Element

pub(crate) fn is_html_element(&self) -> bool

pub(crate) fn is_svg_element(&self) -> bool

pub(crate) fn html_element_in_html_document(&self) -> bool

pub(crate) fn local_name(&self) -> &LocalName

pub(crate) fn parsed_name(&self, name: DOMString) -> LocalName

pub(crate) fn namespace(&self) -> &Namespace

pub(crate) fn prefix(&self) -> Ref<'_, Option<Prefix>>

pub(crate) fn set_prefix(&self, prefix: Option<Prefix>)

pub(crate) fn set_custom_element_registry( &self, registry: Option<&CustomElementRegistry>, no_gc: &NoGC, )

pub(crate) fn custom_element_registry( &self, ) -> Option<DomRoot<CustomElementRegistry>>

pub(crate) fn attrs(&self) -> &AttributeStorage

pub(crate) fn dom_attrs(&self, cx: &mut JSContext) -> &AttributeStorage

pub(crate) fn locate_namespace(&self, prefix: Option<DOMString>) -> Namespace

Element branch of https://dom.spec.whatwg.org/#locate-a-namespace

pub(crate) fn name_attribute(&self) -> Option<Atom>

pub(crate) fn style_attribute( &self, ) -> &DomRefCell<Option<ServoArc<Locked<PropertyDeclarationBlock>>>>

pub(crate) fn summarize(&self) -> Vec<AttrInfo>

pub(crate) fn is_void(&self) -> bool

pub(crate) fn root_element(&self) -> DomRoot<Element>

pub(crate) fn lookup_prefix(&self, namespace: Namespace) -> Option<DOMString>

pub(crate) fn is_document_element(&self) -> bool

https://dom.spec.whatwg.org/#document-element

pub(crate) fn is_active_element(&self) -> bool

https://html.spec.whatwg.org/multipage/#dom-document-activeelement

pub(crate) fn is_editing_host(&self) -> bool

pub(crate) fn is_actually_disabled(&self) -> bool

pub(crate) fn push_new_attribute( &self, cx: &mut JSContext, local_name: LocalName, value: AttrValue, name: LocalName, namespace: Namespace, prefix: Option<Prefix>, reason: AttributeMutationReason, )

fn handle_attribute_changes( &self, cx: &mut JSContext, attr: AttrRef<'_>, old_value: Option<&AttrValue>, new_value: Option<&AttrValue>, reason: AttributeMutationReason, )

https://dom.spec.whatwg.org/#handle-attribute-changes

pub(crate) fn change_attribute( &self, cx: &mut JSContext, attr: &Attr, value: AttrValue, )

https://dom.spec.whatwg.org/#concept-element-attributes-change

pub(crate) fn push_attribute( &self, cx: &mut JSContext, attr: &Attr, reason: AttributeMutationReason, )

https://dom.spec.whatwg.org/#concept-element-attributes-append

pub(crate) fn with_attribute<R, F>( &self, namespace: &Namespace, local_name: &LocalName, map_func: F, ) -> Option<R>
where F: FnOnce(AttrRef<'_>) -> R,

pub(crate) fn get_attribute_with_namespace( &self, cx: &mut JSContext, namespace: &Namespace, local_name: &LocalName, ) -> Option<DomRoot<Attr>>

This is the inner logic for: https://dom.spec.whatwg.org/#concept-element-attributes-get-by-namespace

In addition to taking a namespace argument, this version does not require the attribute to be lowercase ASCII, in accordance with the specification.

pub(crate) fn get_attribute_by_name( &self, cx: &mut JSContext, name: DOMString, ) -> Option<DomRoot<Attr>>

https://dom.spec.whatwg.org/#concept-element-attributes-get-by-name

pub(crate) fn set_attribute_from_parser( &self, cx: &mut JSContext, qname: QualName, value: DOMString, )

pub(crate) fn set_attribute( &self, cx: &mut JSContext, name: &LocalName, value: AttrValue, )

pub(crate) fn set_attribute_with_namespace( &self, cx: &mut JSContext, local_name: LocalName, value: AttrValue, name: LocalName, namespace: Namespace, prefix: Option<Prefix>, )

fn set_first_matching_attribute<F>( &self, cx: &mut JSContext, local_name: LocalName, value: AttrValue, name: LocalName, namespace: Namespace, prefix: Option<Prefix>, find: F, )
where F: Fn(AttrRef<'_>) -> bool,

https://dom.spec.whatwg.org/#concept-element-attributes-set-value

pub(crate) fn parse_attribute( &self, namespace: &Namespace, local_name: &LocalName, value: DOMString, ) -> AttrValue

pub(crate) fn remove_attribute( &self, cx: &mut JSContext, namespace: &Namespace, local_name: &LocalName, ) -> Option<DomRoot<Attr>>

pub(crate) fn remove_attribute_by_name( &self, cx: &mut JSContext, name: &LocalName, ) -> Option<DomRoot<Attr>>

fn remove_first_matching_attribute<F>( &self, cx: &mut JSContext, find: F, ) -> Option<DomRoot<Attr>>
where F: Fn(AttrRef<'_>) -> bool,

https://dom.spec.whatwg.org/#concept-element-attributes-remove

pub(crate) fn has_class( &self, name: &Atom, case_sensitivity: CaseSensitivity, ) -> bool

pub(crate) fn has_attribute(&self, local_name: &LocalName) -> bool

pub(crate) fn will_mutate_attr(&self, attr: AttrRef<'_>)

fn update_style_attribute( &self, cx: &mut JSContext, attr: AttrRef<'_>, mutation: AttributeMutation<'_>, )

https://html.spec.whatwg.org/multipage/#the-style-attribute

fn set_attribute_node( &self, cx: &mut JSContext, attr: &Attr, ) -> Fallible<Option<DomRoot<Attr>>>

https://dom.spec.whatwg.org/#concept-element-attributes-set including steps of https://dom.spec.whatwg.org/#concept-element-attributes-replace

pub(crate) fn update_nonce_internal_slot(&self, nonce: String, no_gc: &NoGC)

https://html.spec.whatwg.org/multipage/#nonce-attributes

pub(crate) fn nonce_value(&self) -> String

https://html.spec.whatwg.org/multipage/#nonce-attributes

pub(crate) fn update_nonce_post_connection(&self, cx: &mut JSContext)

https://html.spec.whatwg.org/multipage/#nonce-attributes

pub(crate) fn is_nonceable(&self) -> bool

https://www.w3.org/TR/CSP/#is-element-nonceable

pub(crate) fn insert_adjacent( &self, cx: &mut JSContext, where_: AdjacentPosition, node: &Node, ) -> Fallible<Option<DomRoot<Node>>>

pub(crate) fn scroll( &self, cx: &mut JSContext, x: f64, y: f64, behavior: ScrollBehavior, )

https://drafts.csswg.org/cssom-view/#dom-element-scroll

TODO(stevennovaryo): Need to update the scroll API to follow the spec since it is quite outdated.

pub(crate) fn parse_fragment( &self, markup: DOMString, cx: &mut JSContext, ) -> Fallible<DomRoot<DocumentFragment>>

https://html.spec.whatwg.org/multipage/#fragment-parsing-algorithm-steps

pub(crate) fn fragment_parsing_context( cx: &mut JSContext, owner_doc: &Document, element: Option<&Self>, ) -> DomRoot<Self>

Step 4 of https://html.spec.whatwg.org/multipage/#dom-element-insertadjacenthtml and step 6. of https://html.spec.whatwg.org/multipage/#dom-range-createcontextualfragment

pub(crate) fn is_in_same_home_subtree<T>(&self, other: &T) -> bool
where T: DerivedFrom<Element> + DomObject,

pub(crate) fn get_id(&self) -> Option<Atom>

pub(crate) fn get_name(&self) -> Option<Atom>

pub(crate) fn get_element_internals(&self) -> Option<DomRoot<ElementInternals>>

pub(crate) fn outer_html(&self, cx: &mut JSContext) -> Fallible<DOMString>

pub(crate) fn compute_source_position(&self, line_number: u32) -> SourcePosition

pub(crate) fn explicitly_set_tab_index(&self) -> Option<i32>

pub(crate) fn tab_index(&self) -> i32

https://html.spec.whatwg.org/multipage/#dom-tabindex

fn insert_selector_flags(&self, flags: ElementSelectorFlags)

fn get_selector_flags(&self) -> ElementSelectorFlags

pub(crate) fn needs_preserved_style_attribute_after_change(&self) -> bool

Custom Elements and those with mutation observers sometimes need the old style attribute to persist after modification, because it is sent as an argument to script callbacks. Normally that’s not the case, so this method tries to detect when it’s required so it can be avoided for performance reasons.

pub(crate) fn register_current_id_and_name_attribute(&self, cx: &mut JSContext)

pub(crate) fn unregister_current_id_and_name_attribute( &self, cx: &mut JSContext, )

pub(crate) fn qualified_name(&self) -> Cow<'_, str>

An element’s qualified name is its local name if its namespace prefix is null; otherwise its namespace prefix, followed by “:”, followed by its local name. https://dom.spec.whatwg.org/#concept-element-qualified-name

impl Element

pub(crate) fn client_rect(&self, no_gc: &NoGC) -> Rect<i32, CSSPixel>

pub(crate) fn as_maybe_activatable(&self) -> Option<&dyn Activatable>

pub(crate) fn as_stylesheet_owner(&self) -> Option<&dyn StylesheetOwner>

pub(crate) fn as_maybe_validatable(&self) -> Option<&dyn Validatable>

pub(crate) fn is_invalid(&self, cx: &mut JSContext, needs_update: bool) -> bool

pub(crate) fn is_instance_validatable(&self) -> bool

pub(crate) fn init_state_for_internals(&self)

pub(crate) fn click_in_progress(&self) -> bool

pub(crate) fn set_click_in_progress(&self, click: bool)

pub fn state(&self) -> ElementState

pub(crate) fn set_state(&self, which: ElementState, value: bool)

pub(crate) fn set_active_state(&self, value: bool)

https://html.spec.whatwg.org/multipage/#concept-selector-active

pub(crate) fn focus_state(&self) -> bool

pub(crate) fn set_focus_state(&self, value: bool)

pub(crate) fn set_hover_state(&self, value: bool)

pub(crate) fn enabled_state(&self) -> bool

pub(crate) fn set_enabled_state(&self, value: bool)

pub(crate) fn disabled_state(&self) -> bool

pub(crate) fn set_disabled_state(&self, value: bool)

pub(crate) fn read_write_state(&self) -> bool

pub(crate) fn set_read_write_state(&self, value: bool)

pub(crate) fn set_open_state(&self, value: bool)

pub(crate) fn set_placeholder_shown_state(&self, value: bool)

pub(crate) fn set_modal_state(&self, value: bool)

pub(crate) fn set_target_state(&self, value: bool)

pub(crate) fn set_fullscreen_state(&self, value: bool)

pub(crate) fn is_connected(&self) -> bool

https://dom.spec.whatwg.org/#connected

pub(crate) fn cannot_navigate(&self) -> bool

https://html.spec.whatwg.org/multipage/#cannot-navigate

impl Element

pub(crate) fn check_ancestors_disabled_state_for_form_control(&self)

pub(crate) fn check_parent_disabled_state_for_option(&self)

pub(crate) fn check_disabled_attribute(&self)

pub(crate) fn update_read_write_state_from_readonly_attribute(&self)

impl Element

pub(crate) fn focusable_area_kind(&self, no_gc: &NoGC) -> FocusableAreaKind

https://html.spec.whatwg.org/multipage/#focusable-area

The list of focusable areas at this point in the specification is both incomplete and leaves a lot up to the user agent. In addition, the specifications for “click focusable” and “sequentially focusable” are written in a way that they are subsets of all focusable areas. In order to avoid having to first determine whether an element is a focusable area and then work backwards to figure out what kind it is, this function attempts to classify the different types of focusable areas ahead of time so that the logic is useful for answering both “Is this element a focusable area?” and “Is this element click (or sequentially) focusable.”

pub(crate) fn is_sequentially_focusable(&self, no_gc: &NoGC) -> bool

https://html.spec.whatwg.org/multipage/#sequentially-focusable.

pub(crate) fn is_focusable_area(&self, no_gc: &NoGC) -> bool

https://html.spec.whatwg.org/multipage/#focusable-area

impl Element

pub(crate) fn resolved_display_value(&self) -> Option<DisplayOutside>

pub(crate) fn specified_command_value( &self, command: &CommandName, ) -> Option<DOMString>

https://w3c.github.io/editing/docs/execCommand/#specified-command-value

pub(crate) fn is_modifiable_element(&self) -> bool

https://w3c.github.io/editing/docs/execCommand/#modifiable-element

pub(crate) fn is_indentation_element(&self) -> bool

pub(crate) fn has_empty_style_attribute(&self) -> bool

pub(crate) fn is_non_list_single_line_container(&self) -> bool

https://w3c.github.io/editing/docs/execCommand/#non-list-single-line-container

pub(crate) fn is_simple_modifiable_element(&self) -> bool

https://w3c.github.io/editing/docs/execCommand/#simple-modifiable-element

pub(crate) fn is_simple_indentation_element(&self) -> bool

pub(crate) fn set_the_tag_name( &self, cx: &mut JSContext, new_name: &str, ) -> DomRoot<Node>

https://w3c.github.io/editing/docs/execCommand/#set-the-tag-name

impl Element

pub(crate) fn fullscreen_element_ready_check(&self) -> bool

impl Element

pub(crate) fn is_resettable(&self) -> bool

pub(crate) fn reset(&self, cx: &mut JSContext)

impl Element

fn determine_auto_margins(&self) -> AutoMargins

impl Element

fn is_keyboard_interactable(&self, no_gc: &NoGC) -> bool

https://w3c.github.io/webdriver/#dfn-keyboard-interactable