Struct script::dom::element::TagName

struct TagName {
    ptr: DOMRefCell<Option<LocalName>>,
}

A holder for an element's "tag name", which will be lazily resolved and cached. Should be reset when the document owner changes.

Fields
ptr: DOMRefCell<Option<LocalName>>
Methods

impl TagName

fn new() -> TagName

fn or_init<F>(&self, cb: F) -> LocalName where F: FnOnce() -> LocalName

Retrieve a copy of the current inner value. If it is None, it is initialized with the result of cb first.

fn clear(&self)

Clear the cached tag name, so that it will be re-calculated the next time that or_init() is called.