script::dom

Module eventtarget

Structs

EventListenerEntry 🔒

A listener in a collection of event listeners.

EventListeners 🔒

A mix of potentially uncompiled and compiled event listeners of the same type.

EventTarget

InternalRawUncompiledHandler 🔒

https://html.spec.whatwg.org/multipage/#internal-raw-uncompiled-handler

Enums

CommonEventHandler 🔒

CompiledEventListener 🔒

A representation of an EventListener/EventHandler object that has previously been compiled successfully, if applicable.

EventListenerType 🔒

InlineEventListener 🔒

A representation of an event handler, either compiled or uncompiled raw source, or null.

ListenerPhase 🔒

Constants

_IMPL_DOMOBJECT_FOR_EventTarget 🔒

Statics

CONTENT_EVENT_HANDLER_NAMES 🔒

https://html.spec.whatwg.org/multipage/#event-handler-content-attributes containing the values from https://html.spec.whatwg.org/multipage/#globaleventhandlers and https://html.spec.whatwg.org/multipage/#windoweventhandlers as well as specific attributes for elements

Functions

get_compiled_handler 🔒

Get a compiled representation of this event handler, compiling it from its raw source if necessary. https://html.spec.whatwg.org/multipage/#getting-the-current-value-of-the-event-handler