Enum script::dom::eventtarget::InlineEventListener

pub enum InlineEventListener {
    Uncompiled(InternalRawUncompiledHandler),
    Compiled(CommonEventHandler),
    Null,
}

A representation of an event handler, either compiled or uncompiled raw source, or null.

Variants

Uncompiled

Compiled

Null

Methods

impl InlineEventListener

fn get_compiled_handler(&mut self, owner: &EventTarget, ty: &Atom) -> Option<CommonEventHandler>

Get a compiled representation of this event handler, compiling it from its raw source if necessary. https://html.spec.whatwg.org/multipage/#getting-the-current-value-of-the-event-handler