script::dom::eventtarget

Struct EventTarget

#[repr(C)]
pub struct EventTarget {
    reflector_: Reflector,
    handlers: DomRefCell<HashMapTracedValues<Atom, EventListeners, FxBuildHasher>>,
}

Fields
reflector_: Reflectorhandlers: DomRefCell<HashMapTracedValues<Atom, EventListeners, FxBuildHasher>>
Implementations

impl EventTarget

fn __assert_parent_type(&self)

impl EventTarget

pub(crate) fn type_id(&self) -> &'static EventTargetTypeId

impl EventTarget

pub(crate) fn new_inherited() -> EventTarget

fn new( global: &GlobalScope, proto: Option<HandleObject<'_>>, can_gc: CanGc, ) -> DomRoot<EventTarget>

pub(crate) fn has_listeners_for(&self, type_: &Atom) -> bool

Determine if there are any listeners for a given event type. See https://github.com/whatwg/dom/issues/453.

pub(crate) fn get_listeners_for(&self, type_: &Atom) -> EventListeners

pub(crate) fn dispatch_event(&self, event: &Event, can_gc: CanGc) -> bool

pub(crate) fn remove_all_listeners(&self)

fn default_passive_value(&self, ty: &Atom) -> bool

https://dom.spec.whatwg.org/#default-passive-value

fn set_inline_event_listener( &self, ty: Atom, listener: Option<InlineEventListener>, )

https://html.spec.whatwg.org/multipage/#event-handler-attributes:event-handlers-11

pub(crate) fn remove_listener( &self, ty: &Atom, entry: &Rc<RefCell<EventListenerEntry>>, )

pub(crate) fn is_passive( &self, listener: &Rc<RefCell<EventListenerEntry>>, ) -> bool

Determines the passive attribute of an associated event listener

fn get_inline_event_listener( &self, ty: &Atom, can_gc: CanGc, ) -> Option<CommonEventHandler>

pub(crate) fn set_event_handler_uncompiled( &self, url: ServoUrl, line: usize, ty: &str, source: &str, )

Store the raw uncompiled event handler for on-demand compilation later. https://html.spec.whatwg.org/multipage/#event-handler-attributes:event-handler-content-attributes-3

fn get_compiled_event_handler( &self, handler: InternalRawUncompiledHandler, ty: &Atom, can_gc: CanGc, ) -> Option<CommonEventHandler>

pub(crate) fn set_event_handler_common<T: CallbackContainer<DomTypeHolder>>( &self, ty: &str, listener: Option<Rc<T>>, )

pub(crate) fn set_error_event_handler<T: CallbackContainer<DomTypeHolder>>( &self, ty: &str, listener: Option<Rc<T>>, )

pub(crate) fn set_beforeunload_event_handler<T: CallbackContainer<DomTypeHolder>>( &self, ty: &str, listener: Option<Rc<T>>, )

pub(crate) fn get_event_handler_common<T: CallbackContainer<DomTypeHolder>>( &self, ty: &str, can_gc: CanGc, ) -> Option<Rc<T>>

pub(crate) fn has_handlers(&self) -> bool

pub(crate) fn fire_event(&self, name: Atom, can_gc: CanGc) -> bool

pub(crate) fn fire_bubbling_event(&self, name: Atom, can_gc: CanGc) -> bool

pub(crate) fn fire_cancelable_event(&self, name: Atom, can_gc: CanGc) -> bool

pub(crate) fn fire_bubbling_cancelable_event( &self, name: Atom, can_gc: CanGc, ) -> bool

pub(crate) fn fire_event_with_params( &self, name: Atom, bubbles: EventBubbles, cancelable: EventCancelable, composed: EventComposed, can_gc: CanGc, ) -> bool

https://dom.spec.whatwg.org/#concept-event-fire

pub(crate) fn add_event_listener( &self, ty: DOMString, listener: Option<Rc<EventListener<DomTypeHolder>>>, options: AddEventListenerOptions<DomTypeHolder>, )

https://dom.spec.whatwg.org/#dom-eventtarget-addeventlistener and https://dom.spec.whatwg.org/#add-an-event-listener

pub(crate) fn remove_event_listener( &self, ty: DOMString, listener: &Option<Rc<EventListener<DomTypeHolder>>>, options: &EventListenerOptions, )

https://dom.spec.whatwg.org/#dom-eventtarget-removeeventlistener and https://dom.spec.whatwg.org/#remove-an-event-listener

pub(crate) fn get_the_parent( &self, event: &Event, ) -> Option<DomRoot<EventTarget>>

https://dom.spec.whatwg.org/#get-the-parent

pub(crate) fn retarget(&self, b: &Self) -> DomRoot<EventTarget>

https://dom.spec.whatwg.org/#retarget

pub(crate) fn is_content_event_handler(name: &str) -> bool

https://html.spec.whatwg.org/multipage/#event-handler-content-attributes