script::dom::execcommand::commands

Module fontsize

Enums

ParsingMode 🔒

Functions

css_font_to_font_size 🔒

execute_fontsize_command 🔒

https://w3c.github.io/editing/docs/execCommand/#the-fontsize-command

font_size_loosely_equivalent 🔒

Handles fontsize command part of https://w3c.github.io/editing/docs/execCommand/#loosely-equivalent-values

font_size_to_css_font 🔒

legacy_font_size_for 🔒

https://w3c.github.io/editing/docs/execCommand/#legacy-font-size-for

maybe_normalize_pixels 🔒

Only in the case we have resolved to actual pixels, we need to do its conversion. In other cases, we already have the relevant font size or corresponding css value. This avoids expensive conversions of pixels to other values.

value_for_fontsize_command 🔒

https://w3c.github.io/editing/docs/execCommand/#the-fontsize-command