script::dom::fetch::response

Struct Response

#[repr(C)]
pub(crate) struct Response {
    reflector_: Reflector,
    headers_reflector: MutNullableDom<Headers>,
    status: DomRefCell<HttpStatus>,
    response_type: DomRefCell<ResponseType>,
    url: DomRefCell<Option<ServoUrl>>,
    url_list: DomRefCell<Vec<ServoUrl>>,
    body_stream: MutNullableDom<ReadableStream>,
    fetch_body_stream: MutNullableDom<ReadableStream>,
    stream_consumer: DomRefCell<Option<StreamConsumer>>,
    redirected: Cell<bool>,
}

Fields
reflector_: Reflectorheaders_reflector: MutNullableDom<Headers>status: DomRefCell<HttpStatus>response_type: DomRefCell<ResponseType>url: DomRefCell<Option<ServoUrl>>

https://fetch.spec.whatwg.org/#concept-response-url FIXME: This should always point to the last entry of url_list. Somehow we directly set it instead.

url_list: DomRefCell<Vec<ServoUrl>>body_stream: MutNullableDom<ReadableStream>

The stream of https://fetch.spec.whatwg.org/#body.

fetch_body_stream: MutNullableDom<ReadableStream>

The stream that receives network delivered bytes for Fetch responses. This must remain stable even if body_stream is replaced by tee() branches during clone().

stream_consumer: DomRefCell<Option<StreamConsumer>>redirected: Cell<bool>

FIXME: This should be removed.

Implementations

impl Response

fn __assert_parent_type(&self)

impl Response

pub(crate) fn new_inherited(stream: DomRoot<ReadableStream>) -> Response

pub(crate) fn new(cx: &mut JSContext, global: &GlobalScope) -> DomRoot<Response>

https://fetch.spec.whatwg.org/#dom-response

fn new_with_proto( cx: &mut JSContext, global: &GlobalScope, proto: Option<HandleObject<'_>>, ) -> DomRoot<Response>

pub(crate) fn error_stream(&self, cx: &mut JSContext, error: Error)

pub(crate) fn is_disturbed(&self) -> bool

pub(crate) fn is_locked(&self) -> bool

impl Response

pub(crate) fn set_type( &self, cx: &mut JSContext, new_response_type: DOMResponseType, )

pub(crate) fn set_headers( &self, cx: &mut JSContext, option_hyper_headers: Option<Serde<HyperHeaders>>, )

pub(crate) fn set_status(&self, status: &HttpStatus)

pub(crate) fn set_final_url(&self, final_url: ServoUrl)

pub(crate) fn set_redirected(&self, is_redirected: bool)

fn set_response_members_by_type( &self, cx: &mut JSContext, response_type: DOMResponseType, )

pub(crate) fn set_stream_consumer(&self, sc: Option<StreamConsumer>)

pub(crate) fn stream_chunk(&self, cx: &mut JSContext, chunk: Bytes)

pub(crate) fn finish(&self, cx: &mut JSContext)