script::dom::file::filereader

Struct FileReader

#[repr(C)]
pub(crate) struct FileReader {
    eventtarget: EventTarget,
    ready_state: Cell<FileReaderReadyState>,
    error: MutNullableDom<DOMException>,
    result: DomRefCell<Option<FileReaderResult>>,
    generation_id: Cell<GenerationId>,
}

Fields
eventtarget: EventTargetready_state: Cell<FileReaderReadyState>error: MutNullableDom<DOMException>result: DomRefCell<Option<FileReaderResult>>generation_id: Cell<GenerationId>
Implementations

impl FileReader

fn __assert_parent_type(&self)

impl FileReader

pub(crate) fn new_inherited() -> FileReader

fn new( cx: &mut JSContext, global: &GlobalScope, proto: Option<HandleObject<'_>>, ) -> DomRoot<FileReader>

pub(crate) fn process_read_error( cx: &mut JSContext, filereader: Trusted<FileReader>, gen_id: GenerationId, error: DOMErrorName, )

pub(crate) fn process_read_data( cx: &mut JSContext, filereader: Trusted<FileReader>, gen_id: GenerationId, )

pub(crate) fn process_read( cx: &mut JSContext, filereader: Trusted<FileReader>, gen_id: GenerationId, )

pub(crate) fn process_read_eof( cx: &mut JSContext, filereader: Trusted<FileReader>, gen_id: GenerationId, data: ReadMetaData, blob_contents: Vec<u8>, )

fn perform_readastext( result: &DomRefCell<Option<FileReaderResult>>, data: ReadMetaData, blob_bytes: &[u8], )

https://w3c.github.io/FileAPI/#packaging-data

fn perform_readasdataurl( result: &DomRefCell<Option<FileReaderResult>>, data: ReadMetaData, bytes: &[u8], )

https://w3c.github.io/FileAPI/#packaging-data

fn perform_readasbinarystring( result: &DomRefCell<Option<FileReaderResult>>, bytes: &[u8], )

https://w3c.github.io/FileAPI/#packaging-data

Return bytes as a binary string, in which every byte is represented by a code unit of equal value [0..255].

fn perform_readasarraybuffer( cx: &mut JSContext, result: &DomRefCell<Option<FileReaderResult>>, bytes: &[u8], )

https://w3c.github.io/FileAPI/#packaging-data

Return a new ArrayBuffer whose contents are bytes.

impl FileReader

fn dispatch_progress_event( &self, cx: &mut JSContext, type_: Atom, loaded: u64, total: Option<u64>, )

fn terminate_ongoing_reading(&self)

fn read( &self, cx: &mut JSContext, function: FileReaderFunction, blob: &Blob, encoding: Option<DOMString>, ) -> ErrorResult

https://w3c.github.io/FileAPI/#readOperation

fn change_ready_state(&self, state: FileReaderReadyState)