script::dom::gamepad::gamepad

Struct Gamepad

#[repr(C)]
pub(crate) struct Gamepad {

    reflector_: Reflector,
    gamepad_id: u32,
    id: String,
    index: Cell<i32>,
    connected: Cell<bool>,
    timestamp: Cell<f64>,
    mapping_type: String,
    frozen_buttons: CachedFrozenArray,
    buttons: Vec<Dom<GamepadButton>>,
    frozen_axes: CachedFrozenArray,
    axes: RefCell<Vec<f64>>,
    pose: Option<Dom<GamepadPose>>,
    hand: GamepadHand,
    axis_bounds: (f64, f64),
    button_bounds: (f64, f64),
    exposed: Cell<bool>,
    vibration_actuator: Dom<GamepadHapticActuator>,

}

Fields
reflector_: Reflectorgamepad_id: u32id: Stringindex: Cell<i32>connected: Cell<bool>timestamp: Cell<f64>mapping_type: Stringfrozen_buttons: CachedFrozenArraybuttons: Vec<Dom<GamepadButton>>frozen_axes: CachedFrozenArrayaxes: RefCell<Vec<f64>>pose: Option<Dom<GamepadPose>>hand: GamepadHandaxis_bounds: (f64, f64)button_bounds: (f64, f64)exposed: Cell<bool>vibration_actuator: Dom<GamepadHapticActuator>
Implementations

impl Gamepad

fn __assert_parent_type(&self)

impl Gamepad

fn new_inherited( gamepad_id: u32, id: String, index: i32, connected: bool, timestamp: f64, mapping_type: String, buttons: &[&GamepadButton], pose: Option<&GamepadPose>, hand: GamepadHand, axis_bounds: (f64, f64), button_bounds: (f64, f64), vibration_actuator: &GamepadHapticActuator, ) -> Gamepad

pub(crate) fn new( cx: &mut JSContext, window: &Window, gamepad_id: u32, id: String, mapping_type: String, axis_bounds: (f64, f64), button_bounds: (f64, f64), supported_haptic_effects: GamepadSupportedHapticEffects, xr: bool, ) -> DomRoot<Gamepad>

From: https://www.w3.org/TR/gamepad/#fingerprinting-mitigation The user agent MAY alter the device information exposed through the API to reduce the fingerprinting surface. As an example, an implementation can require that a Gamepad object have exactly the number of buttons and axes defined in the Standard Gamepad layout even if more or fewer inputs are present on the connected device.

So, we initialize the number of buttons and axes corresponding to the “standard” gamepad mapping and happens to simplify implementation. https://www.w3.org/TR/gamepad/#dfn-a-new-gamepad

impl Gamepad

pub(crate) fn gamepad_id(&self) -> u32

fn init_buttons( cx: &mut JSContext, window: &Window, ) -> Vec<DomRoot<GamepadButton>>

Initialize the standard buttons for a gamepad. https://www.w3.org/TR/gamepad/#dfn-initializing-buttons

pub(crate) fn update_connected(&self, connected: bool)

pub(crate) fn index(&self) -> i32

pub(crate) fn update_index(&self, index: i32)

pub(crate) fn update_timestamp(&self, timestamp: f64)

pub(crate) fn notify_event( &self, cx: &mut JSContext, event_type: GamepadEventType, )

fn init_axes(&self)

Initialize the number of axes in the “standard” gamepad mapping. https://www.w3.org/TR/gamepad/#dfn-initializing-axes

pub(crate) fn map_and_normalize_axes(&self, axis_index: usize, value: f64)

https://www.w3.org/TR/gamepad/#dfn-map-and-normalize-axes

pub(crate) fn map_and_normalize_buttons(&self, button_index: usize, value: f64)

https://www.w3.org/TR/gamepad/#dfn-map-and-normalize-buttons

pub(crate) fn connected(&self) -> bool

pub(crate) fn exposed(&self) -> bool

https://www.w3.org/TR/gamepad/#dfn-exposed

pub(crate) fn set_exposed(&self, exposed: bool)

https://www.w3.org/TR/gamepad/#dfn-exposed

pub(crate) fn vibration_actuator(&self) -> &GamepadHapticActuator