script::dom

Module headers

Structs

Headers 🔒

Enums

Guard 🔒

https://fetch.spec.whatwg.org/#concept-headers-guard

Constants

_IMPL_DOMOBJECT_FOR_Headers 🔒

Functions

is_field_name 🔒

http://tools.ietf.org/html/rfc7230#section-3.2

is_forbidden_request_header 🔒

This function will internally convert name to lowercase for matching, so explicitly converting before calling is not necessary

is_forbidden_response_header 🔒

https://fetch.spec.whatwg.org/#forbidden-response-header-name

is_legal_header_value 🔒

https://fetch.spec.whatg.org/#concept-header-value

is_obs_text 🔒

http://tools.ietf.org/html/rfc7230#section-3.2.6

is_vchar 🔒

https://tools.ietf.org/html/rfc5234#appendix-B.1

validate_name 🔒