script::dom::html::document_metadata::htmllinkelement

Struct HTMLLinkElement

#[repr(C)]
pub(crate) struct HTMLLinkElement {

    htmlelement: HTMLElement,
    rel_list: MutNullableDom<DOMTokenList>,
    relations: Cell<LinkRelations>,
    stylesheet: DomRefCell<Option<Arc<Stylesheet>>>,
    cssom_stylesheet: MutNullableDom<CSSStyleSheet>,
    parser_inserted: Cell<bool>,
    pending_loads: Cell<u32>,
    any_failed_load: Cell<bool>,
    request_generation_id: Cell<RequestGenerationId>,
    is_explicitly_enabled: Cell<bool>,
    previous_type_matched: Cell<bool>,
    previous_media_environment_matched: Cell<bool>,
    line_number: u64,
    source_set: DomRefCell<SourceSet>,
    blocking: MutNullableDom<DOMTokenList>,

}

Fields
htmlelement: HTMLElementrel_list: MutNullableDom<DOMTokenList>

The relations as specified by the “rel” attribute

relations: Cell<LinkRelations>

The link relations as they are used in practice.

The reason this is separate from HTMLLinkElement::rel_list is that a literal list is a bit unwieldy and that there are corner cases to consider (Like rev="made" implying an author relationship that is not represented in rel_list)

stylesheet: DomRefCell<Option<Arc<Stylesheet>>>cssom_stylesheet: MutNullableDom<CSSStyleSheet>parser_inserted: Cell<bool>

https://html.spec.whatwg.org/multipage/#a-style-sheet-that-is-blocking-scripts

pending_loads: Cell<u32>

The number of loads that this link element has triggered (could be more than one because of imports) and have not yet finished.

any_failed_load: Cell<bool>

Whether any of the loads have failed.

request_generation_id: Cell<RequestGenerationId>

A monotonically increasing counter that keeps track of which stylesheet to apply.

is_explicitly_enabled: Cell<bool>

https://html.spec.whatwg.org/multipage/#explicitly-enabled

previous_type_matched: Cell<bool>

Whether the previous type matched with the destination

previous_media_environment_matched: Cell<bool>

Whether the previous media environment matched with the media query

line_number: u64

Line number this element was created on

source_set: DomRefCell<SourceSet>

https://html.spec.whatwg.org/multipage/#source-set

blocking: MutNullableDom<DOMTokenList>

https://html.spec.whatwg.org/multipage/#dom-link-blocking

Implementations

impl HTMLLinkElement

fn __assert_parent_type(&self)

impl HTMLLinkElement

fn new_inherited( local_name: LocalName, prefix: Option<Prefix>, document: &Document, creator: ElementCreator, ) -> HTMLLinkElement

pub(crate) fn new( cx: &mut JSContext, local_name: LocalName, prefix: Option<Prefix>, document: &Document, proto: Option<HandleObject<'_>>, creator: ElementCreator, ) -> DomRoot<HTMLLinkElement>

pub(crate) fn get_request_generation_id(&self) -> RequestGenerationId

fn remove_stylesheet(&self, no_gc: &NoGC)

pub(crate) fn set_stylesheet( &self, no_gc: &NoGC, new_stylesheet: Arc<Stylesheet>, )

pub(crate) fn get_stylesheet(&self) -> Option<Arc<Stylesheet>>

pub(crate) fn get_cssom_stylesheet( &self, cx: &mut JSContext, ) -> Option<DomRoot<CSSStyleSheet>>

pub(crate) fn is_alternate(&self) -> bool

pub(crate) fn is_effectively_disabled(&self) -> bool

fn clean_stylesheet_ownership(&self)

impl HTMLLinkElement

fn compute_destination_for_attribute(&self) -> Option<Destination>

fn processing_options(&self) -> LinkProcessingOptions

https://html.spec.whatwg.org/multipage/#create-link-options-from-element

fn default_fetch_and_process_the_linked_resource( &self, ) -> Option<RequestBuilder>

https://html.spec.whatwg.org/multipage/#default-fetch-and-process-the-linked-resource

This method does not implement Step 7 (fetching the request) and instead returns the RequestBuilder, as the fetch context that should be used depends on the link type.

fn linked_resource_fetch_setup(&self, request: &mut RequestBuilder) -> bool

https://html.spec.whatwg.org/multipage/#linked-resource-fetch-setup-steps

fn fetch_and_process_prefetch_link(&self, href: &str)

The fetch and process the linked resource algorithm for rel="prefetch"

fn handle_stylesheet_url(&self, cx: &mut JSContext)

https://html.spec.whatwg.org/multipage/#concept-link-obtain

fn handle_disabled_attribute_change(&self, no_gc: &NoGC, is_removal: bool)

https://html.spec.whatwg.org/multipage/#attr-link-disabled

fn handle_favicon_url(&self, href: &str)

fn register_image_cache_callback( &self, id: PendingImageId, ) -> ImageCacheResponseCallback

fn process_favicon_response(&self, image: Image)

Rasterizes a loaded favicon file if necessary and notifies the embedder about it.

fn handle_preload_url(&self)

https://html.spec.whatwg.org/multipage/#link-type-preload:fetch-and-process-the-linked-resource-2 and type matching destination steps of https://html.spec.whatwg.org/multipage/#preload

pub(crate) fn fire_event_after_response( &self, cx: &mut JSContext, response: Result<(), NetworkError>, )

https://html.spec.whatwg.org/multipage/#link-type-preload:fetch-and-process-the-linked-resource-2

fn fetch_and_process_modulepreload(&self, cx: &mut JSContext)

https://html.spec.whatwg.org/multipage/#link-type-modulepreload:fetch-and-process-the-linked-resource-2