script::dom::html::document_metadata::htmlmetaelement

Trait IsColorSchemeValue

trait IsColorSchemeValue {
    // Required method
    fn is_color_scheme(&self) -> bool;
}

Trait to make it easier to make sure all callers lowercase to ASCII before comparing to the color-scheme value. Otherwise it is easy to miss one usage and compare case-sensitively.

Required Methods

fn is_color_scheme(&self) -> bool