script::dom::html::embedded_content::htmlcanvaselement

Struct HTMLCanvasElement

#[repr(C)]
pub(crate) struct HTMLCanvasElement {
    htmlelement: HTMLElement,
    context_mode: DomRefCell<Option<RenderingContext>>,
    callback_id: Cell<u32>,
    blob_callbacks: RefCell<FxHashMap<u32, Rc<BlobCallback<DomTypeHolder>>>>,
    image_key: Cell<Option<ImageKey>>,
}

https://html.spec.whatwg.org/multipage/#htmlcanvaselement

Fields
htmlelement: HTMLElementcontext_mode: DomRefCell<Option<RenderingContext>>

https://html.spec.whatwg.org/multipage/#concept-canvas-context-mode

callback_id: Cell<u32>

This id along with Self::blob_callbacks are used to keep track of ongoing toBlob() calls.

blob_callbacks: RefCell<FxHashMap<u32, Rc<BlobCallback<DomTypeHolder>>>>

This hashmap along with Self::callback_id are used to keep track of ongoing toBlob() calls.

image_key: Cell<Option<ImageKey>>

The ImageKey used to render this HTMLCanvasElement to the WebRender scene, if it has a RenderingContext, otherwise None. Note that this key is owned by the RenderingContext itself which will take care of cleaning it up.

Implementations

impl HTMLCanvasElement

fn __assert_parent_type(&self)

impl HTMLCanvasElement

fn new_inherited( local_name: LocalName, prefix: Option<Prefix>, document: &Document, ) -> HTMLCanvasElement

pub(crate) fn new( cx: &mut JSContext, local_name: LocalName, prefix: Option<Prefix>, document: &Document, proto: Option<HandleObject<'_>>, ) -> DomRoot<HTMLCanvasElement>

fn recreate_contexts_after_resize(&self)

pub(crate) fn get_size(&self) -> Size2D<u32>

pub(crate) fn origin_is_clean(&self) -> bool

pub(crate) fn mark_as_dirty(&self)

pub(crate) fn set_natural_width(&self, cx: &mut JSContext, value: u32)

pub(crate) fn set_natural_height(&self, cx: &mut JSContext, value: u32)

impl HTMLCanvasElement

pub(crate) fn context(&self) -> Option<Ref<'_, RenderingContext>>

fn set_rendering_context( &self, no_gc: &NoGC, make_rendering_context: impl FnOnce() -> RenderingContext, )

fn get_or_init_2d_context( &self, cx: &mut JSContext, ) -> Option<DomRoot<CanvasRenderingContext2D>>

fn get_or_init_bitmaprenderer_context( &self, cx: &mut JSContext, ) -> Option<DomRoot<ImageBitmapRenderingContext>>

https://html.spec.whatwg.org/multipage/#canvas-context-bitmaprenderer

fn get_or_init_webgl_context( &self, cx: &mut JSContext, options: HandleValue<'_>, ) -> Option<DomRoot<WebGLRenderingContext>>

fn get_or_init_webgl2_context( &self, cx: &mut JSContext, options: HandleValue<'_>, ) -> Option<DomRoot<WebGL2RenderingContext>>

fn get_or_init_webgpu_context( &self, cx: &mut JSContext, ) -> Option<DomRoot<GPUCanvasContext>>

fn get_gl_attributes( cx: &mut JSContext, options: HandleValue<'_>, ) -> Option<GLContextAttributes>

pub(crate) fn is_valid(&self) -> bool

pub(crate) fn get_image_data(&self) -> Option<Snapshot>

fn maybe_quality(quality: HandleValue<'_>) -> Option<f64>

pub(crate) fn update_rendering(&self, epoch: Epoch) -> Option<ImageKey>