script::dom::html::embedded_content::htmliframeelement

Struct HTMLIFrameElement

#[repr(C)]
pub(crate) struct HTMLIFrameElement {

    htmlelement: HTMLElement,
    webview_id: Cell<Option<WebViewId>>,
    browsing_context_id: Cell<Option<BrowsingContextId>>,
    pipeline_id: Cell<Option<PipelineId>>,
    pending_pipeline_id: Cell<Option<PipelineId>>,
    about_blank_pipeline_id: Cell<Option<PipelineId>>,
    sandbox: MutNullableDom<DOMTokenList>,
    sandboxing_flag_set: Cell<Option<SandboxingFlagSet>>,
    load_blocker: DomRefCell<Option<LoadBlocker>>,
    throttled: Cell<bool>,
    script_window_proxies: Rc<ScriptWindowProxies>,
    current_navigation_was_lazy_loaded: Cell<bool>,
    lazy_load_resumption_steps: Cell<LazyLoadResumptionSteps>,
    pending_navigation: Cell<bool>,
    frozen_name: DomRefCell<Option<String>>,

}

Fields
htmlelement: HTMLElementwebview_id: Cell<Option<WebViewId>>browsing_context_id: Cell<Option<BrowsingContextId>>pipeline_id: Cell<Option<PipelineId>>pending_pipeline_id: Cell<Option<PipelineId>>about_blank_pipeline_id: Cell<Option<PipelineId>>sandbox: MutNullableDom<DOMTokenList>sandboxing_flag_set: Cell<Option<SandboxingFlagSet>>load_blocker: DomRefCell<Option<LoadBlocker>>throttled: Cell<bool>script_window_proxies: Rc<ScriptWindowProxies>current_navigation_was_lazy_loaded: Cell<bool>

https://html.spec.whatwg.org/multipage/#current-navigation-was-lazy-loaded

lazy_load_resumption_steps: Cell<LazyLoadResumptionSteps>

https://html.spec.whatwg.org/multipage/#lazy-load-resumption-steps

pending_navigation: Cell<bool>

Keeping track of whether the iframe will be navigated outside of the processing of it’s attribute(for example: form navigation). This is necessary to prevent the iframe load event steps from asynchronously running for the initial blank document while script at this point(when the flag is set) expects those to run only for the navigated documented.

frozen_name: DomRefCell<Option<String>>

Initial name set by the content of the name attribute on the iframe. This is frozen in time, since it is only processed once on the initial creation of the iframe contents. If the iframe itself changes the window.name, that takes precedence.

Implementations

impl HTMLIFrameElement

fn __assert_parent_type(&self)

impl HTMLIFrameElement

fn shared_attribute_processing_steps_for_iframe_and_frame_elements( &self, _mode: ProcessingMode, ) -> Option<ServoUrl>

https://html.spec.whatwg.org/multipage/#shared-attribute-processing-steps-for-iframe-and-frame-elements,

pub(crate) fn navigate_or_reload_child_browsing_context( &self, load_data: LoadData, history_handling: NavigationHistoryBehavior, mode: ProcessingMode, target_snapshot_params: TargetSnapshotParams, cx: &mut JSContext, )

fn start_new_pipeline( &self, cx: &mut JSContext, load_data: LoadData, pipeline_type: PipelineType, history_handling: NavigationHistoryBehavior, mode: ProcessingMode, target_snapshot_params: TargetSnapshotParams, )

fn continue_navigation( &self, cx: &mut JSContext, load_data: LoadData, pipeline_type: PipelineType, history_handling: NavigationHistoryBehavior, target_snapshot_params: TargetSnapshotParams, )

pub(crate) fn is_initial_blank_document(&self) -> bool

When an iframe is first inserted into the document, an “about:blank” document is created, and synchronously processed by the script thread. This initial synchronous load should have no noticeable effect in script. See the note in iframe_load_event_steps.

fn navigate_an_iframe_or_frame( &self, cx: &mut JSContext, load_data: LoadData, mode: ProcessingMode, )

https://html.spec.whatwg.org/multipage/#navigate-an-iframe-or-frame

fn will_lazy_load_element_steps(&self) -> bool

https://html.spec.whatwg.org/multipage/#will-lazy-load-element-steps

fn navigate_to_the_srcdoc_resource( &self, mode: ProcessingMode, cx: &mut JSContext, )

Step 1.3. of https://html.spec.whatwg.org/multipage/#process-the-iframe-attributes

fn mark_navigation_as_lazy_loaded(&self, cx: &mut JSContext)

https://html.spec.whatwg.org/multipage/#the-iframe-element:potentially-delays-the-load-event

fn process_the_iframe_attributes( &self, mode: ProcessingMode, cx: &mut JSContext, )

https://html.spec.whatwg.org/multipage/#process-the-iframe-attributes

fn create_nested_browsing_context(&self, cx: &mut JSContext)

https://html.spec.whatwg.org/multipage/#create-a-new-child-navigable Synchronously create a new browsing context; this is not a navigation.

fn destroy_nested_browsing_context(&self)

pub(crate) fn update_pipeline_id( &self, new_pipeline_id: PipelineId, reason: UpdatePipelineIdReason, cx: &mut JSContext, ) -> bool

Returns true if the contained pipeline was updated, false otherwise. This can occur if the iframe’s nested browsing context has changed since the asynchronous update was started.

fn new_inherited( local_name: LocalName, prefix: Option<Prefix>, document: &Document, ) -> HTMLIFrameElement

pub(crate) fn new( cx: &mut JSContext, local_name: LocalName, prefix: Option<Prefix>, document: &Document, proto: Option<HandleObject<'_>>, ) -> DomRoot<HTMLIFrameElement>

pub(crate) fn pipeline_id(&self) -> Option<PipelineId>

pub(crate) fn browsing_context_id(&self) -> Option<BrowsingContextId>

pub(crate) fn webview_id(&self) -> Option<WebViewId>

pub(crate) fn sandboxing_flag_set(&self) -> SandboxingFlagSet

pub(crate) fn set_throttled(&self, throttled: bool)

pub(crate) fn note_pending_navigation(&self)

Note a pending navigation. This is used to ignore the async load event steps for the initial blank document if those haven’t run yet.

pub(crate) fn iframe_load_event_steps( &self, loaded_pipeline: PipelineId, cx: &mut JSContext, )

https://html.spec.whatwg.org/multipage/#iframe-load-event-steps

pub(crate) fn run_iframe_load_event_steps(&self, cx: &mut JSContext)

https://html.spec.whatwg.org/multipage/#iframe-load-event-steps

fn parse_sandbox_attribute(&self)

Parse the sandbox attribute value given the [Attr]. This sets the sandboxing_flag_set property or clears it is the value isn’t specified. Notably, an unspecified sandboxing attribute (no sandboxing) is different from an empty one (full sandboxing).

pub(crate) fn destroy_document_and_its_descendants(&self, cx: &mut JSContext)

Step 4.2. of https://html.spec.whatwg.org/multipage/#destroy-a-document-and-its-descendants

fn destroy_child_navigable(&self, cx: &mut JSContext)

https://html.spec.whatwg.org/multipage/#destroy-a-child-navigable