script::dom::html::embedded_content

Module htmlmediaelement

Structs

BufferedDataSource 🔒

FrameHolder 🔒

HTMLMediaElement 🔒

HTMLMediaElementEventHandler 🔒

The HTMLMediaElementEventHandler is a structure responsible for handling media events for the HTMLMediaElement and exists to decouple ownership of the HTMLMediaElement from IPC router callback.

HTMLMediaElementFetchContext 🔒

HTMLMediaElementFetchListener 🔒

MediaFrameRenderer 🔒

SourceChildrenPointer 🔒

https://html.spec.whatwg.org/multipage/#loading-the-media-resource:media-element-29

Enums

CancelReason 🔒

Indicates the reason why a fetch request was cancelled.

DataBuffer 🔒

FrameStatus 🔒

LoadState 🔒

LoopCondition 🔒

Generally the presence of the loop attribute should be considered to mean playback has not “ended”, as “ended” and “looping” are mutually exclusive. https://html.spec.whatwg.org/multipage/#ended-playback

MediaElementMicrotask 🔒

NetworkState 🔒

https://html.spec.whatwg.org/multipage/#dom-media-networkstate

PlaybackDirection 🔒

https://html.spec.whatwg.org/multipage/#direction-of-playback

PlaybackPositionWasMoved 🔒

Used to determine whether the current playback position was changed during normal playback or not in https://html.spec.whatwg.org/multipage/#time-marches-on

ReadyState 🔒

https://html.spec.whatwg.org/multipage/#dom-media-readystate

Resource 🔒

SrcObject 🔒

Constants

SEEK_POSITION_THRESHOLD 🔒

The media engine may report a seek-done position that differs slightly from the requested position (e.g. snapping to the nearest keyframe), so we use a threshold instead of strict equality. (Unit is second)

_IMPL_DOMOBJECT_FOR_HTMLMediaElement 🔒

Statics

MEDIA_CONTROL_CSS 🔒

A CSS file to style the media controls.

MEDIA_CONTROL_JS 🔒

A JS file to control the media controls.