script::dom::html::embedded_content::htmlmediaelement

Struct HTMLMediaElement

#[repr(C)]
pub(crate) struct HTMLMediaElement {

    htmlelement: HTMLElement,
    network_state: Cell<NetworkState>,
    ready_state: Cell<ReadyState>,
    src_object: DomRefCell<Option<SrcObject>>,
    current_src: DomRefCell<String>,
    generation_id: Cell<u32>,
    fired_loadeddata_event: Cell<bool>,
    error: MutNullableDom<MediaError>,
    paused: Cell<bool>,
    default_playback_rate: Cell<f64>,
    playback_rate: Cell<f64>,
    autoplaying: Cell<bool>,
    delaying_the_load_event_flag: DomRefCell<Option<LoadBlocker>>,
    pending_play_promises: DomRefCell<Vec<Rc<Promise>>>,
    in_flight_play_promises_queue: DomRefCell<VecDeque<(Box<[Rc<Promise>]>, ErrorResult)>>,
    player: DomRefCell<Option<Arc<Mutex<dyn Player>>>>,
    video_renderer: Arc<Mutex<MediaFrameRenderer>>,
    audio_renderer: DomRefCell<Option<Arc<Mutex<dyn AudioRenderer>>>>,
    event_handler: RefCell<Option<Arc<Mutex<HTMLMediaElementEventHandler>>>>,
    show_poster: Cell<bool>,
    duration: Cell<f64>,
    current_playback_position: Cell<f64>,
    official_playback_position: Cell<f64>,
    default_playback_start_position: Cell<f64>,
    volume: Cell<f64>,
    seeking: Cell<bool>,
    current_seek_position: Cell<f64>,
    muted: Cell<bool>,
    load_state: Cell<LoadState>,
    source_children_pointer: DomRefCell<Option<SourceChildrenPointer>>,
    current_source_child: MutNullableDom<HTMLSourceElement>,
    resource_url: DomRefCell<Option<ServoUrl>>,
    blob_url: DomRefCell<Option<ServoUrl>>,
    played: DomRefCell<TimeRangesContainer>,
    audio_tracks_list: MutNullableDom<AudioTrackList>,
    video_tracks_list: MutNullableDom<VideoTrackList>,
    text_tracks_list: MutNullableDom<TextTrackList>,
    next_timeupdate_event: Cell<Instant>,
    current_fetch_context: RefCell<Option<HTMLMediaElementFetchContext>>,
    media_controls_id: DomRefCell<Option<String>>,
    did_perform_automatic_track_selection: Cell<bool>,
    position_when_time_marches_on_ran: Cell<Option<f64>>,
    playback_was_moved: Cell<PlaybackPositionWasMoved>,
    newly_introduced_cues: DomRefCell<Vec<Dom<TextTrackCue>>>,

}

Fields
htmlelement: HTMLElementnetwork_state: Cell<NetworkState>

https://html.spec.whatwg.org/multipage/#dom-media-networkstate

ready_state: Cell<ReadyState>

https://html.spec.whatwg.org/multipage/#dom-media-readystate

src_object: DomRefCell<Option<SrcObject>>

https://html.spec.whatwg.org/multipage/#dom-media-srcobject

current_src: DomRefCell<String>

https://html.spec.whatwg.org/multipage/#dom-media-currentsrc

generation_id: Cell<u32>

Incremented whenever tasks associated with this element are cancelled.

fired_loadeddata_event: Cell<bool>

https://html.spec.whatwg.org/multipage/#fire-loadeddata

Reset to false every time the load algorithm is invoked.

error: MutNullableDom<MediaError>

https://html.spec.whatwg.org/multipage/#dom-media-error

paused: Cell<bool>

https://html.spec.whatwg.org/multipage/#dom-media-paused

default_playback_rate: Cell<f64>

https://html.spec.whatwg.org/multipage/#dom-media-defaultplaybackrate

playback_rate: Cell<f64>

https://html.spec.whatwg.org/multipage/#dom-media-playbackrate

autoplaying: Cell<bool>

https://html.spec.whatwg.org/multipage/#attr-media-autoplay

delaying_the_load_event_flag: DomRefCell<Option<LoadBlocker>>

https://html.spec.whatwg.org/multipage/#delaying-the-load-event-flag

pending_play_promises: DomRefCell<Vec<Rc<Promise>>>

https://html.spec.whatwg.org/multipage/#list-of-pending-play-promises

in_flight_play_promises_queue: DomRefCell<VecDeque<(Box<[Rc<Promise>]>, ErrorResult)>>

Play promises which are soon to be fulfilled by a queued task.

player: DomRefCell<Option<Arc<Mutex<dyn Player>>>>video_renderer: Arc<Mutex<MediaFrameRenderer>>audio_renderer: DomRefCell<Option<Arc<Mutex<dyn AudioRenderer>>>>event_handler: RefCell<Option<Arc<Mutex<HTMLMediaElementEventHandler>>>>show_poster: Cell<bool>

https://html.spec.whatwg.org/multipage/#show-poster-flag

duration: Cell<f64>

https://html.spec.whatwg.org/multipage/#dom-media-duration

current_playback_position: Cell<f64>

https://html.spec.whatwg.org/multipage/#current-playback-position

official_playback_position: Cell<f64>

https://html.spec.whatwg.org/multipage/#official-playback-position

default_playback_start_position: Cell<f64>

https://html.spec.whatwg.org/multipage/#default-playback-start-position

volume: Cell<f64>

https://html.spec.whatwg.org/multipage/#dom-media-volume

seeking: Cell<bool>

https://html.spec.whatwg.org/multipage/#dom-media-seeking

current_seek_position: Cell<f64>

The latest seek position (in seconds) is used to distinguish whether the seek request was initiated by a script or by the user agent itself, rather than by the media engine and to abort other running instance of the seek algorithm.

muted: Cell<bool>

https://html.spec.whatwg.org/multipage/#dom-media-muted

load_state: Cell<LoadState>

Loading state from source, if any.

source_children_pointer: DomRefCell<Option<SourceChildrenPointer>>current_source_child: MutNullableDom<HTMLSourceElement>resource_url: DomRefCell<Option<ServoUrl>>

URL of the media resource, if any.

blob_url: DomRefCell<Option<ServoUrl>>

URL of the media resource, if the resource is set through the src_object attribute and it is a blob.

played: DomRefCell<TimeRangesContainer>

https://html.spec.whatwg.org/multipage/#dom-media-played

audio_tracks_list: MutNullableDom<AudioTrackList>video_tracks_list: MutNullableDom<VideoTrackList>text_tracks_list: MutNullableDom<TextTrackList>

https://html.spec.whatwg.org/multipage/#list-of-text-tracks

next_timeupdate_event: Cell<Instant>

Time of last timeupdate notification.

current_fetch_context: RefCell<Option<HTMLMediaElementFetchContext>>

Latest fetch request context.

media_controls_id: DomRefCell<Option<String>>

Media controls id. In order to workaround the lack of privileged JS context, we secure the the access to the “privileged” document.servoGetMediaControls(id) API by keeping a whitelist of media controls identifiers.

did_perform_automatic_track_selection: Cell<bool>

https://html.spec.whatwg.org/multipage/#did-perform-automatic-track-selection

position_when_time_marches_on_ran: Cell<Option<f64>>

Used to track the https://html.spec.whatwg.org/multipage/#current-playback-position when https://html.spec.whatwg.org/multipage/#time-marches-on was last invoked (if any)

playback_was_moved: Cell<PlaybackPositionWasMoved>

Used to track whether the only reason https://html.spec.whatwg.org/multipage/#time-marches-on was called, because it was called as a normal monotic increase.

newly_introduced_cues: DomRefCell<Vec<Dom<TextTrackCue>>>

https://html.spec.whatwg.org/multipage/#list-of-newly-introduced-cues

Implementations

impl HTMLMediaElement

pub(crate) fn new_inherited( tag_name: LocalName, prefix: Option<Prefix>, document: &Document, ) -> Self

pub(crate) fn network_state(&self) -> NetworkState

pub(crate) fn get_ready_state(&self) -> ReadyState

fn media_type_id(&self) -> HTMLMediaElementTypeId

fn update_media_state(&self)

pub(crate) fn delay_load_event(&self, delay: bool, cx: &mut JSContext)

Marks that element as delaying the load event or not.

Nothing happens if the element was already delaying the load event and we pass true to that method again.

https://html.spec.whatwg.org/multipage/#delaying-the-load-event-flag

fn time_marches_on( &self, cx: &mut JSContext, playback_was_moved: PlaybackPositionWasMoved, )

https://html.spec.whatwg.org/multipage/#time-marches-on

fn internal_play_steps(&self, cx: &mut JSContext)

https://html.spec.whatwg.org/multipage/#internal-play-steps

fn internal_pause_steps(&self)

https://html.spec.whatwg.org/multipage/#internal-pause-steps

fn is_allowed_to_play(&self) -> bool

https://html.spec.whatwg.org/multipage/#allowed-to-play

fn notify_about_playing(&self)

https://html.spec.whatwg.org/multipage/#notify-about-playing

fn change_ready_state(&self, cx: &mut JSContext, ready_state: ReadyState)

https://html.spec.whatwg.org/multipage/#ready-states

fn invoke_resource_selection_algorithm(&self, cx: &mut JSContext)

https://html.spec.whatwg.org/multipage/#concept-media-load-algorithm

fn resource_selection_algorithm_sync( &self, base_url: ServoUrl, cx: &mut JSContext, )

https://html.spec.whatwg.org/multipage/#concept-media-load-algorithm

fn load_from_src_object(&self, cx: &JSContext)

https://html.spec.whatwg.org/multipage/#concept-media-load-algorithm

fn load_from_src_attribute(&self, cx: &JSContext, base_url: ServoUrl, src: &str)

https://html.spec.whatwg.org/multipage/#concept-media-load-algorithm

fn load_from_source_child(&self, cx: &JSContext, source: &HTMLSourceElement)

https://html.spec.whatwg.org/multipage/#concept-media-load-algorithm

fn load_from_source_child_failure_steps( &self, cx: &JSContext, source: &HTMLSourceElement, )

https://html.spec.whatwg.org/multipage/#concept-media-load-algorithm

fn select_next_source_child(&self, cx: &mut JSContext)

https://html.spec.whatwg.org/multipage/#concept-media-load-algorithm

fn resource_selection_algorithm_failure_steps(&self, cx: &JSContext)

https://html.spec.whatwg.org/multipage/#concept-media-load-algorithm

fn fetch_request( &self, cx: &JSContext, offset: Option<u64>, seek_lock: Option<SeekLock>, )

fn eligible_for_autoplay(&self) -> bool

https://html.spec.whatwg.org/multipage/#eligible-for-autoplay

fn resource_fetch_algorithm(&self, cx: &JSContext, resource: Resource)

https://html.spec.whatwg.org/multipage/#concept-media-load-resource

fn queue_dedicated_media_source_failure_steps(&self)

Queues a task to run the dedicated media source failure steps.

fn in_error_state(&self) -> bool

fn is_potentially_playing(&self) -> bool

https://html.spec.whatwg.org/multipage/#potentially-playing

fn is_blocked_media_element(&self) -> bool

https://html.spec.whatwg.org/multipage/#blocked-media-element

fn is_paused_for_user_interaction(&self) -> bool

https://html.spec.whatwg.org/multipage/#paused-for-user-interaction

fn is_paused_for_in_band_content(&self) -> bool

https://html.spec.whatwg.org/multipage/#paused-for-in-band-content

fn media_element_load_algorithm(&self, cx: &mut JSContext)

https://html.spec.whatwg.org/multipage/#media-element-load-algorithm

fn queue_media_element_task_to_fire_event(&self, name: Atom)

Queue a media element task given the media element to fire an event at the media element. https://html.spec.whatwg.org/multipage/#queue-a-media-element-task

fn push_pending_play_promise(&self, promise: &Rc<Promise>)

Appends a promise to the list of pending play promises.

fn take_pending_play_promises(&self, result: ErrorResult)

Takes the pending play promises.

The result with which these promises will be fulfilled is passed here and this method returns nothing because we actually just move the current list of pending play promises to the in_flight_play_promises_queue field.

Each call to this method must be followed by a call to fulfill_in_flight_play_promises, to actually fulfill the promises which were taken and moved to the in-flight queue.

fn fulfill_in_flight_play_promises<F>(&self, cx: &mut JSContext, f: F)
where F: FnOnce(&mut JSContext),

Fulfills the next in-flight play promises queue after running a closure.

See the comment on take_pending_play_promises for why this method does not take a list of promises to fulfill. Callers cannot just pop the front list off of in_flight_play_promises_queue and later fulfill the promises because that would mean putting #[cfg_attr(crown, expect(crown::unrooted_must_root))] on even more functions, potentially hiding actual safety bugs.

pub(crate) fn handle_source_child_insertion( &self, source: &HTMLSourceElement, cx: &mut JSContext, )

fn select_next_source_child_after_wait(&self, cx: &mut JSContext)

https://html.spec.whatwg.org/multipage/#concept-media-load-algorithm

fn media_data_processing_failure_steps(&self, cx: &JSContext)

https://html.spec.whatwg.org/multipage/#media-data-processing-steps-list => “If the media data cannot be fetched at all, due to network errors…” => “If the media data can be fetched but is found by inspection to be in an unsupported format, or can otherwise not be rendered at all”

fn media_data_processing_fatal_steps(&self, error: u16, cx: &mut JSContext)

https://html.spec.whatwg.org/multipage/#media-data-processing-steps-list => “If the connection is interrupted after some media data has been received…” => “If the media data is corrupted”

fn seek(&self, time: f64, _approximate_for_speed: bool)

https://html.spec.whatwg.org/multipage/#dom-media-seek

fn seek_end(&self, cx: &mut JSContext)

https://html.spec.whatwg.org/multipage/#dom-media-seek

pub(crate) fn set_poster_frame( &self, no_gc: &NoGC, image: Option<Arc<RasterImage>>, )

https://html.spec.whatwg.org/multipage/#poster-frame

fn player_id(&self) -> Option<usize>

fn create_media_player(&self, resource: &Resource) -> Result<(), ()>

fn reset_media_player(&self, no_gc: &NoGC)

pub(crate) fn set_audio_track(&self, idx: usize, enabled: bool)

pub(crate) fn set_video_track(&self, idx: usize, enabled: bool)

fn direction_of_playback(&self) -> PlaybackDirection

https://html.spec.whatwg.org/multipage/#direction-of-playback

fn ended_playback(&self, loop_condition: LoopCondition) -> bool

https://html.spec.whatwg.org/multipage/#ended-playback

fn end_of_playback_in_forwards_direction(&self, cx: &mut JSContext)

https://html.spec.whatwg.org/multipage/#reaches-the-end

fn end_of_playback_in_backwards_direction(&self)

https://html.spec.whatwg.org/multipage/#reaches-the-end

fn playback_end(&self, cx: &mut JSContext)

fn playback_error(&self, error: &str, cx: &mut JSContext)

fn playback_metadata_updated(&self, cx: &mut JSContext, metadata: &Metadata)

fn playback_duration_changed(&self, duration: Option<Duration>)

fn playback_video_frame_updated(&self, no_gc: &NoGC)

fn playback_need_data(&self)

fn playback_enough_data(&self)

fn playback_position_changed(&self, cx: &mut JSContext, position: f64)

fn playback_seek_done(&self, cx: &JSContext, position: f64)

fn playback_state_changed(&self, cx: &mut JSContext, state: &PlaybackState)

fn seekable(&self) -> TimeRangesContainer

fn earliest_possible_position(&self) -> f64

https://html.spec.whatwg.org/multipage/#earliest-possible-position

fn render_controls(&self, cx: &mut JSContext)

fn remove_controls(&self)

pub(crate) fn get_current_frame(&self) -> Option<VideoFrame>

Gets the video frame at the current playback position.

pub(crate) fn get_current_frame_to_present(&self) -> Option<MediaFrame>

Gets the current frame of the video element to present, if any. https://html.spec.whatwg.org/multipage/#the-video-element:the-video-element-7

pub(crate) fn set_audio_renderer( &self, audio_renderer: Option<Arc<Mutex<dyn AudioRenderer>>>, cx: &mut JSContext, )

By default the audio is rendered through the audio sink automatically selected by the servo-media Player instance. However, in some cases, like the WebAudio MediaElementAudioSourceNode, we need to set a custom audio renderer.

fn send_media_session_event(&self, cx: &mut JSContext, event: MediaSessionEvent)

pub(crate) fn origin_is_clean(&self) -> bool

https://html.spec.whatwg.org/multipage/#concept-media-load-resource

pub(crate) fn add_newly_added_cue(&self, cx: &mut JSContext, cue: &TextTrackCue)

https://html.spec.whatwg.org/multipage/#list-of-newly-introduced-cues

pub(crate) fn was_added_to_list_of_text_tracks( &self, cx: &mut JSContext, track: &TextTrack, )

https://html.spec.whatwg.org/multipage/#list-of-newly-introduced-cues

fn honor_user_preferences_for_automatic_text_track_selection( &self, cx: &mut JSContext, )

https://html.spec.whatwg.org/multipage/#honor-user-preferences-for-automatic-text-track-selection

fn perform_automatic_text_track_selection( &self, cx: &mut JSContext, text_track_kinds: Vec<TextTrackKind>, )

https://html.spec.whatwg.org/multipage/#perform-automatic-text-track-selection