script::dom::html::embedded_content::htmlvideoelement

Struct HTMLVideoElement

#[repr(C)]
pub(crate) struct HTMLVideoElement {
    htmlmediaelement: HTMLMediaElement,
    video_width: Cell<Option<u32>>,
    video_height: Cell<Option<u32>>,
    poster_url: DomRefCell<Option<UrlWithBlobClaim>>,
    generation_id: Cell<u32>,
    load_blocker: DomRefCell<Option<LoadBlocker>>,
    last_frame: DomRefCell<Option<VideoFrame>>,
    planar_texture: DomRefCell<Option<Rc<PlanarTexture>>>,
}

Fields
htmlmediaelement: HTMLMediaElementvideo_width: Cell<Option<u32>>

https://html.spec.whatwg.org/multipage/#dom-video-videowidth

video_height: Cell<Option<u32>>

https://html.spec.whatwg.org/multipage/#dom-video-videoheight

poster_url: DomRefCell<Option<UrlWithBlobClaim>>

https://html.spec.whatwg.org/multipage/#dom-video-poster

generation_id: Cell<u32>

Incremented whenever tasks associated with this element are cancelled.

load_blocker: DomRefCell<Option<LoadBlocker>>

Load event blocker. Will block the load event while the poster frame is being fetched.

last_frame: DomRefCell<Option<VideoFrame>>

A copy of the last frame

planar_texture: DomRefCell<Option<Rc<PlanarTexture>>>

Planar texture for WebGPU

Implementations

impl HTMLVideoElement

fn __assert_parent_type(&self)

impl HTMLVideoElement

fn new_inherited( local_name: LocalName, prefix: Option<Prefix>, document: &Document, ) -> HTMLVideoElement

pub(crate) fn new( cx: &mut JSContext, local_name: LocalName, prefix: Option<Prefix>, document: &Document, proto: Option<HandleObject<'_>>, ) -> DomRoot<HTMLVideoElement>

pub(crate) fn get_video_width(&self) -> Option<u32>

pub(crate) fn get_video_height(&self) -> Option<u32>

pub(crate) fn set_natural_dimensions( &self, no_gc: &NoGC, width: Option<u32>, height: Option<u32>, ) -> bool

pub(crate) fn get_current_frame_data(&self) -> Option<Snapshot>

Gets the copy of the video frame at the current playback position, if that is available, or else (e.g. when the video is seeking or buffering) its previous appearance, if any.

pub(crate) fn planar_video_for_webgpu( &self, device: &GPUDevice, ) -> Fallible<(Size2D<u32>, Option<Rc<PlanarTexture>>)>

fn update_poster_frame( &self, poster_attribute: Option<&str>, cx: &mut JSContext, )

https://html.spec.whatwg.org/multipage/#poster-frame

fn do_fetch_poster_frame( &self, poster_url: UrlWithBlobClaim, id: PendingImageId, cx: &mut JSContext, )

https://html.spec.whatwg.org/multipage/#poster-frame

fn generation_id(&self) -> u32

fn process_image_response(&self, response: ImageResponse, cx: &mut JSContext)

https://html.spec.whatwg.org/multipage/#poster-frame

pub(crate) fn is_usable(&self) -> bool

https://html.spec.whatwg.org/multipage/#check-the-usability-of-the-image-argument

pub(crate) fn origin_is_clean(&self) -> bool

pub(crate) fn is_network_state_empty(&self) -> bool