script::dom::html::form_controls::htmlinputelement

Struct HTMLInputElement

#[repr(C)]
pub(crate) struct HTMLInputElement {

    htmlelement: HTMLElement,
    input_type: DomRefCell<InputType>,
    is_textual_or_password: Cell<bool>,
    checked_changed: Cell<bool>,
    placeholder: DomRefCell<DOMString>,
    size: Cell<u32>,
    maxlength: Cell<i32>,
    minlength: Cell<i32>,
    textinput: DomRefCell<TextInput<EmbedderClipboardProvider>>,
    value_dirty: Cell<bool>,
    form_owner: MutNullableDom<HTMLFormElement>,
    labels_node_list: MutNullableDom<NodeList>,
    validity_state: MutNullableDom<ValidityState>,
    pending_webdriver_response: RefCell<Option<PendingWebDriverResponse>>,
    has_scheduled_selectionchange_event: Cell<bool>,

}

Fields
htmlelement: HTMLElementinput_type: DomRefCell<InputType>is_textual_or_password: Cell<bool>

Whether or not the InputType for this HTMLInputElement renders as textual input. This is cached so that it can be read during layout.

checked_changed: Cell<bool>

https://html.spec.whatwg.org/multipage/#concept-input-checked-dirty-flag

placeholder: DomRefCell<DOMString>size: Cell<u32>maxlength: Cell<i32>minlength: Cell<i32>textinput: DomRefCell<TextInput<EmbedderClipboardProvider>>value_dirty: Cell<bool>

https://html.spec.whatwg.org/multipage/#concept-input-value-dirty-flag

form_owner: MutNullableDom<HTMLFormElement>labels_node_list: MutNullableDom<NodeList>validity_state: MutNullableDom<ValidityState>pending_webdriver_response: RefCell<Option<PendingWebDriverResponse>>has_scheduled_selectionchange_event: Cell<bool>

https://w3c.github.io/selection-api/#dfn-has-scheduled-selectionchange-event

Implementations

impl HTMLInputElement

fn __assert_parent_type(&self)

impl HTMLInputElement

fn new_inherited( local_name: LocalName, prefix: Option<Prefix>, document: &Document, ) -> HTMLInputElement

pub(crate) fn new( cx: &mut JSContext, local_name: LocalName, prefix: Option<Prefix>, document: &Document, proto: Option<HandleObject<'_>>, ) -> DomRoot<HTMLInputElement>

pub(crate) fn auto_directionality(&self) -> Option<String>

pub(crate) fn directionality_from_value(value: &str) -> String

fn is_first_strong_character_rtl(value: &str) -> bool

pub(crate) fn value_mode(&self) -> ValueMode

https://html.spec.whatwg.org/multipage/#concept-input-apply

pub(crate) fn input_type(&self) -> Ref<'_, InputType>

pub(crate) fn is_nontypeable(&self) -> bool

https://w3c.github.io/webdriver/#dfn-non-typeable-form-control

pub(crate) fn is_submit_button(&self) -> bool

pub(crate) fn is_auto_directionality_form_associated_element(&self) -> bool

https://html.spec.whatwg.org/multipage/#auto-directionality-form-associated-elements

fn does_minmaxlength_apply(&self) -> bool

fn does_pattern_apply(&self) -> bool

fn does_multiple_apply(&self) -> bool

fn does_value_as_number_apply(&self) -> bool

fn does_value_as_date_apply(&self) -> bool

pub(crate) fn allowed_value_step(&self) -> Option<f64>

https://html.spec.whatwg.org/multipage#concept-input-step

pub(crate) fn minimum(&self) -> Option<f64>

https://html.spec.whatwg.org/multipage#concept-input-min

pub(crate) fn maximum(&self) -> Option<f64>

https://html.spec.whatwg.org/multipage#concept-input-max

pub(crate) fn stepped_minimum(&self) -> Option<f64>

when allowed_value_step and minimum both exist, this is the smallest value >= minimum that lies on an integer step

pub(crate) fn stepped_maximum(&self) -> Option<f64>

when allowed_value_step and maximum both exist, this is the smallest value <= maximum that lies on an integer step

fn default_minimum(&self) -> Option<f64>

https://html.spec.whatwg.org/multipage#concept-input-min-default

fn default_maximum(&self) -> Option<f64>

https://html.spec.whatwg.org/multipage#concept-input-max-default

pub(crate) fn default_range_value(&self) -> f64

https://html.spec.whatwg.org/multipage#concept-input-value-default-range

fn default_step(&self) -> Option<f64>

https://html.spec.whatwg.org/multipage#concept-input-step-default

fn step_scale_factor(&self) -> f64

https://html.spec.whatwg.org/multipage#concept-input-step-scale

pub(crate) fn step_base(&self) -> f64

https://html.spec.whatwg.org/multipage#concept-input-min-zero

fn default_step_base(&self) -> Option<f64>

https://html.spec.whatwg.org/multipage#concept-input-step-default-base

fn step_up_or_down( &self, cx: &mut JSContext, n: i32, dir: StepDirection, ) -> ErrorResult

https://html.spec.whatwg.org/multipage/#dom-input-stepup

https://html.spec.whatwg.org/multipage/#dom-input-stepdown

fn suggestions_source_element(&self) -> Option<DomRoot<HTMLDataListElement>>

https://html.spec.whatwg.org/multipage/#concept-input-list

fn suffers_from_being_missing(&self, value: &DOMString) -> bool

https://html.spec.whatwg.org/multipage/#suffering-from-being-missing

fn suffers_from_type_mismatch(&self, value: &DOMString) -> bool

https://html.spec.whatwg.org/multipage/#suffering-from-a-type-mismatch

fn suffers_from_pattern_mismatch( &self, cx: &mut JSContext, value: &DOMString, ) -> bool

https://html.spec.whatwg.org/multipage/#suffering-from-a-pattern-mismatch

fn suffers_from_bad_input(&self, value: &DOMString) -> bool

https://html.spec.whatwg.org/multipage/#suffering-from-bad-input

fn suffers_from_length_issues(&self, value: &DOMString) -> ValidationFlags

https://html.spec.whatwg.org/multipage/#suffering-from-being-too-short

fn suffers_from_range_issues(&self, value: &DOMString) -> ValidationFlags

https://html.spec.whatwg.org/multipage/#suffering-from-an-underflow

https://html.spec.whatwg.org/multipage/#suffering-from-an-overflow

https://html.spec.whatwg.org/multipage/#suffering-from-a-step-mismatch

pub(crate) fn is_textual_or_password(&self) -> bool

Whether this input type renders as a basic text input widget.

fn may_have_embedder_control(&self) -> bool

fn handle_key_reaction( &self, cx: &mut JSContext, action: KeyReaction, event: &Event, )

pub(crate) fn value_for_shadow_dom(&self) -> DOMString

Return a string that represents the contents of the element in its displayed shadow DOM.

pub(crate) fn textinput_mut( &self, ) -> RefMut<'_, TextInput<EmbedderClipboardProvider>>

fn schedule_a_selection_change_event(&self)

https://w3c.github.io/selection-api/#dfn-schedule-a-selectionchange-event

impl HTMLInputElement

pub(crate) fn form_datums( &self, submitter: Option<FormSubmitterElement<'_>>, encoding: Option<&'static Encoding>, ) -> (Vec<FormDatum>, bool)

https://html.spec.whatwg.org/multipage/#constructing-the-form-data-set Steps range from 5.1 to 5.10 (specific to HTMLInputElement)

pub(crate) fn radio_group_name(&self) -> Option<Atom>

https://html.spec.whatwg.org/multipage/#radio-button-group

fn update_checked_state(&self, cx: &mut JSContext, checked: bool, dirty: bool)

pub(crate) fn is_mutable(&self) -> bool

pub(crate) fn reset(&self, cx: &mut JSContext)

https://html.spec.whatwg.org/multipage/#the-input-element:concept-form-reset-control:

The reset algorithm for input elements is to set its user validity, dirty value flag, and dirty checkedness flag back to false, set the value of the element to the value of the value content attribute, if there is one, or the empty string otherwise, set the checkedness of the element to true if the element has a checked content attribute and false if it does not, empty the list of selected files, and then invoke the value sanitization algorithm, if the type attribute’s current state defines one.

pub(crate) fn clear(&self, cx: &mut JSContext)

https://w3c.github.io/webdriver/#ref-for-dfn-clear-algorithm-3 Used by WebDriver to clear the input element.

fn update_placeholder_shown_state(&self)

pub(crate) fn select_files_for_webdriver( &self, test_paths: Vec<DOMString>, response_sender: GenericSender<Result<bool, ErrorStatus>>, )

pub(crate) fn take_pending_webdriver_response( &self, ) -> Option<PendingWebDriverResponse>

fn sanitize_value(&self, value: &mut DOMString)

https://html.spec.whatwg.org/multipage/#value-sanitization-algorithm

fn selection(&self) -> TextControlSelection<'_, Self>

fn implicit_submission(&self, cx: &mut JSContext)

https://html.spec.whatwg.org/multipage/#implicit-submission

pub(crate) fn convert_string_to_number(&self, value: &str) -> Option<f64>

https://html.spec.whatwg.org/multipage/#concept-input-value-string-number

fn convert_number_to_string(&self, value: f64) -> Option<DOMString>

https://html.spec.whatwg.org/multipage/#concept-input-value-string-number

fn update_related_validity_states(&self, cx: &mut JSContext)

fn value_changed(&self, cx: &mut JSContext)

pub(crate) fn show_the_picker_if_applicable(&self)

https://html.spec.whatwg.org/multipage/#show-the-picker,-if-applicable

pub(crate) fn handle_color_picker_response( &self, cx: &mut JSContext, response: Option<RgbColor>, )

pub(crate) fn handle_file_picker_response( &self, cx: &mut JSContext, response: Option<Vec<SelectedFile>>, )

fn handle_focus_event(&self, event: &FocusEvent)