script::dom::html::form_controls::input_type

Enum InputType

pub(crate) enum InputType {

    Button(ButtonInputType),
    Checkbox(CheckboxInputType),
    Color(ColorInputType),
    Date(DateInputType),
    DatetimeLocal(DatetimeLocalInputType),
    Email(EmailInputType),
    File(FileInputType),
    Hidden(HiddenInputType),
    Image(ImageInputType),
    Month(MonthInputType),
    Number(NumberInputType),
    Password(PasswordInputType),
    Radio(RadioInputType),
    Range(RangeInputType),
    Reset(ResetInputType),
    Search(SearchInputType),
    Submit(SubmitInputType),
    Tel(TelInputType),
    Text(TextInputType),
    Time(TimeInputType),
    Url(UrlInputType),
    Week(WeekInputType),

}

https://html.spec.whatwg.org/multipage/#attr-input-type

Variants

Button(ButtonInputType)

https://html.spec.whatwg.org/multipage/#button-state-(type=button)

Checkbox(CheckboxInputType)

https://html.spec.whatwg.org/multipage/#checkbox-state-(type=checkbox)

Color(ColorInputType)

https://html.spec.whatwg.org/multipage/#color-state-(type=color)

Date(DateInputType)

https://html.spec.whatwg.org/multipage/#date-state-(type=date)

DatetimeLocal(DatetimeLocalInputType)

https://html.spec.whatwg.org/multipage/#local-date-and-time-state-(type=datetime-local)

Email(EmailInputType)

https://html.spec.whatwg.org/multipage/#email-state-(type=email)

File(FileInputType)

https://html.spec.whatwg.org/multipage/#file-upload-state-(type=file)

Hidden(HiddenInputType)

https://html.spec.whatwg.org/multipage/#hidden-state-(type=hidden)

Image(ImageInputType)

https://html.spec.whatwg.org/multipage/#image-button-state-(type=image)

Month(MonthInputType)

https://html.spec.whatwg.org/multipage/#month-state-(type=month)

Number(NumberInputType)

https://html.spec.whatwg.org/multipage/#number-state-(type=number)

Password(PasswordInputType)

https://html.spec.whatwg.org/multipage/#password-state-(type=password)

Radio(RadioInputType)

https://html.spec.whatwg.org/multipage/#radio-button-state-(type=radio)

Range(RangeInputType)

https://html.spec.whatwg.org/multipage/#range-state-(type=range)

Reset(ResetInputType)

https://html.spec.whatwg.org/multipage/#reset-button-state-(type=reset)

Search(SearchInputType)

https://html.spec.whatwg.org/multipage/#text-(type=text)-state-and-search-state-(type=search)

Submit(SubmitInputType)

https://html.spec.whatwg.org/multipage/#submit-button-state-(type=submit)

Tel(TelInputType)

https://html.spec.whatwg.org/multipage/#telephone-state-(type=tel)

Text(TextInputType)

https://html.spec.whatwg.org/multipage/#text-(type=text)-state-and-search-state-(type=search)

Time(TimeInputType)

https://html.spec.whatwg.org/multipage/#time-state-(type=time)

Url(UrlInputType)

https://html.spec.whatwg.org/multipage/#url-state-(type=url)

Week(WeekInputType)

https://html.spec.whatwg.org/multipage/#week-state-(type=week)

Implementations

impl InputType

pub(crate) fn new_from_atom(value: &Atom) -> Self

pub(crate) fn new_text() -> Self

pub(crate) fn as_specific(&self) -> &dyn SpecificInputType

pub(crate) fn is_textual(&self) -> bool

Defines which input type that should perform like a text input, specifically when it is interacting with JS. Note that Password is not included here since it is handled slightly differently, with placeholder characters shown rather than the underlying value.

pub(crate) fn is_textual_or_password(&self) -> bool

pub(crate) fn has_periodic_domain(&self) -> bool

https://html.spec.whatwg.org/multipage/#has-a-periodic-domain

pub(crate) fn as_str(&self) -> &str