script::dom::html::form_controls::input_type

Module text_input_widget

Structs

TextInputWidget 🔒

TextInputWidgetShadowTree 🔒

Contains reference to text control inner editor and placeholder container element in the UA shadow tree for text, password, url, tel, and email input. The following is the structure of the shadow tree.

Constants

PASSWORD_REPLACEMENT_CHAR 🔒

Functions

create_ua_widget_div_with_text_node 🔒

Create a div element with a text node within an UA Widget and either append or prepend it to the designated parent. This is used to create the text container for input elements.