script::dom::html::form_controls::input_type

Trait SpecificInputType

pub(crate) trait SpecificInputType {

    // Required method
    fn text_input_widget(&self) -> Option<&DomRefCell<TextInputWidget>>;

    // Provided methods
    fn sanitize_value(&self, _input: &HTMLInputElement, _value: &mut DOMString) { ... }
    fn convert_string_to_number(&self, _value: &str) -> Option<f64> { ... }
    fn convert_number_to_string(&self, _value: f64) -> Option<DOMString> { ... }
    fn convert_string_to_naive_datetime(
        &self,
        _value: DOMString,
    ) -> Option<OffsetDateTime> { ... }
    fn convert_datetime_to_dom_string(
        &self,
        _value: OffsetDateTime,
    ) -> DOMString { ... }
    fn suffers_from_being_missing(
        &self,
        input: &HTMLInputElement,
        value: &DOMString,
    ) -> bool { ... }
    fn suffers_from_bad_input(&self, _value: &DOMString) -> bool { ... }
    fn suffers_from_type_mismatch(
        &self,
        _input: &HTMLInputElement,
        _value: &DOMString,
    ) -> bool { ... }
    fn value_for_shadow_dom(&self, _input: &HTMLInputElement) -> DOMString { ... }
    fn signal_type_change(&self, _cx: &mut JSContext, _input: &HTMLInputElement) { ... }
    fn legacy_pre_activation_behavior(
        &self,
        _cx: &mut JSContext,
        _input: &HTMLInputElement,
    ) -> Option<InputActivationState> { ... }
    fn legacy_canceled_activation_behavior(
        &self,
        _cx: &mut JSContext,
        _input: &HTMLInputElement,
        _cache: InputActivationState,
    ) { ... }
    fn show_the_picker_if_applicable(&self, _input: &HTMLInputElement) { ... }
    fn select_files(
        &self,
        _input: &HTMLInputElement,
        _test_paths: Option<Vec<DOMString>>,
    ) { ... }
    fn get_files(&self) -> Option<DomRoot<FileList>> { ... }
    fn set_files(&self, _filelist: &FileList) { ... }
    fn update_shadow_tree(&self, _cx: &mut JSContext, _input: &HTMLInputElement) { ... }
    fn update_placeholder_contents(
        &self,
        _cx: &mut JSContext,
        _input: &HTMLInputElement,
    ) { ... }
    fn attribute_mutated(
        &self,
        _cx: &mut JSContext,
        _input: &HTMLInputElement,
        _attr: AttrRef<'_>,
        _mutation: AttributeMutation<'_>,
    ) { ... }
    fn bind_to_tree(
        &self,
        _cx: &mut JSContext,
        _input: &HTMLInputElement,
        _context: &BindContext<'_>,
    ) { ... }
    fn unbind_from_tree(
        &self,
        _cx: &mut JSContext,
        _input: &HTMLInputElement,
        _form_owner: Option<DomRoot<HTMLFormElement>>,
        _context: &UnbindContext<'_>,
    ) { ... }

}

Required Methods

fn text_input_widget(&self) -> Option<&DomRefCell<TextInputWidget>>

Provided Methods

fn sanitize_value(&self, _input: &HTMLInputElement, _value: &mut DOMString)

fn convert_string_to_number(&self, _value: &str) -> Option<f64>

fn convert_number_to_string(&self, _value: f64) -> Option<DOMString>

fn convert_string_to_naive_datetime( &self, _value: DOMString, ) -> Option<OffsetDateTime>

https://html.spec.whatwg.org/multipage/#concept-input-value-string-date This does the safe Rust part of conversion; the unsafe JS Date part is in GetValueAsDate

fn convert_datetime_to_dom_string(&self, _value: OffsetDateTime) -> DOMString

https://html.spec.whatwg.org/multipage/#concept-input-value-date-string This does the safe Rust part of conversion; the unsafe JS Date part is in SetValueAsDate

fn suffers_from_being_missing( &self, input: &HTMLInputElement, value: &DOMString, ) -> bool

https://html.spec.whatwg.org/multipage/#the-required-attribute%3Asuffering-from-being-missing

fn suffers_from_bad_input(&self, _value: &DOMString) -> bool

fn suffers_from_type_mismatch( &self, _input: &HTMLInputElement, _value: &DOMString, ) -> bool

fn value_for_shadow_dom(&self, _input: &HTMLInputElement) -> DOMString

fn signal_type_change(&self, _cx: &mut JSContext, _input: &HTMLInputElement)

https://html.spec.whatwg.org/multipage/#signal-a-type-change

fn legacy_pre_activation_behavior( &self, _cx: &mut JSContext, _input: &HTMLInputElement, ) -> Option<InputActivationState>

fn legacy_canceled_activation_behavior( &self, _cx: &mut JSContext, _input: &HTMLInputElement, _cache: InputActivationState, )

fn show_the_picker_if_applicable(&self, _input: &HTMLInputElement)

fn select_files( &self, _input: &HTMLInputElement, _test_paths: Option<Vec<DOMString>>, )

fn get_files(&self) -> Option<DomRoot<FileList>>

fn set_files(&self, _filelist: &FileList)

fn update_shadow_tree(&self, _cx: &mut JSContext, _input: &HTMLInputElement)

fn update_placeholder_contents( &self, _cx: &mut JSContext, _input: &HTMLInputElement, )

fn attribute_mutated( &self, _cx: &mut JSContext, _input: &HTMLInputElement, _attr: AttrRef<'_>, _mutation: AttributeMutation<'_>, )

fn bind_to_tree( &self, _cx: &mut JSContext, _input: &HTMLInputElement, _context: &BindContext<'_>, )

fn unbind_from_tree( &self, _cx: &mut JSContext, _input: &HTMLInputElement, _form_owner: Option<DomRoot<HTMLFormElement>>, _context: &UnbindContext<'_>, )