script::dom::html::form_controls

Module text_input

Common handling of keyboard input and state management for text input controls

Structs

ClipboardEventFlags

Resulting action to be taken by the owner of a text input that is handling a clipboard event.

ClipboardEventReaction

EmbedderClipboardProvider 🔒

SelectionState 🔒

TextInput

Encapsulated state for handling keyboard input in a single or multiline text input control.

TextInputSelectionDragHandler 🔒

Enums

Direction

The direction in which to delete a character.

InputType

https://www.w3.org/TR/input-events-2/#interface-InputEvent-Attributes

IsComposing

KeyReaction

Resulting action to be taken by the owner of a text input that is handling an event.

Lines

Selection

SelectionDirection

Constants

CMD_OR_CONTROL 🔒

Traits

ClipboardProvider

A trait which abstracts access to the embedder’s clipboard in order to allow unit testing clipboard-dependent parts of script.

Functions

len_of_first_n_code_units 🔒

The length in bytes of the first n code units in a string when encoded in UTF-16.