script::dom::html

Module htmldetailselement

Structs

DetailsNameGroups 🔒

Tracks all details name groups within a tree.

HTMLDetailsElement 🔒

ShadowTree 🔒

Holds handles to all slots in the UA shadow tree

Enums

ExclusivityConflictResolution 🔒

Describes how to proceed in case two details elements in the same details name groups are open at the same time

Constants

DEFAULT_SUMMARY 🔒

The summary that should be presented if no <summary> element is present

_IMPL_DOMOBJECT_FOR_HTMLDetailsElement 🔒