script::dom::html::htmlelement

Struct HTMLElement

#[repr(C)]
pub(crate) struct HTMLElement {
    element: Element,
    style_decl: MutNullableDom<CSSStyleDeclaration>,
    dataset: MutNullableDom<DOMStringMap>,
}

Fields
element: Elementstyle_decl: MutNullableDom<CSSStyleDeclaration>dataset: MutNullableDom<DOMStringMap>
Implementations

impl HTMLElement

fn __assert_parent_type(&self)

impl HTMLElement

pub(crate) fn local_name(&self) -> &str

fn remove_value_from_text_decoration(&self, cx: &mut JSContext, value: &str)

pub(crate) fn clear_the_value(&self, cx: &mut JSContext, command: &CommandName)

https://w3c.github.io/editing/docs/execCommand/#clear-the-value

pub(crate) fn handle_focus_state_for_contenteditable(&self, cx: &mut JSContext)

There is no specification for this implementation. Instead, it is reverse-engineered based on the WPT test /selection/contenteditable/initial-selection-on-focus.tentative.html

impl HTMLElement

pub(crate) fn new_inherited( tag_name: LocalName, prefix: Option<Prefix>, document: &Document, ) -> HTMLElement

pub(crate) fn new_inherited_with_state( state: ElementState, tag_name: LocalName, prefix: Option<Prefix>, document: &Document, ) -> HTMLElement

pub(crate) fn new( cx: &mut JSContext, local_name: LocalName, prefix: Option<Prefix>, document: &Document, proto: Option<HandleObject<'_>>, ) -> DomRoot<HTMLElement>

fn is_body_or_frameset(&self) -> bool

pub(crate) fn get_inner_outer_text(&self) -> DOMString

Calls into the layout engine to generate a plain text representation of a HTMLElement as specified when getting the .innerText or .outerText in JavaScript.`

https://html.spec.whatwg.org/multipage/#get-the-text-steps

pub(crate) fn set_inner_text(&self, cx: &mut JSContext, input: DOMString)

https://html.spec.whatwg.org/multipage/#set-the-inner-text-steps

pub(crate) fn media_attribute_matches_media_environment(&self) -> bool

https://html.spec.whatwg.org/multipage/#matches-the-environment

pub(crate) fn is_editing_host(&self) -> bool

https://html.spec.whatwg.org/multipage/#editing-host

pub(crate) fn previously_focused_element( &self, no_gc: &NoGC, ) -> Option<DomRoot<Element>>

pub(crate) fn set_previously_focused_element( &self, element: Option<&Element>, no_gc: &NoGC, )

pub(crate) fn ensure_element_internals( &self, cx: &mut JSContext, ) -> DomRoot<ElementInternals>

impl HTMLElement

pub(crate) fn is_labelable_element(&self) -> bool

https://html.spec.whatwg.org/multipage/#category-label

pub(crate) fn is_form_associated_custom_element(&self) -> bool

https://html.spec.whatwg.org/multipage/#form-associated-custom-element

pub(crate) fn is_listed_element(&self) -> bool

https://html.spec.whatwg.org/multipage/#category-listed

pub(crate) fn is_body_element(&self) -> bool

https://html.spec.whatwg.org/multipage/#the-body-element-2

pub(crate) fn is_submittable_element(&self) -> bool

https://html.spec.whatwg.org/multipage/#category-submit

pub(crate) fn label_at<'a>( &self, no_gc: &'a NoGC, index: u32, ) -> Option<UnrootedDom<'a, Node>>

pub(crate) fn labels_count(&self) -> u32

pub(crate) fn directionality(&self) -> Option<String>

pub(crate) fn summary_activation_behavior(&self, cx: &mut JSContext)

pub(crate) fn is_a_summary_for_its_parent_details(&self) -> bool

https://html.spec.whatwg.org/multipage/#summary-for-its-parent-details

fn is_implicit_summary_element(&self) -> bool

Whether or not this is an implicitly generated <summary> element for a UA <details> shadow tree

fn rendered_text_fragment( &self, cx: &mut JSContext, input: DOMString, ) -> DomRoot<DocumentFragment>

https://html.spec.whatwg.org/multipage/#rendered-text-fragment

fn merge_with_the_next_text_node(cx: &mut JSContext, node: &Node)

Checks whether a given DomRoot<Node> and its next sibling are of type Text, and if so merges them into a single Text node.

https://html.spec.whatwg.org/multipage/#merge-with-the-next-text-node

fn update_assigned_access_key(&self)

https://html.spec.whatwg.org/multipage/#keyboard-shortcuts-processing-model

Whenever an element’s accesskey attribute is set, changed, or removed, the user agent must update the element’s assigned access key by running the following steps: