script::dom::html

Module htmlinputelement

Structs

ColorInputShadowTree 🔒

Contains references to the elements in the shadow tree for <input type=color>.

HTMLInputElement 🔒

InputActivationState 🔒

PendingWebDriverResponse 🔒

When WebDriver asks the HTMLInputElement to do some asynchronous actions, such as selecting files, this stores the details necessary to complete the response when the action is complete.

TextInputWidgetShadowTree 🔒

Contains reference to text control inner editor and placeholder container element in the UA shadow tree for text, password, url, tel, and email input. The following is the structure of the shadow tree.

TextValueShadowTree 🔒

Enums

InputElementShadowTree 🔒

InputType 🔒

https://html.spec.whatwg.org/multipage/#attr-input-type

StepDirection 🔒

ValueMode 🔒

Constants

DEFAULT_FILE_INPUT_VALUE 🔒

DEFAULT_RESET_VALUE 🔒

DEFAULT_SUBMIT_VALUE 🔒

PASSWORD_REPLACEMENT_CHAR 🔒

_IMPL_DOMOBJECT_FOR_HTMLInputElement 🔒

Statics

DEFAULT_INPUT_SIZE 🔒

DEFAULT_MAX_LENGTH 🔒

DEFAULT_MIN_LENGTH 🔒

Traits

LayoutHTMLInputElementHelpers 🔒

Functions

broadcast_radio_checked 🔒

check_js_regex_syntax 🔒

Check if the pattern by itself is valid first, and not that it only becomes valid once we add ^(?: and )$.

compile_pattern 🔒

This is used to compile JS-compatible regex provided in pattern attribute that matches only the entirety of string. https://html.spec.whatwg.org/multipage/#compiled-pattern-regular-expression

create_ua_widget_div_with_text_node 🔒

Create a div element with a text node within an UA Widget and either append or prepend it to the designated parent. This is used to create the text container for input elements.

filter_from_accept 🔒

https://html.spec.whatwg.org/multipage/#attr-input-accept

in_same_group 🔒

https://html.spec.whatwg.org/multipage/#radio-button-group

matches_js_regex 🔒

new_js_regex 🔒

perform_radio_group_validation 🔒

radio_group_iter 🔒

round_halves_positive 🔒