script::dom::html::htmlinputelement

Enum InputType

pub(crate) enum InputType {

    Button,
    Checkbox,
    Color,
    Date,
    DatetimeLocal,
    Email,
    File,
    Hidden,
    Image,
    Month,
    Number,
    Password,
    Radio,
    Range,
    Reset,
    Search,
    Submit,
    Tel,
    Text,
    Time,
    Url,
    Week,

}

https://html.spec.whatwg.org/multipage/#attr-input-type

Variants

Button

https://html.spec.whatwg.org/multipage/#button-state-(type=button)

Checkbox

https://html.spec.whatwg.org/multipage/#checkbox-state-(type=checkbox)

Color

https://html.spec.whatwg.org/multipage/#color-state-(type=color)

Date

https://html.spec.whatwg.org/multipage/#date-state-(type=date)

DatetimeLocal

https://html.spec.whatwg.org/multipage/#local-date-and-time-state-(type=datetime-local)

Email

https://html.spec.whatwg.org/multipage/#email-state-(type=email)

File

https://html.spec.whatwg.org/multipage/#file-upload-state-(type=file)

Hidden

https://html.spec.whatwg.org/multipage/#hidden-state-(type=hidden)

Image

https://html.spec.whatwg.org/multipage/#image-button-state-(type=image)

Month

https://html.spec.whatwg.org/multipage/#month-state-(type=month)

Number

https://html.spec.whatwg.org/multipage/#number-state-(type=number)

Password

https://html.spec.whatwg.org/multipage/#password-state-(type=password)

Radio

https://html.spec.whatwg.org/multipage/#radio-button-state-(type=radio)

Range

https://html.spec.whatwg.org/multipage/#range-state-(type=range)

Reset

https://html.spec.whatwg.org/multipage/#reset-button-state-(type=reset)

Search

https://html.spec.whatwg.org/multipage/#text-(type=text)-state-and-search-state-(type=search)

Submit

https://html.spec.whatwg.org/multipage/#submit-button-state-(type=submit)

Tel

https://html.spec.whatwg.org/multipage/#telephone-state-(type=tel)

Text

https://html.spec.whatwg.org/multipage/#text-(type=text)-state-and-search-state-(type=search)

Time

https://html.spec.whatwg.org/multipage/#time-state-(type=time)

Url

https://html.spec.whatwg.org/multipage/#url-state-(type=url)

Week

https://html.spec.whatwg.org/multipage/#week-state-(type=week)

Implementations

impl InputType

pub(crate) fn is_textual(&self) -> bool

Defines which input type that should perform like a text input, specifically when it is interacting with JS. Note that Password is not included here since it is handled slightly differently, with placeholder characters shown rather than the underlying value.

fn is_textual_or_password(&self) -> bool

fn has_periodic_domain(&self) -> bool

https://html.spec.whatwg.org/multipage/#has-a-periodic-domain

fn as_str(&self) -> &str