script::dom::html::htmlinputelement

Function create_ua_widget_div_with_text_node

fn create_ua_widget_div_with_text_node(
    document: &Document,
    parent: &Node,
    implemented_pseudo: PseudoElement,
    as_first_child: bool,
    can_gc: CanGc,
) -> DomRoot<Element>

Create a div element with a text node within an UA Widget and either append or prepend it to the designated parent. This is used to create the text container for input elements.