script::dom::html::htmlinputelement

Struct HTMLInputElement

#[repr(C)]
pub(crate) struct HTMLInputElement {

    htmlelement: HTMLElement,
    input_type: Cell<InputType>,
    checked_changed: Cell<bool>,
    placeholder: DomRefCell<DOMString>,
    size: Cell<u32>,
    maxlength: Cell<i32>,
    minlength: Cell<i32>,
    textinput: DomRefCell<TextInput<EmbedderClipboardProvider>>,
    value_dirty: Cell<bool>,
    sanitization_flag: Cell<bool>,
    filelist: MutNullableDom<FileList>,
    form_owner: MutNullableDom<HTMLFormElement>,
    labels_node_list: MutNullableDom<NodeList>,
    validity_state: MutNullableDom<ValidityState>,
    shadow_tree: DomRefCell<Option<InputElementShadowTree>>,
    pending_webdriver_response: RefCell<Option<PendingWebDriverResponse>>,

}

Fields
htmlelement: HTMLElementinput_type: Cell<InputType>checked_changed: Cell<bool>

https://html.spec.whatwg.org/multipage/#concept-input-checked-dirty-flag

placeholder: DomRefCell<DOMString>size: Cell<u32>maxlength: Cell<i32>minlength: Cell<i32>textinput: DomRefCell<TextInput<EmbedderClipboardProvider>>value_dirty: Cell<bool>sanitization_flag: Cell<bool>filelist: MutNullableDom<FileList>form_owner: MutNullableDom<HTMLFormElement>labels_node_list: MutNullableDom<NodeList>validity_state: MutNullableDom<ValidityState>shadow_tree: DomRefCell<Option<InputElementShadowTree>>pending_webdriver_response: RefCell<Option<PendingWebDriverResponse>>
Implementations

impl HTMLInputElement

fn __assert_parent_type(&self)

impl HTMLInputElement

fn new_inherited( local_name: LocalName, prefix: Option<Prefix>, document: &Document, ) -> HTMLInputElement

pub(crate) fn new( local_name: LocalName, prefix: Option<Prefix>, document: &Document, proto: Option<HandleObject<'_>>, can_gc: CanGc, ) -> DomRoot<HTMLInputElement>

pub(crate) fn auto_directionality(&self) -> Option<String>

pub(crate) fn directionality_from_value(value: &str) -> String

fn is_first_strong_character_rtl(value: &str) -> bool

fn value_mode(&self) -> ValueMode

https://html.spec.whatwg.org/multipage/#concept-input-apply

pub(crate) fn input_type(&self) -> InputType

pub(crate) fn is_nontypeable(&self) -> bool

https://w3c.github.io/webdriver/#dfn-non-typeable-form-control

pub(crate) fn is_submit_button(&self) -> bool

pub(crate) fn disable_sanitization(&self)

pub(crate) fn enable_sanitization(&self)

fn does_minmaxlength_apply(&self) -> bool

fn does_pattern_apply(&self) -> bool

fn does_multiple_apply(&self) -> bool

fn does_value_as_number_apply(&self) -> bool

fn does_value_as_date_apply(&self) -> bool

fn allowed_value_step(&self) -> Option<f64>

https://html.spec.whatwg.org/multipage#concept-input-step

fn minimum(&self) -> Option<f64>

https://html.spec.whatwg.org/multipage#concept-input-min

fn maximum(&self) -> Option<f64>

https://html.spec.whatwg.org/multipage#concept-input-max

fn stepped_minimum(&self) -> Option<f64>

when allowed_value_step and minimum both exist, this is the smallest value >= minimum that lies on an integer step

fn stepped_maximum(&self) -> Option<f64>

when allowed_value_step and maximum both exist, this is the smallest value <= maximum that lies on an integer step

fn default_minimum(&self) -> Option<f64>

https://html.spec.whatwg.org/multipage#concept-input-min-default

fn default_maximum(&self) -> Option<f64>

https://html.spec.whatwg.org/multipage#concept-input-max-default

fn default_range_value(&self) -> f64

https://html.spec.whatwg.org/multipage#concept-input-value-default-range

fn default_step(&self) -> Option<f64>

https://html.spec.whatwg.org/multipage#concept-input-step-default

fn step_scale_factor(&self) -> f64

https://html.spec.whatwg.org/multipage#concept-input-step-scale

fn step_base(&self) -> f64

https://html.spec.whatwg.org/multipage#concept-input-min-zero

fn default_step_base(&self) -> Option<f64>

https://html.spec.whatwg.org/multipage#concept-input-step-default-base

fn step_up_or_down( &self, n: i32, dir: StepDirection, can_gc: CanGc, ) -> ErrorResult

https://html.spec.whatwg.org/multipage/#dom-input-stepup

https://html.spec.whatwg.org/multipage/#dom-input-stepdown

fn suggestions_source_element(&self) -> Option<DomRoot<HTMLDataListElement>>

https://html.spec.whatwg.org/multipage/#concept-input-list

fn suffers_from_being_missing(&self, value: &DOMString) -> bool

https://html.spec.whatwg.org/multipage/#suffering-from-being-missing

fn suffers_from_type_mismatch(&self, value: &DOMString) -> bool

https://html.spec.whatwg.org/multipage/#suffering-from-a-type-mismatch

fn suffers_from_pattern_mismatch( &self, value: &DOMString, can_gc: CanGc, ) -> bool

https://html.spec.whatwg.org/multipage/#suffering-from-a-pattern-mismatch

fn suffers_from_bad_input(&self, value: &DOMString) -> bool

https://html.spec.whatwg.org/multipage/#suffering-from-bad-input

fn suffers_from_length_issues(&self, value: &DOMString) -> ValidationFlags

https://html.spec.whatwg.org/multipage/#suffering-from-being-too-short

fn suffers_from_range_issues(&self, value: &DOMString) -> ValidationFlags

https://html.spec.whatwg.org/multipage/#suffering-from-an-underflow

https://html.spec.whatwg.org/multipage/#suffering-from-an-overflow

https://html.spec.whatwg.org/multipage/#suffering-from-a-step-mismatch

fn get_or_create_shadow_tree( &self, can_gc: CanGc, ) -> Ref<'_, InputElementShadowTree>

Get the shadow tree for this HTMLInputElement, if it is created and valid, otherwise recreate the shadow tree and return it.

pub(crate) fn renders_as_text_input_widget(&self) -> bool

Whether this input type renders as a basic text input widget.

TODO(#38251): This should eventually only include text, password, url, tel, and email, but the others do not yet have a custom shadow DOM implementation.

fn may_have_embedder_control(&self) -> bool

fn handle_key_reaction(&self, action: KeyReaction, event: &Event, can_gc: CanGc)

fn value_for_shadow_dom(&self) -> DOMString

Return a string that represents the contents of the element in its displayed shadow DOM.

impl HTMLInputElement

fn radio_group_updated(&self, group: Option<&Atom>, can_gc: CanGc)

pub(crate) fn form_datums( &self, submitter: Option<FormSubmitterElement<'_>>, encoding: Option<&'static Encoding>, ) -> Vec<FormDatum>

https://html.spec.whatwg.org/multipage/#constructing-the-form-data-set Steps range from 5.1 to 5.10 (specific to HTMLInputElement)

fn radio_group_name(&self) -> Option<Atom>

https://html.spec.whatwg.org/multipage/#radio-button-group

fn update_checked_state(&self, checked: bool, dirty: bool, can_gc: CanGc)

pub(crate) fn is_mutable(&self) -> bool

pub(crate) fn reset(&self, can_gc: CanGc)

pub(crate) fn clear(&self, can_gc: CanGc)

https://w3c.github.io/webdriver/#ref-for-dfn-clear-algorithm-3 Used by WebDriver to clear the input element.

fn update_placeholder_shown_state(&self)

pub(crate) fn select_files_for_webdriver( &self, test_paths: Vec<DOMString>, response_sender: GenericSender<Result<bool, ErrorStatus>>, )

pub(crate) fn select_files(&self, test_paths: Option<Vec<DOMString>>)

Select files by invoking UI or by passed in argument.

https://html.spec.whatwg.org/multipage/#file-upload-state-(type=file)

fn sanitize_value(&self, value: &mut DOMString)

https://html.spec.whatwg.org/multipage/#value-sanitization-algorithm

fn selection(&self) -> TextControlSelection<'_, Self>

fn implicit_submission(&self, can_gc: CanGc)

fn convert_string_to_number(&self, value: &str) -> Option<f64>

https://html.spec.whatwg.org/multipage/#concept-input-value-string-number

fn convert_number_to_string(&self, value: f64) -> Option<DOMString>

https://html.spec.whatwg.org/multipage/#concept-input-value-string-number

fn convert_string_to_naive_datetime( &self, value: DOMString, ) -> Option<OffsetDateTime>

fn convert_datetime_to_dom_string(&self, value: OffsetDateTime) -> DOMString

https://html.spec.whatwg.org/multipage/#concept-input-value-date-string This does the safe Rust part of conversion; the unsafe JS Date part is in SetValueAsDate

fn update_related_validity_states(&self, can_gc: CanGc)

fn value_changed(&self, can_gc: CanGc)

fn show_the_picker_if_applicable(&self)

https://html.spec.whatwg.org/multipage/#show-the-picker,-if-applicable

pub(crate) fn handle_color_picker_response( &self, response: Option<RgbColor>, can_gc: CanGc, )

pub(crate) fn handle_file_picker_response( &self, response: Option<Vec<SelectedFile>>, can_gc: CanGc, )

fn handle_focus_event(&self, event: &FocusEvent)