script::dom::html::scripting::htmlscriptelement

Struct HTMLScriptElement

#[repr(C)]
pub(crate) struct HTMLScriptElement {

    htmlelement: HTMLElement,
    delaying_the_load_event: DomRefCell<Option<LoadBlocker>>,
    already_started: Cell<bool>,
    parser_inserted: Cell<bool>,
    non_blocking: Cell<bool>,
    parser_document: Dom<Document>,
    preparation_time_document: MutNullableDom<Document>,
    line_number: u64,
    script_text: DomRefCell<DOMString>,
    from_an_external_file: Cell<bool>,
    blocking: MutNullableDom<DOMTokenList>,
    marked_as_render_blocking: Cell<bool>,
    result: DomRefCell<Option<Result<Script, ()>>>,

}

Fields
htmlelement: HTMLElementdelaying_the_load_event: DomRefCell<Option<LoadBlocker>>

https://html.spec.whatwg.org/multipage/#concept-script-delay-load

already_started: Cell<bool>

https://html.spec.whatwg.org/multipage/#already-started

parser_inserted: Cell<bool>

https://html.spec.whatwg.org/multipage/#parser-inserted

non_blocking: Cell<bool>

https://html.spec.whatwg.org/multipage/#non-blocking

(currently unused)

parser_document: Dom<Document>

Document of the parser that created this element https://html.spec.whatwg.org/multipage/#parser-document

preparation_time_document: MutNullableDom<Document>

Prevents scripts that move between documents during preparation from executing. https://html.spec.whatwg.org/multipage/#preparation-time-document

line_number: u64

Track line line_number

script_text: DomRefCell<DOMString>

https://w3c.github.io/trusted-types/dist/spec/#htmlscriptelement-script-text

from_an_external_file: Cell<bool>

https://html.spec.whatwg.org/multipage/#concept-script-external

blocking: MutNullableDom<DOMTokenList>

https://html.spec.whatwg.org/multipage/#dom-script-blocking

marked_as_render_blocking: Cell<bool>

Used to keep track whether we consider this script element render blocking during prepare

result: DomRefCell<Option<Result<Script, ()>>>

https://html.spec.whatwg.org/multipage/#concept-script-result

Implementations

impl HTMLScriptElement

fn __assert_parent_type(&self)

impl HTMLScriptElement

fn new_inherited( local_name: LocalName, prefix: Option<Prefix>, document: &Document, creator: ElementCreator, ) -> HTMLScriptElement

pub(crate) fn new( cx: &mut JSContext, local_name: LocalName, prefix: Option<Prefix>, document: &Document, proto: Option<HandleObject<'_>>, creator: ElementCreator, ) -> DomRoot<HTMLScriptElement>

fn delay_load_event(&self, document: &Document, url: ServoUrl)

Marks that element as delaying the load event or not.

https://html.spec.whatwg.org/multipage/#concept-script-delay-load https://html.spec.whatwg.org/multipage/#delaying-the-load-event-flag

fn get_script_kind(&self, script_type: ScriptType) -> ExternalScriptKind

Helper method to determine the script kind based on attributes and insertion context.

This duplicates the script preparation logic from the HTML spec to determine the script’s active document without full preparation.

https://html.spec.whatwg.org/multipage/#prepare-the-script-element

fn get_script_active_document( &self, script_kind: ExternalScriptKind, ) -> DomRoot<Document>

https://html.spec.whatwg.org/multipage/#prepare-the-script-element

impl HTMLScriptElement

pub(crate) fn set_initial_script_text(&self)

https://w3c.github.io/trusted-types/dist/spec/#setting-slot-values-from-parser

fn prepare_the_script_text(&self, cx: &mut JSContext) -> Fallible<()>

https://w3c.github.io/trusted-types/dist/spec/#abstract-opdef-prepare-the-script-text

fn has_render_blocking_attribute(&self) -> bool

fn potentially_render_blocking(&self) -> bool

https://html.spec.whatwg.org/multipage/#potentially-render-blocking

pub(crate) fn prepare( &self, cx: &mut JSContext, introduction_type_override: Option<&'static CStr>, )

https://html.spec.whatwg.org/multipage/#prepare-the-script-element

pub(crate) fn execute(&self, cx: &mut JSContext, result: Result<Script, ()>)

https://html.spec.whatwg.org/multipage/#execute-the-script-element

pub(crate) fn queue_error_event(&self)

pub(crate) fn get_script_type(&self) -> Option<ScriptType>

pub(crate) fn set_parser_inserted(&self, parser_inserted: bool)

pub(crate) fn set_already_started(&self, already_started: bool)

fn text(&self) -> DOMString