Struct script::dom::htmlelement::HTMLElement

#[repr(C)]
pub struct HTMLElement {
    element: Element,
    style_decl: MutNullableDom<CSSStyleDeclaration>,
    dataset: MutNullableDom<DOMStringMap>,
}

Fields
element: Elementstyle_decl: MutNullableDom<CSSStyleDeclaration>dataset: MutNullableDom<DOMStringMap>
Implementations

impl HTMLElement

fn __assert_parent_type(&self)

impl HTMLElement

pub fn new_inherited( tag_name: LocalName, prefix: Option<Prefix>, document: &Document, ) -> HTMLElement

pub fn new_inherited_with_state( state: ElementState, tag_name: LocalName, prefix: Option<Prefix>, document: &Document, ) -> HTMLElement

pub fn new( local_name: LocalName, prefix: Option<Prefix>, document: &Document, proto: Option<HandleObject<'_>>, can_gc: CanGc, ) -> Root<Dom<HTMLElement>>

fn is_body_or_frameset(&self) -> bool

fn get_inner_outer_text(&self, can_gc: CanGc) -> DOMString

Calls into the layout engine to generate a plain text representation of a HTMLElement as specified when getting the .innerText or .outerText in JavaScript.`

https://html.spec.whatwg.org/multipage/#get-the-text-steps

impl HTMLElement

pub fn set_custom_attr( &self, name: DOMString, value: DOMString, can_gc: CanGc, ) -> Result<(), Error>

pub fn get_custom_attr(&self, local_name: DOMString) -> Option<DOMString>

pub fn delete_custom_attr(&self, local_name: DOMString)

pub fn is_labelable_element(&self) -> bool

https://html.spec.whatwg.org/multipage/#category-label

pub fn is_form_associated_custom_element(&self) -> bool

https://html.spec.whatwg.org/multipage/#form-associated-custom-element

pub fn is_listed_element(&self) -> bool

https://html.spec.whatwg.org/multipage/#category-listed

pub fn is_submittable_element(&self) -> bool

https://html.spec.whatwg.org/multipage/#category-submit

pub fn supported_prop_names_custom_attr(&self) -> Vec<DOMString>

pub fn label_at(&self, index: u32) -> Option<Root<Dom<Node>>>

pub fn labels_count(&self) -> u32

pub fn directionality(&self) -> Option<String>

pub fn summary_activation_behavior(&self)

fn is_summary_for_its_parent_details(&self) -> bool

fn rendered_text_fragment( &self, input: DOMString, can_gc: CanGc, ) -> Root<Dom<DocumentFragment>>

https://html.spec.whatwg.org/multipage/#rendered-text-fragment

fn merge_with_the_next_text_node(node: Root<Dom<Node>>)

Checks whether a given DomRoot<Node> and its next sibling are of type Text, and if so merges them into a single Text node.

https://html.spec.whatwg.org/multipage/#merge-with-the-next-text-node