Struct script::dom::htmlimageelement::HTMLImageElement

pub struct HTMLImageElement {
    htmlelement: HTMLElement,
    current_request: DOMRefCell<ImageRequest>,
    pending_request: DOMRefCell<ImageRequest>,
}

Fields

htmlelement

current_request

pending_request

Methods

impl HTMLImageElement

fn get_url(&self) -> Option<Url>

impl HTMLImageElement

fn update_image(&self, value: Option<(DOMString, Url)>)

Makes the local image member match the status of the src attribute and starts prefetching the image. This method must be called after src is changed.

fn new_inherited(localName: Atom, prefix: Option<DOMString>, document: &Document) -> HTMLImageElement

fn new(localName: Atom, prefix: Option<DOMString>, document: &Document) -> Root<HTMLImageElement>

fn Image(global: GlobalRef, width: Option<u32>, height: Option<u32>) -> Fallible<Root<HTMLImageElement>>