Struct script::dom::htmlmediaelement::HTMLMediaElementContext

struct HTMLMediaElementContext {
    elem: Trusted<HTMLMediaElement>,
    data: Vec<u8>,
    metadata: Option<Metadata>,
    generation_id: u32,
    next_progress_event: Timespec,
    url: ServoUrl,
    have_metadata: bool,
    ignore_response: bool,
}

Fields
elem: Trusted<HTMLMediaElement>

The element that initiated the request.

data: Vec<u8>

The response body received to date.

metadata: Option<Metadata>

The response metadata received to date.

generation_id: u32

The generation of the media element when this fetch started.

next_progress_event: Timespec

Time of last progress notification.

url: ServoUrl

Url of resource requested.

have_metadata: bool

Whether the media metadata has been completely received.

ignore_response: bool

True if this response is invalid and should be ignored.

Methods

impl HTMLMediaElementContext

fn new(elem: &HTMLMediaElement, url: ServoUrl) -> HTMLMediaElementContext

fn check_metadata(&mut self, elem: &HTMLMediaElement)