Module script::dom::htmlscriptelement

Structs

ClassicContext

The context required for asynchronously loading an external script source.

CompiledSourceCode

HTMLScriptElement

OffThreadCompilationContext

OffThreadCompilationToken

A wrapper to mark OffThreadToken as Send, which should be safe according to mozjs/js/public/OffThreadScriptCompilation.h

ScriptId

An unique id for script element.

ScriptOrigin

Enums

ExternalScriptKind

ScriptType

SourceCode

Constants

_IMPL_DOMOBJECT_FOR_HTMLScriptElement

Statics

SCRIPT_JS_MIMES

Supported script types as defined by https://html.spec.whatwg.org/multipage/#javascript-mime-type.

Functions

fetch_a_classic_script

https://html.spec.whatwg.org/multipage/#fetch-a-classic-script

finish_fetching_a_classic_script

Final steps of https://html.spec.whatwg.org/multipage/#fetch-a-classic-script

off_thread_compilation_callback⚠

script_fetch_request

Steps 1-2 of https://html.spec.whatwg.org/multipage/#fetch-a-classic-script

Type Definitions

ScriptResult