script::dom::indexeddb::idbtransaction

Struct IDBTransaction

#[repr(C)]
pub struct IDBTransaction {

    eventtarget: EventTarget,
    object_store_names: Dom<DOMStringList>,
    mode: IDBTransactionMode,
    durability: IDBTransactionDurability,
    db: Dom<IDBDatabase>,
    error: MutNullableDom<DOMException>,
    store_handles: DomRefCell<HashMap<String, Dom<IDBObjectStore>>>,
    requests: DomRefCell<Vec<Dom<IDBRequest>>>,
    active: Cell<bool>,
    finished: Cell<bool>,
    abort_initiated: Cell<bool>,
    abort_requested: Cell<bool>,
    committing: Cell<bool>,
    commit_started: Cell<bool>,
    version_change_old_version: Cell<Option<u64>>,
    version_change_old_object_store_names: DomRefCell<Option<Vec<DOMString>>>,
    cleanup_event_loop: Cell<Option<ScriptEventLoopId>>,
    registered_in_global: Cell<bool>,
    pending_request_count: Cell<usize>,
    next_request_id: Cell<u64>,
    next_unhandled_request_id: Cell<u64>,
    handled_pending: DomRefCell<HashSet<u64>>,
    serial_number: u64,

}

Fields
eventtarget: EventTargetobject_store_names: Dom<DOMStringList>mode: IDBTransactionModedurability: IDBTransactionDurabilitydb: Dom<IDBDatabase>error: MutNullableDom<DOMException>store_handles: DomRefCell<HashMap<String, Dom<IDBObjectStore>>>requests: DomRefCell<Vec<Dom<IDBRequest>>>active: Cell<bool>finished: Cell<bool>abort_initiated: Cell<bool>abort_requested: Cell<bool>committing: Cell<bool>commit_started: Cell<bool>version_change_old_version: Cell<Option<u64>>version_change_old_object_store_names: DomRefCell<Option<Vec<DOMString>>>cleanup_event_loop: Cell<Option<ScriptEventLoopId>>registered_in_global: Cell<bool>pending_request_count: Cell<usize>next_request_id: Cell<u64>next_unhandled_request_id: Cell<u64>handled_pending: DomRefCell<HashSet<u64>>serial_number: u64
Implementations

impl IDBTransaction

fn __assert_parent_type(&self)

impl IDBTransaction

fn new_inherited( connection: &IDBDatabase, mode: IDBTransactionMode, durability: IDBTransactionDurability, scope: &DOMStringList, serial_number: u64, ) -> IDBTransaction

pub fn new( cx: &mut JSContext, global: &GlobalScope, connection: &IDBDatabase, mode: IDBTransactionMode, durability: IDBTransactionDurability, scope: &DOMStringList, ) -> DomRoot<IDBTransaction>

Does a blocking call to create a backend transaction and get its id.

pub(crate) fn new_with_serial( cx: &mut JSContext, global: &GlobalScope, connection: &IDBDatabase, mode: IDBTransactionMode, durability: IDBTransactionDurability, scope: &DOMStringList, serial_number: u64, ) -> DomRoot<IDBTransaction>

fn create_transaction( global: &GlobalScope, db_name: DOMString, mode: IDBTransactionMode, scope: &DOMStringList, ) -> u64

pub fn set_active_flag(&self, status: bool)

https://w3c.github.io/IndexedDB/#transaction-lifecycle

pub fn is_active(&self) -> bool

pub(crate) fn is_usable(&self) -> bool

https://w3c.github.io/IndexedDB/#transaction-lifetime

pub(crate) fn is_inactive(&self) -> bool

pub(crate) fn is_committing(&self) -> bool

pub(crate) fn is_finished(&self) -> bool

pub(crate) fn set_cleanup_event_loop(&self)

pub(crate) fn clear_cleanup_event_loop(&self)

pub(crate) fn cleanup_event_loop_matches_current(&self) -> bool

pub(crate) fn set_registered_in_global(&self)

pub(crate) fn clear_registered_in_global(&self)

pub(crate) fn set_versionchange_old_version(&self, version: u64)

pub(crate) fn register_object_store_handle( &self, name: &DOMString, store: &IDBObjectStore, )

pub(crate) fn rename_object_store_handle_cache( &self, old_name: &DOMString, new_name: &DOMString, store: &IDBObjectStore, )

fn restore_associated_object_store_handles_after_abort( &self, cx: &mut JSContext, )

https://w3c.github.io/IndexedDB/#abort-an-upgrade-transaction

fn attempt_commit(&self) -> bool

pub(crate) fn maybe_commit(&self, cx: &mut JSContext)

fn force_commit(&self)

pub fn get_mode(&self) -> IDBTransactionMode

pub fn get_db_name(&self) -> DOMString

pub(crate) fn get_db(&self) -> &IDBDatabase

pub fn get_serial_number(&self) -> u64

pub(crate) fn issued_count(&self) -> u64

pub(crate) fn allocate_request_id(&self) -> u64

request_id is only required to be unique within this transaction. The backend keys “handled” state by (txn, request_id).

pub(crate) fn mark_request_handled(&self, request_id: u64)

pub fn add_request(&self, request: &IDBRequest)

pub fn request_finished(&self)

pub(crate) fn initiate_abort(&self, cx: &mut JSContext, error: Error)

pub(crate) fn request_backend_abort(&self)

fn notify_backend_transaction_finished(&self)

pub(crate) fn finalize_abort(&self)

pub(crate) fn finalize_commit(&self)

fn dispatch_complete(&self)

fn get_idb_thread(&self) -> GenericSender<IndexedDBThreadMsg>

fn object_store_parameters( &self, object_store_name: &DOMString, ) -> Option<(IDBObjectStoreParameters, Vec<IndexedDBIndex>, Option<i64>)>

pub(crate) fn create_abort_callback(&self) -> GenericCallback<BackendError>