script::dom

Module intersectionobserver

Structs

IntersectionObservationOutput 🔒

The values from computing step 2.2.4-2.2.14 in https://w3c.github.io/IntersectionObserver/#update-intersection-observations-algo. See IntersectionObserver::maybe_compute_intersection_output.

IntersectionObserver 🔒

The Intersection Observer interface

IntersectionObserverRegistration 🔒

https://w3c.github.io/IntersectionObserver/#intersectionobserverregistration

Constants

_IMPL_DOMOBJECT_FOR_IntersectionObserver 🔒

Functions

compute_the_intersection 🔒

https://w3c.github.io/IntersectionObserver/#compute-the-intersection

parse_a_margin 🔒

https://w3c.github.io/IntersectionObserver/#parse-a-margin

Type Aliases

IntersectionRoot

The intersection root for an IntersectionObserver is the value of its root attribute if the attribute is non-null; otherwise, it is the top-level browsing context’s document node, referred to as the implicit root.