script::dom::intersectionobserver::intersectionobserver

Function compute_the_intersection

fn compute_the_intersection(
    target: &Element,
    root: &ElementOrDocument<DomTypeHolder>,
    root_bounds: Rect<Au, CSSPixel>,
    intersection_rect: Rect<Au, CSSPixel>,
    scroll_margin: &IntersectionObserverMargin,
) -> Option<Rect<Au, CSSPixel>>

Compute the intersection rectangle of the target Element returning the results of intersection in the coordinate space of the target’s owning Document. Additionally, we assume that both the target and the root is connected. https://w3c.github.io/IntersectionObserver/#compute-the-intersection