script::dom

Module macros

Macros

beforeunload_event_handler 🔒

define_event_handler 🔒

These are used to generate a event handler which has no special case.

define_window_owned_event_handler 🔒

error_event_handler 🔒

event_handler 🔒

global_event_handlers 🔒

handle_potential_webgl_error 🔒

make_enumerated_getter 🔒

Implements the To determine the state of an attribute steps from https://html.spec.whatwg.org/multipage/#keywords-and-enumerated-attributes

registered_event_handler 🔒

Similar to event_handler!, but also registers/unregisters a [ConstellationInterest] with the global scope when the handler is set or cleared. Use this macro for event handlers whose corresponding events are sent by the constellation only to interested pipelines.

unsafe_no_jsmanaged_fields 🔒

For use on non-jsmanaged types Use #[derive(JSTraceable)] on JS managed types

window_event_handlers 🔒

window_owned_beforeunload_event_handler 🔒

window_owned_event_handler 🔒