Module script::dom::node

The core DOM types. Defines the basic DOM hierarchy as well as all the HTML elements.

Structs

BindContext
The context of the binding to tree of a node.

FollowingNodeIterator

Node
An HTML node.

NodeFlags
Flags for node items.

PrecedingNodeIterator

QuerySelectorIterator

SimpleNodeIterator

TreeIterator

UnbindContext
The context of the unbinding from a tree of a node when one of its inclusive ancestors is removed.

UniqueId
A node's unique ID, for devtools.

Enums

ChildrenMutation

CloneChildrenFlag
Specifies whether children must be recursively cloned or not.

NodeDamage
A summary of the changes that happened to a node.

ShadowIncluding
Whether a tree traversal should pass shadow tree boundaries.

SuppressObserver
suppress observers flag https://dom.spec.whatwg.org/#concept-node-insert https://dom.spec.whatwg.org/#concept-node-remove

Constants

_IMPL_DOMOBJECT_FOR_Node

Traits

LayoutNodeHelpers

VecPreOrderInsertionHelper
Helper trait to insert an element into vector whose elements are maintained in tree order

Functions

as_uintptr

containing_shadow_root

document_from_node

first_node_not_in
Iterate through nodes until we find a Node that is not in not_in

from_untrusted_node_address⚠
If the given untrusted node address represents a valid DOM node in the given runtime, returns it.

stylesheets_owner_from_node

window_from_node