Enum script::dom::node::ChildrenMutation

pub enum ChildrenMutation<'a> {
    Append {
        prev: &'a Node,
        added: &'a [&'a Node],
    },
    Insert {
        prev: &'a Node,
        added: &'a [&'a Node],
        next: &'a Node,
    },
    Prepend {
        added: &'a [&'a Node],
        next: &'a Node,
    },
    Replace {
        prev: Option<&'a Node>,
        removed: &'a Node,
        added: &'a [&'a Node],
        next: Option<&'a Node>,
    },
    ReplaceAll {
        removed: &'a [&'a Node],
        added: &'a [&'a Node],
    },
}

Variants
Append
Fields

prev: &'a Node

added: &'a [&'a Node]

Insert
Fields

prev: &'a Node

added: &'a [&'a Node]

next: &'a Node

Prepend
Fields

added: &'a [&'a Node]

next: &'a Node

Replace
Fields

prev: Option<&'a Node>

removed: &'a Node

added: &'a [&'a Node]

next: Option<&'a Node>

ReplaceAll
Fields

removed: &'a [&'a Node]

added: &'a [&'a Node]

Methods

impl<'a> ChildrenMutation<'a>

fn insert(prev: Option<&'a Node>, added: &'a [&'a Node], next: Option<&'a Node>) -> ChildrenMutation<'a>

fn replace(prev: Option<&'a Node>, removed: &'a Option<&'a Node>, added: &'a [&'a Node], next: Option<&'a Node>) -> ChildrenMutation<'a>

fn replace_all(removed: &'a [&'a Node], added: &'a [&'a Node]) -> ChildrenMutation<'a>

fn next_child(&self) -> Option<&Node>

Get the child that follows the added or removed children.

fn modified_edge_element(&self) -> Option<Root<Node>>

If nodes were added or removed at the start or end of a container, return any previously-existing child whose ":first-child" or ":last-child" status may have changed.

NOTE: This does not check whether the inserted/removed nodes were elements, so in some cases it will return a false positive. This doesn't matter for correctness, because at worst the returned element will be restyled unnecessarily.