script::dom::node

Module iterators

Structs

FollowingNodeIterator 🔒

PrecedingNodeIterator 🔒

SimpleNodeIterator 🔒

TreeIterator 🔒

UnrootedFollowingFlatTreeNodesTraversal 🔒

A traversal of a [Document]’s flat tree. This is both a pre-order and post-order unrooted traversal. This means that an item is returned both when encountering a node for the first time and when leaving a node. In addition, no garbage collection can happen while iterating, allowing returning unrooted values for performance reasons.

UnrootedFollowingNodeIterator 🔒

UnrootedPrecedingNodeIterator 🔒

UnrootedSimpleNodeIterator 🔒

An efficient SimpleNodeIterator because it skips rooting if there are no GC pauses.

UnrootedTreeIterator 🔒

An efficient TreeIterator because it skips rooting if there are no GC pauses.

Enums

PrePostIteration 🔒

An Item in an traversal that is both pre-order and post-order. Each iteration of the traversal is either an record of entering a node or a leaving a node during the course of traversal.

ShadowIncluding 🔒

Whether a tree traversal should pass shadow tree boundaries.