script::dom::node

Module node

The core DOM types. Defines the basic DOM hierarchy as well as all the HTML elements.

Structs

Node

An HTML node.

NodeFlags 🔒

Flags for node items

UniqueId 🔒

A node’s unique ID, for devtools.

Enums

CloneChildrenFlag 🔒

Specifies whether children must be recursively cloned or not.

FlatTreeParent 🔒

The return value of Node::parent_in_flat_tree.

ForceSlottableNodeReconciliation 🔒

NodeDamage 🔒

A summary of the changes that happened to a node.

SuppressObserver 🔒

suppress observers flag https://dom.spec.whatwg.org/#insert-suppressobservers https://dom.spec.whatwg.org/#remove-suppressobservers

Constants

_IMPL_DOMOBJECT_FOR_Node 🔒

Traits

NodeTraits 🔒

VecPreOrderInsertionHelper 🔒

Helper trait to insert an element into vector whose elements are maintained in tree order

Functions

as_uintptr 🔒

first_node_not_in 🔒

Iterate through nodes until we find a Node that is not in not_in

from_untrusted_node_address 🔒 ⚠

If the given untrusted node address represents a valid DOM node in the given runtime, returns it.