script::dom::node::node

Struct Node

#[repr(C)]
pub struct Node {
    eventtarget: EventTarget,
    parent_node: MutNullableDom<Node>,
    first_child: MutNullableDom<Node>,
    last_child: MutNullableDom<Node>,
    next_sibling: MutNullableDom<Node>,
    prev_sibling: MutNullableDom<Node>,
    owner_doc: MutNullableDom<Document>,
    rare_data: DomRefCell<Option<Box<NodeRareData>>>,
    children_count: Cell<u32>,
    flags: Cell<NodeFlags>,
    inclusive_descendants_version: Cell<u64>,
    layout_data: DomRefCell<Option<Box<GenericLayoutData>>>,
}

An HTML node.

Fields
eventtarget: EventTarget

The JavaScript reflector for this node.

parent_node: MutNullableDom<Node>

The parent of this node.

first_child: MutNullableDom<Node>

The first child of this node.

last_child: MutNullableDom<Node>

The last child of this node.

next_sibling: MutNullableDom<Node>

The next sibling of this node.

prev_sibling: MutNullableDom<Node>

The previous sibling of this node.

owner_doc: MutNullableDom<Document>

The document that this node belongs to.

rare_data: DomRefCell<Option<Box<NodeRareData>>>

Rare node data.

children_count: Cell<u32>

The live count of children of this node.

flags: Cell<NodeFlags>

A bitfield of flags for node items.

inclusive_descendants_version: Cell<u64>

The maximum version of any inclusive descendant of this node.

layout_data: DomRefCell<Option<Box<GenericLayoutData>>>

Layout data for this node. This is populated during layout and can be used for incremental relayout and script queries.

Implementations

impl Node

fn as_text_input(&self) -> Option<DomRoot<Element>>

impl Node

pub(crate) fn push_down_values( &self, cx: &mut JSContext, command: &CommandName, new_value: Option<DOMString>, )

https://w3c.github.io/editing/docs/execCommand/#push-down-values

fn reorder_modifiable_descendants( &self, cx: &mut JSContext, command: &CommandName, new_value: &DOMString, )

https://w3c.github.io/editing/docs/execCommand/#reorder-modifiable-descendants

pub(crate) fn force_the_value( &self, cx: &mut JSContext, command: &CommandName, new_value: Option<&DOMString>, )

https://w3c.github.io/editing/docs/execCommand/#force-the-value

pub(crate) fn same_editing_host(&self, other: &Node) -> bool

https://w3c.github.io/editing/docs/execCommand/#in-the-same-editing-host

pub(crate) fn is_block_node(&self) -> bool

https://w3c.github.io/editing/docs/execCommand/#block-node

pub(crate) fn is_inline_node(&self) -> bool

https://w3c.github.io/editing/docs/execCommand/#inline-node

pub(crate) fn block_node_of(&self) -> Option<DomRoot<Node>>

https://w3c.github.io/editing/docs/execCommand/#block-node-of

pub(crate) fn is_visible(&self, no_gc: &NoGC) -> bool

https://w3c.github.io/editing/docs/execCommand/#visible

pub(crate) fn is_invisible(&self, no_gc: &NoGC) -> bool

https://w3c.github.io/editing/docs/execCommand/#invisible

pub(crate) fn is_formattable(&self, no_gc: &NoGC) -> bool

https://w3c.github.io/editing/docs/execCommand/#formattable-node

pub(crate) fn is_single_line_container(&self) -> bool

https://w3c.github.io/editing/docs/execCommand/#single-line-container

pub(crate) fn is_block_start_point(&self, no_gc: &NoGC, offset: usize) -> bool

https://w3c.github.io/editing/docs/execCommand/#block-start-point

pub(crate) fn is_block_end_point(&self, offset: u32, no_gc: &NoGC) -> bool

https://w3c.github.io/editing/docs/execCommand/#block-end-point

pub(crate) fn is_block_boundary_point(&self, no_gc: &NoGC, offset: u32) -> bool

https://w3c.github.io/editing/docs/execCommand/#block-boundary-point

pub(crate) fn is_no_allowed_child_in_same_editing_host( &self, no_gc: &NoGC, ) -> bool

pub(crate) fn is_prohibited_paragraph_child(&self) -> bool

https://w3c.github.io/editing/docs/execCommand/#prohibited-paragraph-child

pub(crate) fn fix_disallowed_ancestors( &self, cx: &mut JSContext, context_object: &Document, )

https://w3c.github.io/editing/docs/execCommand/#fix-disallowed-ancestors

pub(crate) fn is_collapsed_block_prop(&self, no_gc: &NoGC) -> bool

https://w3c.github.io/editing/docs/execCommand/#collapsed-block-prop

fn follows_a_line_break(&self, no_gc: &NoGC) -> bool

https://w3c.github.io/editing/docs/execCommand/#follows-a-line-break

fn precedes_a_line_break(&self, no_gc: &NoGC) -> bool

https://w3c.github.io/editing/docs/execCommand/#precedes-a-line-break

fn canonical_space_sequence( n: usize, non_breaking_start: bool, non_breaking_end: bool, ) -> String

https://w3c.github.io/editing/docs/execCommand/#canonical-space-sequence

pub(crate) fn canonicalize_whitespace( &self, cx: &mut JSContext, offset: u32, fix_collapsed_space: bool, )

https://w3c.github.io/editing/docs/execCommand/#canonicalize-whitespace

fn remove_extraneous_line_breaks_before(&self, cx: &mut JSContext)

https://w3c.github.io/editing/docs/execCommand/#remove-extraneous-line-breaks-before

pub(crate) fn remove_extraneous_line_breaks_at_the_end_of( &self, cx: &mut JSContext, )

https://w3c.github.io/editing/docs/execCommand/#remove-extraneous-line-breaks-at-the-end-of

fn remove_extraneous_line_breaks_from(&self, cx: &mut JSContext)

https://w3c.github.io/editing/docs/execCommand/#remove-extraneous-line-breaks-from

pub(crate) fn remove_preserving_its_descendants(&self, cx: &mut JSContext)

https://w3c.github.io/editing/docs/execCommand/#preserving-its-descendants

pub(crate) fn effective_command_value( &self, command: &CommandName, ) -> Option<DOMString>

https://w3c.github.io/editing/docs/execCommand/#effective-command-value

impl Node

fn is_in_plaintext_only_state(&self) -> bool

impl Node

pub(crate) fn find_click_focusable_area(&self, cx: &JSContext) -> FocusableArea

Returns the appropriate FocusableArea when this Node is clicked on according to https://www.w3.org/TR/pointerevents4/#handle-native-mouse-down.

Note that this is doing more than the specification which says to only take into account the node from the hit test. This isn’t exactly how browsers work though, as they seem to look for the first inclusive ancestor node that has a focusable area associated with it. Note also that this may return FocusableArea::Viewport.

pub(crate) fn get_the_focusable_area( &self, cx: &JSContext, focus_trigger: FocusTrigger, ) -> Option<FocusableArea>

https://html.spec.whatwg.org/multipage/#get-the-focusable-area

There seems to be hole in the specification here. It describes how to get the focusable area for a focus target that isn’t a focuable area, but is ambiguous about how to do this for a focus target that actually is a focusable area. The obvious thing is to just return the focus target, but it’s still odd that this isn’t mentioned in the specification.

fn get_the_focusable_area_if_not_a_focusable_area( &self, cx: &JSContext, focus_trigger: FocusTrigger, ) -> Option<FocusableArea>

https://html.spec.whatwg.org/multipage/#get-the-focusable-area

In addition to returning the DOM anchor of the focusable area for this Node, this method also returns the FocusableAreaKind for efficiency reasons. Note that None is returned if this Node does not have a focusable area or if its focusable area is the Document’s viewport.

TODO: It might be better to distinguish these two cases in the future.

pub(crate) fn focus_delegate( &self, cx: &JSContext, focus_trigger: Option<FocusTrigger>, ) -> Option<FocusableArea>

https://html.spec.whatwg.org/multipage/#focus-delegate

In addition to returning the focus delegate for this Node, this method also returns the FocusableAreaKind for efficiency reasons.

fn autofocus_delegate( &self, cx: &JSContext, focus_trigger: FocusTrigger, ) -> Option<FocusableArea>

https://html.spec.whatwg.org/multipage/#autofocus-delegate

pub(crate) fn run_the_focusing_steps( &self, cx: &mut JSContext, fallback_target: Option<FocusableArea>, focus_trigger: FocusTrigger, ) -> bool

https://html.spec.whatwg.org/multipage/#focusing-steps

This is an initial implementation of the “focusing steps” from the HTML specification. Note that this is currently in a state of transition from Servo’s old internal focus APIs to ones that match the specification. That is why the arguments to this method do not match the specification yet.

Return true if anything was focused or false otherwise.

pub(crate) fn as_focus_navigation_scope_owner( &self, ) -> Option<FocusNavigationScopeOwner>

If this node is a focus navigation scope owner, return the corresponding FocusNavigationScopeOwner that describes it or None if this node is not a focus navigation scope owner.

https://html.spec.whatwg.org/multipage/#focus-navigation-scope-owner

pub(crate) fn containing_focus_navigation_scope_owner( &self, ) -> Option<FocusNavigationScopeOwner>

Find the focus navigation scope owner for this node. If this node is itself a focus navigation scope owner, this will return its containing focus navigation scope owner.

This will return None if this node is the Document element.

https://html.spec.whatwg.org/multipage/#focus-navigation-scope-owner

impl Node

pub(super) fn parent_node(&self) -> &MutNullableDom<Node>

pub(super) fn first_child(&self) -> &MutNullableDom<Node>

pub(super) fn last_child(&self) -> &MutNullableDom<Node>

pub(super) fn next_sibling(&self) -> &MutNullableDom<Node>

pub(super) fn prev_sibling(&self) -> &MutNullableDom<Node>

pub(super) fn get_owner_doc(&self) -> &MutNullableDom<Document>

pub(super) fn get_rare_data(&self) -> &DomRefCell<Option<Box<NodeRareData>>>

pub(super) fn flags(&self) -> &Cell<NodeFlags>

pub(crate) fn layout_data(&self) -> &DomRefCell<Option<Box<GenericLayoutData>>>

fn add_child(&self, cx: &mut JSContext, new_child: &Node, before: Option<&Node>)

Adds a new child to the end of this node’s list of children.

Fails unless new_child is disconnected from the tree.

pub(crate) fn complete_remove_subtree( cx: &mut JSContext, root: &Node, context: &UnbindContext<'_>, )

Clean up flags and runs steps 11-14 of remove a node. https://dom.spec.whatwg.org/#concept-node-remove

pub(crate) fn complete_move_subtree(cx: &mut JSContext, root: &Node)

fn remove_child( &self, cx: &mut JSContext, child: &Node, cached_index: Option<u32>, )

Removes the given child from this node’s list of children.

Fails unless child is a child of this node.

fn move_child(&self, cx: &mut JSContext, child: &Node)

pub(crate) fn to_opaque(&self) -> OpaqueNode

pub(crate) fn as_custom_element(&self) -> Option<DomRoot<Element>>

pub(crate) fn fire_synthetic_pointer_event_not_trusted( &self, cx: &mut JSContext, event_type: Atom, )

https://html.spec.whatwg.org/multipage/#fire-a-synthetic-pointer-event

pub(crate) fn parent_directionality(&self) -> String

pub(crate) fn is_being_rendered_or_delegates_rendering( &self, pseudo_element: Option<PseudoElement>, ) -> bool

Implements the combination of:

https://html.spec.whatwg.org/multipage/#being-rendered

https://html.spec.whatwg.org/multipage/#delegating-its-rendering-to-its-children

pub(crate) fn is_being_rendered( &self, pseudo_element: Option<PseudoElement>, ) -> bool

https://html.spec.whatwg.org/multipage/#being-rendered

pub(crate) fn add_pending_accessibility_damage( &self, damage: AccessibilityDamage, )

impl Node

fn ensure_rare_data(&self) -> RefMut<'_, Box<NodeRareData>>

pub(crate) fn is_before(&self, no_gc: &NoGC, other: &Node) -> bool

Returns true if this node is before other in the same connected DOM tree.

pub(crate) fn registered_mutation_observers_mut( &self, ) -> RefMut<'_, Vec<RegisteredObserver>>

Return all registered mutation observers for this node. Lazily initialize the raredata if it does not exist.

pub(crate) fn registered_mutation_observers( &self, ) -> Option<Ref<'_, Vec<RegisteredObserver>>>

pub(crate) fn add_mutation_observer(&self, observer: RegisteredObserver)

Add a new mutation observer for a given node.

pub(crate) fn remove_mutation_observer(&self, observer: &MutationObserver)

Removes the mutation observer for a given node.

pub(crate) fn debug_str(&self) -> String

Returns a string that describes this node.

pub(crate) fn is_in_a_document_tree(&self) -> bool

https://dom.spec.whatwg.org/#in-a-document-tree

pub(crate) fn is_in_a_shadow_tree(&self) -> bool

Return true iff node’s root is a shadow-root.

pub(crate) fn has_weird_parser_insertion_mode(&self) -> bool

pub(crate) fn set_weird_parser_insertion_mode(&self)

pub(crate) fn is_connected(&self) -> bool

https://dom.spec.whatwg.org/#connected

pub(crate) fn set_in_ua_widget(&self, in_ua_widget: bool)

pub(crate) fn is_in_ua_widget(&self) -> bool

pub(crate) fn type_id(&self) -> NodeTypeId

Returns the type ID of this node.

pub(crate) fn len(&self) -> u32

https://dom.spec.whatwg.org/#concept-node-length

pub(crate) fn is_empty(&self) -> bool

pub(crate) fn index(&self) -> u32

https://dom.spec.whatwg.org/#concept-tree-index

pub(crate) fn has_parent(&self) -> bool

Returns true if this node has a parent.

pub(crate) fn children_count(&self) -> u32

pub(crate) fn ensure_weak_ranges(&self) -> RefMut<'_, WeakRangeVec>

pub(crate) fn has_live_ranges(&self) -> bool

Whether or not this node has any live ranges.

pub(crate) fn live_ranges(&self) -> SmallVec<[DomRoot<Range>; 4]>

Return a SmallVec of all live ranges registered on this node.

pub(crate) fn is_doctype(&self) -> bool

pub(crate) fn get_flag(&self, flag: NodeFlags) -> bool

pub(crate) fn set_flag(&self, flag: NodeFlags, value: bool)

pub(crate) fn rev_version(&self, no_gc: &NoGC)

pub(crate) fn clear_layout_data(&self)

pub(crate) fn dirty(&self, no_gc: &NoGC, damage: NodeDamage)

pub(crate) fn inclusive_descendants_version(&self) -> u64

The maximum version number of this node’s descendants, including itself

pub(crate) fn traverse_preorder( &self, shadow_including: ShadowIncluding, ) -> TreeIterator

Iterates over this node and all its descendants, in preorder.

pub(crate) fn traverse_preorder_non_rooting<'b>( &self, no_gc: &'b NoGC, shadow_including: ShadowIncluding, ) -> UnrootedTreeIterator<'b>

Iterates over this node and all its descendants, in preorder. We take &NoGC to prevent GC which allows us to avoid rooting.

pub(crate) fn inclusively_following_siblings( &self, ) -> impl Iterator<Item = DomRoot<Node>> + use<>

pub(crate) fn inclusively_following_siblings_unrooted<'b>( &self, no_gc: &'b NoGC, ) -> impl Iterator<Item = UnrootedDom<'b, Node>> + use<'b>

pub(crate) fn inclusively_preceding_siblings_unrooted<'b>( &self, no_gc: &'b NoGC, ) -> impl Iterator<Item = UnrootedDom<'b, Node>> + use<'b>

pub(crate) fn common_ancestor( &self, other: &Node, shadow_including: ShadowIncluding, ) -> Option<DomRoot<Node>>

pub(crate) fn common_ancestor_in_flat_tree( &self, no_gc: &NoGC, other: &Node, ) -> Option<DomRoot<Node>>

pub(crate) fn following_flat_tree_nodes_unrooted<'no_gc>( &self, no_gc: &'no_gc NoGC, ) -> UnrootedFollowingFlatTreeNodesTraversal<'no_gc>

pub(crate) fn is_inclusive_ancestor_of(&self, child: &Node) -> bool

https://dom.spec.whatwg.org/#concept-tree-inclusive-ancestor

pub(crate) fn is_ancestor_of(&self, possible_descendant: &Node) -> bool

https://dom.spec.whatwg.org/#concept-tree-ancestor

fn is_host_including_inclusive_ancestor(&self, child: &Node) -> bool

https://dom.spec.whatwg.org/#concept-tree-host-including-inclusive-ancestor

pub(crate) fn is_shadow_including_inclusive_ancestor_of( &self, node: &Node, ) -> bool

https://dom.spec.whatwg.org/#concept-shadow-including-inclusive-ancestor

pub(crate) fn following_siblings( &self, ) -> impl Iterator<Item = DomRoot<Node>> + use<>

pub(crate) fn preceding_siblings( &self, ) -> impl Iterator<Item = DomRoot<Node>> + use<>

pub(crate) fn following_nodes( &self, root: &Node, shadow_including: ShadowIncluding, ) -> FollowingNodeIterator

pub(crate) fn following_nodes_unrooted<'b>( &self, no_gc: &'b NoGC, root: &Node, shadow_including: ShadowIncluding, ) -> UnrootedFollowingNodeIterator<'b>

pub(crate) fn preceding_nodes(&self, root: &Node) -> PrecedingNodeIterator

pub(crate) fn preceding_nodes_unrooted<'b>( &self, no_gc: &'b NoGC, root: &Node, ) -> UnrootedPrecedingNodeIterator<'b>

pub(crate) fn descending_last_children( &self, ) -> impl Iterator<Item = DomRoot<Node>> + use<>

Return an iterator that moves from self down the tree, choosing the last child at each step of the way.

pub(crate) fn descending_last_children_unrooted<'b>( &self, no_gc: &'b NoGC, ) -> impl Iterator<Item = UnrootedDom<'b, Node>>

pub(crate) fn is_parent_of(&self, child: &Node) -> bool

pub(crate) fn to_trusted_node_address(&self) -> TrustedNodeAddress

pub(crate) fn containing_block_node_without_reflow( &self, ) -> Option<DomRoot<Node>>

Return the node that establishes a containing block for this node.

pub(crate) fn padding(&self) -> Option<PhysicalSides>

pub(crate) fn content_box(&self) -> Option<Rect<Au, CSSPixel>>

pub(crate) fn border_box(&self) -> Option<Rect<Au, CSSPixel>>

pub(crate) fn border_box_without_reflow(&self) -> Option<Rect<Au, CSSPixel>>

pub(crate) fn padding_box(&self) -> Option<Rect<Au, CSSPixel>>

pub(crate) fn padding_box_without_reflow(&self) -> Option<Rect<Au, CSSPixel>>

pub(crate) fn border_boxes(&self) -> CSSPixelRectVec

pub(crate) fn client_rect(&self) -> Rect<i32, CSSPixel>

pub(crate) fn scroll_area(&self) -> Rect<i32, CSSPixel>

https://drafts.csswg.org/cssom-view/#dom-element-scrollwidth https://drafts.csswg.org/cssom-view/#dom-element-scrollheight

pub(crate) fn effective_overflow(&self) -> Option<AxesOverflow>

pub(crate) fn effective_overflow_without_reflow(&self) -> Option<AxesOverflow>

pub(crate) fn before( &self, cx: &mut JSContext, nodes: Vec<NodeOrString<DomTypeHolder>>, ) -> ErrorResult

https://dom.spec.whatwg.org/#dom-childnode-before

pub(crate) fn after( &self, cx: &mut JSContext, nodes: Vec<NodeOrString<DomTypeHolder>>, ) -> ErrorResult

https://dom.spec.whatwg.org/#dom-childnode-after

pub(crate) fn replace_with( &self, cx: &mut JSContext, nodes: Vec<NodeOrString<DomTypeHolder>>, ) -> ErrorResult

https://dom.spec.whatwg.org/#dom-childnode-replacewith

pub(crate) fn prepend( &self, cx: &mut JSContext, nodes: Vec<NodeOrString<DomTypeHolder>>, ) -> ErrorResult

https://dom.spec.whatwg.org/#dom-parentnode-prepend

pub(crate) fn append( &self, cx: &mut JSContext, nodes: Vec<NodeOrString<DomTypeHolder>>, ) -> ErrorResult

https://dom.spec.whatwg.org/#dom-parentnode-append

pub(crate) fn replace_children( &self, cx: &mut JSContext, nodes: Vec<NodeOrString<DomTypeHolder>>, ) -> ErrorResult

https://dom.spec.whatwg.org/#dom-parentnode-replacechildren

pub(crate) fn move_before( &self, cx: &mut JSContext, node: &Node, child: Option<&Node>, ) -> ErrorResult

https://dom.spec.whatwg.org/#dom-parentnode-movebefore

fn move_fn( cx: &mut JSContext, node: &Node, new_parent: &Node, child: Option<&Node>, ) -> ErrorResult

https://dom.spec.whatwg.org/#move

pub(crate) fn query_selector( &self, no_gc: &NoGC, selectors: DOMString, ) -> Fallible<Option<DomRoot<Element>>>

https://dom.spec.whatwg.org/#dom-parentnode-queryselector

pub(crate) fn query_selector_all( &self, cx: &mut JSContext, selectors: DOMString, ) -> Fallible<DomRoot<NodeList>>

https://dom.spec.whatwg.org/#dom-parentnode-queryselectorall

pub(crate) fn ancestors(&self) -> impl Iterator<Item = DomRoot<Node>> + use<>

pub(crate) fn ancestors_unrooted<'a>( &self, no_gc: &'a NoGC, ) -> impl Iterator<Item = UnrootedDom<'a, Node>> + use<'a>

pub(crate) fn inclusive_ancestors( &self, shadow_including: ShadowIncluding, ) -> impl Iterator<Item = DomRoot<Node>> + use<>

https://dom.spec.whatwg.org/#concept-shadow-including-inclusive-ancestor

pub(crate) fn inclusive_ancestors_unrooted<'a>( &self, no_gc: &'a NoGC, shadow_including: ShadowIncluding, ) -> impl Iterator<Item = UnrootedDom<'a, Node>> + use<'a>

pub(crate) fn owner_doc(&self) -> DomRoot<Document>

pub(crate) fn owner_doc_unrooted<'a>( &self, no_gc: &'a NoGC, ) -> UnrootedDom<'a, Document>

pub(crate) fn set_owner_doc(&self, document: &Document)

pub(crate) fn containing_shadow_root(&self) -> Option<DomRoot<ShadowRoot>>

pub(crate) fn set_containing_shadow_root( &self, shadow_root: Option<&ShadowRoot>, )

pub(crate) fn is_in_html_doc(&self) -> bool

pub(crate) fn is_connected_with_browsing_context(&self) -> bool

pub(crate) fn children(&self) -> impl Iterator<Item = DomRoot<Node>> + use<>

pub(crate) fn children_unrooted<'a>( &self, no_gc: &'a NoGC, ) -> impl Iterator<Item = UnrootedDom<'a, Node>> + use<'a>

pub(crate) fn rev_children(&self) -> impl Iterator<Item = DomRoot<Node>> + use<>

pub(crate) fn child_elements( &self, ) -> impl Iterator<Item = DomRoot<Element>> + use<>

Returns the children that are Elements

pub(crate) fn child_elements_unrooted<'a>( &self, no_gc: &'a NoGC, ) -> impl Iterator<Item = UnrootedDom<'a, Element>> + use<'a>

pub(crate) fn remove_self(&self, cx: &mut JSContext)

pub(crate) fn unique_id_if_already_present(&self) -> Option<String>

Returns the node’s unique_id if it has been computed before and None otherwise.

pub(crate) fn unique_id(&self, pipeline: PipelineId) -> String

pub(crate) fn summarize(&self, cx: &mut JSContext) -> NodeInfo

pub(crate) fn insert_cell_or_row<F, G, I>( &self, cx: &mut JSContext, index: i32, get_items: F, new_child: G, ) -> Fallible<DomRoot<HTMLElement>>
where F: Fn(&mut JSContext) -> DomRoot<HTMLCollection>, G: Fn(&mut JSContext) -> DomRoot<I>, I: DerivedFrom<Node> + DerivedFrom<HTMLElement> + DomObject,

Used by HTMLTableSectionElement::InsertRow and HTMLTableRowElement::InsertCell

pub(crate) fn delete_cell_or_row<F, G>( &self, cx: &mut JSContext, index: i32, get_items: F, is_delete_type: G, ) -> ErrorResult
where F: Fn(&mut JSContext) -> DomRoot<HTMLCollection>, G: Fn(&Element) -> bool,

Used by HTMLTableSectionElement::DeleteRow and HTMLTableRowElement::DeleteCell

pub(crate) fn get_cssom_stylesheet( &self, cx: &mut JSContext, ) -> Option<DomRoot<CSSStyleSheet>>

pub(crate) fn get_lang(&self) -> Option<String>

https://html.spec.whatwg.org/multipage/#language

pub(crate) fn assign_slottables_for_a_tree( &self, cx: &JSContext, force: ForceSlottableNodeReconciliation, )

https://dom.spec.whatwg.org/#assign-slotables-for-a-tree

pub(crate) fn assigned_slot(&self) -> Option<DomRoot<HTMLSlotElement>>

pub(crate) fn assigned_slot_unrooted<'a>( &self, no_gc: &'a NoGC, ) -> Option<UnrootedDom<'a, HTMLSlotElement>>

pub(crate) fn set_assigned_slot(&self, assigned_slot: Option<&HTMLSlotElement>)

pub(crate) fn manual_slot_assignment(&self) -> Option<DomRoot<HTMLSlotElement>>

pub(crate) fn set_manual_slot_assignment( &self, manually_assigned_slot: Option<&HTMLSlotElement>, )

pub(crate) fn parent_in_flat_tree<'b>( &self, no_gc: &'b NoGC, ) -> FlatTreeParent<'b>

Gets the parent of this node from the perspective of layout and style.

If the node and its parent have a flat tree relationship, this returns:

The node’s assigned slot.

The parent node’s shadow host if it’s a shadow root.

Or the node’s parent.

The parent might not have a flat tree relationship with the node if

It’s a light tree child of a shadow host.

It’s fallback content for an assigned slot.

pub(crate) fn inclusive_ancestors_in_flat_tree_unrooted<'a>( &self, no_gc: &'a NoGC, ) -> impl Iterator<Item = UnrootedDom<'a, Node>> + use<'a>

pub(crate) fn set_implemented_pseudo_element( &self, pseudo_element: PseudoElement, )

We are marking this as an implemented pseudo element.

pub(crate) fn implemented_pseudo_element(&self) -> Option<PseudoElement>

pub(crate) fn editing_host_of(&self) -> Option<DomRoot<Node>>

https://w3c.github.io/editing/docs/execCommand/#editing-host-of

pub(crate) fn is_editable_or_editing_host(&self) -> bool

pub(crate) fn is_editing_host(&self) -> bool

https://html.spec.whatwg.org/multipage/#editing-host

pub(crate) fn is_editable(&self) -> bool

https://w3c.github.io/editing/docs/execCommand/#editable

impl Node

pub(crate) fn reflect_node<N>( cx: &mut JSContext, node: Box<N>, document: &Document, ) -> DomRoot<N>
where N: DerivedFrom<Node> + DomObject + DomObjectWrap<DomTypeHolder>,

pub(crate) fn reflect_node_with_proto<N>( cx: &mut JSContext, node: Box<N>, document: &Document, proto: Option<HandleObject<'_>>, ) -> DomRoot<N>
where N: DerivedFrom<Node> + DomObject + DomObjectWrap<DomTypeHolder>,

pub(crate) fn reflect_weak_referenceable_node_with_proto<N>( cx: &mut JSContext, node: Rc<N>, document: &Document, proto: Option<HandleObject<'_>>, ) -> DomRoot<N>
where N: DerivedFrom<Node> + DomObject + WeakReferenceableDomObjectWrap<DomTypeHolder>,

pub(crate) fn new_inherited(doc: &Document) -> Node

pub(crate) fn new_document_node() -> Node

fn new_(flags: NodeFlags, doc: Option<&Document>) -> Node

pub(crate) fn adopt(cx: &mut JSContext, node: &Node, document: &Document)

https://dom.spec.whatwg.org/#concept-node-adopt

pub(crate) fn ensure_pre_insertion_validity( no_gc: &NoGC, node: &Node, parent: &Node, child: Option<&Node>, ) -> ErrorResult

https://dom.spec.whatwg.org/#concept-node-ensure-pre-insertion-validity

pub(crate) fn pre_insert( cx: &mut JSContext, node: &Node, parent: &Node, child: Option<&Node>, ) -> Fallible<DomRoot<Node>>

https://dom.spec.whatwg.org/#concept-node-pre-insert

pub(crate) fn insert( cx: &mut JSContext, node: &Node, parent: &Node, child: Option<&Node>, suppress_observers: SuppressObserver, )

https://dom.spec.whatwg.org/#concept-node-insert

pub(crate) fn maybe_dirty_visible_selection_for_newly_inserted_nodes( no_gc: &NoGC, parent: &Node, inserted_nodes: &[&Node], )

If insertion of any of the given nodes happened within an existing visible selection, mark the Document’s visible selection as dirty.

pub(crate) fn replace_all( cx: &mut JSContext, node: Option<&Node>, parent: &Node, )

https://dom.spec.whatwg.org/#concept-node-replace-all

pub(crate) fn string_replace_all( cx: &mut JSContext, string: DOMString, parent: &Node, )

https://dom.spec.whatwg.org/multipage/#string-replace-all

pub(super) fn pre_remove( cx: &mut JSContext, child: &Node, parent: &Node, ) -> Fallible<DomRoot<Node>>

https://dom.spec.whatwg.org/#concept-node-pre-remove

pub(super) fn remove( cx: &mut JSContext, node: &Node, parent: &Node, suppress_observers: SuppressObserver, )

https://dom.spec.whatwg.org/#concept-node-remove

pub(crate) fn clone( cx: &mut JSContext, node: &Node, maybe_doc: Option<&Document>, clone_children: CloneChildrenFlag, registry: Option<DomRoot<CustomElementRegistry>>, ) -> DomRoot<Node>

https://dom.spec.whatwg.org/#concept-node-clone

pub(crate) fn child_text_content(&self) -> DOMString

https://html.spec.whatwg.org/multipage/#child-text-content

pub(crate) fn descendant_text_content(&self) -> DOMString

https://html.spec.whatwg.org/multipage/#descendant-text-content

pub(crate) fn collect_text_contents<T: Iterator<Item = DomRoot<Node>>>( iterator: T, ) -> DOMString

pub(crate) fn set_text_content_for_element( &self, cx: &mut JSContext, value: Option<DOMString>, )

https://dom.spec.whatwg.org/#string-replace-all

pub(crate) fn namespace_to_string(namespace: Namespace) -> Option<DOMString>

pub(crate) fn locate_namespace( node: &Node, prefix: Option<DOMString>, ) -> Namespace

https://dom.spec.whatwg.org/#locate-a-namespace

pub(crate) unsafe fn from_untrusted_node_address( candidate: UntrustedNodeAddress, ) -> &'static Self

If the given untrusted node address represents a valid DOM node in the given runtime, returns it.

Safety

Callers should ensure they pass an UntrustedNodeAddress that points to a valid JSObject in memory that represents a Node.

pub(crate) fn html_serialize( &self, cx: &mut JSContext, traversal_scope: TraversalScope, serialize_shadow_roots: bool, shadow_roots: Vec<DomRoot<ShadowRoot>>, ) -> DOMString

pub(crate) fn xml_serialize( &self, traversal_scope: TraversalScope, ) -> Fallible<DOMString>

https://w3c.github.io/DOM-Parsing/#dfn-xml-serialization

pub(crate) fn fragment_serialization_algorithm( &self, cx: &mut JSContext, require_well_formed: bool, ) -> Fallible<DOMString>

https://html.spec.whatwg.org/multipage/#fragment-serializing-algorithm-steps

pub(crate) fn get_next_sibling_unrooted<'a>( &self, no_gc: &'a NoGC, ) -> Option<UnrootedDom<'a, Node>>

pub(crate) fn next_flat_tree_sibling_unrooted<'a>( &self, no_gc: &'a NoGC, ) -> Option<UnrootedDom<'a, Node>>

pub(crate) fn get_previous_sibling_unrooted<'a>( &self, no_gc: &'a NoGC, ) -> Option<UnrootedDom<'a, Node>>

pub(crate) fn get_first_child_unrooted<'a>( &self, no_gc: &'a NoGC, ) -> Option<UnrootedDom<'a, Node>>

pub(crate) fn first_flat_tree_child_unrooted<'a>( &self, no_gc: &'a NoGC, ) -> Option<UnrootedDom<'a, Node>>

fn get_last_child_unrooted<'b>( &self, no_gc: &'b NoGC, ) -> Option<UnrootedDom<'b, Node>>

pub(crate) fn get_parent_node_unrooted<'a>( &self, no_gc: &'a NoGC, ) -> Option<UnrootedDom<'a, Node>>

pub(crate) fn compare_dom_tree_position( &self, other: &Node, common_ancestor: &Node, shadow_including: ShadowIncluding, ) -> Ordering

Compares other with self in tree order.

impl Node

fn to_sibling_or_utf16_offset(&self, offset: Utf32CodeUnitsOrNodeOffset) -> u32

Get the Utf16CodeUnits offset for the given offset if self is a CharacterData or else return the offset in the child list.