Struct script::dom::node::LayoutDataRef

pub struct LayoutDataRef {
    pub data_cell: RefCell<Option<LayoutData>>,
}

Fields

data_cell

Methods

impl LayoutDataRef

fn new() -> LayoutDataRef

fn dispose(&self)

Sends layout data, if any, back to the layout task to be destroyed.

unsafe fn borrow_unchecked(&self) -> *const Option<LayoutData>

Borrows the layout data immutably, asserting that there are no mutators. Bad things will happen if you try to mutate the layout data while this is held. This is the only thread- safe layout data accessor.

fn borrow<'a>(&'a self) -> Ref<'a, Option<LayoutData>>

Borrows the layout data immutably. This function is not thread-safe.

fn borrow_mut<'a>(&'a self) -> RefMut<'a, Option<LayoutData>>

Borrows the layout data mutably. This function is not thread-safe.

FIXME(pcwalton): We should really put this behind a MutLayoutView phantom type, to prevent CSS selector matching from mutably accessing nodes it's not supposed to and racing on it. This has already resulted in one bug!