script::dom::node

Struct Node

#[repr(C)]
pub struct Node {

    eventtarget: EventTarget,
    parent_node: MutNullableDom<Node>,
    first_child: MutNullableDom<Node>,
    last_child: MutNullableDom<Node>,
    next_sibling: MutNullableDom<Node>,
    prev_sibling: MutNullableDom<Node>,
    owner_doc: MutNullableDom<Document>,
    rare_data: DomRefCell<Option<Box<NodeRareData>>>,
    children_count: Cell<u32>,
    flags: Cell<NodeFlags>,
    inclusive_descendants_version: Cell<u64>,
    style_data: DomRefCell<Option<Box<StyleData>>>,
    layout_data: DomRefCell<Option<Box<GenericLayoutData>>>,

}

An HTML node.

Fields
eventtarget: EventTarget

The JavaScript reflector for this node.

parent_node: MutNullableDom<Node>

The parent of this node.

first_child: MutNullableDom<Node>

The first child of this node.

last_child: MutNullableDom<Node>

The last child of this node.

next_sibling: MutNullableDom<Node>

The next sibling of this node.

prev_sibling: MutNullableDom<Node>

The previous sibling of this node.

owner_doc: MutNullableDom<Document>

The document that this node belongs to.

rare_data: DomRefCell<Option<Box<NodeRareData>>>

Rare node data.

children_count: Cell<u32>

The live count of children of this node.

flags: Cell<NodeFlags>

A bitfield of flags for node items.

inclusive_descendants_version: Cell<u64>

The maximum version of any inclusive descendant of this node.

style_data: DomRefCell<Option<Box<StyleData>>>

Style data for this node. This is accessed and mutated by style passes and is used to lay out this node and populate layout data.

layout_data: DomRefCell<Option<Box<GenericLayoutData>>>

Layout data for this node. This is populated during layout and can be used for incremental relayout and script queries.

Implementations

impl Node

fn as_text_input(&self) -> Option<DomRoot<Element>>

impl Node

fn add_child(&self, new_child: &Node, before: Option<&Node>, can_gc: CanGc)

Adds a new child to the end of this node’s list of children.

Fails unless new_child is disconnected from the tree.

pub(crate) fn unsafely_set_html( target: &Node, context_element: &Element, html: DOMString, can_gc: CanGc, )

Implements the “unsafely set HTML” algorithm as specified in: https://html.spec.whatwg.org/multipage/#concept-unsafely-set-html

pub(crate) fn clean_up_style_and_layout_data(&self)

pub(crate) fn complete_remove_subtree( root: &Node, context: &UnbindContext<'_>, can_gc: CanGc, )

Clean up flags and runs steps 11-14 of remove a node. https://dom.spec.whatwg.org/#concept-node-remove

fn remove_child(&self, child: &Node, cached_index: Option<u32>, can_gc: CanGc)

Removes the given child from this node’s list of children.

Fails unless child is a child of this node.

pub(crate) fn to_untrusted_node_address(&self) -> UntrustedNodeAddress

pub(crate) fn to_opaque(&self) -> OpaqueNode

pub(crate) fn as_custom_element(&self) -> Option<DomRoot<Element>>

pub(crate) fn fire_synthetic_pointer_event_not_trusted( &self, name: DOMString, can_gc: CanGc, )

https://html.spec.whatwg.org/multipage/#fire-a-synthetic-pointer-event

pub(crate) fn parent_directionality(&self) -> String

impl Node

fn rare_data(&self) -> Ref<'_, Option<Box<NodeRareData>>>

fn ensure_rare_data(&self) -> RefMut<'_, Box<NodeRareData>>

pub(crate) fn is_before(&self, other: &Node) -> bool

Returns true if this node is before other in the same connected DOM tree.

pub(crate) fn registered_mutation_observers_mut( &self, ) -> RefMut<'_, Vec<RegisteredObserver>>

Return all registered mutation observers for this node. Lazily initialize the raredata if it does not exist.

pub(crate) fn registered_mutation_observers( &self, ) -> Option<Ref<'_, Vec<RegisteredObserver>>>

pub(crate) fn add_mutation_observer(&self, observer: RegisteredObserver)

Add a new mutation observer for a given node.

pub(crate) fn remove_mutation_observer(&self, observer: &MutationObserver)

Removes the mutation observer for a given node.

pub(crate) fn dump(&self)

Dumps the subtree rooted at this node, for debugging.

pub(crate) fn dump_indent(&self, indent: u32)

Dumps the node tree, for debugging, with indentation.

pub(crate) fn debug_str(&self) -> String

Returns a string that describes this node.

pub(crate) fn is_in_a_document_tree(&self) -> bool

https://dom.spec.whatwg.org/#in-a-document-tree

pub(crate) fn is_in_a_shadow_tree(&self) -> bool

Return true iff node’s root is a shadow-root.

pub(crate) fn has_weird_parser_insertion_mode(&self) -> bool

pub(crate) fn set_weird_parser_insertion_mode(&self)

pub(crate) fn is_connected(&self) -> bool

https://dom.spec.whatwg.org/#connected

pub(crate) fn set_in_ua_widget(&self, in_ua_widget: bool)

pub(crate) fn is_in_ua_widget(&self) -> bool

pub(crate) fn type_id(&self) -> NodeTypeId

Returns the type ID of this node.

pub(crate) fn len(&self) -> u32

https://dom.spec.whatwg.org/#concept-node-length

pub(crate) fn is_empty(&self) -> bool

pub(crate) fn index(&self) -> u32

https://dom.spec.whatwg.org/#concept-tree-index

pub(crate) fn has_parent(&self) -> bool

Returns true if this node has a parent.

pub(crate) fn children_count(&self) -> u32

pub(crate) fn ranges(&self) -> RefMut<'_, WeakRangeVec>

pub(crate) fn ranges_is_empty(&self) -> bool

pub(crate) fn is_doctype(&self) -> bool

pub(crate) fn get_flag(&self, flag: NodeFlags) -> bool

pub(crate) fn set_flag(&self, flag: NodeFlags, value: bool)

pub(crate) fn note_dirty_descendants(&self)

pub(crate) fn has_dirty_descendants(&self) -> bool

pub(crate) fn rev_version(&self)

pub(crate) fn dirty(&self, damage: NodeDamage)

pub(crate) fn inclusive_descendants_version(&self) -> u64

The maximum version number of this node’s descendants, including itself

pub(crate) fn traverse_preorder( &self, shadow_including: ShadowIncluding, ) -> TreeIterator

Iterates over this node and all its descendants, in preorder.

pub(crate) fn inclusively_following_siblings( &self, ) -> impl Iterator<Item = DomRoot<Node>> + use<>

pub(crate) fn inclusively_preceding_siblings( &self, ) -> impl Iterator<Item = DomRoot<Node>> + use<>

pub(crate) fn common_ancestor( &self, other: &Node, shadow_including: ShadowIncluding, ) -> Option<DomRoot<Node>>

pub(crate) fn common_ancestor_in_flat_tree( &self, other: &Node, ) -> Option<DomRoot<Node>>

pub(crate) fn is_inclusive_ancestor_of(&self, child: &Node) -> bool

pub(crate) fn is_ancestor_of(&self, child: &Node) -> bool

pub(crate) fn is_shadow_including_inclusive_ancestor_of( &self, node: &Node, ) -> bool

pub(crate) fn following_siblings( &self, ) -> impl Iterator<Item = DomRoot<Node>> + use<>

pub(crate) fn preceding_siblings( &self, ) -> impl Iterator<Item = DomRoot<Node>> + use<>

pub(crate) fn following_nodes(&self, root: &Node) -> FollowingNodeIterator

pub(crate) fn preceding_nodes(&self, root: &Node) -> PrecedingNodeIterator

pub(crate) fn descending_last_children( &self, ) -> impl Iterator<Item = DomRoot<Node>> + use<>

Return an iterator that moves from self down the tree, choosing the last child at each step of the way.

pub(crate) fn is_parent_of(&self, child: &Node) -> bool

pub(crate) fn to_trusted_node_address(&self) -> TrustedNodeAddress

pub(crate) fn padding(&self) -> Option<PhysicalSides>

pub(crate) fn content_box(&self) -> Option<Rect<Au>>

pub(crate) fn border_box(&self) -> Option<Rect<Au>>

pub(crate) fn padding_box(&self) -> Option<Rect<Au>>

pub(crate) fn border_boxes(&self) -> Vec<Rect<Au>>

pub(crate) fn client_rect(&self) -> Rect<i32>

pub(crate) fn scroll_area(&self) -> Rect<i32>

https://drafts.csswg.org/cssom-view/#dom-element-scrollwidth https://drafts.csswg.org/cssom-view/#dom-element-scrollheight

pub(crate) fn before( &self, nodes: Vec<NodeOrString<DomTypeHolder>>, can_gc: CanGc, ) -> ErrorResult

https://dom.spec.whatwg.org/#dom-childnode-before

pub(crate) fn after( &self, nodes: Vec<NodeOrString<DomTypeHolder>>, can_gc: CanGc, ) -> ErrorResult

https://dom.spec.whatwg.org/#dom-childnode-after

pub(crate) fn replace_with( &self, nodes: Vec<NodeOrString<DomTypeHolder>>, can_gc: CanGc, ) -> ErrorResult

https://dom.spec.whatwg.org/#dom-childnode-replacewith

pub(crate) fn prepend( &self, nodes: Vec<NodeOrString<DomTypeHolder>>, can_gc: CanGc, ) -> ErrorResult

https://dom.spec.whatwg.org/#dom-parentnode-prepend

pub(crate) fn append( &self, nodes: Vec<NodeOrString<DomTypeHolder>>, can_gc: CanGc, ) -> ErrorResult

https://dom.spec.whatwg.org/#dom-parentnode-append

pub(crate) fn replace_children( &self, nodes: Vec<NodeOrString<DomTypeHolder>>, can_gc: CanGc, ) -> ErrorResult

https://dom.spec.whatwg.org/#dom-parentnode-replacechildren

pub(crate) fn query_selector( &self, selectors: DOMString, ) -> Fallible<Option<DomRoot<Element>>>

https://dom.spec.whatwg.org/#dom-parentnode-queryselector

pub(crate) fn query_selector_iter( &self, selectors: DOMString, ) -> Fallible<QuerySelectorIterator>

https://dom.spec.whatwg.org/#scope-match-a-selectors-string Get an iterator over all nodes which match a set of selectors Be careful not to do anything which may manipulate the DOM tree whilst iterating, otherwise the iterator may be invalidated.

pub(crate) fn query_selector_all( &self, selectors: DOMString, ) -> Fallible<DomRoot<NodeList>>

https://dom.spec.whatwg.org/#dom-parentnode-queryselectorall

pub(crate) fn ancestors(&self) -> impl Iterator<Item = DomRoot<Node>> + use<>

pub(crate) fn inclusive_ancestors( &self, shadow_including: ShadowIncluding, ) -> impl Iterator<Item = DomRoot<Node>> + use<>

https://dom.spec.whatwg.org/#concept-shadow-including-inclusive-ancestor

pub(crate) fn owner_doc(&self) -> DomRoot<Document>

pub(crate) fn set_owner_doc(&self, document: &Document)

pub(crate) fn containing_shadow_root(&self) -> Option<DomRoot<ShadowRoot>>

pub(crate) fn set_containing_shadow_root( &self, shadow_root: Option<&ShadowRoot>, )

pub(crate) fn is_in_html_doc(&self) -> bool

pub(crate) fn is_connected_with_browsing_context(&self) -> bool

pub(crate) fn children(&self) -> impl Iterator<Item = DomRoot<Node>> + use<>

pub(crate) fn rev_children(&self) -> impl Iterator<Item = DomRoot<Node>> + use<>

pub(crate) fn child_elements( &self, ) -> impl Iterator<Item = DomRoot<Element>> + use<>

pub(crate) fn remove_self(&self, can_gc: CanGc)

pub(crate) fn unique_id(&self, pipeline: PipelineId) -> String

pub(crate) fn summarize(&self, can_gc: CanGc) -> NodeInfo

pub(crate) fn insert_cell_or_row<F, G, I>( &self, index: i32, get_items: F, new_child: G, can_gc: CanGc, ) -> Fallible<DomRoot<HTMLElement>>
where F: Fn() -> DomRoot<HTMLCollection>, G: Fn() -> DomRoot<I>, I: DerivedFrom<Node> + DerivedFrom<HTMLElement> + DomObject,

Used by HTMLTableSectionElement::InsertRow and HTMLTableRowElement::InsertCell

pub(crate) fn delete_cell_or_row<F, G>( &self, index: i32, get_items: F, is_delete_type: G, can_gc: CanGc, ) -> ErrorResult
where F: Fn() -> DomRoot<HTMLCollection>, G: Fn(&Element) -> bool,

Used by HTMLTableSectionElement::DeleteRow and HTMLTableRowElement::DeleteCell

pub(crate) fn get_stylesheet(&self) -> Option<Arc<Stylesheet>>

pub(crate) fn get_cssom_stylesheet(&self) -> Option<DomRoot<CSSStyleSheet>>

pub(crate) fn is_styled(&self) -> bool

pub(crate) fn is_display_none(&self) -> bool

pub(crate) fn style(&self) -> Option<Arc<ComputedValues>>

pub(crate) fn get_lang(&self) -> Option<String>

https://html.spec.whatwg.org/multipage/#language

pub(crate) fn assign_slottables_for_a_tree(&self)

https://dom.spec.whatwg.org/#assign-slotables-for-a-tree

pub(crate) fn assigned_slot(&self) -> Option<DomRoot<HTMLSlotElement>>

pub(crate) fn set_assigned_slot(&self, assigned_slot: Option<&HTMLSlotElement>)

pub(crate) fn manual_slot_assignment(&self) -> Option<DomRoot<HTMLSlotElement>>

pub(crate) fn set_manual_slot_assignment( &self, manually_assigned_slot: Option<&HTMLSlotElement>, )

pub(crate) fn parent_in_flat_tree(&self) -> Option<DomRoot<Node>>

Gets the parent of this node from the perspective of layout and style.

The returned node is the node’s assigned slot, if any, or the shadow host if it’s a shadow root. Otherwise, it is the node’s parent.

pub(crate) fn inclusive_ancestors_in_flat_tree( &self, ) -> impl Iterator<Item = DomRoot<Node>> + use<>

pub(crate) fn set_implemented_pseudo_element( &self, pseudo_element: PseudoElement, )

We are marking this as an implemented pseudo element.

pub(crate) fn implemented_pseudo_element(&self) -> Option<PseudoElement>

impl Node

pub(crate) fn reflect_node<N>( node: Box<N>, document: &Document, can_gc: CanGc, ) -> DomRoot<N>
where N: DerivedFrom<Node> + DomObject + DomObjectWrap<DomTypeHolder>,

pub(crate) fn reflect_node_with_proto<N>( node: Box<N>, document: &Document, proto: Option<HandleObject<'_>>, can_gc: CanGc, ) -> DomRoot<N>
where N: DerivedFrom<Node> + DomObject + DomObjectWrap<DomTypeHolder>,

pub(crate) fn new_inherited(doc: &Document) -> Node

pub(crate) fn new_document_node() -> Node

fn new_(flags: NodeFlags, doc: Option<&Document>) -> Node

pub(crate) fn adopt(node: &Node, document: &Document, can_gc: CanGc)

https://dom.spec.whatwg.org/#concept-node-adopt

pub(crate) fn ensure_pre_insertion_validity( node: &Node, parent: &Node, child: Option<&Node>, ) -> ErrorResult

https://dom.spec.whatwg.org/#concept-node-ensure-pre-insertion-validity

pub(crate) fn pre_insert( node: &Node, parent: &Node, child: Option<&Node>, can_gc: CanGc, ) -> Fallible<DomRoot<Node>>

https://dom.spec.whatwg.org/#concept-node-pre-insert

fn insert( node: &Node, parent: &Node, child: Option<&Node>, suppress_observers: SuppressObserver, can_gc: CanGc, )

https://dom.spec.whatwg.org/#concept-node-insert

pub(crate) fn replace_all(node: Option<&Node>, parent: &Node, can_gc: CanGc)

https://dom.spec.whatwg.org/#concept-node-replace-all

pub(crate) fn string_replace_all( string: DOMString, parent: &Node, can_gc: CanGc, )

https://dom.spec.whatwg.org/multipage/#string-replace-all

fn pre_remove( child: &Node, parent: &Node, can_gc: CanGc, ) -> Fallible<DomRoot<Node>>

https://dom.spec.whatwg.org/#concept-node-pre-remove

fn remove( node: &Node, parent: &Node, suppress_observers: SuppressObserver, can_gc: CanGc, )

https://dom.spec.whatwg.org/#concept-node-remove

fn compute_attribute_value_with_style_fast_path( attr: &Dom<Attr>, elem: &Element, ) -> AttrValue

Ensure that for styles, we clone the already-parsed property declaration block. This does two things:

it uses the same fast-path as CSSStyleDeclaration

it also avoids the CSP checks when cloning (it shouldn’t run any when cloning existing valid attributes)

pub(crate) fn clone( node: &Node, maybe_doc: Option<&Document>, clone_children: CloneChildrenFlag, registry: Option<DomRoot<CustomElementRegistry>>, can_gc: CanGc, ) -> DomRoot<Node>

https://dom.spec.whatwg.org/#concept-node-clone

pub(crate) fn child_text_content(&self) -> DOMString

https://html.spec.whatwg.org/multipage/#child-text-content

pub(crate) fn descendant_text_content(&self) -> DOMString

https://html.spec.whatwg.org/multipage/#descendant-text-content

pub(crate) fn collect_text_contents<T: Iterator<Item = DomRoot<Node>>>( iterator: T, ) -> DOMString

pub(crate) fn set_text_content_for_element( &self, value: Option<DOMString>, can_gc: CanGc, )

https://dom.spec.whatwg.org/#string-replace-all

pub(crate) fn namespace_to_string(namespace: Namespace) -> Option<DOMString>

pub(crate) fn locate_namespace( node: &Node, prefix: Option<DOMString>, ) -> Namespace

https://dom.spec.whatwg.org/#locate-a-namespace

pub(crate) unsafe fn from_untrusted_node_address( candidate: UntrustedNodeAddress, ) -> &'static Self

If the given untrusted node address represents a valid DOM node in the given runtime, returns it.

Safety

Callers should ensure they pass an UntrustedNodeAddress that points to a valid JSObject in memory that represents a Node.

pub(crate) fn html_serialize( &self, traversal_scope: TraversalScope, serialize_shadow_roots: bool, shadow_roots: Vec<DomRoot<ShadowRoot>>, can_gc: CanGc, ) -> DOMString

pub(crate) fn xml_serialize( &self, traversal_scope: TraversalScope, ) -> Fallible<DOMString>

https://w3c.github.io/DOM-Parsing/#dfn-xml-serialization

pub(crate) fn fragment_serialization_algorithm( &self, require_well_formed: bool, can_gc: CanGc, ) -> Fallible<DOMString>

https://html.spec.whatwg.org/multipage/#fragment-serializing-algorithm-steps

pub(crate) fn establishes_scrolling_box(&self) -> bool

Return true if this node establishes a “scrolling box” for the purposes of scrollIntoView.