Struct script::dom::node::UnbindContext

pub struct UnbindContext<'a> {
    index: Cell<Option<u32>>,
    pub parent: &'a Node,
    prev_sibling: Option<&'a Node>,
    pub tree_in_doc: bool,
}

The context of the unbinding from a tree of a node when one of its inclusive ancestors is removed.

Fields

index

The index of the inclusive ancestor that was removed.

parent

The parent of the inclusive ancestor that was removed.

prev_sibling

The previous sibling of the inclusive ancestor that was removed.

tree_in_doc

Whether the tree is in a document.

Methods

impl<'a> UnbindContext<'a>

fn new(parent: &'a Node, prev_sibling: Option<&'a Node>, cached_index: Option<u32>) -> Self

Create a new UnbindContext value.

fn index(&self) -> u32

The index of the inclusive ancestor that was removed from the tree.