Trait script::dom::node::NodeHelpers

pub trait NodeHelpers<'a> {
    fn ancestors(self) -> AncestorIterator<'a>;
    fn inclusive_ancestors(self) -> AncestorIterator<'a>;
    fn children(self) -> NodeChildrenIterator<'a>;
    fn rev_children(self) -> ReverseChildrenIterator;
    fn child_elements(self) -> ChildElementIterator<'a>;
    fn following_siblings(self) -> NodeChildrenIterator<'a>;
    fn is_in_doc(self) -> bool;
    fn is_inclusive_ancestor_of(self, parent: JSRef<Node>) -> bool;
    fn is_parent_of(self, child: JSRef<Node>) -> bool;
    fn type_id(self) -> NodeTypeId;
    fn parent_node(self) -> Option<Temporary<Node>>;
    fn first_child(self) -> Option<Temporary<Node>>;
    fn last_child(self) -> Option<Temporary<Node>>;
    fn prev_sibling(self) -> Option<Temporary<Node>>;
    fn next_sibling(self) -> Option<Temporary<Node>>;
    fn owner_doc(self) -> Temporary<Document>;
    fn set_owner_doc(self, document: JSRef<Document>);
    fn is_in_html_doc(self) -> bool;
    fn is_element(self) -> bool;
    fn is_document(self) -> bool;
    fn is_doctype(self) -> bool;
    fn is_text(self) -> bool;
    fn is_anchor_element(self) -> bool;
    fn get_flag(self, flag: NodeFlags) -> bool;
    fn set_flag(self, flag: NodeFlags, value: bool);
    fn get_hover_state(self) -> bool;
    fn set_hover_state(self, state: bool);
    fn get_disabled_state(self) -> bool;
    fn set_disabled_state(self, state: bool);
    fn get_enabled_state(self) -> bool;
    fn set_enabled_state(self, state: bool);
    fn get_has_changed(self) -> bool;
    fn set_has_changed(self, state: bool);
    fn get_is_dirty(self) -> bool;
    fn set_is_dirty(self, state: bool);
    fn get_has_dirty_siblings(self) -> bool;
    fn set_has_dirty_siblings(self, state: bool);
    fn get_has_dirty_descendants(self) -> bool;
    fn set_has_dirty_descendants(self, state: bool);
    fn dirty(self, damage: NodeDamage);
    fn force_dirty_ancestors(self, damage: NodeDamage);
    fn dirty_impl(self, damage: NodeDamage, force_ancestors: bool);
    fn dump(self);
    fn dump_indent(self, indent: usize);
    fn debug_str(self) -> String;
    fn traverse_preorder(self) -> TreeIterator<'a>;
    fn inclusively_following_siblings(self) -> NodeChildrenIterator<'a>;
    fn to_trusted_node_address(self) -> TrustedNodeAddress;
    fn get_bounding_content_box(self) -> Rect<Au>;
    fn get_content_boxes(self) -> Vec<Rect<Au>>;
    fn query_selector(self, selectors: DOMString) -> Fallible<Option<Temporary<Element>>>;
    unsafe fn query_selector_iter(self, selectors: DOMString) -> Fallible<QuerySelectorIterator<'a>>;
    fn query_selector_all(self, selectors: DOMString) -> Fallible<Temporary<NodeList>>;
    fn remove_self(self);
    fn get_unique_id(self) -> String;
    fn summarize(self) -> NodeInfo;
    fn teardown(self);
}

Required Methods

fn ancestors(self) -> AncestorIterator<'a>

fn inclusive_ancestors(self) -> AncestorIterator<'a>

fn children(self) -> NodeChildrenIterator<'a>

fn rev_children(self) -> ReverseChildrenIterator

fn child_elements(self) -> ChildElementIterator<'a>

fn following_siblings(self) -> NodeChildrenIterator<'a>

fn is_in_doc(self) -> bool

fn is_inclusive_ancestor_of(self, parent: JSRef<Node>) -> bool

fn is_parent_of(self, child: JSRef<Node>) -> bool

fn type_id(self) -> NodeTypeId

fn parent_node(self) -> Option<Temporary<Node>>

fn first_child(self) -> Option<Temporary<Node>>

fn last_child(self) -> Option<Temporary<Node>>

fn prev_sibling(self) -> Option<Temporary<Node>>

fn next_sibling(self) -> Option<Temporary<Node>>

fn owner_doc(self) -> Temporary<Document>

fn set_owner_doc(self, document: JSRef<Document>)

fn is_in_html_doc(self) -> bool

fn is_element(self) -> bool

fn is_document(self) -> bool

fn is_doctype(self) -> bool

fn is_text(self) -> bool

fn is_anchor_element(self) -> bool

fn get_flag(self, flag: NodeFlags) -> bool

fn set_flag(self, flag: NodeFlags, value: bool)

fn get_hover_state(self) -> bool

fn set_hover_state(self, state: bool)

fn get_disabled_state(self) -> bool

fn set_disabled_state(self, state: bool)

fn get_enabled_state(self) -> bool

fn set_enabled_state(self, state: bool)

fn get_has_changed(self) -> bool

fn set_has_changed(self, state: bool)

fn get_is_dirty(self) -> bool

fn set_is_dirty(self, state: bool)

fn get_has_dirty_siblings(self) -> bool

fn set_has_dirty_siblings(self, state: bool)

fn get_has_dirty_descendants(self) -> bool

fn set_has_dirty_descendants(self, state: bool)

fn dirty(self, damage: NodeDamage)

Marks the given node as IS_DIRTY, its siblings as HAS_DIRTY_SIBLINGS (to deal with sibling selectors), its ancestors as HAS_DIRTY_DESCENDANTS, and its descendants as IS_DIRTY. If anything more than the node's style was damaged, this method also sets the HAS_CHANGED flag.

fn force_dirty_ancestors(self, damage: NodeDamage)

Similar to dirty, but will always walk the ancestors to mark them dirty, too. This is useful when a node is reparented. The node will frequently already be marked as changed to skip double-dirties, but the ancestors still need to be marked as HAS_DIRTY_DESCENDANTS.

See #4170

fn dirty_impl(self, damage: NodeDamage, force_ancestors: bool)

fn dump(self)

fn dump_indent(self, indent: usize)

fn debug_str(self) -> String

fn traverse_preorder(self) -> TreeIterator<'a>

fn inclusively_following_siblings(self) -> NodeChildrenIterator<'a>

fn to_trusted_node_address(self) -> TrustedNodeAddress

fn get_bounding_content_box(self) -> Rect<Au>

fn get_content_boxes(self) -> Vec<Rect<Au>>

fn query_selector(self, selectors: DOMString) -> Fallible<Option<Temporary<Element>>>

unsafe fn query_selector_iter(self, selectors: DOMString) -> Fallible<QuerySelectorIterator<'a>>

fn query_selector_all(self, selectors: DOMString) -> Fallible<Temporary<NodeList>>

fn remove_self(self)

fn get_unique_id(self) -> String

fn summarize(self) -> NodeInfo

fn teardown(self)