script::dom::node::virtualmethods

Trait VirtualMethods

pub(crate) trait VirtualMethods {

    // Required method
    fn super_type(&self) -> Option<&dyn VirtualMethods>;

    // Provided methods
    fn attribute_mutated(
        &self,
        cx: &mut JSContext,
        attr: AttrRef<'_>,
        mutation: AttributeMutation<'_>,
    ) { ... }
    fn attribute_affects_presentational_hints(&self, attr: AttrRef<'_>) -> bool { ... }
    fn parse_plain_attribute(
        &self,
        name: &LocalName,
        value: DOMString,
    ) -> AttrValue { ... }
    fn post_connection_steps(&self, cx: &mut JSContext) { ... }
    fn moving_steps(&self, cx: &mut JSContext, context: &MoveContext<'_>) { ... }
    fn bind_to_tree(&self, cx: &mut JSContext, context: &BindContext<'_>) { ... }
    fn unbind_from_tree(&self, cx: &mut JSContext, context: &UnbindContext<'_>) { ... }
    fn children_changed(
        &self,
        cx: &mut JSContext,
        mutation: &ChildrenMutation<'_>,
    ) { ... }
    fn handle_event(&self, cx: &mut JSContext, event: &Event) { ... }
    fn handle_mousedown_event(
        &self,
        cx: &mut JSContext,
        mouse_event: &MouseEvent,
        hit_test_result: &HitTestResult,
    ) { ... }
    fn is_valid_command_steps(&self, command: CommandState) -> bool { ... }
    fn command_steps(
        &self,
        cx: &mut JSContext,
        button: DomRoot<HTMLButtonElement>,
        command: CommandState,
    ) -> bool { ... }
    fn adopting_steps(&self, cx: &mut JSContext, old_doc: &Document) { ... }
    fn cloning_steps(
        &self,
        cx: &mut JSContext,
        copy: &Node,
        maybe_doc: Option<&Document>,
        clone_children: CloneChildrenFlag,
    ) { ... }
    fn pop(&self, cx: &mut JSContext) { ... }

}

Trait to allow DOM nodes to opt-in to overriding (or adding to) common behaviours. Replicates the effect of C++ virtual methods.

Required Methods

fn super_type(&self) -> Option<&dyn VirtualMethods>

Returns self as the superclass of the implementation for this trait, if any.

Provided Methods

fn attribute_mutated( &self, cx: &mut JSContext, attr: AttrRef<'_>, mutation: AttributeMutation<'_>, )

Called when attributes of a node are mutated. https://dom.spec.whatwg.org/#attribute-is-set https://dom.spec.whatwg.org/#attribute-is-removed

fn attribute_affects_presentational_hints(&self, attr: AttrRef<'_>) -> bool

Returns true if given attribute attr affects style of the given element.

fn parse_plain_attribute(&self, name: &LocalName, value: DOMString) -> AttrValue

Returns the right AttrValue variant for the attribute with name name on this element.

fn post_connection_steps(&self, cx: &mut JSContext)

Invoked during a DOM tree mutation after a node becomes connected, once all related DOM tree mutations have been applied. https://dom.spec.whatwg.org/#concept-node-post-connection-ext

fn moving_steps(&self, cx: &mut JSContext, context: &MoveContext<'_>)

https://dom.spec.whatwg.org/#concept-node-move-ext

fn bind_to_tree(&self, cx: &mut JSContext, context: &BindContext<'_>)

Called when a Node is appended to a tree.

fn unbind_from_tree(&self, cx: &mut JSContext, context: &UnbindContext<'_>)

Called when a Node is removed from a tree. Implements removing steps: https://dom.spec.whatwg.org/#concept-node-remove-ext

fn children_changed(&self, cx: &mut JSContext, mutation: &ChildrenMutation<'_>)

Called on the parent when its children are changed.

fn handle_event(&self, cx: &mut JSContext, event: &Event)

Called during event dispatch after the bubbling phase completes.

fn handle_mousedown_event( &self, cx: &mut JSContext, mouse_event: &MouseEvent, hit_test_result: &HitTestResult, )

Called after mousedown/pointerdown event dispatch completes. This needs to be separate from handle_event because it’s only invoked if preventDefault was called on neither the mousedown nor the pointerdown event handler.

fn is_valid_command_steps(&self, command: CommandState) -> bool

https://html.spec.whatwg.org/multipage/#is-valid-command-steps

fn command_steps( &self, cx: &mut JSContext, button: DomRoot<HTMLButtonElement>, command: CommandState, ) -> bool

https://html.spec.whatwg.org/multipage/#command-steps

fn adopting_steps(&self, cx: &mut JSContext, old_doc: &Document)

https://dom.spec.whatwg.org/#concept-node-adopt-ext

fn cloning_steps( &self, cx: &mut JSContext, copy: &Node, maybe_doc: Option<&Document>, clone_children: CloneChildrenFlag, )

https://dom.spec.whatwg.org/#concept-node-clone-ext

fn pop(&self, cx: &mut JSContext)

Called on an element when it is popped off the stack of open elements of a parser.