script::dom::performance::performance

Struct Performance

#[repr(C)]
pub(crate) struct Performance {
    eventtarget: EventTarget,
    buffer: DomRefCell<PerformanceEntryList>,
    observers: DomRefCell<Vec<PerformanceObserver>>,
    pending_notification_observers_task: Cell<bool>,
    time_origin: CrossProcessInstant,
    resource_timing_buffer_size_limit: Cell<usize>,
    resource_timing_buffer_current_size: Cell<usize>,
    resource_timing_buffer_pending_full_event: Cell<bool>,
    resource_timing_secondary_entries: DomRefCell<VecDeque<DomRoot<PerformanceEntry>>>,
}

Fields
eventtarget: EventTargetbuffer: DomRefCell<PerformanceEntryList>observers: DomRefCell<Vec<PerformanceObserver>>pending_notification_observers_task: Cell<bool>time_origin: CrossProcessInstant

The timeOrigin as described in https://html.spec.whatwg.org/multipage/#concept-settings-object-time-origin.

resource_timing_buffer_size_limit: Cell<usize>

https://w3c.github.io/performance-timeline/#dfn-maxbuffersize The max-size of the buffer, set to 0 once the pipeline exits. TODO: have one max-size per entry type.

resource_timing_buffer_current_size: Cell<usize>resource_timing_buffer_pending_full_event: Cell<bool>resource_timing_secondary_entries: DomRefCell<VecDeque<DomRoot<PerformanceEntry>>>
Implementations

impl Performance

fn __assert_parent_type(&self)

impl Performance

fn new_inherited(time_origin: CrossProcessInstant) -> Performance

pub(crate) fn new( global: &GlobalScope, navigation_start: CrossProcessInstant, can_gc: CanGc, ) -> DomRoot<Performance>

pub(crate) fn to_dom_high_res_time_stamp( &self, instant: CrossProcessInstant, ) -> DOMHighResTimeStamp

pub(crate) fn maybe_to_dom_high_res_time_stamp( &self, instant: Option<CrossProcessInstant>, ) -> DOMHighResTimeStamp

pub(crate) fn clear_and_disable_performance_entry_buffer(&self)

Clear all buffered performance entries, and disable the buffer. Called as part of the window’s “clear_js_runtime” workflow, performed when exiting a pipeline.

pub(crate) fn add_multiple_type_observer( &self, observer: &DOMPerformanceObserver, entry_types: Vec<EntryType>, )

pub(crate) fn add_single_type_observer( &self, observer: &DOMPerformanceObserver, entry_type: EntryType, buffered: bool, )

pub(crate) fn remove_observer(&self, observer: &DOMPerformanceObserver)

Remove a PerformanceObserver from the list of observers.

pub(crate) fn queue_entry(&self, entry: &PerformanceEntry) -> Option<usize>

Queue a notification for each performance observer interested in this type of performance entry and queue a low priority task to notify the observers if no other notification task is already queued.

Algorithm spec: https://w3c.github.io/performance-timeline/#queue-a-performanceentry Also this algorithm has been extented according to : https://w3c.github.io/resource-timing/#sec-extensions-performance-interface

pub(crate) fn notify_observers(&self)

Observers notifications task.

Algorithm spec (step 7): https://w3c.github.io/performance-timeline/#queue-a-performanceentry

fn can_add_resource_timing_entry(&self) -> bool

fn copy_secondary_resource_timing_buffer(&self)

https://w3c.github.io/resource-timing/#dfn-copy-secondary-buffer

fn fire_buffer_full_event(&self, can_gc: CanGc)

https://w3c.github.io/resource-timing/#sec-extensions-performance-interface

fn should_queue_resource_entry(&self, entry: &PerformanceEntry) -> bool

https://w3c.github.io/resource-timing/#dfn-add-a-performanceresourcetiming-entry

pub(crate) fn update_entry(&self, index: usize, entry: &PerformanceEntry)