Struct script::dom::performance::Performance

#[repr(C)]
pub struct Performance {
    eventtarget: EventTarget,
    buffer: DomRefCell<PerformanceEntryList>,
    observers: DomRefCell<Vec<PerformanceObserver>>,
    pending_notification_observers_task: Cell<bool>,
    navigation_start_precise: u64,
    resource_timing_buffer_size_limit: Cell<usize>,
    resource_timing_buffer_current_size: Cell<usize>,
    resource_timing_buffer_pending_full_event: Cell<bool>,
    resource_timing_secondary_entries: DomRefCell<VecDeque<Root<Dom<PerformanceEntry>>>>,
}

Fields
eventtarget: EventTargetbuffer: DomRefCell<PerformanceEntryList>observers: DomRefCell<Vec<PerformanceObserver>>pending_notification_observers_task: Cell<bool>navigation_start_precise: u64resource_timing_buffer_size_limit: Cell<usize>

https://w3c.github.io/performance-timeline/#dfn-maxbuffersize The max-size of the buffer, set to 0 once the pipeline exits. TODO: have one max-size per entry type.

resource_timing_buffer_current_size: Cell<usize>resource_timing_buffer_pending_full_event: Cell<bool>resource_timing_secondary_entries: DomRefCell<VecDeque<Root<Dom<PerformanceEntry>>>>
Methods

impl Performance

fn __assert_parent_type(&self)

impl Performance

fn new_inherited(navigation_start_precise: u64) -> Performance

pub fn new(
global: &GlobalScope,
navigation_start_precise: u64
) -> Root<Dom<Performance>>

pub fn clear_and_disable_performance_entry_buffer(&self)

Clear all buffered performance entries, and disable the buffer. Called as part of the window's "clear_js_runtime" workflow, performed when exiting a pipeline.

pub fn add_multiple_type_observer(
&self,
observer: &DOMPerformanceObserver,
entry_types: Vec<DOMString>
)

Add a PerformanceObserver to the list of observers with a set of observed entry types.

pub fn add_single_type_observer(
&self,
observer: &DOMPerformanceObserver,
entry_type: &DOMString,
buffered: bool
)

pub fn remove_observer(&self, observer: &DOMPerformanceObserver)

Remove a PerformanceObserver from the list of observers.

pub fn queue_entry(&self, entry: &PerformanceEntry) -> Option<usize>

Queue a notification for each performance observer interested in this type of performance entry and queue a low priority task to notify the observers if no other notification task is already queued.

Algorithm spec: https://w3c.github.io/performance-timeline/#queue-a-performanceentry Also this algorithm has been extented according to : https://w3c.github.io/resource-timing/#sec-extensions-performance-interface

pub fn notify_observers(&self)

Observers notifications task.

Algorithm spec (step 7): https://w3c.github.io/performance-timeline/#queue-a-performanceentry

fn now(&self) -> f64

fn can_add_resource_timing_entry(&self) -> bool

fn copy_secondary_resource_timing_buffer(&self)

fn fire_buffer_full_event(&self)

fn should_queue_resource_entry(&self, entry: &PerformanceEntry) -> bool

add a PerformanceResourceTiming entry paragraph of https://w3c.github.io/resource-timing/#sec-extensions-performance-interface

pub fn update_entry(&self, index: usize, entry: &PerformanceEntry)