script::dom::promise::promisenativehandler

Trait Callback

pub(crate) trait Callback: JSTraceable + MallocSizeOf {
    // Required method
    fn callback(&self, cx: &mut CurrentRealm<'_>, v: HandleValue<'_>);
}

Types that implement the Callback trait follow the same rooting requirements as types that use the #[dom_struct] attribute. Prefer storing Dom<T> members inside them instead of DomRoot<T> to minimize redundant work by the garbage collector.

Required Methods

fn callback(&self, cx: &mut CurrentRealm<'_>, v: HandleValue<'_>)