Struct script::dom::promise::Promise

pub struct Promise {
    reflector: Reflector,
    permanent_js_root: MutHeapJSVal,
}

Fields
reflector: Reflector permanent_js_root: MutHeapJSVal

Since Promise values are natively reference counted without the knowledge of the SpiderMonkey GC, an explicit root for the reflector is stored while any native instance exists. This ensures that the reflector will never be GCed while native code could still interact with its native representation.

Methods

impl Promise

fn new(global: &GlobalScope) -> Rc<Promise>

fn duplicate(&self) -> Rc<Promise>

unsafe fn new_with_js_promise(obj: HandleObject,
cx: *mut JSContext)
-> Rc<Promise>

unsafe fn create_js_promise(cx: *mut JSContext,
proto: HandleObject,
obj: MutableHandleObject)

fn Resolve(global: &GlobalScope,
cx: *mut JSContext,
value: HandleValue)
-> Fallible<Rc<Promise>>

fn Reject(global: &GlobalScope,
cx: *mut JSContext,
value: HandleValue)
-> Fallible<Rc<Promise>>

fn resolve_native<T>(&self, cx: *mut JSContext, val: &T) where T: ToJSValConvertible

fn resolve(&self, cx: *mut JSContext, value: HandleValue)

fn reject_native<T>(&self, cx: *mut JSContext, val: &T) where T: ToJSValConvertible

fn reject_error(&self, cx: *mut JSContext, error: Error)

fn reject(&self, cx: *mut JSContext, value: HandleValue)

fn then(&self,
cx: *mut JSContext,
_callee: HandleObject,
cb_resolve: AnyCallback,
cb_reject: AnyCallback,
result: MutableHandleObject)

fn promise_obj(&self) -> HandleObject

fn append_native_handler(&self, handler: &PromiseNativeHandler)