script::dom::range::range

Struct Range

#[repr(C)]
pub(crate) struct Range {
    abstract_range: AbstractRange,
    associated_selections: DomRefCell<Vec<Dom<Selection>>>,
}

Fields
abstract_range: AbstractRangeassociated_selections: DomRefCell<Vec<Dom<Selection>>>
Implementations

impl Range

fn __assert_parent_type(&self)

impl Range

fn is_effectively_contained_node(&self, no_gc: &NoGC, node: &Node) -> bool

https://w3c.github.io/editing/docs/execCommand/#effectively-contained

fn ancestor_for_effectively_contained(&self) -> DomRoot<Node>

The definition of “effectively contained” contains the recursion of ancestors of a single fully selected text node. That is to say, that if the selection is a fully selected text node
[foobar]
, then the div would also be considered effectively contained. As such, we can’t use the common ancestor container, since that would be the text node only.

Instead, we traverse all the way up to the editing host, which we know is sufficient to know to include all contained nodes. That way, we also would traverse ancestors such as the parent div.

pub(crate) fn first_formattable_contained_node( &self, no_gc: &NoGC, ) -> Option<DomRoot<Node>>

pub(crate) fn for_each_effectively_contained_child<Callback: FnMut(&mut JSContext, &Node)>( &self, cx: &mut JSContext, callback: Callback, )

pub(crate) fn contained_nodes<'a>( &self, no_gc: &'a NoGC, ) -> impl Iterator<Item = UnrootedDom<'a, Node>>

pub(crate) fn block_extend( &self, cx: &mut JSContext, document: &Document, ) -> DomRoot<Range>

https://w3c.github.io/editing/docs/execCommand/#block-extend

pub(crate) fn record_current_states_and_values( &self, cx: &mut JSContext, ) -> Vec<RecordedStateOfCommand>

https://w3c.github.io/editing/docs/execCommand/#record-current-states-and-values

pub(crate) fn restore_states_and_values( &self, cx: &mut JSContext, selection: &Selection, context_object: &Document, overrides: Vec<RecordedStateOfCommand>, )

https://w3c.github.io/editing/docs/execCommand/#restore-states-and-values

impl Range

fn new_inherited( start_container: &Node, start_offset: u32, end_container: &Node, end_offset: u32, ) -> Range

pub(crate) fn new_with_doc( cx: &mut JSContext, document: &Document, proto: Option<HandleObject<'_>>, ) -> DomRoot<Range>

pub(crate) fn new( cx: &mut JSContext, document: &Document, start_container: &Node, start_offset: u32, end_container: &Node, end_offset: u32, ) -> DomRoot<Range>

fn new_with_proto( cx: &mut JSContext, document: &Document, proto: Option<HandleObject<'_>>, start_container: &Node, start_offset: u32, end_container: &Node, end_offset: u32, ) -> DomRoot<Range>

pub(crate) fn root(&self) -> DomRoot<Node>

https://dom.spec.whatwg.org/#concept-range-root

The root of a live range is the root of its start node.

pub(crate) fn contains(&self, no_gc: &NoGC, node: &Node) -> bool

https://dom.spec.whatwg.org/#contained

fn partially_contains(&self, node: &Node) -> bool

https://dom.spec.whatwg.org/#partially-contained

pub(crate) fn contained_children( &self, no_gc: &NoGC, ) -> Fallible<ContainedChildren>

https://dom.spec.whatwg.org/#concept-range-clone

pub(crate) fn set_start(&self, node: &Node, offset: u32)

https://dom.spec.whatwg.org/#concept-range-bp-set

pub(crate) fn set_start_without_reporting( &self, node: &Node, offset: u32, ) -> bool

pub(crate) fn set_end(&self, node: &Node, offset: u32)

https://dom.spec.whatwg.org/#concept-range-bp-set

pub(crate) fn set_end_without_reporting(&self, node: &Node, offset: u32) -> bool

fn compare_point( &self, no_gc: &NoGC, node: &Node, offset: u32, ) -> Fallible<Ordering>

https://dom.spec.whatwg.org/#dom-range-comparepointnode-offset

pub(crate) fn associate_selection(&self, selection: &Selection)

pub(crate) fn disassociate_selection(&self, selection: &Selection)

pub(crate) fn report_change(&self, notification: SelectionLiveRangeNotification)

fn abstract_range(&self) -> &AbstractRange

pub(crate) fn start(&self) -> &BoundaryPoint

pub(crate) fn end(&self) -> &BoundaryPoint

pub(crate) fn start_container(&self) -> DomRoot<Node>

pub(crate) fn start_offset(&self) -> u32

pub(crate) fn end_container(&self) -> DomRoot<Node>

pub(crate) fn end_offset(&self) -> u32

pub(crate) fn collapsed(&self) -> bool

fn client_rects(&self, no_gc: &NoGC) -> Vec<Rect<Au, CSSPixel>>

https://drafts.csswg.org/cssom-view/#dom-range-getclientrects

fn set_the_start_or_end( &self, no_gc: &NoGC, node: &Node, offset: u32, start_or_end: StartOrEnd, ) -> ErrorResult

https://dom.spec.whatwg.org/#concept-range-bp-set