script::dom::raredata

Struct ElementRareData

pub(crate) struct ElementRareData {

    pub(crate) shadow_root: Option<Dom<ShadowRoot>>,
    pub(crate) custom_element_reaction_queue: Vec<CustomElementReaction>,
    pub(crate) custom_element_definition: Option<Rc<CustomElementDefinition>>,
    pub(crate) custom_element_state: CustomElementState,
    pub(crate) custom_element_registry: Option<Dom<CustomElementRegistry>>,
    pub(crate) name_attribute: Option<Atom>,
    pub(crate) client_rect: Option<LayoutValue<Rect<i32, CSSPixel>>>,
    pub(crate) element_internals: Option<Dom<ElementInternals>>,
    pub(crate) registered_intersection_observers: Vec<IntersectionObserverRegistration>,
    pub(crate) cryptographic_nonce: String,
    pub(crate) forwarded_part_names: Vec<(String, String)>,
    pub(crate) part: MutNullableDom<DOMTokenList>,
    pub(crate) contenteditable_selection_range: MutNullableDom<Range>,
    pub(crate) had_duplicate_attributes: bool,
    pub(crate) previously_focused_element: MutNullableDom<Element>,

}

Fields
shadow_root: Option<Dom<ShadowRoot>>

https://dom.spec.whatwg.org/#dom-element-shadowroot The ShadowRoot this element is host of.

custom_element_reaction_queue: Vec<CustomElementReaction>

https://html.spec.whatwg.org/multipage/#custom-element-reaction-queue

custom_element_definition: Option<Rc<CustomElementDefinition>>

https://dom.spec.whatwg.org/#concept-element-custom-element-definition

custom_element_state: CustomElementState

https://dom.spec.whatwg.org/#concept-element-custom-element-state

custom_element_registry: Option<Dom<CustomElementRegistry>>

https://dom.spec.whatwg.org/#dom-element-customelementregistry

name_attribute: Option<Atom>

The “name” content attribute; not used as frequently as id, but used in named getter loops so it’s worth looking up quickly when present

client_rect: Option<LayoutValue<Rect<i32, CSSPixel>>>

The client rect reported by layout.

element_internals: Option<Dom<ElementInternals>>

https://html.spec.whatwg.org/multipage#elementinternals

registered_intersection_observers: Vec<IntersectionObserverRegistration>

https://w3c.github.io/IntersectionObserver/#element-private-slots

Element objects have an internal [[RegisteredIntersectionObservers]] slot, which is initialized to an empty list. This list holds IntersectionObserverRegistration records, which have:

cryptographic_nonce: Stringforwarded_part_names: Vec<(String, String)>

https://drafts.csswg.org/css-shadow-parts/#element-forwarded-part-name-list

part: MutNullableDom<DOMTokenList>

https://drafts.csswg.org/css-shadow-parts/#dom-element-part

contenteditable_selection_range: MutNullableDom<Range>

https://w3c.github.io/selection-api/#definition

This one selection must be shared by all the content of the document (though not by nested documents), including any editing hosts in the document.

had_duplicate_attributes: bool

Whether this element had duplicate attributes during tokenization. Used for CSP nonce validation (step 3 of “is element nonceable”).

previously_focused_element: MutNullableDom<Element>

https://html.spec.whatwg.org/multipage/#previously-focused-element