script::dom::security

Module sanitizer

Structs

Sanitizer 🔒

Constants

BUILT_IN_ANIMATING_URL_ATTRIBUTES_LIST 🔒

https://wicg.github.io/sanitizer-api/#built-in-animating-url-attributes-list

BUILT_IN_NAVIGATING_URL_ATTRIBUTES_LIST 🔒

https://wicg.github.io/sanitizer-api/#built-in-navigating-url-attributes-list

BUILT_IN_NON_REPLACEABLE_ELEMENTS_LIST 🔒

https://wicg.github.io/sanitizer-api/#built-in-non-replaceable-elements-list

_IMPL_DOMOBJECT_FOR_Sanitizer 🔒

Traits

AttributeMember 🔒

Helper functions for accessing the “attributes” and “removeAttributes” members of SanitizerElementWithAttributes.

Canonicalization 🔒

NameCanonicalization 🔒

NameMember 🔒

Helper functions for accessing the “name” and “namespace” members of SanitizerElementWithAttributes, SanitizerElement and SanitizerAttribute.

NameSlice 🔒

Supporting algorithms on lists of elements and lists of attributes, from the specification.

NameVec 🔒

Supporting algorithms on lists of elements and lists of attributes, from the specification.

SanitizerConfigAlgorithm 🔒

SanitizerMember 🔒

Helper functions for accessing the “sanitizer” members of SetHTMLOptions and SetHTMLUnsafeOptions.

TargetMember 🔒

Helper functions for accessing the “target” members of SanitizerPI.

TargetSlice 🔒

Supporting algorithms on lists of processing instructions, from the specification.

Functions

built_in_safe_baseline_configuration 🔒

https://wicg.github.io/sanitizer-api/#built-in-safe-baseline-configuration

built_in_safe_default_configuration 🔒

https://wicg.github.io/sanitizer-api/#built-in-safe-default-configuration

contains_javascript_url 🔒

https://wicg.github.io/sanitizer-api/#contains-a-javascript-url

sanitize_core 🔒

https://wicg.github.io/sanitizer-api/#sanitize-core