script::dom::security::sanitizer

Trait NameVec

trait NameVec<T>
where
    T: NameMember + Canonicalization + Clone,
{
    // Required methods
    fn remove_item<S: NameMember>(&mut self, item: &S) -> bool;
    fn add_item(&mut self, name: T);
    fn remove_duplicates(&mut self) -> &mut Self;
    fn intersection<S>(&mut self, others: &[S])
       where S: NameMember + Canonicalization + Clone;
    fn difference(&mut self, others: &[T]);
}

Supporting algorithms on lists of elements and lists of attributes, from the specification.

Required Methods

fn remove_item<S: NameMember>(&mut self, item: &S) -> bool

https://wicg.github.io/sanitizer-api/#sanitizerconfig-remove

fn add_item(&mut self, name: T)

https://wicg.github.io/sanitizer-api/#sanitizerconfig-add

fn remove_duplicates(&mut self) -> &mut Self

https://wicg.github.io/sanitizer-api/#sanitizerconfig-remove-duplicates

fn intersection<S>(&mut self, others: &[S])
where S: NameMember + Canonicalization + Clone,

Set itself to the set intersection of itself and another list.

https://infra.spec.whatwg.org/#set-intersection

fn difference(&mut self, others: &[T])

https://infra.spec.whatwg.org/#set-difference

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.