script::dom

Module selection

Structs

Selection 🔒

https://w3c.github.io/selection-api/#dfn-selection

SelectionLiveRangeNotification 🔒

Enums

Direction 🔒

FlatTreeNodePosition 🔒

Constants

_IMPL_DOMOBJECT_FOR_Selection 🔒

Functions

position_in_flat_tree_for_selection 🔒

Find the position of a node and offset in the flat tree for the purposes of selection boundaries. This projects the given position onto the flat tree, accounting for origin nodes that may not actually be in the flat tree at all.