script::dom::selection

Struct Selection

#[repr(C)]
pub(crate) struct Selection {
    reflector_: Reflector,
    document: Dom<Document>,
    range: DomRefCell<Option<SelectionRange>>,
    live_range: MutNullableDom<Range>,
    direction: Cell<Direction>,
    has_scheduled_selectionchange_event: Cell<bool>,
    visible_selection_dirty: Cell<bool>,
}

https://w3c.github.io/selection-api/#dfn-selection

Fields
reflector_: Reflectordocument: Dom<Document>range: DomRefCell<Option<SelectionRange>>

A range that holds the start and end of this selection, which may potentially cross shadow roots.

live_range: MutNullableDom<Range>

The live range version of this selection, which will never cross shadow roots.

direction: Cell<Direction>

The Direction of this Selection which determines which endpoint of Self::range is the anchor and which is the focus.

has_scheduled_selectionchange_event: Cell<bool>

https://w3c.github.io/selection-api/#dfn-has-scheduled-selectionchange-event

visible_selection_dirty: Cell<bool>

Whether or not this Selection needs to remark DOM nodes with selection flags after a change to its underlying Range.

Implementations

impl Selection

fn __assert_parent_type(&self)

impl Selection

pub(crate) fn delete_the_selection( &self, cx: &mut JSContext, context_object: &Document, block_merging: SelectionDeletionBlockMerging, strip_wrappers: SelectionDeletionStripWrappers, direction: SelectionDeleteDirection, )

https://w3c.github.io/editing/docs/execCommand/#delete-the-selection

pub(crate) fn set_the_selection_value( &self, cx: &mut JSContext, new_value: Option<DOMString>, command: CommandName, context_object: &Document, )

https://w3c.github.io/editing/docs/execCommand/#set-the-selection%27s-value

impl Selection

fn new_inherited(document: &Document) -> Selection

pub(crate) fn new(cx: &mut JSContext, document: &Document) -> DomRoot<Selection>

pub(crate) fn visible_selection_dirty(&self) -> bool

fn clear_cached_live_range(&self)

pub(crate) fn update_from_live_range( &self, live_range: &Range, notification: SelectionLiveRangeNotification, )

fn set_range(&self, _no_gc: &NoGC, new_range: Option<SelectionRange>) -> bool

pub(crate) fn set_live_range(&self, no_gc: &NoGC, new_range: Option<&Range>)

fn selection_boundaries_changed(&self)

fn iter_nodes_with_overlaps_document_selection_flag<'no_gc>( &self, no_gc: &'no_gc NoGC, ) -> impl Iterator<Item = UnrootedDom<'no_gc, Node>>

pub(crate) fn update_overlaps_document_selection_flags<'no_gc>( &self, no_gc: &'no_gc NoGC, )

pub(crate) fn queue_selectionchange_task(&self)

https://w3c.github.io/selection-api/#dfn-schedule-a-selectionchange-event

fn is_in_document_of_range(&self, node: &Node) -> bool

pub(crate) fn start_boundary(&self, cx: &mut JSContext) -> (DomRoot<Node>, u32)

pub(crate) fn end_boundary(&self, cx: &mut JSContext) -> (DomRoot<Node>, u32)

fn assert_valid_selection(&self, no_gc: &NoGC)

fn assert_valid_selection_and_live_range(&self, no_gc: &NoGC)

pub(crate) fn active_range(&self, cx: &mut JSContext) -> Option<DomRoot<Range>>

https://w3c.github.io/editing/docs/execCommand/#active-range

The active range is the range of the selection given by calling getSelection() on the context object. (Thus the active range may be null.)

pub(crate) fn expect_active_range(&self, cx: &mut JSContext) -> DomRoot<Range>

pub(crate) fn set_visible_selection_dirty(&self)

fn composed_anchor_position(&self) -> Option<(DomRoot<Node>, u32)>

fn live_anchor_node(&self, cx: &mut JSContext) -> Option<DomRoot<Node>>

https://w3c.github.io/selection-api/#dfn-anchor

fn live_anchor_offset(&self, cx: &mut JSContext) -> u32

https://w3c.github.io/selection-api/#dfn-anchor

fn live_focus_node(&self, cx: &mut JSContext) -> Option<DomRoot<Node>>

https://w3c.github.io/selection-api/#dfn-focus

fn live_focus_offset(&self, cx: &mut JSContext) -> u32

https://w3c.github.io/selection-api/#dfn-focus

pub(crate) fn insert_steps(&self, parent: &Node, child: &Node, count: u32)

https://dom.spec.whatwg.org/#concept-node-insert steps 5.1-5.2 and https://dom.spec.whatwg.org/#move steps 17.1-17.2 adapted for selections.

pub(crate) fn remove_steps_for_removed_subtree( &self, inclusive_descendant_of_removed_node: &Node, parent_of_removed_node: &Node, index_of_removed_node: &mut dyn FnMut() -> u32, )

https://dom.spec.whatwg.org/#live-range-pre-remove-steps steps 4 and 5 adapted for selections.

These steps are run on the inclusive descendants of a removed node, but to avoid having to iterate through those nodes twice, they are run when the inclusive descendants themselves are unbound from the tree.

pub(crate) fn remove_steps_for_parent( &self, parent: &Node, node_index: &mut dyn FnMut() -> u32, )

https://dom.spec.whatwg.org/#live-range-pre-remove-steps steps 6 and 7 adapted for selections.

pub(crate) fn normalization_steps( &self, parent: &Node, node: &Node, current_node: &Node, current_node_index: &dyn Fn() -> u32, length: u32, )

https://dom.spec.whatwg.org/#dom-node-normalize Steps 6.1-6.4 adapted for selections.

parent: The parent of both other node arguments.

node: The node that text is being merged into.

current_node: The node which has text being merged into node and will be removed from the DOM.

length: The length of the text content that was merged into node from siblings before current_node.

pub(crate) fn replace_data_steps( &self, node: &Node, offset: u32, removed_code_units: u32, added_code_units: &mut dyn FnMut() -> u32, )

https://dom.spec.whatwg.org/#concept-cd-replace steps 8-11 adapted for selections.

pub(crate) fn text_split_steps( &self, node: &Node, offset: u32, parent_node: &Node, new_node: &Node, )

https://dom.spec.whatwg.org/#concept-text-split steps 7.2-7.3 adapted for selections.

pub(crate) fn collapse_to_dom_position( &self, cx: &mut JSContext, container: &Node, offset: Utf32CodeUnitsOrNodeOffset, )

pub(crate) fn collapse_or_extend_to_dom_position( &self, cx: &mut JSContext, container: &Node, offset: Utf32CodeUnitsOrNodeOffset, )