script::dom

Module serviceworker

Modules

cache 🔒

cachestorage 🔒

client 🔒

extendableevent 🔒

extendablemessageevent 🔒

navigationpreloadmanager 🔒

notification 🔒

serviceworker 🔒

serviceworker_manager 🔒

The service worker manager persists the descriptor of any registered service workers. It also stores an active workers map, which holds descriptors of running service workers. If an active service worker timeouts, then it removes the descriptor entry from its active_workers map

serviceworkercontainer 🔒

serviceworkerglobalscope 🔒

serviceworkerregistration 🔒

windowclient 🔒