Struct script::dom::serviceworkerglobalscope::ServiceWorkerGlobalScope

#[repr(C)]
pub struct ServiceWorkerGlobalScope {
    workerglobalscope: WorkerGlobalScope,
    task_queue: TaskQueue<ServiceWorkerScriptMsg>,
    own_sender: Sender<ServiceWorkerScriptMsg>,
    time_out_port: Receiver<Instant>,
    swmanager_sender: IpcSender<ServiceWorkerMsg>,
    scope_url: ServoUrl,
    control_receiver: Receiver<ServiceWorkerControlMsg>,
}

Fields
workerglobalscope: WorkerGlobalScopetask_queue: TaskQueue<ServiceWorkerScriptMsg>own_sender: Sender<ServiceWorkerScriptMsg>time_out_port: Receiver<Instant>

A port on which a single “time-out” message can be received, indicating the sw should stop running, while still draining the task-queue

swmanager_sender: IpcSender<ServiceWorkerMsg>scope_url: ServoUrlcontrol_receiver: Receiver<ServiceWorkerControlMsg>

A receiver of control messages, currently only used to signal shutdown.

Implementations

impl ServiceWorkerGlobalScope

fn __assert_parent_type(&self)

impl ServiceWorkerGlobalScope

fn new_inherited( init: WorkerGlobalScopeInit, worker_url: ServoUrl, from_devtools_receiver: Receiver<DevtoolScriptControlMsg>, runtime: Runtime, own_sender: Sender<ServiceWorkerScriptMsg>, receiver: Receiver<ServiceWorkerScriptMsg>, time_out_port: Receiver<Instant>, swmanager_sender: IpcSender<ServiceWorkerMsg>, scope_url: ServoUrl, control_receiver: Receiver<ServiceWorkerControlMsg>, closing: Arc<AtomicBool> ) -> ServiceWorkerGlobalScope

pub fn new( init: WorkerGlobalScopeInit, worker_url: ServoUrl, from_devtools_receiver: Receiver<DevtoolScriptControlMsg>, runtime: Runtime, own_sender: Sender<ServiceWorkerScriptMsg>, receiver: Receiver<ServiceWorkerScriptMsg>, time_out_port: Receiver<Instant>, swmanager_sender: IpcSender<ServiceWorkerMsg>, scope_url: ServoUrl, control_receiver: Receiver<ServiceWorkerControlMsg>, closing: Arc<AtomicBool> ) -> Root<Dom<ServiceWorkerGlobalScope>>

pub fn run_serviceworker_scope( scope_things: ScopeThings, own_sender: Sender<ServiceWorkerScriptMsg>, receiver: Receiver<ServiceWorkerScriptMsg>, devtools_receiver: IpcReceiver<DevtoolScriptControlMsg>, swmanager_sender: IpcSender<ServiceWorkerMsg>, scope_url: ServoUrl, control_receiver: Receiver<ServiceWorkerControlMsg>, context_sender: Sender<ContextForRequestInterrupt>, closing: Arc<AtomicBool> ) -> JoinHandle<()>

fn handle_mixed_message(&self, msg: MixedMessage) -> bool

fn has_timed_out(&self) -> bool

fn handle_script_event(&self, msg: ServiceWorkerScriptMsg)

pub fn script_chan(&self) -> Box<dyn ScriptChan + Send>

fn dispatch_activate(&self)