script::dom::servoparser::encoding

Struct DetectingState

pub(super) struct DetectingState {
    encoding_hint_from_content_type: Option<&'static Encoding>,
    encoding_of_container_document: Option<&'static Encoding>,
    start_timestamp: Instant,
    attempted_bom_sniffing: bool,
    buffered_bytes: Vec<u8>,
}

Fields
encoding_hint_from_content_type: Option<&'static Encoding>

The charset that was specified in the Content-Type header, if any.

encoding_of_container_document: Option<&'static Encoding>

The encoding of a same-origin container document, if this document is in an <iframe>.

start_timestamp: Instantattempted_bom_sniffing: boolbuffered_bytes: Vec<u8>
Implementations

impl DetectingState

const BUFFER_THRESHOLD: usize = 1024

The maximum amount of bytes to buffer before attempting to determine the encoding

const MAX_TIME_TO_BUFFER: Duration

The time threshold after which we will attempt to determine the encoding and start decoding, even if there are less than [BUFFER_THRESHOLD] bytes in the buffer.

fn buffer( &mut self, data: &[u8], document: &Document, is_at_end_of_file: AtEndOfFile, ) -> Option<&'static Encoding>

Appends some data to the internal buffer and attempts to determine the character encoding.

If an encoding was detected then it is returned. A return value of None indicates that more bytes are required.

fn determine_the_character_encoding( &mut self, document: &Document, potentially_wait_for_more_data: bool, is_at_end_of_file: AtEndOfFile, ) -> Option<&'static Encoding>

https://html.spec.whatwg.org/multipage/#determining-the-character-encoding

fn finish(&mut self, document: &Document) -> &'static Encoding