Struct script::dom::servoxmlparser::ServoXMLParser

pub struct ServoXMLParser {
    reflector_: Reflector,
    tokenizer: DOMRefCell<Tokenizer>,
    pending_input: DOMRefCell<Vec<String>>,
    document: JS<Document>,
    suspended: Cell<bool>,
    last_chunk_received: Cell<bool>,
    pipeline: Option<PipelineId>,
}

Fields
reflector_: Reflector tokenizer: DOMRefCell<Tokenizer> pending_input: DOMRefCell<Vec<String>>

Input chunks received but not yet passed to the parser.

document: JS<Document>

The document associated with this parser.

suspended: Cell<bool>

True if this parser should avoid passing any further data to the tokenizer.

last_chunk_received: Cell<bool>

Whether to expect any further input from the associated network request.

pipeline: Option<PipelineId>

The pipeline associated with this parse, unavailable if this parse does not correspond to a page load.

Methods

impl ServoXMLParser

fn new(base_url: Option<Url>, document: &Document, pipeline: Option<PipelineId>) -> Root<ServoXMLParser>

fn window(&self) -> &Window

fn resume(&self)

fn suspend(&self)

fn is_suspended(&self) -> bool

fn parse_sync(&self)

fn pending_input(&self) -> &DOMRefCell<Vec<String>>

fn set_plaintext_state(&self)

fn end_tokenizer(&self)

fn document(&self) -> &Document

fn last_chunk_received(&self) -> &Cell<bool>

fn tokenizer(&self) -> &DOMRefCell<Tokenizer>