script::dom::shadowroot

Struct ShadowRoot

#[repr(C)]
pub(crate) struct ShadowRoot {

    document_fragment: DocumentFragment,
    document_or_shadow_root: DocumentOrShadowRoot,
    document: Dom<Document>,
    author_styles: DomRefCell<AuthorStyles<ServoStylesheetInDocument>>,
    stylesheet_list: MutNullableDom<StyleSheetList>,
    window: Dom<Window>,
    mode: ShadowRootMode,
    slot_assignment_mode: SlotAssignmentMode,
    clonable: bool,
    available_to_element_internals: Cell<bool>,
    slots: DomRefCell<HashMap<DOMString, Vec<Dom<HTMLSlotElement>>>>,
    is_user_agent_widget: bool,
    declarative: Cell<bool>,
    serializable: Cell<bool>,
    delegates_focus: Cell<bool>,
    adopted_stylesheets: DomRefCell<Vec<Dom<CSSStyleSheet>>>,
    adopted_stylesheets_frozen_types: CachedFrozenArray,
    details_name_groups: DomRefCell<Option<DetailsNameGroups>>,

}

https://dom.spec.whatwg.org/#interface-shadowroot

Fields
document_fragment: DocumentFragment

The DocumentFragment that this ShadowRoot inherits from.

document_or_shadow_root: DocumentOrShadowRootdocument: Dom<Document>author_styles: DomRefCell<AuthorStyles<ServoStylesheetInDocument>>

List of author styles associated with nodes in this shadow tree.

stylesheet_list: MutNullableDom<StyleSheetList>window: Dom<Window>mode: ShadowRootMode

https://dom.spec.whatwg.org/#dom-shadowroot-mode

slot_assignment_mode: SlotAssignmentMode

https://dom.spec.whatwg.org/#dom-shadowroot-slotassignment

clonable: bool

https://dom.spec.whatwg.org/#dom-shadowroot-clonable

available_to_element_internals: Cell<bool>

https://dom.spec.whatwg.org/#shadowroot-available-to-element-internals

slots: DomRefCell<HashMap<DOMString, Vec<Dom<HTMLSlotElement>>>>is_user_agent_widget: booldeclarative: Cell<bool>

https://dom.spec.whatwg.org/#shadowroot-declarative

serializable: Cell<bool>

https://dom.spec.whatwg.org/#shadowroot-serializable

delegates_focus: Cell<bool>

https://dom.spec.whatwg.org/#shadowroot-delegates-focus

adopted_stylesheets: DomRefCell<Vec<Dom<CSSStyleSheet>>>

The constructed stylesheet that is adopted by this ShadowRoot. https://drafts.csswg.org/cssom/#dom-documentorshadowroot-adoptedstylesheets

adopted_stylesheets_frozen_types: CachedFrozenArray

Cached frozen array of Self::adopted_stylesheets

details_name_groups: DomRefCell<Option<DetailsNameGroups>>
Implementations

impl ShadowRoot

fn __assert_parent_type(&self)

impl ShadowRoot

fn new_inherited( host: &Element, document: &Document, mode: ShadowRootMode, slot_assignment_mode: SlotAssignmentMode, clonable: bool, is_user_agent_widget: IsUserAgentWidget, ) -> ShadowRoot

pub(crate) fn new( cx: &mut JSContext, host: &Element, document: &Document, mode: ShadowRootMode, slot_assignment_mode: SlotAssignmentMode, clonable: bool, is_user_agent_widget: IsUserAgentWidget, ) -> DomRoot<ShadowRoot>

pub(crate) fn host_unrooted<'a>( &self, no_gc: &'a NoGC, ) -> UnrootedDom<'a, Element>

pub(crate) fn owner_doc(&self) -> &Document

pub(crate) fn stylesheet_count(&self) -> usize

pub(crate) fn stylesheet_at( &self, cx: &mut JSContext, index: usize, ) -> Option<DomRoot<CSSStyleSheet>>

pub(crate) fn add_owned_stylesheet( &self, no_gc: &NoGC, owner_node: &Element, sheet: Arc<Stylesheet>, )

Add a stylesheet owned by owner_node to the list of shadow root sheets, in the correct tree position. Additionally, ensure that owned stylesheet is inserted before any constructed stylesheet.

https://drafts.csswg.org/cssom/#documentorshadowroot-final-css-style-sheets

pub(crate) fn append_constructed_stylesheet( &self, cssom_stylesheet: &CSSStyleSheet, )

Append a constructed stylesheet to the back of shadow root stylesheet set.

pub(crate) fn remove_stylesheet( &self, owner: StylesheetSource, s: &Arc<Stylesheet>, )

Remove a stylesheet owned by owner from the list of shadow root sheets.

pub(crate) fn invalidate_stylesheets(&self, no_gc: &NoGC)

pub(crate) fn unregister_element_id(&self, id: &Atom)

Remove any existing association between the provided id and any elements in this shadow tree.

pub(crate) fn register_element_id(&self, element: &Element, id: &Atom)

Associate an element present in this shadow tree with the provided id.

pub(crate) fn register_slot(&self, slot: &HTMLSlotElement)

pub(crate) fn unregister_slot(&self, name: DOMString, slot: &HTMLSlotElement)

pub(crate) fn slot_for_name( &self, name: &DOMString, ) -> Option<DomRoot<HTMLSlotElement>>

Find the first slot with the given name among this root’s descendants in tree order

pub(crate) fn has_slot_descendants(&self) -> bool

pub(crate) fn set_available_to_element_internals(&self, value: bool)

pub(crate) fn is_available_to_element_internals(&self) -> bool

https://dom.spec.whatwg.org/#shadowroot-available-to-element-internals

pub(crate) fn is_user_agent_widget(&self) -> bool

pub(crate) fn set_declarative(&self, declarative: bool)

pub(crate) fn is_declarative(&self) -> bool

pub(crate) fn shadow_root_mode(&self) -> ShadowRootMode

pub(crate) fn set_serializable(&self, serializable: bool)

pub(crate) fn set_delegates_focus(&self, delegates_focus: bool)

pub(crate) fn details_name_groups<'a: 'b, 'b>( &'a self, no_gc: &'b NoGC, ) -> RefMut<'b, DetailsNameGroups>

pub(crate) fn custom_element_registry( &self, ) -> Option<DomRoot<CustomElementRegistry>>

pub(crate) fn set_custom_element_registry( &self, registry: Option<&CustomElementRegistry>, )