script::dom::stream::byteteeunderlyingsource

Struct ByteTeeUnderlyingSource

#[repr(C)]
pub(crate) struct ByteTeeUnderlyingSource {

    reflector_: Reflector,
    reader: Rc<RefCell<ReaderType>>,
    stream: Dom<ReadableStream>,
    branch_1: MutNullableDom<ReadableStream>,
    branch_2: MutNullableDom<ReadableStream>,
    read_again_for_branch_1: Rc<Cell<bool>>,
    read_again_for_branch_2: Rc<Cell<bool>>,
    reading: Rc<Cell<bool>>,
    canceled_1: Rc<Cell<bool>>,
    canceled_2: Rc<Cell<bool>>,
    reason_1: Rc<Heap<Value>>,
    reason_2: Rc<Heap<Value>>,
    cancel_promise: Rc<Promise>,
    reader_version: Rc<Cell<u64>>,
    tee_cancel_algorithm: ByteTeeCancelAlgorithm,
    byte_tee_pull_algorithm: ByteTeePullAlgorithm,

}

https://streams.spec.whatwg.org/#abstract-opdef-readablestreamdefaulttee

Fields
reflector_: Reflectorreader: Rc<RefCell<ReaderType>>stream: Dom<ReadableStream>branch_1: MutNullableDom<ReadableStream>branch_2: MutNullableDom<ReadableStream>read_again_for_branch_1: Rc<Cell<bool>>read_again_for_branch_2: Rc<Cell<bool>>reading: Rc<Cell<bool>>canceled_1: Rc<Cell<bool>>canceled_2: Rc<Cell<bool>>reason_1: Rc<Heap<Value>>reason_2: Rc<Heap<Value>>cancel_promise: Rc<Promise>reader_version: Rc<Cell<u64>>tee_cancel_algorithm: ByteTeeCancelAlgorithmbyte_tee_pull_algorithm: ByteTeePullAlgorithm
Implementations

impl ByteTeeUnderlyingSource

fn __assert_parent_type(&self)

impl ByteTeeUnderlyingSource

pub(crate) fn new( cx: &mut JSContext, reader: Rc<RefCell<ReaderType>>, stream: &ReadableStream, reading: Rc<Cell<bool>>, read_again_for_branch_1: Rc<Cell<bool>>, read_again_for_branch_2: Rc<Cell<bool>>, canceled_1: Rc<Cell<bool>>, canceled_2: Rc<Cell<bool>>, reason_1: Rc<Heap<Value>>, reason_2: Rc<Heap<Value>>, cancel_promise: Rc<Promise>, reader_version: Rc<Cell<u64>>, tee_cancel_algorithm: ByteTeeCancelAlgorithm, byte_tee_pull_algorithm: ByteTeePullAlgorithm, ) -> DomRoot<ByteTeeUnderlyingSource>

pub(crate) fn set_branch_1(&self, stream: &ReadableStream)

pub(crate) fn set_branch_2(&self, stream: &ReadableStream)

pub(crate) fn forward_reader_error( &self, cx: &mut JSContext, this_reader: Rc<RefCell<ReaderType>>, )

fn pull_with_default_reader( &self, cx: &mut JSContext, global: &GlobalScope, ) -> Fallible<()>

fn pull_with_byob_reader( &self, cx: &mut JSContext, view: &HeapBufferSource<ArrayBufferViewU8>, for_branch2: bool, global: &GlobalScope, )

pub(crate) fn pull_algorithm( &self, cx: &mut JSContext, byte_tee_pull_algorithm: Option<ByteTeePullAlgorithm>, ) -> Rc<Promise>

Let pullAlgorithm be the following steps:

pub(crate) fn cancel_algorithm( &self, cx: &mut JSContext, reason: SafeHandleValue<'_>, ) -> Option<Result<Rc<Promise>, Error>>

https://streams.spec.whatwg.org/#abstract-opdef-readablestreamdefaulttee Let cancel1Algorithm be the following steps, taking a reason argument and Let cancel2Algorithm be the following steps, taking a reason argument

fn resolve_cancel_promise(&self, cx: &mut JSContext)