script::dom::stream::readablestream

Struct PipeTo

pub(crate) struct PipeTo {
    reader: Dom<ReadableStreamDefaultReader>,
    writer: Dom<WritableStreamDefaultWriter>,
    pending_writes: Rc<RefCell<VecDeque<TracedPromise>>>,
    state: Rc<RefCell<PipeToState>>,
    prevent_abort: bool,
    prevent_cancel: bool,
    prevent_close: bool,
    shutting_down: Rc<Cell<bool>>,
    abort_reason: Rc<Heap<JSVal>>,
    shutdown_error: Rc<RefCell<Option<Heap<JSVal>>>>,
    shutdown_action_promise: Rc<RefCell<Option<TracedPromise>>>,
    result_promise: TracedPromise,
}

The “in parallel, but not really” part of https://streams.spec.whatwg.org/#readable-stream-pipe-to

Note: the spec is flexible about how this is done, but requires the following constraints to apply:

Public API must not be used: we’ll only use Rust.

Backpressure must be enforced: we’ll only read from source when dest is ready.

Shutdown must stop activity: we’ll do this together with the below.

Error and close states must be propagated: we’ll do this by checking these states at every step.

Fields
reader: Dom<ReadableStreamDefaultReader>

https://streams.spec.whatwg.org/#ref-for-readablestream%E2%91%A7%E2%91%A0

writer: Dom<WritableStreamDefaultWriter>

https://streams.spec.whatwg.org/#ref-for-acquire-writable-stream-default-writer

pending_writes: Rc<RefCell<VecDeque<TracedPromise>>>

Pending writes are needed when shutting down(with an action), because we can only finalize when all writes are finished.

state: Rc<RefCell<PipeToState>>

The state machine.

prevent_abort: bool

https://streams.spec.whatwg.org/#readablestream-pipe-to-preventabort

prevent_cancel: bool

https://streams.spec.whatwg.org/#readablestream-pipe-to-preventcancel

prevent_close: bool

https://streams.spec.whatwg.org/#readablestream-pipe-to-preventclose

shutting_down: Rc<Cell<bool>>

The shuttingDown variable of https://streams.spec.whatwg.org/#readable-stream-pipe-to

abort_reason: Rc<Heap<JSVal>>

The abort reason of the abort signal, stored here because we must keep it across a microtask.

shutdown_error: Rc<RefCell<Option<Heap<JSVal>>>>

The error potentially passed to shutdown, stored here because we must keep it across a microtask.

shutdown_action_promise: Rc<RefCell<Option<TracedPromise>>>

The promise returned by a shutdown action. We keep it to only continue when it is not pending anymore.

result_promise: TracedPromise

The promise resolved or rejected at https://streams.spec.whatwg.org/#rs-pipeTo-finalize

Implementations

impl PipeTo

pub(crate) fn abort_with_reason( &self, cx: &mut CurrentRealm<'_>, global: &GlobalScope, reason: SafeHandleValue<'_>, )

Run the abortAlgorithm defined at https://streams.spec.whatwg.org/#readable-stream-pipe-to

impl PipeTo

fn set_shutdown_error(&self, error: SafeHandleValue<'_>)

Setting shutdown error in a way that ensures it isn’t moved after it has been set.

fn wait_for_writer_ready(&self, cx: &mut CurrentRealm<'_>, global: &GlobalScope)

Wait for the writer to be ready, which implements the constraint that backpressure must be enforced.

fn read_chunk(&self, cx: &mut CurrentRealm<'_>, global: &GlobalScope)

Read a chunk

fn write_chunk( &self, cx: &mut JSContext, global: &GlobalScope, chunk: SafeHandleValue<'_>, ) -> bool

Try to write a chunk using the jsval, and returns wether it succeeded

fn wait_on_pending_write( &self, cx: &mut CurrentRealm<'_>, global: &GlobalScope, promise: &Promise, )

Only as part of shutting-down do we wait on pending writes (backpressure is communicated not through pending writes but through the readiness of the writer).

fn check_and_propagate_errors_forward( &self, cx: &mut CurrentRealm<'_>, global: &GlobalScope, )

Errors must be propagated forward part of https://streams.spec.whatwg.org/#readable-stream-pipe-to

fn check_and_propagate_errors_backward( &self, cx: &mut CurrentRealm<'_>, global: &GlobalScope, )

Errors must be propagated backward part of https://streams.spec.whatwg.org/#readable-stream-pipe-to

fn check_and_propagate_closing_forward( &self, cx: &mut CurrentRealm<'_>, global: &GlobalScope, )

Closing must be propagated forward part of https://streams.spec.whatwg.org/#readable-stream-pipe-to

fn check_and_propagate_closing_backward( &self, cx: &mut CurrentRealm<'_>, global: &GlobalScope, )

Closing must be propagated backward part of https://streams.spec.whatwg.org/#readable-stream-pipe-to

fn shutdown( &self, cx: &mut CurrentRealm<'_>, global: &GlobalScope, action: Option<ShutdownAction>, )

https://streams.spec.whatwg.org/#rs-pipeTo-shutdown-with-action https://streams.spec.whatwg.org/#rs-pipeTo-shutdown Combined into one method with an optional action.

fn perform_action( &self, cx: &mut CurrentRealm<'_>, global: &GlobalScope, action: ShutdownAction, )

The perform action part of https://streams.spec.whatwg.org/#rs-pipeTo-shutdown-with-action

fn finalize(&self, cx: &mut JSContext, global: &GlobalScope)

https://streams.spec.whatwg.org/#rs-pipeTo-finalize