script::dom::stream::readablestream

Struct ReadableStream

#[repr(C)]
pub(crate) struct ReadableStream {
    reflector_: Reflector,
    controller: RefCell<Option<ControllerType>>,
    stored_error: Heap<JSVal>,
    disturbed: Cell<bool>,
    reader: RefCell<Option<ReaderType>>,
    state: Cell<ReadableStreamState>,
}

https://streams.spec.whatwg.org/#rs-class

Fields
reflector_: Reflectorcontroller: RefCell<Option<ControllerType>>

https://streams.spec.whatwg.org/#readablestream-controller Note: the inner MutNullableDom should really be an Option<Dom>, because it is never unset once set.

stored_error: Heap<JSVal>

https://streams.spec.whatwg.org/#readablestream-storederror

disturbed: Cell<bool>

https://streams.spec.whatwg.org/#readablestream-disturbed

reader: RefCell<Option<ReaderType>>

https://streams.spec.whatwg.org/#readablestream-reader

state: Cell<ReadableStreamState>

https://streams.spec.whatwg.org/#readablestream-state

Implementations

impl ReadableStream

fn __assert_parent_type(&self)

impl ReadableStream

fn new_inherited() -> ReadableStream

https://streams.spec.whatwg.org/#initialize-readable-stream

pub(crate) fn new_with_proto( cx: &mut JSContext, global: &GlobalScope, proto: Option<SafeHandleObject<'_>>, ) -> DomRoot<ReadableStream>

pub(crate) fn set_default_controller( &self, controller: &ReadableStreamDefaultController, )

Used as part of https://streams.spec.whatwg.org/#set-up-readable-stream-default-controller

pub(crate) fn set_byte_controller( &self, controller: &ReadableByteStreamController, )

Used as part of https://streams.spec.whatwg.org/#set-up-readable-byte-stream-controller

pub(crate) fn assert_no_controller(&self)

Used as part of https://streams.spec.whatwg.org/#set-up-readable-stream-default-controller

pub(crate) fn new_from_bytes( cx: &mut JSContext, global: &GlobalScope, bytes: Vec<u8>, ) -> Fallible<DomRoot<ReadableStream>>

Build a stream backed by a Rust source that has already been read into memory.

pub(crate) fn new_empty( cx: &mut JSContext, global: &GlobalScope, ) -> Fallible<DomRoot<ReadableStream>>

Build an empty stream. Used as step 2 of https://fetch.spec.whatwg.org/#dom-body-textstream

pub(crate) fn new_from_bytes_with_byte_reading_support( cx: &mut JSContext, global: &GlobalScope, bytes: Vec<u8>, ) -> Fallible<DomRoot<ReadableStream>>

https://streams.spec.whatwg.org/#readablestream-set-up-with-byte-reading-support

pub(crate) fn new_with_external_underlying_source( cx: &mut JSContext, global: &GlobalScope, source: UnderlyingSourceType<'_>, ) -> Fallible<DomRoot<ReadableStream>>

Build a stream backed by a Rust underlying source. Note: external sources are always paired with a default controller.

pub(crate) fn new_with_external_underlying_byte_source( cx: &mut JSContext, global: &GlobalScope, source: UnderlyingSourceType<'_>, ) -> Fallible<DomRoot<ReadableStream>>

https://streams.spec.whatwg.org/#readablestream-set-up-with-byte-reading-support

pub(crate) fn perform_release_steps(&self) -> Fallible<()>

Call into the release steps of the controller,

pub(crate) fn perform_pull_steps( &self, cx: &mut JSContext, read_request: &ReadRequest, )

Call into the pull steps of the controller, as part of https://streams.spec.whatwg.org/#readable-stream-default-reader-read

pub(crate) fn perform_pull_into( &self, cx: &mut JSContext, read_into_request: &ReadIntoRequest, view: &HeapBufferSource<ArrayBufferViewU8>, min: u64, )

Call into the pull steps of the controller, as part of https://streams.spec.whatwg.org/#readable-stream-byob-reader-read

pub(crate) fn add_read_request(&self, read_request: &ReadRequest)

https://streams.spec.whatwg.org/#readable-stream-add-read-request

pub(crate) fn add_read_into_request(&self, read_request: &ReadIntoRequest)

https://streams.spec.whatwg.org/#readable-stream-add-read-into-request

pub(crate) fn enqueue_native(&self, cx: &mut JSContext, bytes: Vec<u8>)

https://streams.spec.whatwg.org/#readablestream-enqueue

pub(crate) fn error(&self, cx: &mut JSContext, e: SafeHandleValue<'_>)

https://streams.spec.whatwg.org/#readable-stream-error

pub(crate) fn get_stored_error(&self, handle_mut: SafeMutableHandleValue<'_>)

https://streams.spec.whatwg.org/#readablestream-storederror

pub(crate) fn error_native(&self, cx: &mut JSContext, error: Error)

https://streams.spec.whatwg.org/#readable-stream-error Note: in other use cases this call happens via the controller.

pub(crate) fn controller_close_native(&self, cx: &mut JSContext)

Call into the controller’s Close method. https://streams.spec.whatwg.org/#readablestream-close

pub(crate) fn in_memory(&self) -> bool

Returns a boolean reflecting whether the stream has all data in memory. Useful for native source integration only.

pub(crate) fn get_in_memory_bytes( &self, cx: &mut JSContext, ) -> Option<GenericSharedMemory>

Return bytes for synchronous use, if the stream has all data in memory. Useful for native source integration only.

pub(crate) fn acquire_default_reader( &self, cx: &mut JSContext, ) -> Fallible<DomRoot<ReadableStreamDefaultReader>>

Acquires a reader and locks the stream, must be done before read_a_chunk. Native call to https://streams.spec.whatwg.org/#acquire-readable-stream-reader

pub(crate) fn acquire_byob_reader( &self, cx: &mut JSContext, ) -> Fallible<DomRoot<ReadableStreamBYOBReader>>

https://streams.spec.whatwg.org/#acquire-readable-stream-byob-reader

pub(crate) fn get_default_controller( &self, ) -> DomRoot<ReadableStreamDefaultController>

pub(crate) fn get_byte_controller( &self, ) -> DomRoot<ReadableByteStreamController>

pub(crate) fn get_default_reader(&self) -> DomRoot<ReadableStreamDefaultReader>

pub(crate) fn read_a_chunk(&self, cx: &mut JSContext) -> Rc<Promise>

Read a chunk from the stream, must be called after start_reading, and before stop_reading. Native call to https://streams.spec.whatwg.org/#readable-stream-default-reader-read

pub(crate) fn stop_reading(&self, cx: &mut JSContext)

Releases the lock on the reader, must be done after start_reading. Native call to https://streams.spec.whatwg.org/#abstract-opdef-readablestreamdefaultreaderrelease

pub(crate) fn is_locked(&self) -> bool

https://streams.spec.whatwg.org/#is-readable-stream-locked

pub(crate) fn is_disturbed(&self) -> bool

pub(crate) fn set_is_disturbed(&self, disturbed: bool)

pub(crate) fn is_closed(&self) -> bool

pub(crate) fn is_errored(&self) -> bool

pub(crate) fn is_readable(&self) -> bool

pub(crate) fn has_default_reader(&self) -> bool

pub(crate) fn has_byob_reader(&self) -> bool

pub(crate) fn has_byte_controller(&self) -> bool

pub(crate) fn get_num_read_requests(&self) -> usize

https://streams.spec.whatwg.org/#readable-stream-get-num-read-requests

pub(crate) fn get_num_read_into_requests(&self) -> usize

https://streams.spec.whatwg.org/#readable-stream-get-num-read-into-requests

pub(crate) fn fulfill_read_request( &self, cx: &mut JSContext, chunk: SafeHandleValue<'_>, done: bool, )

https://streams.spec.whatwg.org/#readable-stream-fulfill-read-request

pub(crate) fn fulfill_read_into_request( &self, cx: &mut JSContext, chunk: SafeHandleValue<'_>, done: bool, )

https://streams.spec.whatwg.org/#readable-stream-fulfill-read-into-request

pub(crate) fn close(&self, cx: &mut JSContext)

https://streams.spec.whatwg.org/#readable-stream-close

pub(crate) fn cancel( &self, cx: &mut JSContext, global: &GlobalScope, reason: SafeHandleValue<'_>, ) -> RootedPromise

https://streams.spec.whatwg.org/#readable-stream-cancel

pub(crate) fn set_reader(&self, new_reader: Option<ReaderType>)

fn byte_tee(&self, cx: &mut JSContext) -> Fallible<Vec<DomRoot<ReadableStream>>>

https://streams.spec.whatwg.org/#abstract-opdef-readablebytestreamtee

fn default_tee( &self, cx: &mut JSContext, clone_for_branch_2: bool, ) -> Fallible<Vec<DomRoot<ReadableStream>>>

https://streams.spec.whatwg.org/#abstract-opdef-readablestreamdefaulttee

pub(crate) fn pipe_to( &self, cx: &mut CurrentRealm<'_>, global: &GlobalScope, dest: &WritableStream, prevent_close: bool, prevent_abort: bool, prevent_cancel: bool, signal: Option<&AbortSignal>, ) -> RootedPromise

https://streams.spec.whatwg.org/#readable-stream-pipe-to

pub(crate) fn tee( &self, cx: &mut JSContext, clone_for_branch_2: bool, ) -> Fallible<Vec<DomRoot<ReadableStream>>>

https://streams.spec.whatwg.org/#readable-stream-tee

fn set_up_byte_controller( &self, cx: &mut JSContext, global: &GlobalScope, underlying_source_dict: UnderlyingSource<DomTypeHolder>, underlying_source_handle: SafeHandleObject<'_>, stream: &ReadableStream, strategy_hwm: f64, ) -> Fallible<()>

https://streams.spec.whatwg.org/#set-up-readable-byte-stream-controller-from-underlying-source

pub(crate) fn setup_cross_realm_transform_readable( &self, cx: &mut JSContext, port: &MessagePort, )

https://streams.spec.whatwg.org/#abstract-opdef-setupcrossrealmtransformreadable