script::dom::stream::readablestreamdefaultcontroller

Struct ReadableStreamDefaultController

#[repr(C)]
pub(crate) struct ReadableStreamDefaultController {
    reflector_: Reflector,
    queue: QueueWithSizes,
    underlying_source: MutNullableDom<UnderlyingSourceContainer>,
    stream: MutNullableDom<ReadableStream>,
    strategy_hwm: f64,
    strategy_size: RefCell<Option<Rc<QueuingStrategySize<DomTypeHolder>>>>,
    close_requested: Cell<bool>,
    started: Cell<bool>,
    pulling: Cell<bool>,
    pull_again: Cell<bool>,
}

https://streams.spec.whatwg.org/#readablestreamdefaultcontroller

Fields
reflector_: Reflectorqueue: QueueWithSizes

https://streams.spec.whatwg.org/#readablestreamdefaultcontroller-queue

underlying_source: MutNullableDom<UnderlyingSourceContainer>

A mutable reference to the underlying source is used to implement these two internal slots:

https://streams.spec.whatwg.org/#readablestreamdefaultcontroller-pullalgorithm https://streams.spec.whatwg.org/#readablestreamdefaultcontroller-cancelalgorithm

stream: MutNullableDom<ReadableStream>strategy_hwm: f64

https://streams.spec.whatwg.org/#readablestreamdefaultcontroller-strategyhwm

strategy_size: RefCell<Option<Rc<QueuingStrategySize<DomTypeHolder>>>>

https://streams.spec.whatwg.org/#readablestreamdefaultcontroller-strategysizealgorithm

close_requested: Cell<bool>

https://streams.spec.whatwg.org/#readablestreamdefaultcontroller-closerequested

started: Cell<bool>

https://streams.spec.whatwg.org/#readablestreamdefaultcontroller-started

pulling: Cell<bool>

https://streams.spec.whatwg.org/#readablestreamdefaultcontroller-pulling

pull_again: Cell<bool>

https://streams.spec.whatwg.org/#readablestreamdefaultcontroller-pullagain

Implementations

impl ReadableStreamDefaultController

fn __assert_parent_type(&self)

impl ReadableStreamDefaultController

fn new_inherited( strategy_hwm: f64, strategy_size: Rc<QueuingStrategySize<DomTypeHolder>>, underlying_source: &UnderlyingSourceContainer, ) -> ReadableStreamDefaultController

pub(crate) fn new( cx: &mut JSContext, global: &GlobalScope, underlying_source: UnderlyingSourceType<'_>, strategy_hwm: f64, strategy_size: Rc<QueuingStrategySize<DomTypeHolder>>, ) -> DomRoot<ReadableStreamDefaultController>

pub(crate) fn setup( &self, cx: &mut JSContext, stream: &ReadableStream, ) -> Result<(), Error>

https://streams.spec.whatwg.org/#set-up-readable-stream-default-controller

pub(crate) fn set_underlying_source_this_object( &self, this_object: HandleObject<'_>, )

Setting the JS object after the heap has settled down.

fn dequeue_value(&self, cx: &mut JSContext, rval: MutableHandleValue<'_>)

https://streams.spec.whatwg.org/#dequeue-value

fn should_call_pull(&self) -> bool

https://streams.spec.whatwg.org/#readable-stream-default-controller-should-call-pull

fn call_pull_if_needed(&self, cx: &mut JSContext)

https://streams.spec.whatwg.org/#readable-stream-default-controller-call-pull-if-needed

pub(crate) fn perform_cancel_steps( &self, cx: &mut JSContext, global: &GlobalScope, reason: SafeHandleValue<'_>, ) -> Rc<Promise>

https://streams.spec.whatwg.org/#rs-default-controller-private-cancel

pub(crate) fn perform_pull_steps( &self, cx: &mut JSContext, read_request: &ReadRequest, )

https://streams.spec.whatwg.org/#rs-default-controller-private-pull

pub(crate) fn perform_release_steps(&self) -> Fallible<()>

https://streams.spec.whatwg.org/#ref-for-abstract-opdef-readablestreamcontroller-releasesteps

pub(crate) fn enqueue( &self, cx: &mut JSContext, chunk: SafeHandleValue<'_>, ) -> Result<(), Error>

https://streams.spec.whatwg.org/#readable-stream-default-controller-enqueue

pub(crate) fn enqueue_native(&self, cx: &mut JSContext, chunk: Vec<u8>)

Native call to https://streams.spec.whatwg.org/#readable-stream-default-controller-enqueue

pub(crate) fn in_memory(&self) -> bool

Does the stream have all data in memory?

pub(crate) fn get_in_memory_bytes(&self) -> Option<Vec<u8>>

Return bytes synchronously if the stream has all data in memory.

fn clear_algorithms(&self)

https://streams.spec.whatwg.org/#readable-stream-default-controller-clear-algorithms

pub(crate) fn close(&self, cx: &mut JSContext)

https://streams.spec.whatwg.org/#readable-stream-default-controller-close

pub(crate) fn get_desired_size(&self) -> Option<f64>

https://streams.spec.whatwg.org/#readable-stream-default-controller-get-desired-size

pub(crate) fn can_close_or_enqueue(&self) -> bool

https://streams.spec.whatwg.org/#readable-stream-default-controller-can-close-or-enqueue

pub(crate) fn error(&self, cx: &mut JSContext, e: SafeHandleValue<'_>)

https://streams.spec.whatwg.org/#readable-stream-default-controller-error

pub(crate) fn has_backpressure(&self) -> bool

https://streams.spec.whatwg.org/#rs-default-controller-has-backpressure