script::dom::stream

Module transformstream

Structs

BackpressureChangeRejection 🔒

Reacting to rejection of backpressureChangePromise as part of https://streams.spec.whatwg.org/#transform-stream-default-sink-write-algorithm

CancelPromiseFulfillment 🔒

Reacting to fulfillment of the cancelpromise as part of https://streams.spec.whatwg.org/#transform-stream-default-sink-abort-algorithm

CancelPromiseRejection 🔒

Reacting to rejection of cancelpromise as part of https://streams.spec.whatwg.org/#transform-stream-default-sink-abort-algorithm

FlushPromiseFulfillment 🔒

Reacting to fulfillment of the flushpromise as part of https://streams.spec.whatwg.org/#transform-stream-default-sink-close-algorithm

FlushPromiseRejection 🔒

Reacting to rejection of flushpromise as part of https://streams.spec.whatwg.org/#transform-stream-default-sink-close-algorithm

PerformTransformFulfillment 🔒

Reacting to fulfillment of performTransform as part of https://streams.spec.whatwg.org/#transform-stream-default-sink-write-algorithm

PerformTransformRejection 🔒

Reacting to rejection of performTransform as part of https://streams.spec.whatwg.org/#transform-stream-default-sink-write-algorithm

SourceCancelPromiseFulfillment 🔒

Reacting to fulfillment of the cancelpromise as part of https://streams.spec.whatwg.org/#transform-stream-default-source-cancel

SourceCancelPromiseRejection 🔒

Reacting to rejection of cancelpromise as part of https://streams.spec.whatwg.org/#transform-stream-default-source-cancel

TransformBackPressureChangePromiseFulfillment 🔒

Reacting to backpressureChangePromise as part of https://streams.spec.whatwg.org/#transform-stream-default-sink-write-algorithm

TransformStream

https://streams.spec.whatwg.org/#ts-class

Enums

CrossRealmTransform 🔒

A wrapper to handle message and messageerror events for the message port used by the transfered stream.

Constants

_IMPL_DOMOBJECT_FOR_TransformStream 🔒