script::dom::subtlecrypto

Module ecdsa_operation

Constants

P256_PREHASH_LENGTH 🔒

P384_PREHASH_LENGTH 🔒

P521_PREHASH_LENGTH 🔒

Functions

expand_prehash 🔒

A helper function that expand a prehash to a specified length if the prehash is shorter than the specified length.

export_key 🔒

https://w3c.github.io/webcrypto/#ecdsa-operations-export-key

generate_key 🔒

https://w3c.github.io/webcrypto/#ecdsa-operations-generate-key

import_key 🔒

https://w3c.github.io/webcrypto/#ecdsa-operations-import-key

sign 🔒

https://w3c.github.io/webcrypto/#ecdsa-operations-sign

verify 🔒

https://w3c.github.io/webcrypto/#ecdsa-operations-verify