script::dom::subtlecrypto

Enum SupportedAlgorithm

enum SupportedAlgorithm {

    RsassaPkcs1V1_5,
    RsaPss,
    RsaOaep,
    Ecdsa,
    Ecdh,
    Ed25519,
    X25519,
    AesCtr,
    AesCbc,
    AesGcm,
    AesKw,
    Hmac,
    Sha(ShaAlgorithm),
    Hkdf,
    Pbkdf2,
    MlKem(MlKemAlgorithm),
    MlDsa(MlDsaAlgorithm),
    AesOcb,
    ChaCha20Poly1305,
    Sha3(Sha3Algorithm),
    CShake(CShakeAlgorithm),
    Argon2(Argon2Algorithm),

}

https://w3c.github.io/webcrypto/#dfn-supportedAlgorithms

Variants

RsassaPkcs1V1_5

RsaPss

RsaOaep

Ecdsa

Ecdh

Ed25519

X25519

AesCtr

AesCbc

AesGcm

AesKw

Hmac

Sha(ShaAlgorithm)

Hkdf

Pbkdf2

MlKem(MlKemAlgorithm)

MlDsa(MlDsaAlgorithm)

AesOcb

ChaCha20Poly1305

Sha3(Sha3Algorithm)

CShake(CShakeAlgorithm)

Argon2(Argon2Algorithm)

Implementations

impl SupportedAlgorithm

fn as_str(&self) -> &'static str

fn from_ignore_case(alg_name: &str) -> Result<SupportedAlgorithm, Error>

fn support(&self, op: Operation) -> Result<ParameterType, Error>

Check whether the cryptographic algorithm supports the specified operation. If the algorithm supports the operation, then return the desired IDL dictionary type for the operation of the algorithm. Otherwise, throw a NotSupportedError.

This function is also used as the “define an algorithm” algorithm, by adding algorithms, operations and desired IDL dictionary types, to the match block. https://w3c.github.io/webcrypto/#concept-define-an-algorithm