script::dom::subtlecrypto

Module ml_dsa_operation

Structs

Both 🔒

Structure in Rust representing the both SEQUENCE used in the following ASN.1 structures, as defined in [RFC 9881 Section 6].

Enums

MlDsaPrivateKeyStructure 🔒

Structure in Rust representing all the following three structures as defined in [RFC 9881 Section 6].

Constants

ID_ALG_ML_DSA_44 🔒

Object Identifier (OID) of ML-DSA-44 Section 2 of https://datatracker.ietf.org/doc/html/rfc9881

ID_ALG_ML_DSA_65 🔒

Object Identifier (OID) of ML-DSA-65 Section 2 of https://datatracker.ietf.org/doc/html/rfc9881

ID_ALG_ML_DSA_87 🔒

Object Identifier (OID) of ML-DSA-87 Section 2 of https://datatracker.ietf.org/doc/html/rfc9881

Functions

convert_handle_to_public_key 🔒

Convert an ML-DSA public key handle to public key bytes. If the handle is not representing a ML-DSA public key, throw an OperationError.

convert_handle_to_seed_and_public_key 🔒

Convert an ML-DSA private key handle to seed bytes and public key bytes. If the handle is not representing a ML-DSA private key, throw an OperationError.

convert_public_key_to_handle 🔒

Convert public key bytes to an ML-DSA public key handle. If the conversion fails, throw a DataError.

convert_seed_to_handles 🔒

Convert seed bytes to an ML-DSA private key handle and an ML-DSA public key handle. If private key bytes and/or public key bytes are provided, it runs a consistency check against the seed. If the length in bits of seed bytes is not 256, the conversion fails, or the consistency check fails, throw a DataError.

export_key 🔒

https://wicg.github.io/webcrypto-modern-algos/#ml-dsa-operations-export-key

generate_key 🔒

https://wicg.github.io/webcrypto-modern-algos/#ml-dsa-operations-generate-key

import_key 🔒

https://wicg.github.io/webcrypto-modern-algos/#ml-dsa-operations-import-key

sign 🔒

https://wicg.github.io/webcrypto-modern-algos/#ml-dsa-operations-sign

verify 🔒

https://wicg.github.io/webcrypto-modern-algos/#ml-dsa-operations-verify