script::dom::subtlecrypto

Module ml_kem_operation

Structs

Both 🔒

Structure in Rust representing the both SEQUENCE used in the following ASN.1 structures, as defined in [draft-ietf-lamps-kyber-certificates-11 Section 6].

Enums

MlKemPrivateKeyStructure 🔒

Structure in Rust representing all the following three structures as defined in [draft-ietf-lamps-kyber-certificates-11 Section 6].

Constants

ID_ALG_ML_KEM_512 🔒

Object Identifier (OID) of MK-KEM-512 Section 3 of https://datatracker.ietf.org/doc/draft-ietf-lamps-kyber-certificates/

ID_ALG_ML_KEM_768 🔒

Object Identifier (OID) of MK-KEM-768 Section 3 of https://datatracker.ietf.org/doc/draft-ietf-lamps-kyber-certificates/

ID_ALG_ML_KEM_1024 🔒

Object Identifier (OID) of MK-KEM-1024 Section 3 of https://datatracker.ietf.org/doc/draft-ietf-lamps-kyber-certificates/

Functions

convert_handle_to_public_key 🔒

Convert an ML-KEM public key handle to public key bytes. If the handle is not representing a ML-KEM public key, throw an OperationError.

convert_handle_to_seed_and_public_key 🔒

Convert an ML-KEM private key handle to seed bytes and public key bytes. If the handle is not representing a ML-KEM private key, throw an OperationError.

convert_public_key_to_handle 🔒

Convert public key bytes to an ML-KEM public key handle. If the conversion fails, throw a DataError.

convert_seed_to_handles 🔒

Convert seed bytes to an ML-KEM private key handle and an ML-KEM public key handle. If private key bytes and/or public key bytes are provided, it runs a consistency check against the seed. If the length in bits of seed bytes is not 512, the conversion fails, or the consistency check fails, throw a DataError.

decapsulate 🔒

https://wicg.github.io/webcrypto-modern-algos/#ml-kem-operations-decapsulate

encapsulate 🔒

https://wicg.github.io/webcrypto-modern-algos/#ml-kem-operations-encapsulate

export_key 🔒

https://wicg.github.io/webcrypto-modern-algos/#ml-kem-operations-export-key

generate_key 🔒

https://wicg.github.io/webcrypto-modern-algos/#ml-kem-operations-generate-key

import_key 🔒

https://wicg.github.io/webcrypto-modern-algos/#ml-kem-operations-import-key