Function script::dom::virtualmethods::vtable_for

pub fn vtable_for<'a>(node: &'a JSRef<'a, Node>) -> &'a VirtualMethods + 'a

Obtain a VirtualMethods instance for a given Node-derived object. Any method call on the trait object will invoke the corresponding method on the concrete type, propagating up the parent hierarchy unless otherwise interrupted.