script::dom::webcrypto

Module subtlecrypto

Modules

aes_cbc_operation 🔒

aes_common 🔒

aes_ctr_operation 🔒

aes_gcm_operation 🔒

aes_kw_operation 🔒

aes_ocb_operation 🔒

argon2_operation 🔒

chacha20_poly1305_operation 🔒

cshake_operation 🔒

ec_common 🔒

ecdh_operation 🔒

ecdsa_operation 🔒

ed448_operation 🔒

ed25519_operation 🔒

hkdf_operation 🔒

hmac_operation 🔒

kangarootwelve_operation 🔒

kmac_operation 🔒

ml_dsa_operation 🔒

ml_kem_operation 🔒

pbkdf2_operation 🔒

rsa_common 🔒

rsa_oaep_operation 🔒

rsa_pss_operation 🔒

rsassa_pkcs1_v1_5_operation 🔒

sha3_operation 🔒

sha_operation 🔒

turboshake_operation 🔒

x448_operation 🔒

x25519_operation 🔒

Structs

AeadParams 🔒

https://wicg.github.io/webcrypto-modern-algos/#dfn-AeadParams

AesCbcParams 🔒

https://w3c.github.io/webcrypto/#dfn-AesCbcParams

AesCtrParams 🔒

https://w3c.github.io/webcrypto/#dfn-AesCtrParams

AesDerivedKeyParams 🔒

https://w3c.github.io/webcrypto/#dfn-AesDerivedKeyParams

AesGcmParams 🔒

https://w3c.github.io/webcrypto/#dfn-AesGcmParams

AesKeyAlgorithm 🔒

https://w3c.github.io/webcrypto/#dfn-AesKeyAlgorithm

AesKeyGenParams 🔒

https://w3c.github.io/webcrypto/#dfn-AesKeyGenParams

Algorithm 🔒

https://w3c.github.io/webcrypto/#dfn-Algorithm

Argon2Params 🔒

https://wicg.github.io/webcrypto-modern-algos/#dfn-Argon2Params

CShakeParams 🔒

https://wicg.github.io/webcrypto-modern-algos/#dfn-CShakeParams

ContextParams 🔒

https://wicg.github.io/webcrypto-modern-algos/#dfn-ContextParams

DecapsulateOperation 🔒

The value of the key “decapsulate” in the internal object supportedAlgorithms

DecryptOperation 🔒

The value of the key “decrypt” in the internal object supportedAlgorithms

DeriveBitsOperation 🔒

The value of the key “deriveBits” in the internal object supportedAlgorithms

DigestOperation 🔒

The value of the key “digest” in the internal object supportedAlgorithms

EcKeyAlgorithm 🔒

https://w3c.github.io/webcrypto/#dfn-EcKeyAlgorithm

EcKeyGenParams 🔒

https://w3c.github.io/webcrypto/#dfn-EcKeyGenParams

EcKeyImportParams 🔒

https://w3c.github.io/webcrypto/#dfn-EcKeyImportParams

EcdhKeyDeriveParams 🔒

https://w3c.github.io/webcrypto/#dfn-EcdhKeyDeriveParams

EcdsaParams 🔒

https://w3c.github.io/webcrypto/#dfn-EcdsaParams

EncapsulateOperation 🔒

The value of the key “encapsulate” in the internal object supportedAlgorithms

EncapsulatedBits 🔒

https://wicg.github.io/webcrypto-modern-algos/#dfn-EncapsulatedBits

EncapsulatedKey 🔒

https://wicg.github.io/webcrypto-modern-algos/#dfn-EncapsulatedKey

EncryptOperation 🔒

The value of the key “encrypt” in the internal object supportedAlgorithms

ExportKeyOperation 🔒

The value of the key “exportKey” in the internal object supportedAlgorithms

GenerateKeyOperation 🔒

The value of the key “unwrapKey” in the internal object supportedAlgorithms

GetKeyLengthOperation 🔒

The value of the key “get key length” in the internal object supportedAlgorithms

GetPublicKeyOperation 🔒

The value of the key “getPublicKey” in the internal object supportedAlgorithms

HkdfParams 🔒

https://w3c.github.io/webcrypto/#dfn-HkdfParams

HmacImportParams 🔒

https://w3c.github.io/webcrypto/#dfn-HmacImportParams

HmacKeyAlgorithm 🔒

https://w3c.github.io/webcrypto/#dfn-HmacKeyAlgorithm

HmacKeyGenParams 🔒

https://w3c.github.io/webcrypto/#dfn-HmacKeyGenParams

ImportKeyOperation 🔒

The value of the key “importKey” in the internal object supportedAlgorithms

KangarooTwelveParams 🔒

https://wicg.github.io/webcrypto-modern-algos/#dfn-KangarooTwelveParams

KeyAlgorithm 🔒

https://w3c.github.io/webcrypto/#dfn-KeyAlgorithm

KmacImportParams 🔒

https://wicg.github.io/webcrypto-modern-algos/#dfn-KmacImportParams

KmacKeyAlgorithm 🔒

https://wicg.github.io/webcrypto-modern-algos/#dfn-KmacKeyAlgorithm

KmacKeyGenParams 🔒

https://wicg.github.io/webcrypto-modern-algos/#dfn-KmacKeyGenParams

KmacParams 🔒

https://wicg.github.io/webcrypto-modern-algos/#dfn-KmacParams

Pbkdf2Params 🔒

https://w3c.github.io/webcrypto/#dfn-Pbkdf2Params

RsaHashedImportParams 🔒

https://w3c.github.io/webcrypto/#dfn-RsaHashedImportParams

RsaHashedKeyAlgorithm 🔒

https://w3c.github.io/webcrypto/#dfn-RsaHashedKeyAlgorithm

RsaHashedKeyGenParams 🔒

https://w3c.github.io/webcrypto/#dfn-RsaHashedKeyGenParams

RsaOaepParams 🔒

https://w3c.github.io/webcrypto/#dfn-RsaOaepParams

RsaPssParams 🔒

https://w3c.github.io/webcrypto/#dfn-RsaPssParams

SignOperation 🔒

The value of the key “sign” in the internal object supportedAlgorithms

SubtleCrypto 🔒

https://w3c.github.io/webcrypto/#subtlecrypto-interface

SubtleEd448Params 🔒

https://wicg.github.io/webcrypto-secure-curves/#dfn-Ed448Params

TurboShakeParams 🔒

https://wicg.github.io/webcrypto-modern-algos/#dfn-TurboShakeParams

UnwrapKeyOperation 🔒

The value of the key “unwrapKey” in the internal object supportedAlgorithms

VerifyOperation 🔒

The value of the key “verify” in the internal object supportedAlgorithms

WrapKeyOperation 🔒

The value of the key “wrapKey” in the internal object supportedAlgorithms

Enums

CryptoAlgorithm 🔒

DecapsulateAlgorithm 🔒

Normalized algorithm for the “decapsulate” operation, used as output of https://w3c.github.io/webcrypto/#dfn-normalize-an-algorithm

DecryptAlgorithm 🔒

Normalized algorithm for the “decrypt” operation, used as output of https://w3c.github.io/webcrypto/#dfn-normalize-an-algorithm

DeriveBitsAlgorithm 🔒

Normalized algorithm for the “deriveBits” operation, used as output of https://w3c.github.io/webcrypto/#dfn-normalize-an-algorithm

DigestAlgorithm 🔒

Normalized algorithm for the “digest” operation, used as output of https://w3c.github.io/webcrypto/#dfn-normalize-an-algorithm

EncapsulateAlgorithm 🔒

Normalized algorithm for the “encapsulate” operation, used as output of https://w3c.github.io/webcrypto/#dfn-normalize-an-algorithm

EncryptAlgorithm 🔒

Normalized algorithm for the “encrypt” operation, used as output of https://w3c.github.io/webcrypto/#dfn-normalize-an-algorithm

ExportKeyAlgorithm 🔒

Normalized algorithm for the “exportKey” operation, used as output of https://w3c.github.io/webcrypto/#dfn-normalize-an-algorithm

ExportedKey 🔒

The returned type of the successful export key operation. Bytes should be used when the key is exported in “raw”, “spki” or “pkcs8” format. Jwk should be used when the key is exported in “jwk” format.

GenerateKeyAlgorithm 🔒

Normalized algorithm for the “generateKey” operation, used as output of https://w3c.github.io/webcrypto/#dfn-normalize-an-algorithm

GetKeyLengthAlgorithm 🔒

Normalized algorithm for the “get key length” operation, used as output of https://w3c.github.io/webcrypto/#dfn-normalize-an-algorithm

GetPublicKeyAlgorithm 🔒

Normalized algorithm for the “getPublicKey” operation, used as output of https://w3c.github.io/webcrypto/#dfn-normalize-an-algorithm

ImportKeyAlgorithm 🔒

Normalized algorithm for the “importKey” operation, used as output of https://w3c.github.io/webcrypto/#dfn-normalize-an-algorithm

JwkStringField 🔒

KeyAlgorithmAndDerivatives 🔒

Union type of KeyAlgorithm and IDL dictionary types derived from it. Note that we actually use our “subtle” structs of the corresponding IDL dictionary types so that they can be easily passed to another threads.

SignAlgorithm 🔒

Normalized algorithm for the “sign” operation, used as output of https://w3c.github.io/webcrypto/#dfn-normalize-an-algorithm

UnwrapKeyAlgorithm 🔒

Normalized algorithm for the “unwrapKey” operation, used as output of https://w3c.github.io/webcrypto/#dfn-normalize-an-algorithm

VerifyAlgorithm 🔒

Normalized algorithm for the “verify” operation, used as output of https://w3c.github.io/webcrypto/#dfn-normalize-an-algorithm

WrapKeyAlgorithm 🔒

Normalized algorithm for the “wrapKey” operation, used as output of https://w3c.github.io/webcrypto/#dfn-normalize-an-algorithm

Constants

NAMED_CURVE_P256 🔒

NAMED_CURVE_P384 🔒

NAMED_CURVE_P521 🔒

_IMPL_DOMOBJECT_FOR_SubtleCrypto 🔒

Statics

SUPPORTED_CURVES 🔒

Traits

JsonWebKeyExt 🔒

NormalizedAlgorithm 🔒

Operation 🔒

TryFromWithCxAndName 🔒

Alternative to std::convert::TryFrom, with &mut js::context::JSContext

TryIntoWithCxAndName 🔒

Alternative to std::convert::TryInto, with &mut js::context::JSContext

Functions

check_support_for_algorithm 🔒

https://wicg.github.io/webcrypto-modern-algos/#dfn-check-support-for-algorithm

get_optional_buffer_source 🔒

Helper to retrieve an optional paramter in BufferSource from WebIDL dictionary, and get a copy of the bytes held by the buffer source according to https://webidl.spec.whatwg.org/#dfn-get-buffer-source-copy

get_required_buffer_source 🔒

Helper to retrieve a required paramter in BufferSource from WebIDL dictionary, and get a copy of the bytes held by the buffer source according to https://webidl.spec.whatwg.org/#dfn-get-buffer-source-copy

get_required_parameter 🔒

Helper to retrieve a required paramter from WebIDL dictionary.

get_required_parameter_in_box 🔒

Helper to retrieve a required paramter, in RootedTraceableBox, from WebIDL dictionary.

normalize_algorithm 🔒

https://w3c.github.io/webcrypto/#algorithm-normalization-normalize-an-algorithm