script::dom::webgl::webgltexture

Struct WebGLTexture

#[repr(C)]
pub(crate) struct WebGLTexture {
    webgl_object: WebGLObject,
    target: Cell<Option<u32>>,
    image_info_array: DomRefCell<[Option<ImageInfo>; 186]>,
    face_count: Cell<u8>,
    base_mipmap_level: u32,
    min_filter: Cell<u32>,
    mag_filter: Cell<u32>,
    attached_framebuffer: MutNullableDom<WebGLFramebuffer>,
    immutable_levels: Cell<Option<u32>>,
    droppable: DroppableWebGLTexture,
}

Fields
webgl_object: WebGLObjecttarget: Cell<Option<u32>>

The target to which this texture was bound the first time

image_info_array: DomRefCell<[Option<ImageInfo>; 186]>

Stores information about mipmap levels and cubemap faces.

face_count: Cell<u8>

Face count can only be 1 or 6

base_mipmap_level: u32min_filter: Cell<u32>mag_filter: Cell<u32>attached_framebuffer: MutNullableDom<WebGLFramebuffer>

Framebuffer that this texture is attached to.

immutable_levels: Cell<Option<u32>>

Number of immutable levels.

droppable: DroppableWebGLTexture
Implementations

impl WebGLTexture

fn __assert_parent_type(&self)

impl WebGLTexture

fn new_inherited( context: &WebGLRenderingContext, id: WebGLTextureId, owner: Option<&XRSession>, ) -> Self

pub(crate) fn maybe_new( cx: &mut JSContext, context: &WebGLRenderingContext, ) -> Option<DomRoot<Self>>

pub(crate) fn new( cx: &mut JSContext, context: &WebGLRenderingContext, id: WebGLTextureId, ) -> DomRoot<Self>

pub(crate) fn new_webxr( cx: &mut JSContext, context: &WebGLRenderingContext, id: WebGLTextureId, session: &XRSession, ) -> DomRoot<Self>

impl WebGLTexture

pub(crate) fn id(&self) -> WebGLTextureId

pub(crate) fn bind(&self, target: u32) -> WebGLResult<()>

pub(crate) fn initialize( &self, target: TexImageTarget, width: u32, height: u32, depth: u32, internal_format: TexFormat, level: u32, data_type: Option<TexDataType>, ) -> WebGLResult<()>

pub(crate) fn generate_mipmap(&self) -> WebGLResult<()>

pub(crate) fn delete(&self, operation_fallibility: Operation)

pub(crate) fn is_invalid(&self) -> bool

pub(crate) fn is_immutable(&self) -> bool

pub(crate) fn target(&self) -> Option<u32>

pub(crate) fn maybe_get_tex_parameter( &self, param: TexParameter, ) -> Option<TexParameterValue>

pub(crate) fn tex_parameter( &self, param: u32, value: TexParameterValue, ) -> WebGLResult<()>

We have to follow the conversion rules for GLES 2.0. See: https://www.khronos.org/webgl/public-mailing-list/archives/1008/msg00014.html

pub(crate) fn min_filter(&self) -> u32

pub(crate) fn mag_filter(&self) -> u32

pub(crate) fn is_using_linear_filtering(&self) -> bool

pub(crate) fn populate_mip_chain( &self, first_level: u32, last_level: u32, ) -> WebGLResult<()>

fn is_cube_complete(&self) -> bool

fn face_index_for_target(&self, target: &TexImageTarget) -> u8

pub(crate) fn image_info_for_target( &self, target: &TexImageTarget, level: u32, ) -> Option<ImageInfo>

pub(crate) fn image_info_at_face( &self, face: u8, level: u32, ) -> Option<ImageInfo>

fn set_image_infos_at_level(&self, level: u32, image_info: ImageInfo)

fn set_image_infos_at_level_and_face( &self, level: u32, face: u8, image_info: ImageInfo, )

fn update_size(&self)

fn base_image_info(&self) -> Option<ImageInfo>

pub(crate) fn attach_to_framebuffer(&self, fb: &WebGLFramebuffer)

pub(crate) fn detach_from_framebuffer(&self)

pub(crate) fn storage( &self, target: TexImageTarget, levels: u32, internal_format: TexFormat, width: u32, height: u32, depth: u32, ) -> WebGLResult<()>