Struct script::dom::webgltexture::WebGLTexture

pub struct WebGLTexture {
    webgl_object: WebGLObject,
    id: u32,
    target: Cell<Option<u32>>,
    is_deleted: Cell<bool>,
    image_info_array: DOMRefCell<[ImageInfo; MAX_LEVEL_COUNT * MAX_FACE_COUNT]>,
    face_count: Cell<u8>,
    base_mipmap_level: u32,
    renderer: IpcSender<CanvasMsg>,
}

Fields

webgl_object

id

target

The target to which this texture was bound the first time

is_deleted

image_info_array

Stores information about mipmap levels and cubemap faces.

face_count

Face count can only be 1 or 6

base_mipmap_level

renderer

Methods

impl WebGLTexture

fn new_inherited(renderer: IpcSender<CanvasMsg>, id: u32) -> WebGLTexture

fn maybe_new(global: GlobalRef, renderer: IpcSender<CanvasMsg>) -> Option<Root<WebGLTexture>>

fn new(global: GlobalRef, renderer: IpcSender<CanvasMsg>, id: u32) -> Root<WebGLTexture>

impl WebGLTexture

fn id(&self) -> u32

fn bind(&self, target: u32) -> WebGLResult<()>

fn initialize(&self, target: u32, width: u32, height: u32, depth: u32, internal_format: u32, level: u32) -> WebGLResult<()>

fn generate_mipmap(&self) -> WebGLResult<()>

fn delete(&self)

fn is_deleted(&self) -> bool

fn target(&self) -> Option<u32>

fn tex_parameter(&self, target: u32, name: u32, value: TexParameterValue) -> WebGLResult<()>

We have to follow the conversion rules for GLES 2.0. See: https://www.khronos.org/webgl/public-mailing-list/archives/1008/msg00014.html

fn populate_mip_chain(&self, first_level: u32, last_level: u32) -> WebGLResult<()>

fn is_cube_complete(&self) -> bool

fn image_info_at_face(&self, face: u8, level: u32) -> ImageInfo

fn set_image_infos_at_level(&self, level: u32, image_info: ImageInfo)

fn set_image_infos_at_level_and_face(&self, level: u32, face: u8, image_info: ImageInfo)

fn base_image_info(&self) -> Option<ImageInfo>