script::dom::window

Module window

Structs

CSSErrorReporter 🔒

FontCspHandler 🔒

FontFetchListener 🔒

FontNetworkTimingHandler 🔒

LayoutValue 🔒

An instance of a value associated with a particular snapshot of layout. This stored value can only be read as long as the associated layout marker that is considered valid. It will automatically become unavailable when the next layout operation is performed.

OngoingNavigation 🔒

An id used to cancel navigations; for now only used for planned form navigations. Loosely based on https://html.spec.whatwg.org/multipage/#ongoing-navigation.

PendingImageCallback

A callback to call when a response comes back from the ImageCache.

PendingLayoutImageAncillaryData 🔒

Ancillary data of pending image request that was initiated by layout during a reflow. This data is used to faciliate invalidating layout when the image data becomes available at some point in the future.

Window 🔒

Enums

LayoutBlocker 🔒

During loading and parsing, layouts are suppressed to avoid flashing incomplete page contents.

WindowState 🔒

Current state of the window object

Constants

INITIAL_REFLOW_DELAY 🔒

How long we should wait before performing the initial reflow after <body> is parsed, assuming that <body> take this long to parse.

_IMPL_DOMOBJECT_FOR_Window 🔒

Functions

base64_atob 🔒

base64_btoa 🔒

dump_js_stack 🔒 ⚠

Helper for interactive debugging sessions in lldb/gdb.

is_named_element_with_id_attribute 🔒

is_named_element_with_name_attribute 🔒

Type Aliases

PendingImageRasterizationKey 🔒