script::dom::window::window

Struct Window

#[repr(C)]
pub(crate) struct Window {

    globalscope: GlobalScope,
    weak_script_thread: Weak<ScriptThread>,
    webview_id: WebViewId,
    script_chan: Sender<MainThreadScriptMsg>,
    layout: RefCell<Box<dyn Layout>>,
    navigator: MutNullableDom<Navigator>,
    crypto: MutNullableDom<Crypto>,
    image_cache_sender: Sender<ImageCacheResponseMessage>,
    window_proxy: MutNullableDom<WindowProxy>,
    document: MutNullableDom<Document>,
    location: MutNullableDom<Location>,
    performance: MutNullableDom<Performance>,
    navigation_start: Cell<CrossProcessInstant>,
    screen: MutNullableDom<Screen>,
    session_storage: MutNullableDom<Storage>,
    local_storage: MutNullableDom<Storage>,
    cookie_store: MutNullableDom<CookieStore>,
    status: DomRefCell<DOMString>,
    trusted_types: MutNullableDom<TrustedTypePolicyFactory>,
    ongoing_navigation: Cell<OngoingNavigation>,
    caches: MutNullableDom<CacheStorage>,
    devtools_markers: DomRefCell<HashSet<TimelineMarkerType>>,
    devtools_marker_sender: DomRefCell<Option<GenericSender<Option<TimelineMarker>>>>,
    unhandled_resize_event: DomRefCell<Option<(ViewportDetails, WindowSizeType)>>,
    viewport_details_at_last_resize_steps: Cell<ViewportDetails>,
    embedder_theme: Cell<Theme>,
    parent_info: Option<PipelineId>,
    dom_static: GlobalStaticData,
    js_runtime: DomRefCell<Option<Rc<Runtime>>>,
    viewport_details: Cell<ViewportDetails>,
    bluetooth_thread: GenericSender<BluetoothRequest>,
    bluetooth_extra_permission_data: BluetoothExtraPermissionData,
    layout_blocker: Cell<LayoutBlocker>,
    webdriver_load_status_sender: RefCell<Option<GenericSender<WebDriverLoadStatus>>>,
    current_state: Cell<WindowState>,
    error_reporter: CSSErrorReporter,
    media_query_lists: DOMTracker<MediaQueryList>,
    test_runner: MutNullableDom<TestRunner>,
    webgl_chan: Option<WebGLChan>,
    webxr_registry: Option<Registry>,
    pending_image_callbacks: DomRefCell<FxHashMap<PendingImageId, Vec<PendingImageCallback>>>,
    pending_layout_images: DomRefCell<HashMapTracedValues<PendingImageId, Vec<PendingLayoutImageAncillaryData>, FxBuildHasher>>,
    pending_images_for_rasterization: DomRefCell<HashMapTracedValues<(PendingImageId, DeviceIntSize), Vec<Dom<Node>>, FxBuildHasher>>,
    unminified_css_dir: DomRefCell<Option<String>>,
    local_script_source: Option<String>,
    test_worklet: MutNullableDom<Worklet>,
    paint_worklet: MutNullableDom<Worklet>,
    exists_mut_observer: Cell<bool>,
    paint_api: CrossProcessPaintApi,
    user_scripts: Rc<Vec<UserScript>>,
    player_context: WindowGLContext,
    throttled: Cell<bool>,
    layout_marker: DomRefCell<Rc<Cell<bool>>>,
    current_event: DomRefCell<Option<Dom<Event>>>,
    reporting_observer_list: DomRefCell<Vec<Dom<ReportingObserver>>>,
    report_list: DomRefCell<Vec<Report>>,
    endpoints_list: DomRefCell<Vec<ReportingEndpoint>>,
    script_window_proxies: Rc<ScriptWindowProxies>,
    has_pending_screenshot_readiness_request: Cell<bool>,
    visual_viewport: MutNullableDom<VisualViewport>,
    has_changed_visual_viewport_dimension: Cell<bool>,
    pending_media_query_evaluation: Cell<bool>,
    last_activation_timestamp: Cell<UserActivationTimestamp>,
    devtools_wants_updates: Cell<bool>,
    has_dispatched_scroll_event: Cell<bool>,
    has_dispatched_input_event: Cell<bool>,

}

Fields
globalscope: GlobalScopeweak_script_thread: Weak<ScriptThread>

A Weak reference to this ScriptThread used to give to child Windows so they can more easily call methods on the ScriptThread without constantly having to pass it everywhere.

webview_id: WebViewId

The webview that contains this Window.

This may not be the top-level Window, in the case of frames.

script_chan: Sender<MainThreadScriptMsg>layout: RefCell<Box<dyn Layout>>navigator: MutNullableDom<Navigator>crypto: MutNullableDom<Crypto>image_cache_sender: Sender<ImageCacheResponseMessage>window_proxy: MutNullableDom<WindowProxy>document: MutNullableDom<Document>location: MutNullableDom<Location>performance: MutNullableDom<Performance>navigation_start: Cell<CrossProcessInstant>screen: MutNullableDom<Screen>session_storage: MutNullableDom<Storage>local_storage: MutNullableDom<Storage>cookie_store: MutNullableDom<CookieStore>

https://cookiestore.spec.whatwg.org/#globals

status: DomRefCell<DOMString>trusted_types: MutNullableDom<TrustedTypePolicyFactory>ongoing_navigation: Cell<OngoingNavigation>

The start of something resembling https://html.spec.whatwg.org/multipage/#ongoing-navigation

caches: MutNullableDom<CacheStorage>

https://w3c.github.io/ServiceWorker/#global-caches-attribute

devtools_markers: DomRefCell<HashSet<TimelineMarkerType>>

For sending timeline markers. Will be ignored if no devtools server

devtools_marker_sender: DomRefCell<Option<GenericSender<Option<TimelineMarker>>>>unhandled_resize_event: DomRefCell<Option<(ViewportDetails, WindowSizeType)>>

Most recent unhandled resize event, if any.

viewport_details_at_last_resize_steps: Cell<ViewportDetails>

The viewport at the time of the last “run the resize steps”.

This allows us to detect ABA changes, and suppress firing the event in that case.

embedder_theme: Cell<Theme>

Platform theme.

parent_info: Option<PipelineId>

Parent id associated with this page, if any.

dom_static: GlobalStaticData

Global static data related to the DOM.

js_runtime: DomRefCell<Option<Rc<Runtime>>>

The JavaScript runtime.

viewport_details: Cell<ViewportDetails>

The ViewportDetails of this Window’s frame.

bluetooth_thread: GenericSender<BluetoothRequest>

A handle for communicating messages to the bluetooth thread.

bluetooth_extra_permission_data: BluetoothExtraPermissionDatalayout_blocker: Cell<LayoutBlocker>

See the documentation for LayoutBlocker. Essentially, this flag prevents layouts from happening before the first load event, apart from a few exceptional cases.

webdriver_load_status_sender: RefCell<Option<GenericSender<WebDriverLoadStatus>>>

A channel to notify webdriver if there is a navigation

current_state: Cell<WindowState>

The current state of the window object

error_reporter: CSSErrorReportermedia_query_lists: DOMTracker<MediaQueryList>

All the MediaQueryLists we need to update

test_runner: MutNullableDom<TestRunner>webgl_chan: Option<WebGLChan>

A handle for communicating messages to the WebGL thread, if available.

webxr_registry: Option<Registry>pending_image_callbacks: DomRefCell<FxHashMap<PendingImageId, Vec<PendingImageCallback>>>

When an element triggers an image load or starts watching an image load from the ImageCache it adds an entry to this list. When those loads are triggered from layout, they also add an etry to Self::pending_layout_images.

pending_layout_images: DomRefCell<HashMapTracedValues<PendingImageId, Vec<PendingLayoutImageAncillaryData>, FxBuildHasher>>

All of the elements that have an outstanding image request that was initiated by layout during a reflow. They are stored in the ScriptThread to ensure that the element can be marked dirty when the image data becomes available at some point in the future.

pending_images_for_rasterization: DomRefCell<HashMapTracedValues<(PendingImageId, DeviceIntSize), Vec<Dom<Node>>, FxBuildHasher>>

Vector images for which layout has intiated rasterization at a specific size and whose results are not yet available. They are stored in the ScriptThread so that the element can be marked dirty once the rasterization is completed.

unminified_css_dir: DomRefCell<Option<String>>

Directory to store unminified css for this window if unminify-css opt is enabled.

local_script_source: Option<String>

Directory with stored unminified scripts

test_worklet: MutNullableDom<Worklet>

Worklets

paint_worklet: MutNullableDom<Worklet>

https://drafts.css-houdini.org/css-paint-api-1/#paint-worklet

exists_mut_observer: Cell<bool>

Flag to identify whether mutation observers are present(true)/absent(false)

paint_api: CrossProcessPaintApi

Cross-process access to Paint.

user_scripts: Rc<Vec<UserScript>>

The UserScripts added via UserContentManager. These are potentially shared with other WebViews in this ScriptThread.

player_context: WindowGLContext

Window’s GL context from application

throttled: Cell<bool>layout_marker: DomRefCell<Rc<Cell<bool>>>

A shared marker for the validity of any cached layout values. A value of true indicates that any such values remain valid; any new layout that invalidates those values will cause the marker to be set to false.

current_event: DomRefCell<Option<Dom<Event>>>

https://dom.spec.whatwg.org/#window-current-event

reporting_observer_list: DomRefCell<Vec<Dom<ReportingObserver>>>

https://w3c.github.io/reporting/#windoworworkerglobalscope-registered-reporting-observer-list

report_list: DomRefCell<Vec<Report>>

https://w3c.github.io/reporting/#windoworworkerglobalscope-reports

endpoints_list: DomRefCell<Vec<ReportingEndpoint>>

https://w3c.github.io/reporting/#windoworworkerglobalscope-endpoints

script_window_proxies: Rc<ScriptWindowProxies>

The window proxies the script thread knows.

has_pending_screenshot_readiness_request: Cell<bool>

Whether or not this Window has a pending screenshot readiness request.

visual_viewport: MutNullableDom<VisualViewport>

Visual viewport interface that is associated to this Window. https://drafts.csswg.org/cssom-view/#dom-window-visualviewport

has_changed_visual_viewport_dimension: Cell<bool>

VisualViewport dimension changed and we need to process it on the next tick.

pending_media_query_evaluation: Cell<bool>

Whether something has changed since the last “update the rendering” turn that may affect media query results, like a theme change. Consumed together with the resized signal to decide whether to re-evaluate MediaQueryLists and dispatch change events.

last_activation_timestamp: Cell<UserActivationTimestamp>

https://html.spec.whatwg.org/multipage/#last-activation-timestamp

devtools_wants_updates: Cell<bool>

A flag to indicate whether the developer tools has requested live updates from the window.

has_dispatched_scroll_event: Cell<bool>

https://www.w3.org/TR/largest-contentful-paint/#has-dispatched-scroll-event

has_dispatched_input_event: Cell<bool>

https://wicg.github.io/event-timing/#has-dispatched-input-event

Implementations

impl Window

fn __assert_parent_type(&self)

impl Window

pub(crate) fn script_thread(&self) -> Rc<ScriptThread>

pub(crate) fn webview_id(&self) -> WebViewId

pub(crate) fn as_global_scope(&self) -> &GlobalScope

pub(crate) fn mark_has_dispatched_scroll_event(&self)

https://www.w3.org/TR/largest-contentful-paint/#has-dispatched-scroll-event

pub(crate) fn mark_has_dispatched_input_event(&self)

https://wicg.github.io/event-timing/#has-dispatched-input-event

pub(crate) fn layout(&self) -> Ref<'_, Box<dyn Layout>>

pub(crate) fn layout_mut(&self) -> RefMut<'_, Box<dyn Layout>>

pub(crate) fn get_exists_mut_observer(&self) -> bool

pub(crate) fn set_exists_mut_observer(&self)

pub(crate) fn clear_js_runtime_for_script_deallocation(&self)

pub(crate) fn discard_browsing_context(&self)

A convenience method for https://html.spec.whatwg.org/multipage/#a-browsing-context-is-discarded

pub(crate) fn time_profiler_chan(&self) -> &TimeProfilerChan

Get a sender to the time profiler thread.

pub(crate) fn origin(&self) -> MutableOrigin

https://html.spec.whatwg.org/multipage/#script-settings-for-window-objects:concept-settings-object-origin

pub(crate) fn main_thread_script_chan(&self) -> &Sender<MainThreadScriptMsg>

pub(crate) fn parent_info(&self) -> Option<PipelineId>

pub(crate) fn new_script_pair( &self, ) -> (ScriptEventLoopSender, ScriptEventLoopReceiver)

pub(crate) fn event_loop_sender(&self) -> ScriptEventLoopSender

pub(crate) fn image_cache(&self) -> Arc<dyn ImageCache>

pub(crate) fn window_proxy(&self) -> DomRoot<WindowProxy>

This can panic if it is called after the browsing context has been discarded

pub(crate) fn append_reporting_observer( &self, reporting_observer: &ReportingObserver, )

pub(crate) fn remove_reporting_observer( &self, reporting_observer: &ReportingObserver, )

pub(crate) fn registered_reporting_observers( &self, ) -> Vec<DomRoot<ReportingObserver>>

pub(crate) fn append_report(&self, report: Report)

pub(crate) fn buffered_reports(&self) -> Vec<Report>

pub(crate) fn set_endpoints_list(&self, endpoints: Vec<ReportingEndpoint>)

pub(crate) fn undiscarded_window_proxy(&self) -> Option<DomRoot<WindowProxy>>

Returns the window proxy if it has not been discarded. https://html.spec.whatwg.org/multipage/#a-browsing-context-is-discarded

pub(crate) fn top_level_document_if_local(&self) -> Option<DomRoot<Document>>

Get the active Document of top-level browsing context, or return Window’s Document if it’s browing context is the top-level browsing context. Returning none if the WindowProxy is discarded or the Document is in another ScriptThread. https://html.spec.whatwg.org/multipage/#top-level-browsing-context

pub(crate) fn bluetooth_thread(&self) -> GenericSender<BluetoothRequest>

pub(crate) fn bluetooth_extra_permission_data( &self, ) -> &BluetoothExtraPermissionData

pub(crate) fn css_error_reporter(&self) -> &CSSErrorReporter

pub(crate) fn webgl_chan(&self) -> Option<WebGLChan>

pub(crate) fn webgl_chan_value(&self) -> Option<WebGLChan>

pub(crate) fn webxr_registry(&self) -> Option<Registry>

fn new_paint_worklet(&self, cx: &mut JSContext) -> DomRoot<Worklet>

pub(crate) fn register_image_cache_listener( &self, id: PendingImageId, callback: impl Fn(PendingImageResponse, &mut JSContext) + 'static, ) -> ImageCacheResponseCallback

fn pending_layout_image_notification( &self, no_gc: &NoGC, response: PendingImageResponse, )

pub(crate) fn handle_image_rasterization_complete_notification( &self, no_gc: &NoGC, response: RasterizationCompleteResponse, )

pub(crate) fn pending_image_notification( &self, response: PendingImageResponse, cx: &mut JSContext, )

pub(crate) fn paint_api(&self) -> &CrossProcessPaintApi

pub(crate) fn userscripts(&self) -> &[UserScript]

pub(crate) fn get_player_context(&self) -> WindowGLContext

pub(crate) fn dispatch_event_with_target_override( &self, cx: &mut JSContext, event: &Event, )

pub(crate) fn font_context(&self) -> Arc<FontContext>

pub(crate) fn ongoing_navigation(&self) -> OngoingNavigation

pub(crate) fn set_ongoing_navigation(&self) -> OngoingNavigation

https://html.spec.whatwg.org/multipage/#set-the-ongoing-navigation

fn stop_loading(&self, cx: &mut JSContext)

https://html.spec.whatwg.org/multipage/#nav-stop

fn destroy_top_level_traversable(&self, cx: &mut JSContext)

https://html.spec.whatwg.org/multipage/#destroy-a-top-level-traversable

fn definitely_close(&self, cx: &mut JSContext)

https://html.spec.whatwg.org/multipage/#definitely-close-a-top-level-traversable

fn cannot_show_simple_dialogs(&self) -> bool

https://html.spec.whatwg.org/multipage/#cannot-show-simple-dialogs

pub(crate) fn perform_a_microtask_checkpoint(&self, cx: &mut JSContext)

pub(crate) fn web_font_context(&self, no_gc: &NoGC) -> WebFontDocumentContext

pub(crate) fn request_client(&self, no_gc: Option<&NoGC>) -> RequestClient

Part of https://fetch.spec.whatwg.org/#populate-request-from-client

pub(crate) fn gc(&self, cx: &mut JSContext)

pub(crate) fn with_timers<T>(&self, f: impl FnOnce(&OneshotTimers) -> T) -> T

impl Window

pub(crate) fn scroll_offset(&self) -> Vector2D<f32, LayoutPixel>

pub(crate) fn create_named_properties_object( cx: &mut JSContext, proto: HandleObject<'_>, object: MutableHandleObject<'_>, )

pub(crate) fn current_event(&self) -> Option<DomRoot<Event>>

pub(crate) fn set_current_event( &self, event: Option<&Event>, ) -> Option<DomRoot<Event>>

fn post_message_impl( &self, target_origin: &USVString, source_origin: ImmutableOrigin, source: &Window, cx: &mut JSContext, message: HandleValue<'_>, transfer: CustomAutoRooterGuard<'_, Vec<*mut JSObject>>, ) -> ErrorResult

https://html.spec.whatwg.org/multipage/#window-post-message-steps

pub(crate) fn paint_worklet(&self, cx: &mut JSContext) -> DomRoot<Worklet>

pub(crate) fn clear_js_runtime(&self)

pub(crate) fn scroll( &self, cx: &mut JSContext, x: f32, y: f32, behavior: ScrollBehavior, )

https://drafts.csswg.org/cssom-view/#dom-window-scroll

pub(crate) fn perform_a_scroll( &self, cx: &mut JSContext, x: f32, y: f32, scroll_id: ExternalScrollId, _behavior: ScrollBehavior, element: Option<&Element>, )

https://drafts.csswg.org/cssom-view/#perform-a-scroll

pub(crate) fn device_pixel_ratio(&self) -> Scale<f32, CSSPixel, DevicePixel>

fn client_window(&self) -> DeviceIndependentIntRect

pub(crate) fn advance_animation_clock(&self, no_gc: &NoGC, delta: TimeDuration)

Prepares to tick animations and then does a reflow which also advances the layout animation clock.

pub(crate) fn reflow( &self, cx: &mut JSContext, reflow_goal: ReflowGoal, ) -> (ReflowPhasesRun, ReflowStatistics)

Reflows the page unconditionally if possible and not suppressed. This method will wait for the layout to complete. If there is no window size yet, the page is presumed invisible and no reflow is performed. If reflow is suppressed, no reflow will be performed for ForDisplay goals.

NOTE: This method should almost never be called directly! Layout and rendering updates should happen as part of the HTML event loop via update the rendering.

pub(crate) fn request_screenshot_readiness(&self, cx: &mut JSContext)

pub(crate) fn maybe_resolve_pending_screenshot_readiness_requests( &self, cx: &mut JSContext, )

pub(crate) fn reflow_if_reflow_timer_expired(&self, cx: &mut JSContext)

If parsing has taken a long time and reflows are still waiting for the load event, start allowing them. See https://github.com/servo/servo/pull/6028.

pub(crate) fn prevent_layout_until_load_event(&self)

Block layout for this Window until parsing is done. If parsing takes a long time, we want to layout anyway, so schedule a moment in the future for when layouts are allowed even though parsing isn’t finished and we havne’t sent a load event.

pub(crate) fn allow_layout_if_necessary(&self, cx: &mut JSContext)

Inform the Window that layout is allowed either because load has happened or because parsing the <body> took so long that we cannot wait any longer.

pub(crate) fn layout_blocked(&self) -> bool

fn flush_ancestor_layouts_if_necessary(&self, cx: &mut JSContext)

pub(crate) fn layout_reflow(&self, query_msg: QueryMsg)

Trigger a reflow that is required by a certain queries.

pub(crate) fn reflow_for_non_flushing_update_the_rendering_queries( &self, cx: &mut JSContext, )

Trigger a reflow in preparation for subsequent queries that don’t perform a reflow.

pub(crate) fn resolved_font_style_query( &self, node: &Node, value: String, ) -> Option<ServoArc<Font>>

pub(crate) fn containing_block_node_query_without_reflow( &self, node: &Node, ) -> Option<DomRoot<Node>>

Query the ancestor node that establishes the containing block for the given node. https://drafts.csswg.org/css-position-3/#def-cb

pub(crate) fn is_containing_block_descendant_query_without_reflow( &self, possible_ancestor: &Node, possible_descendant: &Node, ) -> bool

Query whether a node is part of another node’s containing block chain. https://drafts.csswg.org/css-display/#containing-block-chain

pub(crate) fn padding_query_without_reflow( &self, node: &Node, ) -> Option<PhysicalSides>

Query the used padding values for the given node, but do not force a reflow. This is used for things like ResizeObserver which should observe the value from the most recent reflow, but do not need it to reflect the current state of the DOM / style.

pub(crate) fn box_area_query_without_reflow( &self, node: &Node, area: BoxAreaType, exclude_transform_and_inline: bool, ) -> Option<Rect<Au, CSSPixel>>

Do the same kind of query as Self::box_area_query, but do not force a reflow. This is used for things like IntersectionObserver which should observe the value from the most recent reflow, but do not need it to reflect the current state of the DOM / style.

pub(crate) fn box_area_query( &self, node: &Node, area: BoxAreaType, exclude_transform_and_inline: bool, ) -> Option<Rect<Au, CSSPixel>>

pub(crate) fn box_areas_query( &self, node: &Node, area: BoxAreaType, ) -> CSSPixelRectVec

pub(crate) fn client_rect_query(&self, node: &Node) -> Rect<i32, CSSPixel>

pub(crate) fn current_css_zoom_query(&self, node: &Node) -> f32

pub(crate) fn document_unrooted<'a>( &self, no_gc: &'a NoGC, ) -> UnrootedDom<'a, Document>

https://html.spec.whatwg.org/multipage/#dom-document-2

pub(crate) fn scrolling_area_query( &self, node: Option<&Node>, ) -> Rect<i32, CSSPixel>

Find the scroll area of the given node, if it is not None. If the node is None, find the scroll area of the viewport.

pub(crate) fn scroll_offset_query( &self, node: &Node, ) -> Vector2D<f32, LayoutPixel>

fn scroll_offset_query_with_external_scroll_id( &self, external_scroll_id: ExternalScrollId, ) -> Vector2D<f32, LayoutPixel>

fn scroll_offset_query_with_external_scroll_id_no_reflow( &self, external_scroll_id: ExternalScrollId, ) -> Vector2D<f32, LayoutPixel>

pub(crate) fn scroll_an_element( &self, cx: &mut JSContext, element: &Element, x: f32, y: f32, behavior: ScrollBehavior, )

https://drafts.csswg.org/cssom-view/#scroll-an-element

pub(crate) fn resolved_style_query( &self, element: TrustedNodeAddress, pseudo: Option<PseudoElement>, property: PropertyId, ) -> DOMString

pub(crate) fn get_iframe_viewport_details_if_known( &self, browsing_context_id: BrowsingContextId, ) -> Option<ViewportDetails>

If the given |browsing_context_id| refers to an <iframe> that is an element in this Window and that <iframe> has been laid out, return its size. Otherwise, return None.

pub(crate) fn offset_parent_query( &self, node: &Node, ) -> (Option<DomRoot<Element>>, Rect<Au, CSSPixel>)

pub(crate) fn scroll_container_query( &self, node: Option<&Node>, flags: ScrollContainerQueryFlags, ) -> Option<ScrollContainerResponse>

pub(crate) fn scrolling_box_query( &self, node: Option<&Node>, flags: ScrollContainerQueryFlags, ) -> Option<ScrollingBox>

pub(crate) fn text_index_query_on_node_for_event( &self, node: &Node, point_in_viewport: Point2D<Au, CSSPixel>, ) -> Option<(DomRoot<Node>, Utf32CodeUnits)>

pub(crate) fn elements_from_point_query( &self, flags: HitTestFlags, point: LayoutPoint, ) -> HitTestResult

pub(crate) fn query_effective_overflow( &self, node: &Node, ) -> Option<AxesOverflow>

pub(crate) fn query_effective_overflow_without_reflow( &self, node: &Node, ) -> Option<AxesOverflow>

pub(crate) fn hit_test_from_input_event( &self, flags: HitTestFlags, input_event: &ConstellationInputEvent, ) -> Option<HitTestResult>

pub(crate) fn hit_test_from_point_in_viewport( &self, flags: HitTestFlags, point_in_frame: Point2D<f32, CSSPixel>, ) -> Option<HitTestResult>

pub(crate) fn init_window_proxy(&self, window_proxy: &WindowProxy)

pub(crate) fn init_document(&self, document: &Document)

pub(crate) fn update_jsprincipals_from_document(&self, document: &Document)

The JSPrincipals object inside a Window stores a mutable origin used for same-origin checks within SpiderMonkey. When the entire origin of a Window’s Document object is replaced or the Document object itself is replaced with a Document with a different origin, the JSPrincipals stored inside the Window also needs to change. This ensures that same origin checks are done against the correct origin.

pub(crate) fn load_data_for_document( &self, url: ServoUrl, pipeline_id: PipelineId, ) -> LoadData

pub(crate) fn set_viewport_details(&self, viewport_details: ViewportDetails)

Handle a potential change to the ViewportDetails of this Window, triggering a reflow if any change occurred.

pub(crate) fn viewport_details(&self) -> ViewportDetails

pub(crate) fn get_or_init_visual_viewport( &self, cx: &mut JSContext, ) -> DomRoot<VisualViewport>

pub(crate) fn maybe_update_visual_viewport( &self, cx: &mut JSContext, pinch_zoom_infos: PinchZoomInfos, )

Update the VisualViewport of this Window if necessary and note the changes to be processed in the event loop.

pub(crate) fn embedder_theme(&self) -> Theme

Get the embedder theme of this Window.

pub(crate) fn set_embedder_theme(&self, new_theme: Theme)

Handle a theme change request, triggering a reflow is any actual change occurred.

pub(crate) fn refresh_theme(&self)

pub(crate) fn take_pending_media_query_evaluation(&self) -> bool

Returns true and clears the flag if a media-feature change has occurred since the last call.

pub(crate) fn has_pending_media_query_evaluation(&self) -> bool

pub(crate) fn get_url(&self) -> ServoUrl

pub(crate) fn windowproxy_handler(&self) -> &'static WindowProxyHandler

pub(crate) fn add_resize_event( &self, event: ViewportDetails, event_type: WindowSizeType, )

pub(crate) fn take_unhandled_resize_event( &self, ) -> Option<(ViewportDetails, WindowSizeType)>

pub(crate) fn has_unhandled_resize_event(&self) -> bool

Whether or not this Window has any resize events that have not been processed.

pub(crate) fn suspend(&self, cx: &mut JSContext)

pub(crate) fn resume(&self, cx: &mut JSContext)

pub(crate) fn need_emit_timeline_marker( &self, timeline_type: TimelineMarkerType, ) -> bool

pub(crate) fn emit_timeline_marker(&self, marker: TimelineMarker)

pub(crate) fn set_devtools_timeline_markers( &self, markers: Vec<TimelineMarkerType>, reply: GenericSender<Option<TimelineMarker>>, )

pub(crate) fn drop_devtools_timeline_markers( &self, markers: Vec<TimelineMarkerType>, )

pub(crate) fn set_webdriver_load_status_sender( &self, sender: Option<GenericSender<WebDriverLoadStatus>>, )

pub(crate) fn webdriver_load_status_sender( &self, ) -> Option<GenericSender<WebDriverLoadStatus>>

pub(crate) fn is_alive(&self) -> bool

pub(crate) fn is_top_level(&self) -> bool

fn run_resize_steps_for_layout_viewport(&self, cx: &mut JSContext) -> bool

Layout viewport part of: https://drafts.csswg.org/cssom-view/#document-run-the-resize-steps

Handle the pending viewport resize.

pub(crate) fn run_the_resize_steps(&self, cx: &mut JSContext) -> bool

An implementation of: https://drafts.csswg.org/cssom-view/#document-run-the-resize-steps

Returns true if there were any pending viewport resize events.

pub(crate) fn evaluate_media_queries_and_report_changes( &self, cx: &mut JSContext, )

Evaluate media query lists and report changes https://drafts.csswg.org/cssom-view/#evaluate-media-queries-and-report-changes

pub(crate) fn set_throttled(&self, throttled: bool)

Set whether to use less resources by running timers at a heavily limited rate.

pub(crate) fn throttled(&self) -> bool

pub(crate) fn unminified_css_dir(&self) -> Option<String>

pub(crate) fn local_script_source(&self) -> &Option<String>

pub(crate) fn set_navigation_start(&self)

pub(crate) fn navigation_start(&self) -> CrossProcessInstant

pub(crate) fn set_last_activation_timestamp( &self, time: UserActivationTimestamp, )

pub(crate) fn send_to_embedder(&self, msg: EmbedderMsg)

pub(crate) fn send_to_constellation(&self, msg: ScriptToConstellationMessage)

pub(crate) fn in_immersive_xr_session(&self) -> bool

fn handle_new_or_removed_web_fonts_post_reflow( &self, cx: &mut JSContext, changed_web_fonts: WebFontSetDifference, )

Adds and removes entries from document.fonts as needed after a reflow.

fn process_lcp_candidate_post_reflow( &self, candidate: &LCPCandidate, node_address: UntrustedNodeAddress, document: &Document, )

Resolve the LCP candidate OpaqueNode to a DOM Element and store it on the document.

fn handle_pending_images_post_reflow( &self, cx: &mut JSContext, pending_images: Vec<PendingImage>, pending_rasterization_images: Vec<PendingRasterizationImage>, pending_svg_element_for_serialization: Vec<UntrustedNodeAddress>, )

pub(crate) fn has_sticky_activation(&self) -> bool

https://html.spec.whatwg.org/multipage/#sticky-activation

pub(crate) fn has_transient_activation(&self) -> bool

https://html.spec.whatwg.org/multipage/#transient-activation

pub(crate) fn consume_last_activation_timestamp(&self)

pub(crate) fn consume_user_activation(&self)

https://html.spec.whatwg.org/multipage/#consume-user-activation

pub(crate) fn new( cx: &mut JSContext, webview_id: WebViewId, runtime: Rc<Runtime>, script_chan: Sender<MainThreadScriptMsg>, layout: Box<dyn Layout>, image_cache_sender: Sender<ImageCacheResponseMessage>, resource_threads: ResourceThreads, storage_threads: StorageThreads, bluetooth_thread: GenericSender<BluetoothRequest>, mem_profiler_chan: MemProfilerChan, time_profiler_chan: TimeProfilerChan, devtools_chan: Option<GenericCallback<ScriptToDevtoolsControlMsg>>, script_to_constellation_sender: ScriptToConstellationSender, embedder_chan: ScriptToEmbedderChan, control_chan: GenericSender<ScriptThreadMessage>, pipeline_id: PipelineId, parent_info: Option<PipelineId>, viewport_details: ViewportDetails, origin: MutableOrigin, creation_url: ServoUrl, top_level_creation_url: ServoUrl, navigation_start: CrossProcessInstant, webgl_chan: Option<WebGLChan>, webxr_registry: Option<Registry>, paint_api: CrossProcessPaintApi, unminify_js: bool, unminify_css: bool, local_script_source: Option<String>, user_scripts: Rc<Vec<UserScript>>, player_context: WindowGLContext, gpu_id_hub: Arc<IdentityHub>, inherited_secure_context: Option<bool>, embedder_theme: Theme, weak_script_thread: Weak<ScriptThread>, ) -> DomRoot<Self>

pub(crate) fn task_manager(&self) -> Rc<TaskManager>

pub(crate) fn pipeline_id(&self) -> PipelineId

pub(crate) fn live_devtools_updates(&self) -> bool

pub(crate) fn set_devtools_wants_updates(&self, value: bool)

pub(crate) fn cache_layout_value<T>(&self, value: T) -> LayoutValue<T>
where T: Copy + MallocSizeOf,

Create a new cached instance of the given value.

pub(crate) fn set_up_a_window_environment_settings_object( &self, layout: Box<dyn Layout>, creation_url: ServoUrl, top_level_creation_url: ServoUrl, navigation_start: CrossProcessInstant, viewport_details: ViewportDetails, )

This method is an approximation of the specification [algorithm]. It exists in this form because we still store some fields in Window/GlobalScope that realistically are specific to the active document. Where possible they should be migrated to Document and WorkerGlobalScope, but currently doing so would result in much more complicated code. This method is the compromise, where we mutate the values in place to match the values that the specification expects.

[algorithm] https://html.spec.whatwg.org/multipage/#set-up-a-window-environment-settings-object

impl Window

pub(crate) fn post_message( &self, target_origin: Option<ImmutableOrigin>, source_origin: ImmutableOrigin, source: &WindowProxy, data: StructuredSerializedData, )