script::dom::window

Module windowproxy

Structs

CreatorBrowsingContextInfo 🔒

A browsing context can have a creator browsing context, the browsing context that was responsible for its creation. If a browsing context has a parent browsing context, then that is its creator browsing context. Otherwise, if the browsing context has an opener browsing context, then that is its creator browsing context. Otherwise, the browsing context has no creator browsing context.

WindowProxy 🔒

WindowProxyHandler 🔒

Proxy handler for a WindowProxy. Has ownership of the inner pointer and deallocates it when it is no longer needed.

Enums

IncludePrototypes 🔒

WindowOrDissimilarOriginWindow 🔒

A wrapper class for either a Window or DissimilarOriginWindow that exposes a consistent interface for both of them.

Constants

_IMPL_DOMOBJECT_FOR_WindowProxy 🔒

Statics

PROXY_TRAPS 🔒

Functions

define_property 🔒 ⚠

https://html.spec.whatwg.org/multipage/#windowproxy-defineownproperty

delete 🔒 ⚠

https://html.spec.whatwg.org/multipage/#windowproxy-delete

finalize 🔒 ⚠

get 🔒 ⚠

https://html.spec.whatwg.org/multipage/#windowproxy-get

get_own_property_descriptor 🔒 ⚠

https://html.spec.whatwg.org/multipage/#windowproxy-getownproperty

get_prototype 🔒 ⚠

https://html.spec.whatwg.org/multipage/#windowproxy-getprototypeof

get_prototype_if_ordinary 🔒 ⚠

has 🔒 ⚠

has_or_has_own 🔒 ⚠

has_own 🔒 ⚠

own_property_keys 🔒 ⚠

https://html.spec.whatwg.org/multipage#windowproxy-ownpropertykeys

parse_open_feature_boolean 🔒

https://html.spec.whatwg.org/multipage/#concept-window-open-features-parse-boolean

set 🔒 ⚠

https://html.spec.whatwg.org/multipage/#windowproxy-set

tokenize_open_features 🔒

https://html.spec.whatwg.org/multipage/#concept-window-open-features-tokenize

trace 🔒 ⚠

window_proxy_target 🔒