script::dom::window::windowproxy

Struct WindowProxy

#[repr(C)]
pub(crate) struct WindowProxy {

    reflector: Reflector,
    browsing_context_id: BrowsingContextId,
    opener: Option<BrowsingContextId>,
    webview_id: WebViewId,
    name: DomRefCell<DOMString>,
    currently_active: Cell<Option<PipelineId>>,
    discarded: Cell<bool>,
    disowned: Cell<bool>,
    is_closing: Cell<bool>,
    frame_element: Option<Dom<Element>>,
    parent: Option<Dom<WindowProxy>>,
    delaying_load_events_mode: Cell<bool>,
    creator_url: Option<ServoUrl>,
    creator_origin: Option<ImmutableOrigin>,
    script_window_proxies: Rc<ScriptWindowProxies>,

}

Fields
reflector: Reflector

The JS WindowProxy object. Unlike other reflectors, we mutate this field because we have to brain-transplant the reflector when the WindowProxy changes Window.

browsing_context_id: BrowsingContextId

The id of the browsing context. In the case that this is a nested browsing context, this is the id of the container.

opener: Option<BrowsingContextId>webview_id: WebViewId

The frame id of the top-level ancestor browsing context. In the case that this is a top-level window, this is our id.

name: DomRefCell<DOMString>

The name of the browsing context (sometimes, but not always, equal to the name of a container element)

currently_active: Cell<Option<PipelineId>>

The pipeline id of the currently active document. May be None, when the currently active document is in another script thread. We do not try to keep the pipeline id for documents in other threads, as this would require the constellation notifying many script threads about the change, which could be expensive.

discarded: Cell<bool>

Has the browsing context been discarded?

disowned: Cell<bool>

Has the browsing context been disowned?

is_closing: Cell<bool>

https://html.spec.whatwg.org/multipage/#is-closing

frame_element: Option<Dom<Element>>

If the containing <iframe> of this WindowProxy is from a same-origin page, this will be the Element of the <iframe> element in the realm of the parent page. Otherwise, it is None.

parent: Option<Dom<WindowProxy>>

The parent browsing context’s window proxy, if this is a nested browsing context

delaying_load_events_mode: Cell<bool>

https://html.spec.whatwg.org/multipage/#delaying-load-events-mode

creator_url: Option<ServoUrl>

The creator browsing context’s url.

creator_origin: Option<ImmutableOrigin>

The creator browsing context’s origin.

script_window_proxies: Rc<ScriptWindowProxies>

The window proxies the script thread knows.

Implementations

impl WindowProxy

fn new_inherited( browsing_context_id: BrowsingContextId, webview_id: WebViewId, currently_active: Option<PipelineId>, frame_element: Option<&Element>, parent: Option<&WindowProxy>, opener: Option<BrowsingContextId>, creator: CreatorBrowsingContextInfo, ) -> WindowProxy

pub(crate) fn new( cx: &mut JSContext, window: &Window, browsing_context_id: BrowsingContextId, webview_id: WebViewId, frame_element: Option<&Element>, parent: Option<&WindowProxy>, opener: Option<BrowsingContextId>, creator: CreatorBrowsingContextInfo, ) -> DomRoot<WindowProxy>

pub(crate) fn new_dissimilar_origin( cx: &mut JSContext, global_to_clone_from: &GlobalScope, browsing_context_id: BrowsingContextId, webview_id: WebViewId, parent: Option<&WindowProxy>, opener: Option<BrowsingContextId>, creator: CreatorBrowsingContextInfo, ) -> DomRoot<WindowProxy>

fn create_auxiliary_browsing_context( &self, cx: &mut JSContext, name: DOMString, noopener: bool, ) -> Option<DomRoot<WindowProxy>>

https://html.spec.whatwg.org/multipage/#auxiliary-browsing-context

pub(crate) fn start_delaying_load_events_mode(&self)

https://html.spec.whatwg.org/multipage/#delaying-load-events-mode

pub(crate) fn stop_delaying_load_events_mode(&self)

https://html.spec.whatwg.org/multipage/#delaying-load-events-mode

pub(crate) fn disown(&self)

pub(crate) fn close(&self)

https://html.spec.whatwg.org/multipage/#dom-window-close Step 3.1, set BCs is_closing to true.

pub(crate) fn is_closing(&self) -> bool

https://html.spec.whatwg.org/multipage/#is-closing

pub(crate) fn opener( &self, cx: &mut CurrentRealm<'_>, retval: MutableHandleValue<'_>, )

pub(crate) fn open( &self, cx: &mut JSContext, url: USVString, target: DOMString, features: DOMString, ) -> Fallible<Option<DomRoot<WindowProxy>>>

pub(crate) fn choose_a_navigable( &self, cx: &mut JSContext, name: DOMString, noopener: bool, ) -> (Option<DomRoot<WindowProxy>>, bool)

https://html.spec.whatwg.org/multipage/#the-rules-for-choosing-a-navigable

fn find_navigable_by_target_name( &self, name: &DOMString, ) -> Option<DomRoot<WindowProxy>>

https://html.spec.whatwg.org/multipage/#find-a-navigable-by-target-name

fn find_navigable_by_target_name_in_descendants( &self, name: &DOMString, ) -> Option<DomRoot<WindowProxy>>

https://html.spec.whatwg.org/multipage/#find-a-navigable-by-target-name step 4 and 7 substeps.

pub(crate) fn is_auxiliary(&self) -> bool

pub(crate) fn discard_browsing_context(&self)

pub(crate) fn is_browsing_context_discarded(&self) -> bool

pub(crate) fn browsing_context_id(&self) -> BrowsingContextId

pub(crate) fn webview_id(&self) -> WebViewId

pub(crate) fn frame_element(&self) -> Option<&Element>

If the containing <iframe> of this WindowProxy is from a same-origin page, this will return an Element of the <iframe> element in the realm of the parent page.

pub(crate) fn document(&self) -> Option<DomRoot<Document>>

pub(crate) fn parent(&self) -> Option<&WindowProxy>

pub(crate) fn top(&self) -> &WindowProxy

pub(crate) fn document_origin(&self) -> Option<OriginSnapshot>

pub(crate) fn internal_ancestor_origin_objects_list( &self, ) -> Option<Vec<ImmutableOrigin>>

pub(crate) fn parent_origin_and_internal_ancestor_origin_objects_list( &self, ) -> Option<(OriginSnapshot, Vec<ImmutableOrigin>)>

https://html.spec.whatwg.org/multipage/#internal-ancestor-origin-objects-list-creation-steps

fn set_window(&self, cx: &mut JSContext, window: &GlobalScope)

Change the Window that this WindowProxy resolves to.

pub(crate) fn set_pipeline_id(&self, pipeline_id: PipelineId)

pub(crate) fn set_currently_active(&self, cx: &mut JSContext, window: &Window)

pub(crate) fn unset_currently_active(&self, cx: &mut JSContext)

pub(crate) fn currently_active(&self) -> Option<PipelineId>

pub(crate) fn get_name(&self) -> DOMString

pub(crate) fn set_name(&self, name: DOMString)