Struct script::dom::windowproxy::WindowProxy

#[repr(C)]
pub struct WindowProxy {

    reflector: Reflector,
    browsing_context_id: BrowsingContextId,
    opener: Option<BrowsingContextId>,
    top_level_browsing_context_id: TopLevelBrowsingContextId,
    name: DomRefCell<DOMString>,
    currently_active: Cell<Option<PipelineId>>,
    discarded: Cell<bool>,
    disowned: Cell<bool>,
    is_closing: Cell<bool>,
    frame_element: Option<Dom<Element>>,
    parent: Option<Dom<WindowProxy>>,
    delaying_load_events_mode: Cell<bool>,
    creator_base_url: Option<ServoUrl>,
    creator_url: Option<ServoUrl>,
    creator_origin: Option<ImmutableOrigin>,

}

Fields
reflector: Reflector

The JS WindowProxy object. Unlike other reflectors, we mutate this field because we have to brain-transplant the reflector when the WindowProxy changes Window.

browsing_context_id: BrowsingContextId

The id of the browsing context. In the case that this is a nested browsing context, this is the id of the container.

opener: Option<BrowsingContextId>top_level_browsing_context_id: TopLevelBrowsingContextId

The frame id of the top-level ancestor browsing context. In the case that this is a top-level window, this is our id.

name: DomRefCell<DOMString>

The name of the browsing context (sometimes, but not always, equal to the name of a container element)

currently_active: Cell<Option<PipelineId>>

The pipeline id of the currently active document. May be None, when the currently active document is in another script thread. We do not try to keep the pipeline id for documents in other threads, as this would require the constellation notifying many script threads about the change, which could be expensive.

discarded: Cell<bool>

Has the browsing context been discarded?

disowned: Cell<bool>

Has the browsing context been disowned?

is_closing: Cell<bool>

https://html.spec.whatwg.org/multipage/#is-closing

frame_element: Option<Dom<Element>>

The containing iframe element, if this is a same-origin iframe

parent: Option<Dom<WindowProxy>>

The parent browsing context’s window proxy, if this is a nested browsing context

delaying_load_events_mode: Cell<bool>

https://html.spec.whatwg.org/multipage/#delaying-load-events-mode

creator_base_url: Option<ServoUrl>

The creator browsing context’s base url.

creator_url: Option<ServoUrl>

The creator browsing context’s url.

creator_origin: Option<ImmutableOrigin>

The creator browsing context’s origin.

Implementations

impl WindowProxy

pub fn new_inherited( browsing_context_id: BrowsingContextId, top_level_browsing_context_id: TopLevelBrowsingContextId, currently_active: Option<PipelineId>, frame_element: Option<&Element>, parent: Option<&WindowProxy>, opener: Option<BrowsingContextId>, creator: CreatorBrowsingContextInfo ) -> WindowProxy

pub fn new( window: &Window, browsing_context_id: BrowsingContextId, top_level_browsing_context_id: TopLevelBrowsingContextId, frame_element: Option<&Element>, parent: Option<&WindowProxy>, opener: Option<BrowsingContextId>, creator: CreatorBrowsingContextInfo ) -> Root<Dom<WindowProxy>>

pub fn new_dissimilar_origin( global_to_clone_from: &GlobalScope, browsing_context_id: BrowsingContextId, top_level_browsing_context_id: TopLevelBrowsingContextId, parent: Option<&WindowProxy>, opener: Option<BrowsingContextId>, creator: CreatorBrowsingContextInfo ) -> Root<Dom<WindowProxy>>

fn create_auxiliary_browsing_context( &self, name: DOMString, noopener: bool ) -> Option<Root<Dom<WindowProxy>>>

pub fn is_delaying_load_events_mode(&self) -> bool

https://html.spec.whatwg.org/multipage/#delaying-load-events-mode

pub fn start_delaying_load_events_mode(&self)

https://html.spec.whatwg.org/multipage/#delaying-load-events-mode

pub fn stop_delaying_load_events_mode(&self)

https://html.spec.whatwg.org/multipage/#delaying-load-events-mode

pub fn disown(&self)

pub fn close(&self)

https://html.spec.whatwg.org/multipage/#dom-window-close Step 3.1, set BCs is_closing to true.

pub fn is_closing(&self) -> bool

https://html.spec.whatwg.org/multipage/#is-closing

pub fn creator_base_url(&self) -> Option<ServoUrl>

https://html.spec.whatwg.org/multipage/#creator-base-url

pub fn has_creator_base_url(&self) -> bool

pub fn creator_url(&self) -> Option<ServoUrl>

https://html.spec.whatwg.org/multipage/#creator-url

pub fn has_creator_url(&self) -> bool

pub fn creator_origin(&self) -> Option<ImmutableOrigin>

https://html.spec.whatwg.org/multipage/#creator-origin

pub fn has_creator_origin(&self) -> bool

pub fn opener(&self, cx: *mut JSContext, in_realm_proof: InRealm<'_>) -> JSVal

pub fn open( &self, url: USVString, target: DOMString, features: DOMString ) -> Result<Option<Root<Dom<WindowProxy>>>, Error>

pub fn choose_browsing_context( &self, name: DOMString, noopener: bool ) -> (Option<Root<Dom<WindowProxy>>>, bool)

pub fn is_auxiliary(&self) -> bool

pub fn discard_browsing_context(&self)

pub fn is_browsing_context_discarded(&self) -> bool

pub fn browsing_context_id(&self) -> BrowsingContextId

pub fn top_level_browsing_context_id(&self) -> TopLevelBrowsingContextId

pub fn frame_element(&self) -> Option<&Element>

pub fn document(&self) -> Option<Root<Dom<Document>>>

pub fn parent(&self) -> Option<&WindowProxy>

pub fn top(&self) -> &WindowProxy

fn set_window(&self, window: &GlobalScope, traps: &ProxyTraps)

Change the Window that this WindowProxy resolves to.

pub fn set_currently_active(&self, window: &Window)

pub fn unset_currently_active(&self)

pub fn currently_active(&self) -> Option<PipelineId>

pub fn get_name(&self) -> DOMString

pub fn set_name(&self, name: DOMString)