Struct script::dom::workerglobalscope::WorkerGlobalScope

#[repr(C)]
pub struct WorkerGlobalScope {

    globalscope: GlobalScope,
    worker_name: DOMString,
    worker_type: WorkerType,
    worker_id: WorkerId,
    worker_url: DomRefCell<ServoUrl>,
    closing: Arc<AtomicBool>,
    runtime: DomRefCell<Option<Runtime>>,
    location: MutNullableDom<WorkerLocation>,
    navigator: MutNullableDom<WorkerNavigator>,
    from_devtools_sender: Option<IpcSender<DevtoolScriptControlMsg>>,
    from_devtools_receiver: Receiver<DevtoolScriptControlMsg>,
    navigation_start_precise: u64,
    performance: MutNullableDom<Performance>,

}

Fields
globalscope: GlobalScopeworker_name: DOMStringworker_type: WorkerTypeworker_id: WorkerIdworker_url: DomRefCell<ServoUrl>closing: Arc<AtomicBool>runtime: DomRefCell<Option<Runtime>>location: MutNullableDom<WorkerLocation>navigator: MutNullableDom<WorkerNavigator>from_devtools_sender: Option<IpcSender<DevtoolScriptControlMsg>>

Optional IpcSender for sending the DevtoolScriptControlMsg to the server from within the worker

from_devtools_receiver: Receiver<DevtoolScriptControlMsg>

This Receiver will be ignored later if the corresponding IpcSender doesn’t exist

navigation_start_precise: u64performance: MutNullableDom<Performance>
Implementations

impl WorkerGlobalScope

pub fn new_inherited( init: WorkerGlobalScopeInit, worker_name: DOMString, worker_type: WorkerType, worker_url: ServoUrl, runtime: Runtime, from_devtools_receiver: Receiver<DevtoolScriptControlMsg>, closing: Arc<AtomicBool>, gpu_id_hub: Arc<Mutex<Identities>> ) -> Self

pub fn clear_js_runtime(&self, cx_for_interrupt: ContextForRequestInterrupt)

Clear various items when the worker event-loop shuts-down.

pub fn runtime_handle(&self) -> ParentRuntime

pub fn from_devtools_sender(&self) -> Option<IpcSender<DevtoolScriptControlMsg>>

pub fn from_devtools_receiver(&self) -> &Receiver<DevtoolScriptControlMsg>

pub fn get_cx(&self) -> JSContext

pub fn is_closing(&self) -> bool

pub fn get_url(&self) -> Ref<'_, ServoUrl>

pub fn set_url(&self, url: ServoUrl)

pub fn get_worker_id(&self) -> WorkerId

pub fn task_canceller(&self) -> TaskCanceller

pub fn pipeline_id(&self) -> PipelineId

impl WorkerGlobalScope

pub fn execute_script(&self, source: DOMString)

pub fn script_chan(&self) -> Box<dyn ScriptChan + Send>

pub fn dom_manipulation_task_source(&self) -> DOMManipulationTaskSource

pub fn file_reading_task_source(&self) -> FileReadingTaskSource

pub fn networking_task_source(&self) -> NetworkingTaskSource

pub fn performance_timeline_task_source(&self) -> PerformanceTimelineTaskSource

pub fn port_message_queue(&self) -> PortMessageQueue

pub fn timer_task_source(&self) -> TimerTaskSource

pub fn remote_event_task_source(&self) -> RemoteEventTaskSource

pub fn websocket_task_source(&self) -> WebsocketTaskSource

pub fn new_script_pair( &self ) -> (Box<dyn ScriptChan + Send>, Box<dyn ScriptPort + Send>)

pub fn process_event(&self, msg: CommonScriptMsg) -> bool

Process a single event as if it were the next event in the queue for this worker event-loop. Returns a boolean indicating whether further events should be processed.

pub fn close(&self)