script::dom::workers::dedicatedworkerglobalscope

Struct DedicatedWorkerGlobalScope

#[repr(C)]
pub(crate) struct DedicatedWorkerGlobalScope {

    workerglobalscope: WorkerGlobalScope,
    webview_id: WebViewId,
    task_queue: TaskQueue<DedicatedWorkerScriptMsg>,
    own_sender: Sender<DedicatedWorkerScriptMsg>,
    worker: DomRefCell<Option<Trusted<Worker>>>,
    parent_event_loop_sender: ScriptEventLoopSender,
    image_cache: Arc<dyn ImageCache>,
    browsing_context: Option<BrowsingContextId>,
    animation_frame_provider_supported: Arc<AtomicBool>,
    animation_frame_provider_registered: Cell<bool>,
    animation_frame_ident: Cell<u32>,
    animation_frame_list: DomRefCell<VecDeque<(u32, Rc<FrameRequestCallback<DomTypeHolder>>)>>,
    current_animation_frame_list: DomRefCell<VecDeque<(u32, Rc<FrameRequestCallback<DomTypeHolder>>)>>,
    running_animation_callbacks: Cell<bool>,
    animation_frame_callbacks_active: Cell<bool>,
    animation_frame_tick_sender: Option<GenericSender<WorkerAnimationFrameTick>>,
    animation_frame_tick_receiver: Option<RoutedReceiver<WorkerAnimationFrameTick>>,
    control_receiver: Receiver<DedicatedWorkerControlMsg>,
    queued_worker_tasks: DomRefCell<Vec<MessageData>>,

}

Fields
workerglobalscope: WorkerGlobalScopewebview_id: WebViewId

The WebViewId of the WebView that this worker is associated with.

task_queue: TaskQueue<DedicatedWorkerScriptMsg>own_sender: Sender<DedicatedWorkerScriptMsg>worker: DomRefCell<Option<Trusted<Worker>>>parent_event_loop_sender: ScriptEventLoopSender

Sender to the parent thread.

image_cache: Arc<dyn ImageCache>browsing_context: Option<BrowsingContextId>animation_frame_provider_supported: Arc<AtomicBool>animation_frame_provider_registered: Cell<bool>animation_frame_ident: Cell<u32>

https://html.spec.whatwg.org/multipage/#animation-frame-callback-identifier

animation_frame_list: DomRefCell<VecDeque<(u32, Rc<FrameRequestCallback<DomTypeHolder>>)>>

Pending animation frame callbacks for a later worker rendering update.

current_animation_frame_list: DomRefCell<VecDeque<(u32, Rc<FrameRequestCallback<DomTypeHolder>>)>>

Callbacks snapshotted for the current worker rendering update.

running_animation_callbacks: Cell<bool>

Whether we’re in the process of running animation callbacks.

animation_frame_callbacks_active: Cell<bool>

Whether Constellation currently treats this worker as having callbacks.

animation_frame_tick_sender: Option<GenericSender<WorkerAnimationFrameTick>>

A sender for animation frame ticks.

animation_frame_tick_receiver: Option<RoutedReceiver<WorkerAnimationFrameTick>>

A receiver for animation frame ticks.

control_receiver: Receiver<DedicatedWorkerControlMsg>

A receiver of control messages.

queued_worker_tasks: DomRefCell<Vec<MessageData>>
Implementations

impl DedicatedWorkerGlobalScope

fn __assert_parent_type(&self)

impl DedicatedWorkerGlobalScope

fn new_inherited( init: WorkerGlobalScopeInit, webview_id: WebViewId, worker_name: DOMString, worker_type: WorkerType, worker_url: ServoUrl, from_devtools_receiver: RoutedReceiver<DevtoolScriptControlMsg>, runtime: Runtime, parent_event_loop_sender: ScriptEventLoopSender, own_sender: Sender<DedicatedWorkerScriptMsg>, receiver: Receiver<DedicatedWorkerScriptMsg>, closing: Arc<AtomicBool>, animation_frame_provider_supported: Arc<AtomicBool>, image_cache: Arc<dyn ImageCache>, browsing_context: Option<BrowsingContextId>, animation_frame_tick_sender: Option<GenericSender<WorkerAnimationFrameTick>>, animation_frame_tick_receiver: Option<RoutedReceiver<WorkerAnimationFrameTick>>, gpu_id_hub: Arc<IdentityHub>, control_receiver: Receiver<DedicatedWorkerControlMsg>, insecure_requests_policy: InsecureRequestsPolicy, font_context: Arc<FontContext>, ) -> DedicatedWorkerGlobalScope

pub(crate) fn new( init: WorkerGlobalScopeInit, webview_id: WebViewId, worker_name: DOMString, worker_type: WorkerType, worker_url: ServoUrl, from_devtools_receiver: RoutedReceiver<DevtoolScriptControlMsg>, runtime: Runtime, parent_event_loop_sender: ScriptEventLoopSender, own_sender: Sender<DedicatedWorkerScriptMsg>, receiver: Receiver<DedicatedWorkerScriptMsg>, closing: Arc<AtomicBool>, animation_frame_provider_supported: Arc<AtomicBool>, image_cache: Arc<dyn ImageCache>, browsing_context: Option<BrowsingContextId>, animation_frame_tick_sender: Option<GenericSender<WorkerAnimationFrameTick>>, animation_frame_tick_receiver: Option<RoutedReceiver<WorkerAnimationFrameTick>>, gpu_id_hub: Arc<IdentityHub>, control_receiver: Receiver<DedicatedWorkerControlMsg>, insecure_requests_policy: InsecureRequestsPolicy, font_context: Arc<FontContext>, debugger_global: &DebuggerGlobalScope, cx: &mut JSContext, ) -> DomRoot<DedicatedWorkerGlobalScope>

pub(crate) fn run_worker_scope( init: WorkerGlobalScopeInit, webview_id: WebViewId, worker_url: UrlWithBlobClaim, from_devtools_receiver: GenericReceiver<DevtoolScriptControlMsg>, worker: Trusted<Worker>, parent_event_loop_sender: ScriptEventLoopSender, own_sender: Sender<DedicatedWorkerScriptMsg>, receiver: Receiver<DedicatedWorkerScriptMsg>, worker_load_origin: WorkerScriptLoadOrigin, worker_options: &WorkerOptions, closing: Arc<AtomicBool>, animation_frame_provider_supported: Arc<AtomicBool>, image_cache: Arc<dyn ImageCache>, browsing_context: Option<BrowsingContextId>, gpu_id_hub: Arc<IdentityHub>, control_receiver: Receiver<DedicatedWorkerControlMsg>, context_sender: Sender<ThreadSafeJSContext>, insecure_requests_policy: InsecureRequestsPolicy, policy_container: PolicyContainer, font_context: Arc<FontContext>, ) -> JoinHandle<()>

https://html.spec.whatwg.org/multipage/#run-a-worker

pub(crate) fn webview_id(&self) -> WebViewId

pub(crate) fn animation_frame_provider_supported(&self) -> bool

pub(crate) fn animation_frame_provider_supported_flag(&self) -> Arc<AtomicBool>

fn register_animation_frame_provider(&self)

fn unregister_animation_frame_provider(&self)

fn send_animation_frame_callbacks_state(&self, active: bool)

fn set_animation_frame_callbacks_active(&self, active: bool)

fn remove_animation_frame_callback_from( list: &DomRefCell<VecDeque<(u32, Rc<FrameRequestCallback<DomTypeHolder>>)>>, ident: u32, )

fn has_animation_frame_callbacks(&self) -> bool

pub(crate) fn request_animation_frame( &self, callback: Rc<FrameRequestCallback<DomTypeHolder>>, ) -> Fallible<u32>

https://html.spec.whatwg.org/multipage/#dom-animationframeprovider-requestanimationframe

pub(crate) fn cancel_animation_frame(&self, ident: u32) -> ErrorResult

https://html.spec.whatwg.org/multipage/#dom-animationframeprovider-cancelanimationframe

pub(crate) fn run_the_animation_frame_callbacks(&self, cx: &mut JSContext)

https://html.spec.whatwg.org/multipage/#run-the-animation-frame-callbacks

pub(crate) fn clear_animation_frame_callbacks_and_unregister(&self)

fn replace_worker( &self, new_worker: Option<Trusted<Worker>>, ) -> Option<Trusted<Worker>>

The non-None value of the worker field can contain a rooted TrustedWorkerAddress version of the main thread’s worker object. This is set while handling messages and then unset otherwise, ensuring that the main thread object can be garbage collected. See AutoWorkerReset.

pub(crate) fn image_cache(&self) -> Arc<dyn ImageCache>

pub(crate) fn event_loop_sender(&self) -> Option<ScriptEventLoopSender>

pub(crate) fn new_script_pair( &self, ) -> (ScriptEventLoopSender, ScriptEventLoopReceiver)

pub(crate) fn fire_queued_messages(&self, cx: &mut JSContext)

fn dispatch_message_event(&self, cx: &mut JSContext, msg: MessageData)

fn handle_script_event(&self, msg: WorkerScriptMsg, cx: &mut JSContext)

fn handle_mixed_message(&self, msg: MixedMessage, cx: &mut JSContext) -> bool

https://html.spec.whatwg.org/multipage/#event-loop-processing-model

pub(crate) fn forward_error_to_worker_object(&self, error_info: ErrorInfo)

Step 7.2 of https://html.spec.whatwg.org/multipage/#report-an-exception

fn post_message_impl( &self, cx: &mut JSContext, message: HandleValue<'_>, transfer: CustomAutoRooterGuard<'_, Vec<*mut JSObject>>, ) -> ErrorResult

https://html.spec.whatwg.org/multipage/#dom-dedicatedworkerglobalscope-postmessage

pub(crate) fn browsing_context(&self) -> Option<BrowsingContextId>

pub(crate) fn report_csp_violations(&self, violations: Vec<Violation>)

pub(crate) fn forward_simple_error_at_worker(&self)