script::dom::worklet::paintworkletglobalscope

Struct PaintWorkletGlobalScope

#[repr(C)]
pub(crate) struct PaintWorkletGlobalScope {
    worklet_global: WorkletGlobalScope,
    image_cache: Arc<dyn ImageCache>,
    paint_definitions: DomRefCell<HashMapTracedValues<Atom, Box<PaintDefinition>>>,
    paint_class_instances: DomRefCell<HashMapTracedValues<Atom, Box<Heap<JSVal>>>>,
    cached_name: DomRefCell<Atom>,
    cached_size: Cell<Size2D<f32, CSSPixel>>,
    cached_device_pixel_ratio: Cell<Scale<f32, CSSPixel, DevicePixel>>,
    cached_properties: DomRefCell<Vec<(Atom, String)>>,
    cached_arguments: DomRefCell<Vec<String>>,
    cached_result: DomRefCell<DrawAPaintImageResult>,
}

https://drafts.css-houdini.org/css-paint-api/#paintworkletglobalscope

Fields
worklet_global: WorkletGlobalScope

The worklet global for this object

image_cache: Arc<dyn ImageCache>

The image cache

paint_definitions: DomRefCell<HashMapTracedValues<Atom, Box<PaintDefinition>>>

https://drafts.css-houdini.org/css-paint-api/#paint-definitions

paint_class_instances: DomRefCell<HashMapTracedValues<Atom, Box<Heap<JSVal>>>>

https://drafts.css-houdini.org/css-paint-api/#paint-class-instances

cached_name: DomRefCell<Atom>

The most recent name the worklet was called with

cached_size: Cell<Size2D<f32, CSSPixel>>

The most recent size the worklet was drawn at

cached_device_pixel_ratio: Cell<Scale<f32, CSSPixel, DevicePixel>>

The most recent device pixel ratio the worklet was drawn at

cached_properties: DomRefCell<Vec<(Atom, String)>>

The most recent properties the worklet was drawn at

cached_arguments: DomRefCell<Vec<String>>

The most recent arguments the worklet was drawn at

cached_result: DomRefCell<DrawAPaintImageResult>

The most recent result

Implementations

impl PaintWorkletGlobalScope

fn __assert_parent_type(&self)

impl PaintWorkletGlobalScope

pub(crate) fn new( cx: &mut JSContext, pipeline_id: PipelineId, base_url: ServoUrl, inherited_secure_context: Option<bool>, executor: WorkletExecutor, init: &WorkletGlobalScopeInit, closing: Arc<AtomicBool>, microtask_queue: Rc<MicrotaskQueue>, ) -> DomRoot<PaintWorkletGlobalScope>

pub(crate) fn image_cache(&self) -> Arc<dyn ImageCache>

fn has_cached_paint_image( &self, name: &Atom, size: Size2D<f32, CSSPixel>, device_pixel_ratio: Scale<f32, CSSPixel, DevicePixel>, properties: &[(Atom, String)], arguments: &[String], ) -> bool

fn set_cached_paint_image( &self, name: Atom, size: Size2D<f32, CSSPixel>, device_pixel_ratio: Scale<f32, CSSPixel, DevicePixel>, properties: Vec<(Atom, String)>, arguments: Vec<String>, result: DrawAPaintImageResult, )

fn draw_a_paint_image( &self, cx: &mut JSContext, name: &Atom, size_in_px: Size2D<f32, CSSPixel>, device_pixel_ratio: Scale<f32, CSSPixel, DevicePixel>, properties: &StylePropertyMapReadOnly, arguments: &[String], ) -> DrawAPaintImageResult

https://drafts.css-houdini.org/css-paint-api/#draw-a-paint-image

fn get_or_create_paint_instance( &self, cx: &mut JSContext, paint_instance: MutableHandle<'_, Value>, name: Atom, class_constructor: HandleValue<'_>, size_in_dpx: Size2D<u32, DevicePixel>, ) -> Result<(), DrawAPaintImageResult>

Gets or creates the self.paint_class_instances given by name and inserts the value into the provided paint_instance handle. Returns Err when this fails.

fn invoke_a_paint_callback( &self, cx: &mut JSContext, name: &Atom, size_in_px: Size2D<f32, CSSPixel>, size_in_dpx: Size2D<u32, DevicePixel>, device_pixel_ratio: Scale<f32, CSSPixel, DevicePixel>, properties: &StylePropertyMapReadOnly, arguments: &[String], ) -> DrawAPaintImageResult

https://drafts.css-houdini.org/css-paint-api/#invoke-a-paint-callback

fn invalid_image( &self, size: Size2D<u32, DevicePixel>, missing_image_urls: Vec<ServoUrl>, ) -> DrawAPaintImageResult

https://drafts.csswg.org/css-images-4/#invalid-image

fn painter(&self, name: Atom) -> Box<dyn Painter>