script::dom::worklet

Module worklet

An implementation of Houdini worklets.

The goal of this implementation is to maximize responsiveness of worklets, and in particular to ensure that the thread performing worklet tasks is never busy GCing or loading worklet code. We do this by providing a custom thread pool implementation, which only performs GC or code loading on a backup thread, not on the primary worklet thread.

Structs

DroppableField 🔒

PendingTasksStruct 🔒

https://drafts.css-houdini.org/worklets/#pending-tasks-struct

StatelessWorkletThreadPool 🔒

The StatelessWorkletThreadPool executes the associated WorkletTasks in a dedicated thread pool with the assumption that the tasks are idempotent. This is useful for the paint worklet, for example.

Worklet 🔒

https://drafts.css-houdini.org/worklets/#worklet

WorkletExecutor 🔒

An executor of worklet tasks

WorkletId 🔒

A guid for worklets.

WorkletThread 🔒

A thread for executing worklets.

WorkletThreadInit 🔒

Data to initialize a worklet thread.

WorkletThreadRole 🔒

A role that a worklet thread can be playing.

Enums

WorkletControl 🔒

The control message sent to worklet threads

WorkletData 🔒

The data messages sent to worklet threads

Constants

MIN_GC_THRESHOLD 🔒

WORKLET_THREAD_POOL_SIZE 🔒

_IMPL_DOMOBJECT_FOR_Worklet 🔒

Traits

WorkletThreadPool

Functions

reject_promise 🔒

This function is an abstraction of steps 6.4.1.1 and 6.4.2.1 of the AddModule spec https://html.spec.whatwg.org/multipage/#dom-worklet-addmodule

Type Aliases

LazyCellWithBoxedInitializer 🔒

WorkletTask 🔒

A task which can be performed in the context of a WorkletGlobalScope.