script::dom::worklet::worklet

Trait WorkletThreadPool

pub trait WorkletThreadPool: JSTraceable {
    // Required methods
    fn fetch_and_invoke_a_worklet_script(
        &self,
        pipeline_id: PipelineId,
        worklet_id: WorkletId,
        global_type: WorkletGlobalScopeType,
        origin: ImmutableOrigin,
        base_url: ServoUrl,
        script_url: ServoUrl,
        policy_container: PolicyContainer,
        credentials: RequestCredentials,
        pending_tasks_struct: PendingTasksStruct,
        promise: &Rc<Promise>,
        inherited_secure_context: Option<bool>,
    );
    fn exit_worklet(&self, worklet_id: WorkletId);
    fn wake_threads(&self);
    fn perform_a_worklet_task(
        &self,
        worklet_id: WorkletId,
        worklet_task: Box<dyn FnOnce(&mut JSContext, &WorkletGlobalScope) + Send>,
    );
}

Required Methods

fn fetch_and_invoke_a_worklet_script( &self, pipeline_id: PipelineId, worklet_id: WorkletId, global_type: WorkletGlobalScopeType, origin: ImmutableOrigin, base_url: ServoUrl, script_url: ServoUrl, policy_container: PolicyContainer, credentials: RequestCredentials, pending_tasks_struct: PendingTasksStruct, promise: &Rc<Promise>, inherited_secure_context: Option<bool>, )

Loads a worklet module into every thread in this thread pool. If all of the threads load successfully, the promise is resolved. If any of the threads fails to load, the promise is rejected. NOTE: The method implements the Step 6 of AddModule https://html.spec.whatwg.org/multipage/#dom-worklet-addmodule

fn exit_worklet(&self, worklet_id: WorkletId)

Request that the WorkletGlobalScope associated with the WorkletId be removed from all the threads in the thread pool.

fn wake_threads(&self)

Signal all the threads in the pool that there may be control messages to process.

fn perform_a_worklet_task( &self, worklet_id: WorkletId, worklet_task: Box<dyn FnOnce(&mut JSContext, &WorkletGlobalScope) + Send>, )

Queue a WorkletTask for execution on this WorkletThreadPool. The task will be executed in the context of the WorkletGlobalScope represented by the [WorketId].