script::dom

Module writablestream

Structs

AbortAlgorithmFulfillmentHandler 🔒

The fulfillment handler for the abort steps of https://streams.spec.whatwg.org/#writable-stream-finish-erroring

AbortAlgorithmRejectionHandler 🔒

The rejection handler for the abort steps of https://streams.spec.whatwg.org/#writable-stream-finish-erroring

CrossRealmTransformWritable 🔒

https://streams.spec.whatwg.org/#abstract-opdef-setupcrossrealmtransformwritable A wrapper to handle message and messageerror events for the port used by the transfered stream.

PendingAbortRequest 🔒

https://streams.spec.whatwg.org/#pending-abort-request

WritableStream

https://streams.spec.whatwg.org/#ws-class

Enums

WritableStreamState 🔒

https://streams.spec.whatwg.org/#writablestream-state

Constants

_IMPL_DOMOBJECT_FOR_WritableStream 🔒

Functions

create_writable_stream 🔒

https://streams.spec.whatwg.org/#create-writable-stream