script::dom::writablestreamdefaultcontroller

Struct WritableStreamDefaultController

#[repr(C)]
pub struct WritableStreamDefaultController {
    reflector_: Reflector,
    underlying_sink_type: UnderlyingSinkType,
    underlying_sink_obj: Heap<*mut JSObject>,
    queue: QueueWithSizes,
    started: Cell<bool>,
    strategy_hwm: f64,
    strategy_size: RefCell<Option<Rc<QueuingStrategySize<DomTypeHolder>>>>,
    stream: MutNullableDom<WritableStream>,
    abort_controller: Dom<AbortController>,
}

https://streams.spec.whatwg.org/#ws-default-controller-class

Fields
reflector_: Reflectorunderlying_sink_type: UnderlyingSinkType

The type of underlying sink used. Besides the default JS one, there will be others for stream transfer, and for transform stream.

underlying_sink_obj: Heap<*mut JSObject>

The JS object used as this when invoking sink algorithms.

queue: QueueWithSizes

https://streams.spec.whatwg.org/#writablestreamdefaultcontroller-queue

started: Cell<bool>

https://streams.spec.whatwg.org/#writablestreamdefaultcontroller-started

strategy_hwm: f64

https://streams.spec.whatwg.org/#writablestreamdefaultcontroller-strategyhwm

strategy_size: RefCell<Option<Rc<QueuingStrategySize<DomTypeHolder>>>>

https://streams.spec.whatwg.org/#writablestreamdefaultcontroller-strategysizealgorithm

stream: MutNullableDom<WritableStream>

https://streams.spec.whatwg.org/#writablestreamdefaultcontroller-stream

abort_controller: Dom<AbortController>

https://streams.spec.whatwg.org/#writablestreamdefaultcontroller-abortcontroller

Implementations

impl WritableStreamDefaultController

fn __assert_parent_type(&self)

impl WritableStreamDefaultController

fn new_inherited( global: &GlobalScope, underlying_sink_type: UnderlyingSinkType, strategy_hwm: f64, strategy_size: Rc<QueuingStrategySize<DomTypeHolder>>, can_gc: CanGc, ) -> WritableStreamDefaultController

https://streams.spec.whatwg.org/#set-up-writable-stream-default-controller-from-underlying-sink

pub(crate) fn new( global: &GlobalScope, underlying_sink_type: UnderlyingSinkType, strategy_hwm: f64, strategy_size: Rc<QueuingStrategySize<DomTypeHolder>>, can_gc: CanGc, ) -> DomRoot<WritableStreamDefaultController>

pub(crate) fn started(&self) -> bool

pub(crate) fn set_underlying_sink_this_object( &self, this_object: SafeHandleObject<'_>, )

Setting the JS object after the heap has settled down.

pub(crate) fn signal_abort( &self, cx: SafeJSContext, reason: SafeHandleValue<'_>, realm: InRealm<'_>, can_gc: CanGc, )

“Signal abort” call from https://streams.spec.whatwg.org/#writable-stream-abort

fn clear_algorithms(&self)

https://streams.spec.whatwg.org/#writable-stream-default-controller-clear-algorithms

pub(crate) fn setup( &self, cx: SafeJSContext, global: &GlobalScope, stream: &WritableStream, can_gc: CanGc, ) -> Result<(), Error>

https://streams.spec.whatwg.org/#set-up-writable-stream-default-controller

pub(crate) fn close( &self, cx: SafeJSContext, global: &GlobalScope, can_gc: CanGc, )

https://streams.spec.whatwg.org/#writable-stream-default-controller-close

fn start_algorithm( &self, cx: SafeJSContext, global: &GlobalScope, can_gc: CanGc, ) -> Fallible<Rc<Promise>>

pub(crate) fn abort_steps( &self, cx: SafeJSContext, global: &GlobalScope, reason: SafeHandleValue<'_>, can_gc: CanGc, ) -> Rc<Promise>

https://streams.spec.whatwg.org/#ref-for-abstract-opdef-writablestreamcontroller-abortsteps

fn call_write_algorithm( &self, cx: SafeJSContext, chunk: SafeHandleValue<'_>, global: &GlobalScope, can_gc: CanGc, ) -> Rc<Promise>

https://streams.spec.whatwg.org/#writablestreamdefaultcontroller-writealgorithm

fn call_close_algorithm( &self, cx: SafeJSContext, global: &GlobalScope, can_gc: CanGc, ) -> Rc<Promise>

https://streams.spec.whatwg.org/#writablestreamdefaultcontroller-closealgorithm

pub(crate) fn process_close( &self, cx: SafeJSContext, global: &GlobalScope, can_gc: CanGc, )

https://streams.spec.whatwg.org/#writable-stream-default-controller-process-close

fn advance_queue_if_needed( &self, cx: SafeJSContext, global: &GlobalScope, can_gc: CanGc, )

https://streams.spec.whatwg.org/#writable-stream-default-controller-advance-queue-if-needed

pub(crate) fn perform_error_steps(&self)

https://streams.spec.whatwg.org/#ws-default-controller-private-error

fn process_write( &self, cx: SafeJSContext, chunk: SafeHandleValue<'_>, global: &GlobalScope, can_gc: CanGc, )

https://streams.spec.whatwg.org/#writable-stream-default-controller-process-write

pub(crate) fn get_desired_size(&self) -> f64

https://streams.spec.whatwg.org/#writable-stream-default-controller-get-desired-size

fn get_backpressure(&self) -> bool

https://streams.spec.whatwg.org/#writable-stream-default-controller-get-backpressure

pub(crate) fn get_chunk_size( &self, cx: SafeJSContext, global: &GlobalScope, chunk: SafeHandleValue<'_>, can_gc: CanGc, ) -> f64

https://streams.spec.whatwg.org/#writable-stream-default-controller-get-chunk-size

pub(crate) fn write( &self, cx: SafeJSContext, global: &GlobalScope, chunk: SafeHandleValue<'_>, chunk_size: f64, can_gc: CanGc, )

https://streams.spec.whatwg.org/#writable-stream-default-controller-write

pub(crate) fn error_if_needed( &self, cx: SafeJSContext, error: SafeHandleValue<'_>, global: &GlobalScope, can_gc: CanGc, )

https://streams.spec.whatwg.org/#writable-stream-default-controller-error-if-needed

pub(crate) fn error( &self, stream: &WritableStream, cx: SafeJSContext, e: SafeHandleValue<'_>, global: &GlobalScope, can_gc: CanGc, )

https://streams.spec.whatwg.org/#writable-stream-default-controller-error