Struct script::dom::xrsession::XRSession

#[repr(C)]
pub struct XRSession {
    eventtarget: EventTarget,
    base_layer: MutNullableDom<XRWebGLLayer>,
    blend_mode: XREnvironmentBlendMode,
    mode: XRSessionMode,
    visibility_state: Cell<XRVisibilityState>,
    viewer_space: MutNullableDom<XRSpace>,
    session: DomRefCell<Session>,
    frame_requested: Cell<bool>,
    pending_render_state: MutNullableDom<XRRenderState>,
    active_render_state: MutDom<XRRenderState>,
    inline_projection_matrix: DomRefCell<Transform3D<f32, Viewer, Display>>,
    next_raf_id: Cell<i32>,
    raf_callback_list: DomRefCell<Vec<(i32, Option<Rc<XRFrameRequestCallback>>)>>,
    input_sources: Dom<XRInputSourceArray>,
    end_promises: DomRefCell<Vec<Rc<Promise>>>,
    ended: Cell<bool>,
    outside_raf: Cell<bool>,
}

Fields
eventtarget: EventTargetbase_layer: MutNullableDom<XRWebGLLayer>blend_mode: XREnvironmentBlendModemode: XRSessionModevisibility_state: Cell<XRVisibilityState>viewer_space: MutNullableDom<XRSpace>session: DomRefCell<Session>frame_requested: Cell<bool>pending_render_state: MutNullableDom<XRRenderState>active_render_state: MutDom<XRRenderState>inline_projection_matrix: DomRefCell<Transform3D<f32, Viewer, Display>>

Cached projection matrix for inline sessions

next_raf_id: Cell<i32>raf_callback_list: DomRefCell<Vec<(i32, Option<Rc<XRFrameRequestCallback>>)>>input_sources: Dom<XRInputSourceArray>end_promises: DomRefCell<Vec<Rc<Promise>>>ended: Cell<bool>

https://immersive-web.github.io/webxr/#ended

outside_raf: Cell<bool>

Opaque framebuffers need to know the session is "outside of a requestAnimationFrame" https://immersive-web.github.io/webxr/#opaque-framebuffer

Methods

impl XRSession

fn __assert_parent_type(&self)

impl XRSession

fn new_inherited(
session: Session,
render_state: &XRRenderState,
input_sources: &XRInputSourceArray,
mode: XRSessionMode
) -> XRSession

pub fn new(
global: &GlobalScope,
session: Session,
mode: XRSessionMode,
frame_receiver: IpcReceiver<Frame>
) -> Root<Dom<XRSession>>

pub fn with_session<R, F: FnOnce(&Session) -> R>(&self, with: F) -> R

pub fn is_ended(&self) -> bool

pub fn is_immersive(&self) -> bool

fn setup_raf_loop(&self, frame_receiver: IpcReceiver<Frame>)

pub fn is_outside_raf(&self) -> bool

fn attach_event_handler(&self)

pub fn setup_initial_inputs(&self)

fn event_callback(&self, event: XREvent)

fn raf_callback(&self, frame: Frame)

https://immersive-web.github.io/webxr/#xr-animation-frame

fn update_inline_projection_matrix(&self)

pub fn inline_view(&self) -> View<Viewer>

Constructs a View suitable for inline sessions using the inlineVerticalFieldOfView and canvas size