script

Module event_loop

Modules

devtools 🔒

document_collection 🔒

document_loader 🔒

Tracking of pending loads in a document.

script_mutation_observers 🔒

script_thread 🔒

The script thread is the thread that owns the DOM in memory, runs JavaScript, and triggers layout. It’s in charge of processing events for all same-origin pages in a frame tree, and manages the entire lifetime of pages in the frame tree from initial request to teardown.

script_window_proxies 🔒

svg_font 🔒

timers 🔒

webdriver_handlers 🔒