script::event_loop

Module script_thread

The script thread is the thread that owns the DOM in memory, runs JavaScript, and triggers layout. It’s in charge of processing events for all same-origin pages in a frame tree, and manages the entire lifetime of pages in the frame tree from initial request to teardown.

Page loads follow a two-step process. When a request for a new page load is received, the network request is initiated and the relevant data pertaining to the new page is stashed. While the non-blocking request is ongoing, the script thread is free to process further events, noting when they pertain to ongoing loads (such as resizes/viewport adjustments). When the initial response is received for an ongoing load, the second phase starts - the frame tree entry is created, along with the Window and Document objects, and the appropriate parser takes over the response body. Once parsing is complete, the document lifecycle for loading a page runs its course and the script thread returns to processing events in the main event loop.

Structs

BHMExitSignal 🔒

IncompleteParserContexts 🔒

ScriptMemoryFailsafe 🔒

In the event of thread panic, all data on the stack runs its destructor. However, there are no reachable, owning pointers to the DOM memory, so it never gets freed by default when the script thread fails. The ScriptMemoryFailsafe uses the destructor bomb pattern to forcibly tear down the JS realms for pages associated with the failing ScriptThread.

ScriptThread

ScriptThreadUserContents 🔒

This is the ScriptThread’s version of UserContents with the difference that user stylesheets are represented as parsed DocumentStyleSheets instead of simple source strings.

ScriptUserInteractingGuard 🔒

A simple guard structure that restore the user interacting state when dropped

SharedRwLocks

Constants

SCRIPT_THREAD_ROOT 🔒

Functions

interrupt_callback 🔒 ⚠

obtain_a_browsing_context 🔒

https://html.spec.whatwg.org/multipage/#obtain-browsing-context-navigation

window_for_replacement 🔒

Steps 1, 5, and 6 of https://html.spec.whatwg.org/multipage/#initialise-the-document-object

with_optional_script_thread 🔒

with_script_thread 🔒

Type Aliases

NodeIdSet 🔒