script::fetch

Module body

Structs

ExtractedBody 🔒

https://fetch.spec.whatwg.org/#body-with-type

TransmitBodyConnectHandler 🔒

The IPC route handler for https://fetch.spec.whatwg.org/#concept-request-transmit-body. This route runs in the script process, and will queue tasks to perform operations on the stream and transmit body chunks over IPC.

TransmitBodyPromiseHandler 🔒

The handler of read promises of body streams used in https://fetch.spec.whatwg.org/#concept-request-transmit-body.

TransmitBodyPromiseRejectionHandler 🔒

The handler of read promises rejection of body streams used in https://fetch.spec.whatwg.org/#concept-request-transmit-body.

Enums

BodySource 🔒

The Dom object, or ReadableStream, that is the source of a body. https://fetch.spec.whatwg.org/#concept-body-source

BodyType 🔒

FetchedData 🔒

StopReading 🔒

The reason to stop reading from the body.

Traits

BodyMixin 🔒

https://fetch.spec.whatwg.org/#body

Extractable 🔒

https://fetch.spec.whatwg.org/#concept-bodyinit-extract

Functions

append_form_data_entry_from_part 🔒

append_multipart_nodes 🔒

body_text_stream 🔒

https://fetch.spec.whatwg.org/#dom-body-textstream

clone_body_stream_for_dom_body 🔒

https://fetch.spec.whatwg.org/#concept-body-clone

consume_body 🔒

https://fetch.spec.whatwg.org/#concept-body-consume-body https://fetch.spec.whatwg.org/#body-fully-read A combination of parts of both algorithms, body-fully-read can be fully implemented, and separated, later, see #36049.

content_type_from_headers 🔒

decode_to_utf16_with_bom_removal 🔒

extract_filename_from_content_disposition 🔒

extract_name_from_content_disposition 🔒

resolve_result_promise 🔒

The success steps of https://fetch.spec.whatwg.org/#concept-body-consume-body.

run_array_buffer_data_algorithm 🔒

https://fetch.spec.whatwg.org/#ref-for-concept-body-consume-body

run_blob_data_algorithm 🔒

https://fetch.spec.whatwg.org/#ref-for-concept-body-consume-body%E2%91%A0

run_bytes_data_algorithm 🔒

https://fetch.spec.whatwg.org/#ref-for-concept-body-consume-body%E2%91%A1

run_form_data_algorithm 🔒

https://fetch.spec.whatwg.org/#ref-for-concept-body-consume-body%E2%91%A2

run_json_data_algorithm 🔒

https://fetch.spec.whatwg.org/#ref-for-concept-body-consume-body%E2%91%A3

run_package_data_algorithm 🔒

The algorithm that takes a byte sequence and returns a JavaScript value or throws an exception of https://fetch.spec.whatwg.org/#concept-body-consume-body.

run_text_data_algorithm 🔒

https://fetch.spec.whatwg.org/#ref-for-concept-body-consume-body%E2%91%A4

stream_from_body_init_bytes 🔒

Part of https://fetch.spec.whatwg.org/#concept-bodyinit-extract