script::fetch

Module fetch

Structs

DeferredFetchRecord 🔒

https://fetch.spec.whatwg.org/#deferred-fetch-record

DeferredFetchRecordId 🔒

An id to differentiate one deferred fetch record from another.

FetchCanceller 🔒

Fetch canceller object. By default initialized to having a request associated with it, which can be aborted or terminated. Calling ignore will sever the relationship with the request, meaning it cannot be cancelled through this canceller from that point on.

FetchContext 🔒

FetchGroup 🔒

https://fetch.spec.whatwg.org/#concept-fetch-group

FetchLaterListener 🔒

FetchRecord 🔒

https://fetch.spec.whatwg.org/#fetch-record

Enums

DeferredFetchRecordInvokeState 🔒

https://fetch.spec.whatwg.org/#deferred-fetch-record-invoke-state

Traits

CspViolationsProcessor 🔒

RequestWithGlobalScope 🔒

Functions

Fetch 🔒

https://fetch.spec.whatwg.org/#dom-global-fetch

FetchLater 🔒

https://fetch.spec.whatwg.org/#dom-window-fetchlater

abort_fetch_call 🔒

https://fetch.spec.whatwg.org/#abort-fetch

create_a_potential_cors_request 🔒

https://html.spec.whatwg.org/multipage/#create-a-potential-cors-request This function is temporary, since it does not ensure that blob URLs are claimed appropriately. All callers must migrate to create_a_potential_cors_request_with_claim.

create_a_potential_cors_request_with_claim 🔒

https://html.spec.whatwg.org/multipage/#create-a-potential-cors-request

fill_headers_with_metadata 🔒

load_whole_resource 🔒

Convenience function for synchronously loading a whole resource.

queue_deferred_fetch 🔒

https://fetch.spec.whatwg.org/#queue-a-deferred-fetch

request_init_from_request 🔒

Type Aliases

QueuedDeferredFetchRecord 🔒