script::fetch::network_listener

Trait FetchResponseListener

pub(crate) trait FetchResponseListener: Send + 'static {
    // Required methods
    fn process_request_body(&mut self, request_id: RequestId);
    fn process_response(
        &mut self,
        cx: &mut JSContext,
        request_id: RequestId,
        metadata: Result<FetchMetadata, NetworkError>,
    );
    fn process_response_chunk(
        &mut self,
        cx: &mut JSContext,
        request_id: RequestId,
        chunk: Bytes,
    );
    fn process_response_eof(
        self,
        cx: &mut JSContext,
        request_id: RequestId,
        response: Result<(), NetworkError>,
        timing: ResourceFetchTiming,
    );
    fn process_csp_violations(
        &mut self,
        cx: &mut JSContext,
        request_id: RequestId,
        violations: Vec<Violation>,
    );

    // Provided methods
    fn should_invoke(&self) -> bool { ... }
    fn process_content_length(&mut self, _request_id: RequestId, _size: usize) { ... }
}

Required Methods

fn process_request_body(&mut self, request_id: RequestId)

fn process_response( &mut self, cx: &mut JSContext, request_id: RequestId, metadata: Result<FetchMetadata, NetworkError>, )

fn process_response_chunk( &mut self, cx: &mut JSContext, request_id: RequestId, chunk: Bytes, )

fn process_response_eof( self, cx: &mut JSContext, request_id: RequestId, response: Result<(), NetworkError>, timing: ResourceFetchTiming, )

fn process_csp_violations( &mut self, cx: &mut JSContext, request_id: RequestId, violations: Vec<Violation>, )

Provided Methods

fn should_invoke(&self) -> bool

A gating mechanism that runs before invoking the listener methods on the target thread. If the should_invoke method returns false, the listener does not receive the notification.

fn process_content_length(&mut self, _request_id: RequestId, _size: usize)