Struct script::layout_dom::element::ServoLayoutElement

pub struct ServoLayoutElement<'dom> {
    element: LayoutDom<'dom, Element>,
}

A wrapper around elements that ensures layout can only ever access safe properties.

Fields
element: LayoutDom<'dom, Element>

The wrapped private DOM Element.

Implementations

impl<'dom> ServoLayoutElement<'dom>

pub(super) fn from_layout_js(el: LayoutDom<'dom, Element>) -> Self

pub(super) fn is_html_element(&self) -> bool

fn get_attr_enum( &self, namespace: &Namespace, name: &LocalName, ) -> Option<&AttrValue>

fn get_attr(&self, namespace: &Namespace, name: &LocalName) -> Option<&str>

fn get_style_data(&self) -> Option<&StyleData>

pub unsafe fn unset_snapshot_flags(&self)

Unset the snapshot flags on the underlying DOM object for this element.

Safety

This function accesses and modifies the underlying DOM object and should not be used by more than a single thread at once.

pub unsafe fn set_has_snapshot(&self)

Unset the snapshot flags on the underlying DOM object for this element.

Safety

This function accesses and modifies the underlying DOM object and should not be used by more than a single thread at once.

fn is_body_element_of_html_element_root(&self) -> bool

Returns true if this element is the body child of an html element root element.

fn parent_element(&self) -> Option<Self>

Returns the parent element of this element, if it has one.

fn is_root(&self) -> bool