script::layout_dom::element

Struct ServoThreadSafeLayoutElement

pub struct ServoThreadSafeLayoutElement<'dom> {
    pub(super) element: ServoLayoutElement<'dom>,
    pub(super) pseudo_element_chain: PseudoElementChain,
}

A wrapper around elements that ensures layout can only ever access safe properties and cannot race on elements.

Fields
element: ServoLayoutElement<'dom>

The wrapped ServoLayoutElement.

pseudo_element_chain: PseudoElementChain

The possibly nested PseudoElementChain for this element.

Implementations

impl<'dom> ServoThreadSafeLayoutElement<'dom>

pub fn shadow_root(&self) -> Option<ServoShadowRoot<'dom>>

The shadow root this element is a host of.

pub fn slotted_nodes(&self) -> &[ServoLayoutNode<'dom>]