Struct script::layout_dom::ServoLayoutElement

pub struct ServoLayoutElement<'dom, LayoutDataType: LayoutDataTrait> {
    element: LayoutDom<'dom, Element>,
    phantom: PhantomData<LayoutDataType>,
}

A wrapper around elements that ensures layout can only ever access safe properties.

Fields
element: LayoutDom<'dom, Element>

The wrapped private DOM Element.

phantom: PhantomData<LayoutDataType>

A PhantomData that is used to track the type of the stored layout data.

Implementations

impl<'dom, LayoutDataType: LayoutDataTrait> ServoLayoutElement<'dom, LayoutDataType>

pub(super) fn from_layout_js(el: LayoutDom<'dom, Element>) -> Self

fn get_attr_enum( &self, namespace: &Namespace, name: &LocalName ) -> Option<&AttrValue>

fn get_attr(&self, namespace: &Namespace, name: &LocalName) -> Option<&str>

fn get_style_data(&self) -> Option<&StyleData>

pub unsafe fn unset_snapshot_flags(&self)

pub unsafe fn set_has_snapshot(&self)

fn is_body_element_of_html_element_root(&self) -> bool

Returns true if this element is the body child of an html element root element.

fn parent_element(&self) -> Option<Self>

Returns the parent element of this element, if it has one.

fn is_root(&self) -> bool