script::layout_dom

Struct ServoLayoutNode

pub struct ServoLayoutNode<'dom> {
    pub(super) node: LayoutDom<'dom, Node>,
    pub(super) pseudo_element_chain: PseudoElementChain,
}

A wrapper around a LayoutDom<Node> which provides a safe interface that can be used during layout. This implements the LayoutNode trait.

Fields
node: LayoutDom<'dom, Node>

The wrapped private DOM node.

pseudo_element_chain: PseudoElementChain

The possibly nested PseudoElementChain for this node.

Implementations

impl<'dom> ServoLayoutNode<'dom>

pub unsafe fn new(address: &TrustedNodeAddress) -> Self

Create a new ServoLayoutNode for this given TrustedNodeAddress.

Safety

The address pointed to by address should point to a valid node in memory.

pub(super) unsafe fn dangerous_first_child(&self) -> Option<Self>

Get the first child of this node.

Safety

This node should never be exposed directly to the layout interface, as that may allow mutating a node that is being laid out in another thread. Thus, this should never be made public or exposed in the LayoutNode trait.

pub(super) unsafe fn dangerous_next_sibling(&self) -> Option<Self>

Get the next sibling of this node.

Safety

This node should never be exposed directly to the layout interface, as that may allow mutating a node that is being laid out in another thread. Thus, this should never be made public or exposed in the LayoutNode trait.

pub(super) unsafe fn dangerous_previous_sibling(&self) -> Option<Self>

Get the previous sibling of this node.

Safety

This node should never be exposed directly to the layout interface, as that may allow mutating a node that is being laid out in another thread. Thus, this should never be made public or exposed in the LayoutNode trait.