Struct script::layout_exports::LayoutDom

pub struct LayoutDom<T> {
    ptr: NonNull<T>,
}

An unrooted reference to a DOM object for use in layout. Layout*Helpers traits must be implemented on this.

Fields
ptr: NonNull<T>
Methods

impl<T: Castable> LayoutDom<T>

pub fn upcast<U>(&self) -> LayoutDom<U> where
U: Castable,
T: DerivedFrom<U>,

Cast a DOM object root upwards to one of the interfaces it derives from.

pub fn downcast<U>(&self) -> Option<LayoutDom<U>> where
U: DerivedFrom<T>,

Cast a DOM object downwards to one of the interfaces it might implement.

impl<T: DomObject> LayoutDom<T>

pub unsafe fn get_jsobject(&self) -> *mut JSObject

Get the reflector.

impl LayoutDom<Node>

pub unsafe fn from_trusted_node_address(
inner: TrustedNodeAddress
) -> LayoutDom<Node>

Create a new JS-owned value wrapped from an address known to be a Node pointer.

impl<T: DomObject> LayoutDom<T>

pub unsafe fn unsafe_get(&self) -> *const T

Returns an unsafe pointer to the interior of this JS object. This is the only method that be safely accessed from layout. (The fact that this is unsafe is what necessitates the layout wrappers.)