Struct script::layout_exports::ShadowRoot

#[repr(C)]pub struct ShadowRoot {
    document_fragment: DocumentFragment,
    document_or_shadow_root: DocumentOrShadowRoot,
    document: Dom<Document>,
    host: MutNullableDom<Element>,
    author_styles: DomRefCell<AuthorStyles<StyleSheetInDocument>>,
    stylesheet_list: MutNullableDom<StyleSheetList>,
    window: Dom<Window>,
}

Fields
document_fragment: DocumentFragmentdocument_or_shadow_root: DocumentOrShadowRootdocument: Dom<Document>host: MutNullableDom<Element>author_styles: DomRefCell<AuthorStyles<StyleSheetInDocument>>

List of author styles associated with nodes in this shadow tree.

stylesheet_list: MutNullableDom<StyleSheetList>window: Dom<Window>
Implementations

impl ShadowRoot

fn __assert_parent_type(&self)

impl ShadowRoot

fn new_inherited(host: &Element, document: &Document) -> ShadowRoot

pub fn new(host: &Element, document: &Document) -> Root<Dom<ShadowRoot>>

pub fn detach(&self)

pub fn get_focused_element(&self) -> Option<Root<Dom<Element>>>

pub fn stylesheet_count(&self) -> usize

pub fn stylesheet_at(&self, index: usize) -> Option<Root<Dom<CSSStyleSheet>>>

pub fn add_stylesheet(&self, owner: &Element, sheet: Arc<Stylesheet>)

Add a stylesheet owned by owner to the list of shadow root sheets, in the correct tree position.

pub fn remove_stylesheet(&self, owner: &Element, s: &Arc<Stylesheet>)

Remove a stylesheet owned by owner from the list of shadow root sheets.

pub fn invalidate_stylesheets(&self)

pub fn unregister_element_id(&self, to_unregister: &Element, id: Atom)

Remove any existing association between the provided id and any elements in this shadow tree.

pub fn register_element_id(&self, element: &Element, id: Atom)

Associate an element present in this shadow tree with the provided id.