Module script::layout_interface

The high-level interface from script to layout. Using this abstract interface helps reduce coupling between these two components, and enables the DOM to be placed in a separate crate from layout.

Reexports

pub use dom::node::TrustedNodeAddress;

Structs

Animation
State relating to an animation.

ContentBoxResponse

ContentBoxesResponse

HitTestResponse

LayoutChan
Encapsulates a channel to the layout task.

MouseOverResponse

Reflow
Information needed for a reflow.

ScriptReflow
Information needed for a script-initiated reflow.

Enums

Msg
Asynchronous messages that script can send to layout.

ReflowGoal
Why we're doing reflow.

ReflowQueryType
Any query to perform with this reflow.

Traits

LayoutRPC
Synchronous messages that script can send to layout.

ScriptLayoutChan
A trait to manage opaque references to script<->layout channels without needing to expose the message type to crates that don't need to know about them.

Type Definitions

OpaqueNode
Type of an opaque node.