Enum script::layout_interface::Msg

pub enum Msg {
    AddStylesheet(Stylesheet, MediaQueryList),
    LoadStylesheet(Url, MediaQueryList),
    SetQuirksMode,
    Reflow(Box<ScriptReflow>),
    GetRPC(Sender<Box<LayoutRPC + Send>>),
    TickAnimations,
    ReapLayoutData(LayoutData),
    CollectReports(ReportsChan),
    PrepareToExit(Sender<()>),
    ExitNow(PipelineExitType),
}

Asynchronous messages that script can send to layout.

Variants

AddStylesheet

Adds the given stylesheet to the document.

LoadStylesheet

Adds the given stylesheet to the document.

SetQuirksMode

Puts a document into quirks mode, causing the quirks mode stylesheet to be loaded.

Reflow

Requests a reflow.

GetRPC

Get an RPC interface.

TickAnimations

Requests that the layout task render the next frame of all animations.

ReapLayoutData

Destroys layout data associated with a DOM node.

TODO(pcwalton): Maybe think about batching to avoid message traffic.

CollectReports

Requests that the layout task measure its memory usage. The resulting reports are sent back via the supplied channel.

PrepareToExit

Requests that the layout task enter a quiescent state in which no more messages are accepted except ExitMsg. A response message will be sent on the supplied channel when this happens.

ExitNow

Requests that the layout task immediately shut down. There must be no more nodes left after this, or layout will crash.