Struct script::layout_interface::Reflow

pub struct Reflow {
    pub document_root: TrustedNodeAddress,
    pub goal: ReflowGoal,
    pub url: Url,
    pub iframe: bool,
    pub script_chan: ScriptControlChan,
    pub window_size: WindowSizeData,
    pub script_join_chan: Sender<()>,
    pub id: usize,
    pub query_type: ReflowQueryType,
    pub page_clip_rect: Rect<Au>,
}

Information needed for a reflow.

Fields

document_root

The document node.

goal

The goal of reflow: either to render to the screen or to flush layout info for script.

url

The URL of the page.

iframe

Is the current reflow of an iframe, as opposed to a root window?

script_chan

The channel through which messages can be sent back to the script task.

window_size

The current window size.

script_join_chan

The channel that we send a notification to.

id

Unique identifier

query_type

The type of query if any to perform during this reflow.

page_clip_rect

A clipping rectangle for the page, an enlarged rectangle containing the viewport.