Struct script::layout_interface::ScriptReflow

pub struct ScriptReflow {
    pub reflow_info: Reflow,
    pub document_root: TrustedNodeAddress,
    pub script_chan: ScriptControlChan,
    pub window_size: WindowSizeData,
    pub script_join_chan: Sender<()>,
    pub id: u32,
    pub query_type: ReflowQueryType,
}

Information needed for a script-initiated reflow.

Fields

reflow_info

General reflow data.

document_root

The document node.

script_chan

The channel through which messages can be sent back to the script task.

window_size

The current window size.

script_join_chan

The channel that we send a notification to.

id

Unique identifier

query_type

The type of query if any to perform during this reflow.