Trait script::layout_interface::LayoutRPC

pub trait LayoutRPC {
    fn content_box(&self) -> ContentBoxResponse;
    fn content_boxes(&self) -> ContentBoxesResponse;
    fn hit_test(&self, node: TrustedNodeAddress, point: Point2D<f32>) -> Result<HitTestResponse, ()>;
    fn mouse_over(&self, node: TrustedNodeAddress, point: Point2D<f32>) -> Result<MouseOverResponse, ()>;
}

Synchronous messages that script can send to layout.

In general, you should use messages to talk to Layout. Use the RPC interface if and only if the work is

1) read-only with respect to LayoutTaskData, 2) small, 3) and really needs to be fast.

Required Methods

fn content_box(&self) -> ContentBoxResponse

Requests the dimensions of the content box, as in the getBoundingClientRect() call.

fn content_boxes(&self) -> ContentBoxesResponse

Requests the dimensions of all the content boxes, as in the getClientRects() call.

fn hit_test(&self, node: TrustedNodeAddress, point: Point2D<f32>) -> Result<HitTestResponse, ()>

Requests the node containing the point of interest

fn mouse_over(&self, node: TrustedNodeAddress, point: Point2D<f32>) -> Result<MouseOverResponse, ()>