Struct script::layout_wrapper::ServoThreadSafeLayoutElement

pub struct ServoThreadSafeLayoutElement<'le> {
    element: ServoLayoutElement<'le>,
    pseudo: PseudoElementType<Option<T>>,
}

A wrapper around elements that ensures layout can only ever access safe properties and cannot race on elements.

Fields
element: ServoLayoutElement<'le> pseudo: PseudoElementType<Option<T>>

The pseudo-element type, with (optionally) a specified display value to override the stylesheet.