script::modules

Module script_module

The script module mod contains common traits and structs related to type=module for script thread or worker threads.

Structs

ModuleContext 🔒

The context required for asynchronously loading an external module script source.

ModuleHandler 🔒

ModuleObject 🔒

ModuleScript 🔒

ModuleTree 🔒

ResolvedModule 🔒

https://html.spec.whatwg.org/multipage/#specifier-resolution-record

RethrowError 🔒

ScriptFetchOptions 🔒

https://html.spec.whatwg.org/multipage/#script-fetch-options

Enums

ModuleStatus 🔒

Constants

MODULE_RECORD_SLOT 🔒

https://searchfox.org/firefox-main/rev/9a8a80db6ce10ffc2fc91a1e25685eed59ce3501/js/loader/ModuleLoaderBase.h#597

Functions

EnsureModuleHooksInitialized 🔒 ⚠

A function to register module hooks (e.g. listening on resolving modules, getting module metadata, getting script private reference and resolving dynamic import)

HostLoadImportedModule 🔒 ⚠

https://tc39.es/ecma262/#sec-HostLoadImportedModule https://html.spec.whatwg.org/multipage/#hostloadimportedmodule

HostPopulateImportMeta 🔒 ⚠

https://tc39.es/ecma262/#sec-hostgetimportmetaproperties https://html.spec.whatwg.org/multipage/#hostgetimportmetaproperties

fetch_a_module_script_graph 🔒

https://html.spec.whatwg.org/multipage/#fetch-a-module-worker-script-tree https://html.spec.whatwg.org/multipage/#fetch-a-worklet/module-worker-script-graph

fetch_a_modulepreload_module 🔒

https://html.spec.whatwg.org/multipage/#fetch-a-modulepreload-module-script-graph

fetch_a_single_module_script 🔒

https://html.spec.whatwg.org/multipage/#fetch-a-single-module-script

fetch_an_external_module_script 🔒

https://html.spec.whatwg.org/multipage/#fetch-a-module-script-tree

fetch_inline_module_script 🔒

https://html.spec.whatwg.org/multipage/#fetch-an-inline-module-script-graph

fetch_the_descendants_and_link_module_script 🔒

https://html.spec.whatwg.org/multipage/#fetch-the-descendants-of-and-link-a-module-script

gen_type_error 🔒

host_add_ref_top_level_script 🔒 ⚠

host_release_top_level_script 🔒 ⚠

import_meta_resolve 🔒 ⚠

module_script_from_reference_private 🔒 ⚠

resolve_imports_match 🔒

https://html.spec.whatwg.org/multipage/#resolving-an-imports-match

Type Aliases

ModuleRequest 🔒