script

Module navigation

The listener that encapsulates all state for an in-progress document request. Any redirects that are encountered are followed. Whenever a non-redirect response is received, it is forwarded to the appropriate script thread.

Structs

InProgressLoad 🔒

A document load that is in the process of fetching the requested resource. Contains data that will need to be present when the document and frame tree entry are created, but is only easily available at initiation of the load and on a push basis (so some data will be updated according to future resize events, viewport changes, etc.)

NavigationListener

Functions

determine_creation_sandboxing_flags 🔒

https://html.spec.whatwg.org/multipage/#determining-the-creation-sandboxing-flags

determine_iframe_element_referrer_policy 🔒

https://html.spec.whatwg.org/multipage/#determining-the-iframe-element-referrer-policy

determine_the_origin 🔒

https://html.spec.whatwg.org/multipage/#determining-the-origin

navigate 🔒

https://html.spec.whatwg.org/multipage/#navigate

navigate_to_fragment 🔒

https://html.spec.whatwg.org/multipage/#navigate-fragid

snapshot_target_snapshot_params 🔒

https://html.spec.whatwg.org/multipage/#snapshotting-target-snapshot-params