Struct script::network_listener::NetworkListener

pub struct NetworkListener<Listener: PreInvoke + Send + 'static> {
    pub context: Arc<Mutex<Listener>>,
    pub task_source: NetworkingTaskSource,
    pub wrapper: Option<RunnableWrapper>,
}

An off-thread sink for async network event runnables. All such events are forwarded to a target thread, where they are invoked on the provided context object.

Fields
context: Arc<Mutex<Listener>> task_source: NetworkingTaskSource wrapper: Option<RunnableWrapper>
Methods

impl<Listener: PreInvoke + Send + 'static> NetworkListener<Listener>

fn notify<A: Action<Listener> + Send + 'static>(&self, action: A)

impl<Listener: FetchResponseListener + PreInvoke + Send + 'static> NetworkListener<Listener>

fn notify_fetch(&self, action: FetchResponseMsg)