script::python::ffi_types
Module copr

Modules

compile

compile script to code object

parse

Structs

AnnotationInfo

Coprocessor

PyQueryEngine

QueryEngineWeakRef

Enums

BackendType

Either 🔒

Functions

check_args_anno_real_type 🔒

match between arguments’ real type and annotation types if type anno is vector[_] then use real type(from RecordBatch’s schema)

exec_parsed

using a parsed Coprocessor struct as input to execute python code

select_from_rb 🔒

select columns according to fetch_names from rb and cast them into a Vec of PyVector

Type Aliases

CoprocessorRef