Module script::python::ffi_types::copr::parse

Structs

DecoratorArgs

Functions

check_annotation_ret_slice 🔒

check for the grammar correctness of annotation, also return the slice of subscript for further parsing

check_bin_op 🔒

check if binary op expr is legal(with one typename and one None)

get_arg_annotations 🔒

get_return_annotations 🔒

parse_and_compile_copr

parse script and return Coprocessor struct with info extract from ast

parse_annotation 🔒

where:

parse_bin_op 🔒

parse a DataType | None or a single DataType

parse_decorator 🔒

returns args and returns in Vec of String

parse_keywords 🔒

parse a list of keyword and return args and returns list from keywords

parse_native_type 🔒

Item => NativeType default to be not nullable

py_str_to_string 🔒

pylist_to_vec 🔒

turn a python list of string in ast form(a ast::Expr) of string into a Vec<String>

ret_parse_error 🔒

Return a CoprParseSnafu for you to chain fail() to return correct err Result type

try_into_datatype 🔒