Struct script::python::ffi_types::vector::PyVector

#[repr(transparent)]
pub struct PyVector {
    pub(crate) vector: VectorRef,
}

The Main FFI type PyVector that is used both in RustPython and PyO3

Fields
vector: VectorRef
Implementations

impl PyVector

pub(crate) fn vector_and(left: &Self, right: &Self) -> Result<Self, String>

pub(crate) fn vector_or(left: &Self, right: &Self) -> Result<Self, String>

pub(crate) fn vector_invert(left: &Self) -> Result<Self, String>

pub fn as_vector_ref(&self) -> VectorRef

create a ref to inner vector

pub fn to_arrow_array(&self) -> ArrayRef

pub(crate) fn scalar_arith_op<F>( &self, right: Value, target_type: Option<ArrowDataType>, op: F ) -> Result<Self, String>where F: Fn(&dyn Array, &dyn Array) -> Result<ArrayRef, String>,

pub(crate) fn rspy_scalar_arith_op<F>( &self, other: PyObjectRef, target_type: Option<ArrowDataType>, op: F, vm: &VirtualMachine ) -> PyResult<PyVector>where F: Fn(&dyn Array, &dyn Array) -> Result<ArrayRef, String>,

fn coerce_types( left_type: &ArrowDataType, right_type: &ArrowDataType, target_type: &Option<ArrowDataType> ) -> ArrowDataType

Returns the type that should be used for the result of an arithmetic operation

pub(crate) fn vector_arith_op<F>( &self, right: &Self, target_type: Option<ArrowDataType>, op: F ) -> Result<PyVector, String>where F: Fn(&dyn Array, &dyn Array) -> Result<ArrayRef, String>,

pub(crate) fn rspy_vector_arith_op<F>( &self, other: PyObjectRef, target_type: Option<ArrowDataType>, op: F, vm: &VirtualMachine ) -> PyResult<PyVector>where F: Fn(&dyn Array, &dyn Array) -> Result<ArrayRef, String>,

pub(crate) fn _getitem( &self, needle: &PyObject, vm: &VirtualMachine ) -> PyResult<PyObjectRef>

pub(crate) fn getitem_by_index( &self, i: isize, vm: &VirtualMachine ) -> PyResult<PyObjectRef>

fn getitem_by_slice( &self, slice: &SaturatedSlice, vm: &VirtualMachine ) -> PyResult<PyObjectRef>

Return a PyVector in PyObjectRef

fn setitem_by_index( zelf: PyRef<Self>, i: isize, value: PyObjectRef, vm: &VirtualMachine ) -> PyResult<()>

Unsupported TODO(discord9): make it work

pub(crate) fn richcompare( &self, other: PyObjectRef, op: PyComparisonOp, vm: &VirtualMachine ) -> PyResult<PyVector>

rich compare, return a boolean array, accept type are vec and vec and vec and number

pub(crate) fn len(&self) -> usize

impl PyVector

PyVector type wraps a greptime vector, impl multiply/div/add/sub opeerators etc.

pub(crate) fn new( iterable: OptionalArg<PyObjectRef>, vm: &VirtualMachine ) -> PyResult<PyVector>

fn add(&self, other: PyObjectRef, vm: &VirtualMachine) -> PyResult<PyVector>

fn sub(&self, other: PyObjectRef, vm: &VirtualMachine) -> PyResult<PyVector>

fn rsub(&self, other: PyObjectRef, vm: &VirtualMachine) -> PyResult<PyVector>

fn mul(&self, other: PyObjectRef, vm: &VirtualMachine) -> PyResult<PyVector>

fn truediv(&self, other: PyObjectRef, vm: &VirtualMachine) -> PyResult<PyVector>

fn rtruediv(&self, other: PyObjectRef, vm: &VirtualMachine) -> PyResult<PyVector>

fn floordiv(&self, other: PyObjectRef, vm: &VirtualMachine) -> PyResult<PyVector>

fn rfloordiv(&self, other: PyObjectRef, vm: &VirtualMachine) -> PyResult<PyVector>

fn and(&self, other: PyRef<PyVector>, vm: &VirtualMachine) -> PyResult<PyVector>

fn or(&self, other: PyRef<PyVector>, vm: &VirtualMachine) -> PyResult<PyVector>

fn invert(&self, vm: &VirtualMachine) -> PyResult<PyVector>

fn len_rspy(&self) -> usize

fn concat( &self, other: PyRef<PyVector>, vm: &VirtualMachine ) -> PyResult<PyVector>

fn filter( &self, other: PyRef<PyVector>, vm: &VirtualMachine ) -> PyResult<PyVector>

take a boolean array and filters the Array, returning elements matching the filter (i.e. where the values are true).

fn doc(&self) -> PyResult<PyStr>