Module script::python::rspython::builtins::greptime_builtin

GrepTime User Define Function module

allow Python Coprocessor Function to use already implemented udf functions from datafusion and GrepTime DB itself

Structs

PyDataFrame

Constants

DOC 🔒

MODULE_NAME 🔒

Functions

abs 🔒

simple math function, the backing implement is datafusion’s abs math function

acos 🔒

simple math function, the backing implement is datafusion’s acos math function

approx_distinct 🔒

approx_percentile_cont 🔒

Not implement in datafusion TODO(discord9): use greptime’s own impl instead

argmax 🔒

argmin 🔒

array_agg 🔒

effectively equals to list(vector)

asin 🔒

simple math function, the backing implement is datafusion’s asin math function

atan 🔒

simple math function, the backing implement is datafusion’s atan math function

avg 🔒

directly port from datafusion’s avg function

ceil 🔒

simple math function, the backing implement is datafusion’s ceil math function

clip 🔒

col 🔒

correlation 🔒

cos 🔒

simple math function, the backing implement is datafusion’s cos math function

count 🔒

dataframe 🔒

get __dataframe__ from globals and return it TODO(discord9): this is a terrible hack, we should find a better way to get __dataframe__

datetime 🔒

diff 🔒

eval_aggr_func 🔒

eval_func 🔒

exp 🔒

simple math function, the backing implement is datafusion’s exp math function

extend_module 🔒

first 🔒

return first element in a PyVector in sliced new PyVector, if vector’s length is zero, return a zero sized slice instead

floor 🔒

simple math function, the backing implement is datafusion’s floor math function

gen_inteveral 🔒

generate interval time point

gen_none_array 🔒

interval 🔒

func: exec on sliding window slice of given arr, expect it to always return a PyVector of one element ts: a vector of time stamp, expect to be Monotonous increase arr: actual data vector duration: the size of sliding window, also is the default step of sliding window’s per step

last 🔒

return last element in a PyVector in sliced new PyVector, if vector’s length is zero, return a zero sized slice instead

lit 🔒

ln 🔒

simple math function, the backing implement is datafusion’s ln math function

log2 🔒

simple math function, the backing implement is datafusion’s log2 math function

log10 🔒

simple math function, the backing implement is datafusion’s log10 math function

make_module 🔒

max 🔒

mean 🔒

min 🔒

next 🔒

percentile 🔒

polyval 🔒

pow 🔒

Pow function, bind from gp’s [PowFunction]

pow_gp 🔒

GrepTime’s own impl of pow function

prev 🔒

query 🔒

get __query__ from globals and return it TODO(discord9): this is a terrible hack, we should find a better way to get __query__

random 🔒

return a random vector range from 0 to 1 and length of len

round 🔒

simple math function, the backing implement is datafusion’s round math function

scipy_stats_norm_cdf 🔒

scipy_stats_norm_pdf 🔒

signum 🔒

simple math function, the backing implement is datafusion’s signum math function

sin 🔒

simple math function, the backing implement is datafusion’s sin math function

sqrt 🔒

simple math function, the backing implement is datafusion’s sqrt math function

stddev 🔒

stddev_pop 🔒

sum 🔒

tan 🔒

simple math function, the backing implement is datafusion’s tan math function

variance 🔒

variance_pop 🔒

vector 🔒