Module script::python::utils

Functions

block_on_async

Please only use this method because you are calling from (optionally first as async) to sync then to a async a terrible hack to call async from sync by:

format_py_error

spawn_blocking_script

just like [tokio::task::spawn_blocking] but using a dedicated runtime(runtime bg) using by scripts crate