script::runtime

Module job_queue

Implementation of microtasks and microtask queues. It is up to implementations of event loops to store a queue and perform checkpoints at appropriate times, as well as enqueue microtasks as required.

Structs

CustomElementReactionMicrotask

JobQueue 🔒

NotifyMutationObserversMicrotask

UserMicrotask 🔒

A microtask that comes from a queueMicrotask() Javascript call

Statics

JOB_QUEUE_TRAPS 🔒

Traits

MicrotaskRunnable 🔒

Functions

enqueue 🔒

Add a new microtask to this queue. It will be invoked as part of the next microtask checkpoint.

get_host_defined_data 🔒 ⚠

https://searchfox.org/firefox-main/rev/446c6e609dbd7c355c2fb27209dfe4833211991f/xpcom/base/CycleCollectedJSContext.cpp#229

get_host_defined_global 🔒 ⚠

https://searchfox.org/firefox-main/rev/446c6e609dbd7c355c2fb27209dfe4833211991f/xpcom/base/CycleCollectedJSContext.cpp#199

job_queue_clear 🔒

job_queue_microtask_checkpoint 🔒

https://html.spec.whatwg.org/multipage/#perform-a-microtask-checkpoint Perform a microtask checkpoint, executing all queued microtasks until the queue is empty.

microtask_from_jsval 🔒

run_jobs 🔒 ⚠

trace_non_gc_things_micro_task 🔒 ⚠