script::runtime

Module script_runtime

The script runtime contains common traits and structs commonly used by the script thread, the dom, and the worker threads.

Structs

GlobalSizeData 🔒

InterfaceSizeData 🔒

IntroductionType 🔒

introductionType values in SpiderMonkey TransitiveCompileOptions.

PerGlobalInterfaceSizes 🔒

A map of globals to size information for those globals. The key is the global’s JS reflector pointer as an integer.

Runnable 🔒

Runtime 🔒

RuntimeCallbackData 🔒

Data that is sent to SpiderMonkey runtime callbacks as a pointer, which allows access to the Runtime state.

StreamConsumer 🔒

ThreadSafeJSContext 🔒

A version of the JSContext that can be used from other threads and is thus Send and Sync. This should only ever expose operations that are marked as thread-safe by the SpiderMonkey API, ie ones that only atomic fields in JSContext.

Enums

ScriptThreadEventCategory 🔒

Constants

ALREADY_COMPUTED_OBJECTS 🔒

DOM_OBJECT_SIZES 🔒

GC_CYCLE_START 🔒

GC_SLICE_START 🔒

MALLOC_SIZE_OF_OPS 🔒

Statics

JOB_QUEUE_TRAPS 🔒

SECURITY_CALLBACKS 🔒

Functions

code_for_eval_gets 🔒 ⚠

compute_size 🔒

consume_stream 🔒 ⚠

Implements the steps to compile webassembly response mentioned here https://webassembly.github.io/spec/web-api/#compile-a-potential-webassembly-response

content_security_policy_allows 🔒 ⚠

debug_gc_callback 🔒 ⚠

drop_interrupt_queues 🔒 ⚠

gc_slice_callback 🔒 ⚠

get_host_defined_data 🔒 ⚠

https://searchfox.org/firefox-main/rev/446c6e609dbd7c355c2fb27209dfe4833211991f/xpcom/base/CycleCollectedJSContext.cpp#229

get_host_defined_global 🔒 ⚠

https://searchfox.org/firefox-main/rev/446c6e609dbd7c355c2fb27209dfe4833211991f/xpcom/base/CycleCollectedJSContext.cpp#199

get_reports 🔒

get_size 🔒 ⚠

in_range 🔒

invoke_script_environment_preparer 🔒 ⚠

notify_about_rejected_promises 🔒

https://html.spec.whatwg.org/multipage/#notify-about-rejected-promises

pop_interrupt_queue 🔒 ⚠

promise_rejection_tracker 🔒 ⚠

https://html.spec.whatwg.org/multipage/#the-hostpromiserejectiontracker-implementation

push_new_interrupt_queue 🔒 ⚠

report_stream_error 🔒 ⚠

run_jobs 🔒 ⚠

safely_convert_null_to_string 🔒

servo_build_id 🔒 ⚠

set_gc_zeal_options 🔒 ⚠

trace_rust_roots 🔒 ⚠