script::runtime::script_runtime

Struct IntroductionType

pub(crate) struct IntroductionType;

introductionType values in SpiderMonkey TransitiveCompileOptions.

Value definitions are based on the SpiderMonkey Debugger API docs: https://firefox-source-docs.mozilla.org/js/Debugger/Debugger.Source.html#introductiontype

Implementations

impl IntroductionType

pub const EVAL: &CStr = c"eval"

introductionType for code passed to eval.

pub const EVAL_STR: &str = "eval"

pub const DEBUGGER_EVAL: &CStr = c"debugger eval"

introductionType for code evaluated by debugger. This includes code run via the devtools repl, even if the thread is not paused.

pub const DEBUGGER_EVAL_STR: &str = "debugger eval"

pub const FUNCTION: &CStr = c"Function"

introductionType for code passed to the Function constructor.

pub const FUNCTION_STR: &str = "Function"

pub const WORKLET: &CStr = c"Worklet"

introductionType for code loaded by worklet.

pub const WORKLET_STR: &str = "Worklet"

pub const EVENT_HANDLER: &CStr = c"eventHandler"

introductionType for code assigned to DOM elements’ event handler IDL attributes as a string.

pub const EVENT_HANDLER_STR: &str = "eventHandler"

pub const SRC_SCRIPT: &CStr = c"srcScript"

introductionType for code belonging to <script src="file.js"> elements. This includes <script type="module" src="...">.

pub const SRC_SCRIPT_STR: &str = "srcScript"

pub const INLINE_SCRIPT: &CStr = c"inlineScript"

introductionType for code belonging to <script>code;</script> elements. This includes <script type="module" src="...">.

pub const INLINE_SCRIPT_STR: &str = "inlineScript"

pub const INJECTED_SCRIPT: &CStr = c"injectedScript"

introductionType for code belonging to scripts that would be "inlineScript" except that they were not part of the initial file itself. For example, scripts created via:

document.write("<script>code;</script>")

var s = document.createElement("script"); s.text = "code";

pub const INJECTED_SCRIPT_STR: &str = "injectedScript"

pub const IMPORTED_MODULE: &CStr = c"importedModule"

introductionType for code that was loaded indirectly by being imported by another script using ESM static or dynamic imports.

pub const IMPORTED_MODULE_STR: &str = "importedModule"

pub const JAVASCRIPT_URL: &CStr = c"javascriptURL"

introductionType for code presented in javascript: URLs.

pub const JAVASCRIPT_URL_STR: &str = "javascriptURL"

pub const DOM_TIMER: &CStr = c"domTimer"

introductionType for code passed to setTimeout/setInterval as a string.

pub const DOM_TIMER_STR: &str = "domTimer"

pub const WORKER: &CStr = c"Worker"

introductionType for web workers. FIXME: only documented in older(?) devtools user docs https://firefox-source-docs.mozilla.org/devtools-user/debugger-api/debugger.source/index.html

pub const WORKER_STR: &str = "Worker"