script::runtime::script_runtime

Struct Runtime

pub(crate) struct Runtime {
    rt: Runtime,
    job_queue: JobQueue,
    runtime_callback_data: Box<RuntimeCallbackData>,
}

Fields
rt: Runtimejob_queue: JobQueueruntime_callback_data: Box<RuntimeCallbackData>

The data that is set on the SpiderMonkey runtime callbacks as a pointer.

Implementations

impl Runtime

pub(crate) fn new(main_thread_sender: Option<ScriptEventLoopSender>) -> Runtime

Create a new runtime, optionally with the given [SendableTaskSource] for networking.

Safety

If panicking does not abort the program, any threads with child runtimes will continue executing after the thread with the parent runtime panics, but they will be in an invalid and undefined state.

This, like many calls to SpiderMoney API, is unsafe.

pub(crate) unsafe fn cx(&self) -> JSContext

Safety

only one JSContext can exist on the thread at a time (see note in JSContext::from_ptr)

the JSContext must not outlive the Runtime

pub(crate) unsafe fn new_with_parent( parent: Option<ParentRuntime>, script_event_loop_sender: Option<ScriptEventLoopSender>, ) -> Runtime

Create a new runtime, optionally with the given ParentRuntime and [SendableTaskSource] for networking.

Safety

If panicking does not abort the program, any threads with child runtimes will continue executing after the thread with the parent runtime panics, but they will be in an invalid and undefined state.

The parent pointer in the ParentRuntime argument must point to a valid object in memory.

This, like many calls to the SpiderMoney API, is unsafe.

pub(crate) fn set_script_thread(&mut self, script_thread: Weak<ScriptThread>)

pub(crate) fn thread_safe_js_context(&self) -> ThreadSafeJSContext