Enum script::script_runtime::CommonScriptMsg

pub enum CommonScriptMsg {
    CollectReports(ReportsChan),
    RefcountCleanup(TrustedReference),
    RunnableMsg(ScriptThreadEventCategory, Box<Runnable + Send>),
}

Common messages used to control the event loops in both the script and the worker

Variants

CollectReports

Requests that the script thread measure its memory usage. The results are sent back via the supplied channel.

RefcountCleanup

A DOM object's last pinned reference was removed (dispatched to all threads).

RunnableMsg

Generic message that encapsulates event handling.