Struct script::script_runtime::Runtime

pub struct Runtime {
    rt: Runtime,
    pub microtask_queue: Rc<MicrotaskQueue>,
    job_queue: *mut JobQueue,
    networking_task_src: Option<Box<NetworkingTaskSource>>,
}

Fields
rt: Runtimemicrotask_queue: Rc<MicrotaskQueue>job_queue: *mut JobQueuenetworking_task_src: Option<Box<NetworkingTaskSource>>
Implementations

impl Runtime

pub(crate) fn new( networking_task_source: Option<NetworkingTaskSource>, ) -> Runtime

Create a new runtime, optionally with the given NetworkingTaskSource.

Safety

If panicking does not abort the program, any threads with child runtimes will continue executing after the thread with the parent runtime panics, but they will be in an invalid and undefined state.

This, like many calls to SpiderMoney API, is unsafe.

pub(crate) unsafe fn new_with_parent( parent: Option<ParentRuntime>, networking_task_source: Option<NetworkingTaskSource>, ) -> Runtime

Create a new runtime, optionally with the given ParentRuntime and NetworkingTaskSource.

Safety

If panicking does not abort the program, any threads with child runtimes will continue executing after the thread with the parent runtime panics, but they will be in an invalid and undefined state.

The parent pointer in the ParentRuntime argument must point to a valid object in memory.

This, like many calls to the SpiderMoney API, is unsafe.

pub(crate) fn thread_safe_js_context(&self) -> ThreadSafeJSContext