Module script::script_task

The script task is the task that owns the DOM in memory, runs JavaScript, and spawns parsing and layout tasks. It's in charge of processing events for all same-origin pages in a frame tree, and manages the entire lifetime of pages in the frame tree from initial request to teardown.

Page loads follow a two-step process. When a request for a new page load is received, the network request is initiated and the relevant data pertaining to the new page is stashed. While the non-blocking request is ongoing, the script task is free to process further events, noting when they pertain to ongoing loads (such as resizes/viewport adjustments). When the initial response is received for an ongoing load, the second phase starts - the frame tree entry is created, along with the Window and Document objects, and the appropriate parser takes over the response body. Once parsing is complete, the document lifecycle for loading a page runs its course and the script task returns to processing events in the main event loop.

Structs

NonWorkerScriptChan
Encapsulates internal communication within the script task.

ScriptTask
Information for an entire page. Pages are top-level browsing contexts and can contain multiple frames.

StackRootTLS

Enums

ScriptMsg
Messages used to control script event loops, such as ScriptTask and DedicatedWorkerGlobalScope.

TimerSource

Statics

STACK_ROOTS

Traits

MainThreadRunnable

Runnable

ScriptChan
A cloneable interface for communicating with an event loop.

Functions

get_page