Enum script::script_task::ScriptMsg

pub enum ScriptMsg {
    TriggerFragment(PipelineId, String),
    Navigate(PipelineId, LoadData),
    FireTimer(TimerSource, TimerId),
    ExitWindow(PipelineId),
    DOMMessage(StructuredCloneData),
    RunnableMsg(Box<Runnable + Send>),
    MainThreadRunnableMsg(Box<MainThreadRunnable + Send>),
    RefcountCleanup(TrustedReference),
    PageFetchComplete(PipelineId, Option<SubpageId>, LoadResponse),
}

Messages used to control script event loops, such as ScriptTask and DedicatedWorkerGlobalScope.

Variants

TriggerFragment

Acts on a fragment URL load on the specified pipeline (only dispatched to ScriptTask).

Navigate

Begins a content-initiated load on the specified pipeline (only dispatched to ScriptTask).

FireTimer

Fires a JavaScript timeout TimerSource must be FromWindow when dispatched to ScriptTask and must be FromWorker when dispatched to a DedicatedGlobalWorkerScope

ExitWindow

Notifies the script that a window associated with a particular pipeline should be closed (only dispatched to ScriptTask).

DOMMessage

Message sent through Worker.postMessage (only dispatched to DedicatedWorkerGlobalScope).

RunnableMsg

Generic message that encapsulates event handling.

MainThreadRunnableMsg

Generic message for running tasks in the ScriptTask

RefcountCleanup

A DOM object's last pinned reference was removed (dispatched to all tasks).

PageFetchComplete

The final network response for a page has arrived.