Enum script::script_thread::MainThreadScriptMsg

pub enum MainThreadScriptMsg {
    Common(CommonScriptMsg),
    DocumentLoadsComplete(PipelineId),
    ExitWindow(PipelineId),
    Navigate(PipelineId, LoadData),
    DOMManipulation(DOMManipulationTask),
    UserInteraction(UserInteractionTask),
}

Messages used to control the script event loop

Variants

Common

Common variants associated with the script messages

DocumentLoadsComplete

Notify a document that all pending loads are complete.

ExitWindow

Notifies the script that a window associated with a particular pipeline should be closed (only dispatched to ScriptThread).

Navigate

Begins a content-initiated load on the specified pipeline (only dispatched to ScriptThread).

DOMManipulation

Tasks that originate from the DOM manipulation task source

UserInteraction

Tasks that originate from the user interaction task source