Module script::serviceworkerjob

A Job is an abstraction of async operation in service worker lifecycle propagation. Each Job is uniquely identified by its scope_url, and is keyed accordingly under the script thread. The script thread contains a JobQueue, which stores all scheduled Jobs by multiple service worker clients in a Vec.

Structs

AsyncJobHandler

FinishJobHandler

Job

JobQueue

Enums

InvokeType

JobType

SettleType