Module script::task

Machinery for tasks.

Structs

CancellableTask
A task that can be cancelled by toggling a shared flag.

TaskCanceller
Encapsulated state required to create cancellable tasks from non-script threads.

Traits

TaskBox
A boxed version of TaskOnce.

TaskOnce
A task that can be run. The name method is for profiling purposes.