script::tasks

Module task

Machinery for tasks.

Macros

task 🔒

Structs

CancellableTask 🔒

A task that can be cancelled by toggling a shared flag.

TaskCanceller 🔒

Encapsulated state required to create cancellable tasks from non-script threads.

Traits

NonSendTaskBox 🔒

A boxed version of NonSendTaskOnce.

NonSendTaskOnce 🔒

A task that must be run on the same thread it originated in.

TaskBox 🔒

A boxed version of TaskOnce.

TaskOnce 🔒

A task that can be sent between threads and run. The name method is for profiling purposes.