Struct script::test::Dom

pub struct Dom<T> {
    ptr: NonNull<T>,
}

A traced reference to a DOM object

This type is critical to making garbage collection work with the DOM, but it is very dangerous; if garbage collection happens with a Dom<T> on the stack, the Dom<T> can point to freed memory.

This should only be used as a field in other DOM objects.

Fields
ptr: NonNull<T>
Implementations

impl<T> Dom<T>

pub unsafe fn to_layout(&self) -> LayoutDom<'_, T>

Returns LayoutDom<T> containing the same pointer.

impl<T: DomObject> Dom<T>

pub fn from_ref(obj: &T) -> Dom<T>

Create a Dom<T> from a &T