Crate script

The script crate contains all matters DOM.

Modules

clipboard_provider

cors
A partial implementation of CORS For now this library is XHR-specific. For stuff involving <img>, <iframe>, <form>, etc please check what the request mode should be and compare with the fetch spec This library will eventually become the core of the Fetch crate with CORSRequest being expanded into FetchRequest (etc)

devtools

document_loader
Tracking of pending loads in a document. https://html.spec.whatwg.org/multipage/#the-end

dom
The implementation of the DOM.

layout_interface
The high-level interface from script to layout. Using this abstract interface helps reduce coupling between these two components, and enables the DOM to be placed in a separate crate from layout.

mem
Routines for handling measuring the memory usage of arbitrary DOM nodes.

network_listener

origin

parse

reporter

script_runtime
The script runtime contains common traits and structs commonly used by the script thread, the dom, and the worker threads.

script_thread
The script thread is the thread that owns the DOM in memory, runs JavaScript, and spawns parsing and layout threads. It's in charge of processing events for all same-origin pages in a frame tree, and manages the entire lifetime of pages in the frame tree from initial request to teardown.

task_source

textinput
Common handling of keyboard input and state management for text input controls

timers

unpremultiplytable

webdriver_handlers

Macros

make_atomic_setter!

make_bool_getter!

make_bool_setter!

make_dimension_setter!

make_enumerated_getter!

make_getter!

make_int_getter!

make_legacy_color_setter!

make_limited_int_setter!

make_limited_uint_setter!

make_nonzero_dimension_setter!

make_setter!

make_uint_getter!

make_uint_setter!

make_url_getter!

make_url_or_base_getter!

make_url_setter!

Functions

init

perform_platform_specific_initialization