Crate script

The script crate contains all matters DOM.

Modules

cors
A partial implementation of CORS For now this library is XHR-specific. For stuff involving <img>, <iframe>, <form>, etc please check what the request mode should be and compare with the fetch spec This library will eventually become the core of the Fetch crate with CORSRequest being expanded into FetchRequest (etc)

dom
The implementation of the DOM.

layout_interface
The high-level interface from script to layout. Using this abstract interface helps reduce coupling between these two components, and enables the DOM to be placed in a separate crate from layout.

page

parse

script_task
The script task is the task that owns the DOM in memory, runs JavaScript, and spawns parsing and layout tasks. It's in charge of processing events for all same-origin pages in a frame tree, and manages the entire lifetime of pages in the frame tree from initial request to teardown.

textinput
Common handling of keyboard input and state management for text input controls

Macros

make_bool_getter!

make_bool_setter!

make_enumerated_getter!

make_getter!

make_setter!

make_uint_getter!

make_uint_setter!

make_url_getter!

make_url_or_base_getter!