Struct script::document_loader::DocumentLoader

pub struct DocumentLoader {
    resource_threads: ResourceThreads,
    pipeline: Option<PipelineId>,
    blocking_loads: Vec<LoadType>,
    events_inhibited: bool,
}

Fields
resource_threads: ResourceThreads pipeline: Option<PipelineId> blocking_loads: Vec<LoadType> events_inhibited: bool
Methods

impl DocumentLoader

fn new(existing: &DocumentLoader) -> DocumentLoader

fn new_with_threads(resource_threads: ResourceThreads, pipeline: Option<PipelineId>, initial_load: Option<Url>) -> DocumentLoader

fn add_blocking_load(&mut self, load: LoadType)

Add a load to the list of blocking loads.

fn prepare_async_load(&mut self, load: LoadType, referrer: &Document) -> PendingAsyncLoad

Create a new pending network request, which can be initiated at some point in the future.

fn load_async(&mut self, load: LoadType, listener: AsyncResponseTarget, referrer: &Document)

Create and initiate a new network request.

fn finish_load(&mut self, load: &LoadType)

Mark an in-progress network request complete.

fn is_blocked(&self) -> bool

fn inhibit_events(&mut self)

fn events_inhibited(&self) -> bool

fn resource_threads(&self) -> &ResourceThreads