Module script::dom::bindings

The code to expose the DOM to JavaScript through IDL bindings.

Exposing a DOM object to JavaScript

As explained earlier, the implementation of an interface Foo involves two objects: the DOM object (implemented in Rust) and the reflector (a JSObject).

In order to expose the interface's members to the web, properties corresponding to the operations and attributes are defined on an object in the reflector's prototype chain or on the reflector itself.

Typically, these properties are either value properties whose value is a function (for operations) or accessor properties that have a getter and optionally a setter function (for attributes, depending on whether they are marked readonly).

All these JavaScript functions are set up such that, when they're called, they call a Rust function in the generated glue code. This glue code does some sanity checks and argument conversions, and calls into API implementation for the DOM object.

Modules

callback
Base classes to work with IDL callbacks.

cell
A shareable mutable container for the DOM.

codegen
Generated JS-Rust bindings.

conversions
Conversions of Rust values to and from JSVal.

error
Utilities to throw exceptions from Rust bindings.

global
Abstractions for global scopes.

js
Smart pointers for the JS-managed DOM objects.

num
The Finite<T> struct.

proxyhandler
Utilities for the implementation of JSAPI proxy handlers.

refcounted
A generic, safe mechanism by which DOM objects can be pinned and transferred between tasks (or intra-task for asynchronous events). Akin to Gecko's nsMainThreadPtrHandle, this uses thread-safe reference counting and ensures that the actual SpiderMonkey GC integration occurs on the script task via message passing. Ownership of a Trusted<T> object means the DOM object of type T to which it points remains alive. Any other behaviour is undefined. To guarantee the lifetime of a DOM object when performing asynchronous operations, obtain a Trusted<T> from that object and pass it along with each operation. A usable pointer to the original DOM object can be obtained on the script task from a Trusted<T> via the to_temporary method.

str
The ByteString struct.

structuredclone
This module implements structured cloning, as defined by HTML.

trace
Utilities for tracing JS-managed values.

utils
Various utilities to glue JavaScript and the DOM implementation together.