Struct script::dom::bindings::callback::CallSetup

pub struct CallSetup {
    exception_compartment: RootedObject,
    cx: *mut JSContext,
    old_compartment: *mut JSCompartment,
    handling: ExceptionHandling,
}

A class that performs whatever setup we need to safely make a call while this class is on the stack. After new returns, the call is safe to make.

Fields

exception_compartment

The compartment for reporting exceptions. As a RootedObject, this must be the first field in order to determine the final address on the stack correctly.

cx

The JSContext used for the call.

old_compartment

The compartment we were in before the call.

handling

The exception handling used for the call.

Methods

impl CallSetup

fn new<T: CallbackContainer>(callback: &T, handling: ExceptionHandling) -> CallSetup

Performs the setup needed to make a call.

fn get_context(&self) -> *mut JSContext

Returns the JSContext used for the call.