Struct script::dom::bindings::cell::DOMRefCell

pub struct DOMRefCell<T> {
    value: RefCell<T>,
}

A mutable field in the DOM.

This extends the API of core::cell::RefCell to allow unsafe access in certain situations, with dynamic checking in debug builds.

Fields

value

Methods

impl<T> DOMRefCell<T>

unsafe fn borrow_for_layout(&self) -> &T

Return a reference to the contents.

For use in the layout thread only.

unsafe fn borrow_for_gc_trace(&self) -> &T

Borrow the contents for the purpose of GC tracing.

This succeeds even if the object is mutably borrowed, so you have to be careful in trace code!

unsafe fn borrow_for_script_deallocation(&self) -> &mut T

Borrow the contents for the purpose of script deallocation.

fn is_mutably_borrowed(&self) -> bool

Is the cell mutably borrowed?

For safety checks in debug builds only.

fn try_borrow(&self) -> Option<Ref<T>>

Attempts to immutably borrow the wrapped value.

The borrow lasts until the returned Ref exits scope. Multiple immutable borrows can be taken out at the same time.

Returns None if the value is currently mutably borrowed.

Panics

Panics if this is called off the script thread.

fn try_borrow_mut(&self) -> Option<RefMut<T>>

Mutably borrows the wrapped value.

The borrow lasts until the returned RefMut exits scope. The value cannot be borrowed while this borrow is active.

Returns None if the value is currently borrowed.

Panics

Panics if this is called off the script thread.

fn borrow_mut_for_layout(&self) -> RefMut<T>

Version of the above that we use during restyle while the script thread is blocked.

impl<T> DOMRefCell<T>

fn new(value: T) -> DOMRefCell<T>

Create a new DOMRefCell containing value.

fn borrow(&self) -> Ref<T>

Immutably borrows the wrapped value.

The borrow lasts until the returned Ref exits scope. Multiple immutable borrows can be taken out at the same time.

Panics

Panics if this is called off the script thread.

Panics if the value is currently mutably borrowed.

fn borrow_mut(&self) -> RefMut<T>

Mutably borrows the wrapped value.

The borrow lasts until the returned RefMut exits scope. The value cannot be borrowed while this borrow is active.

Panics

Panics if this is called off the script thread.

Panics if the value is currently borrowed.