Module script::dom::bindings::conversions

Conversions of Rust values to and from JSVal.

IDL type 	Argument type 	Return type

any 	JSVal

boolean 	bool

byte 	i8

octet 	u8

short 	i16

unsigned short 	u16

long 	i32

unsigned long 	u32

long long 	i64

unsigned long long 	u64

unrestricted float 	f32

float 	Finite<f32>

unrestricted double 	f64

double 	Finite<f64>

DOMString 	DOMString

USVString 	USVString

ByteString 	ByteString

object 	*mut JSObject

interface types 	JSRef<T> 	Temporary<T>

dictionary types 	&T 	unsupported

enumeration types 	T

callback function types 	T

nullable types 	Option<T>

sequences 	Vec<T>

union types 	T

Enums

StringificationBehavior
Behavior for stringification of JSVals.

Constants

DOM_OBJECT_SLOT
The index of the slot wherein a pointer to the reflected DOM object is stored for non-proxy bindings.

Traits

FromJSValConvertible
A trait to convert JSVals to Rust types.

IDLInterface
A trait to retrieve the constants necessary to check if a JSObject implements a given interface.

ToJSValConvertible
A trait to convert Rust types to JSVals.

Functions

dom_object_slot
Returns the index of the slot wherein a pointer to the reflected DOM object is stored.

is_dom_class
Returns whether the given clasp is one for a DOM object.

is_dom_proxy
Returns whether obj is a DOM object implemented as a proxy.

jsid_to_str
Convert the given jsid to a DOMString. Fails if the jsid is not a string, or if the string does not contain valid UTF-16.

jsstring_to_str
Convert the given JSString to a DOMString. Fails if the string does not contain valid UTF-16.

native_from_reflector
Get the DOM object from the given reflector.

native_from_reflector_jsmanaged
Get an Unrooted<T> for the given DOM object, unwrapping any wrapper around it first, and checking if the object is of the correct type.