Enum script::dom::bindings::global::GlobalRef

pub enum GlobalRef<'a> {
    Window(&'a Window),
    Worker(&'a WorkerGlobalScope),
}

A freely-copyable reference to a rooted global object.

Variants

Window

A reference to a Window object.

Worker

A reference to a WorkerGlobalScope object.

Methods

impl<'a> GlobalRef<'a>

fn get_cx(&self) -> *mut JSContext

Get the JSContext for the JSRuntime associated with the thread this global object is on.

fn as_window(&self) -> &Window

Extract a Window, causing thread failure if the global object is not a Window.

fn request_source(&self) -> RequestSource

gets the custom message channel associated with global object

fn pipeline(&self) -> PipelineId

Get the PipelineId for this global scope.

fn mem_profiler_chan(&self) -> &ProfilerChan

Get a mem::ProfilerChan to send messages to the memory profiler thread.

fn time_profiler_chan(&self) -> &ProfilerChan

Get a time::ProfilerChan to send messages to the time profiler thread.

fn constellation_chan(&self) -> &IpcSender<ConstellationMsg>

Get a IpcSender to send messages to the constellation when available.

fn scheduler_chan(&self) -> &IpcSender<TimerEventRequest>

Get the scheduler channel to request timer events.

fn devtools_chan(&self) -> Option<IpcSender<ScriptToDevtoolsControlMsg>>

Get an IpcSender<ScriptToDevtoolsControlMsg> to send messages to Devtools thread when available.

fn resource_threads(&self) -> ResourceThreads

Get the ResourceThreads for this global scope.

fn core_resource_thread(&self) -> CoreResourceThread

Get the CoreResourceThread for this global scope

fn filemanager_thread(&self) -> IpcSender<FileManagerThreadMsg>

Get the port to file manager for this global scope

fn get_worker_id(&self) -> Option<WorkerId>

Get the worker's id.

fn get_next_worker_id(&self) -> WorkerId

Get next worker id.

fn get_url(&self) -> Url

Get the URL for this global scope.

fn api_base_url(&self) -> Url

Get the base url for this global scope.

fn script_chan(&self) -> Box<ScriptChan + Send>

ScriptChan used to send messages to the event loop of this global's thread.

fn dom_manipulation_task_source(&self) -> Box<TaskSource<DOMManipulationTask> + Send>

TaskSource used to queue DOM manipulation messages to the event loop of this global's thread.

fn networking_task_source(&self) -> Box<ScriptChan + Send>

ScriptChan used to send messages to the event loop of this global's thread.

fn history_traversal_task_source(&self) -> Box<ScriptChan + Send>

ScriptChan used to send messages to the event loop of this global's thread.

fn file_reading_task_source(&self) -> Box<ScriptChan + Send>

ScriptChan used to send messages to the event loop of this global's thread.

fn new_script_pair(&self) -> (Box<ScriptChan + Send>, Box<ScriptPort + Send>)

Create a new sender/receiver pair that can be used to implement an on-demand event loop. Used for implementing web APIs that require blocking semantics without resorting to nested event loops.

fn process_event(&self, msg: CommonScriptMsg)

Process a single event as if it were the next event in the thread queue for this global.

fn evaluate_js_on_global_with_result(&self, code: &str, rval: MutableHandleValue)

Evaluate the JS messages on the RootedValue of this global

fn set_devtools_wants_updates(&self, send_updates: bool)

Set the bool value to indicate whether developer tools has requested updates from the global

fn schedule_callback(&self, callback: OneshotTimerCallback, duration: MsDuration) -> OneshotTimerHandle

Schedule the given callback to be invoked after at least duration milliseconds have passed.

fn unschedule_callback(&self, handle: OneshotTimerHandle)

Unschedule a previously-scheduled callback.

fn reflector(&self) -> &Reflector

Returns the receiver's reflector.

fn panic_chan(&self) -> &IpcSender<PanicMsg>

Returns an IpcSender to report panics on.