Module script::dom::bindings::inheritance

The Castable trait.

Re-exports

pub use crate::dom::bindings::codegen::InheritTypes::*;

Traits

Castable
A trait to hold the cast functions of IDL interfaces that either derive or are derived from other interfaces.

HasParent