Trait script::dom::bindings::inheritance::Castable

pub trait Castable: IDLInterface + DomObject + Sized {
    fn is<T>(&self) -> bool
    where
        T: DerivedFrom<Self>,
    { ... }

    fn upcast<T>(&self) -> &T
    where
        T: Castable,
        Self: DerivedFrom<T>,
    { ... }

    fn downcast<T>(&self) -> Option<&T>
    where
        T: DerivedFrom<Self>,
    { ... }
}

A trait to hold the cast functions of IDL interfaces that either derive or are derived from other interfaces.

Provided methods

fn is<T>(&self) -> bool where
T: DerivedFrom<Self>,

Check whether a DOM object implements one of its deriving interfaces.

fn upcast<T>(&self) -> &T where
T: Castable,
Self: DerivedFrom<T>,

Cast a DOM object upwards to one of the interfaces it derives from.

fn downcast<T>(&self) -> Option<&T> where
T: DerivedFrom<Self>,

Cast a DOM object downwards to one of the interfaces it might implement.

Loading content...