Module script::dom::bindings::interface

Machinery to initialise interface prototype objects and interface objects.

Structs

InterfaceConstructorBehavior
The constructor behavior of a non-callback interface object.

NonCallbackInterfaceObjectClass
The class of a non-callback interface object.

Constants

OBJECT_OPS

Functions

create_callback_interface_object⚠
Create and define the interface object of a callback interface.

create_global_object⚠
Create a global object with the given class.

create_interface_prototype_object⚠
Create the interface prototype object of a non-callback interface.

create_named_constructors⚠
Create and define the named constructors of a non-callback interface.

create_noncallback_interface_object⚠
Create and define the interface object of a non-callback interface.

create_object⚠
Create a new object with a unique type.

create_unscopable_object⚠

define_guarded_constants⚠
Conditionally define constants on an object.

define_guarded_methods⚠
Conditionally define methods on an object.

define_guarded_properties⚠
Conditionally define properties on an object.

define_length⚠

define_name⚠

define_on_global_object⚠
Define a property with a given name on the global object. Should be called through the resolve hook.

fun_to_string_hook⚠

has_instance⚠
Return whether a value is an instance of a given prototype. http://heycam.github.io/webidl/#es-interface-hasinstance

has_instance_hook⚠
Hook for instanceof on interface objects.

invalid_constructor⚠

is_exposed_in⚠
Returns whether an interface with exposure set given by globals should be exposed in the global object obj.

non_new_constructor⚠

Type Definitions

ConstructorClassHook
A constructor class hook.

TraceHook
A trace hook.