Module script::dom::bindings::interface

Machinery to initialise interface prototype objects and interface objects.

Structs

ConstantSpec
Representation of an IDL constant.

InterfaceConstructorBehavior
The constructor behavior of a non-callback interface object.

NonCallbackInterfaceObjectClass
The class of a non-callback interface object.

Enums

ConstantVal
Representation of an IDL constant value.

Functions

create_callback_interface_object
Create and define the interface object of a callback interface.

create_interface_prototype_object
Create the interface prototype object of a non-callback interface.

create_named_constructors
Create and define the named constructors of a non-callback interface.

create_noncallback_interface_object
Create and define the interface object of a non-callback interface.

create_object

create_unscopable_object

define_constants
Defines constants on obj. Fails on JSAPI failure.

define_guarded_methods
Conditionally define methods on an object.

define_guarded_properties
Conditionally define properties on an object.

define_length

define_name

define_on_global_object

fun_to_string_hook

has_instance
Return whether a value is an instance of a given prototype. http://heycam.github.io/webidl/#es-interface-hasinstance

has_instance_hook
Hook for instanceof on interface objects.

invalid_constructor

non_new_constructor

Type Definitions

ConstructorClassHook
A constructor class hook.

NonNullJSNative
A JSNative that cannot be null.