Struct script::dom::bindings::interface::InterfaceConstructorBehavior

pub struct InterfaceConstructorBehavior {
    call: JSNative,
    construct: JSNative,
}

The constructor behavior of a non-callback interface object.

Fields
call: JSNative construct: JSNative
Methods

impl InterfaceConstructorBehavior

const fn throw() -> InterfaceConstructorBehavior

An interface constructor that unconditionally throws a type error.

const fn call(hook: ConstructorClassHook) -> InterfaceConstructorBehavior

An interface constructor that calls a native Rust function.