script::dom::bindings::iterable

Trait IteratorDerives

pub trait IteratorDerives {
    // Required method
    fn derives(class: &'static DOMClass) -> bool;
}

A version of the IDLInterface trait that is specific to types that have iterators defined for them. This allows the script crate to define the derives check for the concrete interface type, while the [IteratableIterator] type defined in this module can be parameterized over an unknown generic.

Required Methods

fn derives(class: &'static DOMClass) -> bool

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.