Module script::dom::bindings::js

Smart pointers for the JS-managed DOM objects.

The DOM is made up of DOM objects whose lifetime is entirely controlled by the whims of the SpiderMonkey garbage collector. The types in this module are designed to ensure that any interactions with said Rust types only occur on values that will remain alive the entire time.

Here is a brief overview of the important types:

Root<T>: a stack-based reference to a rooted DOM object.

JS<T>: a reference to a DOM object that can automatically be traced by the GC when encountered as a field of a Rust structure.

JS<T> does not allow access to their inner value without explicitly creating a stack-based root via the root method. This returns a Root<T>, which causes the JS-owned value to be uncollectable for the duration of the Root object's lifetime. A reference to the object can then be obtained from the Root object. These references are not allowed to outlive their originating Root<T>.

Structs

JS
A traced reference to a DOM object

LayoutJS
An unrooted reference to a DOM object for use in layout. Layout*Helpers traits must be implemented on this.

MutHeap
A holder that provides interior mutability for GC-managed values such as JS<T>. Essentially a Cell<JS<T>>, but safer.

MutHeapJSVal
A holder that provides interior mutability for GC-managed JSVals.

MutNullableHeap
A holder that provides interior mutability for GC-managed values such as JS<T>, with nullability represented by an enclosing Option wrapper. Essentially a Cell<Option<JS<T>>>, but safer.

Root
A rooted reference to a DOM object.

RootCollection
A rooting mechanism for reflectors on the stack. LIFO is not required.

RootCollectionPtr
A pointer to a RootCollection, for use in global variables.

Traits

HeapGCValue
A trait to be implemented for JS-managed types that can be stored in mutable member fields.

OptionalRootedReference
Get an Option<Option<&T>> out of an Option<Option<Root<T>>>

RootedReference
Get an Option<&T> out of an Option<Root<T>>

Functions

trace_roots
SM Callback that traces the rooted reflectors