Struct script::dom::bindings::js::JS

pub struct JS<T> {
    ptr: NonZero<*const T>,
}

A traced reference to a DOM object

This type is critical to making garbage collection work with the DOM, but it is very dangerous; if garbage collection happens with a JS<T> on the stack, the JS<T> can point to freed memory.

This should only be used as a field in other DOM objects.

Fields

ptr

Methods

impl<T> JS<T>

unsafe fn to_layout(&self) -> LayoutJS<T>

Returns LayoutJS<T> containing the same pointer.

impl<T: Reflectable> JS<T>

fn from_rooted(root: &Root<T>) -> JS<T>

Create a JS from a Root XXX Not a great API. Should be a call on Root instead

fn from_ref(obj: &T) -> JS<T>

Create a JS from a &T