Struct script::dom::bindings::js::LayoutJS

pub struct LayoutJS<T> {
    ptr: NonZero<*const T>,
}

An unrooted reference to a DOM object for use in layout. Layout*Helpers traits must be implemented on this.

Fields
ptr: NonZero<*const T>
Methods

impl<T: Castable> LayoutJS<T>

fn upcast<U>(&self) -> LayoutJS<U> where U: Castable, T: DerivedFrom<U>

Cast a DOM object root upwards to one of the interfaces it derives from.

fn downcast<U>(&self) -> Option<LayoutJS<U>> where U: DerivedFrom<T>

Cast a DOM object downwards to one of the interfaces it might implement.

impl<T: Reflectable> LayoutJS<T>

unsafe fn get_jsobject(&self) -> *mut JSObject

Get the reflector.

impl LayoutJS<Node>

unsafe fn from_trusted_node_address(inner: TrustedNodeAddress) -> LayoutJS<Node>

Create a new JS-owned value wrapped from an address known to be a Node pointer.

impl<T: Reflectable> LayoutJS<T>

unsafe fn unsafe_get(&self) -> *const T

Returns an unsafe pointer to the interior of this JS object. This is the only method that be safely accessed from layout. (The fact that this is unsafe is what necessitates the layout wrappers.)