Struct script::dom::bindings::js::Root

pub struct Root<T> {
    // some fields omitted
}

A rooted reference to a DOM object.

The JS value is pinned for the duration of this object's lifetime; roots are additive, so this object's destruction will not invalidate other roots for the same JS value. Roots cannot outlive the associated RootCollection object. Attempts to transfer ownership of a Root via moving will trigger dynamic unrooting failures due to incorrect ordering.

Methods

impl<T: Reflectable> Root<T>

fn r<'b>(&'b self) -> JSRef<'b, T>

Obtain a safe reference to the wrapped JS owned-value that cannot outlive the lifetime of this root.

fn get_unsound_ref_forever<'b>(&self) -> JSRef<'b, T>

Obtain an unsafe reference to the wrapped JS owned-value that can outlive the lifetime of this root.

DO NOT CALL.