Module script::dom::bindings::proxyhandler

Utilities for the implementation of JSAPI proxy handlers.

Statics

JSPROXYSLOT_EXPANDO

Functions

define_property
Defines an expando on the given proxy.

delete
Deletes an expando off the given proxy.

ensure_expando_object
Get the expando object, or create it if it doesn't exist yet. Fails on JSAPI failure.

fill_property_descriptor
Set the property descriptor's object to obj and set it to enumerable, and writable if readonly is true.

get_expando_object
Get the expando object, or null if there is none.

get_property_descriptor
Invoke the [[GetOwnProperty]] trap (getOwnPropertyDescriptor) on proxy, with argument id and return the result, if it is not undefined. Otherwise, walk along the prototype chain to find a property with that name.

is_extensible
Reports whether the object is Extensible

prevent_extensions
Controls whether the Extensible bit can be changed