Module script::dom::bindings::proxyhandler

Utilities for the implementation of JSAPI proxy handlers.

Functions

define_property⚠

Defines an expando on the given proxy.

delete⚠

Deletes an expando off the given proxy.

ensure_expando_object⚠

Get the expando object, or create it if it doesn’t exist yet. Fails on JSAPI failure.

fill_property_descriptor

Set the property descriptor’s object to obj and set it to enumerable, and writable if readonly is true.

get_expando_object⚠

Get the expando object, or null if there is none.

get_prototype_if_ordinary⚠

If proxy (underneath any functionally-transparent wrapper proxies) has as its [[GetPrototypeOf]] trap the ordinary [[GetPrototypeOf]] behavior defined for ordinary objects, set *is_ordinary to true and store obj’s prototype in proto. Otherwise set *isOrdinary to false. In case of error, both outparams have unspecified value.

init⚠

Initialize the infrastructure for DOM proxy objects.

is_extensible⚠

Reports whether the object is Extensible

prevent_extensions⚠

Controls whether the Extensible bit can be changed

shadow_check_callback⚠

Determine if this id shadows any existing properties for this proxy.