Module script::dom::bindings::proxyhandler

Utilities for the implementation of JSAPI proxy handlers.

Structs

CrossOriginProperties

Property and method specs that correspond to the elements of CrossOriginProperties(O).

Constants

ALLOWLISTED_SYMBOL_CODES 🔒

Functions

append_cross_origin_allowlisted_prop_keys 🔒 ⚠

Append « "then", @@toStringTag, @@hasInstance, @@isConcatSpreadable » to props. This is used to implement CrossOriginOwnPropertyKeys.

cross_origin_get⚠

Implementation of CrossOriginGet.

cross_origin_get_own_property_helper⚠

Implementation of CrossOriginGetOwnPropertyHelper.

cross_origin_has_own⚠

Evaluate CrossOriginGetOwnPropertyHelper(proxy, id) != null. SpiderMonkey-specific.

cross_origin_own_property_keys⚠

Implementation of CrossOriginOwnPropertyKeys.

cross_origin_property_fallback⚠

Implementation of CrossOriginPropertyFallback.

cross_origin_set⚠

Implementation of CrossOriginSet.

define_property⚠

Defines an expando on the given proxy.

delete⚠

Deletes an expando off the given proxy.

ensure_cross_origin_property_holder 🔒 ⚠

Get the holder for cross-origin properties for the current global of the JSContext, creating one and storing it in a slot of the proxy object if it doesn’t exist yet.

ensure_expando_object⚠

Get the expando object, or create it if it doesn’t exist yet. Fails on JSAPI failure.

get_expando_object⚠

Get the expando object, or null if there is none.

get_getter_object 🔒 ⚠

get_prototype_if_ordinary⚠

If proxy (underneath any functionally-transparent wrapper proxies) has as its [[GetPrototypeOf]] trap the ordinary [[GetPrototypeOf]] behavior defined for ordinary objects, set *is_ordinary to true and store obj’s prototype in proto. Otherwise set *isOrdinary to false. In case of error, both outparams have unspecified value.

get_setter_object 🔒 ⚠

id_to_source 🔒 ⚠

init⚠

Initialize the infrastructure for DOM proxy objects.

is_accessor_descriptor 🔒

https://tc39.es/ecma262/#sec-isaccessordescriptor

is_cross_origin_allowlisted_prop 🔒 ⚠

is_data_descriptor 🔒

https://tc39.es/ecma262/#sec-isdatadescriptor

is_extensible⚠

Reports whether the object is Extensible

is_platform_object_same_origin⚠

https://html.spec.whatwg.org/multipage/#isplatformobjectsameorigin-(-o-)

maybe_cross_origin_get_prototype⚠

Implementation of [[GetPrototypeOf]] for Location.

maybe_cross_origin_get_prototype_if_ordinary_rawcx⚠

maybe_cross_origin_set_prototype_rawcx⚠

Implementation of [[SetPrototypeOf]] for Location and WindowProxy.

maybe_cross_origin_set_rawcx⚠

Implementation of [[Set]] for Location.

prevent_extensions⚠

Controls whether the Extensible bit can be changed

report_cross_origin_denial⚠

Report a cross-origin denial for a property, Always returns false, so it can be used as return report_cross_origin_denial(...);.

set_property_descriptor

Set the property descriptor’s object to obj and set it to enumerable, and writable if readonly is true.

shadow_check_callback⚠

Determine if this id shadows any existing properties for this proxy.