Module script::dom::bindings::refcounted

A generic, safe mechanism by which DOM objects can be pinned and transferred between tasks (or intra-task for asynchronous events). Akin to Gecko's nsMainThreadPtrHandle, this uses thread-safe reference counting and ensures that the actual SpiderMonkey GC integration occurs on the script task via message passing. Ownership of a Trusted<T> object means the DOM object of type T to which it points remains alive. Any other behaviour is undefined. To guarantee the lifetime of a DOM object when performing asynchronous operations, obtain a Trusted<T> from that object and pass it along with each operation. A usable pointer to the original DOM object can be obtained on the script task from a Trusted<T> via the to_temporary method.

The implementation of Trusted is as follows: A hashtable resides in the script task, keyed on the pointer to the Rust DOM object. The values in this hashtable are atomic reference counts. When a Trusted object is created or cloned, this count is increased. When a Trusted is dropped, the count decreases. If the count hits zero, a message is dispatched to the script task to remove the entry from the hashmap if the count is still zero. The JS reflector for the DOM object is rooted when a hashmap entry is first created, and unrooted when the hashmap entry is removed.

Structs

LiveDOMReferences
The set of live, pinned DOM objects that are currently prevented from being garbage collected due to outstanding references.

Trusted
A safe wrapper around a raw pointer to a DOM object that can be shared among tasks for use in asynchronous operations. The underlying DOM object is guaranteed to live at least as long as the last outstanding Trusted<T> instance.

TrustedReference
A pointer to a Rust DOM object that needs to be destroyed.

Statics

LIVE_REFERENCES