Struct script::dom::bindings::refcounted::Trusted

pub struct Trusted<T: Reflectable> {
    ptr: *const c_void,
    refcount: Arc<Mutex<usize>>,
    script_chan: Box<ScriptChan + Send>,
    owner_thread: *const c_void,
    phantom: PhantomData<T>,
}

A safe wrapper around a raw pointer to a DOM object that can be shared among threads for use in asynchronous operations. The underlying DOM object is guaranteed to live at least as long as the last outstanding Trusted<T> instance.

Fields

ptr

A pointer to the Rust DOM object of type T, but void to allow sending Trusted<T> between threads, regardless of T's sendability.

refcount

script_chan

owner_thread

phantom

Methods

impl<T: Reflectable> Trusted<T>

fn new(ptr: &T) -> Trusted<T>

Create a new Trusted<T> instance from an existing DOM pointer. The DOM object will be prevented from being GCed for the duration of the resulting Trusted<T> object's lifetime.

fn root(&self) -> Root<T>

Obtain a usable DOM pointer from a pinned Trusted<T> value. Fails if used on a different thread than the original value from which this Trusted<T> was obtained.