script::dom::bindings::root

Struct LayoutDom

#[repr(transparent)]
pub(crate) struct LayoutDom<'dom, T> {
    value: &'dom T,
}

An unrooted reference to a DOM object for use in layout. Layout*Helpers traits must be implemented on this.

Fields
value: &'dom T
Implementations

impl<'dom, T> LayoutDom<'dom, T>
where T: Castable,

pub(crate) fn upcast<U>(&self) -> LayoutDom<'dom, U>
where U: Castable, T: DerivedFrom<U>,

Cast a DOM object root upwards to one of the interfaces it derives from.

pub(crate) fn downcast<U>(&self) -> Option<LayoutDom<'dom, U>>
where U: DerivedFrom<T>,

Cast a DOM object downwards to one of the interfaces it might implement.

pub(crate) fn is<U>(&self) -> bool
where U: DerivedFrom<T>,

Returns whether this inner object is a U.

impl<T> LayoutDom<'_, T>
where T: DomObject,

pub(crate) unsafe fn get_jsobject(&self) -> *mut JSObject

Get the reflector.

impl LayoutDom<'_, Node>

pub(crate) unsafe fn from_trusted_node_address( inner: TrustedNodeAddress, ) -> Self

Create a new JS-owned value wrapped from an address known to be a Node pointer.

impl<'dom, T> LayoutDom<'dom, T>
where T: 'dom + DomObject,

pub(crate) fn unsafe_get(self) -> &'dom T

Returns a reference to the interior of this JS object. The fact that this is unsafe is what necessitates the layout wrappers.

pub(crate) unsafe fn to_layout_slice( slice: &'dom [Dom<T>], ) -> &'dom [LayoutDom<'dom, T>]

Transforms a slice of Dom<T> into a slice of LayoutDom<T>.

impl LayoutDom<'_, Element>

pub(super) fn focus_state(self) -> bool

impl<'dom> LayoutDom<'dom, HTMLImageElement>

fn current_request(self) -> &'dom ImageRequest

fn dimension_attribute_source(self) -> LayoutDom<'dom, Element>

impl<'dom> LayoutDom<'dom, HTMLInputElement>

fn get_raw_textinput_value(self) -> DOMString

fn input_type(self) -> InputType

fn textinput_sorted_selection_offsets_range(self) -> Range<Utf8CodeUnitLength>

impl<'dom> LayoutDom<'dom, HTMLTextAreaElement>

fn textinput_sorted_selection_offsets_range(self) -> Range<Utf8CodeUnitLength>

impl<'dom> LayoutDom<'dom, Node>

pub(crate) fn parent_node_ref(self) -> Option<LayoutDom<'dom, Node>>