Struct script::dom::bindings::root::Root

pub struct Root<T: StableTraceObject> {
    value: T,
    root_list: *const RootCollection,
}

A rooted value.

Fields
value: T

The value to root.

root_list: *const RootCollection

List that ensures correct dynamic root ordering

Implementations

impl<T> Root<T> where
T: StableTraceObject + 'static,

pub unsafe fn new(value: T) -> Self

Create a new stack-bounded root for the provided value. It cannot outlive its associated RootCollection, and it gives out references which cannot outlive this new Root.

impl<T: Castable> Root<Dom<T>>

pub fn upcast<U>(root: Root<Dom<T>>) -> Root<Dom<U>> where
U: Castable,
T: DerivedFrom<U>,

Cast a DOM object root upwards to one of the interfaces it derives from.

pub fn downcast<U>(root: Root<Dom<T>>) -> Option<Root<Dom<U>>> where
U: DerivedFrom<T>,

Cast a DOM object root downwards to one of the interfaces it might implement.

impl<T: DomObject> Root<Dom<T>>

pub fn from_ref(unrooted: &T) -> Root<Dom<T>>

Generate a new root from a reference

impl<T> Root<MaybeUnreflectedDom<T>> where
T: DomObject,

pub fn as_ptr(&self) -> *const T

impl<T> Root<MaybeUnreflectedDom<T>> where
T: MutDomObject,

pub unsafe fn reflect_with(self, obj: *mut JSObject) -> Root<Dom<T>>