Trait script::dom::bindings::root::StableTraceObject

pub unsafe trait StableTraceObject {
    fn stable_trace_object(&self) -> *const dyn JSTraceable;
}

Represents values that can be rooted through a stable address that will not change for their whole lifetime.

Required methods

fn stable_trace_object(&self) -> *const dyn JSTraceable

Returns a stable trace object which address won’t change for the whole lifetime of the value.