script::dom::bindings

Module str

Structs

ByteString

Encapsulates the IDL ByteString type.

DOMString

This type corresponds to the DOMString type in WebIDL.

USVString

A string that is constructed from a UCS-2 buffer by replacing invalid code points with the replacement character.

Traits

FromInputValueString 🔒

ToInputValueString 🔒

Functions

is_leap_year 🔒

is_token

Returns whether s is a token, as defined by RFC 2616.

max_day_in_month 🔒

max_week_in_year 🔒

https://html.spec.whatwg.org/multipage/#week-number-of-the-last-day

parse_date_component 🔒

https://html.spec.whatwg.org/multipage/#parse-a-date-component

parse_month_component 🔒

https://html.spec.whatwg.org/multipage/#parse-a-month-component

parse_time_component 🔒

https://html.spec.whatwg.org/multipage/#parse-a-time-component

serialize_jsval_to_json_utf8

Because this converts to a DOMString it becomes UTF-8 encoded which is closer to the spec definition of https://infra.spec.whatwg.org/#serialize-a-javascript-value-to-json-bytes but we generally do not operate on anything that is truly a WTF-16 string.