Module script::dom::bindings::str

The ByteString struct.

Structs

ByteString

Encapsulates the IDL ByteString type.

DOMString

A DOMString.

USVString

A string that is constructed from a UCS-2 buffer by replacing invalid code points with the replacement character.

Functions

is_leap_year 🔒

is_token

Returns whether s is a token, as defined by RFC 2616.

max_day_in_month 🔒

max_week_in_year 🔒

https://html.spec.whatwg.org/multipage/#week-number-of-the-last-day

parse_date_component 🔒

https://html.spec.whatwg.org/multipage/#parse-a-date-component

parse_month_component 🔒

https://html.spec.whatwg.org/multipage/#parse-a-month-component

parse_time_component 🔒

https://html.spec.whatwg.org/multipage/#parse-a-time-component