Struct script::dom::bindings::str::ByteString

pub struct ByteString(Vec<u8>);

Encapsulates the IDL ByteString type.

Methods

impl ByteString

fn new(value: Vec<u8>) -> ByteString

Creates a new ByteString.

fn as_str(&self) -> Option<&str>

Returns self as a string, if it encodes valid UTF-8, and None otherwise.

fn bytes(&mut self) -> Vec<u8>

Returns ownership of the underlying Vec and copies an empty vec in its place

fn len(&self) -> usize

Returns the length.

fn eq_ignore_case(&self, other: &ByteString) -> bool

Compare self to other, matching A–Z and a–z as equal.

fn to_lower(&self) -> ByteString

Returns self with A–Z replaced by a–z.