script::dom::bindings

Module structuredclone

This module implements structured cloning, as defined by HTML.

Structs

StructuredDataReader ğŸ”’

Reader and writer structs for results from, and inputs to, structured-data read/write operations. https://html.spec.whatwg.org/multipage/#safe-passing-of-structured-data

StructuredDataWriter ğŸ”’

A data holder for transferred and serialized objects.

Enums

OperationError ğŸ”’

StructuredCloneTags ğŸ”’

StructuredData ğŸ”’

Statics

STRUCTURED_CLONE_CALLBACKS ğŸ”’

Functions

can_transfer_callback ğŸ”’ âš

can_transfer_for_type ğŸ”’ âš

free_transfer_callback ğŸ”’ âš

read ğŸ”’

Read structured serialized data, possibly containing transferred objects. Returns a vec of rooted transfer-received ports, or an error.

read_callback ğŸ”’ âš

read_object ğŸ”’ âš

read_transfer_callback ğŸ”’ âš

reader_for_type ğŸ”’

receive_object ğŸ”’

receiver_for_type ğŸ”’

report_error_callback ğŸ”’ âš

sab_cloned_callback ğŸ”’ âš

serialize_for_type ğŸ”’

transfer_for_type ğŸ”’

try_serialize ğŸ”’ âš

try_transfer ğŸ”’ âš

write ğŸ”’

Writes a structured clone. Returns a DataClone error if that fails.

write_callback ğŸ”’ âš

write_object ğŸ”’ âš

write_transfer_callback ğŸ”’ âš

https://html.spec.whatwg.org/multipage/#structuredserializewithtransfer

Type Aliases

SerializeOperation ğŸ”’

TransferOperation ğŸ”’