Module script::dom::bindings::structuredclone

This module implements structured cloning, as defined by HTML.

Enums

StructuredCloneTags 🔒

StructuredDataHolder

A data holder for results from, and inputs to, structured-data read/write operations. https://html.spec.whatwg.org/multipage/#safe-passing-of-structured-data

Statics

STRUCTURED_CLONE_CALLBACKS 🔒

Functions

can_transfer_callback 🔒 ⚠

free_transfer_callback 🔒 ⚠

read

Read structured serialized data, possibly containing transferred objects. Returns a vec of rooted transfer-received ports, or an error.

read_blob 🔒 ⚠

read_callback 🔒 ⚠

read_transfer_callback 🔒 ⚠

report_error_callback 🔒 ⚠

sab_cloned_callback 🔒 ⚠

write

Writes a structured clone. Returns a DataClone error if that fails.

write_blob 🔒 ⚠

write_callback 🔒 ⚠

write_transfer_callback 🔒 ⚠

https://html.spec.whatwg.org/multipage/#structuredserializewithtransfer