Module script::dom::bindings::trace

Utilities for tracing JS-managed values.

The lifetime of DOM objects is managed by the SpiderMonkey Garbage Collector. A rooted DOM object implementing the interface Foo is traced as follows:

The GC calls _trace defined in FooBinding during the marking phase. (This happens through JSClass.trace for non-proxy bindings, and through ProxyTraps.trace otherwise.)

_trace calls Foo::trace() (an implementation of JSTraceable). This is typically derived via a #[dom_struct] (implies #[jstraceable]) annotation. Non-JS-managed types have an empty inline trace() method, achieved via no_jsmanaged_fields! or similar.

For all fields, Foo::trace() calls trace() on the field. For example, for fields of type JS<T>, JS<T>::trace() calls trace_reflector().

trace_reflector() calls trace_object() with the JSObject for the reflector.

trace_object() calls JS_CallTracer() to notify the GC, which will add the object to the graph, and will trace that object as well.

When the GC finishes tracing, it finalizes any reflectors that were not reachable.

The no_jsmanaged_fields!() macro adds an empty implementation of JSTraceable to a datatype.

Structs

RootedCollectionSet
Holds a set of vectors that need to be rooted

RootedVec
A vector of items that are rooted for the lifetime of this struct

Statics

ROOTED_COLLECTIONS

Traits

JSTraceable
A trait to allow tracing (only) DOM objects.

Functions

trace_collections
SM Callback that traces the rooted collections

trace_jsval
Trace a JSVal.

trace_object
Trace a JSObject.

trace_reflector
Trace the JSObject held by reflector.