Module script::dom::bindings::utils

Various utilities to glue JavaScript and the DOM implementation together.

Structs

DOMClass
The struct that holds inheritance information for DOM object reflectors.

DOMJSClass
The JSClass used for DOM object reflectors.

GlobalStaticData
Static data associated with a global object.

WindowProxyHandler
Proxy handler for a WindowProxy.

Constants

DOM_CALLBACKS

DOM_PROTOTYPE_SLOT
The index of the slot that contains a reference to the ProtoOrIfaceArray.

DOM_PROTO_UNFORGEABLE_HOLDER_SLOT
The index of the slot where the object holder of that interface's unforgeable members are defined.

JSCLASS_DOM_GLOBAL
The flag set on the JSClasses for DOM global objects.

Statics

WRAP_CALLBACKS
Callback table for use with JS_SetWrapObjectCallbacks

Traits

AsCCharPtrPtr

AsVoidPtr

Functions

call_setter⚠

create_typed_array
Creates a Float32 array

delete_property_by_id⚠
Deletes the property id from object.

enumerate_global⚠
Enumerate lazy properties of a global object.

finalize_global⚠
Drop the resources held by reserved slots of a global object

find_enum_value⚠
Find the enum equivelent of a string given by v in pairs. Returns Err(()) on JSAPI failure (there is a pending exception), and Ok((None, value)) if there was no matching string.

generic_call⚠

generic_getter⚠
Generic getter of IDL interface.

generic_lenient_getter⚠
Generic lenient getter of IDL interface.

generic_lenient_setter⚠
Generic lenient setter of IDL interface.

generic_method⚠
Generic method of IDL interface.

generic_setter⚠
Generic setter of IDL interface.

get_array_index_from_id⚠
Get an array index from the given jsid. Returns None if the given jsid is not an integer.

get_dictionary_property
Get the property with name property from object. Returns Err(()) on JSAPI failure (there is a pending exception), and Ok(false) if there was no property with the given name.

get_property_on_prototype⚠
Gets the property id on proxy's prototype. If it exists, *found is set to true and *vp to the value, otherwise *found is set to false.

get_proto_or_iface_array
Returns the ProtoOrIfaceArray for the given global object. Fails if global is not a DOM global object.

has_property_on_prototype⚠
Returns whether proxy has a property id on its prototype.

instance_class_has_proto_at_depth⚠

is_platform_object

is_platform_object_dynamic
Returns wether obj is a platform object using dynamic unwrap https://heycam.github.io/webidl/#dfn-platform-object

is_platform_object_static
Returns wether obj is a platform object using static unwrap https://heycam.github.io/webidl/#dfn-platform-object

pre_wrap⚠

resolve_global⚠
Resolve a lazy global property, for interface objects and named constructors.

set_dictionary_property
Set the property with name property from object. Returns Err(()) on JSAPI failure, or null object, and Ok(()) otherwise

to_frozen_array
Returns a JSVal representing a frozen array of ports

trace_global⚠
Trace the resources held by reserved slots of a global object

wrap⚠

Type Definitions

ProtoOrIfaceArray
An array of *mut JSObject of size PROTO_OR_IFACE_LENGTH.