Module script::dom::bindings::utils

Various utilities to glue JavaScript and the DOM implementation together.

Structs

ConstantSpec
Representation of an IDL constant.

DOMClass
The struct that holds inheritance information for DOM object reflectors.

DOMJSClass
The JSClass used for DOM object reflectors.

GlobalStaticData
Static data associated with a global object.

NativeProperties
Contains references to lists of methods, attributes, and constants for a given interface.

NativePropertyHooks
Helper structure for cross-origin wrappers for DOM binding objects.

Reflector
A struct to store a reference to the reflector of a DOM object.

WindowProxyHandler
Proxy handler for a WindowProxy.

Enums

ConstantVal
Representation of an IDL constant value.

XMLName
Results of xml_name_type.

Constants

DOM_PROTOTYPE_SLOT
The index of the slot that contains a reference to the ProtoOrIfaceArray.

JSCLASS_DOM_GLOBAL
The flag set on the JSClasses for DOM global objects.

Traits

Reflectable
A trait to provide access to the Reflector for a DOM object.

Functions

create_dom_global
Create a DOM global object with the given class.

delete_property_by_id
Deletes the property id from object.

do_create_interface_objects
Creates the interface prototype object and the interface object (if needed). Fails on JSAPI failure.

finalize_global
Drop the resources held by reserved slots of a global object

find_enum_string_index
Find the index of a string given by v in values. Returns Err(()) on JSAPI failure (there is a pending exception), and Ok(None) if there was no matching string.

get_array_index_from_id
Get an array index from the given jsid. Returns None if the given jsid is not an integer.

get_dictionary_property
Get the property with name property from object. Returns Err(()) on JSAPI failure (there is a pending exception), and Ok(None) if there was no property with the given name.

get_property_on_prototype
Gets the property id on proxy's prototype. If it exists, *found is set to true and *vp to the value, otherwise *found is set to false.

get_proto_or_iface_array
Returns the ProtoOrIfaceArray for the given global object. Fails if global is not a DOM global object.

has_property_on_prototype
Returns whether proxy has a property id on its prototype.

initialize_global
Construct and cache the ProtoOrIfaceArray for the given global. Fails if the argument is not a DOM global.

is_platform_object
Returns wether obj is a platform object http://heycam.github.io/webidl/#dfn-platform-object

outerize_global
Callback to outerize windows.

pre_wrap
Callback to outerize windows before wrapping.

reflect_dom_object
Create the reflector for a new DOM object and yield ownership to the reflector.

throwing_constructor
A throwing constructor, for those interfaces that have neither NoInterfaceObject nor Constructor.

wrap_for_same_compartment
Callback to outerize windows when wrapping.

xml_name_type
Check if an element name is valid. See http://www.w3.org/TR/xml/#NT-Name for details.

Type Definitions

NonNullJSNative
A JSNative that cannot be null.