Struct script::dom::blob::Blob

pub struct Blob {
    reflector_: Reflector,
    blob_impl: BlobImpl,
    typeString: String,
    isClosed_: Cell<bool>,
}

Fields

reflector_

blob_impl

typeString

isClosed_

Methods

impl Blob

fn type_id(&self) -> &'static BlobTypeId

impl Blob

fn new(global: GlobalRef, blob_impl: BlobImpl, typeString: &str) -> Root<Blob>

fn new_inherited(blob_impl: BlobImpl, typeString: &str) -> Blob

fn Constructor(global: GlobalRef, blobParts: Option<Vec<BlobOrString>>, blobPropertyBag: &BlobPropertyBag) -> Fallible<Root<Blob>>

fn get_slice(&self) -> Result<DataSlice, ()>

Get a slice to inner data, this might incur synchronous read and caching

fn get_slice_or_empty(&self) -> DataSlice

Try to get a slice, and if any exception happens, return the empty slice