script::dom::canvas::canvas2d::canvas_state

Struct CanvasState

pub(super) struct CanvasState {
    canvas_thread_sender: GenericSender<CanvasMsg>,
    canvas_id: CanvasId,
    size: Cell<Size2D<u64>>,
    state: DomRefCell<CanvasContextState>,
    origin_clean: Cell<bool>,
    image_cache: Arc<dyn ImageCache>,
    base_url: ServoUrl,
    origin: ImmutableOrigin,
    missing_image_urls: DomRefCell<Vec<ServoUrl>>,
    saved_states: DomRefCell<Vec<CanvasContextState>>,
    current_default_path: DomRefCell<Path>,
}

Fields
canvas_thread_sender: GenericSender<CanvasMsg>canvas_id: CanvasIdsize: Cell<Size2D<u64>>state: DomRefCell<CanvasContextState>origin_clean: Cell<bool>image_cache: Arc<dyn ImageCache>base_url: ServoUrl

The base URL for resolving CSS image URL values. Needed because of https://github.com/servo/servo/issues/17625

origin: ImmutableOriginmissing_image_urls: DomRefCell<Vec<ServoUrl>>

Any missing image URLs.

saved_states: DomRefCell<Vec<CanvasContextState>>current_default_path: DomRefCell<Path>

https://html.spec.whatwg.org/multipage/#current-default-path

Implementations

impl CanvasState

pub(super) fn new( global: &GlobalScope, size: Size2D<u64>, ) -> Option<CanvasState>

pub(super) fn set_image_key(&self, image_key: ImageKey)

pub(super) fn get_missing_image_urls(&self) -> &DomRefCell<Vec<ServoUrl>>

pub(super) fn get_canvas_id(&self) -> CanvasId

pub(super) fn is_paintable(&self) -> bool

pub(super) fn send_canvas_2d_msg(&self, msg: Canvas2dMsg)

pub(super) fn update_rendering(&self, canvas_epoch: Option<Epoch>) -> bool

Updates WR image and blocks on completion

pub(super) fn set_bitmap_dimensions(&self, size: Size2D<u64>)

https://html.spec.whatwg.org/multipage/#concept-canvas-set-bitmap-dimensions

pub(super) fn reset(&self)

https://html.spec.whatwg.org/multipage/#reset-the-rendering-context-to-its-default-state

fn reset_to_initial_state(&self)

https://html.spec.whatwg.org/multipage/#reset-the-rendering-context-to-its-default-state

pub(super) fn reset_bitmap(&self)

fn create_drawable_rect( &self, x: f64, y: f64, w: f64, h: f64, ) -> Option<Rect<f32>>

pub(super) fn origin_is_clean(&self) -> bool

fn set_origin_clean(&self, origin_clean: bool)

fn is_origin_clean(&self, source: CanvasImageSource<DomTypeHolder>) -> bool

https://html.spec.whatwg.org/multipage/#the-image-argument-is-not-origin-clean

fn fetch_image_data( &self, url: ServoUrl, cors_setting: Option<CorsSettings>, ) -> Option<Snapshot>

fn request_image_from_cache( &self, url: ServoUrl, cors_setting: Option<CorsSettings>, ) -> ImageResponse

fn draw_image_internal( &self, htmlcanvas: Option<&HTMLCanvasElement>, image: CanvasImageSource<DomTypeHolder>, sx: f64, sy: f64, sw: Option<f64>, sh: Option<f64>, dx: f64, dy: f64, dw: Option<f64>, dh: Option<f64>, ) -> ErrorResult

drawImage coordinates explained

 Source Image      Destination Canvas
+-------------+     +-------------+
|             |     |             |
|(sx,sy)      |     |(dx,dy)      |
|   +----+    |     |   +----+    |
|   |    |    |     |   |    |    |
|   |    |sh  |---->|   |    |dh  |
|   |    |    |     |   |    |    |
|   +----+    |     |   +----+    |
|     sw      |     |     dw      |
|             |     |             |
+-------------+     +-------------+

The rectangle (sx, sy, sw, sh) from the source image is copied on the rectangle (dx, dy, dh, dw) of the destination canvas

https://html.spec.whatwg.org/multipage/#dom-context-2d-drawimage

fn draw_html_image_element( &self, image: &HTMLImageElement, canvas: Option<&HTMLCanvasElement>, sx: f64, sy: f64, sw: Option<f64>, sh: Option<f64>, dx: f64, dy: f64, dw: Option<f64>, dh: Option<f64>, )

https://html.spec.whatwg.org/multipage/#dom-context-2d-drawimage

fn draw_html_video_element( &self, video: &HTMLVideoElement, canvas: Option<&HTMLCanvasElement>, sx: f64, sy: f64, sw: Option<f64>, sh: Option<f64>, dx: f64, dy: f64, dw: Option<f64>, dh: Option<f64>, )

https://html.spec.whatwg.org/multipage/#dom-context-2d-drawimage

fn draw_offscreen_canvas( &self, canvas: &OffscreenCanvas, htmlcanvas: Option<&HTMLCanvasElement>, sx: f64, sy: f64, sw: Option<f64>, sh: Option<f64>, dx: f64, dy: f64, dw: Option<f64>, dh: Option<f64>, ) -> ErrorResult

fn draw_html_canvas_element( &self, canvas: &HTMLCanvasElement, htmlcanvas: Option<&HTMLCanvasElement>, sx: f64, sy: f64, sw: Option<f64>, sh: Option<f64>, dx: f64, dy: f64, dw: Option<f64>, dh: Option<f64>, ) -> ErrorResult

fn fetch_and_draw_image_data( &self, canvas: Option<&HTMLCanvasElement>, url: ServoUrl, cors_setting: Option<CorsSettings>, sx: f64, sy: f64, sw: Option<f64>, sh: Option<f64>, dx: f64, dy: f64, dw: Option<f64>, dh: Option<f64>, ) -> ErrorResult

fn draw_image_bitmap( &self, bitmap: &ImageBitmap, canvas: Option<&HTMLCanvasElement>, sx: f64, sy: f64, sw: Option<f64>, sh: Option<f64>, dx: f64, dy: f64, dw: Option<f64>, dh: Option<f64>, )

https://html.spec.whatwg.org/multipage/#dom-context-2d-drawimage

pub(super) fn mark_as_dirty(&self, canvas: Option<&HTMLCanvasElement>)

fn adjust_source_dest_rects( &self, image_size: Size2D<u32>, sx: f64, sy: f64, sw: f64, sh: f64, dx: f64, dy: f64, dw: f64, dh: f64, ) -> (Rect<f64>, Rect<f64>)

It is used by DrawImage to calculate the size of the source and destination rectangles based on the drawImage call arguments source rectangle = area of the original image to be copied destination rectangle = area of the destination canvas where the source image is going to be drawn

fn update_transform(&self, transform: Transform2D<f64>)

pub(super) fn fill_rect(&self, x: f64, y: f64, width: f64, height: f64)

pub(super) fn clear_rect(&self, x: f64, y: f64, width: f64, height: f64)

pub(super) fn stroke_rect(&self, x: f64, y: f64, width: f64, height: f64)

pub(super) fn shadow_offset_x(&self) -> f64

pub(super) fn set_shadow_offset_x(&self, value: f64)

pub(super) fn shadow_offset_y(&self) -> f64

pub(super) fn set_shadow_offset_y(&self, value: f64)

pub(super) fn shadow_blur(&self) -> f64

pub(super) fn set_shadow_blur(&self, value: f64)

pub(super) fn shadow_color(&self) -> DOMString

pub(super) fn set_shadow_color( &self, canvas: Option<&HTMLCanvasElement>, value: DOMString, )

pub(super) fn stroke_style( &self, ) -> StringOrCanvasGradientOrCanvasPattern<DomTypeHolder>

pub(super) fn set_stroke_style( &self, canvas: Option<&HTMLCanvasElement>, value: StringOrCanvasGradientOrCanvasPattern<DomTypeHolder>, )

pub(super) fn fill_style( &self, ) -> StringOrCanvasGradientOrCanvasPattern<DomTypeHolder>

pub(super) fn set_fill_style( &self, canvas: Option<&HTMLCanvasElement>, value: StringOrCanvasGradientOrCanvasPattern<DomTypeHolder>, )

pub(super) fn create_linear_gradient( &self, global: &GlobalScope, x0: Finite<f64>, y0: Finite<f64>, x1: Finite<f64>, y1: Finite<f64>, can_gc: CanGc, ) -> DomRoot<CanvasGradient>

pub(super) fn create_radial_gradient( &self, global: &GlobalScope, x0: Finite<f64>, y0: Finite<f64>, r0: Finite<f64>, x1: Finite<f64>, y1: Finite<f64>, r1: Finite<f64>, can_gc: CanGc, ) -> Fallible<DomRoot<CanvasGradient>>

https://html.spec.whatwg.org/multipage/#dom-context-2d-createradialgradient

pub(super) fn create_pattern( &self, global: &GlobalScope, image: CanvasImageSource<DomTypeHolder>, repetition: DOMString, can_gc: CanGc, ) -> Fallible<Option<DomRoot<CanvasPattern>>>

https://html.spec.whatwg.org/multipage/#dom-context-2d-createpattern

pub(super) fn save(&self)

pub(super) fn restore(&self)

https://html.spec.whatwg.org/multipage/#dom-context-2d-restore

pub(super) fn global_alpha(&self) -> f64

pub(super) fn set_global_alpha(&self, alpha: f64)

pub(super) fn global_composite_operation(&self) -> DOMString

pub(super) fn set_global_composite_operation(&self, op_str: DOMString)

pub(super) fn image_smoothing_enabled(&self) -> bool

pub(super) fn set_image_smoothing_enabled(&self, value: bool)

pub(super) fn fill_text( &self, global_scope: &GlobalScope, canvas: Option<&HTMLCanvasElement>, text: DOMString, x: f64, y: f64, max_width: Option<f64>, )

pub(super) fn stroke_text( &self, global_scope: &GlobalScope, canvas: Option<&HTMLCanvasElement>, text: DOMString, x: f64, y: f64, max_width: Option<f64>, )

pub(super) fn measure_text( &self, global: &GlobalScope, canvas: Option<&HTMLCanvasElement>, text: DOMString, can_gc: CanGc, ) -> DomRoot<TextMetrics>

https://html.spec.whatwg.org/multipage/#text-preparation-algorithm https://html.spec.whatwg.org/multipage/#dom-context-2d-measuretext https://html.spec.whatwg.org/multipage/#textmetrics

pub(super) fn set_font( &self, canvas: Option<&HTMLCanvasElement>, value: DOMString, )

fn font_style(&self) -> ServoArc<Font>

pub(super) fn font(&self) -> DOMString

pub(super) fn text_align(&self) -> CanvasTextAlign

pub(super) fn set_text_align(&self, value: CanvasTextAlign)

pub(super) fn text_baseline(&self) -> CanvasTextBaseline

pub(super) fn set_text_baseline(&self, value: CanvasTextBaseline)

pub(super) fn direction(&self) -> CanvasDirection

pub(super) fn set_direction(&self, value: CanvasDirection)

pub(super) fn line_width(&self) -> f64

pub(super) fn set_line_width(&self, width: f64)

pub(super) fn line_cap(&self) -> CanvasLineCap

pub(super) fn set_line_cap(&self, cap: CanvasLineCap)

pub(super) fn line_join(&self) -> CanvasLineJoin

pub(super) fn set_line_join(&self, join: CanvasLineJoin)

pub(super) fn miter_limit(&self) -> f64

pub(super) fn set_miter_limit(&self, limit: f64)

pub(super) fn line_dash(&self) -> Vec<f64>

https://html.spec.whatwg.org/multipage/#dom-context-2d-getlinedash

pub(super) fn set_line_dash(&self, segments: Vec<f64>)

https://html.spec.whatwg.org/multipage/#dom-context-2d-setlinedash

pub(super) fn line_dash_offset(&self) -> f64

https://html.spec.whatwg.org/multipage/#dom-context-2d-linedashoffset

pub(super) fn set_line_dash_offset(&self, offset: f64)

<https://html.spec.whatwg.org/multipage/#dom-context-2d-linedashoffset?

pub(super) fn create_image_data( &self, global: &GlobalScope, sw: i32, sh: i32, can_gc: CanGc, ) -> Fallible<DomRoot<ImageData>>

pub(super) fn create_image_data_( &self, global: &GlobalScope, imagedata: &ImageData, can_gc: CanGc, ) -> Fallible<DomRoot<ImageData>>

pub(super) fn get_image_data( &self, canvas_size: Size2D<u32>, global: &GlobalScope, sx: i32, sy: i32, sw: i32, sh: i32, can_gc: CanGc, ) -> Fallible<DomRoot<ImageData>>

pub(super) fn put_image_data( &self, canvas_size: Size2D<u32>, imagedata: &ImageData, dx: i32, dy: i32, )

pub(super) fn put_image_data_( &self, canvas_size: Size2D<u32>, imagedata: &ImageData, dx: i32, dy: i32, dirty_x: i32, dirty_y: i32, dirty_width: i32, dirty_height: i32, )

https://html.spec.whatwg.org/multipage/#dom-context-2d-putimagedata

pub(super) fn draw_image( &self, canvas: Option<&HTMLCanvasElement>, image: CanvasImageSource<DomTypeHolder>, dx: f64, dy: f64, ) -> ErrorResult

pub(super) fn draw_image_( &self, canvas: Option<&HTMLCanvasElement>, image: CanvasImageSource<DomTypeHolder>, dx: f64, dy: f64, dw: f64, dh: f64, ) -> ErrorResult

pub(super) fn draw_image__( &self, canvas: Option<&HTMLCanvasElement>, image: CanvasImageSource<DomTypeHolder>, sx: f64, sy: f64, sw: f64, sh: f64, dx: f64, dy: f64, dw: f64, dh: f64, ) -> ErrorResult

https://html.spec.whatwg.org/multipage/#dom-context-2d-drawimage

pub(super) fn begin_path(&self)

pub(super) fn fill(&self, fill_rule: CanvasFillRule)

pub(super) fn fill_(&self, path: Path, fill_rule: CanvasFillRule)

pub(super) fn stroke(&self)

pub(super) fn stroke_(&self, path: Path)

pub(super) fn clip(&self, fill_rule: CanvasFillRule)

pub(super) fn clip_(&self, path: Path, fill_rule: CanvasFillRule)

pub(super) fn is_point_in_path( &self, global: &GlobalScope, x: f64, y: f64, fill_rule: CanvasFillRule, ) -> bool

pub(super) fn is_point_in_path_( &self, _global: &GlobalScope, path: Path, x: f64, y: f64, fill_rule: CanvasFillRule, ) -> bool

pub(super) fn scale(&self, x: f64, y: f64)

pub(super) fn rotate(&self, angle: f64)

pub(super) fn translate(&self, x: f64, y: f64)

pub(super) fn transform(&self, a: f64, b: f64, c: f64, d: f64, e: f64, f: f64)

pub(super) fn get_transform( &self, global: &GlobalScope, can_gc: CanGc, ) -> DomRoot<DOMMatrix>

pub(super) fn set_transform( &self, a: f64, b: f64, c: f64, d: f64, e: f64, f: f64, )

https://html.spec.whatwg.org/multipage/#dom-context-2d-settransform

pub(super) fn set_transform_(&self, transform: &DOMMatrix2DInit) -> ErrorResult

https://html.spec.whatwg.org/multipage/#dom-context-2d-settransform-matrix

pub(super) fn reset_transform(&self)

pub(super) fn close_path(&self)

pub(super) fn move_to(&self, x: f64, y: f64)

pub(super) fn line_to(&self, x: f64, y: f64)

pub(super) fn rect(&self, x: f64, y: f64, width: f64, height: f64)

pub(super) fn quadratic_curve_to(&self, cpx: f64, cpy: f64, x: f64, y: f64)

pub(super) fn bezier_curve_to( &self, cp1x: f64, cp1y: f64, cp2x: f64, cp2y: f64, x: f64, y: f64, )

pub(super) fn arc( &self, x: f64, y: f64, r: f64, start: f64, end: f64, ccw: bool, ) -> ErrorResult

pub(super) fn arc_to( &self, cp1x: f64, cp1y: f64, cp2x: f64, cp2y: f64, r: f64, ) -> ErrorResult

pub(super) fn ellipse( &self, x: f64, y: f64, rx: f64, ry: f64, rotation: f64, start: f64, end: f64, ccw: bool, ) -> ErrorResult

https://html.spec.whatwg.org/multipage/#dom-context-2d-ellipse

fn text_with_size( &self, global_scope: &GlobalScope, text: &str, origin: Point2D<f64>, size: f64, max_width: Option<f64>, ) -> Option<(Rect<f64>, Vec<TextRun>)>

fn build_unshaped_text_runs<'text>( &self, font_context: &FontContext, text: &'text str, font_group: &FontGroup, ) -> Vec<UnshapedTextRun<'text>>

fn find_anchor_point_for_line_of_text( &self, origin: Point2D<f64>, metrics: &FontMetrics, width: f64, ) -> Point2D<f64>

Find the anchor_point for the given parameters of a line of text. See https://html.spec.whatwg.org/multipage/#text-preparation-algorithm.