Struct script::dom::dissimilaroriginwindow::DissimilarOriginWindow

#[repr(C)]

pub struct DissimilarOriginWindow {
    globalscope: GlobalScope,
    browsing_context: JS<BrowsingContext>,
    location: MutNullableJS<DissimilarOriginLocation>,
}

Represents a dissimilar-origin Window that exists in another script thread.

Since the Window is in a different script thread, we cannot access it directly, but some of its accessors (for example window.parent) still need to function.

In browsingcontext.rs, we create a custom window proxy for these windows, that throws security exceptions for most accessors. This is not a replacement for XOWs, but provides belt-and-braces security.

Fields
globalscope: GlobalScope

The global for this window.

browsing_context: JS<BrowsingContext>

The browsing context this window is part of.

location: MutNullableJS<DissimilarOriginLocation>

The location of this window, initialized lazily.

Methods

impl DissimilarOriginWindow

fn new(browsing_context: &BrowsingContext) -> Root<DissimilarOriginWindow>