Module script::dom::document

Structs

AnchorsFilter

Document

https://dom.spec.whatwg.org/#document

EmbedsFilter

FakeRequestAnimationFrameCallback

A fake requestAnimationFrame() callback—“fake” because it is not triggered by the video refresh but rather a simple timer.

FormsFilter

ImagesFilter

LinksFilter

PendingInOrderScriptVec

PendingScript

ScriptsFilter

Enums

AnimationFrameCallback

DocumentSource

ElementLookupResult

FireMouseEventType

FocusEventType

Focus events

FocusTransaction

FocusType

Specifies the type of focus event that is sent to a pipeline

HasBrowsingContext

IsHTMLDocument

ReflowTriggerCondition

TouchEventResult

Constants

FAKE_REQUEST_ANIMATION_FRAME_DELAY

The amount of time between fake requestAnimationFrame()s.

SPURIOUS_ANIMATION_FRAME_THRESHOLD

The number of times we are allowed to see spurious requestAnimationFrame() calls before falling back to fake ones.

_IMPL_DOMOBJECT_FOR_Document

Traits

LayoutDocumentHelpers

Functions

determine_policy_for_token

https://w3c.github.io/webappsec-referrer-policy/#determine-policy-for-token

get_registrable_domain_suffix_of_or_is_equal_to

is_character_value_key

update_with_current_time_ms

url_has_network_scheme

https://url.spec.whatwg.org/#network-scheme