Struct script::dom::document::Document

#[repr(C)]

pub struct Document {
    node: Node,
    window: JS<Window>,
    implementation: MutNullableJS<DOMImplementation>,
    location: MutNullableJS<Location>,
    content_type: DOMString,
    last_modified: Option<String>,
    encoding: Cell<EncodingRef>,
    has_browsing_context: bool,
    is_html_document: bool,
    activity: Cell<DocumentActivity>,
    url: DOMRefCell<ServoUrl>,
    quirks_mode: Cell<QuirksMode>,
    id_map: DOMRefCell<HashMap<Atom, Vec<JS<Element>>>>,
    tag_map: DOMRefCell<HashMap<LocalName, JS<HTMLCollection>>>,
    tagns_map: DOMRefCell<HashMap<QualName, JS<HTMLCollection>>>,
    classes_map: DOMRefCell<HashMap<Vec<Atom>, JS<HTMLCollection>>>,
    images: MutNullableJS<HTMLCollection>,
    embeds: MutNullableJS<HTMLCollection>,
    links: MutNullableJS<HTMLCollection>,
    forms: MutNullableJS<HTMLCollection>,
    scripts: MutNullableJS<HTMLCollection>,
    anchors: MutNullableJS<HTMLCollection>,
    applets: MutNullableJS<HTMLCollection>,
    stylesheets: DOMRefCell<Option<Vec<StylesheetInDocument>>>,
    stylesheets_changed_since_reflow: Cell<bool>,
    stylesheet_list: MutNullableJS<StyleSheetList>,
    ready_state: Cell<DocumentReadyState>,
    domcontentloaded_dispatched: Cell<bool>,
    possibly_focused: MutNullableJS<Element>,
    focused: MutNullableJS<Element>,
    current_script: MutNullableJS<HTMLScriptElement>,
    pending_parsing_blocking_script: DOMRefCell<Option<PendingScript>>,
    script_blocking_stylesheets_count: Cell<u32>,
    deferred_scripts: PendingInOrderScriptVec,
    asap_in_order_scripts_list: PendingInOrderScriptVec,
    asap_scripts_set: DOMRefCell<Vec<JS<HTMLScriptElement>>>,
    scripting_enabled: bool,
    animation_frame_ident: Cell<u32>,
    animation_frame_list: DOMRefCell<Vec<(u32, Option<AnimationFrameCallback>)>>,
    running_animation_callbacks: Cell<bool>,
    loader: DOMRefCell<DocumentLoader>,
    current_parser: MutNullableJS<ServoParser>,
    reflow_timeout: Cell<Option<u64>>,
    base_element: MutNullableJS<HTMLBaseElement>,
    appropriate_template_contents_owner_document: MutNullableJS<Document>,
    pending_restyles: DOMRefCell<HashMap<JS<Element>, PendingRestyle>>,
    needs_paint: Cell<bool>,
    active_touch_points: DOMRefCell<Vec<JS<Touch>>>,
    dom_loading: Cell<u64>,
    dom_interactive: Cell<u64>,
    dom_content_loaded_event_start: Cell<u64>,
    dom_content_loaded_event_end: Cell<u64>,
    dom_complete: Cell<u64>,
    load_event_start: Cell<u64>,
    load_event_end: Cell<u64>,
    https_state: Cell<HttpsState>,
    touchpad_pressure_phase: Cell<TouchpadPressurePhase>,
    origin: Origin,
    referrer_policy: Cell<Option<ReferrerPolicy>>,
    referrer: Option<String>,
    target_element: MutNullableJS<Element>,
    last_click_info: DOMRefCell<Option<(Instant, Point2D<f32>)>>,
    ignore_destructive_writes_counter: Cell<u32>,
    dom_count: Cell<u32>,
    fullscreen_element: MutNullableJS<Element>,
}

Fields
node: Node window: JS<Window> implementation: MutNullableJS<DOMImplementation> location: MutNullableJS<Location> content_type: DOMString last_modified: Option<String> encoding: Cell<EncodingRef> has_browsing_context: bool is_html_document: bool activity: Cell<DocumentActivity> url: DOMRefCell<ServoUrl> quirks_mode: Cell<QuirksMode> id_map: DOMRefCell<HashMap<Atom, Vec<JS<Element>>>>

Caches for the getElement methods

tag_map: DOMRefCell<HashMap<LocalName, JS<HTMLCollection>>> tagns_map: DOMRefCell<HashMap<QualName, JS<HTMLCollection>>> classes_map: DOMRefCell<HashMap<Vec<Atom>, JS<HTMLCollection>>> images: MutNullableJS<HTMLCollection> embeds: MutNullableJS<HTMLCollection> links: MutNullableJS<HTMLCollection> forms: MutNullableJS<HTMLCollection> scripts: MutNullableJS<HTMLCollection> anchors: MutNullableJS<HTMLCollection> applets: MutNullableJS<HTMLCollection> stylesheets: DOMRefCell<Option<Vec<StylesheetInDocument>>>

List of stylesheets associated with nodes in this document. |None| if the list needs to be refreshed.

stylesheets_changed_since_reflow: Cell<bool>

Whether the list of stylesheets has changed since the last reflow was triggered.

stylesheet_list: MutNullableJS<StyleSheetList> ready_state: Cell<DocumentReadyState> domcontentloaded_dispatched: Cell<bool>

Whether the DOMContentLoaded event has already been dispatched.

possibly_focused: MutNullableJS<Element>

The element that has most recently requested focus for itself.

focused: MutNullableJS<Element>

The element that currently has the document focus context.

current_script: MutNullableJS<HTMLScriptElement>

The script element that is currently executing.

pending_parsing_blocking_script: DOMRefCell<Option<PendingScript>>

https://html.spec.whatwg.org/multipage/#pending-parsing-blocking-script

script_blocking_stylesheets_count: Cell<u32>

Number of stylesheets that block executing the next parser-inserted script

deferred_scripts: PendingInOrderScriptVec

https://html.spec.whatwg.org/multipage/#list-of-scripts-that-will-execute-when-the-document-has-finished-parsing

asap_in_order_scripts_list: PendingInOrderScriptVec

https://html.spec.whatwg.org/multipage/#list-of-scripts-that-will-execute-in-order-as-soon-as-possible

asap_scripts_set: DOMRefCell<Vec<JS<HTMLScriptElement>>>

https://html.spec.whatwg.org/multipage/#set-of-scripts-that-will-execute-as-soon-as-possible

scripting_enabled: bool

https://html.spec.whatwg.org/multipage/#concept-n-noscript True if scripting is enabled for all scripts in this document

animation_frame_ident: Cell<u32>

https://html.spec.whatwg.org/multipage/#animation-frame-callback-identifier Current identifier of animation frame callback

animation_frame_list: DOMRefCell<Vec<(u32, Option<AnimationFrameCallback>)>>

https://html.spec.whatwg.org/multipage/#list-of-animation-frame-callbacks List of animation frame callbacks

running_animation_callbacks: Cell<bool>

Whether we're in the process of running animation callbacks.

Tracking this is not necessary for correctness. Instead, it is an optimization to avoid sending needless ChangeRunningAnimationsState messages to the compositor.

loader: DOMRefCell<DocumentLoader>

Tracks all outstanding loads related to this document.

current_parser: MutNullableJS<ServoParser>

The current active HTML parser, to allow resuming after interruptions.

reflow_timeout: Cell<Option<u64>>

When we should kick off a reflow. This happens during parsing.

base_element: MutNullableJS<HTMLBaseElement>

The cached first base element with an href attribute.

appropriate_template_contents_owner_document: MutNullableJS<Document>

This field is set to the document itself for inert documents. https://html.spec.whatwg.org/multipage/#appropriate-template-contents-owner-document

pending_restyles: DOMRefCell<HashMap<JS<Element>, PendingRestyle>>

Information on elements needing restyle to ship over to the layout thread when the time comes.

needs_paint: Cell<bool>

This flag will be true if layout suppressed a reflow attempt that was needed in order for the page to be painted.

active_touch_points: DOMRefCell<Vec<JS<Touch>>>

http://w3c.github.io/touch-events/#dfn-active-touch-point

dom_loading: Cell<u64>

Navigation Timing properties: https://w3c.github.io/navigation-timing/#sec-PerformanceNavigationTiming

dom_interactive: Cell<u64> dom_content_loaded_event_start: Cell<u64> dom_content_loaded_event_end: Cell<u64> dom_complete: Cell<u64> load_event_start: Cell<u64> load_event_end: Cell<u64> https_state: Cell<HttpsState>

https://html.spec.whatwg.org/multipage/#concept-document-https-state

touchpad_pressure_phase: Cell<TouchpadPressurePhase> origin: Origin

The document's origin.

referrer_policy: Cell<Option<ReferrerPolicy>>

https://w3c.github.io/webappsec-referrer-policy/#referrer-policy-states

referrer: Option<String>

https://html.spec.whatwg.org/multipage/#dom-document-referrer

target_element: MutNullableJS<Element>

https://html.spec.whatwg.org/multipage/#target-element

last_click_info: DOMRefCell<Option<(Instant, Point2D<f32>)>>

https://w3c.github.io/uievents/#event-type-dblclick

ignore_destructive_writes_counter: Cell<u32>

https://html.spec.whatwg.org/multipage/#ignore-destructive-writes-counter

dom_count: Cell<u32>

Track the total number of elements in this DOM's tree. This is sent to the layout thread every time a reflow is done; layout uses this to determine if the gains from parallel layout will be worth the overhead.

See also: https://github.com/servo/servo/issues/10110

fullscreen_element: MutNullableJS<Element>

Entry node for fullscreen.

Methods

impl Document

fn loader(&self) -> Ref<DocumentLoader>

fn mut_loader(&self) -> RefMut<DocumentLoader>

fn browsing_context(&self) -> Option<Root<BrowsingContext>>

https://html.spec.whatwg.org/multipage/#concept-document-bc

fn window(&self) -> &Window

fn is_html_document(&self) -> bool

fn set_https_state(&self, https_state: HttpsState)

fn is_fully_active(&self) -> bool

fn is_active(&self) -> bool

fn set_activity(&self, activity: DocumentActivity)

fn origin(&self) -> &Origin

fn url(&self) -> ServoUrl

fn set_url(&self, url: ServoUrl)

fn fallback_base_url(&self) -> ServoUrl

fn base_url(&self) -> ServoUrl

fn needs_paint(&self) -> bool

fn needs_reflow(&self) -> bool

fn base_element(&self) -> Option<Root<HTMLBaseElement>>

Returns the first base element in the DOM that has an href attribute.

fn refresh_base_element(&self)

Refresh the cached first base element in the DOM. https://github.com/w3c/web-platform-tests/issues/2122

fn dom_count(&self) -> u32

fn increment_dom_count(&self)

This is called by bind_to_tree when a node is added to the DOM. The internal count is used by layout to determine whether to be sequential or parallel. (it's sequential for small DOMs)

fn decrement_dom_count(&self)

This is called by unbind_from_tree when a node is removed from the DOM.

fn quirks_mode(&self) -> QuirksMode

fn set_quirks_mode(&self, mode: QuirksMode)

fn encoding(&self) -> EncodingRef

fn set_encoding(&self, encoding: EncodingRef)

fn content_and_heritage_changed(&self, node: &Node, damage: NodeDamage)

fn reflow_if_reflow_timer_expired(&self)

Reflows and disarms the timer if the reflow timer has expired.

fn set_reflow_timeout(&self, timeout: u64)

Schedules a reflow to be kicked off at the given timeout (in time::precise_time_ns() units). This reflow happens even if the event loop is busy. This is used to display initial page content during parsing.

fn unregister_named_element(&self, to_unregister: &Element, id: Atom)

Remove any existing association between the provided id and any elements in this document.

fn register_named_element(&self, element: &Element, id: Atom)

Associate an element present in this document with the provided id.

fn find_fragment_node(&self, fragid: &str) -> Option<Root<Element>>

Attempt to find a named element in this page's document. https://html.spec.whatwg.org/multipage/#the-indicated-part-of-the-document

fn check_and_scroll_fragment(&self, fragment: &str)

Scroll to the target element, and when we do not find a target and the fragment is empty or "top", scroll to the top. https://html.spec.whatwg.org/multipage/#scroll-to-the-fragment-identifier

fn get_anchor_by_name(&self, name: &str) -> Option<Root<Element>>

fn set_ready_state(&self, state: DocumentReadyState)

fn is_scripting_enabled(&self) -> bool

Return whether scripting is enabled or not

fn get_focused_element(&self) -> Option<Root<Element>>

Return the element that currently has focus.

fn begin_focus_transaction(&self)

Initiate a new round of checking for elements requesting focus. The last element to call request_focus before commit_focus_transaction is called will receive focus.

fn request_focus(&self, elem: &Element)

Request that the given element receive focus once the current transaction is complete.

fn commit_focus_transaction(&self, focus_type: FocusType)

Reassign the focus context to the element that last requested focus during this transaction, or none if no elements requested it.

fn title_changed(&self)

Handles any updates when the document's title has changed.

fn send_title_to_compositor(&self)

Sends this document's title to the compositor.

fn dirty_all_nodes(&self)

fn handle_mouse_event(&self,
js_runtime: *mut JSRuntime,
button: MouseButton,
client_point: Point2D<f32>,
mouse_event_type: MouseEventType)

fn maybe_fire_dblclick(&self, click_pos: Point2D<f32>, target: &Node)

fn handle_touchpad_pressure_event(&self,
js_runtime: *mut JSRuntime,
client_point: Point2D<f32>,
pressure: f32,
phase_now: TouchpadPressurePhase)

fn fire_forcetouch_event(&self,
event_name: String,
target: &EventTarget,
force: f32)

fn fire_mouse_event(&self,
client_point: Point2D<f32>,
target: &EventTarget,
event_name: String)

fn handle_mouse_move_event(&self,
js_runtime: *mut JSRuntime,
client_point: Option<Point2D<f32>>,
prev_mouse_over_target: &MutNullableJS<Element>)

fn handle_touch_event(&self,
js_runtime: *mut JSRuntime,
event_type: TouchEventType,
touch_id: TouchId,
point: Point2D<f32>)
-> TouchEventResult

fn dispatch_key_event(&self,
ch: Option<char>,
key: Key,
state: KeyState,
modifiers: KeyModifiers,
constellation: &IpcSender<ConstellationMsg>)

The entry point for all key processing for web content

fn node_from_nodes_and_strings(&self,
nodes: Vec<NodeOrString>)
-> Result<Root<Node>, Error>

fn get_body_attribute(&self, local_name: &LocalName) -> DOMString

fn set_body_attribute(&self, local_name: &LocalName, value: DOMString)

fn set_current_script(&self, script: Option<&HTMLScriptElement>)

fn get_script_blocking_stylesheets_count(&self) -> u32

fn increment_script_blocking_stylesheet_count(&self)

fn decrement_script_blocking_stylesheet_count(&self)

fn invalidate_stylesheets(&self)

fn get_and_reset_stylesheets_changed_since_reflow(&self) -> bool

fn trigger_mozbrowser_event(&self, event: MozBrowserEvent)

fn request_animation_frame(&self, callback: AnimationFrameCallback) -> u32

https://html.spec.whatwg.org/multipage/#dom-window-requestanimationframe

fn cancel_animation_frame(&self, ident: u32)

https://html.spec.whatwg.org/multipage/#dom-window-cancelanimationframe

fn run_the_animation_frame_callbacks(&self)

https://html.spec.whatwg.org/multipage/#run-the-animation-frame-callbacks

fn fetch_async(&self,
load: LoadType,
request: RequestInit,
fetch_target: IpcSender<FetchResponseMsg>)

fn finish_load(&self, load: LoadType)

fn set_pending_parsing_blocking_script(&self,
script: &HTMLScriptElement,
load: Option<Result<ClassicScript, NetworkError>>)

fn has_pending_parsing_blocking_script(&self) -> bool

fn pending_parsing_blocking_script_loaded(&self,
element: &HTMLScriptElement,
result: Result<ClassicScript, NetworkError>)

https://html.spec.whatwg.org/multipage/#prepare-a-script step 22.d.

fn process_pending_parsing_blocking_script(&self)

fn add_asap_script(&self, script: &HTMLScriptElement)

fn asap_script_loaded(&self,
element: &HTMLScriptElement,
result: Result<ClassicScript, NetworkError>)

https://html.spec.whatwg.org/multipage/#the-end step 5. https://html.spec.whatwg.org/multipage/#prepare-a-script step 22.d.

fn push_asap_in_order_script(&self, script: &HTMLScriptElement)

fn asap_in_order_script_loaded(&self,
element: &HTMLScriptElement,
result: Result<ClassicScript, NetworkError>)

https://html.spec.whatwg.org/multipage/#the-end step 5. https://html.spec.whatwg.org/multipage/#prepare-a-script step 22.c.

fn add_deferred_script(&self, script: &HTMLScriptElement)

fn deferred_script_loaded(&self,
element: &HTMLScriptElement,
result: Result<ClassicScript, NetworkError>)

https://html.spec.whatwg.org/multipage/#the-end step 3. https://html.spec.whatwg.org/multipage/#prepare-a-script step 22.d.

fn process_deferred_scripts(&self)

https://html.spec.whatwg.org/multipage/#the-end step 3.

fn maybe_dispatch_dom_content_loaded(&self)

fn notify_constellation_load(&self)

fn set_current_parser(&self, script: Option<&ServoParser>)

fn get_current_parser(&self) -> Option<Root<ServoParser>>

fn iter_iframes(&self) -> impl Iterator<Item=Root<HTMLIFrameElement>>

Iterate over all iframes in the document.

fn find_iframe(&self, frame_id: FrameId) -> Option<Root<HTMLIFrameElement>>

Find an iframe element in the document.

fn get_dom_loading(&self) -> u64

fn get_dom_interactive(&self) -> u64

fn get_dom_content_loaded_event_start(&self) -> u64

fn get_dom_content_loaded_event_end(&self) -> u64

fn get_dom_complete(&self) -> u64

fn get_load_event_start(&self) -> u64

fn get_load_event_end(&self) -> u64

fn fire_focus_event(&self,
focus_event_type: FocusEventType,
node: &Node,
related_target: Option<&EventTarget>)

fn is_cookie_averse(&self) -> bool

https://html.spec.whatwg.org/multipage/#cookie-averse-document-object

fn nodes_from_point(&self,
client_point: &Point2D<f32>)
-> Vec<UntrustedNodeAddress>

impl Document

fn new_inherited(window: &Window,
has_browsing_context: HasBrowsingContext,
url: Option<ServoUrl>,
origin: Origin,
is_html_document: IsHTMLDocument,
content_type: Option<DOMString>,
last_modified: Option<String>,
activity: DocumentActivity,
source: DocumentSource,
doc_loader: DocumentLoader,
referrer: Option<String>,
referrer_policy: Option<ReferrerPolicy>)
-> Document

fn Constructor(window: &Window) -> Result<Root<Document>, Error>

fn new(window: &Window,
has_browsing_context: HasBrowsingContext,
url: Option<ServoUrl>,
origin: Origin,
doctype: IsHTMLDocument,
content_type: Option<DOMString>,
last_modified: Option<String>,
activity: DocumentActivity,
source: DocumentSource,
doc_loader: DocumentLoader,
referrer: Option<String>,
referrer_policy: Option<ReferrerPolicy>)
-> Root<Document>

fn create_node_list<F: Fn(&Node) -> bool>(&self, callback: F) -> Root<NodeList>

fn get_html_element(&self) -> Option<Root<HTMLHtmlElement>>

fn ensure_stylesheets(&self)

fn stylesheets(&self) -> Vec<Arc<Stylesheet>>

Returns the list of stylesheets associated with nodes in the document.

fn with_style_sheets_in_document<F, T>(&self, f: F) -> T where F: FnMut(&[StylesheetInDocument]) -> T

fn appropriate_template_contents_owner_document(&self) -> Root<Document>

https://html.spec.whatwg.org/multipage/#appropriate-template-contents-owner-document

fn get_element_by_id(&self, id: &Atom) -> Option<Root<Element>>

fn ensure_pending_restyle(&self, el: &Element) -> RefMut<PendingRestyle>

fn ensure_snapshot(&self, el: &Element) -> RefMut<Snapshot>

fn element_state_will_change(&self, el: &Element)

fn element_attr_will_change(&self, el: &Element)

fn set_referrer_policy(&self, policy: Option<ReferrerPolicy>)

fn get_referrer_policy(&self) -> Option<ReferrerPolicy>

fn set_target_element(&self, node: Option<&Element>)

fn incr_ignore_destructive_writes_counter(&self)

fn decr_ignore_destructive_writes_counter(&self)

fn enter_fullscreen(&self, pending: &Element) -> Rc<Promise>

fn exit_fullscreen(&self) -> Rc<Promise>

fn set_fullscreen_element(&self, element: Option<&Element>)

fn get_allow_fullscreen(&self) -> bool