Struct script::dom::document::Document

pub struct Document {
    node: Node,
    window: JS<Window>,
    browsing_context: Option<JS<BrowsingContext>>,
    implementation: MutNullableHeap<JS<DOMImplementation>>,
    location: MutNullableHeap<JS<Location>>,
    content_type: DOMString,
    last_modified: Option<String>,
    encoding: Cell<EncodingRef>,
    is_html_document: bool,
    url: Url,
    quirks_mode: Cell<QuirksMode>,
    id_map: DOMRefCell<HashMap<Atom, Vec<JS<Element>>>>,
    tag_map: DOMRefCell<HashMap<Atom, JS<HTMLCollection>>>,
    tagns_map: DOMRefCell<HashMap<QualName, JS<HTMLCollection>>>,
    classes_map: DOMRefCell<HashMap<Vec<Atom>, JS<HTMLCollection>>>,
    images: MutNullableHeap<JS<HTMLCollection>>,
    embeds: MutNullableHeap<JS<HTMLCollection>>,
    links: MutNullableHeap<JS<HTMLCollection>>,
    forms: MutNullableHeap<JS<HTMLCollection>>,
    scripts: MutNullableHeap<JS<HTMLCollection>>,
    anchors: MutNullableHeap<JS<HTMLCollection>>,
    applets: MutNullableHeap<JS<HTMLCollection>>,
    stylesheets: DOMRefCell<Option<Vec<(JS<Node>, Arc<Stylesheet>)>>>,
    stylesheets_changed_since_reflow: Cell<bool>,
    ready_state: Cell<DocumentReadyState>,
    domcontentloaded_dispatched: Cell<bool>,
    possibly_focused: MutNullableHeap<JS<Element>>,
    focused: MutNullableHeap<JS<Element>>,
    current_script: MutNullableHeap<JS<HTMLScriptElement>>,
    pending_parsing_blocking_script: MutNullableHeap<JS<HTMLScriptElement>>,
    script_blocking_stylesheets_count: Cell<u32>,
    deferred_scripts: DOMRefCell<Vec<JS<HTMLScriptElement>>>,
    asap_in_order_scripts_list: DOMRefCell<Vec<JS<HTMLScriptElement>>>,
    asap_scripts_set: DOMRefCell<Vec<JS<HTMLScriptElement>>>,
    scripting_enabled: Cell<bool>,
    animation_frame_ident: Cell<u32>,
    animation_frame_list: DOMRefCell<BTreeMap<u32, Box<FnBox(f64)>>>,
    running_animation_callbacks: Cell<bool>,
    loader: DOMRefCell<DocumentLoader>,
    current_parser: MutNullableParserField,
    reflow_timeout: Cell<Option<u64>>,
    base_element: MutNullableHeap<JS<HTMLBaseElement>>,
    appropriate_template_contents_owner_document: MutNullableHeap<JS<Document>>,
    modified_elements: DOMRefCell<HashMap<JS<Element>, ElementSnapshot>>,
    active_touch_points: DOMRefCell<Vec<JS<Touch>>>,
    dom_loading: Cell<u64>,
    dom_interactive: Cell<u64>,
    dom_content_loaded_event_start: Cell<u64>,
    dom_content_loaded_event_end: Cell<u64>,
    dom_complete: Cell<u64>,
    load_event_start: Cell<u64>,
    load_event_end: Cell<u64>,
    css_errors_store: DOMRefCell<Vec<CSSError>>,
    https_state: Cell<HttpsState>,
    touchpad_pressure_phase: Cell<TouchpadPressurePhase>,
    origin: Origin,
    referrer_policy: Option<ReferrerPolicy>,
}

Fields

node

window

browsing_context

https://html.spec.whatwg.org/multipage/#concept-document-bc

implementation

location

content_type

last_modified

encoding

is_html_document

url

quirks_mode

id_map

Caches for the getElement methods

tag_map

tagns_map

classes_map

images

embeds

links

forms

scripts

anchors

applets

stylesheets

List of stylesheets associated with nodes in this document. |None| if the list needs to be refreshed.

stylesheets_changed_since_reflow

Whether the list of stylesheets has changed since the last reflow was triggered.

ready_state

domcontentloaded_dispatched

Whether the DOMContentLoaded event has already been dispatched.

possibly_focused

The element that has most recently requested focus for itself.

focused

The element that currently has the document focus context.

current_script

The script element that is currently executing.

pending_parsing_blocking_script

https://html.spec.whatwg.org/multipage/#pending-parsing-blocking-script

script_blocking_stylesheets_count

Number of stylesheets that block executing the next parser-inserted script

deferred_scripts

https://html.spec.whatwg.org/multipage/#list-of-scripts-that-will-execute-when-the-document-has-finished-parsing

asap_in_order_scripts_list

https://html.spec.whatwg.org/multipage/#list-of-scripts-that-will-execute-in-order-as-soon-as-possible

asap_scripts_set

https://html.spec.whatwg.org/multipage/#set-of-scripts-that-will-execute-as-soon-as-possible

scripting_enabled

https://html.spec.whatwg.org/multipage/#concept-n-noscript True if scripting is enabled for all scripts in this document

animation_frame_ident

https://html.spec.whatwg.org/multipage/#animation-frame-callback-identifier Current identifier of animation frame callback

animation_frame_list

https://html.spec.whatwg.org/multipage/#list-of-animation-frame-callbacks List of animation frame callbacks

running_animation_callbacks

Whether we're in the process of running animation callbacks.

Tracking this is not necessary for correctness. Instead, it is an optimization to avoid sending needless ChangeRunningAnimationsState messages to the compositor.

loader

Tracks all outstanding loads related to this document.

current_parser

The current active HTML parser, to allow resuming after interruptions.

reflow_timeout

When we should kick off a reflow. This happens during parsing.

base_element

The cached first base element with an href attribute.

appropriate_template_contents_owner_document

This field is set to the document itself for inert documents. https://html.spec.whatwg.org/multipage/#appropriate-template-contents-owner-document

modified_elements

For each element that has had a state or attribute change since the last restyle, track the original condition of the element.

active_touch_points

http://w3c.github.io/touch-events/#dfn-active-touch-point

dom_loading

Navigation Timing properties: https://w3c.github.io/navigation-timing/#sec-PerformanceNavigationTiming

dom_interactive

dom_content_loaded_event_start

dom_content_loaded_event_end

dom_complete

load_event_start

load_event_end

css_errors_store

Vector to store CSS errors

https_state

https://html.spec.whatwg.org/multipage/#concept-document-https-state

touchpad_pressure_phase

origin

The document's origin.

referrer_policy

https://w3c.github.io/webappsec-referrer-policy/#referrer-policy-states

Methods

impl Document

fn loader(&self) -> Ref<DocumentLoader>

fn mut_loader(&self) -> RefMut<DocumentLoader>

fn browsing_context(&self) -> Option<&BrowsingContext>

https://html.spec.whatwg.org/multipage/#concept-document-bc

fn window(&self) -> &Window

fn is_html_document(&self) -> bool

fn set_https_state(&self, https_state: HttpsState)

fn report_css_error(&self, css_error: CSSError)

fn is_fully_active(&self) -> bool

fn url(&self) -> &Url

fn fallback_base_url(&self) -> Url

fn base_url(&self) -> Url

fn needs_reflow(&self) -> bool

fn base_element(&self) -> Option<Root<HTMLBaseElement>>

Returns the first base element in the DOM that has an href attribute.

fn refresh_base_element(&self)

Refresh the cached first base element in the DOM. https://github.com/w3c/web-platform-tests/issues/2122

fn quirks_mode(&self) -> QuirksMode

fn set_quirks_mode(&self, mode: QuirksMode)

fn encoding(&self) -> EncodingRef

fn set_encoding(&self, encoding: EncodingRef)

fn content_and_heritage_changed(&self, node: &Node, damage: NodeDamage)

fn reflow_if_reflow_timer_expired(&self)

Reflows and disarms the timer if the reflow timer has expired.

fn set_reflow_timeout(&self, timeout: u64)

Schedules a reflow to be kicked off at the given timeout (in time::precise_time_ns() units). This reflow happens even if the event loop is busy. This is used to display initial page content during parsing.

fn disarm_reflow_timeout(&self)

Disables any pending reflow timeouts.

fn unregister_named_element(&self, to_unregister: &Element, id: Atom)

Remove any existing association between the provided id and any elements in this document.

fn register_named_element(&self, element: &Element, id: Atom)

Associate an element present in this document with the provided id.

fn find_fragment_node(&self, fragid: &str) -> Option<Root<Element>>

Attempt to find a named element in this page's document. https://html.spec.whatwg.org/multipage/#the-indicated-part-of-the-document

fn get_anchor_by_name(&self, name: &str) -> Option<Root<Element>>

fn set_ready_state(&self, state: DocumentReadyState)

fn is_scripting_enabled(&self) -> bool

Return whether scripting is enabled or not

fn get_focused_element(&self) -> Option<Root<Element>>

Return the element that currently has focus.

fn begin_focus_transaction(&self)

Initiate a new round of checking for elements requesting focus. The last element to call request_focus before commit_focus_transaction is called will receive focus.

fn request_focus(&self, elem: &Element)

Request that the given element receive focus once the current transaction is complete.

fn commit_focus_transaction(&self, focus_type: FocusType)

Reassign the focus context to the element that last requested focus during this transaction, or none if no elements requested it.

fn title_changed(&self)

Handles any updates when the document's title has changed.

fn send_title_to_compositor(&self)

Sends this document's title to the compositor.

fn dirty_all_nodes(&self)

fn handle_mouse_event(&self, js_runtime: *mut JSRuntime, button: MouseButton, client_point: Point2D<f32>, mouse_event_type: MouseEventType)

fn handle_touchpad_pressure_event(&self, js_runtime: *mut JSRuntime, client_point: Point2D<f32>, pressure: f32, phase_now: TouchpadPressurePhase)

fn fire_forcetouch_event(&self, event_name: String, target: &EventTarget, force: f32)

fn fire_mouse_event(&self, client_point: Point2D<f32>, target: &EventTarget, event_name: String)

fn handle_mouse_move_event(&self, js_runtime: *mut JSRuntime, client_point: Option<Point2D<f32>>, prev_mouse_over_target: &MutNullableHeap<JS<Element>>)

fn handle_touch_event(&self, js_runtime: *mut JSRuntime, event_type: TouchEventType, TouchId: TouchId, point: Point2D<f32>) -> bool

fn dispatch_key_event(&self, key: Key, state: KeyState, modifiers: KeyModifiers, compositor: &mut IpcSender<ScriptToCompositorMsg>)

The entry point for all key processing for web content

fn node_from_nodes_and_strings(&self, nodes: Vec<NodeOrString>) -> Fallible<Root<Node>>

fn get_body_attribute(&self, local_name: &Atom) -> DOMString

fn set_body_attribute(&self, local_name: &Atom, value: DOMString)

fn set_current_script(&self, script: Option<&HTMLScriptElement>)

fn get_script_blocking_stylesheets_count(&self) -> u32

fn increment_script_blocking_stylesheet_count(&self)

fn decrement_script_blocking_stylesheet_count(&self)

fn invalidate_stylesheets(&self)

fn get_and_reset_stylesheets_changed_since_reflow(&self) -> bool

fn set_pending_parsing_blocking_script(&self, script: Option<&HTMLScriptElement>)

fn get_pending_parsing_blocking_script(&self) -> Option<Root<HTMLScriptElement>>

fn add_deferred_script(&self, script: &HTMLScriptElement)

fn add_asap_script(&self, script: &HTMLScriptElement)

fn push_asap_in_order_script(&self, script: &HTMLScriptElement)

fn trigger_mozbrowser_event(&self, event: MozBrowserEvent)

fn request_animation_frame(&self, callback: Box<FnBox(f64)>) -> u32

https://html.spec.whatwg.org/multipage/#dom-window-requestanimationframe

fn cancel_animation_frame(&self, ident: u32)

https://html.spec.whatwg.org/multipage/#dom-window-cancelanimationframe

fn run_the_animation_frame_callbacks(&self)

https://html.spec.whatwg.org/multipage/#run-the-animation-frame-callbacks

fn add_blocking_load(&self, load: LoadType)

Add a load to the list of loads blocking this document's load.

fn prepare_async_load(&self, load: LoadType) -> PendingAsyncLoad

fn load_async(&self, load: LoadType, listener: AsyncResponseTarget)

fn finish_load(&self, load: LoadType)

fn maybe_execute_parser_blocking_script(&self) -> ParserBlockedByScript

If document parsing is blocked on a script, and that script is ready to run, execute it. https://html.spec.whatwg.org/multipage/#ready-to-be-parser-executed

fn process_deferred_scripts(&self)

https://html.spec.whatwg.org/multipage/#the-end step 3

fn process_asap_scripts(&self)

https://html.spec.whatwg.org/multipage/#the-end step 5 and the latter parts of https://html.spec.whatwg.org/multipage/#prepare-a-script 15.d and 15.e.

fn maybe_dispatch_dom_content_loaded(&self)

fn notify_constellation_load(&self)

fn set_current_parser(&self, script: Option<ParserRef>)

fn get_current_parser(&self) -> Option<ParserRoot>

fn find_iframe(&self, subpage_id: SubpageId) -> Option<Root<HTMLIFrameElement>>

Find an iframe element in the document.

fn find_iframe_by_pipeline(&self, pipeline: PipelineId) -> Option<Root<HTMLIFrameElement>>

Find an iframe element in the document.

fn get_dom_loading(&self) -> u64

fn get_dom_interactive(&self) -> u64

fn get_dom_content_loaded_event_start(&self) -> u64

fn get_dom_content_loaded_event_end(&self) -> u64

fn get_dom_complete(&self) -> u64

fn get_load_event_start(&self) -> u64

fn get_load_event_end(&self) -> u64

fn fire_focus_event(&self, focus_event_type: FocusEventType, node: &Node, relatedTarget: Option<&EventTarget>)

fn is_cookie_averse(&self) -> bool

https://html.spec.whatwg.org/multipage/#cookie-averse-document-object

fn nodes_from_point(&self, page_point: &Point2D<f32>) -> Vec<UntrustedNodeAddress>

impl Document

fn new_inherited(window: &Window, browsing_context: Option<&BrowsingContext>, url: Option<Url>, is_html_document: IsHTMLDocument, content_type: Option<DOMString>, last_modified: Option<String>, source: DocumentSource, doc_loader: DocumentLoader) -> Document

fn Constructor(global: GlobalRef) -> Fallible<Root<Document>>

fn new(window: &Window, browsing_context: Option<&BrowsingContext>, url: Option<Url>, doctype: IsHTMLDocument, content_type: Option<DOMString>, last_modified: Option<String>, source: DocumentSource, doc_loader: DocumentLoader) -> Root<Document>

fn create_node_list<F: Fn(&Node) -> bool>(&self, callback: F) -> Root<NodeList>

fn get_html_element(&self) -> Option<Root<HTMLHtmlElement>>

fn stylesheets(&self) -> Vec<Arc<Stylesheet>>

Returns the list of stylesheets associated with nodes in the document.

fn appropriate_template_contents_owner_document(&self) -> Root<Document>

https://html.spec.whatwg.org/multipage/#appropriate-template-contents-owner-document

fn get_element_by_id(&self, id: &Atom) -> Option<Root<Element>>

fn element_state_will_change(&self, el: &Element)

fn element_attr_will_change(&self, el: &Element)

fn get_referrer_policy(&self) -> Option<ReferrerPolicy>