Struct script::dom::element::Element

#[repr(C)]

pub struct Element {
    node: Node,
    local_name: LocalName,
    tag_name: TagName,
    namespace: Namespace,
    prefix: Option<DOMString>,
    attrs: DOMRefCell<Vec<JS<Attr>>>,
    id_attribute: DOMRefCell<Option<Atom>>,
    style_attribute: DOMRefCell<Option<Arc<RwLock<PropertyDeclarationBlock>>>>,
    attr_list: MutNullableJS<NamedNodeMap>,
    class_list: MutNullableJS<DOMTokenList>,
    state: Cell<ElementState>,
    selector_flags: Cell<ElementSelectorFlags>,
}

Fields
node: Node local_name: LocalName tag_name: TagName namespace: Namespace prefix: Option<DOMString> attrs: DOMRefCell<Vec<JS<Attr>>> id_attribute: DOMRefCell<Option<Atom>> style_attribute: DOMRefCell<Option<Arc<RwLock<PropertyDeclarationBlock>>>> attr_list: MutNullableJS<NamedNodeMap> class_list: MutNullableJS<DOMTokenList> state: Cell<ElementState> selector_flags: Cell<ElementSelectorFlags>

These flags are set by the style system to indicate the that certain operations may require restyling this element or its descendants. The flags are not atomic, so the style system takes care of only set them when it has exclusive access to the element.

Methods

impl Element

fn click_event_filter_by_disabled_state(&self) -> bool

impl Element

fn create(name: QualName,
prefix: Option<Prefix>,
document: &Document,
creator: ElementCreator)
-> Root<Element>

fn new_inherited(local_name: LocalName,
namespace: Namespace,
prefix: Option<DOMString>,
document: &Document)
-> Element

fn new_inherited_with_state(state: ElementState,
local_name: LocalName,
namespace: Namespace,
prefix: Option<DOMString>,
document: &Document)
-> Element

fn new(local_name: LocalName,
namespace: Namespace,
prefix: Option<DOMString>,
document: &Document)
-> Root<Element>

fn restyle(&self, damage: NodeDamage)

fn has_css_layout_box(&self) -> bool

fn potentially_scrollable(&self) -> bool

fn overflow_x_is_visible(&self) -> bool

fn overflow_y_is_visible(&self) -> bool

impl Element

fn html_element_in_html_document(&self) -> bool

fn local_name(&self) -> &LocalName

fn parsed_name(&self, name: DOMString) -> LocalName

fn namespace(&self) -> &Namespace

fn prefix(&self) -> Option<&DOMString>

fn attrs(&self) -> Ref<[JS<Attr>]>

fn locate_namespace(&self, prefix: Option<DOMString>) -> Namespace

fn style_attribute(&self)
-> &DOMRefCell<Option<Arc<RwLock<PropertyDeclarationBlock>>>>

fn summarize(&self) -> Vec<AttrInfo>

fn is_void(&self) -> bool

fn serialize(&self, traversal_scope: TraversalScope) -> Result<DOMString, Error>

fn root_element(&self) -> Root<Element>

fn lookup_prefix(&self, namespace: Namespace) -> Option<DOMString>

fn is_focusable_area(&self) -> bool

fn is_actually_disabled(&self) -> bool

fn push_new_attribute(&self,
local_name: LocalName,
value: AttrValue,
name: LocalName,
namespace: Namespace,
prefix: Option<Prefix>)

fn push_attribute(&self, attr: &Attr)

fn get_attribute(&self,
namespace: &Namespace,
local_name: &LocalName)
-> Option<Root<Attr>>

fn get_attribute_by_name(&self, name: DOMString) -> Option<Root<Attr>>

fn set_attribute_from_parser(&self,
qname: QualName,
value: DOMString,
prefix: Option<Prefix>)

fn set_attribute(&self, name: &LocalName, value: AttrValue)

fn set_custom_attribute(&self,
name: DOMString,
value: DOMString)
-> Result<(), Error>

fn set_first_matching_attribute<F>(&self,
local_name: LocalName,
value: AttrValue,
name: LocalName,
namespace: Namespace,
prefix: Option<Prefix>,
find: F) where F: Fn(&Attr) -> bool

fn parse_attribute(&self,
namespace: &Namespace,
local_name: &LocalName,
value: DOMString)
-> AttrValue

fn remove_attribute(&self,
namespace: &Namespace,
local_name: &LocalName)
-> Option<Root<Attr>>

fn remove_attribute_by_name(&self, name: &LocalName) -> Option<Root<Attr>>

fn remove_first_matching_attribute<F>(&self, find: F) -> Option<Root<Attr>> where F: Fn(&Attr) -> bool

fn has_class(&self, name: &Atom) -> bool

fn set_atomic_attribute(&self, local_name: &LocalName, value: DOMString)

fn has_attribute(&self, local_name: &LocalName) -> bool

fn set_bool_attribute(&self, local_name: &LocalName, value: bool)

fn get_url_attribute(&self, local_name: &LocalName) -> DOMString

fn set_url_attribute(&self, local_name: &LocalName, value: DOMString)

fn get_string_attribute(&self, local_name: &LocalName) -> DOMString

fn set_string_attribute(&self, local_name: &LocalName, value: DOMString)

fn get_tokenlist_attribute(&self, local_name: &LocalName) -> Vec<Atom>

fn set_tokenlist_attribute(&self, local_name: &LocalName, value: DOMString)

fn set_atomic_tokenlist_attribute(&self,
local_name: &LocalName,
tokens: Vec<Atom>)

fn get_int_attribute(&self, local_name: &LocalName, default: i32) -> i32

fn set_int_attribute(&self, local_name: &LocalName, value: i32)

fn get_uint_attribute(&self, local_name: &LocalName, default: u32) -> u32

fn set_uint_attribute(&self, local_name: &LocalName, value: u32)

fn will_mutate_attr(&self, attr: &Attr)

fn insert_adjacent(&self,
where_: AdjacentPosition,
node: &Node)
-> Result<Option<Root<Node>>, Error>

fn scroll(&self, x_: f64, y_: f64, behavior: ScrollBehavior)

fn parse_fragment(&self,
markup: DOMString)
-> Result<Root<DocumentFragment>, Error>

fn fragment_parsing_context(owner_doc: &Document,
element: Option<&Self>)
-> Root<Self>

fn fullscreen_element_ready_check(&self) -> bool

impl Element

fn as_maybe_activatable(&self) -> Option<&Activatable>

fn as_stylesheet_owner(&self) -> Option<&StylesheetOwner>

fn as_maybe_validatable(&self) -> Option<&Validatable>

fn click_in_progress(&self) -> bool

fn set_click_in_progress(&self, click: bool)

fn nearest_activable_element(&self) -> Option<Root<Element>>

fn is_link(&self) -> bool

fn authentic_click_activation(&self, event: &Event)

Please call this method only for real click events

https://html.spec.whatwg.org/multipage/#run-authentic-click-activation-steps

Use an element's synthetic click activation (or handle_event) for any script-triggered clicks. If the spec says otherwise, check with Manishearth first

fn state(&self) -> ElementState

fn set_state(&self, which: ElementState, value: bool)

fn active_state(&self) -> bool

fn set_active_state(&self, value: bool)

https://html.spec.whatwg.org/multipage/#concept-selector-active

fn focus_state(&self) -> bool

fn set_focus_state(&self, value: bool)

fn hover_state(&self) -> bool

fn set_hover_state(&self, value: bool)

fn enabled_state(&self) -> bool

fn set_enabled_state(&self, value: bool)

fn disabled_state(&self) -> bool

fn set_disabled_state(&self, value: bool)

fn read_write_state(&self) -> bool

fn set_read_write_state(&self, value: bool)

fn placeholder_shown_state(&self) -> bool

fn set_placeholder_shown_state(&self, value: bool)

fn target_state(&self) -> bool

fn set_target_state(&self, value: bool)

fn fullscreen_state(&self) -> bool

fn set_fullscreen_state(&self, value: bool)

fn is_connected(&self) -> bool

https://dom.spec.whatwg.org/#connected

impl Element

fn check_ancestors_disabled_state_for_form_control(&self)

fn check_parent_disabled_state_for_option(&self)

fn check_disabled_attribute(&self)