Struct script::dom::element::Element

pub struct Element {
    node: Node,
    local_name: Atom,
    namespace: Namespace,
    prefix: Option<DOMString>,
    attrs: DOMRefCell<Vec<JS<Attr>>>,
    id_attribute: DOMRefCell<Option<Atom>>,
    style_attribute: DOMRefCell<Option<PropertyDeclarationBlock>>,
    attr_list: MutNullableHeap<JS<NamedNodeMap>>,
    class_list: MutNullableHeap<JS<DOMTokenList>>,
    state: Cell<ElementState>,
    atomic_flags: AtomicElementFlags,
}

Fields

node

local_name

namespace

prefix

attrs

id_attribute

style_attribute

attr_list

class_list

state

atomic_flags

Methods

impl Element

fn click_event_filter_by_disabled_state(&self) -> bool

impl Element

fn create(name: QualName, prefix: Option<Atom>, document: &Document, creator: ElementCreator) -> Root<Element>

fn new_inherited(local_name: Atom, namespace: Namespace, prefix: Option<DOMString>, document: &Document) -> Element

fn new_inherited_with_state(state: ElementState, local_name: Atom, namespace: Namespace, prefix: Option<DOMString>, document: &Document) -> Element

fn new(local_name: Atom, namespace: Namespace, prefix: Option<DOMString>, document: &Document) -> Root<Element>

fn has_css_layout_box(&self) -> bool

fn potentially_scrollable(&self) -> bool

fn overflow_x_is_visible(&self) -> bool

fn overflow_y_is_visible(&self) -> bool

impl Element

fn html_element_in_html_document(&self) -> bool

fn local_name(&self) -> &Atom

fn parsed_name(&self, name: DOMString) -> Atom

fn namespace(&self) -> &Namespace

fn prefix(&self) -> &Option<DOMString>

fn attrs(&self) -> Ref<Vec<JS<Attr>>>

fn style_attribute(&self) -> &DOMRefCell<Option<PropertyDeclarationBlock>>

fn summarize(&self) -> Vec<AttrInfo>

fn is_void(&self) -> bool

fn sync_property_with_attrs_style(&self)

fn remove_inline_style_property(&self, property: &str)

fn update_inline_style(&self, declarations: Vec<PropertyDeclaration>, style_priority: StylePriority)

fn set_inline_style_property_priority(&self, properties: &[&str], style_priority: StylePriority)

fn get_inline_style_declaration(&self, property: &Atom) -> Option<Ref<PropertyDeclaration>>

fn get_important_inline_style_declaration(&self, property: &Atom) -> Option<Ref<PropertyDeclaration>>

fn serialize(&self, traversal_scope: TraversalScope) -> Fallible<DOMString>

fn root_element(&self) -> Root<Element>

fn lookup_prefix(&self, namespace: Namespace) -> Option<DOMString>

impl Element

fn is_focusable_area(&self) -> bool

fn is_actually_disabled(&self) -> bool

impl Element

fn push_new_attribute(&self, local_name: Atom, value: AttrValue, name: Atom, namespace: Namespace, prefix: Option<Atom>)

fn push_attribute(&self, attr: &Attr)

fn get_attribute(&self, namespace: &Namespace, local_name: &Atom) -> Option<Root<Attr>>

fn get_attribute_by_name(&self, name: DOMString) -> Option<Root<Attr>>

fn set_attribute_from_parser(&self, qname: QualName, value: DOMString, prefix: Option<Atom>)

fn set_attribute(&self, name: &Atom, value: AttrValue)

fn set_custom_attribute(&self, name: DOMString, value: DOMString) -> ErrorResult

fn set_first_matching_attribute<F>(&self, local_name: Atom, value: AttrValue, name: Atom, namespace: Namespace, prefix: Option<Atom>, find: F) where F: Fn(&Attr) -> bool

fn parse_attribute(&self, namespace: &Namespace, local_name: &Atom, value: DOMString) -> AttrValue

fn remove_attribute(&self, namespace: &Namespace, local_name: &Atom) -> Option<Root<Attr>>

fn remove_attribute_by_name(&self, name: &Atom) -> Option<Root<Attr>>

fn remove_first_matching_attribute<F>(&self, find: F) -> Option<Root<Attr>> where F: Fn(&Attr) -> bool

fn has_class(&self, name: &Atom) -> bool

fn set_atomic_attribute(&self, local_name: &Atom, value: DOMString)

fn has_attribute(&self, local_name: &Atom) -> bool

fn set_bool_attribute(&self, local_name: &Atom, value: bool)

fn get_url_attribute(&self, local_name: &Atom) -> DOMString

fn set_url_attribute(&self, local_name: &Atom, value: DOMString)

fn get_string_attribute(&self, local_name: &Atom) -> DOMString

fn set_string_attribute(&self, local_name: &Atom, value: DOMString)

fn get_tokenlist_attribute(&self, local_name: &Atom) -> Vec<Atom>

fn set_tokenlist_attribute(&self, local_name: &Atom, value: DOMString)

fn set_atomic_tokenlist_attribute(&self, local_name: &Atom, tokens: Vec<Atom>)

fn get_int_attribute(&self, local_name: &Atom, default: i32) -> i32

fn set_int_attribute(&self, local_name: &Atom, value: i32)

fn get_uint_attribute(&self, local_name: &Atom, default: u32) -> u32

fn set_uint_attribute(&self, local_name: &Atom, value: u32)

fn will_mutate_attr(&self)

fn insert_adjacent(&self, where_: DOMString, node: &Node) -> Fallible<Option<Root<Node>>>

fn scroll(&self, x_: f64, y_: f64, behavior: ScrollBehavior)

impl Element

fn as_maybe_activatable(&self) -> Option<&Activatable>

fn as_maybe_validatable(&self) -> Option<&Validatable>

fn click_in_progress(&self) -> bool

fn set_click_in_progress(&self, click: bool)

fn nearest_activable_element(&self) -> Option<Root<Element>>

fn is_link(&self) -> bool

fn authentic_click_activation(&self, event: &Event)

Please call this method only for real click events

https://html.spec.whatwg.org/multipage/#run-authentic-click-activation-steps

Use an element's synthetic click activation (or handle_event) for any script-triggered clicks. If the spec says otherwise, check with Manishearth first

fn state(&self) -> ElementState

fn set_state(&self, which: ElementState, value: bool)

fn active_state(&self) -> bool

fn set_active_state(&self, value: bool)

fn focus_state(&self) -> bool

fn set_focus_state(&self, value: bool)

fn hover_state(&self) -> bool

fn set_hover_state(&self, value: bool)

fn enabled_state(&self) -> bool

fn set_enabled_state(&self, value: bool)

fn disabled_state(&self) -> bool

fn set_disabled_state(&self, value: bool)

fn read_write_state(&self) -> bool

fn set_read_write_state(&self, value: bool)

impl Element

fn check_ancestors_disabled_state_for_form_control(&self)

fn check_parent_disabled_state_for_option(&self)

fn check_disabled_attribute(&self)