script::dom::element

Struct Element

#[repr(C)]
pub struct Element {

    node: Node,
    local_name: LocalName,
    tag_name: TagName,
    namespace: Namespace,
    prefix: DomRefCell<Option<Prefix>>,
    attrs: DomRefCell<Vec<Dom<Attr>>>,
    id_attribute: DomRefCell<Option<Atom>>,
    is: DomRefCell<Option<LocalName>>,
    style_attribute: DomRefCell<Option<Arc<Locked<PropertyDeclarationBlock>>>>,
    attr_list: MutNullableDom<NamedNodeMap>,
    class_list: MutNullableDom<DOMTokenList>,
    state: Cell<ElementState>,
    selector_flags: Cell<ElementSelectorFlags>,
    rare_data: DomRefCell<Option<Box<ElementRareData>>>,

}

https://dom.spec.whatwg.org/#element

Fields
node: Nodelocal_name: LocalNametag_name: TagNamenamespace: Namespaceprefix: DomRefCell<Option<Prefix>>attrs: DomRefCell<Vec<Dom<Attr>>>id_attribute: DomRefCell<Option<Atom>>is: DomRefCell<Option<LocalName>>

https://dom.spec.whatwg.org/#concept-element-is-value

style_attribute: DomRefCell<Option<Arc<Locked<PropertyDeclarationBlock>>>>attr_list: MutNullableDom<NamedNodeMap>class_list: MutNullableDom<DOMTokenList>state: Cell<ElementState>selector_flags: Cell<ElementSelectorFlags>

These flags are set by the style system to indicate the that certain operations may require restyling this element or its descendants. The flags are not atomic, so the style system takes care of only set them when it has exclusive access to the element.

rare_data: DomRefCell<Option<Box<ElementRareData>>>
Implementations

impl Element

fn __assert_parent_type(&self)

impl Element

fn has_selection(&self) -> bool

fn has_selectable_text(&self) -> bool

fn select_all(&self)

impl Element

pub(crate) fn create( name: QualName, is: Option<LocalName>, document: &Document, creator: ElementCreator, mode: CustomElementCreationMode, proto: Option<HandleObject<'_>>, can_gc: CanGc, ) -> DomRoot<Element>

pub(crate) fn new_inherited( local_name: LocalName, namespace: Namespace, prefix: Option<Prefix>, document: &Document, ) -> Element

pub(crate) fn new_inherited_with_state( state: ElementState, local_name: LocalName, namespace: Namespace, prefix: Option<Prefix>, document: &Document, ) -> Element

pub(crate) fn new( local_name: LocalName, namespace: Namespace, prefix: Option<Prefix>, document: &Document, proto: Option<HandleObject<'_>>, can_gc: CanGc, ) -> DomRoot<Element>

fn rare_data(&self) -> Ref<'_, Option<Box<ElementRareData>>>

fn rare_data_mut(&self) -> RefMut<'_, Option<Box<ElementRareData>>>

fn ensure_rare_data(&self) -> RefMut<'_, Box<ElementRareData>>

pub(crate) fn restyle(&self, damage: NodeDamage)

pub(crate) fn set_is(&self, is: LocalName)

pub(crate) fn get_is(&self) -> Option<LocalName>

https://dom.spec.whatwg.org/#concept-element-is-value

pub(crate) fn set_initial_custom_element_state_to_uncustomized(&self)

This is a performance optimization. Element::create can simply call element.set_custom_element_state(CustomElementState::Uncustomized) to initialize uncustomized, built-in elements with the right state, which currently just means that the DEFINED state should be true for styling. However set_custom_element_state has a high performance cost and it is unnecessary if the element is being created as an uncustomized built-in element.

See https://github.com/servo/servo/issues/37745 for more details.

pub(crate) fn set_custom_element_state(&self, state: CustomElementState)

https://dom.spec.whatwg.org/#concept-element-custom-element-state

pub(crate) fn get_custom_element_state(&self) -> CustomElementState

pub(crate) fn is_custom(&self) -> bool

https://dom.spec.whatwg.org/#concept-element-custom

pub(crate) fn set_custom_element_definition( &self, definition: Rc<CustomElementDefinition>, )

pub(crate) fn get_custom_element_definition( &self, ) -> Option<Rc<CustomElementDefinition>>

pub(crate) fn clear_custom_element_definition(&self)

pub(crate) fn push_callback_reaction( &self, function: Rc<Function<DomTypeHolder>>, args: Box<[Heap<JSVal>]>, )

pub(crate) fn push_upgrade_reaction( &self, definition: Rc<CustomElementDefinition>, )

pub(crate) fn clear_reaction_queue(&self)

pub(crate) fn invoke_reactions(&self, can_gc: CanGc)

pub(crate) fn style(&self) -> Option<Arc<ComputedValues>>

style will be None for elements in a display: none subtree. otherwise, the element has a layout box iff it doesn’t have display: none.

pub(crate) fn has_css_layout_box(&self) -> bool

pub(crate) fn is_potentially_scrollable_body(&self) -> bool

https://drafts.csswg.org/cssom-view/#potentially-scrollable

pub(crate) fn is_potentially_scrollable_body_for_scrolling_element( &self, ) -> bool

https://drafts.csswg.org/cssom-view/#potentially-scrollable

fn is_potentially_scrollable_body_shared_logic( &self, treat_overflow_clip_on_parent_as_hidden: bool, ) -> bool

https://drafts.csswg.org/cssom-view/#potentially-scrollable

fn has_scrolling_box(&self) -> bool

https://drafts.csswg.org/cssom-view/#scrolling-box

fn has_overflow(&self) -> bool

pub(crate) fn shadow_root(&self) -> Option<DomRoot<ShadowRoot>>

pub(crate) fn is_shadow_host(&self) -> bool

pub(crate) fn attach_shadow( &self, is_ua_widget: IsUserAgentWidget, mode: ShadowRootMode, clonable: bool, serializable: bool, delegates_focus: bool, slot_assignment_mode: SlotAssignmentMode, can_gc: CanGc, ) -> Fallible<DomRoot<ShadowRoot>>

https://dom.spec.whatwg.org/#dom-element-attachshadow

pub(crate) fn attach_ua_shadow_root( &self, use_ua_widget_styling: bool, can_gc: CanGc, ) -> DomRoot<ShadowRoot>

Attach a UA widget shadow root with its default parameters. Additionally mark ShadowRoot to use styling configuration for a UA widget.

The general trait of these elements is that it would hide the implementation. Thus, we would make it inaccessible (i.e., closed mode, not cloneable, and not serializable).

With UA shadow root element being assumed as one element, any focus should be delegated to its host.

pub(crate) fn is_translate_enabled(&self) -> bool

pub(crate) fn directionality(&self) -> String

pub(crate) fn is_root(&self) -> bool

pub(crate) fn registered_intersection_observers_mut( &self, ) -> RefMut<'_, Vec<IntersectionObserverRegistration>>

Return all IntersectionObserverRegistration for this element. Lazily initialize the raredata if it does not exist.

pub(crate) fn registered_intersection_observers( &self, ) -> Option<Ref<'_, Vec<IntersectionObserverRegistration>>>

pub(crate) fn get_intersection_observer_registration( &self, observer: &IntersectionObserver, ) -> Option<Ref<'_, IntersectionObserverRegistration>>

pub(crate) fn add_initial_intersection_observer_registration( &self, observer: &IntersectionObserver, )

Add a new IntersectionObserverRegistration with initial value to the element.

pub(crate) fn remove_intersection_observer( &self, observer: &IntersectionObserver, )

Removes a certain IntersectionObserver.

pub(crate) fn scrolling_box( &self, flags: ScrollContainerQueryFlags, ) -> Option<ScrollingBox>

Get the ScrollingBox that contains this element, if one does. position: fixed elements do not have a containing ScrollingBox.

pub(crate) fn scroll_into_view_with_options( &self, behavior: ScrollBehavior, block: ScrollAxisState, inline: ScrollAxisState, container: Option<&Element>, inner_target_rect: Option<Rect<Au>>, )

https://drafts.csswg.org/cssom-view/#scroll-a-target-into-view

impl Element

pub(crate) fn is_html_element(&self) -> bool

pub(crate) fn html_element_in_html_document(&self) -> bool

pub(crate) fn local_name(&self) -> &LocalName

pub(crate) fn parsed_name(&self, name: DOMString) -> LocalName

pub(crate) fn namespace(&self) -> &Namespace

pub(crate) fn prefix(&self) -> Ref<'_, Option<Prefix>>

pub(crate) fn set_prefix(&self, prefix: Option<Prefix>)

pub(crate) fn set_custom_element_registry( &self, registry: Option<DomRoot<CustomElementRegistry>>, )

pub(crate) fn custom_element_registry( &self, ) -> Option<DomRoot<CustomElementRegistry>>

pub(crate) fn attrs(&self) -> Ref<'_, [Dom<Attr>]>

pub(crate) fn locate_namespace(&self, prefix: Option<DOMString>) -> Namespace

Element branch of https://dom.spec.whatwg.org/#locate-a-namespace

pub(crate) fn name_attribute(&self) -> Option<Atom>

pub(crate) fn style_attribute( &self, ) -> &DomRefCell<Option<Arc<Locked<PropertyDeclarationBlock>>>>

pub(crate) fn summarize(&self) -> Vec<AttrInfo>

pub(crate) fn is_void(&self) -> bool

pub(crate) fn root_element(&self) -> DomRoot<Element>

pub(crate) fn lookup_prefix(&self, namespace: Namespace) -> Option<DOMString>

pub(crate) fn is_document_element(&self) -> bool

https://dom.spec.whatwg.org/#document-element

pub(crate) fn is_active_element(&self) -> bool

https://html.spec.whatwg.org/multipage/#dom-document-activeelement

pub(crate) fn is_focusable_area(&self) -> bool

pub(crate) fn find_focusable_shadow_host_if_necessary( &self, ) -> Option<DomRoot<Element>>

Returns the focusable shadow host if this is a text control inner editor. This is a workaround for the focus delegation of shadow DOM and should be used only to delegate focusable inner editor of HTMLInputElement and HTMLTextAreaElement.

pub(crate) fn is_actually_disabled(&self) -> bool

pub(crate) fn push_new_attribute( &self, local_name: LocalName, value: AttrValue, name: LocalName, namespace: Namespace, prefix: Option<Prefix>, reason: AttributeMutationReason, can_gc: CanGc, )

fn handle_attribute_changes( &self, attr: &Attr, old_value: Option<&AttrValue>, new_value: Option<DOMString>, reason: AttributeMutationReason, can_gc: CanGc, )

https://dom.spec.whatwg.org/#handle-attribute-changes

pub(crate) fn change_attribute( &self, attr: &Attr, value: AttrValue, can_gc: CanGc, )

https://dom.spec.whatwg.org/#concept-element-attributes-change

pub(crate) fn push_attribute( &self, attr: &Attr, reason: AttributeMutationReason, can_gc: CanGc, )

https://dom.spec.whatwg.org/#concept-element-attributes-append

pub(crate) fn get_attribute( &self, namespace: &Namespace, local_name: &LocalName, ) -> Option<DomRoot<Attr>>

pub(crate) fn get_attribute_by_name( &self, name: DOMString, ) -> Option<DomRoot<Attr>>

https://dom.spec.whatwg.org/#concept-element-attributes-get-by-name

pub(crate) fn set_attribute_from_parser( &self, qname: QualName, value: DOMString, prefix: Option<Prefix>, can_gc: CanGc, )

pub(crate) fn set_attribute( &self, name: &LocalName, value: AttrValue, can_gc: CanGc, )

pub(crate) fn set_custom_attribute( &self, name: DOMString, value: DOMString, can_gc: CanGc, ) -> ErrorResult

fn set_first_matching_attribute<F>( &self, local_name: LocalName, value: AttrValue, name: LocalName, namespace: Namespace, prefix: Option<Prefix>, find: F, can_gc: CanGc, )
where F: Fn(&Attr) -> bool,

https://dom.spec.whatwg.org/#concept-element-attributes-set-value

pub(crate) fn parse_attribute( &self, namespace: &Namespace, local_name: &LocalName, value: DOMString, ) -> AttrValue

pub(crate) fn remove_attribute( &self, namespace: &Namespace, local_name: &LocalName, can_gc: CanGc, ) -> Option<DomRoot<Attr>>

pub(crate) fn remove_attribute_by_name( &self, name: &LocalName, can_gc: CanGc, ) -> Option<DomRoot<Attr>>

fn remove_first_matching_attribute<F>( &self, find: F, can_gc: CanGc, ) -> Option<DomRoot<Attr>>
where F: Fn(&Attr) -> bool,

https://dom.spec.whatwg.org/#concept-element-attributes-remove

pub(crate) fn has_class( &self, name: &Atom, case_sensitivity: CaseSensitivity, ) -> bool

pub(crate) fn is_part( &self, name: &Atom, case_sensitivity: CaseSensitivity, ) -> bool

pub(crate) fn set_atomic_attribute( &self, local_name: &LocalName, value: DOMString, can_gc: CanGc, )

pub(crate) fn has_attribute(&self, local_name: &LocalName) -> bool

pub(crate) fn set_bool_attribute( &self, local_name: &LocalName, value: bool, can_gc: CanGc, )

pub(crate) fn get_url_attribute(&self, local_name: &LocalName) -> USVString

pub(crate) fn set_url_attribute( &self, local_name: &LocalName, value: USVString, can_gc: CanGc, )

pub(crate) fn get_trusted_type_url_attribute( &self, local_name: &LocalName, ) -> TrustedScriptURLOrUSVString<DomTypeHolder>

pub(crate) fn get_trusted_html_attribute( &self, local_name: &LocalName, ) -> TrustedHTMLOrString<DomTypeHolder>

pub(crate) fn get_string_attribute(&self, local_name: &LocalName) -> DOMString

pub(crate) fn set_string_attribute( &self, local_name: &LocalName, value: DOMString, can_gc: CanGc, )

fn get_nullable_string_attribute( &self, local_name: &LocalName, ) -> Option<DOMString>

Used for string attribute reflections where absence of the attribute returns null, e.g. element.ariaLabel returning null when the aria-label attribute is absent.

fn set_nullable_string_attribute( &self, local_name: &LocalName, value: Option<DOMString>, can_gc: CanGc, )

Used for string attribute reflections where setting null/undefined removes the attribute, e.g. element.ariaLabel = null removing the aria-label attribute.

pub(crate) fn get_tokenlist_attribute( &self, local_name: &LocalName, ) -> Vec<Atom>

pub(crate) fn set_tokenlist_attribute( &self, local_name: &LocalName, value: DOMString, can_gc: CanGc, )

pub(crate) fn set_atomic_tokenlist_attribute( &self, local_name: &LocalName, tokens: Vec<Atom>, can_gc: CanGc, )

pub(crate) fn get_int_attribute( &self, local_name: &LocalName, default: i32, ) -> i32

pub(crate) fn set_int_attribute( &self, local_name: &LocalName, value: i32, can_gc: CanGc, )

pub(crate) fn get_uint_attribute( &self, local_name: &LocalName, default: u32, ) -> u32

pub(crate) fn set_uint_attribute( &self, local_name: &LocalName, value: u32, can_gc: CanGc, )

pub(crate) fn will_mutate_attr(&self, attr: &Attr)

fn update_style_attribute(&self, attr: &Attr, mutation: AttributeMutation<'_>)

https://html.spec.whatwg.org/multipage/#the-style-attribute

fn set_attribute_node( &self, attr: &Attr, can_gc: CanGc, ) -> Fallible<Option<DomRoot<Attr>>>

https://dom.spec.whatwg.org/#concept-element-attributes-set including steps of https://dom.spec.whatwg.org/#concept-element-attributes-replace

pub(crate) fn update_nonce_internal_slot(&self, nonce: String)

https://html.spec.whatwg.org/multipage/#nonce-attributes

pub(crate) fn nonce_value(&self) -> String

https://html.spec.whatwg.org/multipage/#nonce-attributes

pub(crate) fn update_nonce_post_connection(&self)

https://html.spec.whatwg.org/multipage/#nonce-attributes

pub(crate) fn nonce_value_if_nonceable(&self) -> Option<String>

https://www.w3.org/TR/CSP/#is-element-nonceable

pub(crate) fn insert_adjacent( &self, where_: AdjacentPosition, node: &Node, can_gc: CanGc, ) -> Fallible<Option<DomRoot<Node>>>

pub(crate) fn scroll(&self, x: f64, y: f64, behavior: ScrollBehavior)

https://drafts.csswg.org/cssom-view/#dom-element-scroll

TODO(stevennovaryo): Need to update the scroll API to follow the spec since it is quite outdated.

pub(crate) fn parse_fragment( &self, markup: DOMString, can_gc: CanGc, ) -> Fallible<DomRoot<DocumentFragment>>

https://html.spec.whatwg.org/multipage/#fragment-parsing-algorithm-steps

pub(crate) fn fragment_parsing_context( owner_doc: &Document, element: Option<&Self>, can_gc: CanGc, ) -> DomRoot<Self>

Step 4 of https://html.spec.whatwg.org/multipage/#dom-element-insertadjacenthtml and step 6. of https://html.spec.whatwg.org/multipage/#dom-range-createcontextualfragment

pub(crate) fn fullscreen_element_ready_check(&self) -> bool

pub(crate) fn is_in_same_home_subtree<T>(&self, other: &T) -> bool
where T: DerivedFrom<Element> + DomObject,

pub(crate) fn get_id(&self) -> Option<Atom>

pub(crate) fn get_name(&self) -> Option<Atom>

fn is_sequentially_focusable(&self) -> bool

pub(crate) fn update_sequentially_focusable_status(&self, can_gc: CanGc)

pub(crate) fn get_element_internals(&self) -> Option<DomRoot<ElementInternals>>

pub(crate) fn ensure_element_internals( &self, can_gc: CanGc, ) -> DomRoot<ElementInternals>

pub(crate) fn outer_html(&self, can_gc: CanGc) -> Fallible<DOMString>

pub(crate) fn compute_source_position(&self, line_number: u32) -> SourcePosition

impl Element

fn each_custom_state<F>(&self, callback: F)
where F: FnMut(&AtomIdent),

pub(crate) fn client_rect(&self) -> Rect<i32>

pub(crate) fn as_maybe_activatable(&self) -> Option<&dyn Activatable>

pub(crate) fn as_stylesheet_owner(&self) -> Option<&dyn StylesheetOwner>

pub(crate) fn as_maybe_validatable(&self) -> Option<&dyn Validatable>

pub(crate) fn is_invalid(&self, needs_update: bool, can_gc: CanGc) -> bool

pub(crate) fn is_instance_validatable(&self) -> bool

pub(crate) fn init_state_for_internals(&self)

pub(crate) fn click_in_progress(&self) -> bool

pub(crate) fn set_click_in_progress(&self, click: bool)

pub(crate) fn nearest_activable_element(&self) -> Option<DomRoot<Element>>

pub fn state(&self) -> ElementState

pub(crate) fn set_state(&self, which: ElementState, value: bool)

pub(crate) fn set_active_state(&self, value: bool)

https://html.spec.whatwg.org/multipage/#concept-selector-active

pub(crate) fn focus_state(&self) -> bool

pub(crate) fn set_focus_state(&self, value: bool)

pub(crate) fn hover_state(&self) -> bool

pub(crate) fn set_hover_state(&self, value: bool)

pub(crate) fn enabled_state(&self) -> bool

pub(crate) fn set_enabled_state(&self, value: bool)

pub(crate) fn disabled_state(&self) -> bool

pub(crate) fn set_disabled_state(&self, value: bool)

pub(crate) fn read_write_state(&self) -> bool

pub(crate) fn set_read_write_state(&self, value: bool)

pub(crate) fn placeholder_shown_state(&self) -> bool

pub(crate) fn set_placeholder_shown_state(&self, value: bool)

pub(crate) fn set_target_state(&self, value: bool)

pub(crate) fn set_fullscreen_state(&self, value: bool)

pub(crate) fn is_connected(&self) -> bool

https://dom.spec.whatwg.org/#connected

pub(crate) fn cannot_navigate(&self) -> bool

impl Element

pub(crate) fn check_ancestors_disabled_state_for_form_control(&self)

pub(crate) fn check_parent_disabled_state_for_option(&self)

pub(crate) fn check_disabled_attribute(&self)

pub(crate) fn update_read_write_state_from_readonly_attribute(&self)