script::dom::gamepad::gamepad

Struct Gamepad

#[repr(C)]
pub(crate) struct Gamepad {

    reflector_: Reflector,
    gamepad_id: u32,
    id: String,
    index: Cell<i32>,
    connected: Cell<bool>,
    timestamp: Cell<f64>,
    mapping_type: String,
    axes: HeapBufferSource<Float64>,
    buttons: Dom<GamepadButtonList>,
    pose: Option<Dom<GamepadPose>>,
    hand: GamepadHand,
    axis_bounds: (f64, f64),
    button_bounds: (f64, f64),
    exposed: Cell<bool>,
    vibration_actuator: Dom<GamepadHapticActuator>,

}

Fields
reflector_: Reflectorgamepad_id: u32id: Stringindex: Cell<i32>connected: Cell<bool>timestamp: Cell<f64>mapping_type: Stringaxes: HeapBufferSource<Float64>buttons: Dom<GamepadButtonList>pose: Option<Dom<GamepadPose>>hand: GamepadHandaxis_bounds: (f64, f64)button_bounds: (f64, f64)exposed: Cell<bool>vibration_actuator: Dom<GamepadHapticActuator>
Implementations

impl Gamepad

fn __assert_parent_type(&self)

impl Gamepad

fn new_inherited( gamepad_id: u32, id: String, index: i32, connected: bool, timestamp: f64, mapping_type: String, buttons: &GamepadButtonList, pose: Option<&GamepadPose>, hand: GamepadHand, axis_bounds: (f64, f64), button_bounds: (f64, f64), vibration_actuator: &GamepadHapticActuator, ) -> Gamepad

pub(crate) fn new( window: &Window, gamepad_id: u32, id: String, mapping_type: String, axis_bounds: (f64, f64), button_bounds: (f64, f64), supported_haptic_effects: GamepadSupportedHapticEffects, xr: bool, can_gc: CanGc, ) -> DomRoot<Gamepad>

When we construct a new gamepad, we initialize the number of buttons and axes corresponding to the “standard” gamepad mapping. The spec says UAs may do this for fingerprint mitigation, and it also happens to simplify implementation https://www.w3.org/TR/gamepad/#fingerprinting-mitigation

impl Gamepad

pub(crate) fn gamepad_id(&self) -> u32

pub(crate) fn update_connected( &self, connected: bool, has_gesture: bool, can_gc: CanGc, )

pub(crate) fn index(&self) -> i32

pub(crate) fn update_index(&self, index: i32)

pub(crate) fn update_timestamp(&self, timestamp: f64)

pub(crate) fn notify_event(&self, event_type: GamepadEventType, can_gc: CanGc)

fn init_axes(&self, can_gc: CanGc)

Initialize the number of axes in the “standard” gamepad mapping. https://www.w3.org/TR/gamepad/#dfn-initializing-axes

pub(crate) fn map_and_normalize_axes(&self, axis_index: usize, value: f64)

https://www.w3.org/TR/gamepad/#dfn-map-and-normalize-axes

pub(crate) fn map_and_normalize_buttons(&self, button_index: usize, value: f64)

https://www.w3.org/TR/gamepad/#dfn-map-and-normalize-buttons

pub(crate) fn exposed(&self) -> bool

https://www.w3.org/TR/gamepad/#dfn-exposed

pub(crate) fn set_exposed(&self, exposed: bool)

https://www.w3.org/TR/gamepad/#dfn-exposed

pub(crate) fn vibration_actuator(&self) -> &GamepadHapticActuator