Struct script::dom::globalscope::GlobalScope

#[repr(C)]

pub struct GlobalScope {
    eventtarget: EventTarget,
    crypto: MutNullableJS<Crypto>,
    next_worker_id: Cell<WorkerId>,
    pipeline_id: PipelineId,
    devtools_wants_updates: Cell<bool>,
    console_timers: DOMRefCell<HashMap<DOMString, u64>>,
    devtools_chan: Option<IpcSender<ScriptToDevtoolsControlMsg>>,
    mem_profiler_chan: ProfilerChan,
    time_profiler_chan: ProfilerChan,
    constellation_chan: IpcSender<ConstellationMsg>,
    scheduler_chan: IpcSender<TimerEventRequest>,
    in_error_reporting_mode: Cell<bool>,
    resource_threads: ResourceThreads,
    timers: OneshotTimers,
}

Fields
eventtarget: EventTarget crypto: MutNullableJS<Crypto> next_worker_id: Cell<WorkerId> pipeline_id: PipelineId

Pipeline id associated with this global.

devtools_wants_updates: Cell<bool>

A flag to indicate whether the developer tools has requested live updates from the worker.

console_timers: DOMRefCell<HashMap<DOMString, u64>>

Timers used by the Console API.

devtools_chan: Option<IpcSender<ScriptToDevtoolsControlMsg>>

For providing instructions to an optional devtools server.

mem_profiler_chan: ProfilerChan

For sending messages to the memory profiler.

time_profiler_chan: ProfilerChan

For sending messages to the time profiler.

constellation_chan: IpcSender<ConstellationMsg>

A handle for communicating messages to the constellation thread.

scheduler_chan: IpcSender<TimerEventRequest> in_error_reporting_mode: Cell<bool>

https://html.spec.whatwg.org/multipage/#in-error-reporting-mode

resource_threads: ResourceThreads

Associated resource threads for use by DOM objects like XMLHttpRequest, including resource_thread, filemanager_thread and storage_thread

timers: OneshotTimers
Methods

impl GlobalScope

fn new_inherited(pipeline_id: PipelineId,
devtools_chan: Option<IpcSender<ScriptToDevtoolsControlMsg>>,
mem_profiler_chan: ProfilerChan,
time_profiler_chan: ProfilerChan,
constellation_chan: IpcSender<ConstellationMsg>,
scheduler_chan: IpcSender<TimerEventRequest>,
resource_threads: ResourceThreads,
timer_event_chan: IpcSender<TimerEvent>)
-> Self

fn from_reflector<T: DomObject>(reflector: &T) -> Root<Self>

Returns the global scope of the realm that the given DOM object's reflector was created in.

unsafe fn from_object(obj: *mut JSObject) -> Root<Self>

Returns the global scope of the realm that the given JS object was created in.

unsafe fn from_context(cx: *mut JSContext) -> Root<Self>

Returns the global scope for the given JSContext

unsafe fn from_object_maybe_wrapped(obj: *mut JSObject) -> Root<Self>

Returns the global object of the realm that the given JS object was created in, after unwrapping any wrappers.

fn get_cx(&self) -> *mut JSContext

fn crypto(&self) -> Root<Crypto>

fn get_next_worker_id(&self) -> WorkerId

Get next worker id.

fn live_devtools_updates(&self) -> bool

fn set_devtools_wants_updates(&self, value: bool)

fn time(&self, label: DOMString) -> Result<(), ()>

fn time_end(&self, label: &str) -> Result<u64, ()>

fn devtools_chan(&self) -> Option<&IpcSender<ScriptToDevtoolsControlMsg>>

Get an &IpcSender<ScriptToDevtoolsControlMsg> to send messages to the devtools thread when available.

fn mem_profiler_chan(&self) -> &ProfilerChan

Get a sender to the memory profiler thread.

fn time_profiler_chan(&self) -> &ProfilerChan

Get a sender to the time profiler thread.

fn constellation_chan(&self) -> &IpcSender<ConstellationMsg>

Get a sender to the constellation thread.

fn scheduler_chan(&self) -> &IpcSender<TimerEventRequest>

fn pipeline_id(&self) -> PipelineId

Get the PipelineId for this global scope.

fn api_base_url(&self) -> ServoUrl

Get the base url for this global scope.

fn get_url(&self) -> ServoUrl

Get the URL for this global scope.

fn as_window(&self) -> &Window

Extract a Window, panic if the global object is not a Window.

fn report_an_error(&self, error_info: ErrorInfo, value: HandleValue)

https://html.spec.whatwg.org/multipage/#report-the-error

fn resource_threads(&self) -> &ResourceThreads

Get the &ResourceThreads for this global scope.

fn core_resource_thread(&self) -> CoreResourceThread

Get the CoreResourceThread for this global scope.

fn script_chan(&self) -> Box<ScriptChan + Send>

ScriptChan to send messages to the event loop of this global scope.

fn networking_task_source(&self) -> NetworkingTaskSource

ScriptChan to send messages to the networking task source of this of this global scope.

fn evaluate_js_on_global_with_result(&self,
code: &str,
rval: MutableHandleValue)

Evaluate JS code on this global scope.

fn evaluate_script_on_global_with_result(&self,
code: &str,
filename: &str,
rval: MutableHandleValue,
line_number: u32)

Evaluate a JS script on this global scope.

fn schedule_callback(&self,
callback: OneshotTimerCallback,
duration: MsDuration)
-> OneshotTimerHandle

fn unschedule_callback(&self, handle: OneshotTimerHandle)

fn set_timeout_or_interval(&self,
callback: TimerCallback,
arguments: Vec<HandleValue>,
timeout: i32,
is_interval: IsInterval)
-> i32

fn clear_timeout_or_interval(&self, handle: i32)

fn fire_timer(&self, handle: TimerEventId)

fn resume(&self)

fn suspend(&self)

fn slow_down_timers(&self)

fn speed_up_timers(&self)

fn timer_source(&self) -> TimerSource

fn get_runnable_wrapper(&self) -> RunnableWrapper

Returns a wrapper for runnables to ensure they are cancelled if the global scope is being destroyed.

fn flush_promise_jobs(&self)

Start the process of executing the pending promise callbacks. They will be invoked in FIFO order, synchronously, at some point in the future.

fn enqueue_promise_job(&self, job: EnqueuedPromiseCallback)

Enqueue a promise callback for subsequent execution.

fn new_script_pair(&self) -> (Box<ScriptChan + Send>, Box<ScriptPort + Send>)

Create a new sender/receiver pair that can be used to implement an on-demand event loop. Used for implementing web APIs that require blocking semantics without resorting to nested event loops.

fn process_event(&self, msg: CommonScriptMsg)

Process a single event as if it were the next event in the thread queue for this global scope.

fn file_reading_task_source(&self) -> FileReadingTaskSource

Channel to send messages to the file reading task source of this of this global scope.

fn current() -> Root<Self>

Returns the "current" global object.

fn entry() -> Root<Self>

Returns the "entry" global object.

fn incumbent() -> Option<Root<Self>>

Returns the "incumbent" global object.