script::dom::globalscope

Struct GlobalScope

#[repr(C)]
pub(crate) struct GlobalScope {

    eventtarget: EventTarget,
    crypto: MutNullableDom<Crypto>,
    task_manager: OnceCell<TaskManager>,
    message_port_state: DomRefCell<MessagePortState>,
    broadcast_channel_state: DomRefCell<BroadcastChannelState>,
    blob_state: DomRefCell<HashMapTracedValues<BlobId, BlobInfo, FxBuildHasher>>,
    registration_map: DomRefCell<HashMapTracedValues<ServiceWorkerRegistrationId, Dom<ServiceWorkerRegistration>, FxBuildHasher>>,
    cookie_store: MutNullableDom<CookieStore>,
    worker_map: DomRefCell<HashMapTracedValues<ServiceWorkerId, Dom<ServiceWorker>, FxBuildHasher>>,
    pipeline_id: PipelineId,
    devtools_wants_updates: Cell<bool>,
    console_timers: DomRefCell<HashMap<DOMString, Instant>>,
    module_map: DomRefCell<HashMapTracedValues<ServoUrl, Rc<ModuleTree>>>,
    inline_module_map: DomRefCell<FxHashMap<ScriptId, Rc<ModuleTree>>>,
    devtools_chan: Option<GenericCallback<ScriptToDevtoolsControlMsg>>,
    mem_profiler_chan: ProfilerChan,
    time_profiler_chan: ProfilerChan,
    script_to_constellation_chan: ScriptToConstellationChan,
    script_to_embedder_chan: ScriptToEmbedderChan,
    in_error_reporting_mode: Cell<bool>,
    resource_threads: ResourceThreads,
    storage_threads: StorageThreads,
    timers: OnceCell<OneshotTimers>,
    origin: MutableOrigin,
    creation_url: DomRefCell<ServoUrl>,
    top_level_creation_url: Option<ServoUrl>,
    permission_state_invocation_results: DomRefCell<HashMap<PermissionName, PermissionState>>,
    list_auto_close_worker: DomRefCell<Vec<AutoCloseWorker>>,
    event_source_tracker: DOMTracker<EventSource>,
    abort_signal_dependents: DomRefCell<IndexSet<Dom<AbortSignal>>>,
    uncaught_rejections: DomRefCell<Vec<Box<Heap<*mut JSObject>>>>,
    consumed_rejections: DomRefCell<Vec<Box<Heap<*mut JSObject>>>>,
    gpu_id_hub: Arc<IdentityHub>,
    gpu_devices: DomRefCell<HashMapTracedValues<WebGPUDevice, WeakRef<GPUDevice>, FxBuildHasher>>,
    frozen_supported_performance_entry_types: CachedFrozenArray,
    https_state: Cell<HttpsState>,
    console_group_stack: DomRefCell<Vec<DOMString>>,
    console_count_map: DomRefCell<HashMap<DOMString, usize>>,
    dynamic_modules: DomRefCell<DynamicModuleList>,
    inherited_secure_context: Option<bool>,
    unminified_js_dir: Option<String>,
    byte_length_queuing_strategy_size_function: OnceCell<Rc<Function<DomTypeHolder>>>,
    count_queuing_strategy_size_function: OnceCell<Rc<Function<DomTypeHolder>>>,
    notification_permission_request_callback_map: DomRefCell<HashMap<String, Rc<NotificationPermissionCallback<DomTypeHolder>>>>,
    import_map: DomRefCell<ImportMap>,
    resolved_module_set: DomRefCell<HashSet<ResolvedModule>>,
    font_context: Option<Arc<FontContext>>,
    fetch_group: RefCell<FetchGroup>,

}

Fields
eventtarget: EventTargetcrypto: MutNullableDom<Crypto>task_manager: OnceCell<TaskManager>

A TaskManager for this GlobalScope.

message_port_state: DomRefCell<MessagePortState>

The message-port router id for this global, if it is managing ports.

broadcast_channel_state: DomRefCell<BroadcastChannelState>

The broadcast channels state this global, if it is managing any.

blob_state: DomRefCell<HashMapTracedValues<BlobId, BlobInfo, FxBuildHasher>>

The blobs managed by this global, if any.

registration_map: DomRefCell<HashMapTracedValues<ServiceWorkerRegistrationId, Dom<ServiceWorkerRegistration>, FxBuildHasher>>

https://w3c.github.io/ServiceWorker/#environment-settings-object-service-worker-registration-object-map

cookie_store: MutNullableDom<CookieStore>

https://cookiestore.spec.whatwg.org/#globals

worker_map: DomRefCell<HashMapTracedValues<ServiceWorkerId, Dom<ServiceWorker>, FxBuildHasher>>

https://w3c.github.io/ServiceWorker/#environment-settings-object-service-worker-object-map

pipeline_id: PipelineId

Pipeline id associated with this global.

devtools_wants_updates: Cell<bool>

A flag to indicate whether the developer tools has requested live updates from the worker.

console_timers: DomRefCell<HashMap<DOMString, Instant>>

Timers (milliseconds) used by the Console API.

module_map: DomRefCell<HashMapTracedValues<ServoUrl, Rc<ModuleTree>>>

module map is used when importing JavaScript modules https://html.spec.whatwg.org/multipage/#concept-settings-object-module-map

inline_module_map: DomRefCell<FxHashMap<ScriptId, Rc<ModuleTree>>>devtools_chan: Option<GenericCallback<ScriptToDevtoolsControlMsg>>

For providing instructions to an optional devtools server.

mem_profiler_chan: ProfilerChan

For sending messages to the memory profiler.

time_profiler_chan: ProfilerChan

For sending messages to the time profiler.

script_to_constellation_chan: ScriptToConstellationChan

A handle for communicating messages to the constellation thread.

script_to_embedder_chan: ScriptToEmbedderChan

A handle for communicating messages to the Embedder.

in_error_reporting_mode: Cell<bool>

https://html.spec.whatwg.org/multipage/#in-error-reporting-mode

resource_threads: ResourceThreads

Associated resource threads for use by DOM objects like XMLHttpRequest, including resource_thread and filemanager_thread

storage_threads: StorageThreads

Associated resource threads for use by DOM objects like XMLHttpRequest, including indexeddb thread and storage_thread

timers: OnceCell<OneshotTimers>

The mechanism by which time-outs and intervals are scheduled. https://html.spec.whatwg.org/multipage/#timers

origin: MutableOrigin

The origin of the globalscope

creation_url: DomRefCell<ServoUrl>

https://html.spec.whatwg.org/multipage/#concept-environment-creation-url

top_level_creation_url: Option<ServoUrl>

https://html.spec.whatwg.org/multipage/#concept-environment-top-level-creation-url

permission_state_invocation_results: DomRefCell<HashMap<PermissionName, PermissionState>>

A map for storing the previous permission state read results.

list_auto_close_worker: DomRefCell<Vec<AutoCloseWorker>>

Vector storing closing references of all workers

event_source_tracker: DOMTracker<EventSource>

Vector storing references of all eventsources.

abort_signal_dependents: DomRefCell<IndexSet<Dom<AbortSignal>>>

Dependent AbortSignals that must be kept alive per <https://dom.spec.whatwg.org/#abort-signal-garbage-collection?

uncaught_rejections: DomRefCell<Vec<Box<Heap<*mut JSObject>>>>

Storage for watching rejected promises waiting for some client to consume their rejection. Promises in this list have been rejected in the last turn of the event loop without the rejection being handled. Note that this can contain nullptrs in place of promises removed because theyre consumed before itd be reported.

https://html.spec.whatwg.org/multipage/#about-to-be-notified-rejected-promises-list

consumed_rejections: DomRefCell<Vec<Box<Heap<*mut JSObject>>>>

Promises in this list have previously been reported as rejected (because they were in the above list), but the rejection was handled in the last turn of the event loop.

https://html.spec.whatwg.org/multipage/#outstanding-rejected-promises-weak-set

gpu_id_hub: Arc<IdentityHub>

Identity Manager for WebGPU resources

gpu_devices: DomRefCell<HashMapTracedValues<WebGPUDevice, WeakRef<GPUDevice>, FxBuildHasher>>

WebGPU devices

frozen_supported_performance_entry_types: CachedFrozenArrayhttps_state: Cell<HttpsState>

currect https state (from previous request)

console_group_stack: DomRefCell<Vec<DOMString>>

The stack of active group labels for the Console APIs.

console_count_map: DomRefCell<HashMap<DOMString, usize>>

The count map for the Console APIs.

https://console.spec.whatwg.org/#count

dynamic_modules: DomRefCell<DynamicModuleList>

List of ongoing dynamic module imports.

inherited_secure_context: Option<bool>

Is considered in a secure context

unminified_js_dir: Option<String>

Directory to store unminified scripts for this window if unminify-js opt is enabled.

byte_length_queuing_strategy_size_function: OnceCell<Rc<Function<DomTypeHolder>>>

The byte length queuing strategy size function that will be initialized once size getter of ByteLengthQueuingStrategy is called.

https://streams.spec.whatwg.org/#byte-length-queuing-strategy-size-function

count_queuing_strategy_size_function: OnceCell<Rc<Function<DomTypeHolder>>>

The count queuing strategy size function that will be initialized once size getter of CountQueuingStrategy is called.

https://streams.spec.whatwg.org/#count-queuing-strategy-size-function

notification_permission_request_callback_map: DomRefCell<HashMap<String, Rc<NotificationPermissionCallback<DomTypeHolder>>>>import_map: DomRefCell<ImportMap>

An import map allows control over module specifier resolution. For now, only Window global objects have their import map modified from the initial empty one.

https://html.spec.whatwg.org/multipage/#import-maps

resolved_module_set: DomRefCell<HashSet<ResolvedModule>>

https://html.spec.whatwg.org/multipage/#resolved-module-set

font_context: Option<Arc<FontContext>>

The FontContext for this GlobalScope if it has one. This is used for canvas and layout, so if this GlobalScope doesnt need to use either, this might be None.

fetch_group: RefCell<FetchGroup>

https://fetch.spec.whatwg.org/#environment-settings-object-fetch-group

Implementations

impl GlobalScope

pub(crate) fn webview_id(&self) -> Option<WebViewId>

A sender to the event loop of this global scope. This either sends to the Worker event loop or the ScriptThread event loop in the case of a Window. This can be None for dedicated workers that are not currently handling a message.

pub(crate) fn new_inherited( pipeline_id: PipelineId, devtools_chan: Option<GenericCallback<ScriptToDevtoolsControlMsg>>, mem_profiler_chan: ProfilerChan, time_profiler_chan: ProfilerChan, script_to_constellation_chan: ScriptToConstellationChan, script_to_embedder_chan: ScriptToEmbedderChan, resource_threads: ResourceThreads, storage_threads: StorageThreads, origin: MutableOrigin, creation_url: ServoUrl, top_level_creation_url: Option<ServoUrl>, gpu_id_hub: Arc<IdentityHub>, inherited_secure_context: Option<bool>, unminify_js: bool, font_context: Option<Arc<FontContext>>, ) -> Self

fn port_router_id(&self) -> Option<MessagePortRouterId>

The message-port router Id of the global, if any

fn is_managing_port(&self, port_id: &MessagePortId) -> bool

Is this global managing a given port?

fn timers(&self) -> &OneshotTimers

pub(crate) fn font_context(&self) -> Option<&Arc<FontContext>>

pub(crate) fn get_serviceworker_registration( &self, script_url: &ServoUrl, scope: &ServoUrl, registration_id: ServiceWorkerRegistrationId, installing_worker: Option<ServiceWorkerId>, _waiting_worker: Option<ServiceWorkerId>, _active_worker: Option<ServiceWorkerId>, can_gc: CanGc, ) -> DomRoot<ServiceWorkerRegistration>

https://w3c.github.io/ServiceWorker/#get-the-service-worker-registration-object

pub(crate) fn get_serviceworker( &self, script_url: &ServoUrl, scope: &ServoUrl, worker_id: ServiceWorkerId, can_gc: CanGc, ) -> DomRoot<ServiceWorker>

https://w3c.github.io/ServiceWorker/#get-the-service-worker-object

fn complete_port_transfer( &self, port_id: MessagePortId, tasks: VecDeque<PortMessageTask>, disentangled: bool, can_gc: CanGc, )

Complete the transfer of a message-port.

fn try_complete_disentanglement(&self, port_id: MessagePortId, can_gc: CanGc)

The closing of otherPort, if it is in a different global. https://html.spec.whatwg.org/multipage/#disentangle

pub(crate) fn perform_a_dom_garbage_collection_checkpoint(&self)

Clean-up DOM related resources

pub(crate) fn remove_web_messaging_and_dedicated_workers_infra(&self)

Remove the routers for ports and broadcast-channels. Drain the list of workers.

fn remove_message_ports_router(&self)

Update our state to un-managed, and tell the constellation to drop the sender to our message-port router.

fn remove_broadcast_channel_router(&self)

Update our state to un-managed, and tell the constellation to drop the sender to our broadcast router.

pub(crate) fn disentangle_port(&self, port: &MessagePort, can_gc: CanGc)

https://html.spec.whatwg.org/multipage/#disentangle

pub(crate) fn entangle_ports(&self, port1: MessagePortId, port2: MessagePortId)

https://html.spec.whatwg.org/multipage/#entangle

pub(crate) fn mark_port_as_transferred( &self, port_id: &MessagePortId, ) -> MessagePortImpl

Handle the transfer of a port in the current task.

pub(crate) fn start_message_port(&self, port_id: &MessagePortId, can_gc: CanGc)

https://html.spec.whatwg.org/multipage/#dom-messageport-start

pub(crate) fn close_message_port(&self, port_id: &MessagePortId)

https://html.spec.whatwg.org/multipage/#dom-messageport-close

pub(crate) fn post_messageport_msg( &self, port_id: MessagePortId, task: PortMessageTask, )

https://html.spec.whatwg.org/multipage/#message-port-post-message-steps

fn re_route_port_task(&self, port_id: MessagePortId, task: PortMessageTask)

If we dont know about the port, send the message to the constellation for routing.

pub(crate) fn schedule_broadcast( &self, msg: BroadcastChannelMsg, channel_id: &Uuid, )

https://html.spec.whatwg.org/multipage/#dom-broadcastchannel-postmessage Step 7 and following steps.

pub(crate) fn broadcast_message_event( &self, event: BroadcastChannelMsg, channel_id: Option<&Uuid>, )

https://html.spec.whatwg.org/multipage/#dom-broadcastchannel-postmessage Step 7 and following steps.

pub(crate) fn note_cross_realm_transform_readable( &self, cross_realm_transform_readable: &CrossRealmTransformReadable, port_id: &MessagePortId, )

https://streams.spec.whatwg.org/#abstract-opdef-setupcrossrealmtransformreadable The Add a handler for ports message event with the following steps: and Add a handler for ports messageerror event with the following steps: part.

pub(crate) fn note_cross_realm_transform_writable( &self, cross_realm_transform_writable: &CrossRealmTransformWritable, port_id: &MessagePortId, )

https://streams.spec.whatwg.org/#abstract-opdef-setupcrossrealmtransformwritable The Add a handler for ports message event with the following steps: and Add a handler for ports messageerror event with the following steps: part.

pub(crate) fn route_task_to_port( &self, port_id: MessagePortId, task: PortMessageTask, can_gc: CanGc, )

Custom routing logic, followed by the task steps of https://html.spec.whatwg.org/multipage/#message-port-post-message-steps

pub(crate) fn maybe_add_pending_ports(&self)

Check all ports that have been transfer-received in the previous task, and complete their transfer if they havent been re-transferred.

pub(crate) fn perform_a_message_port_garbage_collection_checkpoint(&self)

https://html.spec.whatwg.org/multipage/#ports-and-garbage-collection

pub(crate) fn perform_a_broadcast_channel_garbage_collection_checkpoint(&self)

Remove broadcast-channels that are closed. TODO: Also remove them if they do not have an event-listener. see https://github.com/servo/servo/issues/25772

pub(crate) fn register_dependent_abort_signal(&self, signal: &AbortSignal)

Register a dependent AbortSignal that may need to be kept alive https://dom.spec.whatwg.org/#abort-signal-garbage-collection

pub(crate) fn perform_an_abort_signal_garbage_collection_checkpoint(&self)

Clean up dependent AbortSignals that no longer satisfy the GC predicate.

pub(crate) fn track_broadcast_channel(&self, dom_channel: &BroadcastChannel)

Start tracking a broadcast-channel.

pub(crate) fn track_message_port( &self, dom_port: &MessagePort, port_impl: Option<MessagePortImpl>, )

Start tracking a message-port

pub(crate) fn serialize_blob(&self, blob_id: &BlobId) -> BlobImpl

https://html.spec.whatwg.org/multipage/#serialization-steps defined at https://w3c.github.io/FileAPI/#blob-section. Get the snapshot state and underlying bytes of the blob.

fn track_blob_info(&self, blob_info: BlobInfo, blob_id: BlobId)

pub(crate) fn track_blob(&self, dom_blob: &Blob, blob_impl: BlobImpl)

Start tracking a blob

pub(crate) fn track_file(&self, file: &File, blob_impl: BlobImpl)

Start tracking a file

fn perform_a_blob_garbage_collection_checkpoint(&self)

Clean-up any file or blob that is unreachable from script, unless it has an oustanding blob url. https://w3c.github.io/FileAPI/#lifeTime

pub(crate) fn clean_up_all_file_resources(&self)

Clean-up all file related resources on document unload. https://w3c.github.io/FileAPI/#lifeTime

fn decrement_file_ref(&self, id: Uuid)

pub(crate) fn get_blob_bytes(&self, blob_id: &BlobId) -> Result<Vec<u8>, ()>

Get a slice to the inner data of a Blob, In the case of a File-backed blob, this might incur synchronous read and caching.

pub(crate) fn get_blob_data<'a>(&'a self, blob_id: &BlobId) -> Ref<'a, BlobData>

Retrieve information about a specific blob from the blob store

Panics

This function panics if there is no blob with the given ID.

fn get_blob_bytes_non_sliced(&self, blob_id: &BlobId) -> Result<Vec<u8>, ()>

Get bytes from a non-sliced blob

fn get_blob_bytes_or_file_id(&self, blob_id: &BlobId) -> BlobResult

Get a slice to the inner data of a Blob, if its a memory blob, or its file-id and file-size otherwise.

Note: this is almost a duplicate of get_blob_bytes, tweaked for integration with streams. TODO: merge with get_blob_bytes by way of broader integration with blob streams.

fn get_blob_bytes_non_sliced_or_file_id(&self, blob_id: &BlobId) -> BlobResult

Get bytes from a non-sliced blob if in memory, or its file-id and file-size.

Note: this is almost a duplicate of get_blob_bytes_non_sliced, tweaked for integration with streams. TODO: merge with get_blob_bytes by way of broader integration with blob streams.

pub(crate) fn get_blob_type_string(&self, blob_id: &BlobId) -> String

Get a copy of the type_string of a blob.

pub(crate) fn get_blob_size(&self, blob_id: &BlobId) -> u64

https://w3c.github.io/FileAPI/#dfn-size

pub(crate) fn get_blob_url_id(&self, blob_id: &BlobId) -> Uuid

fn create_sliced_url_id( &self, blob_info: &mut BlobInfo, parent_file_id: &Uuid, rel_pos: &RelativePos, parent_len: u64, ) -> Uuid

Get a FileID representing sliced parent-blob content

pub(crate) fn promote(&self, blob_info: &mut BlobInfo, set_valid: bool) -> Uuid

Promote non-Slice blob:

Memory-based: The bytes in data slice will be transferred to file manager thread.

File-based: If set_valid, then activate the FileID so it can serve as URL Depending on set_valid, the returned FileID can be part of valid or invalid Blob URL.

fn send_to_file_manager(&self, msg: FileManagerThreadMsg)

fn read_file(&self, id: Uuid) -> Result<Vec<u8>, ()>

pub(crate) fn get_blob_stream( &self, blob_id: &BlobId, can_gc: CanGc, ) -> Fallible<DomRoot<ReadableStream>>

https://w3c.github.io/FileAPI/#blob-get-stream

pub(crate) fn read_file_async( &self, id: Uuid, promise: Rc<Promise>, callback: Box<dyn Fn(Rc<Promise>, Fallible<Vec<u8>>) + Send>, )

fn send_msg(&self, id: Uuid) -> IpcReceiver<FileManagerResult<ReadFileProgress>>

fn read_msg( receiver: IpcReceiver<FileManagerResult<ReadFileProgress>>, ) -> Result<Vec<u8>, ()>

pub(crate) fn permission_state_invocation_results( &self, ) -> &DomRefCell<HashMap<PermissionName, PermissionState>>

pub(crate) fn track_worker( &self, closing: Arc<AtomicBool>, join_handle: JoinHandle<()>, control_sender: Sender<DedicatedWorkerControlMsg>, context: ThreadSafeJSContext, )

pub(crate) fn track_event_source(&self, event_source: &EventSource)

pub(crate) fn close_event_sources(&self) -> bool

pub(crate) fn from_reflector<T: DomObject>( reflector: &T, _realm: InRealm<'_>, ) -> DomRoot<Self>

Returns the global scope of the realm that the given DOM objects reflector was created in.

pub(crate) unsafe fn from_object(obj: *mut JSObject) -> DomRoot<Self>

Returns the global scope of the realm that the given JS object was created in.

pub(crate) unsafe fn from_context( cx: *mut JSContext, _realm: InRealm<'_>, ) -> DomRoot<Self>

Returns the global scope for the given JSContext

pub(crate) fn from_current_realm(realm: &CurrentRealm<'_>) -> DomRoot<Self>

Return global scope asociated with current realm

Eventually we could return Handle here as global is already rooted by realm.

pub(crate) fn from_safe_context( cx: SafeJSContext, realm: InRealm<'_>, ) -> DomRoot<Self>

Returns the global scope for the given SafeJSContext

pub(crate) unsafe fn from_object_maybe_wrapped( obj: *mut JSObject, cx: *mut JSContext, ) -> DomRoot<Self>

Returns the global object of the realm that the given JS object was created in, after unwrapping any wrappers.

pub(crate) fn add_uncaught_rejection(&self, rejection: HandleObject)

pub(crate) fn remove_uncaught_rejection(&self, rejection: HandleObject)

pub(crate) fn get_uncaught_rejections( &self, ) -> &DomRefCell<Vec<Box<Heap<*mut JSObject>>>>

pub(crate) fn add_consumed_rejection(&self, rejection: HandleObject)

pub(crate) fn remove_consumed_rejection(&self, rejection: HandleObject)

pub(crate) fn get_consumed_rejections( &self, ) -> &DomRefCell<Vec<Box<Heap<*mut JSObject>>>>

pub(crate) fn set_module_map(&self, url: ServoUrl, module: ModuleTree)

pub(crate) fn get_module_map( &self, ) -> &DomRefCell<HashMapTracedValues<ServoUrl, Rc<ModuleTree>>>

pub(crate) fn set_inline_module_map( &self, script_id: ScriptId, module: ModuleTree, )

pub(crate) fn get_inline_module_map( &self, ) -> &DomRefCell<FxHashMap<ScriptId, Rc<ModuleTree>>>

pub(crate) fn get_cx() -> SafeJSContext

pub(crate) fn crypto(&self, can_gc: CanGc) -> DomRoot<Crypto>

pub(crate) fn cookie_store(&self, can_gc: CanGc) -> DomRoot<CookieStore>

pub(crate) fn live_devtools_updates(&self) -> bool

pub(crate) fn set_devtools_wants_updates(&self, value: bool)

pub(crate) fn time(&self, label: DOMString) -> Result<(), ()>

pub(crate) fn time_log(&self, label: &DOMString) -> Result<u64, ()>

Computes the delta time since a label has been created

Returns an error if the label does not exist.

pub(crate) fn time_end(&self, label: &DOMString) -> Result<u64, ()>

Computes the delta time since a label has been created and stops tracking the label.

Returns an error if the label does not exist.

pub(crate) fn devtools_chan( &self, ) -> Option<&GenericCallback<ScriptToDevtoolsControlMsg>>

Get an &IpcSender<ScriptToDevtoolsControlMsg> to send messages to the devtools thread when available.

pub(crate) fn issue_page_warning(&self, warning: &str)

pub(crate) fn mem_profiler_chan(&self) -> &ProfilerChan

Get a sender to the memory profiler thread.

pub(crate) fn time_profiler_chan(&self) -> &ProfilerChan

Get a sender to the time profiler thread.

pub(crate) fn script_to_constellation_chan(&self) -> &ScriptToConstellationChan

Get a sender to the constellation thread.

pub(crate) fn script_to_embedder_chan(&self) -> &ScriptToEmbedderChan

pub(crate) fn send_to_embedder(&self, msg: EmbedderMsg)

pub(crate) fn send_to_constellation(&self, msg: ScriptToConstellationMessage)

pub(crate) fn pipeline_id(&self) -> PipelineId

Get the PipelineId for this global scope.

pub(crate) fn origin(&self) -> &MutableOrigin

Get the origin for this global scope

pub(crate) fn creation_url(&self) -> ServoUrl

Get the creation_url for this global scope

pub(crate) fn set_creation_url(&self, creation_url: ServoUrl)

pub(crate) fn top_level_creation_url(&self) -> &Option<ServoUrl>

Get the top_level_creation_url for this global scope

pub(crate) fn image_cache(&self) -> Arc<dyn ImageCache>

pub(crate) fn schedule_timer( &self, request: TimerEventRequest, ) -> Option<TimerId>

Schedule a TimerEventRequest on this GlobalScopes timers::TimerScheduler. Every Worker has its own scheduler, which handles events in the Worker event loop, but Windows use a shared scheduler associated with their ScriptThread.

pub(crate) fn is_nested_browsing_context(&self) -> bool

https://html.spec.whatwg.org/multipage/#nested-browsing-context

pub(crate) fn total_size_of_in_flight_keep_alive_records(&self) -> u64

Obtain the size of in flight keep alive records from the resource thread. If we cant communicate with the thread, we return u64::MAX to ensure the limit is higher than what is allowed. This ensures that whenever we want to initiate a keep alive request and the thread doesnt communicate, we block additional keep alive requests.

pub(crate) fn request_client(&self) -> RequestClient

Part of https://fetch.spec.whatwg.org/#populate-request-from-client

pub(crate) fn policy_container(&self) -> PolicyContainer

https://html.spec.whatwg.org/multipage/#concept-settings-object-policy-container

pub(crate) fn api_base_url(&self) -> ServoUrl

Get the base url for this global scope.

pub(crate) fn get_url(&self) -> ServoUrl

Get the URL for this global scope.

pub(crate) fn get_referrer_policy(&self) -> ReferrerPolicy

Get the Referrer Policy for this global scope.

pub(crate) fn get_referrer(&self) -> Referrer

Step 3.client of https://w3c.github.io/webappsec-referrer-policy/#determine-requests-referrer Determine the Referrer for a request whose Referrer is client

pub(crate) fn as_window(&self) -> &Window

Extract a Window, panic if the global object is not a Window.

pub(crate) fn insecure_requests_policy(&self) -> InsecureRequestsPolicy

Returns a policy that should be used for fetches initiated from this global.

pub(crate) fn has_trustworthy_ancestor_origin(&self) -> bool

Whether this document has ancestor navigables that are trustworthy

pub(crate) fn has_trustworthy_ancestor_or_current_origin(&self) -> bool

pub(crate) fn report_an_exception( &self, cx: SafeJSContext, error: HandleValue<'_>, can_gc: CanGc, )

https://html.spec.whatwg.org/multipage/#report-an-exception

pub(crate) fn report_an_error( &self, error_info: ErrorInfo, value: HandleValue<'_>, can_gc: CanGc, )

Steps 6-7 of https://html.spec.whatwg.org/multipage/#report-an-exception

pub(crate) fn resource_threads(&self) -> &ResourceThreads

Get the &ResourceThreads for this global scope.

pub(crate) fn core_resource_thread(&self) -> CoreResourceThread

Get the CoreResourceThread for this global scope.

pub(crate) fn storage_threads(&self) -> &StorageThreads

Get a reference to the StorageThreads for this GlobalScope.

pub(crate) fn event_loop_sender(&self) -> Option<ScriptEventLoopSender>

A sender to the event loop of this global scope. This either sends to the Worker event loop or the ScriptThread event loop in the case of a Window. This can be None for dedicated workers that are not currently handling a message.

pub(crate) fn task_manager(&self) -> &TaskManager

A reference to the TaskManager used to schedule tasks for this GlobalScope.

pub(crate) fn evaluate_js_on_global( &self, code: Cow<'_, str>, filename: &str, introduction_type: Option<&'static CStr>, rval: MutableHandleValue<'_>, can_gc: CanGc, ) -> Result<(), JavaScriptEvaluationError>

Evaluate JS code on this global scope.

pub(crate) fn schedule_callback( &self, callback: OneshotTimerCallback, duration: Duration, ) -> OneshotTimerHandle

https://html.spec.whatwg.org/multipage/#timer-initialisation-steps

pub(crate) fn unschedule_callback(&self, handle: OneshotTimerHandle)

pub(crate) fn set_timeout_or_interval( &self, callback: TimerCallback, arguments: Vec<HandleValue<'_>>, timeout: Duration, is_interval: IsInterval, can_gc: CanGc, ) -> Fallible<i32>

https://html.spec.whatwg.org/multipage/#timer-initialisation-steps

pub(crate) fn clear_timeout_or_interval(&self, handle: i32)

pub(crate) fn fire_timer(&self, handle: TimerEventId, can_gc: CanGc)

pub(crate) fn resume(&self)

pub(crate) fn suspend(&self)

pub(crate) fn slow_down_timers(&self)

pub(crate) fn speed_up_timers(&self)

fn timer_source(&self) -> TimerSource

pub(crate) fn can_continue_running(&self) -> bool

Returns a boolean indicating whether the event-loop where this global is running on can continue running JS.

pub(crate) fn get_indexeddb(&self) -> DomRoot<IDBFactory>

Returns the idb factory for this global. TODO: move the idb to the global itself.

pub(crate) fn perform_a_microtask_checkpoint(&self, can_gc: CanGc)

Perform a microtask checkpoint.

pub(crate) fn enqueue_microtask(&self, job: Microtask)

Enqueue a microtask for subsequent execution.

pub(crate) fn new_script_pair( &self, ) -> (ScriptEventLoopSender, ScriptEventLoopReceiver)

Create a new sender/receiver pair that can be used to implement an on-demand event loop. Used for implementing web APIs that require blocking semantics without resorting to nested event loops.

pub(crate) fn process_event( &self, msg: CommonScriptMsg, cx: &mut JSContext, ) -> bool

Process a single event as if it were the next event in the queue for the event-loop where this global scope is running on. Returns a boolean indicating whether further events should be processed.

pub(crate) fn runtime_handle(&self) -> ParentRuntime

pub(crate) fn current() -> Option<DomRoot<Self>>

Returns the current global object.

pub(crate) fn entry() -> DomRoot<Self>

Returns the entry global object.

pub(crate) fn incumbent() -> Option<DomRoot<Self>>

Returns the incumbent global object.

pub(crate) fn performance(&self) -> DomRoot<Performance>

pub(crate) fn supported_performance_entry_types( &self, cx: SafeJSContext, retval: MutableHandleValue<'_>, can_gc: CanGc, )

https://w3c.github.io/performance-timeline/#supportedentrytypes-attribute

pub(crate) fn get_https_state(&self) -> HttpsState

pub(crate) fn set_https_state(&self, https_state: HttpsState)

pub(crate) fn inherited_secure_context(&self) -> Option<bool>

pub(crate) fn is_secure_context(&self) -> bool

https://html.spec.whatwg.org/multipage/#secure-context

pub(crate) fn get_csp_list(&self) -> Option<CspList>

https://www.w3.org/TR/CSP/#get-csp-of-object

pub(crate) fn status_code(&self) -> Option<u16>

pub(crate) fn wgpu_id_hub(&self) -> Arc<IdentityHub>

pub(crate) fn add_gpu_device(&self, device: &GPUDevice)

pub(crate) fn remove_gpu_device(&self, device: WebGPUDevice)

pub(crate) fn gpu_device_lost( &self, device: WebGPUDevice, reason: DeviceLostReason, msg: String, )

pub(crate) fn handle_uncaptured_gpu_error( &self, device: WebGPUDevice, error: Error, )

pub(crate) fn current_group_label(&self) -> Option<DOMString>

pub(crate) fn push_console_group(&self, group: DOMString)

pub(crate) fn pop_console_group(&self)

pub(crate) fn increment_console_count(&self, label: &DOMString) -> usize

pub(crate) fn reset_console_count(&self, label: &DOMString) -> Result<(), ()>

pub(crate) fn dynamic_module_list(&self) -> RefMut<'_, DynamicModuleList>

pub(crate) fn structured_clone( &self, cx: SafeJSContext, value: HandleValue<'_>, options: RootedTraceableBox<StructuredSerializeOptions>, retval: MutableHandleValue<'_>, can_gc: CanGc, ) -> Fallible<()>

pub(crate) fn fetch<Listener: FetchResponseListener>( &self, request_builder: RequestBuilder, context: Listener, task_source: SendableTaskSource, )

pub(crate) fn fetch_with_network_listener<Listener: FetchResponseListener>( &self, request_builder: RequestBuilder, network_listener: NetworkListener<Listener>, )

pub(crate) fn unminify_js(&self) -> bool

pub(crate) fn unminified_js_dir(&self) -> Option<String>

pub(crate) fn set_byte_length_queuing_strategy_size( &self, function: Rc<Function<DomTypeHolder>>, )

pub(crate) fn get_byte_length_queuing_strategy_size( &self, ) -> Option<Rc<Function<DomTypeHolder>>>

pub(crate) fn set_count_queuing_strategy_size( &self, function: Rc<Function<DomTypeHolder>>, )

pub(crate) fn get_count_queuing_strategy_size( &self, ) -> Option<Rc<Function<DomTypeHolder>>>

pub(crate) fn add_notification_permission_request_callback( &self, callback_id: String, callback: Rc<NotificationPermissionCallback<DomTypeHolder>>, )

pub(crate) fn remove_notification_permission_request_callback( &self, callback_id: String, ) -> Option<Rc<NotificationPermissionCallback<DomTypeHolder>>>

pub(crate) fn trusted_types( &self, can_gc: CanGc, ) -> DomRoot<TrustedTypePolicyFactory>

pub(crate) fn append_reporting_observer( &self, reporting_observer: &ReportingObserver, )

pub(crate) fn remove_reporting_observer( &self, reporting_observer: &ReportingObserver, )

pub(crate) fn registered_reporting_observers( &self, ) -> Vec<DomRoot<ReportingObserver>>

pub(crate) fn append_report(&self, report: Report)

pub(crate) fn buffered_reports(&self) -> Vec<Report>

pub(crate) fn append_deferred_fetch( &self, deferred_fetch: Rc<DeferredFetchRecord>, ) -> DeferredFetchRecordId

pub(crate) fn deferred_fetches(&self) -> Vec<Rc<DeferredFetchRecord>>

pub(crate) fn deferred_fetch_record_for_id( &self, deferred_fetch_record_id: &DeferredFetchRecordId, ) -> Rc<DeferredFetchRecord>

pub(crate) fn process_deferred_fetches(&self)

https://fetch.spec.whatwg.org/#process-deferred-fetches

pub(crate) fn import_map(&self) -> Ref<'_, ImportMap>

pub(crate) fn import_map_mut(&self) -> RefMut<'_, ImportMap>

pub(crate) fn resolved_module_set(&self) -> Ref<'_, HashSet<ResolvedModule>>

pub(crate) fn resolved_module_set_mut( &self, ) -> RefMut<'_, HashSet<ResolvedModule>>

pub(crate) fn add_module_to_resolved_module_set( &self, base_url: &str, specifier: &str, specifier_url: Option<ServoUrl>, )

https://html.spec.whatwg.org/multipage/#add-module-to-resolved-module-set

pub(crate) fn create_a_classic_script( &self, source: Cow<'_, str>, url: ServoUrl, fetch_options: ScriptFetchOptions, muted_errors: ErrorReporting, introduction_type: Option<&'static CStr>, line_number: u32, external: bool, ) -> ClassicScript

https://html.spec.whatwg.org/multipage/#creating-a-classic-script

pub(crate) fn run_a_classic_script( &self, script: ClassicScript, rethrow_errors: RethrowErrors, can_gc: CanGc, ) -> ErrorResult

https://html.spec.whatwg.org/multipage/#run-a-classic-script

fn can_run_script(&self) -> bool

https://html.spec.whatwg.org/multipage/#check-if-we-can-run-script

pub(crate) fn run_steps_after_a_timeout<F>( &self, ordering_identifier: DOMString, milliseconds: i64, completion_steps: F, ) -> i32
where F: 'static + FnOnce(&GlobalScope, CanGc),

https://html.spec.whatwg.org/multipage/#run-steps-after-a-timeout TODO: This should end-up being used in the other timer mechanism integrate as per <https://html.spec.whatwg.org/multipage/#timers:run-steps-after-a-timeout?