script::dom::html::htmlelement

Struct HTMLElement

#[repr(C)]
pub(crate) struct HTMLElement {
    element: Element,
    style_decl: MutNullableDom<CSSStyleDeclaration>,
    dataset: MutNullableDom<DOMStringMap>,
}

Fields
element: Elementstyle_decl: MutNullableDom<CSSStyleDeclaration>dataset: MutNullableDom<DOMStringMap>
Implementations

impl HTMLElement

fn __assert_parent_type(&self)

impl HTMLElement

pub(crate) fn new_inherited( tag_name: LocalName, prefix: Option<Prefix>, document: &Document, ) -> HTMLElement

pub(crate) fn new_inherited_with_state( state: ElementState, tag_name: LocalName, prefix: Option<Prefix>, document: &Document, ) -> HTMLElement

pub(crate) fn new( local_name: LocalName, prefix: Option<Prefix>, document: &Document, proto: Option<HandleObject<'_>>, can_gc: CanGc, ) -> DomRoot<HTMLElement>

fn is_body_or_frameset(&self) -> bool

pub(crate) fn get_inner_outer_text(&self) -> DOMString

Calls into the layout engine to generate a plain text representation of a HTMLElement as specified when getting the .innerText or .outerText in JavaScript.`

https://html.spec.whatwg.org/multipage/#get-the-text-steps

pub(crate) fn set_inner_text(&self, input: DOMString, can_gc: CanGc)

https://html.spec.whatwg.org/multipage/#set-the-inner-text-steps

impl HTMLElement

pub(crate) fn set_custom_attr( &self, name: DOMString, value: DOMString, can_gc: CanGc, ) -> ErrorResult

pub(crate) fn get_custom_attr(&self, local_name: DOMString) -> Option<DOMString>

pub(crate) fn delete_custom_attr(&self, local_name: DOMString, can_gc: CanGc)

pub(crate) fn is_labelable_element(&self) -> bool

https://html.spec.whatwg.org/multipage/#category-label

pub(crate) fn is_form_associated_custom_element(&self) -> bool

https://html.spec.whatwg.org/multipage/#form-associated-custom-element

pub(crate) fn is_listed_element(&self) -> bool

https://html.spec.whatwg.org/multipage/#category-listed

pub(crate) fn is_body_element(&self) -> bool

https://html.spec.whatwg.org/multipage/#the-body-element-2

pub(crate) fn is_submittable_element(&self) -> bool

https://html.spec.whatwg.org/multipage/#category-submit

pub(crate) fn supported_prop_names_custom_attr(&self) -> Vec<DOMString>

pub(crate) fn label_at(&self, index: u32) -> Option<DomRoot<Node>>

pub(crate) fn labels_count(&self) -> u32

pub(crate) fn directionality(&self) -> Option<String>

pub(crate) fn summary_activation_behavior(&self)

fn is_a_summary_for_its_parent_details(&self) -> bool

https://html.spec.whatwg.org/multipage/#summary-for-its-parent-details

fn is_implicit_summary_element(&self) -> bool

Whether or not this is an implicitly generated <summary> element for a UA <details> shadow tree

fn rendered_text_fragment( &self, input: DOMString, can_gc: CanGc, ) -> DomRoot<DocumentFragment>

https://html.spec.whatwg.org/multipage/#rendered-text-fragment

fn merge_with_the_next_text_node(node: DomRoot<Node>, can_gc: CanGc)

Checks whether a given DomRoot<Node> and its next sibling are of type Text, and if so merges them into a single Text node.

https://html.spec.whatwg.org/multipage/#merge-with-the-next-text-node