Struct script::dom::htmllinkelement::HTMLLinkElement

#[repr(C)]
pub struct HTMLLinkElement {
    htmlelement: HTMLElement,
    rel_list: MutNullableDom<DOMTokenList>,
    stylesheet: DomRefCell<Option<Arc<Stylesheet>>>,
    cssom_stylesheet: MutNullableDom<CSSStyleSheet>,
    parser_inserted: Cell<bool>,
    pending_loads: Cell<u32>,
    any_failed_load: Cell<bool>,
    request_generation_id: Cell<RequestGenerationId>,
}

Fields
htmlelement: HTMLElementrel_list: MutNullableDom<DOMTokenList>stylesheet: DomRefCell<Option<Arc<Stylesheet>>>cssom_stylesheet: MutNullableDom<CSSStyleSheet>parser_inserted: Cell<bool>

https://html.spec.whatwg.org/multipage/#a-style-sheet-that-is-blocking-scripts

pending_loads: Cell<u32>

The number of loads that this link element has triggered (could be more than one because of imports) and have not yet finished.

any_failed_load: Cell<bool>

Whether any of the loads have failed.

request_generation_id: Cell<RequestGenerationId>

A monotonically increasing counter that keeps track of which stylesheet to apply.

Methods

impl HTMLLinkElement

fn __assert_parent_type(&self)

impl HTMLLinkElement

fn new_inherited(
local_name: LocalName,
prefix: Option<Prefix>,
document: &Document,
creator: ElementCreator
) -> HTMLLinkElement

pub fn new(
local_name: LocalName,
prefix: Option<Prefix>,
document: &Document,
creator: ElementCreator
) -> Root<Dom<HTMLLinkElement>>

pub fn get_request_generation_id(&self) -> RequestGenerationId

pub fn set_stylesheet(&self, s: Arc<Stylesheet>)

pub fn get_stylesheet(&self) -> Option<Arc<Stylesheet>>

pub fn get_cssom_stylesheet(&self) -> Option<Root<Dom<CSSStyleSheet>>>

pub fn is_alternate(&self) -> bool

impl HTMLLinkElement

fn handle_stylesheet_url(&self, href: &str)

https://html.spec.whatwg.org/multipage/#concept-link-obtain

fn handle_favicon_url(&self, _rel: &str, href: &str, _sizes: &Option<String>)