Struct script::dom::htmllinkelement::HTMLLinkElement

#[repr(C)]
pub struct HTMLLinkElement {
    htmlelement: HTMLElement,
    rel_list: MutNullableDom<DOMTokenList>,
    relations: Cell<LinkRelations>,
    stylesheet: DomRefCell<Option<Arc<Stylesheet>>>,
    cssom_stylesheet: MutNullableDom<CSSStyleSheet>,
    parser_inserted: Cell<bool>,
    pending_loads: Cell<u32>,
    any_failed_load: Cell<bool>,
    request_generation_id: Cell<RequestGenerationId>,
}

Fields
htmlelement: HTMLElementrel_list: MutNullableDom<DOMTokenList>

The relations as specified by the “rel” attribute

relations: Cell<LinkRelations>

The link relations as they are used in practice.

The reason this is seperate from HTMLLinkElement::rel_list is that a literal list is a bit unwieldy and that there are corner cases to consider (Like rev="made" implying an author relationship that is not represented in rel_list)

stylesheet: DomRefCell<Option<Arc<Stylesheet>>>cssom_stylesheet: MutNullableDom<CSSStyleSheet>parser_inserted: Cell<bool>

https://html.spec.whatwg.org/multipage/#a-style-sheet-that-is-blocking-scripts

pending_loads: Cell<u32>

The number of loads that this link element has triggered (could be more than one because of imports) and have not yet finished.

any_failed_load: Cell<bool>

Whether any of the loads have failed.

request_generation_id: Cell<RequestGenerationId>

A monotonically increasing counter that keeps track of which stylesheet to apply.

Implementations

impl HTMLLinkElement

fn __assert_parent_type(&self)

impl HTMLLinkElement

fn new_inherited( local_name: LocalName, prefix: Option<Prefix>, document: &Document, creator: ElementCreator, ) -> HTMLLinkElement

pub fn new( local_name: LocalName, prefix: Option<Prefix>, document: &Document, proto: Option<HandleObject<'_>>, creator: ElementCreator, can_gc: CanGc, ) -> Root<Dom<HTMLLinkElement>>

pub fn get_request_generation_id(&self) -> RequestGenerationId

pub fn set_stylesheet(&self, s: Arc<Stylesheet>)

pub fn get_stylesheet(&self) -> Option<Arc<Stylesheet>>

pub fn get_cssom_stylesheet(&self) -> Option<Root<Dom<CSSStyleSheet>>>

pub fn is_alternate(&self) -> bool

fn clean_stylesheet_ownership(&self)

impl HTMLLinkElement

fn processing_options(&self) -> LinkProcessingOptions

https://html.spec.whatwg.org/multipage/#create-link-options-from-element

fn fetch_and_process_prefetch_link(&self, href: &str)

The fetch and process the linked resource algorithm for rel="prefetch"

fn handle_stylesheet_url(&self, href: &str)

https://html.spec.whatwg.org/multipage/#concept-link-obtain

fn handle_favicon_url(&self, href: &str, _sizes: &Option<String>)