Struct script::dom::htmlmediaelement::HTMLMediaElement

#[repr(C)]
pub struct HTMLMediaElement {
    htmlelement: HTMLElement,
    network_state: Cell<NetworkState>,
    ready_state: Cell<ReadyState>,
    src_object: DomRefCell<Option<SrcObject>>,
    current_src: DomRefCell<String>,
    generation_id: Cell<u32>,
    fired_loadeddata_event: Cell<bool>,
    error: MutNullableDom<MediaError>,
    paused: Cell<bool>,
    defaultPlaybackRate: Cell<f64>,
    playbackRate: Cell<f64>,
    autoplaying: Cell<bool>,
    delaying_the_load_event_flag: DomRefCell<Option<LoadBlocker>>,
    pending_play_promises: DomRefCell<Vec<Rc<Promise>>>,
    in_flight_play_promises_queue: DomRefCell<VecDeque<(Box<[Rc<Promise>]>, Result<(), Error>)>>,
    player: DomRefCell<Option<Arc<Mutex<dyn Player>>>>,
    video_renderer: Arc<Mutex<MediaFrameRenderer>>,
    audio_renderer: DomRefCell<Option<Arc<Mutex<dyn AudioRenderer>>>>,
    show_poster: Cell<bool>,
    duration: Cell<f64>,
    playback_position: Cell<f64>,
    default_playback_start_position: Cell<f64>,
    volume: Cell<f64>,
    seeking: Cell<bool>,
    muted: Cell<bool>,
    resource_url: DomRefCell<Option<ServoUrl>>,
    blob_url: DomRefCell<Option<ServoUrl>>,
    played: DomRefCell<TimeRangesContainer>,
    audio_tracks_list: MutNullableDom<AudioTrackList>,
    video_tracks_list: MutNullableDom<VideoTrackList>,
    text_tracks_list: MutNullableDom<TextTrackList>,
    next_timeupdate_event: Cell<Timespec>,
    current_fetch_context: DomRefCell<Option<HTMLMediaElementFetchContext>>,
    id: Cell<u64>,
    media_controls_id: DomRefCell<Option<String>>,
    player_context: WindowGLContext,
}

Fields
htmlelement: HTMLElementnetwork_state: Cell<NetworkState>

https://html.spec.whatwg.org/multipage/#dom-media-networkstate

ready_state: Cell<ReadyState>

https://html.spec.whatwg.org/multipage/#dom-media-readystate

src_object: DomRefCell<Option<SrcObject>>

https://html.spec.whatwg.org/multipage/#dom-media-srcobject

current_src: DomRefCell<String>

https://html.spec.whatwg.org/multipage/#dom-media-currentsrc

generation_id: Cell<u32>

Incremented whenever tasks associated with this element are cancelled.

fired_loadeddata_event: Cell<bool>

https://html.spec.whatwg.org/multipage/#fire-loadeddata

Reset to false every time the load algorithm is invoked.

error: MutNullableDom<MediaError>

https://html.spec.whatwg.org/multipage/#dom-media-error

paused: Cell<bool>

https://html.spec.whatwg.org/multipage/#dom-media-paused

defaultPlaybackRate: Cell<f64>

https://html.spec.whatwg.org/multipage/#dom-media-defaultplaybackrate

playbackRate: Cell<f64>

https://html.spec.whatwg.org/multipage/#dom-media-playbackrate

autoplaying: Cell<bool>

https://html.spec.whatwg.org/multipage/#attr-media-autoplay

delaying_the_load_event_flag: DomRefCell<Option<LoadBlocker>>

https://html.spec.whatwg.org/multipage/#delaying-the-load-event-flag

pending_play_promises: DomRefCell<Vec<Rc<Promise>>>

https://html.spec.whatwg.org/multipage/#list-of-pending-play-promises

in_flight_play_promises_queue: DomRefCell<VecDeque<(Box<[Rc<Promise>]>, Result<(), Error>)>>

Play promises which are soon to be fulfilled by a queued task.

player: DomRefCell<Option<Arc<Mutex<dyn Player>>>>video_renderer: Arc<Mutex<MediaFrameRenderer>>audio_renderer: DomRefCell<Option<Arc<Mutex<dyn AudioRenderer>>>>show_poster: Cell<bool>

https://html.spec.whatwg.org/multipage/#show-poster-flag

duration: Cell<f64>

https://html.spec.whatwg.org/multipage/#dom-media-duration

playback_position: Cell<f64>

https://html.spec.whatwg.org/multipage/#official-playback-position

default_playback_start_position: Cell<f64>

https://html.spec.whatwg.org/multipage/#default-playback-start-position

volume: Cell<f64>

https://html.spec.whatwg.org/multipage/#dom-media-volume

seeking: Cell<bool>

https://html.spec.whatwg.org/multipage/#dom-media-seeking

muted: Cell<bool>

https://html.spec.whatwg.org/multipage/#dom-media-muted

resource_url: DomRefCell<Option<ServoUrl>>

URL of the media resource, if any.

blob_url: DomRefCell<Option<ServoUrl>>

URL of the media resource, if the resource is set through the src_object attribute and it is a blob.

played: DomRefCell<TimeRangesContainer>

https://html.spec.whatwg.org/multipage/#dom-media-played

audio_tracks_list: MutNullableDom<AudioTrackList>video_tracks_list: MutNullableDom<VideoTrackList>text_tracks_list: MutNullableDom<TextTrackList>

https://html.spec.whatwg.org/multipage/#dom-media-texttracks

next_timeupdate_event: Cell<Timespec>

Time of last timeupdate notification.

current_fetch_context: DomRefCell<Option<HTMLMediaElementFetchContext>>

Latest fetch request context.

id: Cell<u64>

Player Id reported the player thread

media_controls_id: DomRefCell<Option<String>>

Media controls id. In order to workaround the lack of privileged JS context, we secure the the access to the "privileged" document.servoGetMediaControls(id) API by keeping a whitelist of media controls identifiers.

player_context: WindowGLContext
Methods

impl HTMLMediaElement

pub fn new_inherited(
tag_name: LocalName,
prefix: Option<Prefix>,
document: &Document
) -> Self

pub fn get_ready_state(&self) -> ReadyState

fn media_type_id(&self) -> HTMLMediaElementTypeId

fn play_media(&self)

pub fn delay_load_event(&self, delay: bool)

Marks that element as delaying the load event or not.

Nothing happens if the element was already delaying the load event and we pass true to that method again.

https://html.spec.whatwg.org/multipage/#delaying-the-load-event-flag

fn time_marches_on(&self)

https://html.spec.whatwg.org/multipage/#time-marches-on

fn internal_pause_steps(&self)

https://html.spec.whatwg.org/multipage/#internal-pause-steps

fn is_allowed_to_play(&self) -> bool

fn notify_about_playing(&self)

fn change_ready_state(&self, ready_state: ReadyState)

fn invoke_resource_selection_algorithm(&self)

fn resource_selection_algorithm_sync(&self, base_url: ServoUrl)

fn fetch_request(&self, offset: Option<u64>, seek_lock: Option<SeekLock>)

fn resource_fetch_algorithm(&self, resource: Resource)

fn queue_dedicated_media_source_failure_steps(&self)

Queues a task to run the dedicated media source failure steps.

fn queue_ratechange_event(&self)

fn is_potentially_playing(&self) -> bool

fn is_blocked_media_element(&self) -> bool

fn is_paused_for_user_interaction(&self) -> bool

fn is_paused_for_in_band_content(&self) -> bool

fn media_element_load_algorithm(&self)

fn push_pending_play_promise(&self, promise: &Rc<Promise>)

Appends a promise to the list of pending play promises.

fn take_pending_play_promises(&self, result: Result<(), Error>)

Takes the pending play promises.

The result with which these promises will be fulfilled is passed here and this method returns nothing because we actually just move the current list of pending play promises to the in_flight_play_promises_queue field.

Each call to this method must be followed by a call to fulfill_in_flight_play_promises, to actually fulfill the promises which were taken and moved to the in-flight queue.

fn fulfill_in_flight_play_promises<F>(&self, f: F) where
F: FnOnce(),

Fulfills the next in-flight play promises queue after running a closure.

See the comment on take_pending_play_promises for why this method does not take a list of promises to fulfill. Callers cannot just pop the front list off of in_flight_play_promises_queue and later fulfill the promises because that would mean putting #[allow(unrooted_must_root)] on even more functions, potentially hiding actual safety bugs.

pub fn handle_source_child_insertion(&self)

Handles insertion of source children.

https://html.spec.whatwg.org/multipage/#the-source-element:nodes-are-inserted

fn seek(&self, time: f64, _approximate_for_speed: bool)

fn seek_end(&self)

pub fn process_poster_response(&self, image: ImageResponse)

https://html.spec.whatwg.org/multipage/#poster-frame

fn setup_media_player(&self, resource: &Resource) -> Result<(), ()>

pub fn set_audio_track(&self, idx: usize, enabled: bool)

pub fn set_video_track(&self, idx: usize, enabled: bool)

fn handle_player_event(&self, event: &PlayerEvent)

fn earliest_possible_position(&self) -> f64

fn render_controls(&self)

fn remove_controls(&self)

pub fn get_current_frame(&self) -> Option<VideoFrame>

pub fn set_audio_renderer(&self, audio_renderer: Arc<Mutex<dyn AudioRenderer>>)

By default the audio is rendered through the audio sink automatically selected by the servo-media Player instance. However, in some cases, like the WebAudio MediaElementAudioSourceNode, we need to set a custom audio renderer.

fn send_media_session_event(&self, event: MediaSessionEvent)

pub fn set_duration(&self, duration: f64)

pub fn reset(&self)

impl HTMLMediaElement

fn direction_of_playback(&self) -> PlaybackDirection