Module script::dom::macros

Macros

beforeunload_event_handler ğŸ”’

define_event_handler ğŸ”’

These are used to generate a event handler which has no special case.

define_window_owned_event_handler ğŸ”’

document_and_element_event_handlers ğŸ”’

error_event_handler ğŸ”’

event_handler ğŸ”’

global_event_handlers ğŸ”’

handle_potential_webgl_error ğŸ”’

impl_performance_entry_struct ğŸ”’

DOM struct implementation for simple interfaces inheriting from PerformanceEntry.

impl_rare_data ğŸ”’

unsafe_no_jsmanaged_fields ğŸ”’

For use on non-jsmanaged types Use #[derive(JSTraceable)] on JS managed types

window_event_handlers ğŸ”’

window_owned_beforeunload_event_handler ğŸ”’

window_owned_event_handler ğŸ”’