Module script::dom::macros

Macros

beforeunload_event_handler

define_event_handler

These are used to generate a event handler which has no special case.

define_window_owned_event_handler

document_and_element_event_handlers

error_event_handler

event_handler

global_event_handlers

handle_potential_webgl_error

impl_performance_entry_struct

DOM struct implementation for simple interfaces inheriting from PerformanceEntry.

impl_rare_data

jsmanaged_array

unsafe_no_jsmanaged_fields

For use on non-jsmanaged types Use #[derive(JSTraceable)] on JS managed types

window_event_handlers

window_owned_beforeunload_event_handler

window_owned_event_handler