Module script::dom::node

The core DOM types. Defines the basic DOM hierarchy as well as all the HTML elements.

Structs

AncestorIterator

FollowingNodeIterator

LastChildIterator

Node
An HTML node.

NodeFlags
Flags for node items.

NodeSiblingIterator

OpaqueStyleAndLayoutData

PrecedingNodeIterator

QuerySelectorIterator

ReverseSiblingIterator

TreeIterator

TrustedNodeAddress
The address of a node known to be valid. These are sent from script to layout, and are also used in the HTML parser interface.

UnbindContext
The context of the unbinding from a tree of a node when one of its inclusive ancestors is removed.

UniqueId
A node's unique ID, for devtools.

Enums

ChildrenMutation

CloneChildrenFlag
Specifies whether children must be recursively cloned or not.

NodeDamage
A summary of the changes that happened to a node.

SuppressObserver
suppress observers flag https://dom.spec.whatwg.org/#concept-node-insert https://dom.spec.whatwg.org/#concept-node-remove

Constants

CAN_BE_FRAGMENTED
Whether any ancestor is a fragmentation container

CLICK_IN_PROGRESS
Specifies whether or not there is an authentic click in progress on this element.

HAS_CHANGED
Specifies whether this node must be reflowed regardless of style differences.

HAS_DIRTY_DESCENDANTS
Specifies whether this node has descendants (inclusive of itself) which have changed since the last reflow.

IS_DIRTY
Specifies whether this node needs style recalc on next reflow.

IS_IN_DOC
Specifies whether this node is in a document.

SEQUENTIALLY_FOCUSABLE
Specifies whether this node is focusable and whether it is supposed to be reachable with using sequential focus navigation.

Traits

LayoutNodeHelpers

Functions

as_uintptr

document_from_node

first_node_not_in
Iterate through nodes until we find a Node that is not in not_in

from_untrusted_node_address
If the given untrusted node address represents a valid DOM node in the given runtime, returns it.

window_from_node

Type Definitions

ChildElementIterator