Module script::dom::node

The core DOM types. Defines the basic DOM hierarchy as well as all the HTML elements.

Structs

AncestorIterator

LayoutData
Encapsulates the abstract layout data.

LayoutDataRef

Node
An HTML node.

NodeChildrenIterator

NodeFlags
Flags for node items.

QuerySelectorIterator

ReverseChildrenIterator

SharedLayoutData
Layout data that is shared between the script and layout tasks.

TreeIterator

TrustedNodeAddress
The address of a node known to be valid. These are sent from script to layout, and are also used in the HTML parser interface.

Enums

CloneChildrenFlag
Specifies whether children must be recursively cloned or not.

NodeDamage
A summary of the changes that happened to a node.

NodeTypeId
The different types of nodes.

Constants

CLICK_IN_PROGRESS
Specifies whether or not there is an authentic click in progress on this element.

HAS_CHANGED
Specifies whether this node must be reflowed regardless of style differences.

HAS_DIRTY_DESCENDANTS
Specifies whether this node has descendants (inclusive of itself) which have changed since the last reflow.

HAS_DIRTY_SIBLINGS
Specifies whether this node has siblings (inclusive of itself) which changed since the last reflow.

IN_DISABLED_STATE
Specifies whether this node is in disabled state.

IN_ENABLED_STATE
Specifies whether this node is in enabled state.

IN_HOVER_STATE
Specifies whether this node is in hover state.

IS_DIRTY
Specifies whether this node needs style recalc on next reflow.

IS_IN_DOC
Specifies whether this node is in a document.

Traits

DisabledStateHelpers

LayoutNodeHelpers

NodeHelpers

RawLayoutNodeHelpers

Functions

document_from_node

from_untrusted_node_address
If the given untrusted node address represents a valid DOM node in the given runtime, returns it.

window_from_node

Type Definitions

ChildElementIterator