Struct script::dom::node::Node

pub struct Node {
    eventtarget: EventTarget,
    parent_node: MutNullableHeap<JS<Node>>,
    first_child: MutNullableHeap<JS<Node>>,
    last_child: MutNullableHeap<JS<Node>>,
    next_sibling: MutNullableHeap<JS<Node>>,
    prev_sibling: MutNullableHeap<JS<Node>>,
    owner_doc: MutNullableHeap<JS<Document>>,
    child_list: MutNullableHeap<JS<NodeList>>,
    children_count: Cell<u32>,
    flags: Cell<NodeFlags>,
    inclusive_descendants_version: Cell<u64>,
    ranges: WeakRangeVec,
    style_and_layout_data: Cell<Option<OpaqueStyleAndLayoutData>>,
    unique_id: UniqueId,
}

An HTML node.

Fields

eventtarget

The JavaScript reflector for this node.

parent_node

The parent of this node.

first_child

The first child of this node.

last_child

The last child of this node.

next_sibling

The next sibling of this node.

prev_sibling

The previous sibling of this node.

owner_doc

The document that this node belongs to.

child_list

The live list of children return by .childNodes.

children_count

The live count of children of this node.

flags

A bitfield of flags for node items.

inclusive_descendants_version

The maximum version of any inclusive descendant of this node.

ranges

A vector of weak references to Range instances of which the start or end containers are this node. No range should ever be found twice in this vector, even if both the start and end containers are this node.

style_and_layout_data

Style+Layout information. Only the layout thread may touch this data.

Must be sent back to the layout thread to be destroyed when this node is finalized.

unique_id

Methods

impl Node

fn add_child(&self, new_child: &Node, before: Option<&Node>)

Adds a new child to the end of this node's list of children.

Fails unless new_child is disconnected from the tree.

fn remove_child(&self, child: &Node, cached_index: Option<u32>)

Removes the given child from this node's list of children.

Fails unless child is a child of this node.

impl Node

fn teardown(&self)

fn dump(&self)

Dumps the subtree rooted at this node, for debugging.

fn dump_indent(&self, indent: u32)

Dumps the node tree, for debugging, with indentation.

fn debug_str(&self) -> String

Returns a string that describes this node.

fn is_in_doc(&self) -> bool

fn type_id(&self) -> NodeTypeId

Returns the type ID of this node.

fn len(&self) -> u32

fn index(&self) -> u32

fn children_count(&self) -> u32

fn ranges(&self) -> &WeakRangeVec

fn is_doctype(&self) -> bool

fn get_flag(&self, flag: NodeFlags) -> bool

fn set_flag(&self, flag: NodeFlags, value: bool)

fn has_changed(&self) -> bool

fn set_has_changed(&self, state: bool)

fn is_dirty(&self) -> bool

fn set_is_dirty(&self, state: bool)

fn has_dirty_descendants(&self) -> bool

fn set_has_dirty_descendants(&self, state: bool)

fn force_dirty_ancestors(&self, damage: NodeDamage)

fn rev_version(&self)

fn dirty(&self, damage: NodeDamage)

fn dirty_impl(&self, damage: NodeDamage, force_ancestors: bool)

fn inclusive_descendants_version(&self) -> u64

The maximum version number of this node's descendants, including itself

fn traverse_preorder(&self) -> TreeIterator

Iterates over this node and all its descendants, in preorder.

fn inclusively_following_siblings(&self) -> NodeSiblingIterator

fn inclusively_preceding_siblings(&self) -> ReverseSiblingIterator

fn is_inclusive_ancestor_of(&self, parent: &Node) -> bool

fn is_ancestor_of(&self, parent: &Node) -> bool

fn following_siblings(&self) -> NodeSiblingIterator

fn preceding_siblings(&self) -> ReverseSiblingIterator

fn following_nodes(&self, root: &Node) -> FollowingNodeIterator

fn preceding_nodes(&self, root: &Node) -> PrecedingNodeIterator

fn descending_last_children(&self) -> LastChildIterator

fn is_parent_of(&self, child: &Node) -> bool

fn to_trusted_node_address(&self) -> TrustedNodeAddress

fn bounding_content_box(&self) -> Rect<Au>

fn content_boxes(&self) -> Vec<Rect<Au>>

fn client_rect(&self) -> Rect<i32>

fn scroll_area(&self) -> Rect<i32>

fn scroll_offset(&self) -> Point2D<f32>

fn before(&self, nodes: Vec<NodeOrString>) -> ErrorResult

fn after(&self, nodes: Vec<NodeOrString>) -> ErrorResult

fn replace_with(&self, nodes: Vec<NodeOrString>) -> ErrorResult

fn prepend(&self, nodes: Vec<NodeOrString>) -> ErrorResult

fn append(&self, nodes: Vec<NodeOrString>) -> ErrorResult

fn query_selector(&self, selectors: DOMString) -> Fallible<Option<Root<Element>>>

fn query_selector_iter(&self, selectors: DOMString) -> Fallible<QuerySelectorIterator>

https://dom.spec.whatwg.org/#scope-match-a-selectors-string Get an iterator over all nodes which match a set of selectors Be careful not to do anything which may manipulate the DOM tree whilst iterating, otherwise the iterator may be invalidated.

fn query_selector_all(&self, selectors: DOMString) -> Fallible<Root<NodeList>>

fn ancestors(&self) -> AncestorIterator

fn inclusive_ancestors(&self) -> AncestorIterator

fn owner_doc(&self) -> Root<Document>

fn set_owner_doc(&self, document: &Document)

fn is_in_html_doc(&self) -> bool

fn children(&self) -> NodeSiblingIterator

fn rev_children(&self) -> ReverseSiblingIterator

fn child_elements(&self) -> ChildElementIterator

fn remove_self(&self)

fn unique_id(&self) -> String

fn summarize(&self) -> NodeInfo

fn parse_fragment(&self, markup: DOMString) -> Fallible<Root<DocumentFragment>>

fn insert_cell_or_row<F, G, I>(&self, index: i32, get_items: F, new_child: G) -> Fallible<Root<HTMLElement>> where F: Fn() -> Root<HTMLCollection>, G: Fn() -> Root<I>, I: DerivedFrom<Node> + DerivedFrom<HTMLElement> + Reflectable

Used by HTMLTableSectionElement::InsertRow and HTMLTableRowElement::InsertCell

fn delete_cell_or_row<F, G>(&self, index: i32, get_items: F, is_delete_type: G) -> ErrorResult where F: Fn() -> Root<HTMLCollection>, G: Fn(&Element) -> bool

Used by HTMLTableSectionElement::DeleteRow and HTMLTableRowElement::DeleteCell

impl Node

fn reflect_node<N: DerivedFrom<Node> + Reflectable>(node: Box<N>, document: &Document, wrap_fn: fn(*mut JSContext, GlobalRef, Box<N>) -> Root<N>) -> Root<N>

fn new_inherited(doc: &Document) -> Node

fn new_document_node() -> Node

fn new_(flags: NodeFlags, doc: Option<&Document>) -> Node

fn adopt(node: &Node, document: &Document)

fn ensure_pre_insertion_validity(node: &Node, parent: &Node, child: Option<&Node>) -> ErrorResult

fn pre_insert(node: &Node, parent: &Node, child: Option<&Node>) -> Fallible<Root<Node>>

fn insert(node: &Node, parent: &Node, child: Option<&Node>, suppress_observers: SuppressObserver)

fn replace_all(node: Option<&Node>, parent: &Node)

fn pre_remove(child: &Node, parent: &Node) -> Fallible<Root<Node>>

fn remove(node: &Node, parent: &Node, suppress_observers: SuppressObserver)

fn clone(node: &Node, maybe_doc: Option<&Document>, clone_children: CloneChildrenFlag) -> Root<Node>

fn collect_text_contents<T: Iterator<Item=Root<Node>>>(iterator: T) -> DOMString

fn namespace_to_string(namespace: Namespace) -> Option<DOMString>

fn locate_namespace(node: &Node, prefix: Option<DOMString>) -> Namespace