Module script::dom::promise

Native representation of JS Promise values.

This implementation differs from the traditional Rust DOM object, because the reflector is provided by SpiderMonkey and has no knowledge of an associated native representation (ie. dom::Promise). This means that native instances use native reference counting (Rc) to ensure that no memory is leaked, which means that there can be multiple instances of native Promise values that refer to the same JS value yet are distinct native objects (ie. address equality for the native objects is meaningless).

Structs

Promise

Enums

NativeHandlerTask 🔒

Constants

SLOT_NATIVEHANDLER 🔒

SLOT_NATIVEHANDLER_TASK 🔒

_IMPL_DOMOBJECT_FOR_Promise 🔒

Traits

PromiseHelper 🔒

Private helper to enable adding new methods to Rc<Promise>.

Functions

create_native_handler_function 🔒

do_nothing_promise_executor 🔒 ⚠

native_handler_callback 🔒 ⚠