Module script::dom::readablestream

Structs

ExternalUnderlyingSourceController 🔒

ReadableStream

Enums

ExternalUnderlyingSource

Constants

_IMPL_DOMOBJECT_FOR_ReadableStream 🔒

Statics

UNDERLYING_SOURCE_TRAPS 🔒

Functions

cancel 🔒 ⚠

close 🔒 ⚠

error 🔒 ⚠

finalize 🔒 ⚠

get_read_promise_bytes

Get the value property of an object that a read promise resolved to.

get_read_promise_done

Get the done property of an object that a read promise resolved to.

request_data 🔒 ⚠

write_into_read_request_buffer 🔒 ⚠