script::dom

Module servoparser

Modules

async_html ğŸ”’

encoding ğŸ”’

html ğŸ”’

prefetch ğŸ”’

xml ğŸ”’

Structs

ElementAttribute ğŸ”’

FragmentContext ğŸ”’

FragmentParsingResult ğŸ”’

NavigationParams ğŸ”’

https://html.spec.whatwg.org/multipage/#navigation-params This does not have the relevant fields, but mimics the intent of the struct when used in loading document spec algorithms.

ParserContext ğŸ”’

The context required for asynchronously fetching a document and parsing it progressively.

ServoParser ğŸ”’

The parser maintains two input streams: one for input from script through document.write(), and one for input from network.

Sink ğŸ”’

Enums

ParserKind ğŸ”’

ParsingAlgorithm ğŸ”’

Tokenizer ğŸ”’

Constants

_IMPL_DOMOBJECT_FOR_ServoParser ğŸ”’

Functions

attach_declarative_shadow_inner ğŸ”’

create_element_for_token ğŸ”’

https://html.spec.whatwg.org/multipage/#create-an-element-for-the-token

insert ğŸ”’