Struct script::dom::servoparser::ServoParser

pub struct ServoParser {
    reflector: Reflector,
    document: JS<Document>,
    pipeline: Option<PipelineId>,
    pending_input: DOMRefCell<VecDeque<String>>,
    tokenizer: DOMRefCell<Tokenizer>,
    last_chunk_received: Cell<bool>,
    suspended: Cell<bool>,
}

Fields
reflector: Reflector document: JS<Document>

The document associated with this parser.

pipeline: Option<PipelineId>

The pipeline associated with this parse, unavailable if this parse does not correspond to a page load.

pending_input: DOMRefCell<VecDeque<String>>

Input chunks received but not yet passed to the parser.

tokenizer: DOMRefCell<Tokenizer>

The tokenizer of this parser.

last_chunk_received: Cell<bool>

Whether to expect any further input from the associated network request.

suspended: Cell<bool>

Whether this parser should avoid passing any further data to the tokenizer.

Methods

impl ServoParser

fn new_inherited(document: &Document,
pipeline: Option<PipelineId>,
tokenizer: Tokenizer,
last_chunk_state: LastChunkState)
-> Self

fn new(document: &Document,
pipeline: Option<PipelineId>,
tokenizer: Tokenizer,
last_chunk_state: LastChunkState)
-> Root<Self>

fn document(&self) -> &Document

fn pipeline(&self) -> Option<PipelineId>

fn has_pending_input(&self) -> bool

fn push_input_chunk(&self, chunk: String)

fn take_next_input_chunk(&self) -> Option<String>

fn last_chunk_received(&self) -> bool

fn mark_last_chunk_received(&self)

fn set_plaintext_state(&self)

fn end_tokenizer(&self)

fn suspend(&self)

fn resume(&self)

fn is_suspended(&self) -> bool

fn parse_sync(&self)

fn do_parse_sync(&self)

fn parse_chunk(&self, input: String)

fn finish(&self)