script::dom::servoparser

Struct ServoParser

#[repr(C)]
pub(crate) struct ServoParser {

    reflector: Reflector,
    document: Dom<Document>,
    network_decoder: DomRefCell<NetworkDecoderState>,
    network_input: BufferQueue,
    script_input: BufferQueue,
    tokenizer: Tokenizer,
    last_chunk_received: Cell<bool>,
    suspended: Cell<bool>,
    script_nesting_level: Cell<usize>,
    aborted: Cell<bool>,
    script_created_parser: bool,
    prefetch_decoder: RefCell<LossyDecoder<NetworkSink>>,
    prefetch_tokenizer: Tokenizer,
    prefetch_input: BufferQueue,
    content_for_devtools: Option<DomRefCell<String>>,

}

The parser maintains two input streams: one for input from script through document.write(), and one for input from network.

There is no concrete representation of the insertion point, instead it always points to just before the next character from the network input, with all of the script input before itself.

    ... script input ... | ... network input ...
                         ^
                insertion point

Fields
reflector: Reflectordocument: Dom<Document>

The document associated with this parser.

network_decoder: DomRefCell<NetworkDecoderState>

The decoder used for the network input.

network_input: BufferQueue

Input received from network.

script_input: BufferQueue

Input received from script. Used only to support document.write().

tokenizer: Tokenizer

The tokenizer of this parser.

last_chunk_received: Cell<bool>

Whether to expect any further input from the associated network request.

suspended: Cell<bool>

Whether this parser should avoid passing any further data to the tokenizer.

script_nesting_level: Cell<usize>

https://html.spec.whatwg.org/multipage/#script-nesting-level

aborted: Cell<bool>

https://html.spec.whatwg.org/multipage/#abort-a-parser

script_created_parser: bool

https://html.spec.whatwg.org/multipage/#script-created-parser

prefetch_decoder: RefCell<LossyDecoder<NetworkSink>>

A decoder exclusively for input to the prefetch tokenizer.

Unlike the actual decoder, this one takes a best guess at the encoding and starts decoding immediately.

prefetch_tokenizer: Tokenizer

We do a quick-and-dirty parse of the input looking for resources to prefetch.

prefetch_input: BufferQueuecontent_for_devtools: Option<DomRefCell<String>>
Implementations

impl ServoParser

fn __assert_parent_type(&self)

impl ServoParser

pub(crate) fn parser_is_not_active(&self) -> bool

pub(crate) fn parse_html_document( document: &Document, input: Option<DOMString>, url: ServoUrl, encoding_hint_from_content_type: Option<&'static Encoding>, encoding_of_container_document: Option<&'static Encoding>, can_gc: CanGc, )

https://html.spec.whatwg.org/multipage/#parse-html-from-a-string

pub(crate) fn parse_html_fragment( context: &Element, input: DOMString, allow_declarative_shadow_roots: bool, can_gc: CanGc, ) -> impl Iterator<Item = DomRoot<Node>> + use<'_>

https://html.spec.whatwg.org/multipage/#parsing-html-fragments

pub(crate) fn parse_html_script_input(document: &Document, url: ServoUrl)

pub(crate) fn parse_xml_document( document: &Document, input: Option<DOMString>, url: ServoUrl, encoding_hint_from_content_type: Option<&'static Encoding>, can_gc: CanGc, )

pub(crate) fn script_nesting_level(&self) -> usize

pub(crate) fn is_script_created(&self) -> bool

pub(crate) fn resume_with_pending_parsing_blocking_script( &self, script: &HTMLScriptElement, result: Result<Script, NoTrace<NetworkError>>, can_gc: CanGc, )

Corresponds to the latter part of the “Otherwise” branch of the ‘An end tag whose tag name is “script”’ of https://html.spec.whatwg.org/multipage/#parsing-main-incdata

This first moves everything from the script input to the beginning of the network input, effectively resetting the insertion point to just before the next character to be consumed.

    | ... script input ... network input ...
    ^
    insertion point

pub(crate) fn can_write(&self) -> bool

pub(crate) fn write(&self, text: DOMString, can_gc: CanGc)

Steps 6-8 of https://html.spec.whatwg.org/multipage/#document.write()

pub(crate) fn close(&self, can_gc: CanGc)

pub(crate) fn abort(&self, can_gc: CanGc)

pub(crate) fn is_active(&self) -> bool

pub(crate) fn get_current_line(&self) -> u32

fn new_inherited( document: &Document, tokenizer: Tokenizer, kind: ParserKind, encoding_hint_from_content_type: Option<&'static Encoding>, encoding_of_container_document: Option<&'static Encoding>, ) -> Self

fn new( document: &Document, tokenizer: Tokenizer, kind: ParserKind, encoding_hint_from_content_type: Option<&'static Encoding>, encoding_of_container_document: Option<&'static Encoding>, can_gc: CanGc, ) -> DomRoot<Self>

fn push_tendril_input_chunk(&self, chunk: StrTendril)

fn push_bytes_input_chunk(&self, chunk: Vec<u8>)

fn should_prefetch(&self) -> bool

fn push_string_input_chunk(&self, chunk: String)

fn parse_sync(&self, can_gc: CanGc)

fn parse_complete_string_chunk(&self, input: String, can_gc: CanGc)

fn parse_bytes_chunk(&self, input: Vec<u8>, can_gc: CanGc)

fn tokenize<F>(&self, feed: F, can_gc: CanGc)
where F: Fn(&Tokenizer) -> TokenizerResult<DomRoot<HTMLScriptElement>>,

fn finish(&self, can_gc: CanGc)

https://html.spec.whatwg.org/multipage/#the-end