Trait script::dom::virtualmethods::VirtualMethods

pub trait VirtualMethods {
    fn super_type(&self) -> Option<&VirtualMethods>;

    fn attribute_mutated(&self, attr: &Attr, mutation: AttributeMutation) { ... }
    fn parse_plain_attribute(&self, name: &Atom, value: DOMString) -> AttrValue { ... }
    fn bind_to_tree(&self, tree_in_doc: bool) { ... }
    fn unbind_from_tree(&self, context: &UnbindContext) { ... }
    fn children_changed(&self, mutation: &ChildrenMutation) { ... }
    fn handle_event(&self, event: &Event) { ... }
    fn adopting_steps(&self, old_doc: &Document) { ... }
    fn cloning_steps(&self, copy: &Node, maybe_doc: Option<&Document>, clone_children: CloneChildrenFlag) { ... }
}

Trait to allow DOM nodes to opt-in to overriding (or adding to) common behaviours. Replicates the effect of C++ virtual methods.

Required Methods

fn super_type(&self) -> Option<&VirtualMethods>

Returns self as the superclass of the implementation for this trait, if any.

Provided Methods

fn attribute_mutated(&self, attr: &Attr, mutation: AttributeMutation)

Called when attributes of a node are mutated. https://dom.spec.whatwg.org/#attribute-is-set https://dom.spec.whatwg.org/#attribute-is-removed

fn parse_plain_attribute(&self, name: &Atom, value: DOMString) -> AttrValue

Returns the right AttrValue variant for the attribute with name name on this element.

fn bind_to_tree(&self, tree_in_doc: bool)

Called when a Node is appended to a tree, where 'tree_in_doc' indicates whether the tree is part of a Document.

fn unbind_from_tree(&self, context: &UnbindContext)

Called when a Node is removed from a tree, where 'tree_in_doc' indicates whether the tree is part of a Document. Implements removing steps: https://dom.spec.whatwg.org/#concept-node-remove-ext

fn children_changed(&self, mutation: &ChildrenMutation)

Called on the parent when its children are changed.

fn handle_event(&self, event: &Event)

Called during event dispatch after the bubbling phase completes.

fn adopting_steps(&self, old_doc: &Document)

https://dom.spec.whatwg.org/#concept-node-adopt-ext

fn cloning_steps(&self, copy: &Node, maybe_doc: Option<&Document>, clone_children: CloneChildrenFlag)

https://dom.spec.whatwg.org/#concept-node-clone-ext