script::dom::virtualmethods

Trait VirtualMethods

pub(crate) trait VirtualMethods {
    // Required method
    fn super_type(&self) -> Option<&dyn VirtualMethods>;

    // Provided methods
    fn attribute_mutated(
        &self,
        attr: &Attr,
        mutation: AttributeMutation<'_>,
        can_gc: CanGc,
    ) { ... }
    fn attribute_affects_presentational_hints(&self, attr: &Attr) -> bool { ... }
    fn parse_plain_attribute(
        &self,
        name: &LocalName,
        value: DOMString,
    ) -> AttrValue { ... }
    fn post_connection_steps(&self, can_gc: CanGc) { ... }
    fn bind_to_tree(&self, context: &BindContext<'_>, can_gc: CanGc) { ... }
    fn unbind_from_tree(&self, context: &UnbindContext<'_>, can_gc: CanGc) { ... }
    fn children_changed(&self, mutation: &ChildrenMutation<'_>, can_gc: CanGc) { ... }
    fn handle_event(&self, event: &Event, can_gc: CanGc) { ... }
    fn adopting_steps(&self, old_doc: &Document, can_gc: CanGc) { ... }
    fn cloning_steps(
        &self,
        copy: &Node,
        maybe_doc: Option<&Document>,
        clone_children: CloneChildrenFlag,
        can_gc: CanGc,
    ) { ... }
    fn pop(&self) { ... }
}

Trait to allow DOM nodes to opt-in to overriding (or adding to) common behaviours. Replicates the effect of C++ virtual methods.

Required Methods

fn super_type(&self) -> Option<&dyn VirtualMethods>

Returns self as the superclass of the implementation for this trait, if any.

Provided Methods

fn attribute_mutated( &self, attr: &Attr, mutation: AttributeMutation<'_>, can_gc: CanGc, )

Called when attributes of a node are mutated. https://dom.spec.whatwg.org/#attribute-is-set https://dom.spec.whatwg.org/#attribute-is-removed

fn attribute_affects_presentational_hints(&self, attr: &Attr) -> bool

Returns true if given attribute attr affects style of the given element.

fn parse_plain_attribute(&self, name: &LocalName, value: DOMString) -> AttrValue

Returns the right AttrValue variant for the attribute with name name on this element.

fn post_connection_steps(&self, can_gc: CanGc)

Invoked during a DOM tree mutation after a node becomes connected, once all related DOM tree mutations have been applied. https://dom.spec.whatwg.org/#concept-node-post-connection-ext

fn bind_to_tree(&self, context: &BindContext<'_>, can_gc: CanGc)

Called when a Node is appended to a tree.

fn unbind_from_tree(&self, context: &UnbindContext<'_>, can_gc: CanGc)

Called when a Node is removed from a tree. Implements removing steps: https://dom.spec.whatwg.org/#concept-node-remove-ext

fn children_changed(&self, mutation: &ChildrenMutation<'_>, can_gc: CanGc)

Called on the parent when its children are changed.

fn handle_event(&self, event: &Event, can_gc: CanGc)

Called during event dispatch after the bubbling phase completes.

fn adopting_steps(&self, old_doc: &Document, can_gc: CanGc)

https://dom.spec.whatwg.org/#concept-node-adopt-ext

fn cloning_steps( &self, copy: &Node, maybe_doc: Option<&Document>, clone_children: CloneChildrenFlag, can_gc: CanGc, )

https://dom.spec.whatwg.org/#concept-node-clone-ext

fn pop(&self)

Called on an element when it is popped off the stack of open elements of a parser.