Trait script::dom::virtualmethods::VirtualMethods

pub trait VirtualMethods {
    fn super_type<'b>(&'b self) -> Option<&'b VirtualMethods>;

    fn after_set_attr(&self, attr: JSRef<Attr>) { ... }
    fn before_remove_attr(&self, attr: JSRef<Attr>) { ... }
    fn parse_plain_attribute(&self, name: &Atom, value: DOMString) -> AttrValue { ... }
    fn bind_to_tree(&self, tree_in_doc: bool) { ... }
    fn unbind_from_tree(&self, tree_in_doc: bool) { ... }
    fn child_inserted(&self, child: JSRef<Node>) { ... }
    fn handle_event(&self, event: JSRef<Event>) { ... }
    fn cloning_steps(&self, copy: JSRef<Node>, maybe_doc: Option<JSRef<Document>>, clone_children: CloneChildrenFlag) { ... }
}

Trait to allow DOM nodes to opt-in to overriding (or adding to) common behaviours. Replicates the effect of C++ virtual methods.

Required Methods

fn super_type<'b>(&'b self) -> Option<&'b VirtualMethods>

Returns self as the superclass of the implementation for this trait, if any.

Provided Methods

fn after_set_attr(&self, attr: JSRef<Attr>)

Called when changing or adding attributes, after the attribute's value has been updated.

fn before_remove_attr(&self, attr: JSRef<Attr>)

Called when changing or removing attributes, before any modification has taken place.

fn parse_plain_attribute(&self, name: &Atom, value: DOMString) -> AttrValue

Returns the right AttrValue variant for the attribute with name name on this element.

fn bind_to_tree(&self, tree_in_doc: bool)

Called when a Node is appended to a tree, where 'tree_in_doc' indicates whether the tree is part of a Document.

fn unbind_from_tree(&self, tree_in_doc: bool)

Called when a Node is removed from a tree, where 'tree_in_doc' indicates whether the tree is part of a Document.

fn child_inserted(&self, child: JSRef<Node>)

Called on the parent when a node is added to its child list.

fn handle_event(&self, event: JSRef<Event>)

Called during event dispatch after the bubbling phase completes.

fn cloning_steps(&self, copy: JSRef<Node>, maybe_doc: Option<JSRef<Document>>, clone_children: CloneChildrenFlag)

https://dom.spec.whatwg.org/#concept-node-clone (step 5)