Enum script::dom::webgl_validations::tex_image_2d::TexImageValidationError

pub enum TexImageValidationError {
    InvalidTextureTarget(u32),
    TextureTargetNotBound(u32),
    InvalidCubicTextureDimensions,
    NegativeLevel,
    LevelTooHigh,
    NegativeDimension,
    TextureTooBig,
    InvalidDataType,
    InvalidTextureFormat,
    TextureFormatMismatch,
    InvalidTypeForFormat,
    InvalidBorder,
    NonPotTexture,
}

The errors that the texImage* family of functions can generate.

Variants
InvalidTextureTarget(u32)

An invalid texture target was passed, it contains the invalid target.

TextureTargetNotBound(u32)

The passed texture target was not bound.

InvalidCubicTextureDimensions

Invalid texture dimensions were given.

NegativeLevel

A negative level was passed.

LevelTooHigh

A level too high to be allowed by the implementation was passed.

NegativeDimension

A negative width and height was passed.

TextureTooBig

A bigger with and height were passed than what the implementation allows.

InvalidDataType

An invalid data type was passed.

InvalidTextureFormat

An invalid texture format was passed.

TextureFormatMismatch

Format did not match internal_format.

InvalidTypeForFormat

Invalid data type for the given format.

InvalidBorder

Invalid border

NonPotTexture

Expected a power of two texture.