script::dom::webxr::xrsession

Struct XRSession

#[repr(C)]
pub(crate) struct XRSession {

    eventtarget: EventTarget,
    blend_mode: XREnvironmentBlendMode,
    mode: XRSessionMode,
    visibility_state: Cell<XRVisibilityState>,
    viewer_space: MutNullableDom<XRSpace>,
    session: DomRefCell<Session>,
    frame_requested: Cell<bool>,
    pending_render_state: MutNullableDom<XRRenderState>,
    active_render_state: MutDom<XRRenderState>,
    inline_projection_matrix: DomRefCell<Transform3D<f32, Viewer, Display>>,
    next_raf_id: Cell<i32>,
    raf_callback_list: DomRefCell<Vec<(i32, Option<Rc<XRFrameRequestCallback<DomTypeHolder>>>)>>,
    current_raf_callback_list: DomRefCell<Vec<(i32, Option<Rc<XRFrameRequestCallback<DomTypeHolder>>>)>>,
    input_sources: Dom<XRInputSourceArray>,
    end_promises: DomRefCell<Vec<Rc<Promise>>>,
    ended: Cell<bool>,
    next_hit_test_id: Cell<HitTestId>,
    pending_hit_test_promises: DomRefCell<HashMapTracedValues<HitTestId, Rc<Promise>, FxBuildHasher>>,
    outside_raf: Cell<bool>,
    input_frames: DomRefCell<HashMap<InputId, InputFrame>>,
    framerate: Cell<f32>,
    update_framerate_promise: DomRefCell<Option<Rc<Promise>>>,
    reference_spaces: DomRefCell<Vec<Dom<XRReferenceSpace>>>,

}

Fields
eventtarget: EventTargetblend_mode: XREnvironmentBlendModemode: XRSessionModevisibility_state: Cell<XRVisibilityState>viewer_space: MutNullableDom<XRSpace>session: DomRefCell<Session>frame_requested: Cell<bool>pending_render_state: MutNullableDom<XRRenderState>active_render_state: MutDom<XRRenderState>inline_projection_matrix: DomRefCell<Transform3D<f32, Viewer, Display>>

Cached projection matrix for inline sessions

next_raf_id: Cell<i32>raf_callback_list: DomRefCell<Vec<(i32, Option<Rc<XRFrameRequestCallback<DomTypeHolder>>>)>>current_raf_callback_list: DomRefCell<Vec<(i32, Option<Rc<XRFrameRequestCallback<DomTypeHolder>>>)>>input_sources: Dom<XRInputSourceArray>end_promises: DomRefCell<Vec<Rc<Promise>>>ended: Cell<bool>

https://immersive-web.github.io/webxr/#ended

next_hit_test_id: Cell<HitTestId>pending_hit_test_promises: DomRefCell<HashMapTracedValues<HitTestId, Rc<Promise>, FxBuildHasher>>outside_raf: Cell<bool>

Opaque framebuffers need to know the session is “outside of a requestAnimationFrame” https://immersive-web.github.io/webxr/#opaque-framebuffer

input_frames: DomRefCell<HashMap<InputId, InputFrame>>framerate: Cell<f32>update_framerate_promise: DomRefCell<Option<Rc<Promise>>>reference_spaces: DomRefCell<Vec<Dom<XRReferenceSpace>>>
Implementations

impl XRSession

fn __assert_parent_type(&self)

impl XRSession

fn new_inherited( session: Session, render_state: &XRRenderState, input_sources: &XRInputSourceArray, mode: XRSessionMode, ) -> XRSession

pub(crate) fn new( window: &Window, session: Session, mode: XRSessionMode, frame_receiver: IpcReceiver<Frame>, can_gc: CanGc, ) -> DomRoot<XRSession>

pub(crate) fn with_session<R, F: FnOnce(&Session) -> R>(&self, with: F) -> R

pub(crate) fn is_ended(&self) -> bool

pub(crate) fn is_immersive(&self) -> bool

pub(crate) fn has_layers_feature(&self) -> bool

fn setup_raf_loop(&self, frame_receiver: IpcReceiver<Frame>)

pub(crate) fn is_outside_raf(&self) -> bool

fn attach_event_handler(&self)

pub(crate) fn setup_initial_inputs(&self)

fn event_callback(&self, event: XREvent, can_gc: CanGc)

fn raf_callback(&self, frame: Frame, time: CrossProcessInstant, can_gc: CanGc)

https://immersive-web.github.io/webxr/#xr-animation-frame

fn update_inline_projection_matrix(&self)

pub(crate) fn inline_view(&self) -> View<Viewer>

Constructs a View suitable for inline sessions using the inlineVerticalFieldOfView and canvas size

pub(crate) fn session_id(&self) -> SessionId

pub(crate) fn dirty_layers(&self)

fn layers_begin_frame(&self, frame: &XRFrame)

fn layers_end_frame(&self, frame: &XRFrame)

fn apply_frame_updates(&self, _frame: &XRFrame)

https://immersive-web.github.io/webxr/#xrframe-apply-frame-updates

fn handle_frame_event(&self, event: FrameUpdateEvent, can_gc: CanGc)

fn apply_nominal_framerate(&self, rate: f32, can_gc: CanGc)

https://www.w3.org/TR/webxr/#apply-the-nominal-frame-rate