Struct script::dom::window::Window

pub struct Window {
    globalscope: GlobalScope,
    script_chan: MainThreadScriptChan,
    dom_manipulation_task_source: DOMManipulationTaskSource,
    user_interaction_task_source: UserInteractionTaskSource,
    networking_task_source: NetworkingTaskSource,
    history_traversal_task_source: HistoryTraversalTaskSource,
    file_reading_task_source: FileReadingTaskSource,
    navigator: MutNullableHeap<JS<Navigator>>,
    image_cache_thread: ImageCacheThread,
    image_cache_chan: ImageCacheChan,
    browsing_context: MutNullableHeap<JS<BrowsingContext>>,
    history: MutNullableHeap<JS<History>>,
    performance: MutNullableHeap<JS<Performance>>,
    navigation_start: u64,
    navigation_start_precise: f64,
    screen: MutNullableHeap<JS<Screen>>,
    session_storage: MutNullableHeap<JS<Storage>>,
    local_storage: MutNullableHeap<JS<Storage>>,
    status: DOMRefCell<DOMString>,
    devtools_markers: DOMRefCell<HashSet<TimelineMarkerType>>,
    devtools_marker_sender: DOMRefCell<Option<IpcSender<Option<TimelineMarker>>>>,
    resize_event: Cell<Option<(WindowSizeData, WindowSizeType)>>,
    parent_info: Option<(PipelineId, FrameType)>,
    dom_static: GlobalStaticData,
    js_runtime: DOMRefCell<Option<Rc<Runtime>>>,
    layout_chan: Sender<Msg>,
    layout_rpc: Box<LayoutRPC + 'static>,
    window_size: Cell<Option<WindowSizeData>>,
    bluetooth_thread: IpcSender<BluetoothMethodMsg>,
    fragment_name: DOMRefCell<Option<String>>,
    page_clip_rect: Cell<Rect<Au>>,
    suppress_reflow: Cell<bool>,
    pending_reflow_count: Cell<u32>,
    webdriver_script_chan: DOMRefCell<Option<IpcSender<WebDriverJSResult>>>,
    current_state: Cell<WindowState>,
    current_viewport: Cell<Rect<Au>>,
    ignore_further_async_events: Arc<AtomicBool>,
    error_reporter: CSSErrorReporter,
    scroll_offsets: DOMRefCell<HashMap<UntrustedNodeAddress, Point2D<f32>>>,
    media_query_lists: WeakMediaQueryListVec,
    test_runner: MutNullableHeap<JS<TestRunner>>,
}

Fields
globalscope: GlobalScope script_chan: MainThreadScriptChan dom_manipulation_task_source: DOMManipulationTaskSource user_interaction_task_source: UserInteractionTaskSource networking_task_source: NetworkingTaskSource history_traversal_task_source: HistoryTraversalTaskSource file_reading_task_source: FileReadingTaskSource navigator: MutNullableHeap<JS<Navigator>> image_cache_thread: ImageCacheThread image_cache_chan: ImageCacheChan browsing_context: MutNullableHeap<JS<BrowsingContext>> history: MutNullableHeap<JS<History>> performance: MutNullableHeap<JS<Performance>> navigation_start: u64 navigation_start_precise: f64 screen: MutNullableHeap<JS<Screen>> session_storage: MutNullableHeap<JS<Storage>> local_storage: MutNullableHeap<JS<Storage>> status: DOMRefCell<DOMString> devtools_markers: DOMRefCell<HashSet<TimelineMarkerType>>

For sending timeline markers. Will be ignored if no devtools server

devtools_marker_sender: DOMRefCell<Option<IpcSender<Option<TimelineMarker>>>> resize_event: Cell<Option<(WindowSizeData, WindowSizeType)>>

Pending resize event, if any.

parent_info: Option<(PipelineId, FrameType)>

Parent id associated with this page, if any.

dom_static: GlobalStaticData

Global static data related to the DOM.

js_runtime: DOMRefCell<Option<Rc<Runtime>>>

The JavaScript runtime.

layout_chan: Sender<Msg>

A handle for communicating messages to the layout thread.

layout_rpc: Box<LayoutRPC + 'static>

A handle to perform RPC calls into the layout, quickly.

window_size: Cell<Option<WindowSizeData>>

The current size of the window, in pixels.

bluetooth_thread: IpcSender<BluetoothMethodMsg>

A handle for communicating messages to the bluetooth thread.

fragment_name: DOMRefCell<Option<String>>

Pending scroll to fragment event, if any

page_clip_rect: Cell<Rect<Au>>

An enlarged rectangle around the page contents visible in the viewport, used to prevent creating display list items for content that is far away from the viewport.

suppress_reflow: Cell<bool>

Flag to suppress reflows. The first reflow will come either with RefreshTick or with FirstLoad. Until those first reflows, we want to suppress others like MissingExplicitReflow.

pending_reflow_count: Cell<u32>

A counter of the number of pending reflows for this window.

webdriver_script_chan: DOMRefCell<Option<IpcSender<WebDriverJSResult>>>

A channel for communicating results of async scripts back to the webdriver server

current_state: Cell<WindowState>

The current state of the window object

current_viewport: Cell<Rect<Au>> ignore_further_async_events: Arc<AtomicBool>

A flag to prevent async events from attempting to interact with this window.

error_reporter: CSSErrorReporter scroll_offsets: DOMRefCell<HashMap<UntrustedNodeAddress, Point2D<f32>>>

A list of scroll offsets for each scrollable element.

media_query_lists: WeakMediaQueryListVec

All the MediaQueryLists we need to update

test_runner: MutNullableHeap<JS<TestRunner>>
Methods

impl Window

fn clear_js_runtime_for_script_deallocation(&self)

fn get_cx(&self) -> *mut JSContext

fn dom_manipulation_task_source(&self) -> DOMManipulationTaskSource

fn user_interaction_task_source(&self) -> UserInteractionTaskSource

fn networking_task_source(&self) -> Box<ScriptChan + Send>

fn history_traversal_task_source(&self) -> Box<ScriptChan + Send>

fn file_reading_task_source(&self) -> FileReadingTaskSource

fn main_thread_script_chan(&self) -> &Sender<MainThreadScriptMsg>

fn image_cache_chan(&self) -> ImageCacheChan

fn parent_info(&self) -> Option<(PipelineId, FrameType)>

fn new_script_pair(&self) -> (Box<ScriptChan + Send>, Box<ScriptPort + Send>)

fn image_cache_thread(&self) -> &ImageCacheThread

fn browsing_context(&self) -> Root<BrowsingContext>

fn bluetooth_thread(&self) -> IpcSender<BluetoothMethodMsg>

fn css_error_reporter(&self) -> Box<ParseErrorReporter + Send>

fn set_scroll_offsets(&self,
offsets: HashMap<UntrustedNodeAddress, Point2D<f32>>)

Sets a new list of scroll offsets.

This is called when layout gives us new ones and WebRender is in use.

fn current_viewport(&self) -> Rect<Au>

impl Window

fn get_runnable_wrapper(&self) -> RunnableWrapper

fn clear_js_runtime(&self)

fn scroll(&self, x_: f64, y_: f64, behavior: ScrollBehavior)

https://drafts.csswg.org/cssom-view/#dom-window-scroll

fn perform_a_scroll(&self,
x: f32,
y: f32,
behavior: ScrollBehavior,
element: Option<&Element>)

https://drafts.csswg.org/cssom-view/#perform-a-scroll

fn update_viewport_for_scroll(&self, x: f32, y: f32)

fn client_window(&self) -> (Size2D<u32>, Point2D<i32>)

fn advance_animation_clock(&self, delta: i32, tick: bool)

Advances the layout animation clock by delta milliseconds, and then forces a reflow if tick is true.

fn force_reflow(&self,
goal: ReflowGoal,
query_type: ReflowQueryType,
reason: ReflowReason)
-> bool

Reflows the page unconditionally if possible and not suppressed. This method will wait for the layout thread to complete (but see the TODO below). If there is no window size yet, the page is presumed invisible and no reflow is performed. If reflow is suppressed, no reflow will be performed for ForDisplay goals.

TODO(pcwalton): Only wait for style recalc, since we have off-main-thread layout.

Returns true if layout actually happened, false otherwise.

fn reflow(&self,
goal: ReflowGoal,
query_type: ReflowQueryType,
reason: ReflowReason)
-> bool

Reflows the page if it's possible to do so and the page is dirty. This method will wait for the layout thread to complete (but see the TODO below). If there is no window size yet, the page is presumed invisible and no reflow is performed.

TODO(pcwalton): Only wait for style recalc, since we have off-main-thread layout.

Returns true if layout actually happened, false otherwise. This return value is useful for script queries, that wait for a lock that layout might hold if the first layout hasn't happened yet (which may happen in the only case a query reflow may bail out, that is, if the viewport size is not present). See #11223 for an example of that.

fn layout(&self) -> &LayoutRPC

fn content_box_query(&self, content_box_request: TrustedNodeAddress) -> Rect<Au>

fn content_boxes_query(&self,
content_boxes_request: TrustedNodeAddress)
-> Vec<Rect<Au>>

fn client_rect_query(&self,
node_geometry_request: TrustedNodeAddress)
-> Rect<i32>

fn hit_test_query(&self,
client_point: Point2D<f32>,
update_cursor: bool)
-> Option<UntrustedNodeAddress>

fn scroll_area_query(&self, node: TrustedNodeAddress) -> Rect<i32>

fn overflow_query(&self, node: TrustedNodeAddress) -> Point2D<T>

fn scroll_offset_query(&self, node: &Node) -> Point2D<f32>

fn scroll_node(&self,
_node: TrustedNodeAddress,
x_: f64,
y_: f64,
behavior: ScrollBehavior)

fn resolved_style_query(&self,
element: TrustedNodeAddress,
pseudo: Option<PseudoElement>,
property: &Atom)
-> Option<DOMString>

fn offset_parent_query(&self,
node: TrustedNodeAddress)
-> (Option<Root<Element>>, Rect<Au>)

fn margin_style_query(&self, node: TrustedNodeAddress) -> MarginStyleResponse

fn init_browsing_context(&self, browsing_context: &BrowsingContext)

fn load_url(&self,
url: Url,
replace: bool,
referrer_policy: Option<ReferrerPolicy>)

Commence a new URL load which will either replace this window or scroll to a fragment.

fn handle_fire_timer(&self, timer_id: TimerEventId)

fn set_fragment_name(&self, fragment: Option<String>)

fn steal_fragment_name(&self) -> Option<String>

fn set_window_size(&self, size: WindowSizeData)

fn window_size(&self) -> Option<WindowSizeData>

fn get_url(&self) -> Url

fn layout_chan(&self) -> &Sender<Msg>

fn windowproxy_handler(&self) -> WindowProxyHandler

fn get_pending_reflow_count(&self) -> u32

fn add_pending_reflow(&self)

fn set_resize_event(&self, event: WindowSizeData, event_type: WindowSizeType)

fn steal_resize_event(&self) -> Option<(WindowSizeData, WindowSizeType)>

fn set_page_clip_rect_with_new_viewport(&self, viewport: Rect<f32>) -> bool

fn IndexedGetter(&self, _index: u32, _found: &mut bool) -> Option<Root<Window>>

fn thaw(&self)

fn need_emit_timeline_marker(&self, timeline_type: TimelineMarkerType) -> bool

fn emit_timeline_marker(&self, marker: TimelineMarker)

fn set_devtools_timeline_markers(&self,
markers: Vec<TimelineMarkerType>,
reply: IpcSender<Option<TimelineMarker>>)

fn drop_devtools_timeline_markers(&self, markers: Vec<TimelineMarkerType>)

fn set_webdriver_script_chan(&self, chan: Option<IpcSender<WebDriverJSResult>>)

fn is_alive(&self) -> bool

fn is_top_level(&self) -> bool

fn parent(&self) -> Option<Root<Window>>

fn is_mozbrowser(&self) -> bool

Returns whether this window is mozbrowser.

unsafe fn global_is_mozbrowser(_: *mut JSContext, obj: HandleObject) -> bool

Returns whether mozbrowser is enabled and obj has been created in a top-level Window global.

fn dispatch_mozbrowser_event(&self, event: MozBrowserEvent)

fn evaluate_media_queries_and_report_changes(&self)

impl Window

fn new(runtime: Rc<Runtime>,
script_chan: MainThreadScriptChan,
dom_task_source: DOMManipulationTaskSource,
user_task_source: UserInteractionTaskSource,
network_task_source: NetworkingTaskSource,
history_task_source: HistoryTraversalTaskSource,
file_task_source: FileReadingTaskSource,
image_cache_chan: ImageCacheChan,
image_cache_thread: ImageCacheThread,
resource_threads: ResourceThreads,
bluetooth_thread: IpcSender<BluetoothMethodMsg>,
mem_profiler_chan: ProfilerChan,
time_profiler_chan: ProfilerChan,
devtools_chan: Option<IpcSender<ScriptToDevtoolsControlMsg>>,
constellation_chan: IpcSender<ConstellationMsg>,
control_chan: IpcSender<ConstellationControlMsg>,
scheduler_chan: IpcSender<TimerEventRequest>,
timer_event_chan: IpcSender<TimerEvent>,
layout_chan: Sender<Msg>,
id: PipelineId,
parent_info: Option<(PipelineId, FrameType)>,
window_size: Option<WindowSizeData>)
-> Root<Window>