Struct script::dom::window::Window

pub struct Window {
    eventtarget: EventTarget,
    script_chan: MainThreadScriptChan,
    dom_manipulation_task_source: DOMManipulationTaskSource,
    user_interaction_task_source: UserInteractionTaskSource,
    networking_task_source: NetworkingTaskSource,
    history_traversal_task_source: HistoryTraversalTaskSource,
    file_reading_task_source: FileReadingTaskSource,
    console: MutNullableHeap<JS<Console>>,
    crypto: MutNullableHeap<JS<Crypto>>,
    navigator: MutNullableHeap<JS<Navigator>>,
    image_cache_thread: ImageCacheThread,
    image_cache_chan: ImageCacheChan,
    custom_message_chan: IpcSender<CustomResponseSender>,
    browsing_context: MutNullableHeap<JS<BrowsingContext>>,
    performance: MutNullableHeap<JS<Performance>>,
    navigation_start: u64,
    navigation_start_precise: f64,
    screen: MutNullableHeap<JS<Screen>>,
    session_storage: MutNullableHeap<JS<Storage>>,
    local_storage: MutNullableHeap<JS<Storage>>,
    status: DOMRefCell<DOMString>,
    scheduler_chan: IpcSender<TimerEventRequest>,
    timers: OneshotTimers,
    next_worker_id: Cell<WorkerId>,
    mem_profiler_chan: ProfilerChan,
    time_profiler_chan: ProfilerChan,
    devtools_chan: Option<IpcSender<ScriptToDevtoolsControlMsg>>,
    devtools_markers: DOMRefCell<HashSet<TimelineMarkerType>>,
    devtools_marker_sender: DOMRefCell<Option<IpcSender<TimelineMarker>>>,
    devtools_wants_updates: Cell<bool>,
    next_subpage_id: Cell<SubpageId>,
    resize_event: Cell<Option<(WindowSizeData, WindowSizeType)>>,
    id: PipelineId,
    parent_info: Option<(PipelineId, SubpageId)>,
    dom_static: GlobalStaticData,
    js_runtime: DOMRefCell<Option<Rc<Runtime>>>,
    layout_chan: Sender<Msg>,
    layout_rpc: Box<LayoutRPC + 'static>,
    window_size: Cell<Option<WindowSizeData>>,
    resource_threads: ResourceThreads,
    bluetooth_thread: IpcSender<BluetoothMethodMsg>,
    constellation_chan: IpcSender<ConstellationMsg>,
    fragment_name: DOMRefCell<Option<String>>,
    page_clip_rect: Cell<Rect<Au>>,
    suppress_reflow: Cell<bool>,
    pending_reflow_count: Cell<u32>,
    webdriver_script_chan: DOMRefCell<Option<IpcSender<WebDriverJSResult>>>,
    current_state: Cell<WindowState>,
    current_viewport: Cell<Rect<Au>>,
    ignore_further_async_events: Arc<AtomicBool>,
    error_reporter: CSSErrorReporter,
    panic_chan: IpcSender<PanicMsg>,
}

Fields

eventtarget

script_chan

dom_manipulation_task_source

user_interaction_task_source

networking_task_source

history_traversal_task_source

file_reading_task_source

console

crypto

navigator

image_cache_thread

image_cache_chan

custom_message_chan

browsing_context

performance

navigation_start

navigation_start_precise

screen

session_storage

local_storage

status

scheduler_chan

timers

next_worker_id

mem_profiler_chan

For sending messages to the memory profiler.

time_profiler_chan

For sending messages to the memory profiler.

devtools_chan

For providing instructions to an optional devtools server.

devtools_markers

For sending timeline markers. Will be ignored if no devtools server

devtools_marker_sender

devtools_wants_updates

A flag to indicate whether the developer tools have requested live updates of page changes.

next_subpage_id

resize_event

Pending resize event, if any.

id

Pipeline id associated with this page.

parent_info

Subpage id associated with this page, if any.

dom_static

Global static data related to the DOM.

js_runtime

The JavaScript runtime.

layout_chan

A handle for communicating messages to the layout thread.

layout_rpc

A handle to perform RPC calls into the layout, quickly.

window_size

The current size of the window, in pixels.

resource_threads

Associated resource threads for use by DOM objects like XMLHttpRequest, including resource_thread, filemanager_thread and storage_thread

bluetooth_thread

A handle for communicating messages to the bluetooth thread.

constellation_chan

A handle for communicating messages to the constellation thread.

fragment_name

Pending scroll to fragment event, if any

page_clip_rect

An enlarged rectangle around the page contents visible in the viewport, used to prevent creating display list items for content that is far away from the viewport.

suppress_reflow

Flag to suppress reflows. The first reflow will come either with RefreshTick or with FirstLoad. Until those first reflows, we want to suppress others like MissingExplicitReflow.

pending_reflow_count

A counter of the number of pending reflows for this window.

webdriver_script_chan

A channel for communicating results of async scripts back to the webdriver server

current_state

The current state of the window object

current_viewport

ignore_further_async_events

A flag to prevent async events from attempting to interact with this window.

error_reporter

panic_chan

Methods

impl Window

fn clear_js_runtime_for_script_deallocation(&self)

fn get_cx(&self) -> *mut JSContext

fn dom_manipulation_task_source(&self) -> DOMManipulationTaskSource

fn user_interaction_task_source(&self) -> UserInteractionTaskSource

fn networking_task_source(&self) -> Box<ScriptChan + Send>

fn history_traversal_task_source(&self) -> Box<ScriptChan + Send>

fn file_reading_task_source(&self) -> Box<ScriptChan + Send>

fn main_thread_script_chan(&self) -> &Sender<MainThreadScriptMsg>

fn image_cache_chan(&self) -> ImageCacheChan

fn custom_message_chan(&self) -> IpcSender<CustomResponseSender>

fn get_next_worker_id(&self) -> WorkerId

fn pipeline(&self) -> PipelineId

fn subpage(&self) -> Option<SubpageId>

fn parent_info(&self) -> Option<(PipelineId, SubpageId)>

fn new_script_pair(&self) -> (Box<ScriptChan + Send>, Box<ScriptPort + Send>)

fn image_cache_thread(&self) -> &ImageCacheThread

fn browsing_context(&self) -> Root<BrowsingContext>

fn bluetooth_thread(&self) -> IpcSender<BluetoothMethodMsg>

fn css_error_reporter(&self) -> Box<ParseErrorReporter + Send>

impl Window

fn get_runnable_wrapper(&self) -> RunnableWrapper

fn clear_js_runtime(&self)

fn scroll(&self, x_: f64, y_: f64, behavior: ScrollBehavior)

https://drafts.csswg.org/cssom-view/#dom-window-scroll

fn perform_a_scroll(&self, x: f32, y: f32, layer_id: LayerId, behavior: ScrollBehavior, element: Option<&Element>)

https://drafts.csswg.org/cssom-view/#perform-a-scroll

fn update_viewport_for_scroll(&self, x: f32, y: f32)

fn client_window(&self) -> (Size2D<u32>, Point2D<i32>)

fn force_reflow(&self, goal: ReflowGoal, query_type: ReflowQueryType, reason: ReflowReason) -> bool

Reflows the page unconditionally if possible and not suppressed. This method will wait for the layout thread to complete (but see the TODO below). If there is no window size yet, the page is presumed invisible and no reflow is performed. If reflow is suppressed, no reflow will be performed for ForDisplay goals.

TODO(pcwalton): Only wait for style recalc, since we have off-main-thread layout.

Returns true if layout actually happened, false otherwise.

fn reflow(&self, goal: ReflowGoal, query_type: ReflowQueryType, reason: ReflowReason) -> bool

Reflows the page if it's possible to do so and the page is dirty. This method will wait for the layout thread to complete (but see the TODO below). If there is no window size yet, the page is presumed invisible and no reflow is performed.

TODO(pcwalton): Only wait for style recalc, since we have off-main-thread layout.

Returns true if layout actually happened, false otherwise. This return value is useful for script queries, that wait for a lock that layout might hold if the first layout hasn't happened yet (which may happen in the only case a query reflow may bail out, that is, if the viewport size is not present). See #11223 for an example of that.

fn layout(&self) -> &LayoutRPC

fn content_box_query(&self, content_box_request: TrustedNodeAddress) -> Rect<Au>

fn content_boxes_query(&self, content_boxes_request: TrustedNodeAddress) -> Vec<Rect<Au>>

fn client_rect_query(&self, node_geometry_request: TrustedNodeAddress) -> Rect<i32>

fn hit_test_query(&self, hit_test_request: Point2D<f32>, update_cursor: bool) -> Option<UntrustedNodeAddress>

fn scroll_area_query(&self, node: TrustedNodeAddress) -> Rect<i32>

fn overflow_query(&self, node: TrustedNodeAddress) -> Point2D<T>

fn scroll_offset_query(&self, node: TrustedNodeAddress) -> Point2D<f32>

fn scroll_node(&self, node: TrustedNodeAddress, x_: f64, y_: f64, behavior: ScrollBehavior)

fn resolved_style_query(&self, element: TrustedNodeAddress, pseudo: Option<PseudoElement>, property: &Atom) -> Option<DOMString>

fn offset_parent_query(&self, node: TrustedNodeAddress) -> (Option<Root<Element>>, Rect<Au>)

fn margin_style_query(&self, node: TrustedNodeAddress) -> MarginStyleResponse

fn init_browsing_context(&self, browsing_context: &BrowsingContext)

fn load_url(&self, url: Url)

Commence a new URL load which will either replace this window or scroll to a fragment.

fn handle_fire_timer(&self, timer_id: TimerEventId)

fn set_fragment_name(&self, fragment: Option<String>)

fn steal_fragment_name(&self) -> Option<String>

fn set_window_size(&self, size: WindowSizeData)

fn window_size(&self) -> Option<WindowSizeData>

fn get_url(&self) -> Url

fn resource_threads(&self) -> &ResourceThreads

fn mem_profiler_chan(&self) -> &ProfilerChan

fn time_profiler_chan(&self) -> &ProfilerChan

fn devtools_chan(&self) -> Option<IpcSender<ScriptToDevtoolsControlMsg>>

fn layout_chan(&self) -> &Sender<Msg>

fn constellation_chan(&self) -> &IpcSender<ConstellationMsg>

fn scheduler_chan(&self) -> &IpcSender<TimerEventRequest>

fn panic_chan(&self) -> &IpcSender<PanicMsg>

fn schedule_callback(&self, callback: OneshotTimerCallback, duration: MsDuration) -> OneshotTimerHandle

fn unschedule_callback(&self, handle: OneshotTimerHandle)

fn windowproxy_handler(&self) -> WindowProxyHandler

fn get_next_subpage_id(&self) -> SubpageId

fn get_pending_reflow_count(&self) -> u32

fn add_pending_reflow(&self)

fn set_resize_event(&self, event: WindowSizeData, event_type: WindowSizeType)

fn steal_resize_event(&self) -> Option<(WindowSizeData, WindowSizeType)>

fn set_page_clip_rect_with_new_viewport(&self, viewport: Rect<f32>) -> bool

fn set_devtools_wants_updates(&self, value: bool)

fn IndexedGetter(&self, _index: u32, _found: &mut bool) -> Option<Root<Window>>

fn thaw(&self)

fn freeze(&self)

fn need_emit_timeline_marker(&self, timeline_type: TimelineMarkerType) -> bool

fn emit_timeline_marker(&self, marker: TimelineMarker)

fn set_devtools_timeline_markers(&self, markers: Vec<TimelineMarkerType>, reply: IpcSender<TimelineMarker>)

fn drop_devtools_timeline_markers(&self, markers: Vec<TimelineMarkerType>)

fn set_webdriver_script_chan(&self, chan: Option<IpcSender<WebDriverJSResult>>)

fn is_alive(&self) -> bool

fn parent(&self) -> Option<Root<Window>>

fn is_mozbrowser(&self) -> bool

Returns whether this window is mozbrowser.

unsafe fn global_is_mozbrowser(_: *mut JSContext, obj: HandleObject) -> bool

Returns whether mozbrowser is enabled and obj has been created in a top-level Window global.

impl Window

fn new(runtime: Rc<Runtime>, script_chan: MainThreadScriptChan, dom_task_source: DOMManipulationTaskSource, user_task_source: UserInteractionTaskSource, network_task_source: NetworkingTaskSource, history_task_source: HistoryTraversalTaskSource, file_task_source: FileReadingTaskSource, image_cache_chan: ImageCacheChan, custom_message_chan: IpcSender<CustomResponseSender>, image_cache_thread: ImageCacheThread, resource_threads: ResourceThreads, bluetooth_thread: IpcSender<BluetoothMethodMsg>, mem_profiler_chan: ProfilerChan, time_profiler_chan: ProfilerChan, devtools_chan: Option<IpcSender<ScriptToDevtoolsControlMsg>>, constellation_chan: IpcSender<ConstellationMsg>, control_chan: IpcSender<ConstellationControlMsg>, scheduler_chan: IpcSender<TimerEventRequest>, panic_chan: IpcSender<PanicMsg>, timer_event_chan: IpcSender<TimerEvent>, layout_chan: Sender<Msg>, id: PipelineId, parent_info: Option<(PipelineId, SubpageId)>, window_size: Option<WindowSizeData>) -> Root<Window>

fn live_devtools_updates(&self) -> bool