Module script::dom::windowproxy

Structs

CreatorBrowsingContextInfo

A browsing context can have a creator browsing context, the browsing context that was responsible for its creation. If a browsing context has a parent browsing context, then that is its creator browsing context. Otherwise, if the browsing context has an opener browsing context, then that is its creator browsing context. Otherwise, the browsing context has no creator browsing context.

WindowProxy

Constants

_IMPL_DOMOBJECT_FOR_WindowProxy

Statics

PROXY_HANDLER

XORIGIN_PROXY_HANDLER

Functions

GetSubframeWindowProxy⚠

defineProperty⚠

defineProperty_xorigin⚠

delete_xorigin⚠

finalize⚠

get⚠

getOwnPropertyDescriptor⚠

getOwnPropertyDescriptor_xorigin⚠

get_prototype_if_ordinary⚠

get_xorigin⚠

has⚠

has_xorigin⚠

new_window_proxy_handler

parse_open_feature_boolean

preventExtensions_xorigin⚠

set⚠

set_xorigin⚠

throw_security_error⚠

tokenize_open_features

trace⚠