Struct script::dom::workerglobalscope::WorkerGlobalScope

pub struct WorkerGlobalScope {
    eventtarget: EventTarget,
    worker_id: WorkerId,
    worker_url: Url,
    closing: Arc<AtomicBool>,
    runtime: Runtime,
    next_worker_id: Cell<WorkerId>,
    core_resource_thread: CoreResourceThread,
    location: MutNullableHeap<JS<WorkerLocation>>,
    navigator: MutNullableHeap<JS<WorkerNavigator>>,
    console: MutNullableHeap<JS<Console>>,
    crypto: MutNullableHeap<JS<Crypto>>,
    timers: OneshotTimers,
    mem_profiler_chan: ProfilerChan,
    time_profiler_chan: ProfilerChan,
    to_devtools_sender: Option<IpcSender<ScriptToDevtoolsControlMsg>>,
    from_devtools_sender: Option<IpcSender<DevtoolScriptControlMsg>>,
    from_devtools_receiver: Receiver<DevtoolScriptControlMsg>,
    devtools_wants_updates: Cell<bool>,
    constellation_chan: IpcSender<ConstellationMsg>,
    scheduler_chan: IpcSender<TimerEventRequest>,
    panic_chan: IpcSender<PanicMsg>,
    custom_msg_chan: IpcSender<CustomResponseSender>,
    custom_msg_port: Receiver<CustomResponseSender>,
}

Fields

eventtarget

worker_id

worker_url

closing

runtime

next_worker_id

core_resource_thread

location

navigator

console

crypto

timers

mem_profiler_chan

time_profiler_chan

to_devtools_sender

from_devtools_sender

Optional IpcSender for sending the DevtoolScriptControlMsg to the server from within the worker

from_devtools_receiver

This Receiver will be ignored later if the corresponding IpcSender doesn't exist

devtools_wants_updates

A flag to indicate whether the developer tools has requested live updates from the worker

constellation_chan

scheduler_chan

panic_chan

custom_msg_chan

custom_msg_port

Methods

impl WorkerGlobalScope

fn new_inherited(init: WorkerGlobalScopeInit, worker_url: Url, runtime: Runtime, from_devtools_receiver: Receiver<DevtoolScriptControlMsg>, timer_event_chan: IpcSender<TimerEvent>) -> WorkerGlobalScope

fn mem_profiler_chan(&self) -> &ProfilerChan

fn time_profiler_chan(&self) -> &ProfilerChan

fn devtools_chan(&self) -> Option<IpcSender<ScriptToDevtoolsControlMsg>>

fn from_devtools_sender(&self) -> Option<IpcSender<DevtoolScriptControlMsg>>

fn from_devtools_receiver(&self) -> &Receiver<DevtoolScriptControlMsg>

fn constellation_chan(&self) -> &IpcSender<ConstellationMsg>

fn scheduler_chan(&self) -> &IpcSender<TimerEventRequest>

fn schedule_callback(&self, callback: OneshotTimerCallback, duration: MsDuration) -> OneshotTimerHandle

fn unschedule_callback(&self, handle: OneshotTimerHandle)

fn runtime(&self) -> *mut JSRuntime

fn get_cx(&self) -> *mut JSContext

fn custom_message_chan(&self) -> IpcSender<CustomResponseSender>

fn custom_message_port(&self) -> &Receiver<CustomResponseSender>

fn is_closing(&self) -> bool

fn core_resource_thread(&self) -> &CoreResourceThread

fn get_url(&self) -> &Url

fn get_worker_id(&self) -> WorkerId

fn get_next_worker_id(&self) -> WorkerId

fn panic_chan(&self) -> &IpcSender<PanicMsg>

impl WorkerGlobalScope

fn execute_script(&self, source: DOMString)

fn script_chan(&self) -> Box<ScriptChan + Send>

fn pipeline(&self) -> PipelineId

fn new_script_pair(&self) -> (Box<ScriptChan + Send>, Box<ScriptPort + Send>)

fn process_event(&self, msg: CommonScriptMsg)

fn handle_fire_timer(&self, timer_id: TimerEventId)

fn set_devtools_wants_updates(&self, value: bool)