script::dom::workers::workerglobalscope

Struct WorkerGlobalScope

#[repr(C)]
pub(crate) struct WorkerGlobalScope {

    globalscope: GlobalScope,
    microtask_queue: Rc<MicrotaskQueue>,
    worker_name: DOMString,
    worker_type: WorkerType,
    worker_id: WorkerId,
    worker_url: DomRefCell<ServoUrl>,
    closing: Arc<AtomicBool>,
    execution_ready: AtomicBool,
    runtime: DomRefCell<Option<Runtime>>,
    location: MutNullableDom<WorkerLocation>,
    navigator: MutNullableDom<WorkerNavigator>,
    policy_container: DomRefCell<PolicyContainer>,
    _devtools_sender: Option<GenericSender<DevtoolScriptControlMsg>>,
    devtools_receiver: Option<RoutedReceiver<DevtoolScriptControlMsg>>,
    navigation_start: CrossProcessInstant,
    performance: MutNullableDom<Performance>,
    indexeddb: MutNullableDom<IDBFactory>,
    trusted_types: MutNullableDom<TrustedTypePolicyFactory>,
    timer_scheduler: RefCell<TimerScheduler>,
    insecure_requests_policy: InsecureRequestsPolicy,
    reporting_observer_list: DomRefCell<Vec<DomRoot<ReportingObserver>>>,
    report_list: DomRefCell<Vec<Report>>,
    endpoints_list: DomRefCell<Vec<ReportingEndpoint>>,

}

Fields
globalscope: GlobalScopemicrotask_queue: Rc<MicrotaskQueue>

https://html.spec.whatwg.org/multipage/#microtask-queue

worker_name: DOMStringworker_type: WorkerTypeworker_id: WorkerIdworker_url: DomRefCell<ServoUrl>closing: Arc<AtomicBool>execution_ready: AtomicBoolruntime: DomRefCell<Option<Runtime>>location: MutNullableDom<WorkerLocation>navigator: MutNullableDom<WorkerNavigator>policy_container: DomRefCell<PolicyContainer>

https://html.spec.whatwg.org/multipage/#the-workerglobalscope-common-interface:policy-container

_devtools_sender: Option<GenericSender<DevtoolScriptControlMsg>>

A Sender for sending messages to devtools. This is unused but is stored here to keep the channel alive.

devtools_receiver: Option<RoutedReceiver<DevtoolScriptControlMsg>>

A Receiver for receiving messages from devtools.

navigation_start: CrossProcessInstantperformance: MutNullableDom<Performance>indexeddb: MutNullableDom<IDBFactory>trusted_types: MutNullableDom<TrustedTypePolicyFactory>timer_scheduler: RefCell<TimerScheduler>

A TimerScheduler used to schedule timers for this WorkerGlobalScope. Timers are handled in the service worker event loop.

insecure_requests_policy: InsecureRequestsPolicyreporting_observer_list: DomRefCell<Vec<DomRoot<ReportingObserver>>>

https://w3c.github.io/reporting/#windoworworkerglobalscope-registered-reporting-observer-list

report_list: DomRefCell<Vec<Report>>

https://w3c.github.io/reporting/#windoworworkerglobalscope-reports

endpoints_list: DomRefCell<Vec<ReportingEndpoint>>

https://w3c.github.io/reporting/#windoworworkerglobalscope-endpoints

Implementations

impl WorkerGlobalScope

pub(crate) fn new_inherited( init: WorkerGlobalScopeInit, worker_name: DOMString, worker_type: WorkerType, worker_url: ServoUrl, runtime: Runtime, devtools_receiver: RoutedReceiver<DevtoolScriptControlMsg>, closing: Arc<AtomicBool>, gpu_id_hub: Arc<IdentityHub>, insecure_requests_policy: InsecureRequestsPolicy, font_context: Option<Arc<FontContext>>, ) -> Self

pub(crate) fn enqueue_microtask(&self, job: Microtask)

pub(crate) fn perform_a_microtask_checkpoint(&self, can_gc: CanGc)

Perform a microtask checkpoint.

pub(crate) fn insecure_requests_policy(&self) -> InsecureRequestsPolicy

Returns a policy value that should be used by fetches initiated by this worker.

pub(crate) fn clear_js_runtime(&self)

Clear various items when the worker event-loop shuts-down.

pub(crate) fn runtime_handle(&self) -> ParentRuntime

pub(crate) fn devtools_receiver( &self, ) -> Option<&RoutedReceiver<DevtoolScriptControlMsg>>

pub(crate) fn get_cx(&self) -> JSContext

pub(crate) fn is_closing(&self) -> bool

pub(crate) fn is_execution_ready(&self) -> bool

pub(crate) fn get_url(&self) -> Ref<'_, ServoUrl>

pub(crate) fn set_url(&self, url: ServoUrl)

pub(crate) fn worker_name(&self) -> DOMString

pub(crate) fn worker_id(&self) -> WorkerId

pub(crate) fn pipeline_id(&self) -> PipelineId

pub(crate) fn policy_container(&self) -> Ref<'_, PolicyContainer>

pub(crate) fn set_csp_list(&self, csp_list: Option<CspList>)

pub(crate) fn set_referrer_policy(&self, referrer_policy: ReferrerPolicy)

pub(crate) fn append_reporting_observer( &self, reporting_observer: DomRoot<ReportingObserver>, )

pub(crate) fn remove_reporting_observer( &self, reporting_observer: &ReportingObserver, )

pub(crate) fn registered_reporting_observers( &self, ) -> Vec<DomRoot<ReportingObserver>>

pub(crate) fn append_report(&self, report: Report)

pub(crate) fn buffered_reports(&self) -> Vec<Report>

pub(crate) fn set_endpoints_list( &self, endpoints: Option<Vec<ReportingEndpoint>>, )

pub(crate) fn timer_scheduler(&self) -> RefMut<'_, TimerScheduler>

Get a mutable reference to the TimerScheduler for this [ServiceWorkerGlobalScope].

pub(crate) fn shared_task_canceller(&self) -> TaskCanceller

Return a copy to the shared task canceller that is used to cancel all tasks when this worker is closing.

fn initialize_policy_container_for_worker_global_scope( &self, metadata: &Metadata, parent_policy_container: &PolicyContainer, )

https://html.spec.whatwg.org/multipage/#initialize-worker-policy-container and https://html.spec.whatwg.org/multipage/#creating-a-policy-container-from-a-fetch-response

fn on_complete( &self, script: Option<ClassicScript>, worker: Trusted<Worker>, cx: &mut JSContext, )

onComplete algorithm defined inside https://html.spec.whatwg.org/multipage/#run-a-worker

impl WorkerGlobalScope

pub(crate) fn new_script_pair( &self, ) -> (ScriptEventLoopSender, ScriptEventLoopReceiver)

pub(crate) fn process_event( &self, msg: CommonScriptMsg, cx: &mut JSContext, ) -> bool

Process a single event as if it were the next event in the queue for this worker event-loop. Returns a boolean indicating whether further events should be processed.

pub(crate) fn close(&self)