Module script::dom::worklet

An implementation of Houdini worklets.

The goal of this implementation is to maximize responsiveness of worklets, and in particular to ensure that the thread performing worklet tasks is never busy GCing or loading worklet code. We do this by providing a custom thread pool implementation, which only performs GC or code loading on a backup thread, not on the primary worklet thread.

Structs

PendingTasksStruct

https://drafts.css-houdini.org/worklets/#pending-tasks-struct

Worklet

https://drafts.css-houdini.org/worklets/#worklet

WorkletExecutor

An executor of worklet tasks

WorkletId

A guid for worklets.

WorkletThread

A thread for executing worklets.

WorkletThreadInit

Data to initialize a worklet thread.

WorkletThreadPool

Worklets execute in a dedicated thread pool.

WorkletThreadRole

A role that a worklet thread can be playing.

Enums

WorkletControl

The control message sent to worklet threads

WorkletData

The data messages sent to worklet threads

Constants

MIN_GC_THRESHOLD

WORKLET_THREAD_POOL_SIZE

_IMPL_DOMOBJECT_FOR_Worklet