Module script::fetch

Structs

FetchCanceller

RAII fetch canceller object. By default initialized to not having a canceller in it, however you can ask it for a cancellation receiver to send to Fetch in which case it will store the sender. You can manually cancel it or let it cancel on Drop in that case.

FetchContext

Functions

Fetch

create_a_potential_cors_request

https://html.spec.whatwg.org/multipage/#create-a-potential-cors-request

fill_headers_with_metadata

load_whole_resource

Convenience function for synchronously loading a whole resource.

request_init_from_request