Struct script::layout_dom::ServoThreadSafeLayoutElement

pub struct ServoThreadSafeLayoutElement<'dom> {
    pub(super) element: ServoLayoutElement<'dom>,
    pub(super) pseudo: PseudoElementType,
}

A wrapper around elements that ensures layout can only ever access safe properties and cannot race on elements.

Fields
element: ServoLayoutElement<'dom>pseudo: PseudoElementType

The pseudo-element type, with (optionally) a specified display value to override the stylesheet.