Struct script::layout_exports::Element

#[repr(C)]
pub struct Element {

    node: Node,
    local_name: LocalName,
    tag_name: TagName,
    namespace: Namespace,
    prefix: DomRefCell<Option<Prefix>>,
    attrs: DomRefCell<Vec<Dom<Attr>>>,
    id_attribute: DomRefCell<Option<Atom>>,
    is: DomRefCell<Option<LocalName>>,
    style_attribute: DomRefCell<Option<Arc<Locked<PropertyDeclarationBlock>>>>,
    attr_list: MutNullableDom<NamedNodeMap>,
    class_list: MutNullableDom<DOMTokenList>,
    state: Cell<ElementState>,
    selector_flags: Cell<ElementSelectorFlags>,
    rare_data: DomRefCell<Option<Box<ElementRareData>>>,

}

Fields
node: Nodelocal_name: LocalNametag_name: TagNamenamespace: Namespaceprefix: DomRefCell<Option<Prefix>>attrs: DomRefCell<Vec<Dom<Attr>>>id_attribute: DomRefCell<Option<Atom>>is: DomRefCell<Option<LocalName>>style_attribute: DomRefCell<Option<Arc<Locked<PropertyDeclarationBlock>>>>attr_list: MutNullableDom<NamedNodeMap>class_list: MutNullableDom<DOMTokenList>state: Cell<ElementState>selector_flags: Cell<ElementSelectorFlags>

These flags are set by the style system to indicate the that certain operations may require restyling this element or its descendants. The flags are not atomic, so the style system takes care of only set them when it has exclusive access to the element.

rare_data: DomRefCell<Option<Box<ElementRareData>>>
Implementations

impl Element

fn __assert_parent_type(&self)

impl Element

fn click_event_filter_by_disabled_state(&self) -> bool

impl Element

pub fn create(
name: QualName,
is: Option<LocalName>,
document: &Document,
creator: ElementCreator,
mode: CustomElementCreationMode
) -> Root<Dom<Element>>

pub fn new_inherited(
local_name: LocalName,
namespace: Namespace,
prefix: Option<Prefix>,
document: &Document
) -> Element

pub fn new_inherited_with_state(
state: ElementState,
local_name: LocalName,
namespace: Namespace,
prefix: Option<Prefix>,
document: &Document
) -> Element

pub fn new(
local_name: LocalName,
namespace: Namespace,
prefix: Option<Prefix>,
document: &Document
) -> Root<Dom<Element>>

fn rare_data(&self) -> Ref<'_, Option<Box<ElementRareData>>>

fn rare_data_mut(&self) -> RefMut<'_, Option<Box<ElementRareData>>>

fn ensure_rare_data(&self) -> RefMut<'_, Box<ElementRareData>>

pub fn restyle(&self, damage: NodeDamage)

pub fn set_is(&self, is: LocalName)

pub fn get_is(&self) -> Option<LocalName>

pub fn set_custom_element_state(&self, state: CustomElementState)

pub fn get_custom_element_state(&self) -> CustomElementState

pub fn set_custom_element_definition(
&self,
definition: Rc<CustomElementDefinition>
)

pub fn get_custom_element_definition(
&self
) -> Option<Rc<CustomElementDefinition>>

pub fn clear_custom_element_definition(&self)

pub fn push_callback_reaction(
&self,
function: Rc<Function>,
args: Box<[Heap<JSVal>]>
)

pub fn push_upgrade_reaction(&self, definition: Rc<CustomElementDefinition>)

pub fn clear_reaction_queue(&self)

pub fn invoke_reactions(&self)

pub fn style(&self) -> Option<Arc<ComputedValues>>

style will be None for elements in a display: none subtree. otherwise, the element has a layout box iff it doesn’t have display: none.

pub fn has_css_layout_box(&self) -> bool

fn potentially_scrollable(&self) -> bool

fn has_scrolling_box(&self) -> bool

fn has_overflow(&self) -> bool

fn has_any_visible_overflow(&self) -> bool

Computed value of overflow-x or overflow-y is “visible”

fn has_any_hidden_overflow(&self) -> bool

Computed value of overflow-x or overflow-y is “hidden”

fn shadow_root(&self) -> Option<Root<Dom<ShadowRoot>>>

pub fn is_shadow_host(&self) -> bool

pub fn attach_shadow(
&self,
is_ua_widget: IsUserAgentWidget
) -> Result<Root<Dom<ShadowRoot>>, Error>

https://dom.spec.whatwg.org/#dom-element-attachshadow XXX This is not exposed to web content yet. It is meant to be used for UA widgets only.

pub fn detach_shadow(&self)

pub fn is_translate_enabled(&self) -> bool

pub fn directionality(&self) -> String

impl Element

pub fn is_html_element(&self) -> bool

pub fn html_element_in_html_document(&self) -> bool

pub fn local_name(&self) -> &LocalName

pub fn parsed_name(&self, name: DOMString) -> LocalName

pub fn namespace(&self) -> &Namespace

pub fn prefix(&self) -> Ref<'_, Option<Prefix>>

pub fn set_prefix(&self, prefix: Option<Prefix>)

pub fn attrs(&self) -> Ref<'_, [Dom<Attr>]>

pub fn locate_namespace(&self, prefix: Option<DOMString>) -> Namespace

pub fn name_attribute(&self) -> Option<Atom>

pub fn style_attribute(
&self
) -> &DomRefCell<Option<Arc<Locked<PropertyDeclarationBlock>>>>

pub fn summarize(&self) -> Vec<AttrInfo>

pub fn is_void(&self) -> bool

pub fn serialize(
&self,
traversal_scope: TraversalScope
) -> Result<DOMString, Error>

pub fn xmlSerialize(
&self,
traversal_scope: XmlTraversalScope
) -> Result<DOMString, Error>

pub fn root_element(&self) -> Root<Dom<Element>>

pub fn lookup_prefix(&self, namespace: Namespace) -> Option<DOMString>

pub fn input_method_type(&self) -> Option<InputMethodType>

pub fn is_focusable_area(&self) -> bool

pub fn is_actually_disabled(&self) -> bool

pub fn push_new_attribute(
&self,
local_name: LocalName,
value: AttrValue,
name: LocalName,
namespace: Namespace,
prefix: Option<Prefix>
)

pub fn push_attribute(&self, attr: &Attr)

pub fn get_attribute(
&self,
namespace: &Namespace,
local_name: &LocalName
) -> Option<Root<Dom<Attr>>>

pub fn get_attribute_by_name(&self, name: DOMString) -> Option<Root<Dom<Attr>>>

pub fn set_attribute_from_parser(
&self,
qname: QualName,
value: DOMString,
prefix: Option<Prefix>
)

pub fn set_attribute(&self, name: &LocalName, value: AttrValue)

pub fn set_custom_attribute(
&self,
name: DOMString,
value: DOMString
) -> Result<(), Error>

fn set_first_matching_attribute<F>(
&self,
local_name: LocalName,
value: AttrValue,
name: LocalName,
namespace: Namespace,
prefix: Option<Prefix>,
find: F
) where
F: Fn(&Attr) -> bool,

pub fn parse_attribute(
&self,
namespace: &Namespace,
local_name: &LocalName,
value: DOMString
) -> AttrValue

pub fn remove_attribute(
&self,
namespace: &Namespace,
local_name: &LocalName
) -> Option<Root<Dom<Attr>>>

pub fn remove_attribute_by_name(
&self,
name: &LocalName
) -> Option<Root<Dom<Attr>>>

fn remove_first_matching_attribute<F>(&self, find: F) -> Option<Root<Dom<Attr>>> where
F: Fn(&Attr) -> bool,

pub fn has_class(&self, name: &Atom, case_sensitivity: CaseSensitivity) -> bool

pub fn set_atomic_attribute(&self, local_name: &LocalName, value: DOMString)

pub fn has_attribute(&self, local_name: &LocalName) -> bool

pub fn set_bool_attribute(&self, local_name: &LocalName, value: bool)

pub fn get_url_attribute(&self, local_name: &LocalName) -> USVString

pub fn set_url_attribute(&self, local_name: &LocalName, value: USVString)

pub fn get_string_attribute(&self, local_name: &LocalName) -> DOMString

pub fn set_string_attribute(&self, local_name: &LocalName, value: DOMString)

pub fn get_tokenlist_attribute(&self, local_name: &LocalName) -> Vec<Atom>

pub fn set_tokenlist_attribute(&self, local_name: &LocalName, value: DOMString)

pub fn set_atomic_tokenlist_attribute(
&self,
local_name: &LocalName,
tokens: Vec<Atom>
)

pub fn get_int_attribute(&self, local_name: &LocalName, default: i32) -> i32

pub fn set_int_attribute(&self, local_name: &LocalName, value: i32)

pub fn get_uint_attribute(&self, local_name: &LocalName, default: u32) -> u32

pub fn set_uint_attribute(&self, local_name: &LocalName, value: u32)

pub fn will_mutate_attr(&self, attr: &Attr)

pub fn insert_adjacent(
&self,
where_: AdjacentPosition,
node: &Node
) -> Result<Option<Root<Dom<Node>>>, Error>

pub fn scroll(&self, x_: f64, y_: f64, behavior: ScrollBehavior)

pub fn parse_fragment(
&self,
markup: DOMString
) -> Result<Root<Dom<DocumentFragment>>, Error>

pub fn fragment_parsing_context(
owner_doc: &Document,
element: Option<&Self>
) -> Root<Dom<Self>>

pub fn fullscreen_element_ready_check(&self) -> bool

pub fn is_in_same_home_subtree<T>(&self, other: &T) -> bool where
T: DerivedFrom<Element> + DomObject,

pub fn get_id(&self) -> Option<Atom>

pub fn get_name(&self) -> Option<Atom>

fn is_sequentially_focusable(&self) -> bool

pub(crate) fn update_sequentially_focusable_status(&self)

impl Element

fn client_rect(&self) -> Rect<i32>

pub fn as_maybe_activatable(&self) -> Option<&dyn Activatable>

pub fn as_stylesheet_owner(&self) -> Option<&dyn StylesheetOwner>

pub fn as_maybe_validatable(&self) -> Option<&dyn Validatable>

pub fn click_in_progress(&self) -> bool

pub fn set_click_in_progress(&self, click: bool)

pub fn nearest_activable_element(&self) -> Option<Root<Dom<Element>>>

pub fn get_lang(&self) -> String

pub fn state(&self) -> ElementState

pub fn set_state(&self, which: ElementState, value: bool)

pub fn set_active_state(&self, value: bool)

https://html.spec.whatwg.org/multipage/#concept-selector-active

pub fn focus_state(&self) -> bool

pub fn set_focus_state(&self, value: bool)

pub fn hover_state(&self) -> bool

pub fn set_hover_state(&self, value: bool)

pub fn enabled_state(&self) -> bool

pub fn set_enabled_state(&self, value: bool)

pub fn disabled_state(&self) -> bool

pub fn set_disabled_state(&self, value: bool)

pub fn read_write_state(&self) -> bool

pub fn set_read_write_state(&self, value: bool)

pub fn placeholder_shown_state(&self) -> bool

pub fn set_placeholder_shown_state(&self, value: bool)

pub fn set_target_state(&self, value: bool)

pub fn set_fullscreen_state(&self, value: bool)

pub fn is_connected(&self) -> bool

https://dom.spec.whatwg.org/#connected

pub fn cannot_navigate(&self) -> bool

impl Element

pub fn check_ancestors_disabled_state_for_form_control(&self)

pub fn check_parent_disabled_state_for_option(&self)

pub fn check_disabled_attribute(&self)