Module script::layout_interface

The high-level interface from script to layout. Using this abstract interface helps reduce coupling between these two components, and enables the DOM to be placed in a separate crate from layout.

Reexports

pub use dom::bindings::inheritance::{CharacterDataTypeId, ElementTypeId};

pub use dom::bindings::inheritance::{HTMLElementTypeId, NodeTypeId};

pub use dom::bindings::js::LayoutJS;

pub use dom::characterdata::LayoutCharacterDataHelpers;

pub use dom::document::{Document, LayoutDocumentHelpers};

pub use dom::element::{Element, LayoutElementHelpers, RawLayoutElementHelpers};

pub use dom::htmlcanvaselement::HTMLCanvasData;

pub use dom::htmlobjectelement::is_image_data;

pub use dom::node::{CAN_BE_FRAGMENTED, HAS_CHANGED, HAS_DIRTY_DESCENDANTS, IS_DIRTY};

pub use dom::node::LayoutNodeHelpers;

pub use dom::node::Node;

pub use dom::node::OpaqueStyleAndLayoutData;

pub use dom::node::TrustedNodeAddress;

pub use dom::text::Text;

Structs

ContentBoxResponse

ContentBoxesResponse

HitTestResponse

MarginStyleResponse

NewLayoutThreadInfo

NodeGeometryResponse

NodeLayerIdResponse

NodeOverflowResponse

OffsetParentResponse

Reflow
Information needed for a reflow.

ResolvedStyleResponse

ScriptReflow
Information needed for a script-initiated reflow.

Enums

Msg
Asynchronous messages that script can send to layout.

ReflowQueryType
Any query to perform with this reflow.

Traits

LayoutRPC
Synchronous messages that script can send to layout.