Module script::layout_interface

The high-level interface from script to layout. Using this abstract interface helps reduce coupling between these two components, and enables the DOM to be placed in a separate crate from layout.

Reexports

pub use dom::node::TrustedNodeAddress;

Structs

ContentBoxResponse

ContentBoxesResponse

HitTestResponse

MarginStyleResponse

NewLayoutThreadInfo

NodeGeometryResponse

NodeLayerIdResponse

NodeOverflowResponse

OffsetParentResponse

Reflow
Information needed for a reflow.

ResolvedStyleResponse

ScriptReflow
Information needed for a script-initiated reflow.

Enums

Msg
Asynchronous messages that script can send to layout.

ReflowQueryType
Any query to perform with this reflow.

Traits

LayoutRPC
Synchronous messages that script can send to layout.