Enum script::layout_interface::Msg

pub enum Msg {
    AddStylesheet(Arc<Stylesheet>),
    SetQuirksMode,
    Reflow(ScriptReflow),
    GetRPC(Sender<Box<LayoutRPC + Send>>),
    TickAnimations,
    ReflowWithNewlyLoadedWebFont,
    SetVisibleRects(Vec<(LayerId, Rect<Au>)>),
    ReapStyleAndLayoutData(OpaqueStyleAndLayoutData),
    CollectReports(ReportsChan),
    PrepareToExit(Sender<()>),
    ExitNow,
    GetCurrentEpoch(IpcSender<Epoch>),
    GetWebFontLoadState(IpcSender<bool>),
    CreateLayoutThread(NewLayoutThreadInfo),
    SetFinalUrl(Url),
    SetStackingContextScrollStates(Vec<StackingContextScrollState>),
}

Asynchronous messages that script can send to layout.

Variants

AddStylesheet

Adds the given stylesheet to the document.

SetQuirksMode

Puts a document into quirks mode, causing the quirks mode stylesheet to be loaded.

Reflow

Requests a reflow.

GetRPC

Get an RPC interface.

TickAnimations

Requests that the layout thread render the next frame of all animations.

ReflowWithNewlyLoadedWebFont

Requests that the layout thread reflow with a newly-loaded Web font.

SetVisibleRects

Updates the layout visible rects, affecting the area that display lists will be constructed for.

ReapStyleAndLayoutData

Destroys layout data associated with a DOM node.

TODO(pcwalton): Maybe think about batching to avoid message traffic.

CollectReports

Requests that the layout thread measure its memory usage. The resulting reports are sent back via the supplied channel.

PrepareToExit

Requests that the layout thread enter a quiescent state in which no more messages are accepted except ExitMsg. A response message will be sent on the supplied channel when this happens.

ExitNow

Requests that the layout thread immediately shut down. There must be no more nodes left after this, or layout will crash.

GetCurrentEpoch

Get the last epoch counter for this layout thread.

GetWebFontLoadState

Asks the layout thread whether any Web fonts have yet to load (if true, loads are pending; false otherwise).

CreateLayoutThread

Creates a new layout thread.

This basically exists to keep the script-layout dependency one-way.

SetFinalUrl

Set the final Url.

SetStackingContextScrollStates

Tells layout about the new scrolling offsets of each scrollable stacking context.