Struct script::layout_interface::Reflow

pub struct Reflow {
    pub goal: ReflowGoal,
    pub url: Url,
    pub iframe: bool,
    pub page_clip_rect: Rect<Au>,
}

Information needed for a reflow.

Fields

goal

The goal of reflow: either to render to the screen or to flush layout info for script.

url

The URL of the page.

iframe

Is the current reflow of an iframe, as opposed to a root window?

page_clip_rect

A clipping rectangle for the page, an enlarged rectangle containing the viewport.