Struct script::layout_interface::ScriptReflow

pub struct ScriptReflow {
    pub reflow_info: Reflow,
    pub document: TrustedNodeAddress,
    pub document_stylesheets: Vec<Arc<Stylesheet>>,
    pub stylesheets_changed: bool,
    pub window_size: WindowSizeData,
    pub script_join_chan: Sender<()>,
    pub query_type: ReflowQueryType,
}

Information needed for a script-initiated reflow.

Fields

reflow_info

General reflow data.

document

The document node.

document_stylesheets

The document's list of stylesheets.

stylesheets_changed

Whether the document's stylesheets have changed since the last script reflow.

window_size

The current window size.

script_join_chan

The channel that we send a notification to.

query_type

The type of query if any to perform during this reflow.