Trait script::layout_interface::LayoutRPC

pub trait LayoutRPC {
    fn content_box(&self) -> ContentBoxResponse;
    fn content_boxes(&self) -> ContentBoxesResponse;
    fn node_geometry(&self) -> NodeGeometryResponse;
    fn node_overflow(&self) -> NodeOverflowResponse;
    fn node_scroll_area(&self) -> NodeGeometryResponse;
    fn node_layer_id(&self) -> NodeLayerIdResponse;
    fn hit_test(&self) -> HitTestResponse;
    fn resolved_style(&self) -> ResolvedStyleResponse;
    fn offset_parent(&self) -> OffsetParentResponse;
    fn margin_style(&self) -> MarginStyleResponse;
    fn nodes_from_point(&self, point: Point2D<f32>) -> Vec<UntrustedNodeAddress>;
}

Synchronous messages that script can send to layout.

In general, you should use messages to talk to Layout. Use the RPC interface if and only if the work is

1) read-only with respect to LayoutThreadData, 2) small, 3) and really needs to be fast.

Required Methods

fn content_box(&self) -> ContentBoxResponse

Requests the dimensions of the content box, as in the getBoundingClientRect() call.

fn content_boxes(&self) -> ContentBoxesResponse

Requests the dimensions of all the content boxes, as in the getClientRects() call.

fn node_geometry(&self) -> NodeGeometryResponse

Requests the geometry of this node. Used by APIs such as clientTop.

fn node_overflow(&self) -> NodeOverflowResponse

Requests the overflow-x and overflow-y of this node. Used by scrollTop etc.

fn node_scroll_area(&self) -> NodeGeometryResponse

Requests the scroll geometry of this node. Used by APIs such as scrollTop.

fn node_layer_id(&self) -> NodeLayerIdResponse

Requests the layer id of this node. Used by APIs such as scrollTop

fn hit_test(&self) -> HitTestResponse

Requests the node containing the point of interest

fn resolved_style(&self) -> ResolvedStyleResponse

Query layout for the resolved value of a given CSS property

fn offset_parent(&self) -> OffsetParentResponse

fn margin_style(&self) -> MarginStyleResponse

Query layout for the resolve values of the margin properties for an element.

fn nodes_from_point(&self, point: Point2D<f32>) -> Vec<UntrustedNodeAddress>