Module script::microtask

Implementation of microtasks and microtask queues. It is up to implementations of event loops to store a queue and perform checkpoints at appropriate times, as well as enqueue microtasks as required.

Structs

EnqueuedPromiseCallback

A promise callback scheduled to run during the next microtask checkpoint (#4283).

MicrotaskQueue

A collection of microtasks in FIFO order.

UserMicrotask

A microtask that comes from a queueMicrotask() Javascript call, identical to EnqueuedPromiseCallback once it’s on the queue

Enums

Microtask

Traits

MicrotaskRunnable