script

Module network_listener

Structs

NetworkListener 🔒

An off-thread sink for async network event tasks. All such events are forwarded to a target thread, where they are invoked on the provided context object.

Traits

FetchResponseListener 🔒

ResourceTimingListener 🔒

Functions

submit_timing 🔒

submit_timing_data 🔒