Struct script::network_listener::NetworkListener

pub struct NetworkListener<T: AsyncResponseListener + PreInvoke + Send + 'static> {
    pub context: Arc<Mutex<T>>,
    pub script_chan: Box<ScriptChan + Send>,
}

An off-thread sink for async network event runnables. All such events are forwarded to a target thread, where they are invoked on the provided context object.

Fields

context

script_chan

Methods

impl<T: AsyncResponseListener + PreInvoke + Send + 'static> NetworkListener<T>

fn notify(&self, action: ResponseAction)