Trait script::network_listener::PreInvoke

pub trait PreInvoke {
    fn should_invoke(&self) -> bool { ... }
}

A gating mechanism that runs before invoking the task on the target thread. If the should_invoke method returns false, the task is discarded without being invoked.

Provided methods

fn should_invoke(&self) -> bool