script::python
Module error

Macros

ensure

Ensure a condition is true. If it is not, return from the function with an error.

Structs

ArrowSnafu 🔒

SNAFU context selector for the Error::Arrow variant

CoprParseSnafu 🔒

SNAFU context selector for the Error::CoprParse variant

DataFusionSnafu 🔒

SNAFU context selector for the Error::DataFusion variant

DatabaseQuerySnafu 🔒

SNAFU context selector for the Error::DatabaseQuery variant

NewRecordBatchSnafu 🔒

SNAFU context selector for the Error::NewRecordBatch variant

OtherSnafu 🔒

SNAFU context selector for the Error::Other variant

PyCompileSnafu 🔒

SNAFU context selector for the Error::PyCompile variant

PyParseSnafu 🔒

SNAFU context selector for the Error::PyParse variant

PyRuntimeSnafu 🔒

SNAFU context selector for the Error::PyRuntime variant

RecordBatchSnafu 🔒

SNAFU context selector for the Error::RecordBatch variant

TokioJoinSnafu 🔒

SNAFU context selector for the Error::TokioJoin variant

TypeCastSnafu 🔒

SNAFU context selector for the Error::TypeCast variant

UnsupportedSqlSnafu 🔒

SNAFU context selector for the Error::UnsupportedSql variant

Enums

Error

Functions

get_error_reason_loc

extract a reason for Error in string format, also return a location if possible

pretty_print_error_in_src

pretty print Error in given script, basically print a arrow which point to where error occurs(if possible to get a location)

ret_other_error_with 🔒

visualize_loc

pretty print a location in script with desc.

Type Aliases

Result