script::python::ffi_types
Module vector

Structs

PyVector

The Main FFI type PyVector that is used both in RustPython and PyO3

Functions

arrow_rfloordiv 🔒

Performs val / arr, but cast to i64.

arrow_rsub 🔒

Performs val - arr.

arrow_rtruediv 🔒

Performs val / arr

cast 🔒

get_arrow_op 🔒

get corresponding arrow op function according to given PyComaprsionOp

get_arrow_scalar_op 🔒

get corresponding arrow scalar op function according to given PyComaprsionOp

is_float 🔒

is_signed 🔒

is_unsigned 🔒

rspy_is_pyobj_scalar 🔒

if this pyobj can be cast to a scalar value(i.e Null/Int/Float/Bool)

to_type_error 🔒

val_to_pyobj

convert a DataType Value into a PyObjectRef

wrap_result 🔒

Type Aliases

PyVectorRef 🔒