Module script::python::rspython::copr_impl

Constants

INTERPRETER 🔒

Functions

exec_with_cached_vm 🔒

init_interpreter 🔒

init interpreter with type PyVector and Module: greptime

rspy_exec_parsed 🔒

Using RustPython to run a parsed Coprocessor struct as input to execute python code

set_items_in_scope 🔒

set arguments with given name and values in python scopes

set_query_engine_in_scope 🔒

try_into_columns 🔒

convert a tuple of PyVector or one PyVector(wrapped in a Python Object RefPyObjectRef) to a Vec<VectorRef> by default, a constant(int/float/bool) gives the a constant array of same length with input args