Module script::python::rspython::vector_impl

Functions

get_concrete_type 🔒

pyobj_try_to_typed_val 🔒

convert a PyObjectRef into a datatypes::Value(is that ok?) if obj can be convert to given ConcreteDataType then return inner Value else return None if dtype is None, return types with highest precision Not used for now but may be use in future

to_type_error 🔒

PyVectors’ rustpython specify methods

Type Aliases

PyVectorRef 🔒