script

Module script_module

The script module mod contains common traits and structs related to type=module for script thread or worker threads.

Structs

DynamicModule 🔒

DynamicModuleList 🔒

ImportMap 🔒

https://html.spec.whatwg.org/multipage/#import-map-processing-model

ModuleContext 🔒

The context required for asynchronously loading an external module script source.

ModuleHandler 🔒

ModuleObject 🔒

ModuleScript 🔒

ModuleSource 🔒

ModuleTree 🔒

ResolvedModule 🔒

https://html.spec.whatwg.org/multipage/#specifier-resolution-record

RethrowError 🔒

ScriptFetchOptions 🔒

https://html.spec.whatwg.org/multipage/#script-fetch-options

Enums

ModuleIdentity 🔒

Identity for a module which will be used to retrieve the module when we’d like to get it from module map.

ModuleOwner 🔒

The owner of the module It can be worker or script element

ModuleStatus 🔒

Functions

EnsureModuleHooksInitialized 🔒 ⚠

A function to register module hooks (e.g. listening on resolving modules, getting module metadata, getting script private reference and resolving dynamic import)

HostPopulateImportMeta 🔒 ⚠

https://tc39.es/ecma262/#sec-hostgetimportmetaproperties https://html.spec.whatwg.org/multipage/#hostgetimportmetaproperties

HostResolveImportedModule 🔒 ⚠

https://tc39.es/ecma262/#sec-HostLoadImportedModule https://html.spec.whatwg.org/multipage/#hostloadimportedmodule

fetch_an_import_module_script_graph 🔒

https://html.spec.whatwg.org/multipage/#fetch-an-import()-module-script-graph

fetch_external_module_script 🔒

https://html.spec.whatwg.org/multipage/#fetch-a-module-script-tree

fetch_inline_module_script 🔒

https://html.spec.whatwg.org/multipage/#fetch-an-inline-module-script-graph

fetch_single_module_script 🔒

https://html.spec.whatwg.org/multipage/#fetch-a-single-module-script

gen_type_error 🔒

host_add_ref_top_level_script 🔒 ⚠

host_import_module_dynamically 🔒 ⚠

https://tc39.es/ecma262/#sec-hostimportmoduledynamically https://html.spec.whatwg.org/multipage/#hostimportmoduledynamically(referencingscriptormodule,-specifier,-promisecapability)

host_release_top_level_script 🔒 ⚠

merge_existing_and_new_import_maps 🔒

https://html.spec.whatwg.org/multipage/#merge-existing-and-new-import-maps

merge_module_specifier_maps 🔒

https://html.spec.whatwg.org/multipage/#merge-module-specifier-maps

module_script_from_reference_private 🔒 ⚠

normalize_module_integrity_map 🔒

https://html.spec.whatwg.org/multipage/#normalizing-a-module-integrity-map

normalize_specifier_key 🔒

https://html.spec.whatwg.org/multipage/#normalizing-a-specifier-key

parse_an_import_map_string 🔒

https://html.spec.whatwg.org/multipage/#parse-an-import-map-string

register_import_map 🔒

https://html.spec.whatwg.org/multipage/#register-an-import-map

resolve_imports_match 🔒

https://html.spec.whatwg.org/multipage/#resolving-an-imports-match

sort_and_normalize_module_specifier_map 🔒

https://html.spec.whatwg.org/multipage/#sorting-and-normalizing-a-module-specifier-map

sort_and_normalize_scopes 🔒

https://html.spec.whatwg.org/multipage/#sorting-and-normalizing-scopes

Type Aliases

ModuleIntegrityMap 🔒

ModuleSpecifierMap 🔒