Module script::script_module

The script module mod contains common traits and structs related to type=module for script thread or worker threads.

Structs

DynamicModule

DynamicModuleList

ModuleContext

The context required for asynchronously loading an external module script source.

ModuleHandler

ModuleObject

ModuleScript

ModuleTree

RethrowError

ScriptFetchOptions

https://html.spec.whatwg.org/multipage/#script-fetch-options

Enums

ModuleIdentity

Identity for a module which will be used to retrieve the module when we’d like to get it from module map.

ModuleOwner

The owner of the module It can be worker or script element

ModuleStatus

Functions

EnsureModuleHooksInitialized⚠

A function to register module hooks (e.g. listening on resolving modules, getting module metadata, getting script private reference and resolving dynamic import)

HostPopulateImportMeta⚠

https://tc39.es/ecma262/#sec-hostgetimportmetaproperties https://html.spec.whatwg.org/multipage/#hostgetimportmetaproperties

HostResolveImportedModule⚠

https://tc39.github.io/ecma262/#sec-hostresolveimportedmodule https://html.spec.whatwg.org/multipage/#hostresolveimportedmodule(referencingscriptormodule%2C-specifier)

fetch_an_import_module_script_graph

https://html.spec.whatwg.org/multipage/#fetch-an-import()-module-script-graph

fetch_external_module_script

https://html.spec.whatwg.org/multipage/#fetch-a-module-script-tree

fetch_inline_module_script

https://html.spec.whatwg.org/multipage/#fetch-an-inline-module-script-graph

fetch_single_module_script

https://html.spec.whatwg.org/multipage/#fetch-a-single-module-script

gen_type_error⚠

host_add_ref_top_level_script⚠

host_import_module_dynamically⚠

https://html.spec.whatwg.org/multipage/#hostimportmoduledynamically(referencingscriptormodule,-specifier,-promisecapability)

host_release_top_level_script⚠

module_script_from_reference_private⚠