script::script_module

Struct ModuleTree

pub(crate) struct ModuleTree {
    url: ServoUrl,
    text: DomRefCell<Rc<DOMString>>,
    record: DomRefCell<Option<ModuleObject>>,
    status: DomRefCell<ModuleStatus>,
    parent_identities: DomRefCell<IndexSet<ModuleIdentity>>,
    descendant_urls: DomRefCell<IndexSet<ServoUrl>>,
    incomplete_fetch_urls: DomRefCell<IndexSet<ServoUrl>>,
    visited_urls: DomRefCell<HashSet<ServoUrl>>,
    rethrow_error: DomRefCell<Option<RethrowError>>,
    network_error: DomRefCell<Option<NetworkError>>,
    promise: DomRefCell<Option<Rc<Promise>>>,
    external: bool,
}

Fields
url: ServoUrltext: DomRefCell<Rc<DOMString>>record: DomRefCell<Option<ModuleObject>>status: DomRefCell<ModuleStatus>parent_identities: DomRefCell<IndexSet<ModuleIdentity>>descendant_urls: DomRefCell<IndexSet<ServoUrl>>incomplete_fetch_urls: DomRefCell<IndexSet<ServoUrl>>visited_urls: DomRefCell<HashSet<ServoUrl>>rethrow_error: DomRefCell<Option<RethrowError>>network_error: DomRefCell<Option<NetworkError>>promise: DomRefCell<Option<Rc<Promise>>>external: bool
Implementations

impl ModuleTree

pub(crate) fn new( url: ServoUrl, external: bool, visited_urls: HashSet<ServoUrl>, ) -> Self

pub(crate) fn get_status(&self) -> ModuleStatus

pub(crate) fn set_status(&self, status: ModuleStatus)

pub(crate) fn get_record(&self) -> &DomRefCell<Option<ModuleObject>>

pub(crate) fn set_record(&self, record: ModuleObject)

pub(crate) fn get_rethrow_error(&self) -> &DomRefCell<Option<RethrowError>>

pub(crate) fn set_rethrow_error(&self, rethrow_error: RethrowError)

pub(crate) fn get_network_error(&self) -> &DomRefCell<Option<NetworkError>>

pub(crate) fn set_network_error(&self, network_error: NetworkError)

pub(crate) fn get_text(&self) -> &DomRefCell<Rc<DOMString>>

pub(crate) fn set_text(&self, module_text: Rc<DOMString>)

pub(crate) fn get_incomplete_fetch_urls( &self, ) -> &DomRefCell<IndexSet<ServoUrl>>

pub(crate) fn get_descendant_urls(&self) -> &DomRefCell<IndexSet<ServoUrl>>

pub(crate) fn get_parent_urls(&self) -> IndexSet<ServoUrl>

pub(crate) fn insert_parent_identity(&self, parent_identity: ModuleIdentity)

pub(crate) fn insert_incomplete_fetch_url(&self, dependency: &ServoUrl)

pub(crate) fn remove_incomplete_fetch_url(&self, dependency: &ServoUrl)

fn recursive_check_descendants( module_tree: &ModuleTree, module_map: &HashMap<ServoUrl, Rc<ModuleTree>>, discovered_urls: &mut HashSet<ServoUrl>, ) -> bool

recursively checks if all of the transitive descendants are in the FetchingDescendants or later status

fn has_all_ready_descendants(&self, global: &GlobalScope) -> bool

fn append_handler( &self, owner: ModuleOwner, module_identity: ModuleIdentity, can_gc: CanGc, )

fn append_dynamic_module_handler( &self, owner: ModuleOwner, module_identity: ModuleIdentity, dynamic_module: RootedTraceableBox<DynamicModule>, can_gc: CanGc, )

impl ModuleTree

fn compile_module_script( &self, global: &GlobalScope, owner: ModuleOwner, module_script_text: Rc<DOMString>, url: &ServoUrl, options: ScriptFetchOptions, module_script: RustMutableHandleObject<'_>, inline: bool, line_number: u64, introduction_type: Option<&'static CStr>, can_gc: CanGc, ) -> Result<(), RethrowError>

https://html.spec.whatwg.org/multipage/#creating-a-module-script Step 7-11. Although the CanGc argument appears unused, it represents the GC operations that can occur as part of compiling a script.

pub(crate) fn instantiate_module_tree( &self, global: &GlobalScope, module_record: HandleObject, ) -> Result<(), RethrowError>

https://html.spec.whatwg.org/multipage/#fetch-the-descendants-of-and-link-a-module-script Step 5-2.

pub(crate) fn execute_module( &self, global: &GlobalScope, module_record: HandleObject, eval_result: MutableHandleValue, _can_gc: CanGc, ) -> Result<(), RethrowError>

Execute the provided module, storing the evaluation return value in the provided mutable handle. Although the CanGc appears unused, it represents the GC operations possible when evluating arbitrary JS.

pub(crate) fn report_error(&self, global: &GlobalScope, can_gc: CanGc)

fn resolve_requested_module_specifiers( &self, global: &GlobalScope, module_object: HandleObject, can_gc: CanGc, ) -> Result<IndexSet<ServoUrl>, RethrowError>

fn resolve_module_specifier( global: &GlobalScope, script: Option<&ModuleScript>, specifier: DOMString, can_gc: CanGc, ) -> Fallible<ServoUrl>

https://html.spec.whatwg.org/multipage/#resolve-a-module-specifier

fn resolve_url_like_module_specifier( specifier: &DOMString, base_url: &ServoUrl, ) -> Option<ServoUrl>

https://html.spec.whatwg.org/multipage/#resolving-a-url-like-module-specifier

fn find_first_parse_error( &self, global: &GlobalScope, discovered_urls: &mut HashSet<ServoUrl>, ) -> (Option<NetworkError>, Option<RethrowError>)

https://html.spec.whatwg.org/multipage/#finding-the-first-parse-error

fn fetch_module_descendants( &self, owner: &ModuleOwner, destination: Destination, options: &ScriptFetchOptions, parent_identity: ModuleIdentity, can_gc: CanGc, )

https://html.spec.whatwg.org/multipage/#fetch-the-descendants-of-a-module-script

fn advance_finished_and_link(&self, global: &GlobalScope, can_gc: CanGc)

https://html.spec.whatwg.org/multipage/#fetch-the-descendants-of-and-link-a-module-script step 4-7.