Struct script::script_module::ModuleTree

pub struct ModuleTree {
    url: ServoUrl,
    text: DomRefCell<DOMString>,
    record: DomRefCell<Option<ModuleObject>>,
    status: DomRefCell<ModuleStatus>,
    parent_urls: DomRefCell<IndexSet<ServoUrl>>,
    descendant_urls: DomRefCell<IndexSet<ServoUrl>>,
    visited_urls: DomRefCell<HashSet<ServoUrl>>,
    error: DomRefCell<Option<ModuleError>>,
    promise: DomRefCell<Option<Rc<Promise>>>,
}

Fields
url: ServoUrltext: DomRefCell<DOMString>record: DomRefCell<Option<ModuleObject>>status: DomRefCell<ModuleStatus>parent_urls: DomRefCell<IndexSet<ServoUrl>>descendant_urls: DomRefCell<IndexSet<ServoUrl>>visited_urls: DomRefCell<HashSet<ServoUrl>>error: DomRefCell<Option<ModuleError>>promise: DomRefCell<Option<Rc<Promise>>>
Methods

impl ModuleTree

pub fn new(url: ServoUrl) -> Self

pub fn get_promise(&self) -> &DomRefCell<Option<Rc<Promise>>>

pub fn set_promise(&self, promise: Rc<Promise>)

pub fn get_status(&self) -> ModuleStatus

pub fn set_status(&self, status: ModuleStatus)

pub fn get_record(&self) -> &DomRefCell<Option<ModuleObject>>

pub fn set_record(&self, record: ModuleObject)

pub fn get_error(&self) -> &DomRefCell<Option<ModuleError>>

pub fn set_error(&self, error: Option<ModuleError>)

pub fn get_text(&self) -> &DomRefCell<DOMString>

pub fn set_text(&self, module_text: DOMString)

pub fn get_parent_urls(&self) -> &DomRefCell<IndexSet<ServoUrl>>

pub fn insert_parent_url(&self, parent_url: ServoUrl)

pub fn append_parent_urls(&self, parent_urls: IndexSet<ServoUrl>)

pub fn get_descendant_urls(&self) -> &DomRefCell<IndexSet<ServoUrl>>

pub fn append_descendant_urls(&self, descendant_urls: IndexSet<ServoUrl>)

fn recursive_check_descendants(
module_tree: &ModuleTree,
module_map: &HashMap<ServoUrl, Rc<ModuleTree>>,
discovered_urls: &mut HashSet<ServoUrl>
) -> bool

recursively checks if all of the transitive descendants are in the FetchingDescendants or later status

fn has_all_ready_descendants(
&self,
module_map: &HashMap<ServoUrl, Rc<ModuleTree>>
) -> bool

pub fn get_visited_urls(&self) -> &DomRefCell<HashSet<ServoUrl>>

pub fn append_handler(
&self,
owner: ModuleOwner,
module_url: ServoUrl,
is_top_level: bool
)

impl ModuleTree

fn compile_module_script(
&self,
global: &GlobalScope,
module_script_text: DOMString,
url: ServoUrl
) -> Result<ModuleObject, ModuleError>

https://html.spec.whatwg.org/multipage/#creating-a-module-script Step 7-11.

pub fn instantiate_module_tree(
&self,
global: &GlobalScope,
module_record: HandleObject
) -> Result<(), ModuleError>

pub fn execute_module(
&self,
global: &GlobalScope,
module_record: HandleObject
) -> Result<(), ModuleError>

pub fn report_error(&self, global: &GlobalScope)

pub fn resolve_requested_modules(
&self,
global: &GlobalScope
) -> Result<IndexSet<ServoUrl>, ModuleError>

https://html.spec.whatwg.org/multipage/#fetch-the-descendants-of-a-module-script Step 5.

fn resolve_requested_module_specifiers(
&self,
global: &GlobalScope,
module_object: HandleObject,
base_url: ServoUrl
) -> Result<IndexSet<ServoUrl>, ModuleError>

fn resolve_module_specifier(
cx: *mut JSContext,
url: &ServoUrl,
specifier: RawHandle<*mut JSString>
) -> Result<ServoUrl, UrlParseError>

The following module specifiers are allowed by the spec:

a valid absolute URL

a valid relative URL that starts with "/", "./" or "../"

Bareword module specifiers are currently disallowed as these may be given special meanings in the future. https://html.spec.whatwg.org/multipage/#resolve-a-module-specifier

fn find_first_parse_error(
global: &GlobalScope,
module_tree: &ModuleTree,
discovered_urls: &mut HashSet<ServoUrl>
) -> Option<ModuleError>

https://html.spec.whatwg.org/multipage/#finding-the-first-parse-error