Module script::script_runtime

The script runtime contains common traits and structs commonly used by the script thread, the dom, and the worker threads.

Structs

StackRootTLS

Enums

CommonScriptMsg
Common messages used to control the event loops in both the script and the worker

ScriptThreadEventCategory

Statics

GC_CYCLE_START

GC_SLICE_START

Traits

ScriptChan
A cloneable interface for communicating with an event loop.

ScriptPort
An interface for receiving ScriptMsg values in an event loop. Used for synchronous DOM APIs that need to abstract over multiple kinds of event loops (worker/main thread) with different Receiver interfaces.

Functions

debug_gc_callback

gc_slice_callback

get_reports

new_rt_and_cx

set_gc_zeal_options

trace_rust_roots