script

Module script_runtime

The script runtime contains common traits and structs commonly used by the script thread, the dom, and the worker threads.

Structs

CanGc 🔒

A compile-time marker that there are operations that could trigger a JS garbage collection operation within the current stack frame. It is trivially copyable, so it should be passed as a function argument and reused when calling other functions whenever possible. Since it is only meaningful within the current stack frame, it is impossible to move it to a different thread or into a task that will execute asynchronously.

IntroductionType 🔒

introductionType values in SpiderMonkey TransitiveCompileOptions.

JSContext 🔒

JSEngineSetup

Runnable 🔒

Runtime 🔒

StreamConsumer 🔒

ThreadSafeJSContext 🔒

A version of the JSContext that can be used from other threads and is thus Send and Sync. This should only ever expose operations that are marked as thread-safe by the SpiderMonkey API, ie ones that only atomic fields in JSContext.

Enums

ScriptThreadEventCategory 🔒

Constants

GC_CYCLE_START 🔒

GC_SLICE_START 🔒

HOST_DEFINED_DATA_SLOTS 🔒

INCUMBENT_SETTING_SLOT 🔒

MALLOC_SIZE_OF_OPS 🔒

Statics

HOST_DEFINED_DATA 🔒

HOST_DEFINED_DATA_CLASS 🔒

JOB_QUEUE_TRAPS 🔒

JS_ENGINE 🔒

SECURITY_CALLBACKS 🔒

Traits

JSContextHelper 🔒

Extra methods for the JSContext type defined in script_bindings, when the methods are only called by code in the script crate.

Functions

code_for_eval_gets 🔒 ⚠

consume_stream 🔒 ⚠

Implements the steps to compile webassembly response mentioned here https://webassembly.github.io/spec/web-api/#compile-a-potential-webassembly-response

content_security_policy_allows 🔒 ⚠

debug_gc_callback 🔒 ⚠

drop_interrupt_queues 🔒 ⚠

empty 🔒 ⚠

enqueue_promise_job 🔒 ⚠

https://searchfox.org/mozilla-central/rev/2a8a30f4c9b918b726891ab9d2d62b76152606f1/xpcom/base/CycleCollectedJSContext.cpp#355 SM callback for promise job resolution. Adds a promise callback to the current global’s microtask queue.

gc_slice_callback 🔒 ⚠

get_host_defined_data 🔒 ⚠

https://searchfox.org/mozilla-central/rev/2a8a30f4c9b918b726891ab9d2d62b76152606f1/xpcom/base/CycleCollectedJSContext.cpp#316

get_size 🔒 ⚠

in_range 🔒

notify_about_rejected_promises 🔒

https://html.spec.whatwg.org/multipage/#notify-about-rejected-promises

pop_interrupt_queue 🔒 ⚠

promise_rejection_tracker 🔒 ⚠

https://html.spec.whatwg.org/multipage/#the-hostpromiserejectiontracker-implementation

push_new_interrupt_queue 🔒 ⚠

report_stream_error 🔒 ⚠

run_jobs 🔒 ⚠

safely_convert_null_to_string 🔒

servo_build_id 🔒 ⚠

set_gc_zeal_options 🔒 ⚠

trace_rust_roots 🔒 ⚠