Module script::script_runtime

The script runtime contains common traits and structs commonly used by the script thread, the dom, and the worker threads.

Structs

CanGc

A compile-time marker that there are operations that could trigger a JS garbage collection operation within the current stack frame. It is trivially copyable, so it should be passed as a function argument and reused when calling other functions whenever possible. Since it is only meaningful within the current stack frame, it is impossible to move it to a different thread or into a task that will execute asynchronously.

JSContext

JSEngineSetup

Runnable

Runtime

StreamConsumer

ThreadSafeJSContext

A version of the JSContext that can be used from other threads and is thus Send and Sync. This should only ever expose operations that are marked as thread-safe by the SpiderMonkey API, ie ones that only atomic fields in JSContext.

Enums

CommonScriptMsg

Common messages used to control the event loops in both the script and the worker

ScriptThreadEventCategory

Constants

GC_CYCLE_START 🔒

GC_SLICE_START 🔒

THREAD_ACTIVE 🔒

Statics

JOB_QUEUE_TRAPS 🔒

JS_ENGINE 🔒

SECURITY_CALLBACKS 🔒

Traits

ScriptChan

A cloneable interface for communicating with an event loop.

ScriptPort

An interface for receiving ScriptMsg values in an event loop. Used for synchronous DOM APIs that need to abstract over multiple kinds of event loops (worker/main thread) with different Receiver interfaces.

Functions

consume_stream 🔒 ⚠

Implements the steps to compile webassembly response mentioned here https://webassembly.github.io/spec/web-api/#compile-a-potential-webassembly-response

content_security_policy_allows 🔒 ⚠

debug_gc_callback 🔒 ⚠

empty 🔒 ⚠

enqueue_promise_job 🔒 ⚠

SM callback for promise job resolution. Adds a promise callback to the current global’s microtask queue.

gc_slice_callback 🔒 ⚠

get_incumbent_global 🔒 ⚠

get_size 🔒 ⚠

in_range 🔒

notify_about_rejected_promises

https://html.spec.whatwg.org/multipage/#notify-about-rejected-promises

promise_rejection_tracker 🔒 ⚠

https://html.spec.whatwg.org/multipage/#the-hostpromiserejectiontracker-implementation

report_stream_error 🔒 ⚠

runtime_is_alive 🔒

servo_build_id 🔒 ⚠

set_gc_zeal_options 🔒 ⚠

trace_rust_roots 🔒 ⚠