Struct script::script_thread::InProgressLoad

struct InProgressLoad {
    pipeline_id: PipelineId,
    parent_info: Option<(PipelineId, SubpageId)>,
    window_size: Option<WindowSizeData>,
    layout_chan: Sender<Msg>,
    clip_rect: Option<Rect<f32>>,
    is_frozen: bool,
    url: Url,
}

A document load that is in the process of fetching the requested resource. Contains data that will need to be present when the document and frame tree entry are created, but is only easily available at initiation of the load and on a push basis (so some data will be updated according to future resize events, viewport changes, etc.)

Fields

pipeline_id

The pipeline which requested this load.

parent_info

The parent pipeline and child subpage associated with this load, if any.

window_size

The current window size associated with this pipeline.

layout_chan

Channel to the layout thread associated with this pipeline.

clip_rect

The current viewport clipping rectangle applying to this pipeline, if any.

is_frozen

Window is frozen (navigated away while loading for example).

url

The requested URL of the load.

Methods

impl InProgressLoad

fn new(id: PipelineId, parent_info: Option<(PipelineId, SubpageId)>, layout_chan: Sender<Msg>, window_size: Option<WindowSizeData>, url: Url) -> InProgressLoad

Create a new InProgressLoad object.