Struct script::script_thread::InProgressLoad

struct InProgressLoad {
    pipeline_id: PipelineId,
    browsing_context_id: BrowsingContextId,
    top_level_browsing_context_id: TopLevelBrowsingContextId,
    parent_info: Option<PipelineId>,
    opener: Option<BrowsingContextId>,
    window_size: WindowSizeData,
    layout_chan: Sender<Msg>,
    activity: DocumentActivity,
    is_visible: bool,
    url: ServoUrl,
    origin: MutableOrigin,
    navigation_start: u64,
    navigation_start_precise: u64,
    canceller: FetchCanceller,
    layout_is_busy: Arc<AtomicBool>,
}

A document load that is in the process of fetching the requested resource. Contains data that will need to be present when the document and frame tree entry are created, but is only easily available at initiation of the load and on a push basis (so some data will be updated according to future resize events, viewport changes, etc.)

Fields
pipeline_id: PipelineId

The pipeline which requested this load.

browsing_context_id: BrowsingContextId

The browsing context being loaded into.

top_level_browsing_context_id: TopLevelBrowsingContextId

The top level ancestor browsing context.

parent_info: Option<PipelineId>

The parent pipeline and frame type associated with this load, if any.

opener: Option<BrowsingContextId>

The opener, if this is an auxiliary.

window_size: WindowSizeData

The current window size associated with this pipeline.

layout_chan: Sender<Msg>

Channel to the layout thread associated with this pipeline.

activity: DocumentActivity

The activity level of the document (inactive, active or fully active).

is_visible: bool

Window is visible.

url: ServoUrl

The requested URL of the load.

origin: MutableOrigin

The origin for the document

navigation_start: u64

Timestamp reporting the time when the browser started this load.

navigation_start_precise: u64

High res timestamp reporting the time when the browser started this load.

canceller: FetchCanceller

For cancelling the fetch

layout_is_busy: Arc<AtomicBool>

Flag for sharing with the layout thread that is not yet created.

Methods

impl InProgressLoad

fn new(
id: PipelineId,
browsing_context_id: BrowsingContextId,
top_level_browsing_context_id: TopLevelBrowsingContextId,
parent_info: Option<PipelineId>,
opener: Option<BrowsingContextId>,
layout_chan: Sender<Msg>,
window_size: WindowSizeData,
url: ServoUrl,
origin: MutableOrigin,
layout_is_busy: Arc<AtomicBool>
) -> InProgressLoad

Create a new InProgressLoad object.