Struct script::script_thread::ScriptThread

pub struct ScriptThread {
    documents: DOMRefCell<Documents>,
    browsing_contexts: DOMRefCell<HashMap<FrameId, JS<BrowsingContext>>>,
    incomplete_loads: DOMRefCell<Vec<InProgressLoad>>,
    registration_map: DOMRefCell<HashMap<ServoUrl, JS<ServiceWorkerRegistration>>>,
    job_queue_map: Rc<JobQueue>,
    image_cache_thread: ImageCacheThread,
    resource_threads: ResourceThreads,
    bluetooth_thread: IpcSender<BluetoothRequest>,
    port: Receiver<MainThreadScriptMsg>,
    chan: MainThreadScriptChan,
    dom_manipulation_task_source: DOMManipulationTaskSource,
    user_interaction_task_source: UserInteractionTaskSource,
    networking_task_source: NetworkingTaskSource,
    history_traversal_task_source: HistoryTraversalTaskSource,
    file_reading_task_source: FileReadingTaskSource,
    control_chan: IpcSender<ConstellationControlMsg>,
    control_port: Receiver<ConstellationControlMsg>,
    constellation_chan: IpcSender<ConstellationMsg>,
    layout_to_constellation_chan: IpcSender<LayoutMsg>,
    image_cache_port: Receiver<ImageCacheMsg>,
    image_cache_channel: Sender<ImageCacheMsg>,
    time_profiler_chan: ProfilerChan,
    mem_profiler_chan: ProfilerChan,
    devtools_chan: Option<IpcSender<ScriptToDevtoolsControlMsg>>,
    devtools_port: Receiver<DevtoolScriptControlMsg>,
    devtools_sender: IpcSender<DevtoolScriptControlMsg>,
    js_runtime: Rc<Runtime>,
    topmost_mouse_over_target: MutNullableJS<Element>,
    closed_pipelines: DOMRefCell<HashSet<PipelineId>>,
    scheduler_chan: IpcSender<TimerEventRequest>,
    timer_event_chan: Sender<TimerEvent>,
    timer_event_port: Receiver<TimerEvent>,
    content_process_shutdown_chan: IpcSender<()>,
    microtask_queue: MicrotaskQueue,
    webvr_thread: Option<IpcSender<WebVRMsg>>,
    docs_with_no_blocking_loads: DOMRefCell<HashSet<JS<Document>>>,
}

Fields
documents: DOMRefCell<Documents>

The documents for pipelines managed by this thread

browsing_contexts: DOMRefCell<HashMap<FrameId, JS<BrowsingContext>>>

The browsing contexts known by this thread TODO: this map grows, but never shrinks. Issue #15258.

incomplete_loads: DOMRefCell<Vec<InProgressLoad>>

A list of data pertaining to loads that have not yet received a network response

registration_map: DOMRefCell<HashMap<ServoUrl, JS<ServiceWorkerRegistration>>>

A map to store service worker registrations for a given origin

job_queue_map: Rc<JobQueue>

A job queue for Service Workers keyed by their scope url

image_cache_thread: ImageCacheThread

A handle to the image cache thread.

resource_threads: ResourceThreads

A handle to the resource thread. This is an Arc to avoid running out of file descriptors if there are many iframes.

bluetooth_thread: IpcSender<BluetoothRequest>

A handle to the bluetooth thread.

port: Receiver<MainThreadScriptMsg>

The port on which the script thread receives messages (load URL, exit, etc.)

chan: MainThreadScriptChan

A channel to hand out to script thread-based entities that need to be able to enqueue events in the event queue.

dom_manipulation_task_source: DOMManipulationTaskSource user_interaction_task_source: UserInteractionTaskSource networking_task_source: NetworkingTaskSource history_traversal_task_source: HistoryTraversalTaskSource file_reading_task_source: FileReadingTaskSource control_chan: IpcSender<ConstellationControlMsg>

A channel to hand out to threads that need to respond to a message from the script thread.

control_port: Receiver<ConstellationControlMsg>

The port on which the constellation and layout threads can communicate with the script thread.

constellation_chan: IpcSender<ConstellationMsg>

For communicating load url messages to the constellation

layout_to_constellation_chan: IpcSender<LayoutMsg>

A sender for new layout threads to communicate to the constellation.

image_cache_port: Receiver<ImageCacheMsg>

The port on which we receive messages from the image cache

image_cache_channel: Sender<ImageCacheMsg>

The channel on which the image cache can send messages to ourself.

time_profiler_chan: ProfilerChan

For providing contact with the time profiler.

mem_profiler_chan: ProfilerChan

For providing contact with the memory profiler.

devtools_chan: Option<IpcSender<ScriptToDevtoolsControlMsg>>

For providing instructions to an optional devtools server.

devtools_port: Receiver<DevtoolScriptControlMsg>

For receiving commands from an optional devtools server. Will be ignored if no such server exists.

devtools_sender: IpcSender<DevtoolScriptControlMsg> js_runtime: Rc<Runtime>

The JavaScript runtime.

topmost_mouse_over_target: MutNullableJS<Element>

The topmost element over the mouse.

closed_pipelines: DOMRefCell<HashSet<PipelineId>>

List of pipelines that have been owned and closed by this script thread.

scheduler_chan: IpcSender<TimerEventRequest> timer_event_chan: Sender<TimerEvent> timer_event_port: Receiver<TimerEvent> content_process_shutdown_chan: IpcSender<()> microtask_queue: MicrotaskQueue webvr_thread: Option<IpcSender<WebVRMsg>>

A handle to the webvr thread, if available

docs_with_no_blocking_loads: DOMRefCell<HashSet<JS<Document>>>

A list of pipelines containing documents that finished loading all their blocking resources during a turn of the event loop.

Methods

impl ScriptThread

fn mark_document_with_no_blocked_loads(doc: &Document)

fn invoke_perform_a_microtask_checkpoint()

fn page_headers_available(id: &PipelineId,
metadata: Option<Metadata>)
-> Option<Root<ServoParser>>

fn schedule_job(job: Job, global: &GlobalScope)

fn process_event(msg: CommonScriptMsg)

fn await_stable_state<T: Runnable + Send + 'static>(task: T)

fn process_attach_layout(new_layout_info: NewLayoutInfo, origin: MutableOrigin)

fn find_document(id: PipelineId) -> Option<Root<Document>>

fn find_browsing_context(id: FrameId) -> Option<Root<BrowsingContext>>

fn new(state: InitialScriptState,
port: Receiver<MainThreadScriptMsg>,
chan: Sender<MainThreadScriptMsg>)
-> ScriptThread

Creates a new script thread.

fn get_cx(&self) -> *mut JSContext

fn start(&self)

Starts the script thread. After calling this method, the script thread will loop receiving messages on its port.

fn handle_msgs(&self) -> bool

Handle incoming control messages.

fn categorize_msg(&self, msg: &MixedMessage) -> ScriptThreadEventCategory

fn profile_event<F, R>(&self, category: ScriptThreadEventCategory, f: F) -> R where F: FnOnce() -> R

fn handle_msg_from_constellation(&self, msg: ConstellationControlMsg)

fn handle_msg_from_script(&self, msg: MainThreadScriptMsg)

fn handle_timer_event(&self, timer_event: TimerEvent)

fn handle_msg_from_devtools(&self, msg: DevtoolScriptControlMsg)

fn handle_msg_from_image_cache(&self,
(id, response): (PipelineId, PendingImageResponse))

fn handle_webdriver_msg(&self,
pipeline_id: PipelineId,
msg: WebDriverScriptCommand)

fn handle_resize(&self,
id: PipelineId,
size: WindowSizeData,
size_type: WindowSizeType)

fn handle_viewport(&self, id: PipelineId, rect: Rect<f32>)

fn handle_set_scroll_state(&self,
id: PipelineId,
scroll_states: &[(UntrustedNodeAddress, Point2D<f32>)])

fn handle_new_layout(&self,
new_layout_info: NewLayoutInfo,
origin: MutableOrigin)

fn collect_reports(&self, reports_chan: ReportsChan)

fn handle_visibility_change_complete_msg(&self,
parent_pipeline_id: PipelineId,
id: FrameId,
visible: bool)

Updates iframe element after a change in visibility

fn handle_visibility_change_msg(&self, id: PipelineId, visible: bool)

Handle visibility change message

fn handle_set_document_activity_msg(&self,
id: PipelineId,
activity: DocumentActivity)

Handles activity change message

fn handle_focus_iframe_msg(&self,
parent_pipeline_id: PipelineId,
frame_id: FrameId)

fn handle_framed_content_changed(&self,
parent_pipeline_id: PipelineId,
frame_id: FrameId)

fn handle_post_message_msg(&self,
pipeline_id: PipelineId,
origin: Option<ImmutableOrigin>,
data: Vec<u8>)

fn handle_mozbrowser_event_msg(&self,
parent_pipeline_id: PipelineId,
frame_id: Option<FrameId>,
event: MozBrowserEvent)

Handles a mozbrowser event, for example see: https://developer.mozilla.org/en-US/docs/Web/Events/mozbrowserloadstart

fn handle_update_pipeline_id(&self,
parent_pipeline_id: PipelineId,
frame_id: FrameId,
new_pipeline_id: PipelineId)

fn handle_resize_inactive_msg(&self, id: PipelineId, new_size: WindowSizeData)

Window was resized, but this script was not active, so don't reflow yet

fn handle_exit_window_msg(&self, _: PipelineId)

We have gotten a window.close from script, which we pass on to the compositor. We do not shut down the script thread now, because the compositor will ask the constellation to shut down the pipeline, which will clean everything up normally. If we do exit, we will tear down the DOM nodes, possibly at a point where layout is still accessing them.

fn handle_page_headers_available(&self,
id: &PipelineId,
metadata: Option<Metadata>)
-> Option<Root<ServoParser>>

We have received notification that the response associated with a load has completed. Kick off the document and frame tree creation process using the result.

fn handle_get_registration(&self,
scope_url: &ServoUrl)
-> Option<Root<ServiceWorkerRegistration>>

fn handle_serviceworker_registration(&self,
scope: &ServoUrl,
registration: &ServiceWorkerRegistration,
pipeline_id: PipelineId)

fn dispatch_job_queue(&self, job_handler: Box<AsyncJobHandler>)

fn dom_manipulation_task_source(&self) -> &DOMManipulationTaskSource

fn handle_get_title_msg(&self, pipeline_id: PipelineId)

Handles a request for the window title.

fn handle_exit_pipeline_msg(&self,
id: PipelineId,
discard_bc: DiscardBrowsingContext)

Handles a request to exit a pipeline and shut down layout.

fn handle_exit_script_thread_msg(&self)

Handles a request to exit the script thread and shut down layout.

fn handle_tick_all_animations(&self, id: PipelineId)

Handles when layout thread finishes all animation in one tick

fn handle_transition_event(&self,
unsafe_node: UnsafeNode,
name: String,
duration: f64)

Handles firing of transition events.

fn handle_web_font_loaded(&self, pipeline_id: PipelineId)

Handles a Web font being loaded. Does nothing if the page no longer exists.

fn handle_storage_event(&self,
pipeline_id: PipelineId,
storage_type: StorageType,
url: ServoUrl,
key: Option<String>,
old_value: Option<String>,
new_value: Option<String>)

Notify a window of a storage event

fn handle_frame_load_event(&self,
parent_id: PipelineId,
frame_id: FrameId,
child_id: PipelineId)

Notify the containing document of a child frame that has completed loading.

fn load(&self,
metadata: Metadata,
incomplete: InProgressLoad)
-> Root<ServoParser>

The entry point to document loading. Defines bindings, sets up the window and document objects, parses HTML and CSS, and kicks off initial layout.

fn notify_devtools(&self,
title: DOMString,
url: ServoUrl,
ids: (PipelineId, Option<WorkerId>))

fn rebuild_and_force_reflow(&self, document: &Document, reason: ReflowReason)

Reflows non-incrementally, rebuilding the entire layout tree in the process.

fn handle_event(&self, pipeline_id: PipelineId, event: CompositorEvent)

This is the main entry point for receiving and dispatching DOM events.

TODO: Actually perform DOM event dispatch.

fn handle_mouse_event(&self,
pipeline_id: PipelineId,
mouse_event_type: MouseEventType,
button: MouseButton,
point: Point2D<f32>)

fn handle_touch_event(&self,
pipeline_id: PipelineId,
event_type: TouchEventType,
identifier: TouchId,
point: Point2D<f32>)
-> TouchEventResult

fn handle_navigate(&self,
parent_pipeline_id: PipelineId,
frame_id: Option<FrameId>,
load_data: LoadData,
replace: bool)

https://html.spec.whatwg.org/multipage/#navigating-across-documents The entry point for content to notify that a new load has been requested for the given pipeline (specifically the "navigate" algorithm).

fn handle_resize_event(&self,
pipeline_id: PipelineId,
new_size: WindowSizeData,
size_type: WindowSizeType)

fn start_page_load(&self, incomplete: InProgressLoad, load_data: LoadData)

Initiate a non-blocking fetch for a specified resource. Stores the InProgressLoad argument until a notification is received that the fetch is complete.

fn start_page_load_about_blank(&self, incomplete: InProgressLoad)

Synchronously fetch about:blank. Stores the InProgressLoad argument until a notification is received that the fetch is complete.

fn handle_css_error_reporting(&self,
pipeline_id: PipelineId,
filename: String,
line: usize,
column: usize,
msg: String)

fn handle_reload(&self, pipeline_id: PipelineId)

fn handle_webvr_event(&self, pipeline_id: PipelineId, event: WebVREventMsg)

fn enqueue_microtask(job: Microtask)

fn perform_a_microtask_checkpoint(&self)