Struct script::script_thread::ScriptThread

pub struct ScriptThread {
    browsing_context: MutNullableHeap<JS<BrowsingContext>>,
    incomplete_loads: DOMRefCell<Vec<InProgressLoad>>,
    registration_map: DOMRefCell<HashMap<Url, JS<ServiceWorkerRegistration>>>,
    image_cache_thread: ImageCacheThread,
    resource_threads: ResourceThreads,
    bluetooth_thread: IpcSender<BluetoothMethodMsg>,
    port: Receiver<MainThreadScriptMsg>,
    chan: MainThreadScriptChan,
    custom_message_chan: IpcSender<CustomResponseSender>,
    custom_message_port: Receiver<CustomResponseSender>,
    dom_manipulation_task_source: DOMManipulationTaskSource,
    user_interaction_task_source: UserInteractionTaskSource,
    networking_task_source: NetworkingTaskSource,
    history_traversal_task_source: HistoryTraversalTaskSource,
    file_reading_task_source: FileReadingTaskSource,
    control_chan: IpcSender<ConstellationControlMsg>,
    control_port: Receiver<ConstellationControlMsg>,
    constellation_chan: IpcSender<ConstellationMsg>,
    image_cache_port: Receiver<ImageCacheResult>,
    image_cache_channel: ImageCacheChan,
    time_profiler_chan: ProfilerChan,
    mem_profiler_chan: ProfilerChan,
    devtools_chan: Option<IpcSender<ScriptToDevtoolsControlMsg>>,
    devtools_port: Receiver<DevtoolScriptControlMsg>,
    devtools_sender: IpcSender<DevtoolScriptControlMsg>,
    js_runtime: Rc<Runtime>,
    topmost_mouse_over_target: MutNullableHeap<JS<Element>>,
    closed_pipelines: DOMRefCell<HashSet<PipelineId>>,
    scheduler_chan: IpcSender<TimerEventRequest>,
    timer_event_chan: Sender<TimerEvent>,
    timer_event_port: Receiver<TimerEvent>,
    content_process_shutdown_chan: IpcSender<()>,
    panic_chan: IpcSender<PanicMsg>,
}

Information for an entire page. Pages are top-level browsing contexts and can contain multiple frames.

Fields

browsing_context

A handle to the information pertaining to page layout

incomplete_loads

A list of data pertaining to loads that have not yet received a network response

registration_map

A map to store service worker registrations for a given origin

image_cache_thread

A handle to the image cache thread.

resource_threads

A handle to the resource thread. This is an Arc to avoid running out of file descriptors if there are many iframes.

bluetooth_thread

A handle to the bluetooth thread.

port

The port on which the script thread receives messages (load URL, exit, etc.)

chan

A channel to hand out to script thread-based entities that need to be able to enqueue events in the event queue.

custom_message_chan

A handle to network event messages

custom_message_port

The port which receives a sender from the network

dom_manipulation_task_source

user_interaction_task_source

networking_task_source

history_traversal_task_source

file_reading_task_source

control_chan

A channel to hand out to threads that need to respond to a message from the script thread.

control_port

The port on which the constellation and layout threads can communicate with the script thread.

constellation_chan

For communicating load url messages to the constellation

image_cache_port

The port on which we receive messages from the image cache

image_cache_channel

The channel on which the image cache can send messages to ourself.

time_profiler_chan

For providing contact with the time profiler.

mem_profiler_chan

For providing contact with the memory profiler.

devtools_chan

For providing instructions to an optional devtools server.

devtools_port

For receiving commands from an optional devtools server. Will be ignored if no such server exists.

devtools_sender

js_runtime

The JavaScript runtime.

topmost_mouse_over_target

The topmost element over the mouse.

closed_pipelines

List of pipelines that have been owned and closed by this script thread.

scheduler_chan

timer_event_chan

timer_event_port

content_process_shutdown_chan

panic_chan

Methods

impl ScriptThread

fn page_headers_available(id: &PipelineId, subpage: Option<&SubpageId>, metadata: Option<Metadata>) -> Option<ParserRoot>

fn set_registration(scope_url: Url, registration: &ServiceWorkerRegistration)

fn parsing_complete(id: PipelineId)

fn process_event(msg: CommonScriptMsg)

fn await_stable_state<T: Runnable + Send + 'static>(task: T)

fn new(state: InitialScriptState, port: Receiver<MainThreadScriptMsg>, chan: Sender<MainThreadScriptMsg>) -> ScriptThread

Creates a new script thread.

fn root_browsing_context(&self) -> Root<BrowsingContext>

fn root_browsing_context_exists(&self) -> bool

fn find_child_context(&self, pipeline_id: PipelineId) -> Option<Root<BrowsingContext>>

Find a child browsing context of the root context by pipeline id. Returns None if the root context does not exist or the child context cannot be found.

fn get_cx(&self) -> *mut JSContext

fn start(&self)

Starts the script thread. After calling this method, the script thread will loop receiving messages on its port.

fn handle_msgs(&self) -> bool

Handle incoming control messages.

fn categorize_msg(&self, msg: &MixedMessage) -> ScriptThreadEventCategory

fn profile_event<F, R>(&self, category: ScriptThreadEventCategory, f: F) -> R where F: FnOnce() -> R

fn handle_msg_from_constellation(&self, msg: ConstellationControlMsg)

fn handle_msg_from_script(&self, msg: MainThreadScriptMsg)

fn handle_timer_event(&self, timer_event: TimerEvent)

fn handle_msg_from_devtools(&self, msg: DevtoolScriptControlMsg)

fn handle_msg_from_image_cache(&self, msg: ImageCacheResult)

fn handle_msg_from_network(&self, msg: IpcSender<Option<CustomResponse>>)

fn handle_webdriver_msg(&self, pipeline_id: PipelineId, msg: WebDriverScriptCommand)

fn handle_resize(&self, id: PipelineId, size: WindowSizeData, size_type: WindowSizeType)

fn handle_viewport(&self, id: PipelineId, rect: Rect<f32>)

fn handle_set_scroll_state(&self, id: PipelineId, scroll_state: &Point2D<f32>)

fn handle_new_layout(&self, new_layout_info: NewLayoutInfo)

fn handle_loads_complete(&self, pipeline: PipelineId)

fn collect_reports(&self, reports_chan: ReportsChan)

fn handle_freeze_msg(&self, id: PipelineId)

Handles freeze message

fn handle_thaw_msg(&self, id: PipelineId)

Handles thaw message

fn handle_focus_iframe_msg(&self, parent_pipeline_id: PipelineId, subpage_id: SubpageId)

fn handle_framed_content_changed(&self, parent_pipeline_id: PipelineId, subpage_id: SubpageId)

fn handle_mozbrowser_event_msg(&self, parent_pipeline_id: PipelineId, subpage_id: SubpageId, event: MozBrowserEvent)

Handles a mozbrowser event, for example see: https://developer.mozilla.org/en-US/docs/Web/Events/mozbrowserloadstart

fn handle_update_subpage_id(&self, containing_pipeline_id: PipelineId, old_subpage_id: SubpageId, new_subpage_id: SubpageId, new_pipeline_id: PipelineId)

fn handle_resize_inactive_msg(&self, id: PipelineId, new_size: WindowSizeData)

Window was resized, but this script was not active, so don't reflow yet

fn handle_exit_window_msg(&self, _: PipelineId)

We have gotten a window.close from script, which we pass on to the compositor. We do not shut down the script thread now, because the compositor will ask the constellation to shut down the pipeline, which will clean everything up normally. If we do exit, we will tear down the DOM nodes, possibly at a point where layout is still accessing them.

fn handle_page_headers_available(&self, id: &PipelineId, subpage: Option<&SubpageId>, metadata: Option<Metadata>) -> Option<ParserRoot>

We have received notification that the response associated with a load has completed. Kick off the document and frame tree creation process using the result.

fn handle_serviceworker_registration(&self, scope: Url, registration: &ServiceWorkerRegistration)

fn handle_get_title_msg(&self, pipeline_id: PipelineId)

Handles a request for the window title.

fn handle_exit_pipeline_msg(&self, id: PipelineId) -> bool

Handles a request to exit the script thread and shut down layout. Returns true if the script thread should shut down and false otherwise.

fn handle_tick_all_animations(&self, id: PipelineId)

Handles when layout thread finishes all animation in one tick

fn handle_web_font_loaded(&self, pipeline_id: PipelineId)

Handles a Web font being loaded. Does nothing if the page no longer exists.

fn handle_frame_load_event(&self, containing_pipeline: PipelineId, id: PipelineId)

Notify the containing document of a child frame that has completed loading.

fn load(&self, metadata: Metadata, incomplete: InProgressLoad) -> ParserRoot

The entry point to document loading. Defines bindings, sets up the window and document objects, parses HTML and CSS, and kicks off initial layout.

fn notify_devtools(&self, title: DOMString, url: Url, ids: (PipelineId, Option<WorkerId>))

fn scroll_fragment_point(&self, pipeline_id: PipelineId, element: &Element)

fn rebuild_and_force_reflow(&self, context: &BrowsingContext, reason: ReflowReason)

Reflows non-incrementally, rebuilding the entire layout tree in the process.

fn handle_event(&self, pipeline_id: PipelineId, event: CompositorEvent)

This is the main entry point for receiving and dispatching DOM events.

TODO: Actually perform DOM event dispatch.

fn handle_mouse_event(&self, pipeline_id: PipelineId, mouse_event_type: MouseEventType, button: MouseButton, point: Point2D<f32>)

fn handle_touch_event(&self, pipeline_id: PipelineId, event_type: TouchEventType, identifier: TouchId, point: Point2D<f32>) -> bool

fn handle_navigate(&self, pipeline_id: PipelineId, subpage_id: Option<SubpageId>, load_data: LoadData)

https://html.spec.whatwg.org/multipage/#navigating-across-documents The entry point for content to notify that a new load has been requested for the given pipeline (specifically the "navigate" algorithm).

fn handle_resize_event(&self, pipeline_id: PipelineId, new_size: WindowSizeData, size_type: WindowSizeType)

fn start_page_load(&self, incomplete: InProgressLoad, load_data: LoadData)

Initiate a non-blocking fetch for a specified resource. Stores the InProgressLoad argument until a notification is received that the fetch is complete.

fn handle_parsing_complete(&self, id: PipelineId)

fn handle_css_error_reporting(&self, pipeline_id: PipelineId, filename: String, line: usize, column: usize, msg: String)