script::script_thread

Struct ScriptThread

pub struct ScriptThread {

    this: Weak<ScriptThread>,
    last_render_opportunity_time: Cell<Option<Instant>>,
    documents: DomRefCell<DocumentCollection>,
    window_proxies: Rc<ScriptWindowProxies>,
    incomplete_loads: DomRefCell<Vec<InProgressLoad>>,
    incomplete_parser_contexts: IncompleteParserContexts,
    image_cache_factory: Arc<dyn ImageCacheFactory>,
    receivers: ScriptThreadReceivers,
    senders: ScriptThreadSenders,
    resource_threads: ResourceThreads,
    storage_threads: StorageThreads,
    task_queue: TaskQueue<MainThreadScriptMsg>,
    background_hang_monitor: Box<dyn BackgroundHangMonitor>,
    closing: Arc<AtomicBool>,
    timer_scheduler: RefCell<TimerScheduler>,
    system_font_service: Arc<SystemFontServiceProxy>,
    js_runtime: Rc<Runtime>,
    closed_pipelines: DomRefCell<FxHashSet<PipelineId>>,
    microtask_queue: Rc<MicrotaskQueue>,
    mutation_observers: Rc<ScriptMutationObservers>,
    webgl_chan: Option<WebGLPipeline>,
    webxr_registry: Option<Registry>,
    worklet_thread_pool: DomRefCell<Option<Rc<WorkletThreadPool>>>,
    docs_with_no_blocking_loads: DomRefCell<HashSet<Dom<Document>>>,
    custom_element_reaction_stack: Rc<CustomElementReactionStack>,
    paint_api: CrossProcessPaintApi,
    profile_script_events: bool,
    print_pwm: bool,
    unminify_js: bool,
    local_script_source: Option<String>,
    unminify_css: bool,
    user_contents_for_manager_id: RefCell<FxHashMap<UserContentManagerId, Rc<UserContents>>>,
    player_context: WindowGLContext,
    pipeline_to_node_ids: DomRefCell<FxHashMap<PipelineId, HashSet<String>>>,
    is_user_interacting: Rc<Cell<bool>>,
    gpu_id_hub: Arc<IdentityHub>,
    layout_factory: Arc<dyn LayoutFactory>,
    scheduled_update_the_rendering: RefCell<Option<TimerId>>,
    needs_rendering_update: Arc<AtomicBool>,
    debugger_global: Dom<DebuggerGlobalScope>,
    privileged_urls: Vec<ServoUrl>,

}

Fields
this: Weak<ScriptThread>

A reference to the currently operating ScriptThread. This should always be upgradable to an Rc as long as the ScriptThread is running.

last_render_opportunity_time: Cell<Option<Instant>>

https://html.spec.whatwg.org/multipage/#last-render-opportunity-time

documents: DomRefCell<DocumentCollection>

The documents for pipelines managed by this thread

window_proxies: Rc<ScriptWindowProxies>

The window proxies known by this thread

incomplete_loads: DomRefCell<Vec<InProgressLoad>>

A list of data pertaining to loads that have not yet received a network response

incomplete_parser_contexts: IncompleteParserContexts

A vector containing parser contexts which have not yet been fully processed

image_cache_factory: Arc<dyn ImageCacheFactory>

An ImageCacheFactory to use for creating ImageCaches for all of the child Pipelines.

receivers: ScriptThreadReceivers

A ScriptThreadReceivers holding all of the incoming Receivers for messages to this ScriptThread.

senders: ScriptThreadSenders

A ScriptThreadSenders that holds all outgoing sending channels necessary to communicate to other parts of Servo.

resource_threads: ResourceThreads

A handle to the resource thread. This is an Arc to avoid running out of file descriptors if there are many iframes.

storage_threads: StorageThreadstask_queue: TaskQueue<MainThreadScriptMsg>

A queue of tasks to be executed in this script-thread.

background_hang_monitor: Box<dyn BackgroundHangMonitor>

The dedicated means of communication with the background-hang-monitor for this script-thread.

closing: Arc<AtomicBool>

A flag set to true by the BHM on exit, and checked from within the interrupt handler.

timer_scheduler: RefCell<TimerScheduler>

A TimerScheduler used to schedule timers for this ScriptThread. Timers are handled in the ScriptThread event loop.

system_font_service: Arc<SystemFontServiceProxy>

A proxy to the SystemFontService to use for accessing system font lists.

js_runtime: Rc<Runtime>

The JavaScript runtime.

closed_pipelines: DomRefCell<FxHashSet<PipelineId>>

List of pipelines that have been owned and closed by this script thread.

microtask_queue: Rc<MicrotaskQueue>

https://html.spec.whatwg.org/multipage/#microtask-queue

mutation_observers: Rc<ScriptMutationObservers>webgl_chan: Option<WebGLPipeline>

A handle to the WebGL thread

webxr_registry: Option<Registry>

The WebXR device registry

worklet_thread_pool: DomRefCell<Option<Rc<WorkletThreadPool>>>

The worklet thread pool

docs_with_no_blocking_loads: DomRefCell<HashSet<Dom<Document>>>

A list of pipelines containing documents that finished loading all their blocking resources during a turn of the event loop.

custom_element_reaction_stack: Rc<CustomElementReactionStack>

https://html.spec.whatwg.org/multipage/#custom-element-reactions-stack

paint_api: CrossProcessPaintApi

Cross-process access to Paint’s API.

profile_script_events: bool

Periodically print out on which events script threads spend their processing time.

print_pwm: bool

Print Progressive Web Metrics to console.

unminify_js: bool

Unminify Javascript.

local_script_source: Option<String>

Directory with stored unminified scripts

unminify_css: bool

Unminify Css.

user_contents_for_manager_id: RefCell<FxHashMap<UserContentManagerId, Rc<UserContents>>>

A map from UserContentManagerId to its UserContents. This is initialized with a copy of the map in constellation (via the InitialScriptState). After that, the constellation forwards any mutations to this ScriptThread using messages.

player_context: WindowGLContext

Application window’s GL Context for Media player

pipeline_to_node_ids: DomRefCell<FxHashMap<PipelineId, HashSet<String>>>

A map from pipelines to all owned nodes ever created in this script thread

is_user_interacting: Rc<Cell<bool>>

Code is running as a consequence of a user interaction

gpu_id_hub: Arc<IdentityHub>

Identity manager for WebGPU resources

layout_factory: Arc<dyn LayoutFactory>

A factory for making new layouts. This allows layout to depend on script.

scheduled_update_the_rendering: RefCell<Option<TimerId>>

The TimerId of a ScriptThread-scheduled “update the rendering” call, if any. The ScriptThread schedules calls to “update the rendering,” but the renderer can also do this when animating. Renderer-based calls always take precedence.

needs_rendering_update: Arc<AtomicBool>

Whether an animation tick or ScriptThread-triggered rendering update is pending. This might either be because the Servo renderer is managing animations and the ScriptThread has received a ScriptThreadMessage::TickAllAnimations message, because the ScriptThread itself is managing animations the timer fired triggering a ScriptThread-based animation tick, or if there are no animations running and the ScriptThread has noticed a change that requires a rendering update.

debugger_global: Dom<DebuggerGlobalScope>privileged_urls: Vec<ServoUrl>

A list of URLs that can access privileged internal APIs.

Implementations

impl ScriptThread

pub(crate) fn runtime_handle() -> ParentRuntime

pub(crate) fn can_continue_running() -> bool

pub(crate) fn prepare_for_shutdown()

pub(crate) fn mutation_observers() -> Rc<ScriptMutationObservers>

pub(crate) fn microtask_queue() -> Rc<MicrotaskQueue>

pub(crate) fn mark_document_with_no_blocked_loads(doc: &Document)

pub(crate) fn page_headers_available( webview_id: WebViewId, pipeline_id: PipelineId, metadata: Option<Metadata>, can_gc: CanGc, ) -> Option<DomRoot<ServoParser>>

pub(crate) fn process_event(msg: CommonScriptMsg, cx: &mut JSContext) -> bool

Process a single event as if it were the next event in the queue for this window event-loop. Returns a boolean indicating whether further events should be processed.

pub(crate) fn schedule_timer(&self, request: TimerEventRequest) -> TimerId

Schedule a TimerEventRequest on this ScriptThread’s TimerScheduler.

pub(crate) fn cancel_timer(&self, timer_id: TimerId)

Cancel a the TimerEventRequest for the given TimerId on this ScriptThread’s TimerScheduler.

pub(crate) fn await_stable_state(task: Microtask)

pub(crate) fn check_load_origin( source: &LoadOrigin, target: &ImmutableOrigin, ) -> bool

Check that two origins are “similar enough”, for now only used to prevent cross-origin JS url evaluation.

https://github.com/whatwg/html/issues/2591

pub(crate) fn set_needs_rendering_update(&self)

Inform the ScriptThread that it should make a call to ScriptThread::update_the_rendering as soon as possible, as the rendering update timer has fired or the renderer has asked us for a new rendering update.

pub(crate) fn navigate( webview_id: WebViewId, pipeline_id: PipelineId, load_data: LoadData, history_handling: NavigationHistoryBehavior, )

Step 13 of https://html.spec.whatwg.org/multipage/#navigate

pub(crate) fn can_navigate_to_javascript_url( global: &GlobalScope, load_data: &mut LoadData, container: Option<&Element>, can_gc: CanGc, ) -> bool

https://html.spec.whatwg.org/multipage/#navigate-to-a-javascript:-url

pub(crate) fn navigate_to_javascript_url( global: &GlobalScope, containing_global: &GlobalScope, load_data: &mut LoadData, container: Option<&Element>, can_gc: CanGc, ) -> bool

pub(crate) fn get_top_level_for_browsing_context( sender_webview_id: WebViewId, sender_pipeline_id: PipelineId, browsing_context_id: BrowsingContextId, ) -> Option<WebViewId>

pub(crate) fn find_document(id: PipelineId) -> Option<DomRoot<Document>>

pub(crate) fn user_interacting_guard() -> ScriptUserInteractingGuard

Creates a guard that sets user_is_interacting to true and returns the state of user_is_interacting on drop of the guard. Notice that you need to use let _guard = ... as let _ = ... is not enough

pub(crate) fn is_user_interacting() -> bool

pub(crate) fn get_fully_active_document_ids(&self) -> FxHashSet<PipelineId>

pub(crate) fn window_proxies() -> Rc<ScriptWindowProxies>

pub(crate) fn find_window_proxy_by_name( name: &DOMString, ) -> Option<DomRoot<WindowProxy>>

pub(crate) fn worklet_thread_pool( image_cache: Arc<dyn ImageCache>, ) -> Rc<WorkletThreadPool>

The worklet will use the given ImageCache.

fn handle_register_paint_worklet( &self, pipeline_id: PipelineId, name: Atom, properties: Vec<Atom>, painter: Box<dyn Painter>, )

pub(crate) fn custom_element_reaction_stack() -> Rc<CustomElementReactionStack>

pub(crate) fn enqueue_callback_reaction( element: &Element, reaction: CallbackReaction, definition: Option<Rc<CustomElementDefinition>>, )

pub(crate) fn enqueue_upgrade_reaction( element: &Element, definition: Rc<CustomElementDefinition>, )

pub(crate) fn invoke_backup_element_queue(can_gc: CanGc)

pub(crate) fn save_node_id(pipeline: PipelineId, node_id: String)

pub(crate) fn has_node_id(pipeline: PipelineId, node_id: &str) -> bool

pub(crate) fn new( state: InitialScriptState, layout_factory: Arc<dyn LayoutFactory>, image_cache_factory: Arc<dyn ImageCacheFactory>, background_hang_monitor_register: Box<dyn BackgroundHangMonitorRegister>, ) -> (Rc<ScriptThread>, JSContext)

Creates a new script thread.

pub(crate) fn get_cx(&self) -> JSContext

fn can_continue_running_inner(&self) -> bool

Check if we are closing.

fn prepare_for_shutdown_inner(&self)

We are closing, ensure no script can run and potentially hang.

pub(crate) fn start(&self, cx: &mut JSContext)

Starts the script thread. After calling this method, the script thread will loop receiving messages on its port.

fn process_pending_input_events(&self, pipeline_id: PipelineId, can_gc: CanGc)

Process input events as part of a “update the rendering task”.

fn cancel_scheduled_update_the_rendering(&self)

fn schedule_update_the_rendering_timer_if_necessary(&self, delay: Duration)

pub(crate) fn update_the_rendering(&self, can_gc: CanGc) -> bool

https://html.spec.whatwg.org/multipage/#update-the-rendering

Attempt to update the rendering and then do a microtask checkpoint if rendering was actually updated.

Returns true if any reflows produced a new display list.

fn maybe_schedule_rendering_opportunity_after_ipc_message( &self, built_any_display_lists: bool, )

Schedule a rendering update (“update the rendering”), if necessary. This can be necessary for a couple reasons. For instance, when the DOM changes a scheduled rendering update becomes necessary if one isn’t scheduled already. Another example is if rAFs are running but no display lists are being produced. In that case the ScriptThread is responsible for scheduling animation ticks.

fn maybe_fulfill_font_ready_promises(&self, can_gc: CanGc)

Fulfill the possibly-pending pending document.fonts.ready promise if all web fonts have loaded.

fn maybe_resolve_pending_screenshot_readiness_requests(&self, can_gc: CanGc)

If any Pipelines are waiting to become ready for the purpose of taking a screenshot, check to see if the Pipeline is now ready and send a message to the Constellation, if so.

fn handle_msgs(&self, cx: &mut JSContext) -> bool

Handle incoming messages from other tasks and the task queue.

fn categorize_msg(&self, msg: &MixedMessage) -> ScriptThreadEventCategory

fn profile_event<F, R>( &self, category: ScriptThreadEventCategory, pipeline_id: Option<PipelineId>, f: F, ) -> R
where F: FnOnce() -> R,

fn handle_msg_from_constellation( &self, msg: ScriptThreadMessage, cx: &mut JSContext, )

fn handle_set_scroll_states( &self, pipeline_id: PipelineId, scroll_states: FxHashMap<ExternalScrollId, LayoutVector2D>, )

fn handle_msg_from_webgpu_server(&self, msg: WebGPUMsg, cx: &mut JSContext)

fn handle_msg_from_script(&self, msg: MainThreadScriptMsg, cx: &mut JSContext)

fn handle_msg_from_devtools( &self, msg: DevtoolScriptControlMsg, cx: &mut JSContext, )

fn handle_msg_from_image_cache(&self, response: ImageCacheResponseMessage)

fn handle_webdriver_msg( &self, pipeline_id: PipelineId, msg: WebDriverScriptCommand, can_gc: CanGc, )

pub(crate) fn handle_resize_message( &self, id: PipelineId, viewport_details: ViewportDetails, size_type: WindowSizeType, )

Batch window resize operations into a single “update the rendering” task, or, if a load is in progress, set the window size directly.

fn handle_theme_change_msg(&self, theme: Theme)

Handle changes to the theme, triggering reflow if the theme actually changed.

fn handle_exit_fullscreen(&self, id: PipelineId, cx: &mut JSContext)

pub(crate) fn spawn_pipeline(&self, new_pipeline_info: NewPipelineInfo)

fn collect_reports(&self, reports_chan: ReportsChan)

fn handle_set_throttled_in_containing_iframe_msg( &self, parent_pipeline_id: PipelineId, browsing_context_id: BrowsingContextId, throttled: bool, )

Updates iframe element after a change in visibility

fn handle_set_throttled_msg( &self, webview_id: WebViewId, pipeline_id: PipelineId, throttled: bool, )

fn handle_set_document_activity_msg( &self, id: PipelineId, activity: DocumentActivity, can_gc: CanGc, )

Handles activity change message

fn handle_focus_iframe_msg( &self, parent_pipeline_id: PipelineId, browsing_context_id: BrowsingContextId, sequence: FocusSequenceNumber, can_gc: CanGc, )

fn handle_focus_document_msg( &self, pipeline_id: PipelineId, sequence: FocusSequenceNumber, can_gc: CanGc, )

fn handle_unfocus_msg( &self, pipeline_id: PipelineId, sequence: FocusSequenceNumber, can_gc: CanGc, )

fn handle_post_message_msg( &self, pipeline_id: PipelineId, source_webview: WebViewId, source_with_ancestry: Vec<BrowsingContextId>, origin: Option<ImmutableOrigin>, source_origin: ImmutableOrigin, data: StructuredSerializedData, )

https://html.spec.whatwg.org/multipage/#window-post-message-steps

fn handle_stop_delaying_load_events_mode(&self, pipeline_id: PipelineId)

fn handle_unload_document(&self, pipeline_id: PipelineId, can_gc: CanGc)

fn handle_update_pipeline_id( &self, parent_pipeline_id: PipelineId, browsing_context_id: BrowsingContextId, webview_id: WebViewId, new_pipeline_id: PipelineId, reason: UpdatePipelineIdReason, can_gc: CanGc, )

fn handle_update_history_state_msg( &self, pipeline_id: PipelineId, history_state_id: Option<HistoryStateId>, url: ServoUrl, can_gc: CanGc, )

fn handle_remove_history_states( &self, pipeline_id: PipelineId, history_states: Vec<HistoryStateId>, )

fn handle_resize_inactive_msg( &self, id: PipelineId, new_viewport_details: ViewportDetails, )

Window was resized, but this script was not active, so don’t reflow yet

fn handle_page_headers_available( &self, webview_id: WebViewId, pipeline_id: PipelineId, metadata: Option<Metadata>, can_gc: CanGc, ) -> Option<DomRoot<ServoParser>>

We have received notification that the response associated with a load has completed. Kick off the document and frame tree creation process using the result.

fn handle_get_title_msg(&self, pipeline_id: PipelineId)

Handles a request for the window title.

fn handle_exit_pipeline_msg( &self, webview_id: WebViewId, pipeline_id: PipelineId, discard_bc: DiscardBrowsingContext, can_gc: CanGc, )

Handles a request to exit a pipeline and shut down layout.

fn handle_exit_script_thread_msg(&self, can_gc: CanGc)

Handles a request to exit the script thread and shut down layout.

pub(crate) fn handle_tick_all_animations_for_testing(id: PipelineId)

Handles animation tick requested during testing.

fn handle_web_font_loaded(&self, pipeline_id: PipelineId, _success: bool)

Handles a Web font being loaded. Does nothing if the page no longer exists.

fn handle_worklet_loaded(&self, pipeline_id: PipelineId)

Handles a worklet being loaded by triggering a relayout of the page. Does nothing if the page no longer exists.

fn handle_storage_event( &self, pipeline_id: PipelineId, storage_type: WebStorageType, url: ServoUrl, key: Option<String>, old_value: Option<String>, new_value: Option<String>, )

Notify a window of a storage event

fn handle_iframe_load_event( &self, parent_id: PipelineId, browsing_context_id: BrowsingContextId, child_id: PipelineId, can_gc: CanGc, )

Notify the containing document of a child iframe that has completed loading.

fn ask_constellation_for_top_level_info( &self, sender_webview_id: WebViewId, sender_pipeline_id: PipelineId, browsing_context_id: BrowsingContextId, ) -> Option<WebViewId>

fn load( &self, metadata: Metadata, incomplete: InProgressLoad, can_gc: CanGc, ) -> DomRoot<ServoParser>

The entry point to document loading. Defines bindings, sets up the window and document objects, parses HTML and CSS, and kicks off initial layout.

fn notify_devtools( &self, title: DOMString, url: ServoUrl, is_top_level_global: bool, (browsing_context_id, pipeline_id, worker_id, webview_id): (BrowsingContextId, PipelineId, Option<WorkerId>, WebViewId), )

fn handle_input_event( &self, webview_id: WebViewId, pipeline_id: PipelineId, event: ConstellationInputEvent, )

Queue input events for later dispatching as part of a update_the_rendering task.

fn handle_navigate_iframe( &self, parent_pipeline_id: PipelineId, browsing_context_id: BrowsingContextId, load_data: LoadData, history_handling: NavigationHistoryBehavior, can_gc: CanGc, )

Handle a “navigate an iframe” message from the constellation.

pub(crate) fn eval_js_url( global_scope: &GlobalScope, load_data: &mut LoadData, can_gc: CanGc, )

Turn javascript: URL into JS code to eval, according to the steps in https://html.spec.whatwg.org/multipage/#javascript-protocol

fn pre_page_load(&self, incomplete: InProgressLoad)

Instructs the constellation to fetch the document that will be loaded. Stores the InProgressLoad argument until a notification is received that the fetch is complete.

fn handle_navigation_response( &self, pipeline_id: PipelineId, message: FetchResponseMsg, )

fn handle_fetch_metadata( &self, id: PipelineId, request_id: RequestId, fetch_metadata: Result<FetchMetadata, NetworkError>, )

fn handle_fetch_chunk( &self, pipeline_id: PipelineId, request_id: RequestId, chunk: Vec<u8>, )

fn handle_fetch_eof( &self, id: PipelineId, request_id: RequestId, eof: Result<ResourceFetchTiming, NetworkError>, )

fn handle_csp_violations( &self, id: PipelineId, _: RequestId, violations: Vec<Violation>, )

fn handle_navigation_redirect(&self, id: PipelineId, metadata: &Metadata)

fn start_page_load_about_blank(&self, incomplete: InProgressLoad)

Synchronously fetch about:blank. Stores the InProgressLoad argument until a notification is received that the fetch is complete.

fn page_load_about_srcdoc(&self, incomplete: InProgressLoad)

Synchronously parse a srcdoc document from a giving HTML string.

fn handle_css_error_reporting( &self, pipeline_id: PipelineId, filename: String, line: u32, column: u32, msg: String, )

fn handle_reload(&self, pipeline_id: PipelineId, can_gc: CanGc)

fn handle_paint_metric( &self, pipeline_id: PipelineId, metric_type: ProgressiveWebMetricType, metric_value: CrossProcessInstant, first_reflow: bool, can_gc: CanGc, )

fn handle_media_session_action( &self, pipeline_id: PipelineId, action: MediaSessionActionType, can_gc: CanGc, )

pub(crate) fn enqueue_microtask(job: Microtask)

pub(crate) fn perform_a_microtask_checkpoint(&self, can_gc: CanGc)

fn handle_evaluate_javascript( &self, webview_id: WebViewId, pipeline_id: PipelineId, evaluation_id: JavaScriptEvaluationId, script: String, can_gc: CanGc, )

fn handle_refresh_cursor(&self, pipeline_id: PipelineId)

pub(crate) fn is_servo_privileged(url: ServoUrl) -> bool

fn handle_request_screenshot_readiness( &self, webview_id: WebViewId, pipeline_id: PipelineId, can_gc: CanGc, )

fn handle_embedder_control_response( &self, id: EmbedderControlId, response: EmbedderControlResponse, can_gc: CanGc, )