Trait script::stylesheet_loader::StylesheetOwner

pub trait StylesheetOwner {
    fn parser_inserted(&self) -> bool;
    fn referrer_policy(&self) -> Option<ReferrerPolicy>;
    fn increment_pending_loads_count(&self);
    fn load_finished(&self, successful: bool) -> Option<bool>;
}

Required Methods

fn parser_inserted(&self) -> bool

Returns whether this element was inserted by the parser (i.e., it should trigger a document-load-blocking load).

fn referrer_policy(&self) -> Option<ReferrerPolicy>

Which referrer policy should loads triggered by this owner follow, or None for the default.

fn increment_pending_loads_count(&self)

Notes that a new load is pending to finish.

fn load_finished(&self, successful: bool) -> Option<bool>

Returns None if there are still pending loads, or whether any load has failed since the loads started.