script

Module test

Modules

area

encoding_detection

size_of

srcset

textinput

timeranges

Structs

ByteString

Encapsulates the IDL ByteString type.

DOMString

This type corresponds to the DOMString type in WebIDL.

TrustedPromise

A safe wrapper around a DOM Promise object that can be shared among threads for use in asynchronous operations. The underlying DOM object is guaranteed to live at least as long as the last outstanding TrustedPromise instance. These values cannot be cloned, only created from existing Rc<Promise> values.