Struct script::test::DOMString

pub struct DOMString(String, PhantomData<*const ()>);

A DOMString.

This type corresponds to the DOMString type in WebIDL.

Conceptually, a DOMString has the same value space as a JavaScript String, i.e., an array of 16-bit code units representing UTF-16, potentially with unpaired surrogates present (also sometimes called WTF-16).

Currently, this type stores a Rust String, in order to avoid issues when integrating with the rest of the Rust ecosystem and even the rest of the browser itself.

However, Rust Strings are guaranteed to be valid UTF-8, and as such have a smaller value space than WTF-16 (i.e., some JavaScript String values can not be represented as a Rust String). This introduces the question of what to do with values being passed from JavaScript to Rust that contain unpaired surrogates.

The hypothesis is that it does not matter much how exactly those values are transformed, because passing unpaired surrogates into the DOM is very rare. In order to test this hypothesis, Servo will panic when encountering any unpaired surrogates on conversion to DOMString by default. (The command line option -Z replace-surrogates instead causes Servo to replace the unpaired surrogate by a U+FFFD replacement character.)

Currently, the lack of crash reports about this issue provides some evidence to support the hypothesis. This evidence will hopefully be used to convince other browser vendors that it would be safe to replace unpaired surrogates at the boundary between JavaScript and native code. (This would unify the DOMString and USVString types, both in the WebIDL standard and in Servo.)

This type is currently !Send, in order to help with an independent experiment to store JSStrings rather than Rust Strings.

Tuple Fields
0: String1: PhantomData<*const ()>
Implementations

impl DOMString

pub fn new() -> DOMString

Creates a new DOMString.

pub fn from_string(s: String) -> DOMString

Creates a new DOMString from a String.

pub fn push_str(&mut self, string: &str)

Appends a given string slice onto the end of this String.

pub fn clear(&mut self)

Clears this DOMString, removing all contents.

pub fn truncate(&mut self, new_len: usize)

Shortens this String to the specified length.

pub fn strip_newlines(&mut self)

Removes newline characters according to https://infra.spec.whatwg.org/#strip-newlines.

pub fn strip_leading_and_trailing_ascii_whitespace(&mut self)

Removes leading and trailing ASCII whitespaces according to https://infra.spec.whatwg.org/#strip-leading-and-trailing-ascii-whitespace.

pub fn is_valid_time_string(&self) -> bool

Validates this DOMString is a time string according to https://html.spec.whatwg.org/multipage/#valid-time-string.

pub fn is_valid_date_string(&self) -> bool

A valid date string should be “YYYY-MM-DD” YYYY must be four or more digits, MM and DD both must be two digits https://html.spec.whatwg.org/multipage/#valid-date-string

pub fn parse_date_string(&self) -> Result<(i32, u32, u32), ()>

https://html.spec.whatwg.org/multipage/#parse-a-date-string

pub fn parse_time_string(&self) -> Result<(u32, u32, f64), ()>

https://html.spec.whatwg.org/multipage/#parse-a-time-string

pub fn is_valid_month_string(&self) -> bool

A valid month string should be “YYYY-MM” YYYY must be four or more digits, MM both must be two digits https://html.spec.whatwg.org/multipage/#valid-month-string

pub fn parse_month_string(&self) -> Result<(i32, u32), ()>

https://html.spec.whatwg.org/multipage/#parse-a-month-string

pub fn is_valid_week_string(&self) -> bool

A valid week string should be like {YYYY}-W{WW}, such as “2017-W52” YYYY must be four or more digits, WW both must be two digits https://html.spec.whatwg.org/multipage/#valid-week-string

pub fn parse_week_string(&self) -> Result<(i32, u32), ()>

https://html.spec.whatwg.org/multipage/#parse-a-week-string

pub fn is_valid_floating_point_number_string(&self) -> bool

https://html.spec.whatwg.org/multipage/#valid-floating-point-number

pub fn parse_floating_point_number(&self) -> Result<f64, ()>

https://html.spec.whatwg.org/multipage/#rules-for-parsing-floating-point-number-values

pub fn set_best_representation_of_the_floating_point_number(&mut self)

https://html.spec.whatwg.org/multipage/#best-representation-of-the-number-as-a-floating-point-number

pub fn convert_valid_normalized_local_date_and_time_string(
&mut self
) -> Result<(), ()>

A valid normalized local date and time string should be “{date}T{time}” where date and time are both valid, and the time string must be as short as possible https://html.spec.whatwg.org/multipage/#valid-normalised-local-date-and-time-string

pub fn parse_local_date_and_time_string(
&self
) -> Result<((i32, u32, u32), (u32, u32, f64)), ()>

https://html.spec.whatwg.org/multipage/#parse-a-local-date-and-time-string

pub fn is_valid_email_address_string(&self) -> bool

https://html.spec.whatwg.org/multipage/#valid-e-mail-address

pub fn is_valid_simple_color_string(&self) -> bool

https://html.spec.whatwg.org/multipage/#valid-simple-colour