Module script::textinput

Common handling of keyboard input and state management for text input controls

Structs

TextInput
Encapsulated state for handling keyboard input in a single or multiline text input control.

TextPoint

Enums

Direction
The direction in which to delete a character.

KeyReaction
Resulting action to be taken by the owner of a text input that is handling an event.

Lines
Control whether this control should allow multiple lines.

Selection

SelectionDirection

Functions

is_control_key

is_printable_key

len_of_first_n_chars
The length in bytes of the first n characters in a UTF-8 string.

len_of_first_n_code_units
The length in bytes of the first n code units a string when encoded in UTF-16.