Struct script::textinput::TextInput

pub struct TextInput<T: ClipboardProvider> {
    lines: Vec<DOMString>,
    pub edit_point: TextPoint,
    pub selection_begin: Option<TextPoint>,
    multiline: bool,
    clipboard_provider: T,
    pub max_length: Option<usize>,
    pub selection_direction: SelectionDirection,
}

Encapsulated state for handling keyboard input in a single or multiline text input control.

Fields
lines: Vec<DOMString>

Current text input content, split across lines without trailing '\n'

edit_point: TextPoint

Current cursor input point

selection_begin: Option<TextPoint>

Beginning of selection range with edit_point as end that can span multiple lines.

multiline: bool

Is this a multiline input?

clipboard_provider: T max_length: Option<usize>

The maximum number of UTF-16 code units this text input is allowed to hold.

https://html.spec.whatwg.org/multipage/#attr-fe-maxlength

selection_direction: SelectionDirection
Methods

impl<T: ClipboardProvider> TextInput<T>

fn new(lines: Lines, initial: DOMString, clipboard_provider: T, max_length: Option<usize>, selection_direction: SelectionDirection) -> TextInput<T>

Instantiate a new text input control

fn delete_char(&mut self, dir: Direction)

Remove a character at the current editing point

fn insert_char(&mut self, ch: char)

Insert a character at the current editing point

fn insert_string<S: Into<String>>(&mut self, s: S)

Insert a string at the current editing point

fn get_sorted_selection(&self) -> Option<(TextPoint, TextPoint)>

fn get_absolute_selection_range(&self) -> Range<usize>

Return the selection range as UTF-8 byte offsets from the start of the content.

If there is no selection, returns an empty range at the insertion point.

fn get_selection_text(&self) -> Option<String>

fn selection_utf16_len(&self) -> usize

The length of the selected text in UTF-16 code units.

fn fold_selection_slices<B, F: FnMut(&mut B, &str)>(&self, acc: B, f: F) -> B

Run the callback on a series of slices that, concatenated, make up the selected text.

The accumulator acc can be mutated by the callback, and will be returned at the end.

fn replace_selection(&mut self, insert: DOMString)

fn current_line_length(&self) -> usize

Return the length in UTF-8 bytes of the current line under the editing point.

fn adjust_vertical(&mut self, adjust: isize, select: Selection)

Adjust the editing point position by a given of lines. The resulting column is as close to the original column position as possible.

fn adjust_horizontal(&mut self, adjust: isize, select: Selection)

Adjust the editing point position by a given number of bytes. If the adjustment requested is larger than is available in the current line, the editing point is adjusted vertically and the process repeats with the remaining adjustment requested.

fn adjust_horizontal_by_one(&mut self, direction: Direction, select: Selection)

fn adjust_selection_for_horizontal_change(&mut self, adjust: Direction, select: Selection) -> bool

Return whether to cancel the caret move

fn perform_horizontal_adjustment(&mut self, adjust: isize, select: Selection)

fn handle_return(&mut self) -> KeyReaction

Deal with a newline input.

fn select_all(&mut self)

Select all text in the input control.

fn clear_selection(&mut self)

Remove the current selection.

fn handle_keydown(&mut self, event: &KeyboardEvent) -> KeyReaction

Process a given KeyboardEvent and return an action for the caller to execute.

fn handle_keydown_aux(&mut self, key: Key, mods: KeyModifiers) -> KeyReaction

fn is_empty(&self) -> bool

Whether the content is empty.

fn len(&self) -> usize

The length of the content in bytes.

fn utf16_len(&self) -> usize

The length of the content in bytes.

fn char_count(&self) -> usize

The length of the content in chars.

fn get_content(&self) -> DOMString

Get the current contents of the text input. Multiple lines are joined by \n.

fn set_content(&mut self, content: DOMString)

Set the current contents of the text input. If this is control supports multiple lines, any \n encountered will be stripped and force a new logical line.

fn get_absolute_insertion_point(&self) -> usize

Get the insertion point as a byte offset from the start of the content.

fn get_absolute_point_for_text_point(&self, text_point: &TextPoint) -> usize

Convert a TextPoint into a byte offset from the start of the content.

fn get_text_point_for_absolute_point(&self, abs_point: usize) -> TextPoint

Convert a byte offset from the start of the content into a TextPoint.

fn set_selection_range(&mut self, start: u32, end: u32)

fn get_selection_start(&self) -> u32