Struct script::textinput::TextInput

pub struct TextInput<T: ClipboardProvider> {
    lines: Vec<DOMString>,
    edit_point: TextPoint,
    selection_origin: Option<TextPoint>,
    selection_direction: SelectionDirection,
    multiline: bool,
    clipboard_provider: T,
    max_length: Option<UTF16CodeUnits>,
    min_length: Option<UTF16CodeUnits>,
}

Encapsulated state for handling keyboard input in a single or multiline text input control.

Fields
lines: Vec<DOMString>

Current text input content, split across lines without trailing '\n'

edit_point: TextPoint

Current cursor input point

selection_origin: Option<TextPoint>

The current selection goes from the selection_origin until the edit_point. Note that the selection_origin may be after the edit_point, in the case of a backward selection.

selection_direction: SelectionDirectionmultiline: bool

Is this a multiline input?

clipboard_provider: Tmax_length: Option<UTF16CodeUnits>

The maximum number of UTF-16 code units this text input is allowed to hold.

https://html.spec.whatwg.org/multipage/#attr-fe-maxlength

min_length: Option<UTF16CodeUnits>
Methods

impl<T: ClipboardProvider> TextInput<T>

pub fn new(
lines: Lines,
initial: DOMString,
clipboard_provider: T,
max_length: Option<UTF16CodeUnits>,
min_length: Option<UTF16CodeUnits>,
selection_direction: SelectionDirection
) -> TextInput<T>

Instantiate a new text input control

pub fn edit_point(&self) -> TextPoint

pub fn selection_origin(&self) -> Option<TextPoint>

pub fn selection_origin_or_edit_point(&self) -> TextPoint

The selection origin, or the edit point if there is no selection. Note that the selection origin may be after the edit point, in the case of a backward selection.

pub fn selection_direction(&self) -> SelectionDirection

pub fn set_max_length(&mut self, length: Option<UTF16CodeUnits>)

pub fn set_min_length(&mut self, length: Option<UTF16CodeUnits>)

pub fn delete_char(&mut self, dir: Direction)

Remove a character at the current editing point

pub fn insert_char(&mut self, ch: char)

Insert a character at the current editing point

pub fn insert_string<S: Into<String>>(&mut self, s: S)

Insert a string at the current editing point

pub fn selection_start(&self) -> TextPoint

The start of the selection (or the edit point, if there is no selection). Always less than or equal to selection_end(), regardless of the selection direction.

pub fn selection_start_offset(&self) -> UTF8Bytes

The byte offset of the selection_start()

pub fn selection_end(&self) -> TextPoint

The end of the selection (or the edit point, if there is no selection). Always greater than or equal to selection_start(), regardless of the selection direction.

pub fn selection_end_offset(&self) -> UTF8Bytes

The byte offset of the selection_end()

pub fn has_selection(&self) -> bool

Whether or not there is an active selection (the selection may be zero-length)

pub fn sorted_selection_bounds(&self) -> (TextPoint, TextPoint)

Returns a tuple of (start, end) giving the bounds of the current selection. start is always less than or equal to end.

pub fn sorted_selection_offsets_range(&self) -> Range<UTF8Bytes>

Return the selection range as byte offsets from the start of the content.

If there is no selection, returns an empty range at the edit point.

pub fn selection_state(&self) -> SelectionState

The state of the current selection. Can be used to compare whether selection state has changed.

fn assert_ok_selection(&self)

pub fn get_selection_text(&self) -> Option<String>

fn selection_utf16_len(&self) -> UTF16CodeUnits

The length of the selected text in UTF-16 code units.

fn fold_selection_slices<B, F: FnMut(&mut B, &str)>(&self, acc: B, f: F) -> B

Run the callback on a series of slices that, concatenated, make up the selected text.

The accumulator acc can be mutated by the callback, and will be returned at the end.

pub fn replace_selection(&mut self, insert: DOMString)

pub fn current_line_length(&self) -> UTF8Bytes

Return the length in bytes of the current line under the editing point.

pub fn adjust_vertical(&mut self, adjust: isize, select: Selection)

Adjust the editing point position by a given number of lines. The resulting column is as close to the original column position as possible.

pub fn adjust_horizontal(
&mut self,
adjust: UTF8Bytes,
direction: Direction,
select: Selection
)

Adjust the editing point position by a given number of bytes. If the adjustment requested is larger than is available in the current line, the editing point is adjusted vertically and the process repeats with the remaining adjustment requested.

pub fn adjust_horizontal_by_one(
&mut self,
direction: Direction,
select: Selection
)

Adjust the editing point position by exactly one grapheme cluster. If the edit point is at the beginning of the line and the direction is "Backward" or the edit point is at the end of the line and the direction is "Forward", a vertical adjustment is made

fn adjust_selection_for_horizontal_change(
&mut self,
adjust: Direction,
select: Selection
) -> bool

Return whether to cancel the caret move

fn update_selection_direction(&mut self)

Update the field selection_direction.

When the edit_point (or focus) is before the selection_origin (or anchor) you have a backward selection. Otherwise you have a forward selection.

fn perform_horizontal_adjustment(
&mut self,
adjust: UTF8Bytes,
direction: Direction,
select: Selection
)

pub fn handle_return(&mut self) -> KeyReaction

Deal with a newline input.

pub fn select_all(&mut self)

Select all text in the input control.

pub fn clear_selection(&mut self)

Remove the current selection.

pub fn clear_selection_to_limit(&mut self, direction: Direction)

Remove the current selection and set the edit point to the end of the content.

pub fn adjust_horizontal_by_word(
&mut self,
direction: Direction,
select: Selection
)

pub fn adjust_horizontal_to_line_end(
&mut self,
direction: Direction,
select: Selection
)

pub fn adjust_horizontal_to_limit(
&mut self,
direction: Direction,
select: Selection
)

pub fn handle_keydown(&mut self, event: &KeyboardEvent) -> KeyReaction

Process a given KeyboardEvent and return an action for the caller to execute.

pub fn handle_keydown_aux(
&mut self,
key: Key,
mods: Modifiers,
macos: bool
) -> KeyReaction

pub fn handle_compositionend(&mut self, event: &CompositionEvent) -> KeyReaction

pub fn is_empty(&self) -> bool

Whether the content is empty.

pub fn len_utf8(&self) -> UTF8Bytes

The length of the content in bytes.

pub fn utf16_len(&self) -> UTF16CodeUnits

The total number of code units required to encode the content in utf16.

pub fn char_count(&self) -> usize

The length of the content in Unicode code points.

pub fn get_content(&self) -> DOMString

Get the current contents of the text input. Multiple lines are joined by \n.

pub fn single_line_content(&self) -> &DOMString

Get a reference to the contents of a single-line text input. Panics if self is a multiline input.

pub fn set_content(&mut self, content: DOMString)

Set the current contents of the text input. If this is control supports multiple lines, any \n encountered will be stripped and force a new logical line.

fn text_point_to_offset(&self, text_point: &TextPoint) -> UTF8Bytes

Convert a TextPoint into a byte offset from the start of the content.

fn offset_to_text_point(&self, abs_point: UTF8Bytes) -> TextPoint

Convert a byte offset from the start of the content into a TextPoint.

pub fn set_selection_range(
&mut self,
start: u32,
end: u32,
direction: SelectionDirection
)

pub fn set_edit_point_index(&mut self, index: usize)

Set the edit point index position based off of a given grapheme cluster offset