Struct script::textinput::TextInput

pub struct TextInput {
    // some fields omitted
}

Encapsulated state for handling keyboard input in a single or multiline text input control.

Methods

impl TextInput

fn new(lines: Lines, initial: DOMString) -> TextInput

Instantiate a new text input control

fn handle_keydown(&mut self, event: JSRef<KeyboardEvent>) -> KeyReaction

Process a given KeyboardEvent and return an action for the caller to execute.

fn get_content(&self) -> DOMString

Get the current contents of the text input. Multiple lines are joined by \n.

fn set_content(&mut self, content: DOMString)

Set the current contents of the text input. If this is control supports multiple lines, any \n encountered will be stripped and force a new logical line.