Struct script::timers::OneshotTimers

pub struct OneshotTimers {
    js_timers: JsTimers,
    timer_event_chan: IpcSender<TimerEvent>,
    scheduler_chan: IpcSender<TimerEventRequest>,
    next_timer_handle: Cell<OneshotTimerHandle>,
    timers: DOMRefCell<Vec<OneshotTimer>>,
    suspended_since: Cell<Option<MsDuration>>,
    suspension_offset: Cell<MsDuration>,
    expected_event_id: Cell<TimerEventId>,
}

Fields

js_timers

timer_event_chan

scheduler_chan

next_timer_handle

timers

suspended_since

suspension_offset

Initially 0, increased whenever the associated document is reactivated by the amount of ms the document was inactive. The current time can be offset back by this amount for a coherent time across document activations.

expected_event_id

Calls to fire_timer with a different argument than this get ignored. They were previously scheduled and got invalidated when - timers were suspended, - the timer it was scheduled for got canceled or - a timer was added with an earlier callback time. In this case the original timer is rescheduled when it is the next one to get called.

Methods

impl OneshotTimers

fn new(timer_event_chan: IpcSender<TimerEvent>, scheduler_chan: IpcSender<TimerEventRequest>) -> OneshotTimers

fn schedule_callback(&self, callback: OneshotTimerCallback, duration: MsDuration, source: TimerSource) -> OneshotTimerHandle

fn unschedule_callback(&self, handle: OneshotTimerHandle)

fn is_next_timer(&self, handle: OneshotTimerHandle) -> bool

fn fire_timer<T: Reflectable>(&self, id: TimerEventId, this: &T)

fn base_time(&self) -> MsDuration

fn suspend(&self)

fn resume(&self)

fn schedule_timer_call(&self)

fn invalidate_expected_event_id(&self) -> TimerEventId

fn set_timeout_or_interval(&self, global: GlobalRef, callback: TimerCallback, arguments: Vec<HandleValue>, timeout: i32, is_interval: IsInterval, source: TimerSource) -> i32

fn clear_timeout_or_interval(&self, global: GlobalRef, handle: i32)