Struct script::timers::OneshotTimers

pub struct OneshotTimers {
    js_timers: JsTimers,
    timer_event_chan: DomRefCell<Option<IpcSender<TimerEvent>>>,
    scheduler_chan: IpcSender<TimerSchedulerMsg>,
    next_timer_handle: Cell<OneshotTimerHandle>,
    timers: DomRefCell<Vec<OneshotTimer>>,
    suspended_since: Cell<Option<MsDuration>>,
    suspension_offset: Cell<MsDuration>,
    expected_event_id: Cell<TimerEventId>,
}

Fields
js_timers: JsTimerstimer_event_chan: DomRefCell<Option<IpcSender<TimerEvent>>>

The sender, to be cloned for each timer, on which the timer scheduler in the constellation can send an event when the timer is due.

scheduler_chan: IpcSender<TimerSchedulerMsg>

The sender to the timer scheduler in the constellation.

next_timer_handle: Cell<OneshotTimerHandle>timers: DomRefCell<Vec<OneshotTimer>>suspended_since: Cell<Option<MsDuration>>suspension_offset: Cell<MsDuration>

Initially 0, increased whenever the associated document is reactivated by the amount of ms the document was inactive. The current time can be offset back by this amount for a coherent time across document activations.

expected_event_id: Cell<TimerEventId>

Calls to fire_timer with a different argument than this get ignored. They were previously scheduled and got invalidated when

timers were suspended,

the timer it was scheduled for got canceled or

a timer was added with an earlier callback time. In this case the original timer is rescheduled when it is the next one to get called.

Implementations

impl OneshotTimers

pub fn new(scheduler_chan: IpcSender<TimerSchedulerMsg>) -> OneshotTimers

pub fn setup_scheduling(&self, timer_event_chan: IpcSender<TimerEvent>)

pub fn schedule_callback(
&self,
callback: OneshotTimerCallback,
duration: MsDuration,
source: TimerSource
) -> OneshotTimerHandle

pub fn unschedule_callback(&self, handle: OneshotTimerHandle)

fn is_next_timer(&self, handle: OneshotTimerHandle) -> bool

pub fn fire_timer(&self, id: TimerEventId, global: &GlobalScope)

fn base_time(&self) -> MsDuration

pub fn slow_down(&self)

pub fn speed_up(&self)

pub fn suspend(&self)

pub fn resume(&self)

fn schedule_timer_call(&self)

fn invalidate_expected_event_id(&self) -> TimerEventId

pub fn set_timeout_or_interval(
&self,
global: &GlobalScope,
callback: TimerCallback,
arguments: Vec<HandleValue>,
timeout: i32,
is_interval: IsInterval,
source: TimerSource
) -> i32

pub fn clear_timeout_or_interval(&self, global: &GlobalScope, handle: i32)